% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_variables.R
\name{a_summary_internal}
\alias{a_summary_internal}
\title{Helper Function to Create Output Rows for \code{\link[=a_summary]{a_summary()}}}
\usage{
a_summary_internal(
  x,
  .N_col,
  .N_row,
  .var,
  .df_row,
  .ref_group,
  .in_ref_col,
  compare,
  type,
  .stats,
  .formats,
  .labels,
  .indent_mods,
  na.rm,
  na_level,
  ...
)
}
\arguments{
\item{compare}{(\code{logical})\cr Whether comparison statistics should be analyzed instead of summary statistics
(\code{compare = TRUE} adds \code{pval} statistic comparing against reference group).}

\item{type}{(\code{character})\cr type of statistics to calculate given \code{x}. If \code{x} is numeric \code{type} should be
\code{"numeric"}, otherwise type should be \code{"counts"}.}
}
\value{
\itemize{
\item \code{a_summary_internal()} returns a corresponding list with formatted \code{\link[rtables:CellValue]{rtables::CellValue()}} used within \code{a_summary}.
}
}
\description{
Helper Function to Create Output Rows for \code{\link[=a_summary]{a_summary()}}
}
\note{
\itemize{
\item To use for comparison (with additional p-value statistic), parameter \code{compare} must be set to \code{TRUE}.
\item Ensure that either all \code{NA} values are converted to an explicit \code{NA} level or all \code{NA} values are left as is.
}
}
\keyword{internal}
