% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_labels}
\alias{add_labels}
\title{re-attach labels to variables}
\usage{
add_labels(df, labels)
}
\arguments{
\item{df}{a data.frame}

\item{labels}{a named vector with the labels. names are variable names}
}
\description{
re-attach labels to variables
}
\examples{

df <- data.frame(c = c(1, 2), b = c("a", "b"), a = c(3,4))

labels <- setNames(c("a var", "b var"), c("a", "b"))


X <- add_labels(df, labels)

View(X)

}
