% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_diff.R
\name{add_risk_diff}
\alias{add_risk_diff}
\title{Split Function to Configure Risk Difference Column}
\usage{
add_risk_diff(arm_x, arm_y, col_label = "Risk Difference (\%) (95\% CI)")
}
\description{
Split Function to Configure Risk Difference Column
}
