% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{afun_selected_stats}
\alias{afun_selected_stats}
\title{Get Selected Statistics Names}
\usage{
afun_selected_stats(.stats, all_stats)
}
\arguments{
\item{.stats}{(\code{vector} or \code{NULL})\cr input to the layout creating function. Note that \code{NULL} means
in this context that all default statistics should be used.}

\item{all_stats}{(\code{character})\cr all statistics which can be selected here potentially.}
}
\value{
Character vector with the selected statistics.
}
\description{
Helper function to be used for creating \code{afun}.
}
\keyword{internal}
