% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_colvars_functions.R
\name{analyze_colvars_functions}
\alias{analyze_colvars_functions}
\title{Analyze Functions on Columns}
\description{
These functions are wrappers of \code{\link[rtables:analyze_colvars]{rtables::analyze_colvars()}} which apply corresponding \code{tern} statistics functions
to add an analysis to a given table layout:
\itemize{
\item \code{\link[=analyze_vars_in_cols]{analyze_vars_in_cols()}} (extended wrapper, never used in other \code{tern} functions)
\item \code{\link[=summarize_colvars]{summarize_colvars()}} (low level wrapper, never used in other \code{tern} functions)
\item \code{\link[=summarize_coxreg]{summarize_coxreg()}} (if the analysis is multivariate)
\item \code{\link[=tabulate_rsp_subgroups]{tabulate_rsp_subgroups()}}
\item \code{\link[=tabulate_survival_subgroups]{tabulate_survival_subgroups()}}
}
}
\seealso{
\itemize{
\item \link{summarize_functions} for functions which are wrappers for \code{\link[rtables:summarize_row_groups]{rtables::summarize_row_groups()}}.
\item \link{analyze_functions} for functions which are wrappers for \code{\link[rtables:analyze]{rtables::analyze()}}.
}
}
