% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{argument_convention}
\alias{argument_convention}
\title{Standard Arguments}
\arguments{
\item{...}{additional arguments for the lower level functions.}

\item{.df_row}{(\code{data.frame})\cr data frame across all of the columns for the given row split.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}

\item{.N_col}{(\code{count})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is passed by \code{rtables}.}

\item{.N_row}{(\code{count})\cr column-wise N (column count) for the full column that is passed by \code{rtables}.}

\item{.ref_group}{(\code{data.frame} or \code{vector})\cr the data corresponding to the reference group.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels. Defaults to 0, which corresponds to the
unmodified default behavior. Can be negative.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.spl_context}{(\code{data.frame})\cr gives information about ancestor split states
that is passed by \code{rtables}.}

\item{add_total_level}{(\code{flag})\cr adds a "total" level after the others which includes all the levels
that constitute the split. A custom label can be set for this level via the \code{custom_label} argument.}

\item{col_by}{(\code{factor})\cr defining column groups.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{data}{(\code{data.frame})\cr the dataset containing the variables to summarize.}

\item{df}{(\code{data.frame})\cr data set containing all analysis variables.}

\item{draw}{(\code{flag})\cr whether the plot should be drawn.}

\item{drop}{(\code{flag})\cr should non appearing occurrence levels be dropped from the resulting table.
Note that in that case the remaining occurrence levels in the table are sorted alphabetically.}

\item{id}{(\code{string})\cr subject variable name.}

\item{is_event}{(\code{logical})\cr \code{TRUE} if event, \code{FALSE} if time to event is censored.}

\item{indent_mod}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use the \code{.indent_mods} argument instead.}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions). See \code{\link[rtables:summarize_row_groups]{rtables::summarize_row_groups()}}
for more information.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{na.rm}{(\code{flag})\cr whether \code{NA} values should be removed from \code{x} prior to analysis.}

\item{na_level}{(\code{string})\cr used to replace all \code{NA} or empty values in factors with custom \code{string}.}

\item{nested}{(\code{flag})\cr whether this layout instruction be applied within the existing layout structure \emph{if
possible} (\code{TRUE}, the default) or as a new top-level element (\code{FALSE}). Ignored if it would nest a split
underneath analyses, which is not allowed.}

\item{newpage}{(\code{flag})\cr whether the plot should be drawn on a new page.
Only considered if \code{draw = TRUE} is used.}

\item{prune_zero_rows}{(\code{flag})\cr whether to prune all zero rows.}

\item{rsp}{(\code{logical})\cr whether each subject is a responder or not.}

\item{show_labels}{(\code{string})\cr label visibility: one of "default", "visible" and "hidden".}

\item{section_div}{(\code{string})\cr string which should be repeated as a section divider after each group
defined by this split instruction, or \code{NA_character_} (the default) for no section divider.}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{tte}{(\code{numeric})\cr contains time-to-event duration values.}

\item{var_labels}{(\code{character})\cr character for label.}

\item{variables}{(named \code{list} of \code{string})\cr list of additional analysis variables.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{var}{(\code{string})\cr single variable name for the primary analysis variable.}

\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}
}
\description{
The documentation to this function lists all the arguments in \code{tern}
that are used repeatedly to express an analysis.
}
\details{
Although this function just returns \code{NULL} it has two uses, for
the \code{tern} users it provides a documentation of arguments that are
commonly and consistently used in the framework. For the developer it adds a
single reference point to import the \code{roxygen} argument description with:
\verb{@inheritParams argument_convention}
}
\keyword{internal}
