% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arm}
\alias{arm}
\title{Indicate Arm Variable in Formula}
\usage{
arm(x)
}
\arguments{
\item{x}{arm information}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

We use \code{arm} to indicate the study arm variable in \code{tern} formulas.
}
