% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot_regular.R
\name{arm_for_model2}
\alias{arm_for_model2}
\title{update Function from tm_forestplot.R. Suggest to put function arm_for_model2 in R script for uitility functions
re-factor arm variable by input arm.ref and arm.rest}
\usage{
arm_for_model2(arm, arm.ref, arm.rest)
}
\description{
update Function from tm_forestplot.R. Suggest to put function arm_for_model2 in R script for uitility functions
re-factor arm variable by input arm.ref and arm.rest
}
