% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_grid.R
\name{arrange_grobs}
\alias{arrange_grobs}
\title{Arrange Multiple Grobs}
\usage{
arrange_grobs(
  ...,
  grobs = list(...),
  ncol = NULL,
  nrow = NULL,
  padding_ht = grid::unit(2, "line"),
  padding_wt = grid::unit(2, "line"),
  vp = NULL,
  gp = NULL,
  name = NULL
)
}
\arguments{
\item{...}{grobs.}

\item{grobs}{list of grobs.}

\item{ncol}{number of columns in layout.}

\item{nrow}{number of rows in layout.}

\item{padding_ht}{unit of length 1, vertical space between each grob.}

\item{padding_wt}{unit of length 1, horizontal space between each grob.}

\item{vp}{a \code{\link[=viewport]{viewport()}} object (or \code{NULL}).}

\item{gp}{A \code{\link[=gpar]{gpar()}} object.}

\item{name}{a character identifier for the grob.}
}
\value{
A \code{grob}.
}
\description{
Arrange grobs as a new grob with \verb{n*m (rows*cols)} layout.
}
\keyword{internal}
