% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_mmrm.R
\name{as.rtable.data.frame}
\alias{as.rtable.data.frame}
\title{Helper}
\usage{
\method{as.rtable}{data.frame}(x, format = "xx.xx")
}
\arguments{
\item{x}{the data frame}

\item{format}{the format}
}
\value{
the \code{rtables} object
}
\description{
Helper method to convert a \code{data.frame} into \code{rtable}.
}
\note{
This could eventually be moved to the \code{rtables} package. Currently we are not doing that
  because of the upcoming major change of that package.
}
