% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as.rtable.table}
\alias{as.rtable.table}
\title{convert a table to an rtable}
\usage{
\method{as.rtable}{table}(x, format = "xx")
}
\description{
convert a table to an rtable
}
\examples{
Viewer(as.rtable(table(iris$Species)))

Viewer(
  as.rtable(table(sample(letters[1:4], 200, TRUE), sample(LETTERS[1:4], 200, TRUE))),
  row.names.bold = TRUE
)

}
