% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_factor_keep_attributes}
\alias{as_factor_keep_attributes}
\title{Convert a Vector to a Factor}
\usage{
as_factor_keep_attributes(x)
}
\arguments{
\item{x}{atomic}
}
\value{
factor with same attributes (except class) as x
}
\description{
Converts \code{x} to a factor and keeps its attributes.
}
\examples{
as_factor_keep_attributes(with_label(c(1,1,2,3), "id"))

}
