% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtable.R
\name{as_html}
\alias{as_html}
\title{Convert an \code{\link{rtable}} object to an \code{shiny.tag} html
representation of the \code{\link{rtable}}}
\usage{
as_html(x, ...)
}
\arguments{
\item{x}{an \code{\link{rtable}} object}

\item{...}{arguments passed as attributes to the table html objet}
}
\value{
an object of class \code{shiny.tag}
}
\description{
The returned html object can be immediately used in shiny and rmarkdown
}
