% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{basic_node_info}
\alias{basic_node_info}
\alias{basic_node_info,node-method}
\title{Object summary}
\usage{
basic_node_info(x, ...)

\S4method{basic_node_info}{node}(x, index = numeric(0))
}
\arguments{
\item{x}{node object}

\item{...}{other arguments}

\item{index}{path to node (excluding node name itself)}
}
\description{
Summarizes an object

Summarizes a tree of a tree using depth-first.
It displays the dimension and class of each node in the tree recursively.
}
\examples{
n11 <- node(name = "A", content = array(c(1:6), dim = c(2,3)), children = list())
n2 <- node(name = "D", content = c(1:3), children = list(
  node(name = "A", content = array(c(1:6), dim = c(2,3)), children = list()),
  node(name = "B", content = array(c(1:6), dim = c(2,3)), children = list()),
  node(name = "C", content = array(c(1:6), dim = c(2,3)), children = list())
))
summary(n11)
summary(n2)

}
