% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_factor.R
\name{bins_percent_labels}
\alias{bins_percent_labels}
\title{Labels for Bins in Percent}
\usage{
bins_percent_labels(probs, digits = 0)
}
\arguments{
\item{probs}{(\code{proportion} vector)\cr the probabilities identifying the quantiles.
This is a sorted vector of unique \code{proportion} values, i.e. between 0 and 1, where
the boundaries 0 and 1 must not be included.}

\item{digits}{(\code{integer})\cr number of decimal places to round the percent numbers.}
}
\value{
Character vector with labels in the format \verb{[0\%,20\%]}, \verb{(20\%,50\%]}, etc.
}
\description{
This creates labels for quantile based bins in percent. This assumes the right-closed
intervals as produced by \code{\link[=cut_quantile_bins]{cut_quantile_bins()}}.
}
\examples{
# Internal function - bins_percent_labels
\dontrun{
# Just pass the internal probability bounds, then 0 and 100\% will be added automatically.
bins_percent_labels(c(0.2, 0.5))

# Determine how to round.
bins_percent_labels(0.35224, digits = 1)

# Passing an empty vector just gives a single bin 0-100\%.
bins_percent_labels(c(0, 1))
}

}
\keyword{internal}
