% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_table.R
\name{calc_diffor}
\alias{calc_diffor}
\title{Calculate difference in rates and odds ratio between two lists produced by calc_rate function}
\usage{
calc_diffor(comp, ref)
}
\arguments{
\item{comp}{list produced by calc_rate function for comparison arm}

\item{ref}{list produced by calc_rate function for reference arm}
}
\description{
Calculate difference in rates and odds ratio between two lists produced by calc_rate function
}
\examples{
a <- LETTERS[sample(1:3, 20, replace=T)]
b <- LETTERS[sample(1:3, 20, replace=T)]
a_rate <- calc_rate(a, value = "A")
b_rate <- calc_rate(b, value = "A")
calc_diffor(b_rate, a_rate)

}
\author{
Chendi Liao (liaoc10), \email{chendi.liao@roche.com}
}
