% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_response_table.R
\name{calc_rate}
\alias{calc_rate}
\title{Function to calculate counts and percentage for rates}
\usage{
calc_rate(x, value)
}
\arguments{
\item{x}{vector with response information}

\item{value}{name of the response value to compute statistics for}
}
\description{
Function to calculate counts and percentage for rates
}
\examples{
calc_rate(LETTERS[sample(1:3, 10, replace=T)], value = "A")

}
\author{
Chendi Liao (liaoc10), \email{chendi.liao@roche.com}
}
