% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_mmrm.R
\name{cbind.rtable}
\alias{cbind.rtable}
\title{Helper function to \code{cbind} multiple \code{rtable} objects together.}
\usage{
\method{cbind}{rtable}(...)
}
\arguments{
\item{...}{all \code{rtable} objects}
}
\value{
the resulting \code{rtable}
}
\description{
Helper function to \code{cbind} multiple \code{rtable} objects together.
}
\note{
This could eventually be moved to the \code{rtables} package. Currently we are not doing that
  because of the upcoming major change of that package.
}
