% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae_ctc.R
\name{cbind_rtables}
\alias{cbind_rtables}
\title{cbind two rtables}
\usage{
cbind_rtables(x, y)
}
\description{
cbind two rtables
}
\examples{
x <- rtable(c("A", "B"), rrow("x row 1", 1,2), rrow("x row 2", 3, 4))

y <- rtable("C", rrow("y row 1", 5), rrow("y row 2", 6))


cbind_rtables(x, y)

}
