% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{cfun_by_flag}
\alias{cfun_by_flag}
\title{Constructor for Content Functions given Data Frame with Flag Input}
\usage{
cfun_by_flag(analysis_var, flag_var, format = "xx")
}
\arguments{
\item{analysis_var}{(\code{string})\cr variable name for the column containing values to be
returned by the content function.}

\item{flag_var}{(\code{string})\cr variable name for the logical column identifying which
row should be returned.}

\item{format}{(\code{string})\cr rtables format to use.}
}
\value{
Content function which just gives \code{df$analysis_var} at the row identified by
\code{.df_row$flag} in the given format.
}
\description{
This can be useful for tabulating model results.
}
