% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_col_by_factor}
\alias{check_col_by_factor}
\title{Check an Arm Variable}
\usage{
check_col_by_factor(x, col_by, col_N, min_num_levels = 2)
}
\arguments{
\item{x}{(\code{data.frame} or \code{vector})\cr input data.}

\item{col_by}{(\code{factor})\cr defining column groups.}

\item{col_N}{(\code{numeric})\cr explicit column counts.}

\item{min_num_levels}{(\code{count})\cr minimum number of levels that \code{col_by} needs to have.}
}
\value{
Nothing, just passes without error if everything is ok.
}
\description{
Check an Arm Variable
}
