% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_level.R
\name{check_conf_level}
\alias{check_conf_level}
\title{Check if Argument Descibes Confidence Levels for Different Confidence Intervals and Tests}
\usage{
check_conf_level(conf_level, type = NULL)
}
\arguments{
\item{conf_level}{a single value or a named vecotr with values between 0 and 1}

\item{type}{names expected in the returned values}
}
\value{
a named vector with the confidence levels
}
\description{
Often a table contains multiple confidence intervals or p-values from tests.
}
\examples{

tern:::check_conf_level(0.95, NULL)
tern:::check_conf_level(0.95, c("survfit", "coxph", "ztest"))
tern:::check_conf_level(conf_level = c(coxph = 0.99, survfit = 0.95, ztest = 0.98),
                        type = c("survfit", "coxph", "ztest"))

\dontrun{
tern:::check_conf_level(c(survfit = 0.95, ztest = 0.98),
                        c("survfit", "coxph", "ztest"))
}
}
