% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data_frame}
\alias{check_data_frame}
\title{Check strata_data}
\usage{
check_data_frame(x, allow_missing = FALSE)
}
\arguments{
\item{x}{data.frame with valid stratification data}

\item{allow_missing}{boolean whether missing data is allowed in the variables}
}
\description{
strata_data can be \code{NULL} or
}
