% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_events_term_grade_id.R
\name{check_id}
\alias{check_id}
\title{checks that each patient appears in at most one col_by column (possibly
several times as AVAL corresponds to several measures and there are >= 1 rows per patient)}
\usage{
check_id(id, col_by)
}
\description{
checks that each patient appears in at most one col_by column (possibly
several times as AVAL corresponds to several measures and there are >= 1 rows per patient)
}
\examples{
check_id(id = ADAE$USUBJID, col_by = ADAE$ARM)
}
