% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mmrm.R
\name{check_mmrm_vars}
\alias{check_mmrm_vars}
\title{Helper function to check the MMRM variables}
\usage{
check_mmrm_vars(vars, data)
}
\arguments{
\item{vars}{variable list as per \code{\link{s_mmrm}}}

\item{data}{data frame that includes the variables}
}
\value{
nothing, this function will stop in case of problems, and pass when everything is ok.
}
\description{
Helper function to check the MMRM variables
}
