% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_numeric_range}
\alias{check_numeric_range}
\title{check all numbers are in desired range}
\usage{
check_numeric_range(x, min = 0, max = 1)
}
\arguments{
\item{x}{a numeric object}

\item{min}{a numeric value specifying the lower boundary (exclusive), with default value \code{0}}

\item{max}{a numeric value specifying the upper boundary (exclusive), with default value \code{1}}
}
\description{
check all numbers are in desired range
}
\examples{
conf.init <- 0.7
tern:::check_numeric_range(conf.init)

# below produces error
quantiles <- c(0, 1.2)
\donttest{
tern:::check_numeric_range(quantiles)
}
}
