% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_same_n}
\alias{check_same_n}
\title{Check element dimension}
\usage{
check_same_n(..., omit_null = TRUE)
}
\arguments{
\item{...}{(\code{data.frame}s or \code{vector}s)\cr any data frames or vectors.}

\item{omit_null}{(\code{flag})\cr whether \code{NULL} elements in \code{...} should be omitted from the check.}
}
\value{
A \code{logical} value.
}
\description{
Checks if the elements in \code{...} have the same dimension.
}
\keyword{internal}
