% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_strata}
\alias{check_strata}
\title{Check Stratification Data}
\usage{
check_strata(strata_data)
}
\arguments{
\item{strata_data}{data.frame with valid stratification data}
}
\description{
Checks whether \code{strata_data} is either \code{NULL} (i.e. not provided) or a valid stratification
data input.
}
