% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chgfbl_table.R
\name{chgfbl_plot}
\alias{chgfbl_plot}
\title{Plotting data generated from \code{chgfbl_data} into change from baseline
mean plot using ggplot2 and gridExtra}
\usage{
chgfbl_plot(data, ytype = "CHG", errbar = "SE", ref_line = NULL,
  yrange = NULL, fontsize = 16, n_rotate = TRUE)
}
\arguments{
\item{data}{data generated by \code{chgfbl_data} function}

\item{ytype}{type of value to plot for y-axis, possible values can be "AVAL",
"CHG" or "PCHG"}

\item{errbar}{type of error bar to draw, possible values can be "SE", "SD",
"95CI"}

\item{ref_line}{horizontal reference lines, must be either \code{NULL}
(default), a single number or a numeric vector}

\item{yrange}{Range of y-axis to display in plot. If empty (default), range
automatically controlled by ggplot2. Or must be a numeric vector of length
of 2, with the first number smaller than second number.}

\item{fontsize}{Font size of text to be displayed on the plot}

\item{n_rotate}{logical value, \code{FALSE} is default, if \code{TRUE} text
in the frequency table is rotated 90 degrees counterclockwise.}
}
\description{
Plotting data generated from \code{chgfbl_data} into change from baseline
mean plot using ggplot2 and gridExtra
}
\examples{

}
\author{
Chendi Liao (liaoc10), \email{chendi.liao@roche.com}
}
