% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chgfbl_table.R
\name{chgfbl_table}
\alias{chgfbl_table}
\title{Wrap data generated from \code{chgfbl_data} into rtable format, to create a
test result and change from baseline by visit summary table}
\usage{
chgfbl_table(data)
}
\arguments{
\item{data}{data generated by \code{chgfbl_data} function}
}
\description{
Wrap data generated from \code{chgfbl_data} into rtable format, to create a
test result and change from baseline by visit summary table
}
\examples{

}
\author{
Chendi Liao (liaoc10), \email{chendi.liao@roche.com}
}
