% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_events_patyear.R
\name{ci_patyear}
\alias{ci_patyear}
\title{confidence interval for event rate per person-years}
\usage{
ci_patyear(
  event_number,
  person_year,
  conf_level = 0.95,
  conf_method = c("normal", "exact", "byar")
)
}
\arguments{
\item{event_number}{an integer}

\item{person_year}{a number}

\item{conf_level}{confidence level, default is 0.95}

\item{conf_method}{confidence method (\code{"normal", "exact", "byar"}),
see details of \code{\link{ci_patyear}}}
}
\description{
applicable methods for computing confidence intervals for event rate over person-years at risk
}
\details{
Confidence interval methods for  observed Y events over
   total of T person-years at risk  at a given \eqn{\alpha} :
  \enumerate{
  \item {Normal approximation confidence interval}
     {\deqn{LB = Y/T - Z(1- \alpha/2)*sqrt(Y/(T^2))}}
     {\deqn{UB = Y/T + Z(1- \alpha/2)*sqrt(Y/(T^2))}}
  \item {Exact confidence interval}
     {\deqn{LB = chisq(p = \alpha/2, df = 2Y)/(2T)}}
     {\deqn{UB= chisq(p = 1- \alpha/2, df = 2(Y + 1))/(2T)}}
  \item {Byar's confidence interval}
     {\deqn{LB = (Y + 0.5)(1- 1/(9(Y + 0.5)) - Z(1- \alpha/2)*sqrt(1/(Y + 0.5))/3)^3/T}}
     {\deqn{LB= (Y + 0.5)(1- 1/(9(Y + 0.5)) + Z(1- \alpha/2)*sqrt(1/(Y + 0.5))/3)^3/T}}
}
}
\examples{
ci_patyear(event_number = 10,
           person_year = 400,
           conf_level = 0.95,
           conf_method = "exact")


}
