% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{col_indices}
\alias{col_indices}
\title{Obtain Column Indices}
\usage{
col_indices(table_tree, col_names)
}
\arguments{
\item{table_tree}{(\code{VTableTree})\cr table to extract the indices from.}

\item{col_names}{(\code{character})\cr vector of column names.}
}
\value{
the vector of column indices.
}
\description{
Helper function to extract column indices from a \code{VTableTree} for a given
vector of column names.
}
