% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_arm.R
\name{combine_arm}
\alias{combine_arm}
\title{Combine Arms to Baseline and Comparison Arms}
\usage{
combine_arm(arm, arm.ref, arm.comp = setdiff(arm, arm.ref))
}
\arguments{
\item{arm}{vector with arm information}

\item{arm.ref}{values in arm that are combined to the reference group}

\item{arm.comp}{values in arm that are combined to the comparison group}
}
\description{
Combine groups in arms to reference and comparison groups in order to use it
in \code{\link[survival]{survfit}} and \code{\link{glm}}.
}
\examples{

combine_arm(arm = LETTERS[1:4], arm.ref = c("A", "B"))

combine_arm(arm = LETTERS[1:4], arm.ref = c("A", "B"), arm.comp = "C")


combine_arm(arm = LETTERS[1:4], arm.ref = c("A", "B", "R"), arm.comp = "C")

}
\author{
songy24
}
