% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_vectors}
\alias{combine_vectors}
\title{Combine Two Vectors Element Wise}
\usage{
combine_vectors(x, y)
}
\arguments{
\item{x}{(\code{vector})\cr first vector to combine.}

\item{y}{(\code{vector})\cr second vector to combine.}
}
\value{
A \code{list} where each element combines corresponding elements of \code{x} and \code{y}.
}
\description{
Combine Two Vectors Element Wise
}
\examples{
combine_vectors(1:3, 4:6)

}
