% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary_variants.R
\name{compare_in_header}
\alias{compare_in_header}
\title{Compare the list items / data.frame columns in the header}
\usage{
compare_in_header(x)
}
\arguments{
\item{x}{(\code{list}) list that contains items to compare in header}
}
\value{
object with attribute \code{compare_in_header}
}
\description{
Each list item will be a separate column in the header. The header will become hierarchical.
A data.frame is a special case of a list, where each column is a list item. All specified columns are compared.
All items must be of identical class
}
\examples{
compare_in_header(data.frame(`AVAL` = as.numeric(1:3),
  `CHG` = as.numeric(4:6), `CHG2` = as.numeric(7:9)))
}
