% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{control_binary_comparison}
\alias{control_binary_comparison}
\title{Control function for binary comparison analyses}
\usage{
control_binary_comparison(
  diff_ci = "waldcc",
  diff_test = "schouten",
  odds_ratio = TRUE
)
}
\arguments{
\item{diff_ci}{The method used for the confidence interval for difference of
proportions (see \code{\link{s_proportion_diff}} for details). Default method is here "waldcc",
i.e. Wald with continuity correction. If \code{NULL}, then no confidence intervals will be calculated.}

\item{diff_test}{Test for a difference between two proportions
(see \code{\link{s_test_proportion_diff}} for details). Default method is
here \code{schouten}, i.e. Chi-Squared Test with Schouten Correction.
If \code{NULL}, then no test will be performed and no p-values reported.}

\item{odds_ratio}{(\code{logical}) Whether the odds ratio estimates and confidence intervals should be
reported (default) or not.}
}
\value{
a list of components named as the arguments
}
\description{
This is an auxiliary function for controlling the arguments for \code{\link{t_binary_outcome}} specifying
how the response proportions should be compared between the arms.
}
