% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_survival.R
\name{control_coxph}
\alias{control_coxph}
\title{Control function for Cox-PH model}
\usage{
control_coxph(
  pval_method = c("log-rank", "wald", "likelihood"),
  ties = c("efron", "breslow", "exact"),
  conf_level = 0.95,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{pval_method}{(\code{string})\cr p-value method for testing hazard ratio = 1.
Default method is \code{"log-rank"}, can also be set to \code{"wald"} or \code{"likelihood"}.}

\item{ties}{(\code{string})\cr string specifying the method for tie handling. Default is \code{"efron"},
can also be set to \code{"breslow"} or \code{"exact"}. See more in \code{\link[survival:coxph]{survival::coxph()}}.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{alternative}{(\code{string})\cr alternative hypothesis for the p-value test. Default is \code{"two.sided"},
can also be set to \code{"less"} or \code{"greater"} for one-sided testing. Note that one-sided testing is not
supported when \code{pval_method = "likelihood"}.}
}
\value{
A list of components with the same names as the arguments.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is an auxiliary function for controlling arguments for Cox-PH model, typically used internally to specify
details of Cox-PH model for \code{\link[=s_coxph_pairwise]{s_coxph_pairwise()}}. \code{conf_level} refers to Hazard Ratio estimation.
}
