% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_lineplot.R
\name{control_lineplot_vars}
\alias{control_lineplot_vars}
\title{Control function for \code{g_lineplot()}}
\usage{
control_lineplot_vars(
  x = "AVISIT",
  y = "AVAL",
  group_var = "ARM",
  paramcd = "PARAMCD",
  y_unit = "AVALU",
  subject_var = "USUBJID",
  strata = lifecycle::deprecated(),
  cohort_id = lifecycle::deprecated()
)
}
\arguments{
\item{x}{(\code{string})\cr x-variable name.}

\item{y}{(\code{string})\cr y-variable name.}

\item{group_var}{(\code{string} or \code{NA})\cr group variable name.}

\item{paramcd}{(\code{string} or \code{NA})\cr parameter code variable name.}

\item{y_unit}{(\code{string} or \code{NA})\cr y-axis unit variable name.}

\item{subject_var}{(\code{string} or \code{NA})\cr subject variable name.}

\item{strata}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} use the \code{group_var} parameter instead.}

\item{cohort_id}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} use the \code{subject_var} parameter instead.}
}
\value{
A named character vector of variable names.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Default values for \code{variables} parameter in \code{g_lineplot} function.
A variable's default value can be overwritten for any variable.
}
\examples{
control_lineplot_vars()
control_lineplot_vars(group_var = NA)

}
