% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_survival.R
\name{control_surv_time}
\alias{control_surv_time}
\title{Control Function for \code{survfit} Model for Survival Time}
\usage{
control_surv_time(
  conf_level = 0.95,
  conf_type = c("plain", "log", "log-log"),
  quantiles = c(0.25, 0.75)
)
}
\arguments{
\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{conf_type}{(\code{string}) \cr "plain" (default), "log", "log-log" for confidence interval type, \cr
see more in \code{\link[survival:survfit]{survival::survfit()}}. Note that the option "none" is no longer supported.}

\item{quantiles}{(\code{numeric}) \cr of length two to specify the quantiles of survival time.}
}
\value{
A list of components with the same names as the arguments
}
\description{
This is an auxiliary function for controlling arguments for \code{survfit} model, typically used internally to specify
details of \code{survfit} model for \link{s_surv_time}. \code{conf_level} refers to survival time estimation.
}
