% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_missed_doses.R
\name{count_missed_doses}
\alias{count_missed_doses}
\alias{s_count_missed_doses}
\alias{a_count_missed_doses}
\title{Count number of patients with missed doses by thresholds}
\usage{
count_missed_doses(
  lyt,
  vars,
  thresholds,
  var_labels = vars,
  show_labels = "visible",
  na_str = default_na_str(),
  nested = TRUE,
  table_names = vars,
  ...,
  na_rm = TRUE,
  .stats = c("n", "count_fraction"),
  .stat_names = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)

s_count_missed_doses(
  x,
  thresholds,
  .N_col,
  .N_row,
  denom = c("N_col", "n", "N_row"),
  ...
)

a_count_missed_doses(
  x,
  ...,
  .stats = NULL,
  .stat_names = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{lyt}{(\code{PreDataTableLayouts})\cr layout that analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{thresholds}{(\code{numeric})\cr minimum number of missed doses the patients had.}

\item{var_labels}{(\code{character})\cr variable labels.}

\item{show_labels}{(\code{string})\cr label visibility: one of "default", "visible" and "hidden".}

\item{na_str}{(\code{string})\cr string used to replace all \code{NA} or empty values in the output.}

\item{nested}{(\code{flag})\cr whether this layout instruction should be applied within the existing layout structure _if
possible (\code{TRUE}, the default) or as a new top-level element (\code{FALSE}). Ignored if it would nest a split.
underneath analyses, which is not allowed.}

\item{table_names}{(\code{character})\cr this can be customized in the case that the same \code{vars} are analyzed multiple
times, to avoid warnings from \code{rtables}.}

\item{...}{additional arguments for the lower level functions.}

\item{na_rm}{(\code{flag})\cr whether \code{NA} values should be removed from \code{x} prior to analysis.}

\item{.stats}{(\code{character})\cr statistics to select for the table.

Options are: \verb{'n', 'count_fraction'}}

\item{.stat_names}{(\code{character})\cr names of the statistics that are passed directly to name single statistics
(\code{.stats}). This option is visible when producing \code{\link[rtables:as_result_df]{rtables::as_result_df()}} with \code{make_ard = TRUE}.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics. See Details in \code{analyze_vars} for more
information on the \code{"auto"} setting.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels. Defaults to 0, which corresponds to the
unmodified default behavior. Can be negative.}

\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{.N_col}{(\code{integer(1)})\cr column-wise N (column count) for the full column being analyzed that is typically
passed by \code{rtables}.}

\item{.N_row}{(\code{integer(1)})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is typically passed by \code{rtables}.}

\item{denom}{(\code{string})\cr choice of denominator for proportion. Options are:
\itemize{
\item \code{n}: number of values in this row and column intersection.
\item \code{N_row}: total number of values in this row across columns.
\item \code{N_col}: total number of values in this column across rows.
}}
}
\value{
\itemize{
\item \code{count_missed_doses()} returns a layout object suitable for passing to further layouting functions,
or to \code{\link[rtables:build_table]{rtables::build_table()}}. Adding this function to an \code{rtable} layout will add formatted rows containing
the statistics from \code{s_count_missed_doses()} to the table layout.
}

\itemize{
\item \code{s_count_missed_doses()} returns the statistics \code{n} and \code{count_fraction} with one element for each threshold.
}

\itemize{
\item \code{a_count_missed_doses()} returns the corresponding list with formatted \code{\link[rtables:CellValue]{rtables::CellValue()}}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The analyze function creates a layout element to calculate cumulative counts of patients with number of missed
doses at least equal to user-specified threshold values.

This function analyzes numeric variable \code{vars}, a variable with numbers of missed doses,
against the threshold values supplied to the \code{thresholds} argument as a numeric vector. This function
assumes that every row of the given data frame corresponds to a unique patient.
}
\section{Functions}{
\itemize{
\item \code{count_missed_doses()}: Layout-creating function which can take statistics function arguments
and additional format arguments. This function is a wrapper for \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{s_count_missed_doses()}: Statistics function to count patients with missed doses.

\item \code{a_count_missed_doses()}: Formatted analysis function which is used as \code{afun}
in \code{count_missed_doses()}.

}}
\examples{
library(dplyr)

anl <- tern_ex_adsl |>
  distinct(STUDYID, USUBJID, ARM) |>
  mutate(
    PARAMCD = "TNDOSMIS",
    PARAM = "Total number of missed doses during study",
    AVAL = sample(0:20, size = nrow(tern_ex_adsl), replace = TRUE),
    AVALC = ""
  )

basic_table() |>
  split_cols_by("ARM") |>
  add_colcounts() |>
  count_missed_doses("AVAL", thresholds = c(1, 5, 10, 15), var_labels = "Missed Doses") |>
  build_table(anl, alt_counts_df = tern_ex_adsl)

}
\seealso{
\itemize{
\item Relevant description function \code{\link[=d_count_missed_doses]{d_count_missed_doses()}} which generates labels for \code{\link[=count_missed_doses]{count_missed_doses()}}.
\item Similar analyze function \code{\link[=count_cumulative]{count_cumulative()}} which more generally counts cumulative values and has more
options for threshold handling, but uses different labels.
}
}
\keyword{internal}
