% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{cox_regression_inter}
\alias{cox_regression_inter}
\alias{h_coxreg_inter_effect}
\alias{h_coxreg_inter_effect.numeric}
\alias{h_coxreg_inter_effect.factor}
\alias{h_coxreg_extract_interaction}
\title{Cox Regression Helper: Interactions}
\usage{
h_coxreg_inter_effect(x, effect, covar, mod, label, control, ...)

\method{h_coxreg_inter_effect}{numeric}(x, effect, covar, mod, label, control, at, ...)

\method{h_coxreg_inter_effect}{factor}(x, effect, covar, mod, label, control, data, ...)

h_coxreg_extract_interaction(effect, covar, mod, data, at, control)
}
\arguments{
\item{x}{(\code{numeric} or \code{factor})\cr the values of the effect to be tested.}

\item{effect}{(\code{string})\cr the name of the effect to be tested and estimated.}

\item{covar}{(\code{string})\cr the name of the covariate in the model.}

\item{mod}{(\code{coxph})\cr the Cox regression model.}

\item{label}{(\code{string})\cr the label to be return as \code{term_label}
(see \code{return}).}

\item{control}{(\code{list})\cr a list of controls as returned by
\code{\link[=control_coxreg]{control_coxreg()}}.}

\item{...}{see methods.}

\item{at}{(\code{list})\cr a list with items named after the covariate, every
item is a vector of levels at which the interaction should be estimated.}

\item{data}{(\verb{data frame})\cr the data frame on which the model was fit.}
}
\description{
Test and estimate the effect of a treatment in interaction with a covariate.
The effect is estimated as the HR of the tested treatment for a given level
of the covariate, in comparison to the treatment control.
}
\section{Functions}{
\itemize{
\item \code{h_coxreg_inter_effect.numeric}: estimate the interaction with a numerical
covariate

\item \code{h_coxreg_inter_effect.factor}: estimates the interaction with a factor
covariate.

\item \code{h_coxreg_extract_interaction}: a higher level function that returns
the test of the interaction test and the estimated values. If
no interaction, \code{\link[=h_coxreg_univ_extract]{h_coxreg_univ_extract()}} is applied.
}}

\examples{

# Testing dataset [survival::bladder].
library(survival)
set.seed(1, kind = "Mersenne-Twister")
dta_bladder <- with(
  data = bladder[bladder$enum < 5, ],
  data.frame(
    time = stop,
    status = event,
    armcd = as.factor(rx),
    covar1 = as.factor(enum),
    covar2 = factor(
      sample(as.factor(enum)), levels = 1:4, labels = c("F", "F", "M", "M")
    )
  )
)
attr(dta_bladder$armcd, "label") <- "ARM"
attr(dta_bladder$covar1, "label") <- "A Covariate Label"
attr(dta_bladder$covar2, "label") <- "Sex (F/M)"
dta_bladder$AGE <- sample(20:60, size = nrow(dta_bladder), replace = TRUE)

plot(
  survfit(Surv(time, status) ~ armcd + covar1, data = dta_bladder),
  lty= 2:4, xlab="Months",
  col = c("blue1", "blue2", "blue3", "blue4", "red1", "red2", "red3", "red4")
)


mod <- coxph(Surv(time, status) ~ armcd * covar1, data = dta_bladder)
h_coxreg_extract_interaction(
  mod = mod, effect = "armcd", covar = "covar1", data = dta_bladder,
  control = control_coxreg()
)

}
