% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot_data.R
\name{coxphAnnoData}
\alias{coxphAnnoData}
\title{Prepare Cox PH model annotation data with rtable format}
\usage{
coxphAnnoData(formula_coxph, data, cox_ties = "exact",
  info_coxph = "Cox Porportional Model")
}
\arguments{
\item{formula_coxph}{formula specified for Cox PH model.}

\item{data}{analysis data set.}

\item{cox_ties}{ties handling method for Cox PH model. options are "efron", "breslow" and "exact".}

\item{info_coxph}{label information for Cox PH model.}
}
\description{
An rtable format of Cox PH model data for further annotation on top of Kaplan-Meier grob
}
\examples{
library(survival)
library(rtables)

OS <- data.frame(AVAL = abs(rnorm(200)), 
                 CNSR = sample(c(0, 1), 200, TRUE), 
                 ARM = sample(LETTERS[1:3], 200, TRUE),
                 SEX = sample(c("M","F"), 200, TRUE),
                 RACE = sample(c("AA", "BB", "CC"), 200, TRUE),
                 ECOG = sample(c(0, 1), 200, TRUE))
coxphAnnoData(Surv(AVAL, 1-CNSR) ~ ARM + strata(RACE), data = OS)
}
