% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot_data.R
\name{coxphAnnoImpower131}
\alias{coxphAnnoImpower131}
\title{Prepare Independent Cox PH model annotation data with rtable format}
\usage{
coxphAnnoImpower131(formula_coxph, data_list, cox_ties = "exact",
  info_coxph = "Cox Porportional Model")
}
\arguments{
\item{formula_coxph}{formula specified for Cox PH model.}

\item{data_list}{a list of analysis data set.}

\item{cox_ties}{ties handling method for Cox PH model. options are "efron", "breslow" and "exact".}

\item{info_coxph}{label information for Cox PH model.}
}
\description{
An rtable format of Indedendent Cox PH model (spcially for IMPower 131 study) data for further annotation on top of Kaplan-Meier grob
}
