% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_variables.R
\name{create_afun_summary}
\alias{create_afun_summary}
\title{Constructor Function for \code{\link[=analyze_vars]{analyze_vars()}} and \code{\link[=summarize_colvars]{summarize_colvars()}}}
\usage{
create_afun_summary(.stats, .formats, .labels, .indent_mods)
}
\arguments{
\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics. See Details in \code{analyze_vars} for more
information on the \code{"auto"} setting.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{vector} of \code{integer})\cr indent modifiers for the labels. Each element of the vector
should be a name-value pair with name corresponding to a statistic specified in \code{.stats} and value the indentation
for that statistic's row label.}
}
\value{
Combined formatted analysis function for use in \code{\link[=analyze_vars]{analyze_vars()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Constructor function which creates a combined formatted analysis function.
}
\note{
This function has been deprecated in favor of direct implementation of \code{a_summary()}.
}
\seealso{
\code{\link[=analyze_vars]{analyze_vars()}}
}
