% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abnormal_by_baseline.R
\name{d_count_abnormal_by_baseline}
\alias{d_count_abnormal_by_baseline}
\title{Description Function for \code{\link[=s_count_abnormal_by_baseline]{s_count_abnormal_by_baseline()}}}
\usage{
d_count_abnormal_by_baseline(abnormal)
}
\arguments{
\item{abnormal}{(\code{character})\cr identifying the abnormal range level(s) in \code{.var}.}
}
\value{
Abnormal category labels for \code{\link[=s_count_abnormal_by_baseline]{s_count_abnormal_by_baseline()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Description function that produces the labels for \code{\link[=s_count_abnormal_by_baseline]{s_count_abnormal_by_baseline()}}.
}
\examples{
d_count_abnormal_by_baseline("LOW")

}
