% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_cumulative.R
\name{d_count_cumulative}
\alias{d_count_cumulative}
\title{Description of Cumulative Count}
\usage{
d_count_cumulative(threshold, lower_tail, include_eq)
}
\arguments{
\item{threshold}{(\code{number})\cr a cutoff value as threshold to count values of \code{x}.}

\item{lower_tail}{(\code{logical})\cr whether to count lower tail, default is \code{TRUE}.}

\item{include_eq}{(\code{logical})\cr whether to include value equal to the \code{threshold} in
count, default is \code{TRUE}.}
}
\value{
a \code{string}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is a helper function that describes analysis in \code{s_count_cumulative}
}
