% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_multinomial_rsp.R
\name{d_onco_rsp_label}
\alias{d_onco_rsp_label}
\title{Description of Standard Oncology Response}
\usage{
d_onco_rsp_label(x)
}
\arguments{
\item{x}{(\code{character})\cr the standard oncology code to be described.}
}
\value{
Response labels.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Describe the oncology response in a standard way.
}
\examples{
d_onco_rsp_label(
  c("CR", "PR", "SD", "NON CR/PD", "PD", "NE", "Missing", "<Missing>", "NE/Missing")
)

# Adding some values not considered in d_onco_rsp_label

d_onco_rsp_label(
  c("CR", "PR", "hello", "hi")
)

}
\seealso{
\code{\link[=estimate_multinomial_rsp]{estimate_multinomial_rsp()}}
}
