% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_pkparam.R
\name{d_pkparam}
\alias{d_pkparam}
\title{Generate PK reference dataset}
\usage{
d_pkparam()
}
\value{
dataframe of PK parameters
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
pk_reference_dataset <- d_pkparam()
}
