% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_diff.R
\name{d_proportion_diff}
\alias{d_proportion_diff}
\title{Description of method used for proportion comparison}
\usage{
d_proportion_diff(conf_level, method, long = FALSE, method_only = FALSE)
}
\arguments{
\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{method}{(\code{string})\cr the method used for the confidence interval estimation.}

\item{long}{(\code{flag})\cr whether a long or a short (default) description is
required.}

\item{method_only}{(\code{flag})\cr whether to return only the method description,
without the confidence interval part of the description.
If \code{TRUE}, \code{conf_level} and \code{long} are ignored.}
}
\value{
A \code{string} describing the analysis.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is an auxiliary function that describes the analysis in
\code{\link[=s_proportion_diff]{s_proportion_diff()}}.
}
\examples{
d_proportion_diff(0.95, "cmh_sato")
d_proportion_diff(0.95, "cmh_sato", long = TRUE)
d_proportion_diff(0.95, "cmh_sato", method_only = TRUE)

}
\seealso{
\code{\link[=prop_diff]{prop_diff()}}, \code{\link[=d_proportion]{d_proportion()}}, \code{\link[=d_test_proportion_diff]{d_test_proportion_diff()}}
}
