% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_subgroups.R
\name{d_rsp_subgroups_colvars}
\alias{d_rsp_subgroups_colvars}
\title{Labels for column variables in binary response by subgroup table}
\usage{
d_rsp_subgroups_colvars(vars, conf_level = NULL, method = NULL)
}
\arguments{
\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{method}{(\code{string} or \code{NULL})\cr specifies the test used to calculate the p-value for the difference between
two proportions. For options, see \code{\link[=test_proportion_diff]{test_proportion_diff()}}. Default is \code{NULL} so no test is performed.}
}
\value{
A \code{list} of variables to tabulate and their labels.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Internal function to check variables included in \code{\link[=tabulate_rsp_subgroups]{tabulate_rsp_subgroups()}} and create column labels.
}
