% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot.R
\name{data_subgroup_tree}
\alias{data_subgroup_tree}
\title{Split the data into a tree of depth 2 with data frames}
\usage{
data_subgroup_tree(adsl, groupvar, anl)
}
\arguments{
\item{adsl}{ADLS data}

\item{groupvar}{variables in adsl that are used to partition the data}

\item{anl}{analysis data to partition

 @return a named list where each element is another named list with data
   frames}
}
\description{
This is used as the analysis is applied to subsets of the data. The first
layer of the tree defines the variable (or ALL) and the second layer
partitions the the data sets according to that variable.
}
