% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{data_tree}
\alias{data_tree}
\title{Create a tree object with \code{data.frame} objects as leaf}
\usage{
data_tree(x)
}
\arguments{
\item{x}{a list or nested list of rtables, if \code{x} is an rtable then the rtable is returned}
}
\description{
Converts a list of \code{data.frame}s or a nested list of rtables into a \code{data_tree} object. Note that
\code{NULL} objects are used interchangably with rtable objects.
}
\details{
\code{data_tree} does not change the underlying data structure of the argument \code{x} other than adding
  the \code{"data_tree"} class attribute. The \code{data_tree} function also checks if the leafs are either
  \code{NULL} objects or objects of class \code{data.frame}
}
\examples{
df <- data.frame(a = 1:2, b = 3:4)

nl1 <- list(df1 = df, df2 = df, df3 = NULL)
tt1 <- data_tree(nl1)
tt1
summary(tt1)

nl2 <- list(A = nl1, B = nl1)
tt2 <- data_tree(nl2)
tt2
summary(tt2)

nl3 <- list(X = nl2, Y = nl2, Z = nl2)
tt3 <- data_tree(nl3)
summary(tt3)

}
