% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{data_tree_add_overall}
\alias{data_tree_add_overall}
\title{Add overall on leaf depth}
\usage{
data_tree_add_overall(x, label = "- Overall -", set_attr = FALSE)
}
\arguments{
\item{x}{a data_tree: nested list of data.frames}

\item{label}{overall label}

\item{set_attr}{set \code{overall} attribute}
}
\description{
Add overall on leaf depth
}
\examples{

df_by <- expand.grid(u = c("ii", "iii"), x = LETTERS[1:4], y = letters[1:3], z = letters[21:22])
x <- data.frame(num = 1:nrow(df_by))
dfs <- rsplit(x, df_by[1:3])
data_tree_add_overall(dfs)

}
