% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rtables.R
\name{default_na_str}
\alias{default_na_str}
\alias{set_default_na_str}
\title{Default string replacement for \code{NA} values}
\usage{
default_na_str()

set_default_na_str(na_str)
}
\arguments{
\item{na_str}{(\code{string})\cr character that will be set in the R environment
options as default for replacing the \code{NA} values. It needs to be
single string Use \code{getOption("tern_default_na_str")} to get its current
value (\code{NULL} if not set).}
}
\value{
the current default value if option is set for \code{"tern_default_na_str"},
or \code{NA_character} otherwise
}
\description{
The default string for \code{NA} (\code{na_str}) which is displayed if it is not
modified by the user in detail, in each function, or generally with \code{set_default_na_str()}.
}
\examples{
default_na_str()
set_default_na_str("<missing>")
default_na_str()

}
