% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_defaults_handling.R
\name{default_stats_and_formats}
\alias{default_stats_and_formats}
\alias{get_stats}
\alias{get_format_from_stats}
\alias{get_label_from_stats}
\alias{tern_default_formats}
\alias{tern_default_labels}
\title{Defaults for stats methods names and their relative formats/labels}
\usage{
get_stats(method_group)

get_format_from_stats(stats)

get_label_from_stats(stats)

tern_default_formats()

tern_default_labels()
}
\arguments{
\item{method_group}{(\code{string}) \cr Indicates the group of statistical methods that
we need the defaults from.}

\item{stats}{(\code{character}) \cr statistical methods to get defaults formats for.}
}
\value{
\itemize{
\item \code{get_stats()} returns a character vector with all default statistical methods.
}

\itemize{
\item \code{get_format_from_stats()} returns a named list of formats, they being a value from
\code{\link[formatters:list_formats]{formatters::list_valid_format_labels()}} or a custom function (e.g. \link{formatting_functions}).
}

\itemize{
\item \code{get_label_from_stats()} returns a named list of default labels (if present
otherwise \code{NULL}).
}

\itemize{
\item \code{tern_default_formats()} returns a complete named list of default formats for \code{tern}.
}

\itemize{
\item \code{tern_default_labels()} returns a complete named list of default labels for \code{tern}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Utility functions to get valid statistic methods for different method groups
(\code{.stats}) and their relative formats (\code{.formats}) and labels (\code{.labels}). This utility
is used across \code{tern}, but some of its working principles can be seen in \code{\link[=analyze_vars]{analyze_vars()}}.
}
\section{Functions}{
\itemize{
\item \code{get_stats()}: Get defaults statistical methods for different
groups of methods.

\item \code{get_format_from_stats()}: Get formats from vector of statistical methods. If not
present \code{NULL} is returned.

\item \code{get_label_from_stats()}: Get labels from vector of statistical methods.

\item \code{tern_default_formats()}: Function that produce the complete named list
of default formats for \code{tern}.

\item \code{tern_default_labels()}: Function that produce the complete named list
of default labels for \code{tern}.

}}
\seealso{
\link{formatting_functions}
}
