% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\docType{methods}
\name{displayable}
\alias{displayable}
\alias{displayable,ANY-method}
\alias{displayable,rtable-method}
\alias{displayable,node-method}
\title{Returns output to be displayed with cat()}
\usage{
displayable(x, ...)

\S4method{displayable}{ANY}(x, indent = 0)

\S4method{displayable}{rtable}(x, indent = 0)

\S4method{displayable}{node}(x, indent = 0)
}
\arguments{
\item{x}{object to apply to}

\item{...}{other arguments}

\item{indent}{indent with which to print object}
}
\description{
For classes like integer, must use setOldClass("integer") so that S4 method dispatch works
and this function gets called.
}
\examples{
library(rtables)
tbl <- rtable(
  header = c("Treatement\\nN=100", "Comparison\\nN=300"),
  format = "xx (xx.xx\%)",
  rrow("A", c(104, .2), c(100, .4)),
  rrow("B", c(23, .4), c(43, .5)),
  rrow(),
  rrow("this is a very long section header"),
  rrow("estimate", rcell(55.23, "xx.xx", colspan = 2)),
  rrow("95\% CI", indent = 1, rcell(c(44.8, 67.4), format = "(xx.x, xx.x)", colspan = 2))
)
cat(displayable(tbl))
cat(displayable(tbl, indent = 1))
cat(displayable(tbl, indent = 2))
n41 <- node(name = "A", content = 4:60, children = list())
n42 <- node(name = "B", content = 7:9, children = list())
n43 <- node(name = "C", content = 10:12, children = list())
n4 <- node(name = "D", content = c(1:3), children = list(n41, n42, n43))
cat(displayable(n4))
}
