% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_equal_float}
\alias{.is_equal_float}
\title{Utility function to check if a float is equal to another float}
\usage{
.is_equal_float(x, y)
}
\arguments{
\item{x}{(\code{float})\cr A single number.}

\item{y}{(\code{float})\cr A single number.}
}
\value{
\code{TRUE}, if identical. \code{FALSE}, otherwise
}
\description{
Uses \code{.Machine$double.eps} as the tolerance for the comparison.
}
\keyword{internal}
