% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_shared_variables}
\alias{drop_shared_variables}
\title{Remove Shared Variables}
\usage{
drop_shared_variables(x, y, keep)
}
\arguments{
\item{x}{a data.frame}

\item{y}{a data.frame}

\item{keep}{optional, a vector with variable names that should not be removed}
}
\value{
a data.frame
}
\description{
Variables are considered shared if they have the same variable name
}
\examples{
drop_shared_variables(iris, iris[, 1:3])
}
