% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_sum}
\alias{dummy_sum}
\title{Dummy Analysis Function}
\usage{
dummy_sum(lyt, vars)
}
\arguments{
\item{lyt}{layout}

\item{vars}{vector of variable names}
}
\description{
Dummy Analysis Function
}
\examples{

basic_table() \%>\%
  split_cols_by("Species") \%>\%
  dummy_sum("Sepal.Length") \%>\%
  build_table(iris)

basic_table() \%>\%
  split_cols_by("Species") \%>\%
  dummy_sum(c("Sepal.Length", "Petal.Length")) \%>\%
  build_table(iris)

}
