% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_example.R
\name{dummy_summation}
\alias{dummy_summation}
\alias{h_dummy_sum}
\alias{s_dummy_sum}
\alias{a_dummy_sum}
\alias{dummy_sum}
\title{Documentation Example for Helper, Statistics, Formatted Analysis and Analyze Functions}
\usage{
h_dummy_sum(x)

s_dummy_sum(x, .var = "x")

a_dummy_sum(x, .var = "x")

dummy_sum(
  lyt,
  vars,
  ...,
  show_labels = "hidden",
  table_names = vars,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{...}{additional arguments for the lower level functions.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
The statistics are:
\itemize{
\item \code{sum}: the sum of all numbers, after removing \code{NA} values.
}
}
\description{
This illustrates the conventions on how to document helper, statistics, formatted analysis
and analyze functions together.
}
\section{Functions}{
\itemize{
\item \code{h_dummy_sum}: Helper Function which does the actual summation.

\item \code{s_dummy_sum}: Statistics Function which sums the numbers, after removing \code{NA} values.

\item \code{a_dummy_sum}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{dummy_sum}: Analyze Function which adds the sum analysis to
the input layout. Note that additional formatting arguments can be used
here.
}}

\examples{
h_dummy_sum(1:10)

s_dummy_sum(c(1, 2))

a_dummy_sum(c(1, 2))

basic_table() \%>\%
  split_cols_by("Species") \%>\%
  dummy_sum("Sepal.Length") \%>\%
  build_table(iris)

basic_table() \%>\%
  split_cols_by("Species") \%>\%
  dummy_sum(c("Sepal.Length", "Petal.Length")) \%>\%
  build_table(iris)

}
