% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_vector_if_na}
\alias{empty_vector_if_na}
\title{Return an empty numeric if all elements are \code{NA}.}
\usage{
empty_vector_if_na(x)
}
\arguments{
\item{x}{(\code{numeric})\cr vector.}
}
\value{
An empty \code{numeric}.
}
\description{
Return an empty numeric if all elements are \code{NA}.
}
\examples{
x <- c(NA, NA, NA)
tern:::empty_vector_if_na(x)

}
\keyword{internal}
