% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape regular expression metacharacters}
\usage{
escape_regex(x)
}
\arguments{
\item{x}{(\code{character})\cr strings to be used literally inside a regular expression.}
}
\value{
A \code{character} \code{vector} where all regular expression metacharacters in \code{x} are escaped.
}
\description{
Escape regular expression metacharacters
}
\keyword{internal}
