% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{esplit.data.frame}
\alias{esplit.data.frame}
\title{Splits the data.frame by rows}
\usage{
\method{esplit}{data.frame}(df, by)
}
\arguments{
\item{by}{split by, either a factor or a col_by matrix}
}
\description{
Splits the data.frame by rows
}
\examples{
by <- factor(c("M", "M", "F", "F", "F"))
esplit(data.frame(x = 1:5, y = 6:10), by)
}
