% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{esplit.default}
\alias{esplit.default}
\title{Default method for atomics}
\usage{
\method{esplit}{default}(x, by)
}
\arguments{
\item{x}{object to split}

\item{by}{split by, either a factor or a col_by matrix}
}
\description{
Also includes NULL
}
\examples{
by <- factor(c("M", "M", "F", "F", "F"))
esplit(1:5, by)

by <- data.frame("Old" = c(TRUE, TRUE, FALSE), "Tall" = c(FALSE, TRUE, TRUE))
esplit(1:3, by)
}
