% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{esplit.list}
\alias{esplit.list}
\title{Splits each list elements}
\usage{
\method{esplit}{list}(lst, by)
}
\arguments{
\item{by}{split by, either a factor or a col_by matrix}
}
\description{
See \code{\link{esplit.non_rsplit}} for alternative behavior.
}
\examples{
by <- factor(c("M", "M", "F", "F", "F"))
esplit(list(x = 1:5, y = 6:10), by)
esplit(list(), by)

esplit(
  list(data.frame(x = 1:5, y = 6:10), data.frame(z1 = 11:15, z2 = 16:20)),
  by
)
esplit(
  list(
    list(data.frame(x = 1:2, y = 3:4), data.frame(x = 5:6, y = 7:8)),
    list(data.frame(x = 11:12, y = 13:14), data.frame(x = 15:16, y = 17:18))
  ),
  by = factor(c("M", "F"))
)


by <- data.frame("Old" = c(TRUE, TRUE, FALSE), "Tall" = c(FALSE, TRUE, TRUE))
esplit(list(x = 1:3, y = 4:6), by)
esplit(
  list(
    x1 = factor(c(1,2,3,4,5,6)),
    x2 = data.frame(x = 7:12, y = 13:18)
  ),
  factor(c("a", "a", "b", "b", "b", "b"))
)
}
