% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{esplit.non_rsplit}
\alias{esplit.non_rsplit}
\title{Class: recursion}
\usage{
\method{esplit}{non_rsplit}(x, by)
}
\arguments{
\item{x}{object to split}

\item{by}{split by, either a factor or a col_by matrix}
}
\description{
Add this class to avoid recursion
}
\examples{
by <- factor(c("M", "M", "F", "F", "F"))
esplit(non_rsplit(as.list(1:5)), by) # not working: esplit(as.list(1:5), by)
esplit(
  non_rsplit(list(
    list(data.frame(x = 1:2, y = 3:4), data.frame(x = 5:6, y = 7:8)),
    list(data.frame(x = 11:12, y = 13:14), data.frame(x = 15:16, y = 17:18))
  )),
  by = factor(c("M", "F"))
)
}
