% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{esplit}
\alias{esplit}
\title{Split objects}
\usage{
esplit(x, by)
}
\arguments{
\item{x}{object to split}

\item{by}{split by, either a factor or a col_by matrix}
}
\value{
list, one item for each category in by
}
\description{
Split objects according to by object
}
