% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_pval}
\alias{f_pval}
\title{Utility function to create label for p-value}
\usage{
f_pval(diff_test)
}
\value{
a \code{string}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
