% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mmrm.R
\name{fit_lme4}
\alias{fit_lme4}
\title{Helper function to fit the MMRM with lme4 and lmerTest}
\usage{
fit_lme4(formula, data)
}
\arguments{
\item{formula}{the MMRM formula (it could also be another lme4 formula)}

\item{data}{the data frame}
}
\value{
the \code{lmerModLmerTest} object
}
\description{
Helper function to fit the MMRM with lme4 and lmerTest
}
