% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mmrm.R
\name{fit_lme4_single_optimizer}
\alias{fit_lme4_single_optimizer}
\title{Fitting}
\usage{
fit_lme4_single_optimizer(
  formula,
  data,
  optimizer = c("automatic", "nloptwrap_neldermead", "nloptwrap_bobyqa", "nlminbwrap",
    "bobyqa", "neldermead", "nmkbw", "optimx_lbfgsb")
)
}
\arguments{
\item{formula}{the lme4 formula}

\item{data}{the data set}

\item{optimizer}{the optimizer to use}
}
\value{
the fitted \code{lmerMerModTest} object, with additional attributes \code{messages} and \code{optimizer}.
}
\description{
Internal helper function to fit an lme4 model with a single optimizer, while capturing messages and warnings.
}
\note{
While we are not directly importing functions from \code{dfoptim} or \code{optimx} here, we rely on them
  being available to \code{lme4} (which has these packages only in "Suggests"). Therefore we note the imports below.
}
