% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{flatten_list}
\alias{flatten_list}
\title{Flatten a list by one level.}
\usage{
flatten_list(x)
}
\arguments{
\item{x}{possibly nested list}
}
\value{
list with one list level flattened out
}
\description{
Internal function used by the automatically created formatting functions.
Elements of the original input list which have been a list themselves are replaced
by the corresponding elements of the lower level list.
}
