% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_rsp.R
\name{forest_rsp}
\alias{forest_rsp}
\title{alternative tablist}
\usage{
forest_rsp(response, event, arm, group_data, covariARSs = NULL)
}
\arguments{
\item{response}{Tumor Response data}

\item{event}{is boolean, \code{TRUE} if responder, \code{FALSE} if non-responder}

\item{arm}{vector with arm information}

\item{group_data}{data frame with one column per sub-group variable}
}
\description{
alternative tablist
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr) 
library(grid)
library(teal.oncology)
'\%needs\%' <- teal.oncology:::'\%needs\%'
ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

resp_tbl_stream <- get_forest_response_table(com.roche.cdt30019.go29436.re)
Viewer(resp_tbl_stream )

ARS_f <- ARS \%>\% filter(PARAMCD == "OVRSPI") \%>\% 
                 filter(ITTWTFL == "Y") \%>\% 
                 filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\%
                 select(c("USUBJID", "STUDYID", "SEX", "ICLEVEL", "TC3IC3", "ARM", "AVAL", "AVALC"))
ASL_f <- ASL \%>\% filter(ITTWTFL == "Y") \%>\% filter(ARM \%in\% c("DUMMY A", "DUMMY C"))

group_data <- ARS_f[c("USUBJID", "STUDYID", "ICLEVEL", "TC3IC3")]
names(group_data) <- labels_over_names(group_data)

head(group_data)

# Dummy C First (comparison in survfit and glm)
arm <- fct_relevel(ARS_f$ARM, "DUMMY C")

tbl <- forest_rsp(
          response = ARS_f$AVAL,
          event = ARS_f$AVALC \%in\% c("CR","PR"),
          arm = arm, 
          group_data = group_data[, -c(1,2), drop=FALSE]
)
Viewer(tbl)

compare_rtables(tbl, surv_tbl_stream, comp.attr = FALSE)
forest_rsp_plot(tbl, levels(arm)[1], levels(arm)[2])

}


  
}
\author{
Yuyao Song (songy24), \email{yuyao.song@roche.com}
}
