% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_tte.R
\name{forest_tte}
\alias{forest_tte}
\title{alternative tablist}
\usage{
forest_tte(time_to_event, event, arm, group_data, covariates = NULL)
}
\arguments{
\item{time_to_event}{time to event data}

\item{event}{is boolean, \code{TRUE} if event, \code{FALSE} if time_to_event
is censored}

\item{arm}{vector with arm information}

\item{group_data}{data frame with one column per grouping}

\item{arm.ref}{a character vector defining which arms in arm should be taken 
as the reference}

\item{arm.comp}{a character vector defining which arms in arm should be taken 
as the comparison}
}
\description{
alternative tablist
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr)
library(forcats)
library(survival)

'\%needs\%' <- teal.oncology:::'\%needs\%'

surv_tbl_stream <- get_forest_survival_table(com.roche.cdt30019.go29436.re)
Viewer(surv_tbl_stream )

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

ATE_f <- ATE \%>\% filter(PARAMCD == "OS") \%>\% 
             filter(ITTWTFL == "Y") \%>\% 
             filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\%
             select(c("USUBJID", "STUDYID", "AVAL", "CNSR", "ARMCD"))
ASL_f <- ASL \%>\% filter(ITTWTFL == "Y") \%>\% 
             filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\% 
             select(c("USUBJID", "STUDYID", "SEX", "MLIVER", "TCICLVL2", "AGE4CAT", "BECOG"))


group_data <- left_join(ASL_f, ATE_f \%>\% select(c("USUBJID", "STUDYID")))
names(group_data) <- labels_over_names(group_data)
head(group_data)

# Dummy C First (comparison in survfit and glm)
arm <- fct_relevel(ATE_f$ARMCD, "C")

# Collapse Factors
# fct_collapse(arm, "ARM A/B" = c("ARM A", "ARM B"))
# rename a level
# arm <- fct_recode(ATE_f$ARM, "Treat ARM" = "DUMMY A")

tbl <- forest_tte(
   time_to_event = ATE_f$AVAL,
   event = ATE_f$CNSR == 0,
   group_data = group_data[, -c(1,2), drop=FALSE],
   arm = arm
)
Viewer(tbl)

compare_rtables(tbl, surv_tbl_stream, comp.attr = FALSE)

forest_tte_plot(tbl, levels(arm)[1], levels(arm)[2])



}

# forest_tte(Surv(AVAL ~ I(CNSR != 'N') ~ ARM + SEX, data = ATE))
}
