% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_tte.R
\name{forest_tte_ADAM}
\alias{forest_tte_ADAM}
\title{Forest Plot Numbers for Survival data with ADAM data structure}
\usage{
forest_tte_ADAM(ASL, ATE, outcome = "Overall Survival", groupvar, arm.ref,
  arm.comp, arm.var = "ARM", time_to_event.var = "AVAL",
  event.var = "CNSR", negate.event.var = TRUE)
}
\arguments{
\item{ASL}{asl data frame}

\item{ATE}{data frame}

\item{arm.ref}{a character vector defining which arms in arm should be taken 
as the reference}

\item{arm.comp}{a character vector defining which arms in arm should be taken 
as the comparison}
}
\description{
Forest Plot Numbers for Survival data with ADAM data structure
}
\examples{
\dontrun{

rm(list = ls())
library(atezo.data)
library(dplyr)

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

ASL$BAGED <- ifelse(ASL$BAGE <= median(ASL$BAGE), "<=median", ">median")

forest_tte_ADAM(
  ASL, ATE,
  groupvar = c("SEX", "BECOG", "COUNTRY"),
  arm.ref = "DUMMY A", arm.comp = "DUMMY B"
)
  
}
}
