% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_tte.R
\name{forest_tte_plot}
\alias{forest_tte_plot}
\title{Time-to-event Forest Plot}
\usage{
forest_tte_plot(x, arm.ref = "ReferenceAAA verylongtitle",
  arm.comp = "TreatmentAAA verylongtitle", padx = unit(0, "lines"),
  cex = 1)
}
\arguments{
\item{x}{time-to-event forest plot table from function \code{forest_tte}}

\item{arm.ref}{Label displayed in the forest plot for reference arm}

\item{padx}{gap between two columns}

\item{cex}{multiplier applied to overall fontsize}

\item{arm.ref}{Label displayed in the forest plot for comparison arm}
}
\description{
Time-to-event Forest Plot
}
\details{
This function can only be applied to the rtable output from \code{forest_tte}.
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr)
library(forcats)
library(survival)
library(teal.oncology)

'\%needs\%' <- teal.oncology:::'\%needs\%'

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

tbl_stream <- get_forest_survival_table(com.roche.cdt30019.go29436.re)
Viewer(tbl_stream)


ATE_f <- ATE \%>\% filter(PARAMCD == "OS") \%>\% 
             filter(ITTWTFL == "Y") \%>\% 
             filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\%
             select(c("USUBJID", "STUDYID", "AVAL", "CNSR", "ARM"))
ASL_f <- ASL \%>\% filter(ITTWTFL == "Y") \%>\% 
             filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\% 
             select(c("USUBJID", "STUDYID", "SEX", "MLIVER", "TCICLVL2", "AGE4CAT", "RACE"))


group_data <- left_join(ASL_f, ATE_f \%>\% select(c("USUBJID", "STUDYID")))
group_data$MLIVER <- factor(group_data$MLIVER, levels = c("Y", "N"), labels = c("Yes", "No"))
group_data$TCICLVL2 <- factor(group_data$TCICLVL2, levels = c("TC3 or IC2/3", "TC0/1/2 and IC0/1"), labels = c("TC3 or IC2/3", "TC0/1/2 and IC0/1"))#' 
group_data$SEX <- factor(group_data$SEX, levels = c("F", "M"), labels = c("FEMALE", "MALE"))
group_data$AGE4CAT <- factor(group_data$AGE4CAT, levels = c("<65", "65 to 74", "75 to 84", ">=85"), labels = c("<65", "65 to 74", "75 to 84", ">=85"))
group_data$RACE <- factor(group_data$RACE, levels = c("ASIAN", "BLACK OR AFRICAN AMERICAN", "WHITE", "UNKNOWN"), labels = c("ASIAN", "BLACK OR AFRICAN AMERICAN", "WHITE", NA))
names(group_data) <- labels_over_names(group_data)
head(group_data)

arm <- fct_relevel(ATE_f$ARM, "DUMMY C")

tbl <- forest_tte(
   time_to_event = ATE_f$AVAL,
   event = ATE_f$CNSR == 0,
   group_data = group_data[, -c(1,2), drop=FALSE],
   arm = arm
)
Viewer(tbl)

Viewer(tbl, tbl_stream)

compare_rtables(tbl, tbl_stream, comp.attr = FALSE)

library(grid)
forest_tte_plot(tbl, levels(arm)[1], levels(arm)[2], padx=unit(0, "lines"))


}

}
\author{
Yuyao Song (songy24), \email{yuyao.song@roche.com}
}
