% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_forest.R
\name{forestGrob}
\alias{forestGrob}
\title{forest plot grob}
\usage{
forestGrob(tbl, x, lower, upper, xlim = NULL, width_row.names = NULL,
  width_columns = NULL, width_forest = unit(1, "null"), name = NULL,
  gp = NULL, vp = NULL)
}
\arguments{
\item{tbl}{an \code{\link[rtables]{rtable}} object}

\item{x}{coordinate of point}

\item{lower}{lower bound of ci}

\item{upper}{upper bound of ci}

\item{xlim}{x limits for x-scales}

\item{width_columns}{\code{\link[grid]{unit}} object with widths for the
table columns. If \code{NULL} the widths get automatically calculated.}

\item{width_forest}{\code{\link[grid]{unit}} object with width for the
forest column. If \code{NULL} the widths get automatically calculated.}

\item{width_row.name}{\code{\link[grid]{unit}} object with width for row
names. If \code{NULL} the widths get automatically calculated.}
}
\description{
forest plot grob
}
\details{
The heights get automatically determined
}
\examples{

tbl <- rtable(
  header = rheader(
   rrow("", "E", rcell("CI", colspan = 2)),
   rrow("", "A", "B", "C")
  ),
  rrow("row 1", 1, 0.8, 1.1),
  rrow("row 2", 1.4, 0.8, 1.6),
  rrow("row 3", 1.2, 0.8, 1.2)
)

}
