% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_count_fraction}
\alias{format_count_fraction}
\title{Formatting Count and Fraction}
\usage{
format_count_fraction(x, ...)
}
\arguments{
\item{x}{(\code{integer})\cr vector of length 2, count and fraction.}

\item{...}{required for \code{rtables} interface.}
}
\value{
a string in the format \verb{count (fraction \%)}. If \code{count} is 0 the format is \code{0}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Formats a count together with fraction with special consideration when count is \code{0}.
}
\examples{
format_count_fraction(x = c(2, 0.6667))
format_count_fraction(x = c(0, 0))

}
\seealso{
Other formatting functions: 
\code{\link{extreme_format}},
\code{\link{format_fraction_threshold}()},
\code{\link{format_fraction}()},
\code{\link{formatting_functions}}
}
\concept{formatting functions}
