% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_fraction}
\alias{format_fraction}
\title{Format: fraction and percentage}
\usage{
format_fraction(x, ...)
}
\arguments{
\item{x}{a vector with integer elements \code{num} and \code{denom}}

\item{...}{required for rtables interface}
}
\value{
a string in the format \verb{num / denom (ratio \%)}
}
\description{
Format a fraction together with ratio in percent.
}
\examples{
format_fraction(x = c(num = 2L, denom = 3L))
}
