% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{format_wrap}
\alias{format_wrap}
\alias{format_wrap_df}
\alias{format_wrap_x}
\title{Construct Formatted Analysis Functions}
\usage{
format_wrap_df(sfun, indent_mods, formats)

format_wrap_x(sfun, indent_mods, formats)
}
\arguments{
\item{sfun}{the original statistics function returning a named list of results}

\item{indent_mods}{named vector with default (nonnegative) indent modifications}

\item{formats}{named vector with default formats}
}
\value{
the constructed analysis function
}
\description{
The returned function has first argument \code{df} or \code{x} and uses any additional arguments for the
original \code{sfun}. It has additional four formatting arguments:
\describe{
\item{.stats}{character vector to select statistics from the \code{sfun} result}
\item{.indent_mods}{named vector with custom (nonnegative) indent modifications}
\item{.formats}{named vector with custom formats}
\item{.labels}{named vector with custom labels}
}
Note that \code{.indent_mods}, \code{.formats} and \code{.labels} don't need to contain an element for each
statistic, as they have defaults initialized either from the arguments below (for indents and formats)
or from the \code{sfun} results (for labels).
}
\details{
For comparison functions it is common to return "empty" results for the
comparison column. This can be specified by returning empty strings \code{""} in the
corresponding statistics from the statistics function, along with the usual labels.
The Formatted Analysis function which is returned from these wrapper constructors
will then replace the old formats with the default format \code{"xx"} that works with these
empty strings. The table cells will then stay empty.
}
\section{Functions}{
\itemize{
\item \code{format_wrap_df}: the wrapper for \code{sfun} having \code{df} as first argument.

\item \code{format_wrap_x}: the wrapper for \code{sfun} having \code{x} as first argument.
}}

