% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_functions.R
\name{format_xx_fixed_dp}
\alias{format_xx_fixed_dp}
\title{Alternative format XX function :a formatting function with fixed decimal precision}
\usage{
format_xx_fixed_dp(str, na_str)
}
\arguments{
\item{str}{(\code{string})\cr template.}
}
\value{
An \code{rtables} formatting function.
}
\description{
Translate a string where x and dots are interpreted as number place
holders, and others as formatting elements.
}
\examples{
test <- list(c(1.658, 0.5761), c(1e1, 785.6))

z <- format_xx_fixed_dp("xx (xx.x)")
sapply(test, z)

z <- format_xx_fixed_dp("xx.x - xx.x")
sapply(test, z)

z <- format_xx_fixed_dp("xx.x, incl. xx.x\% NE")
sapply(test, z)
}
\seealso{
\link{format_xx}
}
