% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{full_apply_at_depth}
\alias{full_apply_at_depth}
\title{Applies f to node at given depth in depth-first order}
\usage{
full_apply_at_depth(x, f, depth = 0)
}
\arguments{
\item{x}{node to apply to}

\item{f}{function(node) -> node to apply to all nodes at given depth}

\item{depth}{depth at which to apply function}
}
\value{
node
}
\description{
We prefix it full because you can modify the entire node, i.e. also the children and format_data
full_apply_from_depth is also possible to be implemented.
}
\details{
Note that we don't make an S4 method here for brevity.
}
\examples{
tree <- node("1", 1, list(
node("10", 10),
node("7", 7),
node("6", 6, list(
  node("2", 2),
  node("1", 1)
))
))
cat(displayable(tree))
updated_tree <- full_apply_at_depth(
  tree,
  function(node) {
    # Note: copy-on-write for node
    node@content <- node@content + 100
    node
  },
  depth = 2
)
cat(displayable(updated_tree))
}
