% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_mmrm.R
\name{g_mmrm_lsmeans}
\alias{g_mmrm_lsmeans}
\title{Plot the LS means of an MMRM model.}
\usage{
g_mmrm_lsmeans(
  object,
  select = c("estimates", "contrasts"),
  titles = c(estimates = paste("Adjusted mean of", object$labels$response,
    "by treatment at visits"), contrasts = paste0("Differences of ",
    object$labels$response, " adjusted means vs. control ('", object$ref_level, "')")),
  ylab = paste0("Estimates with ", round(object$conf_level * 100), "\% CIs"),
  width = 0.6,
  show_pval = TRUE
)
}
\arguments{
\item{object}{the MMRM model result produced by \code{\link{s_mmrm}}.}

\item{select}{character vector to select one or both of "estimates" and "contrasts" plots.}

\item{titles}{character vector with elements \code{estimates} and \code{contrasts} containing the plot titles.}

\item{ylab}{string with the y axis label.}

\item{width}{width of the error bars.}

\item{show_pval}{should the p-values for the differences of LS means vs. control be displayed (not default)?}
}
\value{
a \code{ggplot2} plot
}
\description{
This function summarizes adjusted \code{lsmeans} and standard error, as well as conducts
comparisons between groups' adjusted \code{lsmeans}, where the first level of the group
is the reference level.
}
\details{
If variable labels are available in the original data set, then these are used. Otherwise
  the variable names themselves are used for annotating the plot.
}
\examples{
library(dplyr)
library(random.cdisc.data)

adsl <- radsl(cached = TRUE)
adqs <- radqs(cached = TRUE)
adqs_f <- adqs \%>\%
  dplyr::filter(PARAMCD=="FKSI-FWB" & !AVISIT \%in\% c("BASELINE")) \%>\%
  droplevels() \%>\%
  dplyr::mutate(ARM = factor(ARM, levels = c("B: Placebo", "A: Drug X", "C: Combination"))) \%>\%
  dplyr::mutate(AVISITN = rank(AVISITN) \%>\% as.factor() \%>\% as.numeric() \%>\% as.factor())
var_labels(adqs_f) <- var_labels(adqs)

mmrm_results <- s_mmrm(
  vars = list(
    response = "AVAL",
    covariates = c("STRATA1", "BMRKR2"),
    id = "USUBJID",
    arm = "ARM",
    visit = "AVISIT"
  ),
  data = adqs_f,
  cor_struct = "random-quadratic",
  weights_emmeans = "equal"
)
g_mmrm_lsmeans(mmrm_results)
g_mmrm_lsmeans(mmrm_results, select = "estimates")
g_mmrm_lsmeans(
  mmrm_results,
  select = "contrasts",
  titles = c(contrasts = "Contrasts of FKSI-FWB means"),
  show_pval = TRUE,
  width = 0.8
)
}
\seealso{
\code{\link{t_mmrm_lsmeans}}
}
