% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_summary_by.R
\name{g_summary_by}
\alias{g_summary_by}
\title{Chanage From Baseline Plot}
\usage{
g_summary_by(x)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{group}{factor with group information}

\item{y_lower}{lower bound of error bar, if \code{NULL} no error bar is displayed}

\item{y_upper}{upper bound of error bar, if \code{NULL} no error bar is displayed}

\item{n_visit}{for each element of \code{x}}

\item{ylim}{y limits}

\item{y_line_at}{draw horizontal lines at}

\item{xlab}{xlabel}

\item{ylab}{ylabel}

\item{nlab}{n_visit label}

\item{draw}{boolean should plot be drawn}

\item{newpage}{boolean should plot be drawn on newpage}
}
\description{
Chanage From Baseline Plot
}
\examples{

\dontrun{
g_summary_by(
 x = ANL$VITIT, y = ANL$CHG, group = ANL$ARM,
 y_lower , y_upper , 
 n_visit = ANL$nvisit,
 ylim, yline_at, title 
)
 
}
}
