% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_pkparam.R
\name{d_pkparam}
\alias{d_pkparam}
\title{Generate PK reference dataset}
\usage{
d_pkparam()
}
\value{
data.frame of pk parameter dataset
}
\description{
Generate PK reference dataset
}
\section{Functions}{
\itemize{
\item \code{d_pkparam}: a dataframe of the correct order of the PK parameters
}}

\examples{
pk_reference_dataset <- d_pkparam()
}
