% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_covariates}
\alias{get_covariates}
\title{Utility function to return a named list of covariate names.}
\usage{
get_covariates(covariates)
}
\arguments{
\item{covariates}{(\code{character})\cr a vector that can contain single variable names (such as
\code{"X1"}), and/or interaction terms indicated by \code{"X1 * X2"}.}
}
\value{
a named \code{list} of character vector.
}
\description{
Utility function to return a named list of covariate names.
}
\keyword{internal}
