% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{get_indent_header}
\alias{get_indent_header}
\title{Derive a header with indented tree in top left corner}
\usage{
get_indent_header(old_header, header_labels)
}
\arguments{
\item{old_header}{The header of an rtable}

\item{header_labels}{(\link{node}) object's labels derived by \code{recursive_get_labels}}
}
\value{
A new header with the tree added in top left corner
}
\description{
Derive a header with indented tree in top left corner
}
