% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mmrm.R
\name{get_lme4_diagnostics}
\alias{get_lme4_diagnostics}
\title{Linear Mixed Model Diagnostics}
\usage{
get_lme4_diagnostics(fit, cov_est = get_lme4_cov_estimate(fit))
}
\arguments{
\item{fit}{(\code{lmerModLmerTest})\cr object fit with REML using \code{\link[lmerTest:lmer]{lmerTest::lmer()}}.}

\item{cov_est}{(\code{Matrix})\cr covariance estimate coming from \code{\link[=get_lme4_cov_estimate]{get_lme4_cov_estimate()}}.}
}
\value{
a list with the REML criterion, the AIC, AICc and BIC.
}
\description{
Compute the model diagnostic statistics for a linear mixed model fit.
}
\details{
AICc is here defined as
\deqn{-2 * loglik + 2 * df * (m / (m - df - 1))}
where \code{loglik} is the restricted log likelihood value, \code{df} is the number of covariance parameters
and \code{m} is the number of observations minus the number of fixed effects, or \code{df + 2} if it is smaller
than that. The same \code{df} is used for AIC and BIC. Note that for BIC, the \code{n} used is the number of
subjects (instead of the number of observations as in AIC and AICc). This matches the definitions in SAS.
}
