% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mmrm.R
\name{get_mmrm_lsmeans}
\alias{get_mmrm_lsmeans}
\title{Extract LS means}
\usage{
get_mmrm_lsmeans(fit, vars, conf_level, weights)
}
\arguments{
\item{fit}{result of \code{\link{fit_lme4}}}

\item{vars}{variable list as per \code{\link{s_mmrm}} input}

\item{conf_level}{confidence level}

\item{weights}{string specifying the type of weights to be used for the LS means,
see \code{\link[emmeans]{emmeans}} for details.}
}
\value{
A list with the LS means `estimates` and `contrasts` results between the treatment
  and control arm groups at the different visits.
}
\description{
Helper function to extract the LS means from an MMRM fit.
}
