% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_N}
\alias{get_N}
\title{Get total count}
\usage{
get_N(col_by)
}
\arguments{
\item{col_by}{(factor or matrix) to get counts from}
}
\value{
counts per factor level or column of \code{col_by}
}
\description{
Normally, you don't need to use this function as it is the default
}
\examples{
get_N(data.frame(A = c(TRUE, TRUE, FALSE), B = c(FALSE, FALSE, TRUE)))
}
