% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{get_onco_rsp_label}
\alias{get_onco_rsp_label}
\title{Get label for standard oncology response categories}
\usage{
get_onco_rsp_label()
}
\value{
Named \code{character} vector.
}
\description{
This is a convenience function for labeling the partition section in \code{t_binary_outcome}.
}
\examples{
library(dplyr)
rsp <- c("CR", "NE", "PR")
# Recode values with standard labels.
rsp <- get_onco_rsp_label()[ match(rsp, names(get_onco_rsp_label()))  ]
# Create factor.
rsp <- factor(rsp, levels = get_onco_rsp_label())
}
