% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pool_object.R
\name{get_pool_object}
\alias{get_pool_object}
\title{Get cached \code{pool} class object from \code{rbmi} package}
\usage{
get_pool_object()
}
\description{
Get cached \code{pool} class object from \code{rbmi} package
}
\examples{
get_pool_object()
}
