% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_smooths}
\alias{get_smooths}
\title{Smooth Function with Optional Grouping}
\usage{
get_smooths(df, x, y, groups = NULL, level = 0.95)
}
\arguments{
\item{df}{(\code{data.frame})\cr.}

\item{x}{(\code{character})\cr value with x column name.}

\item{y}{(\code{character})\cr value with y column name.}

\item{groups}{(\code{character})\cr vector with optional grouping variables names.}

\item{level}{(\code{numeric}) level of confidence interval to use (0.95 by default).}
}
\value{
A \code{data.frame} with original \code{x}, smoothed \code{y}, \code{ylow}, \code{yhigh} and
optional \code{groups} variables formatted to factor type.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This produces \code{loess} smoothed estimates of \code{y} with Student confidence intervals.
}
