% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{glm_simple_term_extract}
\alias{glm_simple_term_extract}
\title{Tabulation of logistic regression}
\usage{
glm_simple_term_extract(x, fit_glm)
}
\arguments{
\item{x}{(\code{string})\cr a variable in \code{fit_glm}.}

\item{fit_glm}{logistic regression model fitted by \code{\link[stats:glm]{stats::glm()}}.}
}
\description{
Tabulate the main effect results of a logistic regression model.
}
\examples{
library(random.cdisc.data)
library(dplyr)
adrs <- radrs(cached = TRUE)
adrs_f <- adrs \%>\%
  dplyr::filter(PARAMCD == "BESRSPI") \%>\%
  dplyr::filter(RACE \%in\% c("ASIAN", "WHITE", "BLACK OR AFRICAN AMERICAN")) \%>\%
  dplyr::mutate(
    Response = case_when(AVALC \%in\% c("PR", "CR") ~ 1, TRUE ~ 0),
    RACE = factor(RACE),
    SEX = factor(SEX)
  )
fit_glm <- fit_logistic(
  data = adrs_f,
  variables = list(response = "Response", arm = "ARMCD", covariates = c("AGE", "RACE"))
)
glm_simple_term_extract("AGE", fit_glm)
glm_simple_term_extract("ARMCD", fit_glm)

}
