% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot_rsp.R
\name{glm_subgroup}
\alias{glm_subgroup}
\title{alternative tablist}
\usage{
glm_subgroup(response, event, arm, arm.ref, arm.comp = setdiff(arm, arm.ref),
  group_by, covariARSs = NULL)
}
\arguments{
\item{event}{is boolean, \code{TRUE} if event, \code{FALSE} if time_to_event
is censored}

\item{arm}{vector with arm information}

\item{arm.ref}{a character vector defining which arms in arm should be taken 
as the reference}

\item{arm.comp}{a character vector defining which arms in arm should be taken 
as the comparison}

\item{group_by}{data frame with one column per grouping}

\item{time_to_event}{Response data}
}
\description{
alternative tablist
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr)
'\%needs\%' <- teal.oncology:::'\%needs\%'
ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

ARS_f <- ARS \%>\% filter(PARAMCD == "BESRSPI")

ASL$BAGED <- ifelse(ASL$BAGE <= median(ASL$BAGE), "<=median", ">median")

group_by <- merge(
 ARS_f[c("USUBJID", "STUDYID")],
 ASL[c("USUBJID", "STUDYID", "BAGED", "SEX", "BECOG")],
 all.x = TRUE, all.y = FALSE
)

head(group_by)

glm_subgroup(
   response = ARS_f$AVAL,
   event = ARS_f$AVALC \%in\% c("CR","PR"),
   arm = ARS_f$ARM, 
   group_by = group_by[, -c(1,2), drop=FALSE],
   arm.ref = "DUMMY A",
   arm.comp = "DUMMY B"
)
}

  
}
