% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot_rsp.R
\name{glm_subgroup_ADAM}
\alias{glm_subgroup_ADAM}
\title{Forest Plot Numbers for Survival data with ADAM data structure}
\usage{
glm_subgroup_ADAM(ASL, ARS, outcome = "Overall Survival", groupvar, arm.ref,
  arm.comp, arm.var = "ARM", time_to_event.var = "AVAL",
  event.var = "CNSR", negARS.event.var = TRUE)
}
\arguments{
\item{ASL}{asl data frame}

\item{ARS}{data frame}

\item{arm.ref}{a character vector defining which arms in arm should be taken 
as the reference}

\item{arm.comp}{a character vector defining which arms in arm should be taken 
as the comparison}
}
\description{
Forest Plot Numbers for Survival data with ADAM data structure
}
\examples{
\dontrun{

rm(list = ls())
library(ARSzo.data)
library(dplyr)

ARS <- ARS(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

ASL$BAGED <- ifelse(ASL$BAGE <= median(ASL$BAGE), "<=median", ">median")

glm_subgroup_ADAM(
  ASL, ARS,
  groupvar = c("SEX", "BECOG", "COUNTRY"),
  arm.ref = "DUMMY A", arm.comp = "DUMMY B"
)
  
}
}
