% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%needs\%}
\alias{\%needs\%}
\title{Check if list or data.frame has elements/variables}
\usage{
data \%needs\% names
}
\arguments{
\item{data}{a \code{data.frame} or named list}

\item{names}{vector with names}
}
\value{
\code{TRUE} if all variables exist and an appropriate error if not.
}
\description{
Checks if list names exist and throws an error otherwise
}
\examples{
\dontrun{
# function is not exported
`\%needs\%` <- teal.oncology:::`\%needs\%`

iris \%needs\% c("Sepal.Length", "Petal.Width")

iris \%needs\% "ABC"

}

}
\author{
Adrian Waddell (waddella), \email{adrian.waddell@roche.com}
}
