% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_km.R
\name{grobDraw}
\alias{grobDraw}
\title{Draw plot of a grob or list of grobs}
\usage{
grobDraw(grobs, ncol = 1, newpage = TRUE)
}
\arguments{
\item{grobs}{a grob or a list of grobs.}

\item{ncol}{number of plots per column.}

\item{newpage}{boolean if \code{draw=TRUE} should plot be drawn on a new page}
}
\description{
utility function to draw a grob or a list of grobs
}
\examples{
library(dplyr)
library(gridExtra)
library(grid)
ptext <- textGrob("test text here")
prect <- rectGrob(width = unit(0.5, "npc"), height = unit(0.5, "npc"))
grobDraw(ptext)
grobDraw(prect)
grobDraw(list(ptext, prect))
}
