% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{h_coxreg_inter_estimations}
\alias{h_coxreg_inter_estimations}
\title{Hazard ratio}
\usage{
h_coxreg_inter_estimations(
  variable,
  given,
  lvl_var,
  lvl_given,
  mod,
  conf_level = 0.95
)
}
\arguments{
\item{variable, given}{(\code{string})\cr
the name of variables in interaction. We seek the estimation of the levels
of \code{variable} given the levels of \code{given}.}

\item{lvl_var, lvl_given}{(\code{character})\cr
corresponding levels has given by \code{\link[=levels]{levels()}}.}

\item{mod}{(\code{coxph})\cr a fitted Cox regression model (see \code{\link[survival:coxph]{survival::coxph()}}).}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}
}
\value{
A list of matrix (one per level of variable) with rows corresponding to the combinations of
\code{variable} and \code{given}, with columns:
\describe{
\item{coef_hat}{Estimation of the coefficient}
\item{coef_se}{Standard error of the estimation.}
\item{hr}{Hazard ratio.}
\item{lcl,ucl}{lower/upper confidence limit of the hazard ratio}
}
}
\description{
Hazard ratio estimation in interactions
}
\details{
Given the cox regression investigating the effect of Arm (A, B, C; reference A)
and Sex (F, M; reference Female) and the model being abbreviated: y ~ Arm + Sex + Arm:Sex.
The cox regression estimates the coefficients along with a variance-covariance matrix for:
\itemize{
\item b1 (arm b), b2 (arm c),
\item b3 (sex m),
\item b4 (arm b: sex m), b5 (arm c: sex m)
}

The estimation of the Hazard Ratio for arm C/sex M is given in reference
to arm A/Sex M by exp(b2 + b3 + b5)/ exp(b3) = exp(b2 + b5).
The interaction coefficient is deduced by b2 + b5 while the standard error
is obtained as $sqrt(Var b2 + Var b5 + 2 * covariance (b2,b5))$.
}
\examples{

# Testing dataset [survival::bladder].
library(survival)
set.seed(1, kind = "Mersenne-Twister")
dta_bladder <- with(
  data = bladder[bladder$enum < 5, ],
  data.frame(
    time = stop,
    status = event,
    armcd = as.factor(rx),
    covar1 = as.factor(enum),
    covar2 = factor(
      sample(as.factor(enum)), levels = 1:4, labels = c("F", "F", "M", "M")
    )
  )
)
attr(dta_bladder$armcd, "label") <- "ARM"
attr(dta_bladder$covar1, "label") <- "A Covariate Label"
attr(dta_bladder$covar2, "label") <- "Sex (F/M)"
dta_bladder$AGE <- sample(20:60, size = nrow(dta_bladder), replace = TRUE)

mod <- coxph(Surv(time, status) ~ armcd * covar1, data = dta_bladder)

result <- h_coxreg_inter_estimations(
  variable = "armcd", given = "covar1",
  lvl_var = levels(dta_bladder$armcd),
  lvl_given = levels(dta_bladder$covar1),
  mod = mod, conf_level = .95
)
result

}
