% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{h_coxreg_univ_extract}
\alias{h_coxreg_univ_extract}
\title{Tabulation of Univariate Cox Regressions}
\usage{
h_coxreg_univ_extract(effect, covar, mod, data, control = control_coxreg())
}
\arguments{
\item{effect}{(\code{string})\cr the treatment variable.}

\item{covar}{(\code{string})\cr the name of the covariate in the model.}

\item{mod}{(\code{coxph})\cr Cox regression model fitted by \code{\link[survival:coxph]{survival::coxph()}}.}

\item{data}{(\verb{data frame})\cr the data frame on which the model was fit.}

\item{control}{(\code{list})\cr a list of controls as returned by
\code{\link[=control_coxreg]{control_coxreg()}}.}
}
\description{
Tabulate the result of a Cox regression model.
}
\examples{

library(survival)
dta_simple <-  data.frame(
  time = c(5, 5, 10, 10, 5, 5, 10, 10),
  status = c(0, 0, 1, 0, 0, 1, 1, 1),
  armcd  = factor(LETTERS[c(1, 1, 1, 1, 2, 2, 2, 2)], levels = c("A", "B"))
)
mod <- coxph(Surv(time, status) ~ armcd, data = dta_simple)
result <- h_coxreg_univ_extract(
  effect = "armcd", covar = "armcd", mod = mod, data = dta_simple
)

}
