% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_forest.R
\name{h_default_forest_header}
\alias{h_default_forest_header}
\title{Production of a Default \code{forest_header}}
\usage{
h_default_forest_header(tbl)
}
\arguments{
\item{tbl}{(\code{rtable})\cr input table.}
}
\value{
Either a character vector with two elements, or \code{NULL}.
}
\description{
This helper functions produces a default \code{forest_header} for \code{\link[=g_forest]{g_forest()}}.
This will work well with the usual full table returned by \code{\link[=tabulate_rsp_subgroups]{tabulate_rsp_subgroups()}}
and \code{\link[=tabulate_survival_subgroups]{tabulate_survival_subgroups()}}. If no arm names can be inferred, it will just
return \code{NULL} such that no header will be included by default. If arms names can be inferred,
but not exactly two, an error will be thrown.
}
