% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_glm_count.R
\name{h_glm_count}
\alias{h_glm_count}
\alias{h_glm_poisson}
\alias{h_glm_quasipoisson}
\alias{h_ppmeans}
\title{Helper Functions for Poisson Models.}
\usage{
h_glm_poisson(.var, .df_row, variables, weights)

h_glm_quasipoisson(.var, .df_row, variables, weights)

h_glm_count(.var, .df_row, variables, distribution, weights)

h_ppmeans(obj, .df_row, arm, conf_level)
}
\arguments{
\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.df_row}{(\verb{data frame})\cr data set that includes all the variables that are called
in \code{.var} and \code{variables}.}

\item{variables}{(named \code{list} of \code{strings})\cr list of additional analysis variables, with
expected elements:
\itemize{
\item \code{arm}: (\code{string})\cr group variable, for which the covariate adjusted means of multiple
groups will be summarized. Specifically, the first level of \code{arm} variable is taken as the
reference group.
\item \code{covariates}: (\code{character})\cr a vector that can contain single variable names (such as
\code{"X1"}), and/or interaction terms indicated by \code{"X1 * X2"}.
\item \code{offset}: (\code{numeric})\cr a numeric vector or scalar adding an offset.
}}

\item{obj}{(\code{glm.fit}) fitted model object used to derive the mean rate estimates in each treatment arm.}

\item{conf_level}{(\code{numeric}) value used to derive the confidence interval for the rate.}

\item{`weights`(`character`)\cr}{character vector specifying weights used in averaging predictions.}

\item{`distribution`(`character`)\cr}{a character value specifying the distribution
used in the regression (poisson, quasipoisson).}

\item{`arm`:}{(\code{string})\cr group variable, for which the covariate adjusted means of multiple
groups will be summarized. Specifically, the first level of \code{arm} variable is taken as the
reference group.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Helper functions that can be used to return the results of various Poisson models.
}
\section{Functions}{
\itemize{
\item \code{h_glm_poisson()}: Helper function to return results of a poisson model.

\item \code{h_glm_quasipoisson()}: Helper function to return results of a quasipoisson model.

\item \code{h_glm_count()}: Helper function to return the results of the
selected model (poisson, quasipoisson, negative binomial).

\item \code{h_ppmeans()}: Helper function to return the estimated means.

}}
\examples{

# Internal function - h_glm_poisson
\dontrun{
h_glm_poisson(
  .var = "AVAL",
  .df_row = anl,
  variables = list(arm = "ARM", offset = "lgTMATRSK", covariates = NULL)
)
}


# Internal function - h_glm_quasipoisson
\dontrun{
h_glm_quasipoisson(
  .var = "AVAL",
  .df_row = anl,
  variables = list(arm = "ARM", offset = "lgTMATRSK", covariates = c("REGION1"))
)
}


# Internal function - h_glm_count
\dontrun{
h_glm_count(
  .var = "AVAL",
  .df_row = anl,
  variables = list(arm = "ARMCD", offset = "lgTMATRSK", covariates = NULL),
  distribution = "poisson"
)
}


# Internal function - h_ppmeans
\dontrun{
fits <- h_glm_count(
  .var = "AVAL",
  .df_row = anl,
  variables = list(arm = "ARMCD", offset = "lgTMATRSK", covariates = c("REGION1")),
  distribution = "quasipoisson"
)

h_ppmeans(
  obj = fits$glm_fit,
  .df_row = anl,
  arm = "ARM",
  conf_level = 0.95
)
}

}
\seealso{
\link{summarize_glm_count}
}
\keyword{internal}
