% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{h_logistic_simple_terms}
\alias{h_logistic_simple_terms}
\title{Helper of logistic regression}
\usage{
h_logistic_simple_terms(x, fit_glm, conf_level = 0.95)
}
\arguments{
\item{x}{(\code{string} or vector of \code{string}) \cr variable in \code{fit_glm}.}

\item{fit_glm}{logistic regression model fitted by \code{\link[stats:glm]{stats::glm()}} with "binomial" family.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}
}
\description{
Tabulate the results including odds ratios and confidence intervals of simple terms.
}
\examples{
library(random.cdisc.data)
library(dplyr)
adrs <- radrs(cached = TRUE)
adrs_f <- adrs \%>\%
  dplyr::filter(PARAMCD == "BESRSPI") \%>\%
  dplyr::filter(RACE \%in\% c("ASIAN", "WHITE", "BLACK OR AFRICAN AMERICAN")) \%>\%
  dplyr::mutate(Response = case_when(AVALC \%in\% c("PR", "CR") ~ 1, TRUE ~ 0))
fit_glm <- fit_logistic(
  data = adrs_f,
  variables = list(
    response = "Response", arm = "ARMCD",
    covariates = c("AGE", "RACE"), interaction = "RACE"
  )
)
h_logistic_simple_terms("AGE", fit_glm)

}
