% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_map_for_count_abnormal.R
\name{h_map_for_count_abnormal}
\alias{h_map_for_count_abnormal}
\title{Helper Function to create a map dataframe that can be used in \code{trim_levels_to_map} split function.}
\usage{
h_map_for_count_abnormal(
  df,
  variables = list(anl = "ANRIND", split_rows = c("PARAM"), range_low = "ANRLO",
    range_high = "ANRHI"),
  abnormal = list(low = c("LOW", "LOW LOW"), high = c("HIGH", "HIGH HIGH")),
  method = c("default", "range"),
  na_level = "<Missing>"
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{variables}{(named \code{list} of \code{string})\cr list of additional analysis variables.}

\item{abnormal}{(\verb{named list})\cr identifying the abnormal range level(s) in \code{df}. Default to
\code{list(Low = "LOW LOW", High = "HIGH HIGH")} but you can also group different levels into the name list, for example,
\verb{abnormal = list(Low = "LOW", High = "HIGH"))}}

\item{method}{(\code{string})\cr indicates how the returned map will be constructed. Can be either \code{"default"} or
\code{"range"}.
If method is \code{"default"}, the returned map will only have the abnormal directions that are observed in the \code{df}, and
records with all normal values will be excluded to avoid error in creating layout.
If method is \code{"range"}, the returned map will be based on the rule that at least one observation with low range > 0
for low direction and at least one observation with high range is not missing for high direction.}

\item{na_level}{(\code{string})\cr used to replace all \code{NA} or empty values in factors.}
}
\description{
Helper Function to create a map dataframe from the input dataset, which can be used as an argument in the
\code{trim_levels_to_map} split function. Based on different method, the map is constructed differently.
}
\examples{

library(scda)
adlb <- synthetic_cdisc_data("latest")$adlb
adlb <- df_explicit_na(adlb)

h_map_for_count_abnormal(
  df = adlb,
  variables = list(anl = "ANRIND", split_rows = c("LBCAT", "PARAM"), range_low = "ANRLO", range_high = "ANRHI"),
  abnormal = list(low = c("LOW"), high = c("HIGH")),
  method = "default",
  na_level = "<Missing>"
)

}
