% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_prop.R
\name{h_prop}
\alias{h_prop}
\alias{prop_wilson}
\alias{prop_clopper_pearson}
\alias{prop_wald}
\alias{prop_agresti_coull}
\alias{prop_jeffreys}
\title{Helper Functions for Calculating Proportion Confidence Intervals}
\usage{
prop_wilson(rsp, conf_level, correct = FALSE)

prop_clopper_pearson(rsp, conf_level)

prop_wald(rsp, conf_level, correct = FALSE)

prop_agresti_coull(rsp, conf_level)

prop_jeffreys(rsp, conf_level)
}
\arguments{
\item{rsp}{(\code{logical})\cr whether each subject is a responder or not.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{correct}{(\code{flag})\cr apply continuity correction.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Functions to calculate different proportion confidence intervals for use in \code{\link[=estimate_proportion]{estimate_proportion()}}.
}
\section{Functions}{
\itemize{
\item \code{prop_wilson}: the Wilson interval calls \code{\link[stats:prop.test]{stats::prop.test()}}.
Also referred to as Wilson score interval.

\item \code{prop_clopper_pearson}: the Clopper-Pearson interval calls
\code{\link[stats:binom.test]{stats::binom.test()}}. Also referred to as the \code{exact} method.

\item \code{prop_wald}: the Wald interval follows the usual
textbook definition for a single proportion confidence interval using the
normal approximation.

\item \code{prop_agresti_coull}: the Agresti-Coull interval was created by
Alan Agresti and Brent Coull and can be understood (for 95\% CI) as adding
two successes and two failures to the data, and then using the Wald
formula to construct a CI.

\item \code{prop_jeffreys}: the Jeffreys interval is an equal-tailed
interval based on the non-informative Jeffreys prior for a binomial
proportion.
}}

\examples{
rsp <- c(
  TRUE, TRUE, TRUE, TRUE, TRUE,
  FALSE, FALSE, FALSE, FALSE, FALSE
)
prop_wilson(rsp, conf_level = 0.9)

prop_clopper_pearson(rsp, conf_level = .95)

prop_wald(rsp, conf_level = 0.95)
prop_wald(rsp, conf_level = 0.95, correct = TRUE)

prop_agresti_coull(rsp, conf_level = 0.95)

prop_jeffreys(rsp, conf_level = 0.95)

}
