% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_multinomial_rsp.R
\name{h_prop_ci}
\alias{h_prop_ci}
\title{Helper Function: Proportion Estimations}
\usage{
h_prop_ci(
  rsp,
  denom = length(rsp),
  method = c("waldcc", "wald", "clopper-pearson", "wilson", "agresti-coull",
    "jeffreys"),
  conf_level,
  long = TRUE,
  rsp_lab = "Responders"
)
}
\arguments{
\item{rsp}{(\code{logical})\cr whether each subject is a responder or not.}

\item{denom}{(\code{number})\cr denominator for the responder proportions.}

\item{method}{(\code{string}) \cr
the method used to construct the confidence interval for proportion of
successful outcomes; one of \code{waldcc}, \code{wald}, \code{clopper-pearson}, \code{wilson},
\code{agresti-coull} or \code{jeffreys}.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{long}{(\code{flag})\cr a long description is required.}

\item{rsp_lab}{(\code{string})\cr the label attribute for the responder proportion.}
}
\description{
Helps in estimating the proportion for a given vector of responses.
}
