% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_patient_plot.R
\name{h_set_nest_theme}
\alias{h_set_nest_theme}
\title{Helper Function to Set a \code{ggplot} Graphical Theme}
\usage{
h_set_nest_theme(font_size)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Function that sets a \code{ggplot} graphical theme to control the outlook of a plot.
}
