% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_survival_duration_subgroups.R
\name{h_split_by_subgroups}
\alias{h_split_by_subgroups}
\title{Split Dataframe by Subgroups}
\usage{
h_split_by_subgroups(data, subgroups)
}
\arguments{
\item{data}{(\verb{data frame})\cr dataset to split.}

\item{subgroups}{(\code{character})\cr names of factor variables from \code{data} used to create subsets.
Unused levels not present in \code{data} are dropped.}
}
\value{
A list with subset data (\code{df}) and metadata about the subset (\code{df_labels}).
}
\description{
Split a dataframe into a non-nested list of subsets.
}
\details{
Main functionality is to prepare data for use in forest plot layouts.
}
\examples{
library(rtables)

df <- data.frame(
  x = c(1:5),
  y = factor(c("A","B", "A", "B", "A"), levels = c("A", "B", "C")),
  z = factor(c("C", "C", "D", "D", "D"), levels = c("D", "C"))
)
var_labels(df) <- paste("label for", names(df))

h_split_by_subgroups(
  data  = df,
  subgroups = c("y", "z")
)
}
