% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{h_split_param}
\alias{h_split_param}
\title{Split parameters}
\usage{
h_split_param(param, value, f)
}
\arguments{
\item{param}{(\code{vector})\cr the parameter to be split.}

\item{value}{(\code{vector})\cr the value used to split.}

\item{f}{(\code{list} of \code{vectors})\cr the reference to make the split}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

It divides the data in the vector \code{param} into the groups defined by \code{f}
based on specified \code{values}.
It is relevant in \code{rtables} layers so as to distribute parameters
\code{.stats} or' \code{.formats} into lists with items corresponding to
specific analysis function.
}
\examples{
f <- list(
  surv = c("pt_at_risk", "event_free_rate", "rate_se", "rate_ci"),
  surv_diff = c("rate_diff", "rate_diff_ci", "ztest_pval")
)

.stats <- c("pt_at_risk", "rate_diff")
h_split_param(.stats, .stats, f = f)

# $surv
# [1] "pt_at_risk"
#
# $surv_diff
# [1] "rate_diff"

.formats <- c("pt_at_risk" = "xx", "event_free_rate" = "xxx")
h_split_param(.formats, names(.formats), f = f)

# $surv
# pt_at_risk event_free_rate
# "xx"           "xxx"
#
# $surv_diff
# NULL

}
