% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_km.R
\name{h_tbl_median_surv}
\alias{h_tbl_median_surv}
\title{Helper function for survival estimations}
\usage{
h_tbl_median_surv(fit_km, armval = "All", digits = 4)
}
\arguments{
\item{fit_km}{(\code{survfit})\cr result of \code{\link[survival:survfit]{survival::survfit()}}.}

\item{armval}{(\code{string})\cr used as strata name when treatment arm variable only has one level. Default is \code{"All"}.}

\item{digits}{(\code{integer(1)})\cr number of significant digits for median and CI values.
Defaults to \code{4}.}
}
\value{
A summary table with statistics \code{N}, \code{Median}, and \verb{XX\% CI} (\code{XX} taken from \code{fit_km}).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Transform a survival fit to a table with groups in rows characterized by N, median and confidence interval.
}
\examples{
library(dplyr)
library(survival)

adtte <- tern_ex_adtte |> filter(PARAMCD == "OS")
fit <- survfit(
  formula = Surv(AVAL, 1 - CNSR) ~ ARMCD,
  data = adtte
)
h_tbl_median_surv(fit_km = fit)
h_tbl_median_surv(fit_km = fit, digits = 2)

}
