% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_by_subgroup.R
\name{helpers_prop_by_subgroup}
\alias{helpers_prop_by_subgroup}
\alias{h_proportion_df}
\alias{h_proportion_subgroups_df}
\title{Helper Functions for Tabulating Binary Response by Subgroup}
\usage{
h_proportion_df(rsp, arm)

h_proportion_subgroups_df(variables, data)
}
\arguments{
\item{rsp}{(\code{logical})\cr whether each subject is a responder or not.}

\item{arm}{(\code{factor})\cr the treatment group variable.}

\item{variables}{(named \code{list} of \code{strings})\cr list of additional analysis variables.}

\item{data}{(\verb{data frame})\cr the dataset containing the variables to summarize.}
}
\description{
Helper functions that extract statistics such as response rate and odds ratio
for population subgroups.
}
\section{Functions}{
\itemize{
\item \code{h_proportion_df}: Helper to prepare a data frame of binary responses by arm.

\item \code{h_proportion_subgroups_df}: Summarizes proportion of binary responses by arm and across subgroups
in a data frame. \code{variables} corresponds to the names of variables found in \code{data}, passed as a named list and
requires elements \code{rsp}, \code{arm} and \code{subgroups}.
}}

\examples{

# Testing dataset.
library(random.cdisc.data)
library(dplyr)

adrs <- radrs(cached = TRUE)
adrs_labels <- var_labels(adrs)

adrs <- adrs \%>\%
  filter(PARAMCD == "BESRSPI") \%>\%
  filter(ARM \%in\% c("A: Drug X", "B: Placebo")) \%>\%
  droplevels() \%>\%
  mutate(
    # Reorder levels of factor to make the placebo group the reference arm.
    ARM = forcats::fct_relevel(ARM, "B: Placebo"),
    rsp = AVALC == "CR"
  )
var_labels(adrs) <- c(adrs_labels, "Response")


h_proportion_df(
  c(TRUE, FALSE, FALSE),
  arm = factor(c("A", "A", "B"), levels = c("A", "B"))
)


h_proportion_subgroups_df(
  variables = list(rsp = "rsp", arm = "ARM", subgroups = c("SEX", "BMRKR2")),
  data = adrs
)

}
