% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{identical_without_attr}
\alias{identical_without_attr}
\title{Compare Without Considering Attributes}
\usage{
identical_without_attr(x, y)
}
\arguments{
\item{x}{first object}

\item{y}{second object}
}
\value{
A single logical value, \code{TRUE} or \code{FALSE}, never \code{NA} and never
anything other than a single value.
}
\description{
Helper function used in format wrapper functions.
}
\details{
This function only works on the top level of the objects,
i.e. if a list is compared with another list, then we still consider
attributes of the individual list elements when doing the comparison.
}
\note{
Note that there is a difference of this function compared to the
behavior of \link{all.equal}. For example, the latter will not return \code{TRUE}
if \code{x} and \code{y} just differ in their class attribute, even when using it
with argument \code{check.attributes = FALSE}.
}
