% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{if_not_null}
\alias{if_not_null}
\title{Kind of a complement to `%||%`
Return y if x is not NULL, otherwise NULL (equal to x)}
\usage{
if_not_null(x, y)
}
\arguments{
\item{x}{value checked for NULL}

\item{y}{value to return if \code{x} is not NULL}
}
\value{
\code{NULL} if \code{x} is \code{NULL}, else \code{y}
}
\description{
Kind of a complement to `%||%`
Return y if x is not NULL, otherwise NULL (equal to x)
}
\examples{
tern:::if_not_null(NULL, 1)
tern:::if_not_null(2, 1)
}
