% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_rule.R
\name{imputation_rule}
\alias{imputation_rule}
\title{Apply 1/3 or 1/2 Imputation Rule to Data}
\usage{
imputation_rule(df, x_stats, stat, imp, post = FALSE)
}
\description{
Apply 1/3 or 1/2 Imputation Rule to Data
}
