% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_tern}
\alias{install_tern}
\title{Install a particular version of tern with the recommended versions of terns
dependencies}
\usage{
install_tern(ref = NULL, dependencies = c("all", "depends", "none"),
  eval = TRUE)
}
\arguments{
\item{ref}{either \code{NULL}, \code{'master'}, \code{'devel'}, or a version
tag (i.e. \code{'v0.0.1'}). If \code{NULL} then the current installed
version is used.}

\item{dependencies}{one of \code{'all'}, \code{'depends'} or \code{'none'}}

\item{eval}{execute the installation. This is only useful to install
\code{ref='master'}, \code{'devel'}, or a version the preceeds the current
installed version}
}
\description{
For production use one should install \code{tern} using this function so that
matching versions of tern and \code{random.cdisc.data} are installed. For a
first time installation use the instructions on
\url{GitHub repository README.md}{http://github.roche.com/Rpackages/tern}.
}
\details{
Currently the R packages \code{tern}, \code{rtables},
\code{random.cdisc.data}, \code{teal.oncology}, and \code{teal} are evolving
rapidly and we currently do not guarantee backwards compatibility. Therefore
\code{tern} and \code{teal.oncology} will provide the
\code{install_<pkg>} functions which provide install instructions for
recommended package versions.
}
\examples{
\dontrun{
install_tern(eval = FALSE)
install_tern('master', 'depends', eval = FALSE)
install_tern('devel', 'none', eval = FALSE)
}
}
