% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{invisible_node}
\alias{invisible_node}
\title{Creates an invisible node with \code{NULL} content and invisible name}
\usage{
invisible_node(children, name = "root", content = NULL,
  format_data = NULL)
}
\arguments{
\item{children}{children of node}

\item{name}{name of invisible node (invisible name in \code{rtable} prints)}

\item{content}{content}

\item{format_data}{format_data}
}
\value{
node
}
\description{
The name of the created node is \code{invisible_node_name(name)}.
}
\examples{
n11 <- node(name = "A", content = array(c(1:6), dim = c(2,3)), children = list())
n12 <- node(name = "B", content = array(c(1:6), dim = c(2,3)), children = list())
n13 <- node(name = "C", content = array(c(1:6), dim = c(2,3)), children = list())
summary(invisible_node(list(n11, n12, n13)))
}
