% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{invisible_node_name}
\alias{invisible_node_name}
\title{Create an invisible node}
\usage{
invisible_node_name(name)
}
\arguments{
\item{name}{name of invisible node}
}
\value{
node
}
\description{
Node can be accessed through this name, but name will not be displayed when converted with to_rtable.
node_name that will not be displayed, also true when converted to rtable
The purpose of this is to still allow indexing in the rtable through the name index
}
\examples{
invisible_node_name("root")
}
