% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_fully_named_list}
\alias{is_fully_named_list}
\title{Checks if the object is NULL or a fully named list}
\usage{
is_fully_named_list(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Checks if the object is NULL or a fully named list
}
\examples{
tern:::is_fully_named_list(list())
tern:::is_fully_named_list(list(a = 1, 2))
tern:::is_fully_named_list(list(1, 2))
tern:::is_fully_named_list(list(a = 1, b = 2))
}
