% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{is_nested_by}
\alias{is_nested_by}
\title{Check whether is of class nested_by}
\usage{
is_nested_by(x)
}
\arguments{
\item{x}{a list that specifies recursive split,
e.g. a list of col_bys (which can be matrices or factors)}
}
\value{
boolean whether it is an nested_by object
}
\description{
Check whether is of class nested_by
}
