% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_rtables.R
\name{is_r_by}
\alias{is_r_by}
\title{Check whether is of class r_by}
\usage{
is_r_by(x)
}
\arguments{
\item{x}{a list that specifies recursive split,
e.g. a list of col_bys (which can be matrices or factors)}
}
\value{
boolean whether it is an r_by object
}
\description{
Check whether is of class r_by
}
