% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_rtables.R
\name{is_simple_by}
\alias{is_simple_by}
\title{Check whether is of class simple_by}
\usage{
is_simple_by(x)
}
\arguments{
\item{x}{object to wrap it around}
}
\value{
boolean whether it is an simple_by object
}
\description{
Check whether is of class simple_by
}
