% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_plot.R
\name{kaplan_meier}
\alias{kaplan_meier}
\title{Kaplan-Meier Plot}
\arguments{
\item{df}{(\code{data.frame})\cr data set containing all analysis variables.}

\item{variables}{(named \code{list})\cr variable names. Details are:
\itemize{
\item \code{tte} (\code{numeric})\cr variable indicating time-to-event duration values.
\item \code{is_event} (\code{logical})\cr event variable. \code{TRUE} if event, \code{FALSE} if time to event is censored.
\item \code{arm} (\code{factor})\cr the treatment group variable.
\item \code{strat} (\code{character} or \code{NULL})\cr variable names indicating stratification factors.
}}

\item{control_surv}{(\code{list})\cr parameters for comparison details, specified by using
the helper function \code{\link{control_surv_timepoint}}. Some possible parameter options are:
\itemize{
\item \code{conf_level} (\code{proportion})\cr confidence level of the interval for survival rate.
\item \code{conf_type} (\code{string})\cr "plain" (default), "log", "log-log" for confidence interval type,
see more in \code{\link[survival:survfit]{survival::survfit()}}. Note that the option "none" is no longer supported.
}}

\item{data}{(\code{data.frame})\cr survival data as pre-processed by \code{h_data_plot}.}

\item{xticks}{(\code{numeric}, \code{number}, or \code{NULL})\cr numeric vector of ticks or single number with spacing
between ticks on the x axis. If \code{NULL} (default), \code{\link[labeling:extended]{labeling::extended()}} is used to determine
an optimal tick position on the x axis.}

\item{yval}{(\code{string})\cr value of y-axis. Options are \code{Survival} (default) and \code{Failure} probability.}

\item{censor_show}{(\code{flag})\cr whether to show censored.}

\item{xlab}{(\code{string})\cr label of x-axis.}

\item{ylab}{(\code{string})\cr label of y-axis.}

\item{title}{(\code{string})\cr title for plot.}

\item{footnotes}{(\code{string})\cr footnotes for plot.}

\item{col}{(\code{character})\cr lines colors. Length of a vector should be equal
to number of strata from \code{\link[survival:survfit]{survival::survfit()}}.}

\item{lty}{(\code{numeric})\cr line type. Length of a vector should be equal
to number of strata from \code{\link[survival:survfit]{survival::survfit()}}.}

\item{lwd}{(\code{numeric})\cr line width. Length of a vector should be equal
to number of strata from \code{\link[survival:survfit]{survival::survfit()}}.}

\item{pch}{(\code{numeric}, \code{string})\cr value or character of points symbol to indicate censored cases.}

\item{size}{(\code{numeric})\cr size of censored point, a class of \code{unit}.}

\item{max_time}{(\code{numeric})\cr maximum value to show on X axis. Only data values less than or up to
this threshold value will be plotted (defaults to \code{NULL}).}

\item{font_size}{(\code{number})\cr font size to be used.}

\item{ci_ribbon}{(\code{flag})\cr draw the confidence interval around the Kaplan-Meier curve.}

\item{ggtheme}{(\code{theme})\cr a graphical theme as provided by \code{ggplot2} to control outlook of the Kaplan-Meier curve.}

\item{annot_at_risk}{(\code{flag})\cr compute and add the annotation table reporting the number of patient at risk
matching the main grid of the Kaplan-Meier curve.}

\item{annot_surv_med}{(\code{flag})\cr compute and add the annotation table on the Kaplan-Meier curve estimating the
median survival time per group.}

\item{annot_coxph}{(\code{flag})\cr add the annotation table from a \code{\link[survival:coxph]{survival::coxph()}} model.}

\item{annot_stats}{(\code{string})\cr statistics annotations to add to the plot. Options are
\code{median} (median survival follow-up time) and \code{min} (minimum survival follow-up time).}

\item{annot_stats_vlines}{(\code{flag})\cr add vertical lines corresponding to each of the statistics
specified by \code{annot_stats}. If \code{annot_stats} is \code{NULL} no lines will be added.}

\item{control_coxph_pw}{(\code{list})\cr parameters for comparison details, specified by using
the helper function \code{\link[=control_coxph]{control_coxph()}}. Some possible parameter options are:
\itemize{
\item \code{pval_method} (\code{string})\cr p-value method for testing hazard ratio = 1.
Default method is "log-rank", can also be set to "wald" or "likelihood".
\item \code{ties} (\code{string})\cr method for tie handling. Default is "efron",
can also be set to "breslow" or "exact". See more in \code{\link[survival:coxph]{survival::coxph()}}
\item \code{conf_level} (\code{proportion})\cr confidence level of the interval for HR.
}}

\item{position_coxph}{(\code{numeric})\cr x and y positions for plotting \code{\link[survival:coxph]{survival::coxph()}} model.}

\item{position_surv_med}{(\code{numeric})\cr x and y positions for plotting annotation table estimating median survival
time per group.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
