% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot_data.R
\name{kmAnnoData}
\alias{kmAnnoData}
\title{Prepare K-M model annotation data with rtable format}
\usage{
kmAnnoData(fit_km)
}
\arguments{
\item{fit_km}{a class "survfit" object.}
}
\description{
An rtable format of KM model data for further annotation on top of Kaplan-Meier grob
}
\examples{

OS <- data.frame(AVAL = abs(rnorm(200)), 
                 CNSR = sample(c(0, 1), 200, TRUE), 
                 ARM = sample(LETTERS[1:3], 200, TRUE),
                 SEX = sample(c("M","F"), 200, TRUE),
                 RACE = sample(c("AA", "BB", "CC"), 200, TRUE),
                 ECOG = sample(c(0, 1), 200, TRUE))
fit_km <- survfit(Surv(AVAL, 1-CNSR) ~ ARM, data = OS, conf.type = "plain")
kmAnnoData(fit_km)


}
