% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot_data.R
\name{kmCurveData}
\alias{kmCurveData}
\title{Prepare necessary data for Kaplan-Meier Plot}
\usage{
kmCurveData(formula_km, data, xaxis_by = NULL)
}
\arguments{
\item{formula_km}{formula specified for Kaplan-Meier curves.}

\item{data}{analysis data set.}

\item{xaxis_by}{break interval of x-axis.}
}
\description{
Return a list of data for primitive element of grid drawing
}
\examples{

library(survival)
library(scales)

OS <- data.frame(AVAL = abs(rnorm(200)), 
                 CNSR = sample(c(0, 1), 200, TRUE), 
                 ARM = sample(LETTERS[1:3], 200, TRUE),
                 SEX = sample(c("M","F"), 200, TRUE),
                 RACE = sample(c("AA", "BB", "CC"), 200, TRUE),
                 ECOG = sample(c(0, 1), 200, TRUE))
kmCurveData(Surv(AVAL, 1-CNSR) ~ ARM, data = OS)
}
