% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_km.R
\name{kmGrob}
\alias{kmGrob}
\title{Define a kmplot grob}
\usage{
kmGrob(fit_km, xaxis_break = NULL, title = "Kaplan - Meier Plot",
  gp = NULL, vp = NULL, name = NULL)
}
\arguments{
\item{fit_km}{a class "survfit" object.}

\item{xaxis_break}{break interval of x-axis. It takes a numeric vector or \code{NULL}.}

\item{title}{title for plot.}
}
\description{
create a grid graphical object for basic KM plot from a \code{survfit} object.
}
\examples{
OS <- data.frame(AVAL = abs(rnorm(200)), 
                 CNSR = sample(c(0, 1), 200, TRUE), 
                 ARM = sample(LETTERS[1:3], 200, TRUE),
                 SEX = sample(c("M","F"), 200, TRUE),
                 RACE = sample(c("AA", "BB", "CC"), 200, TRUE),
                 ECOG = sample(c(0, 1), 200, TRUE))
fit_km <- survfit(Surv(AVAL, 1-CNSR) ~ ARM, data = OS, conf.type = "plain")
kmGrob(fit_km = fit_km, xaxis_break = c(0.5, 0.8, 1.5))

}
