% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot_regular.R
\name{kmPlot3}
\alias{kmPlot3}
\title{kmPlot}
\usage{
kmPlot3(formula, data)
}
\arguments{
\item{time_to_event}{time to event values}

\item{is_event}{boolean, \code{TRUE} if event and \code{FALSE} if censored}

\item{ref.arm}{character: reference arm}
}
\description{
descr
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr)

ATE <- ate(com.roche.cdpt7722.wo29637.rl)

ATE_filtered <- ATE \%>\% filter(PARAMCD == "OS")



kmPlot3(Surv(AVAL, I(1-CNSR)) ~ ARM, data = ATE_filtered)

# debug function
library(survival)

with(ATE_filtered, {
  time_to_event <<- AVAL
  event <<- CNSR == 0
  arm <<- ARM
 })
 ref.arm <- arm[1]

} 

}
