% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot_regular.R
\name{kmPlot_anno}
\alias{kmPlot_anno}
\title{Statistics from KM estimators and Cox PH modeling}
\usage{
kmPlot_anno(data, cox.tie = "efron", conf.int = FALSE,
  plot.median = FALSE, plot.nrisk = TRUE, time.interval = 4,
  nrisk.height = 0.25, size.nrisk = 4, plot.cens = TRUE,
  size.cens = 4.5, shape.cens = "+", plot.stats = TRUE, size.stats = 3,
  xystats.up = c(0.7, 0.8), xyinterval.up = c(0.08, 0.06),
  xystats.lo = c(0.05, 0.3), xyinterval.lo = c(0.07, 0.06),
  line.color = NULL, line.type = 1, line.width = 1, xlab = NULL,
  ylab = NULL, xlim = NULL, ylim = NULL,
  plot.title = "Kaplan-Meier Plot", legend.pos = "none")
}
\description{
Statistics from KM estimators and Cox PH modeling
}
