% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot_regular.R
\name{kmplot_survminer}
\alias{kmplot_survminer}
\title{kmplot_survminer}
\usage{
kmplot_survminer(time_to_event, event, arm, arm.ref, arm.rest = setdiff(arm,
  arm.ref), stratum_df = NULL, facet_by = NULL, n_col = 1, ...)
}
\arguments{
\item{time_to_event}{time to event values}

\item{event}{boolean, \code{TRUE} if event and \code{FALSE} if censored}

\item{arm.ref}{character: reference arm}

\item{stratum_df}{a data frame for stratification factor}
}
\description{
descr
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr)

ATE <-  ate(com.roche.cdt30019.go29436.re)

ATE_filtered <- ATE \%>\% filter(PARAMCD == "OS")


 kmplot_survminer(
   time_to_event = ATE_filtered$AVAL,
   event = ATE_filtered$CNSR == 0,
   arm = ATE_filtered$ARM,
   arm.ref =  "DUMMY C" 
   )
   
kmplot_survminer(
   time_to_event = ATE_filtered$AVAL,
   event = ATE_filtered$CNSR == 0,
   arm = ATE_filtered$ARM,
   arm.ref =  "DUMMY C" ,
   stratum_df = ATE_filtered[c("SEX")] 
   )
   
kmplot_survminer(
   time_to_event = ATE_filtered$AVAL,
   event = ATE_filtered$CNSR == 0,
   arm = ATE_filtered$ARM,
   facet_by = ATE_filtered$SEX,
   stratum_df = ATE_filtered[c("RACE", "HIST")], 
   arm.ref =  "DUMMY B" ,
   arm.rest =  c("DUMMY A", "DUMMY C" )
   )
 }
}
