% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_grid.R
\name{labelPanelGrob}
\alias{labelPanelGrob}
\title{Label Panel Grob
Create a panel style plotting of labels}
\usage{
labelPanelGrob(label, x, group, col = NA, name = NULL, gp = NULL,
  vp = NULL)
}
\arguments{
\item{label}{a vector of plot contents.}

\item{x}{a vector of xticks for each label to be plotted.}

\item{group}{a vector of group indicators each label content belongs to. It's a factor.}

\item{name}{name of grob.}

\item{gp}{graphical paramter.}

\item{vp}{a \code{\link{viewport}}.}
}
\description{
Label Panel Grob
Create a panel style plotting of labels
}
\examples{
label <- c("a", "b", "c", "d", "T", "R", "E", "G")
x <- rep(1:4, 2)
group <- factor(rep(c("GROUP1", "GROUP_new"), each = 4))
grid.newpage()
grid.draw(labelPanelGrob(label, x, group))
grid.newpage()
pushViewport(plotViewport())
grid.draw(labelPanelGrob(label, x, group))

label <- 1:18
x <- c(rep(1:4, 2), rep(1:5, 2))
group <- factor(c(rep(c("G1", "G2"), each = 4), rep(c("G3", "G4"), each = 5)))
col <- c( "black", "blue", "red","green" )

g <- labelPanelGrob(label, x, group, col)
grid.newpage()
pushViewport(plotViewport())
grid.draw(g)

g <- addGrob(g, xaxisGrob(at = unique(x), vp = "labelPlot"))
g <- addGrob(g, textGrob(levels(group), x = unit(-3, "lines"),
                         y = unit(1:nlevels(group)/(nlevels(group)+1), "npc"),
                         just = "left",  gp = gpar(col = col),
                          vp = "labelPlot"))
grid.newpage()
pushViewport(plotViewport(margins = c(5, 4, 4, 2)))
grid.draw(g)

}
