% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{labels_or_names}
\alias{labels_or_names}
\title{Get the labels or names if there are no labels from a list.}
\usage{
labels_or_names(x)
}
\arguments{
\item{x}{a list}
}
\value{
a character vector with the labels or names for the list elements
}
\description{
Internal helper function for working with nested statistic function results which typically
don't have labels but names that we can use.
}
