% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labels_over_names}
\alias{labels_over_names}
\title{return a vector with the variable labels and names if variable labels do not exist}
\usage{
labels_over_names(df)
}
\arguments{
\item{df}{data.frame object}
}
\description{
return a vector with the variable labels and names if variable labels do not exist
}
\examples{

X <- data.frame(
 a = structure(1:4, label = "label for a"),
 b = 3:6,
 d = structure(letters[1:4], label = "label for d"),
 stringsAsFactors = FALSE
)

View(X)

names(X)

}
