% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{list_length}
\alias{list_length}
\title{List Length Definition}
\usage{
list_length(x)
}
\arguments{
\item{x}{list or vector}
}
\value{
the length of the list, or 1 in case of a vector
}
\description{
Internal helper function to deal with possibly nested lists.
We might call this result "list length" since only lists count here.
}
