% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{list_lengths_in_list}
\alias{list_lengths_in_list}
\title{Get the list lengths of all list elements.}
\usage{
list_lengths_in_list(x)
}
\arguments{
\item{x}{a list}
}
\value{
the list lengths of all elements
}
\description{
Internal helper function to deal with possibly nested lists.
}
\seealso{
\code{\link{list_length}}
}
