% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_transpose.R
\name{list_transpose}
\alias{list_transpose}
\title{Transpose a list by changing depht 2 elements with depth 1 elements}
\usage{
list_transpose(x)
}
\arguments{
\item{x}{list of lists or list of vectors}
}
\value{
list of lists
}
\description{
Transpose a list by changing depht 2 elements with depth 1 elements
}
\examples{

x <- list(c1 = 1:3, c2 = 11:13, c3 = 111:113)
list_transpose(x)

x <- list(
  c1 = list(row1 = c(1,4), row2 = c("A"), row3 = "U", row4 = 1),
  c2 = list(row1 = c(9,23), row2 = c("B"), row3 = "V", row4 = 12),
  c3 = list(row1 = c(2,2), row2 = c("C"), row3 = "S", row4 = 99)
)
list_transpose(x)

x <- list(c1 = list(row1 = c(1,4)))
list_transpose(x)

x <- list(c1 = 1:3) 
list_transpose(x)
}
