% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{logistic_regression}
\alias{logistic_regression}
\alias{fit_logistic}
\alias{tidy.glm}
\alias{s_logistic}
\alias{summarize_logistic}
\title{Multi-variable logistic regression table}
\usage{
fit_logistic(
  data,
  variables = list(response = "Response", arm = "ARMCD", covariates = NULL, interaction
    = NULL)
)

\method{tidy}{glm}(fit_glm, conf_level = 0.95)

s_logistic(df, .var)

summarize_logistic(
  lyt,
  conf_level,
  vars = c("df", "estimate", "std_error", "odds_ratio", "ci", "pvalue")
)
}
\arguments{
\item{data}{(\verb{data frame})\cr the data frame on which the model was fit.}

\item{variables}{(named \code{list} of \code{string})\cr list of additional analysis variables.}

\item{fit_glm}{logistic regression model fitted by \code{\link[stats:glm]{stats::glm()}} with "binomial" family.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr the name of statistics to be reported among
\code{df} (degrees of freedom), \code{estimate} (regression coefficient),
\code{std_error} (standard error of coefficient), \code{odds_ratio} (odds ratio) and
\code{ci} (confidence interval of odds ratio), \code{pvalue} (p.value of the effect).}
}
\description{
Logistic regression for binary outcome with categorical/continuous covariates in model statement.
For each covariate category (if categorical) or specified values (if continuous), present degrees of freedom,
regression parameter estimate and standard error (SE) relative to reference group or category.
Report odds ratios for each covariate category or specified values and corresponding Wald
confidence intervals as default but allow user to specify other confidence levels.
Report p-value for Wald chi-square test of the null hypothesis that covariate has no effect on
response in model containing all specified covariates.
Allow option to include one two-way interaction and present similar output for
each interaction degree of freedom.
Note: For \code{glm} formula, the variable names need to be standard dataframe column name without
special characters. The big N is the total number of observations for complete cases.
}
\section{Functions}{
\itemize{
\item \code{fit_logistic}: Fit a logistic regression model.

\item \code{tidy.glm}: Helper method (for \code{\link[broom:reexports]{broom::tidy()}}) to prepare a data frame from an
\code{glm} object with \code{binomial} family.

\item \code{s_logistic}: transforms the tabulated results from \code{tidy}
into a list.

\item \code{summarize_logistic}: layout creating function.
}}

\examples{
library(random.cdisc.data)
library(dplyr)
adrs <- radrs(cached = TRUE)
adrs_f <- adrs \%>\%
  dplyr::filter(PARAMCD == "BESRSPI") \%>\%
  dplyr::filter(RACE \%in\% c("ASIAN", "WHITE", "BLACK OR AFRICAN AMERICAN")) \%>\%
  dplyr::mutate(
    Response = case_when(AVALC \%in\% c("PR", "CR") ~ 1, TRUE ~ 0),
    RACE = factor(RACE),
    SEX = factor(SEX)
  )
mod1 <- fit_logistic(
  data = adrs_f,
  variables = list(
    response = "Response", arm = "ARMCD", covariates = c("AGE", "RACE")
  )
)
mod2 <- fit_logistic(
  data = adrs_f,
  variables = list(
    response = "Response", arm = "ARMCD",
    covariates = c("AGE", "RACE"), interaction = "AGE"
  )
)

df <- broom::tidy(mod1, conf_level = 0.99)

s_logistic(broom::tidy(mod1), .var = "ci")

result <- basic_table() \%>\%
  split_rows_by("variable") \%>\%
  split_rows_by("term", split_fun = drop_split_levels) \%>\%
  summarize_logistic(conf_level = .99) \%>\%
  build_table(df = df)
result

}
