% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{logistic_regression_cols}
\alias{logistic_regression_cols}
\title{Logistic Regression Multivariate Column Layout Function}
\usage{
logistic_regression_cols(lyt, conf_level = 0.95)
}
\arguments{
\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}
}
\value{
A layout object suitable for passing to further layouting functions. Adding this
function to an \code{rtable} layout will split the table into columns corresponding to
statistics \code{df}, \code{estimate}, \code{std_error}, \code{odds_ratio}, \code{ci}, and \code{pvalue}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Layout-creating function which creates a multivariate column layout summarizing logistic
regression results. This function is a wrapper for \code{\link[rtables:split_cols_by_multivar]{rtables::split_cols_by_multivar()}}.
}
