% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{lt_ae_class_term}
\alias{lt_ae_class_term}
\title{By class By Term}
\usage{
lt_ae_class_term(class, term, id, col_by, col_N = tapply(col_by, col_by,
  length), total = "All Patients", class_label, term_label)
}
\description{
By class By Term
}
\examples{
l_tbls <- lt_ae_class_term(
  class = ANL$CLASS,
  term =  ANL$TERM,
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients"
)
stack_rtables_d2(l_tbls)

}
