% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{lt_ae_max_grade_d2}
\alias{lt_ae_max_grade_d2}
\title{list of tables}
\usage{
lt_ae_max_grade_d2(class, term, id, grade, col_by, col_N = tapply(col_by,
  col_by, length), total = "All Patients", grade_levels, class_label,
  term_label, grade_label)
}
\description{
list of tables
}
\examples{
# Simple example
library(tibble)
library(dplyr)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = paste("ARM", LETTERS[rep(c(1,2), c(3,7))])
)


ae_lookup <- tribble(
  ~CLASS,         ~TERM,   ~GRADE,
  "cl A",   "trm A_1/2",        1,
  "cl A",   "trm A_2/2",        2,  
  "cl B",   "trm B_1/3",        2,
  "cl B",   "trm B_2/3",        3,
  "cl B",   "trm B_3/3",        1,
  "cl C",   "trm C_1/1",        1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2,2,2,3,3,4,4,4,4,5,6,6,7,7)]
  ),
  ae_lookup[c(1,1,2,6,4,2,2,3,4,2,1,5,4,6),]
)

ANL <- left_join(AAE, ASL[, c("USUBJID", "ARM")], by ="USUBJID")

l_tbls <- lt_ae_max_grade_d2(
  class = ANL$CLASS,
  term =  ANL$TERM,
  id = ANL$USUBJID,
  grade = ANL$GRADE,
  col_by = factor(ANL$ARM),
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients",
  grade_levels = 1:5,
  class_label = 'MedDRA System Organ Class',
  term_label = 'MedDRA Preferred Term'
)

stack_rtables_d2(l_tbls)


\dontrun{
  ASL <- osprey::rADSL
  AAE <- osprey::rADAE
  
  head(AAE)
  
l_tbls <- lt_ae_max_grade_d2(
  class = AAE$AESOC,
  term =  AAE$AEDECOD,
  id = AAE$USUBJID,
  grade = as.numeric(AAE$AETOXGR),
  col_by = factor(AAE$ARM),
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients",
  grade_levels = 1:5,
  class_label = 'MedDRA System Organ Class',
  term_label = 'MedDRA Preferred Term'
)

tbl <- stack_rtables_d2(l_tbls)

Viewer(tbl  )
  
}
}
