% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{lt_ae_max_grade_term}
\alias{lt_ae_max_grade_term}
\title{List of Adverse Events Terms Tables By Highest Grade (Term only)}
\usage{
lt_ae_max_grade_term(term, id, grade, col_by, col_N, total = "All Patients",
  grade_levels)
}
\arguments{
\item{term}{term information as character or factor, however factor levels
are not repeated by class, only terms with count > 1 are listed per class.
\code{var_relabel} is used as the character string used as a label in the
column header for each term.}

\item{id}{unique subject identifier. If a particular subject has no adverse
event then that information needs to be added to the \code{col_N} argument.}

\item{grade}{grade of adverse event as numeric. \code{var_relabel} is used as the character 
string used as a label in the column header for each grade.}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing. See 'Examples'.}

\item{col_N}{numeric vector with information of the number of patients in the
levels of \code{col_by}. This is useful if there are patients that have no
adverse events can be accounted for with this argument.}

\item{total}{character string that will be used as a label for a column with 
pooled total population, default is "All Patients". If the levels of col_by are 
the only columns of interest then total should be \code{NULL}}

\item{grade_levels}{numeric, ordered values of possible of grades in a form
of \code{x:y}, default is \code{1:5}.}
}
\description{
\code{lt_ae_max_grade_term} returns a nested list of adverse events tables by max
grade (\code{\link{t_max_grade_per_id}}).
}
\examples{

library(dplyr)
library(random.cdisc.data)

ASL <- rasl(10, seed = 1)
AAE <- raae(ASL, 4, seed = 2)

ANL <- left_join(AAE, ASL \%>\% select(USUBJID, STUDYID, ARM), by = c("USUBJID", "STUDYID"))

l_tbls <- tern:::lt_ae_max_grade_term(
  term = ANL$AEDECOD,
  id = ANL$USUBJID,
  grade = ANL$AETOXGR,
  col_by = ANL$ARM,
  col_N = table(ASL$ARM),
  total = "All Patients",
  grade_levels = 1:5
)

do.call(fast_stack_rtables, l_tbls)

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
