% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_names}
\alias{make_names}
\title{Make Names Without Dots}
\usage{
make_names(nams)
}
\arguments{
\item{nams}{(\code{character})\cr vector of original names.}
}
\value{
Character vector of proper names, which does not use dots in contrast to
\code{\link[base:make.names]{base::make.names()}}.
}
\description{
Make Names Without Dots
}
\examples{
make_names(c("foo Bar", "1 2 3 bla"))

}
\keyword{internal}
