% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_available}
\alias{n_available}
\title{Number of available (non-missing entries) in a vector}
\usage{
n_available(x)
}
\arguments{
\item{x}{(\code{vector})\cr vector in which to count non-missing values.}
}
\value{
Number of non-missing values.
}
\description{
Small utility function for better readability.
}
\keyword{internal}
