% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NaReplace}
\alias{NaReplace}
\title{Replace empty string as NA in a data table}
\usage{
NaReplace(data.tbl)
}
\arguments{
\item{data.tbl}{input data table}
}
\description{
Convert factor to character, convert empty string and "." for all columns of a data table
}
