% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NaReplaceVec}
\alias{NaReplaceVec}
\title{Replace empty string and "." as NA in a vector}
\usage{
NaReplaceVec(input.vec)
}
\arguments{
\item{input.vec}{input vector}
}
\description{
Convert "", "   " and "." to NA in a vector
}
