% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary_variants.R
\name{nb_entries}
\alias{nb_entries}
\title{Returns the number of entries in x}
\usage{
nb_entries(x)
}
\arguments{
\item{x}{object}
}
\description{
For a data.frame, it is \code{nrow}. Defaults to \code{length}.
}
\examples{
nb_entries(1:3)
nb_entries(as.numeric(1:3))
nb_entries(data.frame(x = 1:3, y = 4:6))
}
