% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_colors.R
\name{nest_color_palette}
\alias{nest_color_palette}
\title{NEST color palette}
\usage{
nest_color_palette(n = 10, palette = "nest")
}
\arguments{
\item{n}{(\code{numeric}) The number of colors to be returned from
  the color palettes. Please note the colors will be repeated after
  a certain numerical limit per palette:

  \itemize{
     \item{nest }{96}
     \item{stream }{38}
     \item{viridis }{49}
  }}

\item{palette}{(\code{character}) The name of a palette supported
  by this function

  \itemize{
     \item{\code{nest} }{A color palette developed by the NEST team representing
      colors in the NEST logo.
    }
    \item{stream }{A color palette proposed by the stream documentation of stream version \code{2_05}}
       \item{viridis }{A color palette provided by the \code{viridis::plasma} function.}
  }}
}
\description{
A standardized color palette to be used for all plots within
the NEST project.
}
