% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{nested_by}
\alias{nested_by}
\title{Recursive by}
\usage{
nested_by(x)
}
\arguments{
\item{x}{a list that specifies recursive split,
e.g. a list of col_bys (which can be matrices or factors)}
}
\description{
When a by of this type is encountered in tern functions, it recursively splits by it
\code{is.list} still returns TRUE on the returned object.
It is typically used to construct a nested tree from it.
}
\details{
The nested_by class is used to distinguish the simple_by (which can be a matrix
which is a list as well, but in this case a non-recursive interpretation as in col_by)
from the recursive
}
