% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{nested_list_to_tree}
\alias{nested_list_to_tree}
\title{Convert a named nested List to a tree}
\usage{
nested_list_to_tree(x, format_data = NULL)
}
\arguments{
\item{x}{nested list}

\item{format_data}{format data for conversion with \code{\link{to_rtable}},
\code{list(gap_to_children, children_gap, children_indent, content_indent)},
default values if not all given, see that function}
}
\description{
Convert a named nested List to a tree
}
\examples{
lst <- list(
  A = list(b = 1, c = iris),
  B = list(e = 1:3, C = list(f = 5, g = 9, h = "a"))
)
n1 <- nested_list_to_tree(lst)
summary(n1)


}
