% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\docType{class}
\name{node-class}
\alias{node-class}
\alias{node-validity}
\alias{node}
\title{An S4 Class to represent nodes in a tree}
\usage{
node(
  name,
  content,
  children = list(),
  format_data = list(),
  label = character(0)
)
}
\arguments{
\item{name}{name of node}

\item{content}{content of node}

\item{children}{children of node}

\item{format_data}{format data for conversion with \code{\link{to_rtable}},
\code{list(gap_to_children, children_gap, children_indent, content_indent, left_header)},
default values if not all given, see that function}

\item{label}{\code{character} A label for the name}
}
\value{
object of class node
}
\description{
From Wikipedia: A tree data structure can be defined recursively as a collection of nodes (starting at a root node),
where each node is a data structure consisting of a value, together with a list of references to nodes (the
"children"), with the constraints that no reference is duplicated, and none points to the root.

Children must not be \code{NULL}

A node is also a tree. This function makes sure that the children are properly named.
They can then be properly accessed through \code{node_children[['childName']]}.
See the class node for more info about parameters.
}
\details{
Use the function \code{\link{node}} to create objects of this class.

Note: changing the names of the children after calling this function will cause the children list names
to be no longer in-sync because assignment via the \code{@} operator does not check whether the new object is valid,
e.g. names of children nodes agree with names in children list.
Ideally, we should define functions like \code{setChildren} and \code{getChildren}, but this seems to clutter the
code.
Temporarily, we always reassign children names when accessing via \code{`[[`}.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of node}

\item{\code{content}}{of node (e.g. \code{data.frame} or \code{rtable} objects)}

\item{\code{children}}{a list of \code{node} objects}

\item{\code{format_data}}{formatting instructions for conversion with \code{\link{to_rtable}}}

\item{\code{label}}{\code{character} A label for the name}
}}

\examples{

n11 <- node(name = "A", content = array(c(1:6), dim = c(2,3)), children = list())
n12 <- node(name = "B", content = array(c(1:6), dim = c(2,3)), children = list())
n13 <- node(name = "C", content = array(c(1:6), dim = c(2,3)), children = list())
n2 <- node(name = "D", content = c(1:3), children = list(n11, n12, n13))

summary(n11)
summary(n2)

\dontrun{
# incorrect example (children with same name):
node(name = "A", content = c(1:3), children = list("A"))
}

}
