% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{node_format_data}
\alias{node_format_data}
\title{Node Format Data}
\usage{
node_format_data(gap_to_children = NULL, children_gap = NULL,
  children_indent = NULL, content_indent = NULL)
}
\arguments{
\item{gap_to_children}{row gap between content and children}

\item{children_gap}{row gaps between children}

\item{children_indent}{absolute indentation of children}

\item{content_indent}{absolute indentation of}
}
\description{
Node Format Data
}
\examples{
tbl <- function(x) {rtable(c("A", "B"), rrow("r1", "-", "-"))}

to_rtable(node("AAA", tbl("r1")))
to_rtable(node("AAA", tbl("r1"), children = list(
   node("c1", tbl("r2")),
   node("c1", tbl("r2")),
   node("c1", tbl("r2"))
)))


to_rtable(node("AAA", tbl("r1"), children = list(
   node("c1", tbl("r2")),
   node("c1", tbl("r2")),
   node("c1", tbl("r2"))
), format_data = node_format_data(children_indent = 0)))

}
