% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{non_rsplit}
\alias{non_rsplit}
\title{Split list in the classic way instead of applying split to each element}
\usage{
non_rsplit(x)
}
\arguments{
\item{x}{list to protect from splitting}
}
\description{
See \code{\link{esplit.list}} for the alternative behavior.
}
