% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{object_to_node}
\alias{object_to_node}
\title{Create a node with only this object}
\usage{
object_to_node(x,
  node_name = invisible_node_name(deparse(substitute(x))))
}
\arguments{
\item{x}{object to add as node}

\item{node_name}{name of the node}
}
\value{
node
}
\description{
To add children, use the \code{\link{node}} function.
}
\examples{
object_to_node(1:2)
my_obj <- 1:2
object_to_node(my_obj)
object_to_node(my_obj, node_name = "hello")
}
