% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_pkparam_sort.R
\name{h_pkparam_sort}
\alias{h_pkparam_sort}
\title{Sort PK PARAM variable}
\usage{
h_pkparam_sort(pk_data)
}
\arguments{
\item{pk_data}{(\code{data.frame})\cr Pharmacokinetics dataframe}
}
\value{
dataframe with PK PARAM variable ordered
}
\description{
Sort PK PARAM variable
}
\section{Functions}{
\itemize{
\item \code{h_pkparam_sort}: PK PARAM variable based on the TLG order from dataset generated by d_pkparam
}}

\examples{
library(scda)
library(rtables)
library(dplyr)
library(tern)

adpp <- synthetic_cdisc_data("latest")$adpp
adpp <- adpp \%>\% mutate(PKPARAM = factor(paste0(PARAM, " (", AVALU, ")")))

pk_ordered_data <- h_pkparam_sort(adpp)
}
