% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{pairwise}
\alias{pairwise}
\title{Pairwise formula special term}
\usage{
pairwise(x)
}
\arguments{
\item{x}{the variable for which pairwise result is expected}
}
\description{
The special term \code{pairwise} indicate that the model should be fitted individually for
every tested level in comparison to the reference level.
}
\details{
Let's \code{ARM} being a factor with level A, B, C; let's be B the reference level,
a model calling the formula including \code{pairwise(ARM)} will result in two models
\itemize{
\item a model including only levels A and B, and effect of A estimated in reference to B.
\item a model including only levels C and B, the effect of C estimated in reference to B.
}
}
