% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_object.R
\docType{data}
\name{pool_obj}
\alias{pool_obj}
\title{Catched \code{pool} object obtained by \code{rbmi} package}
\format{
An object of class \code{pool} of length 5.
}
\usage{
pool_obj
}
\description{
Catched \code{pool} object obtained by \code{rbmi} package
}
\keyword{datasets}
