% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_response_table.R
\name{print_np}
\alias{print_np}
\title{Helper functions to format rcells for various display}
\usage{
print_np(x)
}
\arguments{
\item{x}{list produced by calc_rate/calc_diffor function}

\item{y}{name of element to display}

\item{z}{colspan for the cell}
}
\description{
Helper functions to format rcells for various display
}
\examples{

a_rate <- calc_rate(LETTERS[sample(1:3, 20, replace=T)], value = "A")
b_rate <- calc_rate(LETTERS[sample(1:3, 20, replace=T)], value = "A")
diffor <- calc_diffor(b_rate, a_rate)

print_np(a_rate)
print_ci(a_rate, "ci", 2)
print_ci(diffor, "diffci", 2)
print_diffor(diffor, "diff", 1)
print_diffor(diffor, "or", 1)

}
\author{
Chendi Liao (liaoc10), \email{chendi.liao@roche.com}
}
