% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{prop}
\alias{prop}
\alias{prop_wilson}
\title{Proportions}
\usage{
prop_wilson(rsp, conf_level)
}
\arguments{
\item{rsp}{(\code{logical})\cr whether each subject is a responder or not.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}
}
\description{
Estimate proportion for the study of binary outcomes.
}
\section{Functions}{
\itemize{
\item \code{prop_wilson}: The Wilson interval calls \code{\link[stats:prop.test]{stats::prop.test()}}
with option \code{correct = FALSE}. Also referred to as Wilson score interval.
}}

\examples{
rsp <- c(
  TRUE, TRUE, TRUE, TRUE, TRUE,
  FALSE, FALSE, FALSE, FALSE, FALSE
)
prop_wilson(rsp, conf_level = 0.9)

}
