% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportion.R
\name{prop_agresti_coull}
\alias{prop_agresti_coull}
\title{(for 95\% CI) adding two successes and two failures to the data and then using the Wald formula to construct a CI.}
\usage{
prop_agresti_coull(rsp, conf_level)
}
\description{
(for 95\% CI) adding two successes and two failures to the data and then using the Wald formula to construct a CI.
}
\examples{
prop_agresti_coull(rsp, conf_level = 0.95)

}
