% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_diff.R
\name{prop_difference}
\alias{prop_difference}
\alias{prop_diff_wald}
\alias{prop_diff_ha}
\alias{prop_diff_nc}
\alias{prop_diff_cmh}
\alias{s_proportion_diff}
\alias{estimate_proportion_diff}
\title{Proportion Difference}
\usage{
prop_diff_wald(rsp, grp, conf_level, correct)

prop_diff_ha(rsp, grp, conf_level)

prop_diff_nc(rsp, grp, conf_level)

prop_diff_cmh(rsp, grp, strata, conf_level = 0.95)

s_proportion_diff(
  df,
  .var,
  .ref_group,
  .in_ref_col,
  variables = list(strata = NULL),
  conf_level = 0.95,
  method = c("waldcc", "wald", "cmh", "ha", "newcombe")
)

estimate_proportion_diff(
  lyt,
  vars,
  var_labels = vars,
  show_labels = "hidden",
  ...
)
}
\arguments{
\item{rsp}{(\code{logical})\cr whether each subject is a responder or not.}

\item{grp}{(\code{factor})\cr
vector assigning observations to one out of two groups
(e.g. reference and treatment group).}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{correct}{\code{logical}\cr
include the continuity correction.}

\item{strata}{(\code{factor})\cr
with one level per stratum and same length as \code{rsp}.}

\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.ref_group}{(\verb{data frame} or \code{vector})\cr the data corresponding to the reference group.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}

\item{variables}{(named \code{list} of \code{strings})\cr list of additional analysis variables.}

\item{method}{(\code{string})\cr
the method used for the confidence interval estimation.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{var_labels}{character for label.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{...}{arguments passed to \code{s_proportion_diff()}.}
}
\description{
Proportion Difference
}
\section{Functions}{
\itemize{
\item \code{prop_diff_wald}: The Wald interval follows the usual textbook
definition for a single proportion confidence interval using the normal
approximation. It is possible to include a continuity correction for Wald's
interval.

\item \code{prop_diff_ha}: Anderson-Hauck confidence interval.

\item \code{prop_diff_nc}: Newcombe confidence interval. It is based on
the Wilson score confidence interval for a single binomial proportion.

\item \code{prop_diff_cmh}: Calculates the weighted difference.
This is defined as the difference in response rates between the
experimental treatment group and the control treatment group, adjusted
for stratification factors by applying Cochran-Mantel-Haenszel (CMH)
weights. For the CMH chi-squared test, use \code{\link[stats:mantelhaen.test]{stats::mantelhaen.test()}}.

\item \code{s_proportion_diff}: Statistics function estimating the difference
in terms of responder proportion.

\item \code{estimate_proportion_diff}: Adds a descriptive analyze layer to \code{rtables}
pipelines. The analysis is applied to a \code{dataframe} and return the
estimations, in \code{rcells}. The ellipsis (\code{...}) conveys arguments to
\code{s_proportion_diff()}, for instance \code{na.rm = FALSE} if missing data
should be accounted for.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{

# Wald confidence interval

set.seed(2)
rsp <- sample(c(TRUE, FALSE), replace = TRUE, size = 20)
grp <- c(rep("A", 10), rep("B", 10))
prop_diff_wald(rsp = rsp, grp = grp, conf_level = 0.90, correct = FALSE)

# Anderson-Hauck confidence interval

## "Mid" case: 3/4 respond in group A, 1/2 respond in group B.
rsp <- c(TRUE, FALSE, FALSE, TRUE, TRUE, TRUE)
grp <- factor(c("A", "B", "A", "B", "A", "A"), levels = c("B", "A"))
prop_diff_ha(rsp = rsp, grp = grp, conf_level = 0.90)

## Edge case: Same proportion of response in A and B.
rsp <- c(TRUE, FALSE, TRUE, FALSE)
grp <- factor(c("A", "A", "B", "B"), levels = c("A", "B"))
prop_diff_ha(rsp = rsp, grp = grp, conf_level = 0.6)


# Newcombe confidence interval

set.seed(1)
rsp <- c(
  sample(c(TRUE, FALSE), size = 40, prob = c(3/4, 1/4), replace = TRUE),
  sample(c(TRUE, FALSE), size = 40, prob = c(1/2, 1/2), replace = TRUE)
)
grp <- factor(rep(c("A", "B"), each = 40), levels = c("B", "A"))
table(rsp, grp)
prop_diff_nc(rsp = rsp, grp = grp, conf_level = 0.9)


# Cochran-Mantel-Haenszel confidence interval

set.seed(2)
rsp <- sample(c(TRUE, FALSE), 100, TRUE)
grp <- sample(c("Placebo", "Treatment"), 100, TRUE)
grp <- factor(grp, levels = c("Placebo", "Treatment"))
strata_data <- data.frame(
  "f1" = sample(c("a", "b"), 100, TRUE),
  "f2" = sample(c("x", "y", "z"), 100, TRUE),
  stringsAsFactors = TRUE)

prop_diff_cmh(
  rsp = rsp, grp = grp, strata = interaction(strata_data),
  conf_level = 0.90
)


# Summary

## "Mid" case: 4/4 respond in group A, 1/2 respond in group B.
dta <- data.frame(
  rsp = c(TRUE, FALSE, FALSE, TRUE, TRUE, TRUE),
  grp = factor(c("A", "B", "A", "B", "A", "A"), levels = c("B", "A"))
)

s_proportion_diff(
  df = subset(dta, grp == "A"),
  .var = "rsp",
  .ref_group = subset(dta, grp == "B"),
  .in_ref_col = FALSE,
  conf_level = 0.90,
  method = "ha"
)


l <- split_cols_by(lyt = NULL, var = "grp", ref_group = "B") \%>\%
  estimate_proportion_diff(
    vars = "rsp",
    conf_level = 0.90,
    method = "ha"
  )

build_table(l, df = dta)

}
