% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{range_noinf}
\alias{range_noinf}
\title{Re-implemented \code{\link[base:range]{base::range()}} Default S3 method for numerical objects}
\usage{
range_noinf(x, na.rm = FALSE, finite = FALSE)
}
\arguments{
\item{x}{(\code{numeric})\cr a sequence of numbers for which the range is computed.}

\item{na.rm}{(\code{logical})\cr indicating if \code{NA} should be omitted.}

\item{finite}{(\code{logical})\cr indicating if non-finite elements should be removed.}
}
\value{
A 2-element vector of class \code{numeric}.
}
\description{
This function returns \code{c(NA, NA)} instead of \code{c(-Inf, Inf)} for zero-length data
without any warnings.
}
\examples{
# Internal function - range_noinf
\dontrun{
range_noinf(1:5)
range_noinf(c(1:5, NA, NA), na.rm = TRUE)
range_noinf(numeric(), na.rm = TRUE)
range_noinf(c(1:5, NA, NA, Inf), na.rm = TRUE, finite = TRUE)
range_noinf(Inf)
range_noinf(Inf, na.rm = TRUE, finite = TRUE)
range_noinf(c(Inf, NA), na.rm = FALSE, finite = TRUE)
range_noinf(c(1, Inf, NA), na.rm = FALSE, finite = TRUE)
}

}
\keyword{internal}
