% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{rapply_data_tree}
\alias{rapply_data_tree}
\title{Recursive Apply A Function to leafs of a Data Tree}
\usage{
rapply_data_tree(x, f, apply_to_NULL = FALSE, ...)
}
\arguments{
\item{x}{a list or nested list of \code{data.frame}, if \code{x} is a \code{data.frame}}

\item{f}{a function to be applied to leafs}

\item{apply_to_NULL}{boolean, should \code{f} also be applied to a \code{NULL} leaf or should \code{NULL} leafs
remain \code{NULL} leafs.}

\item{...}{arguments passed to \code{f}}
}
\description{
The return value of \code{rapply_data_tree} has the same tree structure as the input \code{data_tree} object.
}
