% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_trees.R
\name{rbind_table_tree}
\alias{rbind_table_tree}
\title{Recursive Bind rtable Objects while Retaining Tree Hierarchy}
\usage{
rbind_table_tree(x, gap = 1)
}
\arguments{
\item{x}{a \code{\link{table_tree}} object}

\item{gap}{argument passed to \code{\link[rtables]{rbind}}.}
}
\description{
\code{table_tree} obejcts can be named, when recursivly stacking it is sometimes desired to add these names as rrows
and indent the child elements accordingly to preserve the ancestry information.
}
\examples{
tbl <- rtable(header = 1:3, rrowl("row", letters[1:3]))

nl1 <- list(tbl1 = tbl, tbl2 = tbl, tbl3 = NULL)
rbind_table_tree(table_tree(nl1))

nl2 <- list(A = nl1, B = nl1)
rbind_table_tree(table_tree(nl2))

nl3 <- list(X = nl2, Y = nl2, Z = nl2)
rbind_table_tree(nl3)


}
