% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reapply_varlabels}
\alias{reapply_varlabels}
\title{Reapply Variable Labels}
\usage{
reapply_varlabels(x, varlables, ...)
}
\arguments{
\item{x}{(\code{vector}) vector of elements that needs new labels.}

\item{...}{further parameters to be added to the list.}

\item{varlabels}{(\code{vector}) labels for \code{x}.}
}
\description{
This is an internal function that is used in template tests.
}
