% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{recursive_get_label}
\alias{recursive_get_label}
\alias{recursive_get_label,node-method}
\title{Returns output to be displayed with cat()}
\usage{
recursive_get_label(x)

\S4method{recursive_get_label}{node}(x)
}
\arguments{
\item{x}{object of type \link{node}}

\item{...}{other arguments}
}
\description{
Returns output to be displayed with cat()

Get the header for a nested node tree
}
