% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{recursive_stack_rtables}
\alias{recursive_stack_rtables}
\title{Stack Tables Stored in a nested list of depth 2}
\usage{
recursive_stack_rtables(x)
}
\arguments{
\item{x}{list with lists of rtables}

\item{n}{number of rows to remove from tables that that are not first in the
nested lists}
}
\value{
rtable
}
\description{
This function expects a list with lists of rtables to be stacked. Sometimes
these tables have repeated information at the top and hence the firs n rows
can be optionally removed from the tables that are not first in the lists.
}
\examples{

l_tbls <- list(
  list(
     rtabulate(iris$Sepal.Length, iris$Species, mean),
     rtabulate(iris$Sepal.Length, iris$Species, sd)
  ),
  list(
     rtabulate(iris$Sepal.Width, iris$Species, mean),
     rtabulate(iris$Sepal.Width, iris$Species, sd)
  ),
  list(
     rtabulate(iris$Petal.Length, iris$Species, mean),
     rtabulate(iris$Petal.Length, iris$Species, sd)
  )   
)

recursive_stack_rtables(l_tbls)

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
