% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflast.R
\name{ref_group_last}
\alias{ref_group_last}
\title{Split Function to Place Reference Group Facet Last}
\usage{
ref_group_last(splret, spl, fulldf)
}
\arguments{
\item{splret}{result of the core split}

\item{spl}{split object}

\item{fulldf}{data.frame of incoming data to be split}
}
\description{
Place reference group facet last during post-processing stage in a custom split function.
}
\examples{
library(dplyr)

# Define custom split function
ref_last <- make_split_fun(post = list(ref_group_last))

dat <- data.frame(
  x = factor(letters[1:5], levels = letters[5:1]),
  y = 1:5
)

# With rtables layout functions
basic_table() \%>\%
  split_cols_by("x", ref_group = "c", split_fun = ref_last) \%>\%
  analyze("y") \%>\%
  build_table(dat)

# With tern layout funcitons
adtte_f <- tern_ex_adtte \%>\%
  filter(PARAMCD == "OS") \%>\%
  mutate(
    AVAL = day2month(AVAL),
    is_event = CNSR == 0
  )

basic_table() \%>\%
  split_cols_by(var = "ARMCD", ref_group = "ARM B", split_fun = ref_last) \%>\%
  add_colcounts() \%>\%
  surv_time(
    vars = "AVAL",
    var_labels = "Survival Time (Months)",
    is_event = "is_event",
  ) \%>\%
  build_table(df = adtte_f)

}
\seealso{
\code{\link[rtables:make_split_fun]{rtables::make_split_fun()}}
}
