% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmrm.R
\name{refit_lme4_all_optimizers}
\alias{refit_lme4_all_optimizers}
\title{Re-Fitting \code{lme4} Model}
\usage{
refit_lme4_all_optimizers(original_fit, n_cores = 1L)
}
\arguments{
\item{original_fit}{original model fit coming from \code{\link[=fit_lme4_single_optimizer]{fit_lme4_single_optimizer()}}.}

\item{n_cores}{(\code{count})\cr number of cores which could in principle be used for
parallelizing the computations on Linux or Mac machines.}
}
\value{
The "best" model fit, defined as a converging fit without any warnings or messages, leading
to the highest log-likelihood. If no optimizer succeeds, then an error is thrown.
}
\description{
Refit an lme4 model with all possible optimizers and return the best result.
}
\note{
Currently there are 7 optimizers in total. Since 1 optimizer is already used in the original fit,
only 6 additional optimizer runs need to be done. Thus the maximum number of parallel runs is 6.
For Windows, no parallelization is currently implemented.
}
\keyword{internal}
