% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorder_to_match_id}
\alias{reorder_to_match_id}
\title{Return Ordered Dataset so that a set of variables match exactly}
\usage{
reorder_to_match_id(x, ref, key = c("USUBJID", "STUDYID"))
}
\arguments{
\item{x}{data set to reorder}

\item{ref}{reference data set}

\item{key}{variables that define a unique patient}
}
\description{
This function is useful to ensure that two datasets have the same subjects
ordered in the same way
}
\examples{

A <- data.frame(USUBJID=paste0("id-",1:10), STUDYID = "A", stringsAsFactors = FALSE)
B <- data.frame(USUBJID=paste0("id-",10:1), STUDYID = "A", stringsAsFactors = FALSE)
 
reorder_to_match_id(A, B)

\dontrun{
C <- data.frame(USUBJID=paste0("id-",1:9), STUDYID = "A")
reorder_to_match_id(A, C)

D <- B
D$STUDYID[3] <- "B"
reorder_to_match_id(A, D)

E <- B
E$USUBJID[2] <- "id-10"
reorder_to_match_id(A, E)
}


}
