% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_factor.R
\name{replace_emptys_with_na}
\alias{replace_emptys_with_na}
\title{Replace all occurences of empty string in data frame}
\usage{
replace_emptys_with_na(df, rep_str = "NA")
}
\arguments{
\item{df}{\code{data.frame} table to act upon}

\item{rep_str}{replacement string for empty strings}
}
\description{
fnc to fix for update in Rtables issue \code{tern#593} (\code{NA} alternative -> \code{" "})
This function is similar in scope as \code{df_explicit_na()}. In the future a merge
is to be expected.
}
\details{
this functions relies onto \code{for} loop, \code{levels}, and \code{nchar}. It can
easily be optimized for different cases. For the moment, this fits the purpose
to fix issue \code{tern#593}
}
\seealso{
\code{\link[=df_explicit_na]{df_explicit_na()}}
}
\keyword{internal}
