% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_table.R
\name{response_table}
\alias{response_table}
\title{Response Table Caclulation Function}
\usage{
response_table(response, value.resp = c("CR", "PR"), value.nresp = c("SD",
  "NON CR/PD", "PD", "NE"), arm, style = 1)
}
\arguments{
\item{response}{Tumor response data}

\item{arm}{Arm information data as factor}

\item{style}{Must be 1 or 2, \code{1} if only want to display rates summary
for each response value category, \code{2} if want to display rates,
difference in rate and odds ratio for each response value category.}

\item{value.rsp}{Character vector, defining list of response values to be
used as responders}

\item{value.nrsp}{Character vector, defining list of response values to be
used as non-responders}
}
\description{
Response Table Caclulation Function
}
\details{
We use the \code{PropCIs} package to calculate the CIs for odds ratio
}
\examples{


\dontrun{
library(dplyr)
library(atezo.data)
library(teal.oncology)
library(forcats)

ARS <- ars(com.roche.cdt30019.go29436.re)

tbl_stream <- get_response_table(com.roche.cdt30019.go29436.re)
Viewer(tbl_stream)

ref_arm = "DUMMY C"
comp_arm = c("DUMMY B", "DUMMY A")
combine_arm = FALSE
responders = c("CR", "PR")

ANL <- ARS \%>\% 
  filter(PARAMCD == "OVRSPI", ITTGEFL=='Y', ITTWTFL=='Y', ARM \%in\% c(ref_arm, comp_arm)) \%>\%
  select(c("USUBJID", "STUDYID", "ARM", "PARAMCD", "AVALC"))

# If want to include missing as non-responders
ANL$AVALC[ANL$AVALC==""] <- "NE"

arm1 <- factor(ANL$ARM)
ARM <- fct_relevel(arm1, ref_arm, comp_arm)

tbl <- response_table(
 response = ANL$AVALC,
 value.resp = c("CR", "PR"),
 value.nresp = setdiff(ANL$AVALC, c("CR", "PR")),
 arm = ARM
)

Viewer(tbl)

compare_rtables(tbl, tbl_stream, comp.attr = FALSE)


# different layout
ref_arm = c("DUMMY C", "DUMMY B")
comp_arm = c( "DUMMY A")
combine_arm = FALSE
responders = c("CR", "PR")

refname <- paste0(ref_arm, collapse = "/")
ARM <- do.call(fct_collapse, setNames(list(arm1, ref_arm), c("f", refname))) \%>\%
  fct_relevel(refname, comp_arm)

#style 2 
tbl2 <- response_table(response = ANL$AVALC, 
                       arm = ARM,
                       style = 2)
Viewer(tbl2)

}


'\%needs\%' <- teal.oncology:::'\%needs\%' # for debugging purpuses
}
\author{
Chendi Liao (liaoc10), \email{chendi.liao@roche.com}
}
