% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_response_table.R
\name{response_table_ADAM}
\alias{response_table_ADAM}
\title{Response Table with ADaM data structure}
\usage{
response_table_ADAM(ASL, ARS, paramcd = "BESRSPI", arm.var = "ARM", arm.ref,
  arm.comp, arm.comp.combine = TRUE, value.resp = c("CR", "PR"),
  value.nresp = c("SD", "NON CR/PD", "PD", "NE"), incl.missing = TRUE,
  style = 1)
}
\arguments{
\item{ASL}{dataset with following variables: USUBJID, STUDYID, the specified grouping variable}

\item{ARS}{ARS dataset containing the following variables: USUBJID, STUDYID, PARAMCD, AVALC}

\item{paramcd}{Name of overall response parameter}

\item{arm.var}{Name of variable with arm information}

\item{arm.ref}{Character vector, defining which arm(s) from the list of arms should be used as reference group}

\item{arm.comp}{Character vector, defining which arm(s) from the list of arms should be used as comparison group}

\item{arm.comp.combine}{Logical, \code{TRUE} if want all non-ref arms to be combined into one comparison group, 
\code{FALSE} if want each of the non-ref arms to be a separate comparison group}

\item{style}{Must be 1 or 2, \code{1} if only want to display rates summary for each response value category,
\code{2} if want to display rates, difference in rate and odds ratio for each response value category.}

\item{value.rsp}{Character vector, defining list of response values to be used as responders}

\item{value.nrsp}{Character vector, defining list of response values to be used as non-responders}

\item{incl_missing}{Logical, \code{TRUE} if missing values should be considered non-responders, 
\code{FALSE} if missing response should be removed from analysis}
}
\description{
Response Table with ADaM data structure
}
