% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{risk_diff}
\alias{risk_diff}
\title{Risk Difference}
\usage{
risk_diff(
  df = df,
  .var = .var,
  flag_variables = flag_variables,
  col_trt = col_trt,
  col_placebo = col_placebo,
  .N_col = .N_col,
  .indent_mods = .indent_mods
)
}
\description{
Risk Difference
}
