% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rreplace_format}
\alias{rreplace_format}
\alias{rreplace_format.rtable}
\title{Replace format of \code{rtables} cells}
\usage{
rreplace_format(
  x,
  row = NULL,
  col = NULL,
  old = FALSE,
  new = function(x, output) ""
)

\method{rreplace_format}{rtable}(
  x,
  row = NULL,
  col = NULL,
  old = FALSE,
  new = function(x, output) ""
)
}
\arguments{
\item{x}{A \code{rtable}.}

\item{row}{A vector of row indexes. Default `NULL` applies to all rows.}

\item{col}{A vector of columns indexes.
Default `NULL` applies to all columns.}

\item{old}{A targeted `rtable` format to be replaced. Default `FALSE`,
indicates a replacement independent from current format, otherwise
all `rtables` compatible format are accepted.}

\item{new}{The replacement format.}
}
\description{
A Generic function to replace format of given \code{rtables} cells. This is
convenient when, for instance, some data should rather be blanked during
a post-processing step.

Replace format of `rtables`
}
\examples{
library(random.cdisc.data)
library(tern)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ADLB <- radlb(cached = TRUE)
saved_labels <-  var_labels(ADLB)

# For illustration purpose, the example focuses on "Alanine Aminotransferase
# Measurement" starting from baseline, while excluding visit at week 1 for
# subjects who were prescribed a placebo.
ADLB <- subset(
  ADLB,
  PARAM == "Alanine Aminotransferase Measurement" &
    !(ARM == "B: Placebo" & AVISIT == "WEEK 1 DAY 8") & AVISIT != "SCREENING"
)

# Preprocessing.
ADLB_l <- split(
  ADLB, f = ifelse(ADLB$AVISIT == "BASELINE", "Baseline", "Follow-up")
)
ADLB_l <- lapply(ADLB_l, `var_labels<-`, saved_labels)
tbls1 <- lapply(X = ADLB_l, function(x) {

  tbl <- t_summary_by(
    x = compare_in_header(x[, c("AVAL", "CHG")]),
    row_by = x$AVISIT,
    col_by = x$ARM,
    col_N = table(ADSL$ARM),
    f_numeric = c("count_n", "mean_sd3", "median_t3", "range_t3")
  )
  return(tbl)

})

rreplace_format(tbls1$Baseline, col = 3)
rreplace_format(tbls1$Baseline, row = 2, old = NULL, new = "xx\%")
}
