% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rreplace_format.rtable}
\alias{rreplace_format.rtable}
\title{Replace format of `rtables`}
\usage{
\method{rreplace_format}{rtable}(
  x,
  row = NULL,
  col = NULL,
  old = FALSE,
  new = function(x, output) ""
)
}
\arguments{
\item{x}{A \code{rtable}.}

\item{row}{A vector of row indexes. Default `NULL` applies to all rows.}

\item{col}{A vector of columns indexes.
Default `NULL` applies to all columns.}

\item{old}{A targeted `rtable` format to be replaced. Default `FALSE`,
indicates a replacement independent from current format, otherwise
all `rtables` compatible format are accepted.}

\item{new}{The replacement format.}
}
\description{
Replace format of `rtables`
}
