% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{rsort_tree}
\alias{rsort_tree}
\title{It can also be used to remove children by not returning all indices}
\usage{
rsort_tree(node, f)
}
\arguments{
\item{node}{node to sort}

\item{f}{function(node) that returns the ordered list of new childrens this node will have,
then recurses into the children}
}
\value{
node
}
\description{
It can also be used to remove children by not returning all indices
}
\examples{
tree <- node("1", 1, list(
node("10", 10),
node("7", 7),
node("6", 6, list(
  node("2", 2),
  node("1", 1)
))
))
cat(displayable(tree))
sorted_tree <- rsort_tree(tree, function(node) order(vapply(node@children, function(child) child@content, numeric(1))))
cat(displayable(sorted_tree))
}
