% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{rsplit_to_tree}
\alias{rsplit_to_tree}
\title{Split list recursively according to by and return the associated tree}
\usage{
rsplit_to_tree(lst, by)
}
\arguments{
\item{lst}{list to split, \code{\link{split_subset}} will be applied to all list elements}

\item{by}{list of columns, each of which is a factor to recursively split by}
}
\value{
node object
}
\description{
Split list recursively according to by and return the associated tree
}
\examples{
tree <- rsplit_to_tree(1:5, list(factor(c("M", "M", "F", "F", "F")), factor(c("O", "Y", "Y", "Y", "Y"))))
summary(tree)
cat(displayable(tree))
}
