% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{rtables_prototypes}
\alias{rtables_prototypes}
\title{Prototype Functions to Work with \code{rtables}}
\description{
This summarizes the prototype functions that we currently use to work with
rtables. Note: These are really just prototypes. These are not final and
will likely later be part of rtables.
}
\seealso{
\code{\link[=to_string_matrix]{to_string_matrix()}}, \code{\link[=flatten_list]{flatten_list()}}, \code{\link[=list_length]{list_length()}},
\code{\link[=list_lengths_in_list]{list_lengths_in_list()}}, \code{\link[=labels_or_names]{labels_or_names()}}, \code{\link[=identical_without_attr]{identical_without_attr()}},
\code{\link[=format_wrap_df]{format_wrap_df()}}, \code{\link[=format_wrap_x]{format_wrap_x()}}
}
