% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{s_adj_proportion_diff}
\alias{s_adj_proportion_diff}
\title{Adjusted difference in proportions by CMH weights}
\usage{
s_adj_proportion_diff(x, grp, strat, conf_level = 0.95)
}
\arguments{
\item{x}{\code{logical} vector. \code{TRUE} represents a successful outcome.}

\item{grp}{a \code{factor} vector with 2 levels. Must be the same length as \code{x}.}

\item{strat}{factor with one level per stratum and same length as \code{x}.}

\item{conf_level}{confidence level of the interval.}
}
\value{
Named list with analysis summary:
\describe{
  \item{estimate}{CMH-weight adjusted proportion of successful outcomes in each group.}
  \item{estimate_ci}{Confidence interval for adjusted proportions in \code{estimate}.}
  \item{diff_est}{CMH-weight adjusted difference in proportion between groups.}
  \item{diff_ci}{Confidence interval for adjusted difference in \code{diff_est}.}
  \item{conf_level}{Confidence level used for \code{diff_ci} and \code{estimate_ci}.}
}
}
\description{
This function calculates a weighted difference in proportions for binary outcomes data.
This is defined as the difference in response rates between the experimental treatment
group and the control treatment group, adjusted for stratification factors by applying
Cochran-Mantel-Haenszel (CMH) weights. For the CMH chi-squared test,  use
\code{\link[stats]{mantelhaen.test}}.
}
\examples{
set.seed(1)
rsp <- sample(c(TRUE, FALSE), 100, TRUE)
trt <- sample(c("Placebo", "Treatment"), 100, TRUE)
trt <- factor(trt, levels = c("Placebo", "Treatment"))
strata_data <- data.frame(
  "f1" = sample(c("a", "b"), 100, TRUE),
  "f2" = sample(c("x", "y", "z"), 100, TRUE),
  stringsAsFactors = TRUE)
s_adj_proportion_diff(rsp, trt, interaction(strata_data))

}
\references{
Kim Y, Won S. Adjusted proportion difference and confidence interval in stratified randomized trials.
  [Internet]. PharmaSUG; 2013 [cited 14 May 2020]. (PharmaSUG 2013 Conference Proceedings). Report No.: SP04.
  Available from: https://www.pharmasug.org/proceedings/2013/SP/PharmaSUG-2013-SP04.pdf
}
\seealso{
\code{\link{t_el_proportion_diff}},
  \code{\link{s_proportion_diff}}, \code{\link{s_test_proportion_diff}}
}
