% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_ancova.R
\name{s_ancova}
\alias{s_ancova}
\title{Summary for analysis of covariance (ANCOVA).}
\usage{
s_ancova(formula, data, conf_level = 0.95)
}
\arguments{
\item{formula}{A \code{formula} corresponding to the investigated \code{\link[stats:lm]{linear model}}.
The left-hand side must include the dependent variable. The right-hand side must include the group variable
name wrapped in the special \code{arm()}.}

\item{data}{A \code{data.frame} which includes all the variables that are called in \code{formula}.}

\item{conf_level}{The confidence level of the resulting confidence intervals (default 0.95).}
}
\value{
Named list with analysis results as \code{summary_emm} objects, which are \code{data.frame} objects
with a special print method:
\describe{
\item{sum_fit}{Summary of the estimated marginal means in the groups.}
\item{sum_contrasts}{Summary of the estimated contrasts of the marginal means. Here the first level of the arm
variable is taken as the reference or control group, and all other groups are compared with this control group.
Specifically, the differences of the marginal means are estimated.
}
}
}
\description{
Summary for analysis of covariance (ANCOVA).
}
\examples{

# Estimate the adjusted means of sepal length in the different iris species,
# adjusting for covariates:
result <- s_ancova(
  formula = Sepal.Length ~ arm(Species) + Sepal.Width + Petal.Length + Petal.Width,
  data = iris
)
result
# We can see that the adjusted mean sepal length in the "setosa" species is significantly
# higher than in the "versicolor" and "virginica" species (p-values 0.003 in both differences).
}
