% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_coxph.R
\name{s_coxph}
\alias{s_coxph}
\title{CoxPH model results for unstratified and stratified analysis}
\usage{
s_coxph(
  formula,
  data,
  conf.int = 0.95,
  pval_method = c("log-rank", "wald", "likelihood"),
  ...
)
}
\arguments{
\item{formula}{a survival formula, the arm variable needs to be wrapped in
\code{\link{arm}}. The \code{\link[survival]{strata}} special will only be
used for the stratified analysis. If there is not
\code{\link[survival]{strata}} specification then the stratified analysis
is omitted.}

\item{data}{a \code{data.frame} with all the variable that are used in
\code{formula}}

\item{conf.int}{level for computation of the confidence intervals. If set to {FALSE} no confidence intervals are printed}

\item{pval_method}{the method used to calculate the p-value,
should be one of "wald", "log-rank", "likelihood", default is "log-rank".}

\item{...}{Other arguments will be passed to \code{\link[survival]{coxph}}}
}
\value{
a list of dataframes with unstratified and/or stratified analysis results
}
\description{
CoxPH model results for unstratified and stratified analysis
}
\examples{
library(random.cdisc.data)

ADSL <- cadsl
ADTTE <- cadtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")
\donttest{
formula = Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARM) + strata(RACE)
cox_res <-  s_coxph(formula, data = ADTTE_f, pval_method = "wald",  ties = "exact")
names(cox_res)
cox_res$stratified

formula2 = Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARM)
cox_res2 <- s_coxph(formula2, data = ADTTE_f, 0.8, pval_method = "likelihood",  ties = "breslow")
}

}
