% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{s_logistic_interaction}
\alias{s_logistic_interaction}
\title{Logistic Regression with Two-Way Interaction}
\usage{
s_logistic_interaction(glm_model, conf_level = 0.95, increments = NULL)
}
\arguments{
\item{glm_model}{({\code{\link{glm}}} model object)\cr
The model object can be all main effect model, and a model with one two-way interaction.}

\item{conf_level}{(\code{numeric} value)\cr
Confidence level for Wald odds ratio confidence interval.}

\item{increments}{(\code{named list})\cr
Used to specify numeric values of continuous variables in {\code{glm}} model {\code{formula}}
which interact with other variables. This is used to calculate the odds ratio when comparing
the other interaction variable effect. For example, for a model with ARM and AGE interaction,
{\code{increments = list(AGE = c(18, 65))}} will enable calculation of odds ratios of
comparison ARM vs. reference ARM at AGE = 18 and AGE = 65. If {\code{increments = NULL}}, then
default AGE value is ceiling of median.}
}
\description{
Summary of logistic regression with one two-way interaction term.
}
\examples{
library(random.cdisc.data)
library(dplyr)
library(purrr)

ADSL <- radsl(cached = TRUE)
ADSL <- ADSL \%>\% mutate(
 SEX = as.character(SEX),
 SEX = case_when(!SEX \%in\% c("F", "M") ~ "U",
               TRUE ~ SEX))
ADRS <- radrs(ADSL, seed = 2)
ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI") \%>\%
  mutate(Response = case_when(AVALC \%in\% c("PR", "CR") ~ 1,
                              TRUE ~ 0))
glm_model <- glm(
 formula = Response ~ ARM + AGE + SEX + ARM*SEX,
 data = ADRS_f,
 family = "binomial")
s_logistic_interaction(glm_model, conf_level = 0.90)
glm_model2 <- glm(
 formula = Response ~ ARM + AGE + ARM*AGE,
 data = ADRS_f,
 family = "binomial")
s_logistic_interaction(glm_model2, conf_level = 0.90)
glm_model3 <- glm(
 formula = Response ~ ARM + AGE + BMRKR1 + BMRKR1*AGE,
 data = ADRS_f,
 family = "binomial")
s_logistic_interaction(glm_model3 , conf_level = 0.90)
}
