% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_logistic.R
\name{s_logistic_single}
\alias{s_logistic_single}
\title{Summary}
\usage{
s_logistic_single(glm_model, conf_level = 0.95)
}
\arguments{
\item{glm_model}{({\code{\link{glm}}} model object)\cr
with all main effect model.}

\item{conf_level}{(\code{numeric} value)
confidence level for Wald's odds ratio confidence interval.}
}
\description{
Summary of logistic regression with no interaction term
}
\examples{
library(random.cdisc.data)
library(dplyr)
library(purrr)

ADSL <- radsl(cached = TRUE)
ADSL <- ADSL \%>\% mutate(
 SEX = as.character(SEX),
 SEX = case_when(!SEX \%in\% c("F", "M") ~ "U",
               TRUE ~ SEX))
ADRS <- radrs(ADSL, seed = 2)
ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI") \%>\%
  mutate(Response = case_when(AVALC \%in\% c("PR", "CR") ~ 1,
                              TRUE ~ 0))
glm_model <- glm(
 formula = Response ~ ARM + AGE + SEX,
 data = ADRS_f,
 family = "binomial")
s_logistic_single(glm_model, conf_level = 0.90)
}
