% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mmrm.R
\name{s_mmrm}
\alias{s_mmrm}
\title{Summary function for an MMRM analysis}
\usage{
s_mmrm(
  vars = list(response = "AVAL", covariates = c(), id = "USUBJID", arm = "ARM", visit =
    "AVISIT"),
  data,
  conf_level = 0.95,
  cor_struct = c("unstructured", "random-quadratic", "random-slope",
    "compound-symmetry"),
  weights_emmeans = "proportional"
)
}
\arguments{
\item{vars}{a \code{list} specifying the variables in the MMRM. The following elements need
to be included as character vectors and match corresponding columns in \code{data}:
\describe{
\item{response}{the response variable}
\item{covariates}{the additional covariate terms (might also include interactions)}
\item{id}{the subject ID variable}
\item{arm}{the treatment group variable (factor)}
\item{visit}{the visit variable (factor)}
}
Note that the main effects and interaction of `arm` and `visit` are by default included in the model.}

\item{data}{a \code{data.frame} with all the variables specified in
\code{vars}. Records with missing values in any independent variables
will be excluded.}

\item{conf_level}{confidence level. Must be number greater than 0 and less
than 1. (Default: 0.95)}

\item{cor_struct}{a string specifying the correlation structure, defaults to
\code{"unstructured"}. See the details.}

\item{weights_emmeans}{argument from \code{\link[emmeans]{emmeans}}, "proportional" by default.}
}
\value{
A list with MMRM results:
\describe{
  \item{fit}{The \code{lmerModLmerTest} object which was fitted to the data.}
  \item{lsmeans}{This is a list with data frames \code{estimate} and \code{contrast}.}
  \item{ref_level}{The reference level for the arm variable, which is always the first level.}
  \item{conf_level}{The confidence level which was used to construct the confidence intervals.}
}
}
\description{
Summary function for an MMRM analysis
}
\details{
Only Satterthwaite adjusted degrees of freedom (d.f.) are supported, because they
  match the results obtained in SAS (so far confirmed for unstructured covariance matrix).
  For the correlation structure (\code{cor_struct}), the user can choose among the following options, sorted
  in descending number of variance parameters:
  \describe{
  \item{unstructured}{Unstructured covariance matrix. This is the most flexible choice and default.
     If there are \code{T} visits, then \code{T * (T-1) / 2} variance parameters are used.}
  \item{random-quadratic}{Random quadratic spline for the random effects of the time variable.
     6 variance parameters are used.}
  \item{random-slope}{Random slope for the random effects of the time variable. 3 variance parameters are used.}
  \item{compound-symmetry}{Constant correlation between visits. 2 variance parameters are used.}
  }
}
\examples{
library(random.cdisc.data)
library(dplyr)

adsl <- radsl(cached = TRUE)
adqs <- radqs(cached = TRUE)
adqs_f <- adqs \%>\%
  dplyr::filter(PARAMCD == "FKSI-FWB" & !AVISIT \%in\% c("BASELINE")) \%>\%
  droplevels() \%>\%
  dplyr::mutate(ARMCD = factor(ARMCD, levels = c("ARM B", "ARM A", "ARM C"))) \%>\%
  dplyr::mutate(
    AVISITN = rank(AVISITN) \%>\%
    as.factor() \%>\%
    as.numeric() \%>\%
    as.factor()
  )

mmrm_results <- s_mmrm(
  vars = list(
    response = "AVAL",
    covariates = c("STRATA1", "BMRKR2"),
    id = "USUBJID",
    arm = "ARMCD",
    visit = "AVISIT"
  ),
  data = adqs_f,
  cor_struct = "unstructured",
  weights_emmeans = "equal"
)
}
