% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{s_proportion_diff}
\alias{s_proportion_diff}
\title{Summary of difference between two proportions}
\usage{
s_proportion_diff(
  x,
  grp,
  strat = NULL,
  conf_level = 0.95,
  diff_ci_method = c("wald", "waldcc", "cmh", "anderson-hauck", "ha", "newcombe")
)
}
\arguments{
\item{x}{\code{logical} vector. \code{TRUE} represents a successful outcome.}

\item{grp}{a \code{factor} vector with 2 levels. Must be the same length as \code{x}.}

\item{strat}{factor with one level per stratum and same length as \code{x}. Only required
for stratified analysis when method is "cmh".}

\item{conf_level}{confidence level of the interval.}

\item{diff_ci_method}{one of (\code{"wald"}, \code{"waldcc"}, \code{"cmh"}, \code{"anderson-hauck"} or \code{"ha"},
\code{"newcombe"})\cr
Specifies the method used to construct the confidence interval for the difference
in proportion of outcomes.}
}
\value{
Named list with analysis summary.
\describe{
\item{diff}{Difference in proportion of successful outcomes between groups.}
\item{diff_ci}{Confidence interval for \code{diff}.}
\item{label_ci}{Label for confidence interval in \code{diff_ci}.}
}
}
\description{
This function is a wrapper to calculate the difference between
two proportions or weighed difference if stratification variables are
used. In addition provides various options for types of confidence intervals.
}
\details{
The following options are supported for difference of proportions confidence intervals:
\itemize{
\item{Option \code{wald} calls \code{\link[stats]{prop.test}} with \code{correct=FALSE}}
\item{Option \code{waldcc} includes continuity correction for Wald interval and
calls \code{\link[stats]{prop.test}} with \code{correct=TRUE}}
\item{Option \code{cmh} derives a CI for CMH-weighted difference of proportions and
calls \code{\link{s_adj_proportion_diff}}.}
\item{Options \code{anderson-hauck} or \code{ha} derives the Anderson-Hauck confidence interval.}
\item{Option \code{newcombe} derives the Newcombe confidence interval, which is based on the Wilson score
confidence interval for a single binomial proportion.}
}
}
\examples{
set.seed(1)
rsp <- sample(c(TRUE, FALSE), 100, TRUE)
trt <- sample(c("Placebo", "Treatment"), 100, TRUE)
trt <- factor(trt, levels = c("Placebo", "Treatment"))
strata_data <- data.frame(
  "f1" = sample(c("a", "b"), 100, TRUE),
  "f2" = sample(c("x", "y", "z"), 100, TRUE),
  stringsAsFactors = TRUE)

s_proportion_diff(rsp, trt, diff_ci_method = "wald")
s_proportion_diff(rsp, trt, diff_ci_method = "waldcc")
s_proportion_diff(rsp, trt, strat = interaction(strata_data), diff_ci_method = "cmh")

# case without responders
s_proportion_diff(rep(FALSE, length(trt)), trt, diff_ci_method = "wald")
}
\seealso{
\code{\link{s_adj_proportion_diff}}, \code{\link{s_proportion}},
\code{\link{s_test_proportion_diff}}
}
