% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{s_test_proportion_diff}
\alias{s_test_proportion_diff}
\title{Test for difference between two proportions}
\usage{
s_test_proportion_diff(
  x,
  grp,
  strat = NULL,
  test = c("chisq", "cmh", "fisher", "schouten")
)
}
\arguments{
\item{x}{\code{logical} vector. \code{TRUE} represents a successful outcome.}

\item{grp}{a \code{factor} vector with 2 levels. Must be the same length as \code{x}.}

\item{strat}{a \code{factor} vector with at least 2 levels identifying the strata. Only required if
test is \code{"cmh"}.}

\item{test}{one of (\code{"chisq", "cmh", "fisher", "schouten"})\cr
Specifies the test used to calculate the p-value.}
}
\value{
Named list with analysis summary.
\describe{
  \item{p_value}{P-value for testing null hypothesis that proportions in two groups are the same.}
  \item{test_name}{Label for test used to calculate \code{p_value}.}
}
}
\description{
This function is a wrapper to extract the p-value from various tests for
difference of two proportions.
}
\details{
The following options are supported for testing difference of proportions:
\itemize{
  \item{Option \code{chisq} performs Chi-Squared test. Internally calls \code{\link[stats]{prop.test}}.}
  \item{Option \code{cmh} performs stratified Cochran-Mantel-Haenszel test.
  Internally calls \code{\link[stats]{mantelhaen.test}}.}
  \item{Option \code{fisher} performs the Fisher's exact test.}
  \item{Option \code{schouten} performs the Chi-Squared test with Schouten correction.}
}
}
\examples{
set.seed(1)
x <- sample(c(TRUE, FALSE), 100, TRUE)
grp <- factor(rep(c("A", "B"), each = 50))
strat <- factor(rep(c("V", "W", "X", "Y", "Z"), each = 20))
s_test_proportion_diff(x, grp, test = "chisq")
s_test_proportion_diff(x, grp, strat, test = "cmh")

\dontrun{
# warning occurs because small size in strata
s_test_proportion_diff(
  x = c(TRUE, FALSE, FALSE, TRUE),
  grp = factor(c("A", "B", "A", "B")),
  strat = factor(c("X", "Y", "Y", "X")),
  test = "cmh"
)
}
}
\seealso{
\code{\link{t_el_test_proportion_diff}}
  \code{\link{s_proportion}}, \code{\link{s_proportion_diff}}
}
