% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_join.R
\name{safe_join}
\alias{safe_join}
\title{Safer join for tern}
\usage{
safe_join(x, y, by = NULL, method = full_join, ...)
}
\arguments{
\item{x}{(\code{data.frame} or \code{\link[dplyr]{tbl}}) left hand side of the join}

\item{y}{(\code{data.frame} or \code{\link[dplyr]{tbl}}) right hand side of the join}

\item{by}{(\code{character}) Key columns to use for joining. For a safe join these
are obligatory to guarantee the ability to merge the datasets.}

\item{method}{(\code{function}) \code{dplyr::\link[dplyr]{join}} function to join
the two data sets.}

\item{...}{Additional parameters handed over to the function called in \code{method}}
}
\description{
This is a wrapper function around \code{dplyr::\link[dplyr]{join}}. It gives an error
in case two \code{data.frames} cannot be joined due to wrong keys or dimension mismatches
of equally named columns inside the \code{data.frames}. Additionally columns that are
equal inside the \code{data.frames} do not get merged twice. The join is applied without
duplicates.
}
\references{
\link[dplyr]{join}
}
