% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_rtables.R
\name{simple_by}
\alias{simple_by}
\title{Non-recursive by object}
\usage{
simple_by(x)
}
\arguments{
\item{x}{object to wrap it around}
}
\description{
This can be used both for row and column grouping.
}
\details{
Not called "by" because this function already exists in baseR
}
