% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot.R
\name{split.anl}
\alias{split.anl}
\title{Subgroup ATE by the selected ASL variable categories;}
\usage{
\method{split}{anl}(x, data)
}
\description{
Subgroup ATE by the selected ASL variable categories;
}
