% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_events_term_id.R
\name{split_subset}
\alias{split_subset}
\title{split when lst is not a data.frame, but a list of objects}
\usage{
split_subset(lst, f)
}
\description{
split when lst is not a data.frame, but a list of objects
}
\examples{
split_subset(
  list(
    x1 = factor(c(1,2,3,4,5,6)),
    x2 = data.frame(x = 7:12, y = 13:18)
  ),
  factor(c("a", "a", "b", "b", "b", "b"))
)
# todo: extend to non-disjoint col_by matrix
}
