% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_grob.R
\name{split_text_grob}
\alias{split_text_grob}
\title{Split Text According To Available Text Width}
\usage{
split_text_grob(
  text,
  x = unit(0.5, "npc"),
  y = unit(0.5, "npc"),
  width = unit(1, "npc"),
  just = "centre",
  hjust = NULL,
  vjust = NULL,
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  vp = NULL
)
}
\arguments{
\item{text}{character string}

\item{x}{A numeric vector or unit object specifying x-values.}

\item{y}{A numeric vector or unit object specifying y-values.}

\item{width}{a unit object specifying max width of text}

\item{just}{The justification of the text
    relative to its (x, y) location.  If there are two values, the first
    value specifies horizontal justification and the second value specifies
    vertical justification.  Possible string values are: \code{"left"},
    \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"},
    and \code{"top"}.  For numeric values, 0 means left (bottom) alignment
    and 1 means right (top) alignment.
  }

\item{hjust}{A numeric vector specifying horizontal justification.
    If specified, overrides the \code{just} setting.}

\item{vjust}{A numeric vector specifying vertical justification.
    If specified, overrides the \code{just} setting.}

\item{default.units}{A string indicating the default units to use
    if \code{x} or \code{y}
    are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{gpar}, typically the output
    from a call to the function \code{gpar}.  This is basically
    a list of graphical parameter settings.}

\item{vp}{A Grid viewport object (or NULL).}

\item{...}{passed on to \code{\link{grob}}}
}
\description{
Dynamically wrap text
}
\details{
This code is taken from R Graphics by \code{Paul Murell}, 2nd edition
}
\examples{
sg <- split_text_grob(text = paste(
  "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum vitae",
  "dapibus dolor, ac mattis erat. Nunc metus lectus, imperdiet ut enim eu,",
  "commodo scelerisque urna. Vestibulum facilisis metus vel nibh tempor, sed",
  "elementum sem tempus. Morbi quis arcu condimentum, maximus lorem id,",
  "tristique ante. Nullam a nunc dui. Fusce quis lacus nec ante dignissim",
  "faucibus nec vitae tellus. Suspendisse mollis et sapien eu ornare. Vestibulum",
  "placerat neque nec justo efficitur, ornare varius nulla imperdiet. Nunc justo",
  "sapien, vestibulum eget efficitur eget, porttitor id ante. Nulla tempor",
  "luctus massa id elementum. Praesent dictum, neque vitae vestibulum malesuada,",
  "nunc nisi blandit lacus, sit amet tristique odio dui sit amet velit."
))

grobHeight(sg)

grid.newpage()
pushViewport(plotViewport())
grid.rect()
grid.draw(sg)

grid.rect(
  height = grobHeight(sg), width = unit(1, "cm"), gp = gpar(fill = "red")
)

# stack split_text_grob
grid.newpage()
pushViewport(plotViewport())
grid.rect()
grid.draw(split_text_grob(
  c("Hello, this is a test", "and yet another test"),
  just = c("left", "top"), x = 0, y = 1
))
}
