% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stack_rtables}
\alias{stack_rtables}
\title{Stack rtables with rbind and add empy rows between tables}
\usage{
stack_rtables(..., nrow_pad = 1)
}
\arguments{
\item{...}{rtbale objects}
}
\description{
Stack rtables with rbind and add empy rows between tables
}
