% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helper.R
\name{stat_median_ci}
\alias{stat_median_ci}
\title{Median CI for \code{ggplot2}}
\usage{
stat_median_ci(x, conf_level = 0.95, na.rm = TRUE)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{na.rm}{(\code{flag})\cr whether \code{NA} values should be removed from \code{x} prior to analysis.}
}
\description{
Convenient function for the addition of the median confidence intervals as
error bars.
}
\details{
The function was adapted from \verb{DescTools/versions/0.99.35/source}
}
\examples{
require(ggplot2)
d <- ggplot(mtcars, aes(cyl, mpg)) + geom_point()
d + stat_summary(
  fun.data = stat_median_ci, geom = "errorbar",
  aes(group = 1, linetype = "median")
)
}
