% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_risk_diff_ci}
\alias{stat_risk_diff_ci}
\title{Risk Difference and Confidence Interval}
\usage{
stat_risk_diff_ci(
  frac_x,
  frac_y,
  N_x,
  N_y,
  conf_level = 0.95,
  row_names = NULL,
  pct = TRUE
)
}
\arguments{
\item{frac_x}{(\code{list})\cr list of cumulative incidences (proportions) in arm X.}

\item{frac_y}{(\code{list})\cr list of cumulative incidences (proportions) in arm Y. Must be of equal length to \code{frac_x}.}

\item{N_x}{(\code{numeric})\cr total number of records in arm X.}

\item{N_y}{(\code{numeric})\cr total number of records in arm Y.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{row_names}{(\code{character})\cr names of each variable/level corresponding to pair of proportions in
\code{frac_x} and \code{frac_y}. Must be of equal length to \code{frac_x} and \code{frac_y}.}

\item{pct}{(\code{logical})\cr should output be returned as percentage instead of a fraction. Defaults to \code{TRUE}.}
}
\value{
list of risk differences and CIs corresponding to each pair of proportions in \code{frac_x} and \code{frac_y}.
Each list element consists of 3 statistics: risk difference, CI lower bound, and CI upper bound.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Convenient function for calculating the risk difference and confidence interval between arm X and arm Y.
Risk difference is calculated by subtracting cumulative incidence in arm Y from cumulative incidence in arm X.
}
\examples{
stat_risk_diff_ci(frac_x = list(0.375), frac_y = list(0.01), N_x = 5, N_y = 5, row_names = c("x"), conf_level = 0.9)
stat_risk_diff_ci(frac_x = list(0.5, 0.75, 1), frac_y = list(0.25, 0.05, 0.5), N_x = 10, N_y = 20, pct = FALSE)
}
