% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{study_arm}
\alias{study_arm}
\title{Indicate study arm variable in formula}
\usage{
study_arm(x)
}
\arguments{
\item{x}{arm information}
}
\value{
\code{x}
}
\description{
We use \code{study_arm} to indicate the study arm variable in \code{tern} formulas.
}
\keyword{internal}
