% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{[[,node-method}
\alias{[[,node-method}
\title{Allows to access nodes recursively, each element in index can either be a number or the name of the node}
\usage{
\S4method{[[}{node}(x, i, j, ..., return_content = TRUE)
}
\arguments{
\item{x}{node object}

\item{i}{index (possibly missing)}

\item{j}{index (possibly missing)}

\item{...}{deeper level indexes}

\item{return_content}{whether to return content of node instead of node}
}
\description{
Ensure that children are not named identically. If cannot be avoided, access by index
}
\details{
Can use summary to get index to get to certain element
}
\examples{
n2 <- node(name = "D", content = c(1:3), children = list(
  node(name = "A", content = array(c(1:6), dim = c(2,3)), children = list()),
  node(name = "B", content = array(c(1:6), dim = c(2,3)), children = list()),
  node(name = "C", content = array(c(1:6), dim = c(2,3)), children = list())
))
summary(n2)
summary(n2[[return_content = FALSE]])
n2[[]]
n2[[return_content = TRUE]]
n2[[1]]
n2[["A"]]
summary(n2[["A", return_content = FALSE]])

n3 <- node(name = "D", content = c(1:12), children = list(
node(name = "A", content = c(1:6), children = list(
  node(name = "A1", content = c(1:3), children = list()),
  node(name = "A2", content = c(4:6), children = list())
)),
node(name = "B", content = c(7:12), children = list(
  node(name = "B1", content = c(7:9), children = list()),
  node(name = "B2", content = c(10:12), children = list())
))
))
cat(displayable(n3))
cat(displayable(n3[["A"]]))
cat(displayable(n3[["A", return_content = FALSE]]))
cat(displayable(n3[["A", "A1"]]))
cat(displayable(n3[["A", "A1", return_content = FALSE]]))
}
