% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_functions.R
\name{summarize_functions}
\alias{summarize_functions}
\title{Summarize Functions}
\description{
These functions are wrappers for \code{\link[rtables:summarize_row_groups]{rtables::summarize_row_groups()}}, applying custom \code{tern} content functions
to return the corresponding row summary within a given table layout.
\itemize{
\item \code{\link[=logistic_summary_by_flag]{logistic_summary_by_flag()}}
\item \code{\link[=summarize_num_patients]{summarize_num_patients()}}
\item \code{\link[=summarize_occurrences_by_grade]{summarize_occurrences_by_grade()}}
\item \code{\link[=summarize_patients_events_in_cols]{summarize_patients_events_in_cols()}}
\item \code{\link[=summarize_patients_exposure_in_cols]{summarize_patients_exposure_in_cols()}}
}
}
\seealso{
\link{analyze_functions} for functions which are wrappers for \code{\link[rtables:analyze]{rtables::analyze()}}, and\cr
\code{\link[=analyze_vars_in_cols]{analyze_vars_in_cols()}} which is a wrapper for \code{\link[rtables:analyze_colvars]{rtables::analyze_colvars()}}.
}
