% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_glm_count.R
\name{summarize_glm_count}
\alias{summarize_glm_count}
\alias{s_glm_count}
\alias{a_glm_count}
\title{Summarize Poisson negative binomial regression}
\usage{
summarize_glm_count(
  lyt,
  vars,
  variables,
  distribution,
  conf_level,
  rate_mean_method = c("emmeans", "ppmeans")[1],
  weights = stats::weights,
  scale = 1,
  var_labels,
  na_str = default_na_str(),
  nested = TRUE,
  ...,
  show_labels = "visible",
  table_names = vars,
  .stats = c("n", "rate", "rate_ci", "rate_ratio", "rate_ratio_ci", "pval"),
  .stat_names = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = list(rate_ci = 1L, rate_ratio_ci = 1L, pval = 1L)
)

s_glm_count(
  df,
  .var,
  .df_row,
  .ref_group,
  .in_ref_col,
  variables,
  distribution,
  conf_level,
  rate_mean_method,
  weights,
  scale = 1,
  ...
)

a_glm_count(
  df,
  ...,
  .stats = NULL,
  .stat_names = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{lyt}{(\code{PreDataTableLayouts})\cr layout that analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{variables}{(named \code{list} of \code{string})\cr list of additional analysis variables, with
expected elements:
\itemize{
\item \code{arm} (\code{string})\cr group variable, for which the covariate adjusted means of multiple
groups will be summarized. Specifically, the first level of \code{arm} variable is taken as the
reference group.
\item \code{covariates} (\code{character})\cr a vector that can contain single variable names (such as
\code{"X1"}), and/or interaction terms indicated by \code{"X1 * X2"}.
\item \code{offset} (\code{numeric})\cr a numeric vector or scalar adding an offset.
}}

\item{distribution}{(\code{character})\cr a character value specifying the distribution
used in the regression (Poisson, Quasi-Poisson, negative binomial).}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{rate_mean_method}{(\code{character(1)})\cr method used to estimate the mean odds ratio. Defaults to \code{emmeans}.
see details for more information.}

\item{weights}{(\code{character})\cr a character vector specifying weights used
in averaging predictions. Number of weights must equal the number of levels included in the covariates.
Weights option passed to \code{\link[emmeans:emmeans]{emmeans::emmeans()}}.}

\item{scale}{(\code{numeric(1)})\cr linear scaling factor for rate and confidence intervals. Defaults to \code{1}.}

\item{var_labels}{(\code{character})\cr variable labels.}

\item{na_str}{(\code{string})\cr string used to replace all \code{NA} or empty values in the output.}

\item{nested}{(\code{flag})\cr whether this layout instruction should be applied within the existing layout structure _if
possible (\code{TRUE}, the default) or as a new top-level element (\code{FALSE}). Ignored if it would nest a split.
underneath analyses, which is not allowed.}

\item{...}{additional arguments for the lower level functions.}

\item{show_labels}{(\code{string})\cr label visibility: one of "default", "visible" and "hidden".}

\item{table_names}{(\code{character})\cr this can be customized in the case that the same \code{vars} are analyzed multiple
times, to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.

Options are: \verb{'n', 'rate', 'rate_ci', 'rate_ratio', 'rate_ratio_ci', 'pval'}}

\item{.stat_names}{(\code{character})\cr names of the statistics that are passed directly to name single statistics
(\code{.stats}). This option is visible when producing \code{\link[rtables:as_result_df]{rtables::as_result_df()}} with \code{make_ard = TRUE}.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics. See Details in \code{analyze_vars} for more
information on the \code{"auto"} setting.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels. Defaults to 0, which corresponds to the
unmodified default behavior. Can be negative.}

\item{df}{(\code{data.frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.df_row}{(\code{data.frame})\cr dataset that includes all the variables that are called
in \code{.var} and \code{variables}.}

\item{.ref_group}{(\code{data.frame} or \code{vector})\cr the data corresponding to the reference group.}

\item{.in_ref_col}{(\code{flag})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}
}
\value{
\itemize{
\item \code{summarize_glm_count()} returns a layout object suitable for passing to further layouting functions,
or to \code{\link[rtables:build_table]{rtables::build_table()}}. Adding this function to an \code{rtable} layout will add formatted rows containing
the statistics from \code{s_glm_count()} to the table layout.
}

\itemize{
\item \code{s_glm_count()} returns a named \code{list} of 5 statistics:
\itemize{
\item \code{n}: Count of complete sample size for the group.
\item \code{rate}: Estimated event rate per follow-up time.
\item \code{rate_ci}: Confidence level for estimated rate per follow-up time.
\item \code{rate_ratio}: Ratio of event rates in each treatment arm to the reference arm.
\item \code{rate_ratio_ci}: Confidence level for the rate ratio.
\item \code{pval}: p-value.
}
}

\itemize{
\item \code{a_glm_count()} returns the corresponding list with formatted \code{\link[rtables:CellValue]{rtables::CellValue()}}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Summarize results of a Poisson negative binomial regression.
This can be used to analyze count and/or frequency data using a linear model.
It is specifically useful for analyzing count data (using the Poisson or Negative
Binomial distribution) that is result of a generalized linear model of one (e.g. arm) or more
covariates.
}
\details{
\code{summarize_glm_count()} uses \code{s_glm_count()} to calculate the statistics for the table. This
analysis function uses \code{\link[=h_glm_count]{h_glm_count()}} to estimate the GLM with \code{\link[stats:glm]{stats::glm()}} for Poisson and Quasi-Poisson
distributions or \code{\link[MASS:glm.nb]{MASS::glm.nb()}} for Negative Binomial distribution. All methods assume a
logarithmic link function.

At this point, rates and confidence intervals are estimated from the model using
either \code{\link[emmeans:emmeans]{emmeans::emmeans()}} when \code{rate_mean_method = "emmeans"} or \code{\link[=h_ppmeans]{h_ppmeans()}}
when \code{rate_mean_method = "ppmeans"}.

If a reference group is specified while building the table with \code{split_cols_by(ref_group)},
no rate ratio or \code{p-value} are calculated. Otherwise, we use \code{\link[emmeans:contrast]{emmeans::contrast()}} to
calculate the rate ratio and \code{p-value} for the reference group. Values are always estimated
with \code{method = "trt.vs.ctrl"} and \code{ref} equal to the first \code{arm} value.
}
\section{Functions}{
\itemize{
\item \code{summarize_glm_count()}: Layout-creating function which can take statistics function arguments
and additional format arguments. This function is a wrapper for \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{s_glm_count()}: Statistics function that produces a named list of results
of the investigated Poisson model.

\item \code{a_glm_count()}: Formatted analysis function which is used as \code{afun} in \code{summarize_glm_count()}.

}}
\examples{
library(dplyr)

anl <- tern_ex_adtte |> filter(PARAMCD == "TNE")
anl$AVAL_f <- as.factor(anl$AVAL)

lyt <- basic_table() |>
  split_cols_by("ARM", ref_group = "B: Placebo") |>
  add_colcounts() |>
  analyze_vars(
    "AVAL_f",
    var_labels = "Number of exacerbations per patient",
    .stats = c("count_fraction"),
    .formats = c("count_fraction" = "xx (xx.xx\%)"),
    .labels = c("Number of exacerbations per patient")
  ) |>
  summarize_glm_count(
    vars = "AVAL",
    variables = list(arm = "ARM", offset = "lgTMATRSK", covariates = NULL),
    conf_level = 0.95,
    distribution = "poisson",
    rate_mean_method = "emmeans",
    var_labels = "Adjusted (P) exacerbation rate (per year)",
    table_names = "adjP",
    .stats = c("rate"),
    .labels = c(rate = "Rate")
  ) |>
  summarize_glm_count(
    vars = "AVAL",
    variables = list(arm = "ARM", offset = "lgTMATRSK", covariates = c("REGION1")),
    conf_level = 0.95,
    distribution = "quasipoisson",
    rate_mean_method = "ppmeans",
    var_labels = "Adjusted (QP) exacerbation rate (per year)",
    table_names = "adjQP",
    .stats = c("rate", "rate_ci", "rate_ratio", "rate_ratio_ci", "pval"),
    .labels = c(
      rate = "Rate", rate_ci = "Rate CI", rate_ratio = "Rate Ratio",
      rate_ratio_ci = "Rate Ratio CI", pval = "p value"
    )
  ) |>
  summarize_glm_count(
    vars = "AVAL",
    variables = list(arm = "ARM", offset = "lgTMATRSK", covariates = c("REGION1")),
    conf_level = 0.95,
    distribution = "negbin",
    rate_mean_method = "emmeans",
    var_labels = "Adjusted (NB) exacerbation rate (per year)",
    table_names = "adjNB",
    .stats = c("rate", "rate_ci", "rate_ratio", "rate_ratio_ci", "pval"),
    .labels = c(
      rate = "Rate", rate_ci = "Rate CI", rate_ratio = "Rate Ratio",
      rate_ratio_ci = "Rate Ratio CI", pval = "p value"
    )
  )

build_table(lyt = lyt, df = anl)

}
\keyword{internal}
