% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pk.R
\name{summarize_pk}
\alias{summarize_pk}
\alias{s_summary_pk}
\alias{summarize_pk_in_cols}
\title{Summary of PK datasets}
\usage{
s_summary_pk(x, labelstr = "", custom_label = NULL, ...)

summarize_pk_in_cols(
  lyt,
  var,
  ...,
  .stats = c("n", "mean", "sd", "cv", "geom_mean", "geom_cv", "median", "min", "max"),
  .labels = c(n = "n", mean = "Mean", sd = "SD", cv = "CV \% Mean", geom_mean =
    "Geometric Mean", geom_cv = "CV \% Geometric Mean", median = "Median", min =
    "Minimum", max = "Maximum"),
  .indent_mods = NULL,
  col_split = TRUE
)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions).}

\item{custom_label}{(\code{string} or \code{NULL})\cr if provided and \code{labelstr} is empty then this will
be used as the row label.}

\item{...}{additional arguments for the lower level functions.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{var}{(\code{string})\cr single variable name for the primary analysis variable.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}

\item{col_split}{(\code{flag})\cr whether the columns should be split.}
}
\value{
A named list of 9 statistics:
\itemize{
\item \code{n}: count of complete sample size for the group.
\item \code{mean}: group data mean.
\item \code{sd}: standard deviation of the group data.
\item \code{cv}: coefficient of variation of the group.
\item \code{min}: minimum value among the group.
\item \code{max}: maximum value among the group.
\item \code{median}: median of the group data.
\item \code{geom_mean}: geometric mean of the group.
\item \code{geom_cv}: geometric coefficient of variation of the group.
}
}
\description{
These functions can be used to produce summary tables for PK datasets.
}
\section{Functions}{
\itemize{
\item \code{s_summary_pk}: Statistics function that produces a named list of statistics
to include as columns in the PK table.

\item \code{summarize_pk_in_cols}: Layout creating function which can be used for creating
summary tables for PK data sets.
}}

\examples{

library(scda)
ADSL <- scda::synthetic_cdisc_data("latest")$adsl
s_summary_pk(ADSL$AGE, custom_label = "stats")


lyt <- basic_table() \%>\%
  split_rows_by(var = "ARM") \%>\%
  split_rows_by(var = "SEX") \%>\%
  summarize_pk_in_cols(var = "AGE", col_split = TRUE)
  result <- build_table(lyt, df = ADSL)
  result

lyt <- basic_table() \%>\%
  summarize_pk_in_cols(var = "AGE", col_split = TRUE)
  result <- build_table(lyt, df = ADSL)
  result

}
