% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pk.R
\name{summarize_pk_in_cols}
\alias{summarize_pk_in_cols}
\title{Title}
\usage{
summarize_pk_in_cols(
  lyt,
  var,
  ...,
  .stats = c("n", "mean", "sd", "cv", "median"),
  .labels = c(n = "n", mean = "mean", sd = "sd", cv = "CV \% Mean", median = "Median"),
  .indent_mods = NULL,
  col_split = TRUE
)
}
\arguments{
\item{col_split}{}
}
\value{

}
\description{
Title
}
\examples{

library(random.cdisc.data)
adsl <- radsl(100)

lyt <- basic_table() \%>\%
  split_rows_by(var = "ARM") \%>\%
  split_rows_by(var = "SEX") \%>\%
  summarize_pk_in_cols(var = "AGE", col_split = TRUE)
  result <- build_table(lyt, df = adsl)
  result
}
