% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variables_in_cols.R
\name{summarize_variables_in_columns}
\alias{summarize_variables_in_columns}
\alias{summarize_vars_in_cols}
\title{Summary numeric variables in columns}
\usage{
summarize_vars_in_cols(
  lyt,
  vars,
  ...,
  .stats = c("n", "mean", "sd", "se", "cv", "geom_cv"),
  .labels = c(n = "n", mean = "Mean", sd = "SD", se = "SE", cv = "CV (\%)", geom_cv =
    "CV \% Geometric Mean"),
  labelstr = " ",
  .formats = NULL,
  .indent_mods = NULL,
  na_str = NULL
)
}
\arguments{
\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{...}{additional arguments for the lower level functions.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions).}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}

\item{na_str}{(\code{string})\cr used to replace all \code{NA} with custom \code{string}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function can be used to produce summary tables for PK datasets where
the relevant statistic is on the columns instead of on the rows.
}
\section{Functions}{
\itemize{
\item \code{summarize_vars_in_cols()}: Layout creating
function which can be used for creating summary tables in columns, primarily used for PK data sets.

}}
\examples{
library(scda)
library(dplyr)
adpp <- synthetic_cdisc_dataset("latest", "adpp") \%>\% h_pkparam_sort()

lyt <- basic_table() \%>\%
  split_rows_by(var = "ARM", label_pos = "topleft") \%>\%
  split_rows_by(var = "SEX", label_pos = "topleft") \%>\%
  summarize_vars_in_cols(vars = "AGE")
result <- build_table(lyt = lyt, df = adpp)
result

# By selecting just some statistics and ad-hoc labels
lyt <- basic_table() \%>\%
  split_rows_by(var = "ARM", label_pos = "topleft") \%>\%
  split_rows_by(var = "SEX", label_pos = "topleft") \%>\%
  summarize_vars_in_cols(
    vars = "AGE",
    .stats = c("n", "cv", "geom_mean", "mean_ci", "median", "min", "max"),
    .labels = c(
      n = "myN",
      cv = "myCV",
      geom_mean = "myGeomMean",
      mean_ci = "Mean (95\%CI)",
      median = "Median",
      min = "Minimum",
      max = "Maximum"
    )
  )
result <- build_table(lyt = lyt, df = adpp)
result

lyt <- basic_table() \%>\%
  summarize_vars_in_cols(
    vars = "AGE",
    labelstr = "some custom label"
  )
result <- build_table(lyt, df = adpp)
result

# PKPT03
lyt <- basic_table() \%>\%
  split_rows_by(var = "TLG_DISPLAY", split_label = "PK Parameter", label_pos = "topleft") \%>\%
  summarize_vars_in_cols(
    vars = "AVAL",
    .stats = c("n", "mean", "sd", "cv", "geom_mean", "geom_cv", "median", "min", "max"),
    .labels = c(
      n = "n",
      mean = "Mean",
      sd = "SD",
      cv = "CV (\%)",
      geom_mean = "Geometric Mean",
      geom_cv = "CV \% Geometric Mean",
      median = "Median",
      min = "Minimum",
      max = "Maximum"
    )
  )
result <- build_table(lyt, df = adpp)
result
}
\seealso{
\link{summarize_vars}.
}
