% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{summary.data_tree}
\alias{summary.data_tree}
\title{Summarize a data_tree object}
\usage{
\method{summary}{data_tree}(object, name = "", indent = 0,
  first = TRUE, ...)
}
\arguments{
\item{object}{a \code{\link{table_tree}} object}

\item{name}{list element name of table}

\item{indent}{indentation of table}

\item{first}{logical, whether the nested table should be indented}

\item{...}{arguments which are not used}
}
\value{
an \code{rtable} object with the columns \code{class} and \code{dimension}
}
\description{
Summarize a data_tree object
}
