% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_trees.R
\name{summary.table_tree}
\alias{summary.table_tree}
\title{Summarize a table_tree object}
\usage{
\method{summary}{table_tree}(object, name = "", indent = 0,
  first = TRUE, ...)
}
\arguments{
\item{object}{a \code{\link{table_tree}} object}

\item{name}{list element name of table}

\item{indent}{indentation of table}

\item{first}{logical, whether the nested table should be indented}

\item{...}{arguments which are not used}
}
\value{
an \code{rtable} object with the columns \code{class} and \code{dimension}
}
\description{
Summarize a table_tree object
}
\examples{

tbl <- rtable(header = 1:3, rrowl("row", letters[1:3]))

nl1 <- list(tbl1 = tbl, tbl2 = tbl, tbl3 = NULL)
summary(table_tree(nl1))

nl2 <- list(A = nl1, B = nl1)
summary(table_tree(nl2))

nl3 <- list(X = nl2, Y = nl2, Z = nl2)
summary(table_tree(nl3))

}
