% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_variables.R
\name{summary_custom}
\alias{summary_custom}
\title{Set Defaults Settings for Summary Statistics}
\usage{
summary_custom(
  type = "numeric",
  include_pval = FALSE,
  stats_custom = NULL,
  formats_custom = NULL,
  labels_custom = NULL,
  indents_custom = NULL
)
}
\arguments{
\item{type}{(\code{string})\cr choice of a summary data type. Only \code{counts} and \code{numeric} types are currently supported.}

\item{include_pval}{(\code{logical})\cr whether p-value should be included as a default statistic.}

\item{stats_custom}{(\verb{named vector} of \code{character}) vector of statistics to include if not the defaults. This
argument overrides \code{include_pval} and other custom value arguments such that only settings for these statistics
will be returned.}

\item{formats_custom}{(\verb{named vector} of \code{character}) vector of custom statistics formats to use in place of the
defaults defined in \code{\link[=summary_formats]{summary_formats()}}. Names should be a subset of the statistics defined in \code{stats_custom} (or
default statistics if this is \code{NULL}).}

\item{labels_custom}{(\verb{named vector} of \code{character}) vector of custom statistics labels to use in place of the
defaults defined in \code{\link[=summary_labels]{summary_labels()}}. Names should be a subset of the statistics defined in \code{stats_custom} (or
default statistics if this is \code{NULL}).}

\item{indents_custom}{(\code{integer} or \verb{named vector} of \code{integer}) vector of custom indentation modifiers for
statistics to use instead of the default of \code{0L} for all statistics. Names should be a subset of the statistics
defined in \code{stats_custom} (or default statistics if this is \code{NULL}). Alternatively, the same indentation modifier
can be applied to all statistics by setting \code{indents_custom} to a single integer value.}
}
\description{
Set Defaults Settings for Summary Statistics
}
\examples{
summary_custom()
summary_custom(type = "counts", include_pval = TRUE)
summary_custom(
  include_pval = TRUE, stats_custom = c("n", "mean", "sd", "pval"),
  labels_custom = c(sd = "Std. Dev."), indents_custom = 3L
)

}
