% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_variables.R
\name{summary_formats}
\alias{summary_formats}
\title{Format Function for Descriptive Statistics}
\usage{
summary_formats(type = "numeric", include_pval = FALSE)
}
\arguments{
\item{type}{(\code{string})\cr choice of a summary data type. Only \code{counts} and \code{numeric} types are currently supported.}
}
\value{
A named \code{vector} of default statistic formats for the given data type.
}
\description{
Returns format patterns for descriptive statistics. The format is understood by \code{rtables}.
}
\examples{
summary_formats()
summary_formats(type = "count", include_pval = TRUE)

}
