% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_variables.R
\name{summary_labels}
\alias{summary_labels}
\title{Label Function for Descriptive Statistics}
\usage{
summary_labels()
}
\value{
A named \code{vector} of default statistic labels.
}
\description{
Returns labels of descriptive statistics for numeric variables.
}
\keyword{internal}
