% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_trees.R
\name{summary_tree}
\alias{summary_tree}
\title{Summarize a table_tree object}
\usage{
summary_tree(object, class_name, name = "", indent = 0, first = TRUE)
}
\arguments{
\item{object}{a \code{\link{table_tree}} object}

\item{name}{list element name of table}

\item{indent}{indentation of table}

\item{first}{logical, whether the nested table should be indented}
}
\description{
Summarize a table_tree object
}
