% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{t_ae}
\alias{t_ae}
\title{Adverse Events by System Organ Class and Preferred Term}
\usage{
t_ae(class, term, id, col_by, col_N, total = "All Patients")
}
\arguments{
\item{class}{class information as character or factor vector}

\item{term}{term information as character or factor vector, however factor
levels are not repeated by class, only terms with count > 1 are listed per
class}

\item{id}{unique subject identifier. If a particular subject has no adverse
event then that information needs to be added to the \code{col_N} argument.}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing. See 'Examples'.}

\item{col_N}{numeric vector with information of the number of patients in the
levels of \code{col_by}. This is useful if there are patients that have no
adverse events can be accounted for with this argument.}

\item{total}{character string that will be used as a label for a column with 
pooled total population, default is "All Patients". If the levels of col_by are 
the only columns of interest then total should be \code{NULL}}
}
\description{
\code{t_ae} returns adverse events summary table that corresponds to STREAM template AET02
}
\examples{
# Simple example
library(tibble)
library(dplyr)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = factor(paste("ARM", LETTERS[rep(c(1,2), c(3,7))]))
)


ae_lookup <- tribble(
  ~CLASS,         ~TERM,   ~GRADE,
  "cl A",   "trm A_1/2",        1,
  "cl A",   "trm A_2/2",        2,  
  "cl B",   "trm B_1/3",        2,
  "cl B",   "trm B_2/3",        3,
  "cl B",   "trm B_3/3",        1,
  "cl C",   "trm C_1/1",        1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2,2,2,3,3,4,4,4,4,5,6,6,7,7)]
  ),
  ae_lookup[c(1,1,2,6,4,2,2,3,4,2,1,5,4,6),]
)

ANL <- left_join(AAE, ASL, by = "USUBJID")

tbl <- t_ae(
  class = ANL$CLASS,
  term =  ANL$TERM,
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients"
)

tbl


library(random.cdisc.data)
library(dplyr)

ASL <- radam("ASL", N = 10)
AAE <- radam("AAE", ADSL = ASL)

ANL <- left_join(AAE, ASL \%>\% select(USUBJID, STUDYID, ARM), by = c("STUDYID", "USUBJID"))

tbl <- with(ANL,
            t_ae(
              class = AEBODSYS,
              term =  AEDECOD,
              id = USUBJID,
              col_by = factor(ARM),
              total = "All Patients"
            )
)

tbl

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
