% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_count_true.R
\name{t_count_true.data.frame}
\alias{t_count_true.data.frame}
\title{Tabulate \code{TRUE} counts in a data.frame}
\usage{
\method{t_count_true}{data.frame}(
  x,
  col_by,
  col_N = NULL,
  total = NULL,
  denominator = c("N", "n", "omit"),
  table_tree = FALSE,
  ...
)
}
\arguments{
\item{x}{data frame with only logical variables to be summarized.
If a variable has a \code{label} attribute then it will be used for the row name.}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{(\code{character} value)\cr
Denotes column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{denominator}{(one of \code{("n", "N", "omit"})\cr
Value set to \code{"n"} or \code{"N"} are for calculating the level associated percentage.
With option \code{"N"}, the reference population from \code{col_N} is used as
the denominator. With option \code{"n"}, the number of non-missing records from \code{x} is used as
the denominator. If \code{"omit"} is chosen the percentage is omitted.}

\item{table_tree}{(\code{logical} value)\cr
If \code{FALSE} then a single \code{rtable} gets returned, if \code{TRUE} a
\code{\link{node}} object gets returned with the elementary \code{rtable} objects.}

\item{...}{not used arguments}
}
\description{
Tabulate \code{TRUE} counts in a data.frame
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ANL <- ADSL \%>\%
  dplyr::transmute(
    SAFPOP = SAFFL == 'Y' ,
    DISCSTUD =  !is.na(DCSREAS)
  )

t_count_true(ANL, col_by = ADSL$ARM)

# Add variable labels
ANL <- ANL \%>\%
  var_relabel(
    SAFPOP = "Safety Population",
    DISCSTUD = "Discontinued study"
  )

t_count_true(ANL, col_by = ADSL$ARM)
t_count_true(ANL, col_by = ADSL$ARM, denominator = "omit")
}
\seealso{
\code{\link{t_count_true}},
\code{\link{t_count_true.logical}},
\code{\link{t_summary.logical}}, \code{\link{t_summary.factor}},
}
