% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_count_true.R
\name{t_count_true.logical}
\alias{t_count_true.logical}
\title{Tabulate \code{TRUE} counts in a vector}
\usage{
\method{t_count_true}{logical}(
  x,
  col_by,
  col_N = NULL,
  total = NULL,
  denominator = c("N", "n", "omit"),
  row_name = deparse(substitute(x)),
  ...
)
}
\arguments{
\item{x}{a logical vector}

\item{col_by}{logical, factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{string of column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{denominator}{either n, N or omit. n and N are for calculating the level associated
percentage. With option N, the reference population from \code{col_N} is used as the denominator.
With option n, the number of non-missing records from \code{x} is used as the denominator.
If \code{omit} is chosen the percentage is omitted.}

\item{row_name}{name of row}

\item{...}{not used arguments}
}
\description{
This function is a wrapper for \code{\link{t_summary.logical}} returning only
the counts for \code{TRUE} records. Note that for this function the default
\code{denominator = "N"} and not \code{n} as for the other \code{t_summary}
functions.
}
\examples{
with(iris, t_count_true(Sepal.Length > 5.5, col_by = Species, total = "All Flowers"))

# compare with t_summary.logical
with(iris, t_summary(Sepal.Length > 5.5, col_by = Species, total = "All Flowers"))

# show only counts without percent
with(iris, t_count_true(Sepal.Length > 5.5, col_by = Species, denominator = "omit"))
}
\seealso{
\code{\link{t_count_true}},
\code{\link{t_count_true.data.frame}},
\code{\link{t_summary.logical}}, \code{\link{t_summary.factor}}
}
