% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_count_true.R
\name{t_count_true}
\alias{t_count_true}
\title{Tabulate \code{TRUE} counts}
\usage{
t_count_true(x, col_by, col_N = NULL, total = NULL,
  denominator = c("N", "n", "omit"), ...)
}
\arguments{
\item{x}{an object to dispatch on}

\item{col_by}{logical, factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vector with the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{string of column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{denominator}{either \code{"n"}, \code{"N"} or \code{"omit"}. \code{"n"} and \code{"N"} are for calculating
the level associated percentage. With option \code{"N"}, the reference population from \code{col_N} is used as
the denominator. With option \code{"n"}, the number of non-missing records from \code{x} is used as
the denominator. If \code{"omit"} is chosen the percentage is omitted.}

\item{...}{arguments passed on to methods}
}
\description{
Note that for this function the default \code{denominator = "N"} and
not \code{n} as for the other \code{t_summary} functions.
}
\examples{

t_count_true(c(TRUE, TRUE, FALSE, FALSE), col_by = factor(c("A", "B", "B", "B")),
  row_name = "TRUE")

library(dplyr)
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

t_count_true(x = !is.na(ADSL$DCSREAS), col_by = ADSL$ARM, col_N = table(ADSL$ARM),
  row_name = "Patients who discontinued study" )

# example with missing values
AGE_NA <- ADSL$AGE
AGE_NA[1:50] <- NA
t_count_true(AGE_NA > 35 , col_by = ADSL$ARM, col_N = table(ADSL$ARM),
  row_name = "AGE > 35")
t_count_true(AGE_NA > 35 , col_by = ADSL$ARM, col_N = table(ADSL$ARM),
  row_name = "AGE > 35", denominator = "n")

ADAE <- radae(cached = TRUE)
ANL <- ADAE \%>\%
  dplyr::transmute(
    SER = AESER == 'Y' ,
    GRADE35 =  AETOXGR \%in\% c('3', '4', '5')
  ) \%>\%
  var_relabel(
    SER = "Serious",
    GRADE35 = "Grade 3-5"
  )
tbl <- t_count_true(ANL, col_by = ADAE$ARMCD, col_N = table(ADSL$ARMCD),
  denominator = "omit")
tbl

# Add descriptive row:
insert_rrow(indent(tbl, 1), rrow("Total number of adverse events which are"))
}
\seealso{
\code{\link{t_count_true.logical}},
   \code{\link{t_count_true.data.frame}},
   \code{\link{t_summary.logical}}, \code{\link{t_summary.factor}}
}
