% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_coxreg.R
\name{t_cox_multivariate}
\alias{t_cox_multivariate}
\title{Multivariate Cox Model - tabulated results}
\usage{
t_cox_multivariate(
  formula,
  data,
  conf_level = 0.95,
  pval_method = c("wald", "likelihood"),
  ...
)
}
\arguments{
\item{formula}{A \code{formula} corresponding to the investigated \code{\link[survival:Surv]{survival model}}
including covariates.}

\item{data}{A \code{data.frame} which includes the variable in formula and covariates.}

\item{conf_level}{The level of confidence for the hazard ration interval estimations. Default is 0.95.}

\item{pval_method}{The method used for the estimation of p.values, should be one of \code{"wald"} (default) or
\code{"likelihood"}.}

\item{...}{Optional parameters passed to \code{\link[survival:coxph]{coxph()}}
\itemize{
\item \code{ties} a character string specifying the method for tie handling, one of \code{exact} (default), \code{efron}, \code{breslow}.
}}
}
\description{
Analysis based on multivariate Cox model is usually not performed for the \code{CSR} or regulatory documents and
is for exploratory purposes only (e.g., for publication). In practice, the model usually includes only the main
effects (without interaction terms). It produces the estimates for each of the covariates included in the model.
The analysis follows the same principles (e.g., stratified vs. unstratified analysis and tie handling) as the
usual Cox model analysis. Since there is usually no pre-specified hypothesis testing for such analysis,
the p values need to be interpreted with caution. (\strong{Statistical Analysis of Clinical Trials Data with R},
\verb{NEST's bookdown})
}
\examples{
library(tern)
library(random.cdisc.data)
ADTTE   <- radtte(cached = TRUE)
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")   # _f: filtered
ADTTE_f <- dplyr::filter (
  ADTTE_f,
  PARAMCD == "OS"
  & SEX \%in\% c("F", "M")
  & RACE \%in\% c("ASIAN", "BLACK OR AFRICAN AMERICAN", "WHITE")
);
ADTTE_f$SEX  <- droplevels(ADTTE_f$SEX)
ADTTE_f$RACE <- droplevels(ADTTE_f$RACE)

\dontrun{
t_cox_multivariate(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ ARMCD + AGE + RACE,
  data = ADTTE_f, conf_level = 0.95
)

t_cox_multivariate(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ (ARMCD + RACE + AGE)^2,
  data = ADTTE_f
)
}
}
