% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_coxreg.R
\name{t_coxph_pairwise}
\alias{t_coxph_pairwise}
\title{Proportional Hazards Regression Model Fit Summary Table (pair-wise for multiple arms)}
\usage{
t_coxph_pairwise(
  formula,
  data,
  conf_level = 0.95,
  pval_method = c("log-rank", "wald", "likelihood"),
  ...
)
}
\arguments{
\item{formula}{(\code{formula})\cr
Specifies \code{\link[survival:Surv]{survival model}}.
The arm variable needs to be wrapped in \code{\link{arm}}. The
\code{\link[survival]{strata}} special will only be used for the stratified analysis. If there is not
\code{\link[survival]{strata}} specification then the stratified analysis is omitted.}

\item{data}{(\code{data.frame})\cr
Contains all the variable that are used in \code{formula}}

\item{conf_level}{(\code{numeric} value)\cr
level for computation of the confidence intervals.
If set to \code{FALSE} no confidence intervals are printed}

\item{pval_method}{(one of (\code{"log-rank", "likelihood", "wald"}))\cr
Specifies the method used to calculate the p-value. Default value is \code{"log-rank"}.}

\item{...}{(optional)\cr
Other arguments will be passed to \code{\link[survival]{coxph}}}
}
\description{
An \code{\link[rtables]{rtable}} format of \code{\link[tern]{s_coxph_pairwise}}
results for further annotation on top of Kaplan-Meier grob.
}
\details{
The Cox PH model is evaluated pair-wise (reference to
comparison) and \code{\link[survival]{coxph}} is used to get the p-value,
calculate the hazard ratio and confidence interval.
For example, given an ARM variable has reference group "ARM A" and
two comparison groups "ARM B" and "ARM C",
t_coxph_pairwise will conduct two Cox PH models for the specified \code{formula},
one is for the subset data with only "ARM A" and "ARM B",
the other is for the subset data with only "ARM A" and "ARM C".
}
\examples{
library(random.cdisc.data)
ADSL <- cadsl
ADTTE <- cadtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

tbl <- t_coxph_pairwise(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f
)
tbl

tbl <- t_coxph_pairwise(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + strata(SEX),
  data = ADTTE_f,
  conf_level = 0.8, pval_method = "wald", ties = "exact"
)
tbl
}
\author{
Heng Wang (wangh107) \email{wang.heng@gene.com}
}
