% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_coxreg.R
\name{t_coxreg}
\alias{t_coxreg}
\title{Cox regression for standard outputs}
\usage{
t_coxreg(formula, data, ...)
}
\arguments{
\item{formula}{(\code{formula}) Specifies \code{\link[survival:Surv]{survival model}}.
\itemize{
\item the left-hand side must include \code{Surv(time, event)} with time and event describing occurrence
and censoring (see  \code{\link[survival:Surv]{Surv()}}).
\item the right-hand side includes predictors and \emph{specials} which can be used to specify
a pairwise model or a range of candidate univariate models. See \code{details}.
}}

\item{data}{(\code{data.frame})\cr
Contains all the variables that are used in \code{formula}}

\item{...}{(optional) other argument passed to cox regression:
\itemize{
\item \code{conf_level}   The level of confidence for the hazard ration interval estimations. Default is 0.95.
\item \code{pval_method}  The method used for estimation of p.values, should be one of \code{wald} (default) or
\code{likelihood}.
\item \code{ties} a character string specifying the method for tie handling, one of \code{exact} (default), \code{efron}, \code{breslow}.
\item \code{increments} If a quantitative variable is included, it is possible to provide the expected level
of estimation for the interaction. If provided, it should then be list where
each item is vector giving expected levels and is named after the variable name as it appears
in the formula.
}}
}
\value{
Depending on \code{formula}, returns an object of class \code{tbl} for every level of \code{pairwise()}
if included in \code{COXT01}, a single \code{tbl} otherwise.
}
\description{
Cox models are the most commonly used methods to estimate the magnitude of the effect in survival analysis.
It assumes proportional hazards; that is, it assumes that the ratio of the hazards of the two groups
(e.g., two arms) is constant over time. This ratio is referred to as the "hazard ratio" and is one of the most
commonly reported metrics to describe the effect size in survival analysis.
}
\details{
Possible model specifications:
\itemize{
\item \code{Surv(time, event) ~ pairwise(Pred)}, used to get hazard ratio of each tested level of
treatment \code{Pred} independently, given in reference to the control treatment
(see \link[tern:pairwise]{pairwise}).
\item \code{Surv(time, event) ~ pairwise(Pred) + univariate(Cov1, Cov2, ...)}, add candidate covariates
separated by comas in special \code{univariate()} to test all bivariate combinations of \code{Pred} with
each \code{Cov}. Replacing the symbol \code{+} by \code{*} will result in the additional estimation of the interaction
terms. This is similar to \code{COXT01} standards.
\item \code{Surv(time, event) ~ Pred + Cov1 + ... } estimate the simple effect of multiple Cox regression.
This provides the \code{COXT02} standard output.
}

Known limits:
\itemize{
\item the \emph{special} \code{univariate()} requires \code{pairwise()} being also specified.
\item the interaction terms in multiple regression are not estimated.
\item the third-order interactions are not estimated.
}

For further information about the Cox Proportional Hazards Model, check
"Statistical Analysis of Clinical Trials Data with R", NEST team.
}
\examples{
library(tern)
library(random.cdisc.data)
ADTTE   <- radtte(cached = TRUE)
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")   # _f: filtered
ADTTE_f <- within( # nolint
  data = subset(
    ADTTE_f,
    PARAMCD == "OS"
    & SEX \%in\% c("F", "M")
    & RACE \%in\% c("ASIAN", "BLACK OR AFRICAN AMERICAN", "WHITE")
  ),
  expr = { # nolint start
    set.seed(1)
    ARMCD <- droplevels(ARMCD)
    ARMCD <- relevel(ARMCD, "ARM B")
    SEX <- droplevels(SEX)
    RACE <- droplevels(RACE)
    X <- rnorm(n = length(ARM))
  }  # nolint end
)

# For annotation on top of Kaplan-Meier grob.
tern:::t_coxreg(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ pairwise(ARMCD), data = ADTTE_f,
  conf_level = 0.8, pval_method = "likelihood",  ties = "breslow"
)

## COXT01 - Standard output, no interactions
tern:::t_coxreg(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ pairwise(ARMCD) + univariate(SEX, RACE, AGE),
  data = ADTTE_f
)

\dontrun{
  # For comparison
  mod10 <- coxph(Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD), data = ADTTE_f)
  mod11 <- coxph(Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + SEX, data = ADTTE_f)
  mod12 <- coxph(Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + AGE, data = ADTTE_f)
  mod13 <- coxph(Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + RACE, data = ADTTE_f)

  car::Anova(mod10,  test.statistic = "Wald", type = 'III')
  car::Anova(mod11, test.statistic = "Wald", type = 'III')
  car::Anova(mod12, test.statistic = "Wald", type = 'III')
  car::Anova(mod13, test.statistic = "Wald", type = 'III')

  summary(mod13)$coefficients
}

## COXT01 - options: control ties, set confidence interval level, choose a strata,
## modify pval_method
tern:::t_coxreg(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ pairwise(ARMCD) + strata(SEX) +
    univariate(RACE, AGE),
  data = ADTTE_f, conf_level = 0.80,
  ties = c("exact", "efron", "breslow")[3], pval_method  = c("wald", "likelihood")[2]
)

\dontrun{
  # For comparison
  mod20 <- coxph(Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + strata(SEX), data = ADTTE_f)
  mod21 <- coxph(
    Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + strata(SEX) + AGE, data = ADTTE_f
  )
  mod22 <- coxph(
    Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + strata(SEX) + RACE,
    data = ADTTE_f
  )

  car::Anova(mod20, test.statistic = "Wald", type = 'III')
  car::Anova(mod21, test.statistic = "Wald", type = 'III')
  car::Anova(mod22, test.statistic = "Wald", type = 'III')

  summary(mod22)$coefficients
}

\dontrun{
  ## COXT01 - Addition of the interaction and increments to choose estimated levels
  tern:::t_coxreg(
    formula =  Surv(time = AVAL, event = 1 - CNSR) ~
      pairwise(ARMCD) * univariate(RACE, SEX, AGE, X),
    data = ADTTE_f, conf_level = 0.74, increments = list(X = c(-1, 0, 2))
  )

  ## COXT01: Other example, other data
  library(dplyr)
  ADSL <- radsl(cached = TRUE)
  ADSL <- ADSL \%>\% filter(SEX \%in\% c("F", "M"))

  ADTTE <- radtte(ADSL, seed = 2) \%>\%
    filter(PARAMCD == "PFS" & ARMCD != "ARM C")
  ADTTE$ARMCD <- droplevels(ADTTE$ARMCD)
  ADTTE$SEX <- droplevels(ADTTE$SEX)
  tern:::t_coxreg(
    formula = Surv(time = AVAL, event = 1 - CNSR) ~
      pairwise(ARMCD) * univariate(SEX, BMRKR2, AGE),
    data = ADTTE,
    increments = list(AGE = mean(ADTTE$AGE)),
    conf_level = 0.975)

  # check HRs
  m2 <- coxph(Surv(time = AVAL, event = 1 - CNSR) ~ ARMCD*BMRKR2, data = ADTTE)
  coef(m2)

  # For COXT02
  tern:::t_coxreg(
    formula = Surv(time = AVAL, event = 1 - CNSR) ~ ARMCD + RACE * AGE, data = ADTTE_f,
    conf_level = 0.8, pval_method = "likelihood",  ties = "breslow"
  )
}
}
\seealso{
\code{\link{t_coxph_pairwise}},
\code{\link{t_cox_univariate}},
\code{\link{t_cox_multivariate}}
}
