% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_dsp.R
\name{t_dsp}
\alias{t_dsp}
\title{Single element for disposition table}
\usage{
t_dsp(x = x, col_by, col_N = table(col_by), subset = NULL,
  indent = 0, row.name = NULL)
}
\arguments{
\item{x}{a factor or logical vector.}

\item{col_by}{The group variable to define columns.}

\item{col_N}{The column total for each group that is displayed in the table header with (N=xx).}

\item{subset}{A logical vector with the same length as x that defines the subset}

\item{indent}{An integer specifing the spaces before row.name.}

\item{row.name}{Only applicable when x is a logical vector. A string to label the row name. Default is "TRUE".}

\item{...}{arguments passed on to methods}
}
\description{
Single element for disposition table
}
\examples{

library(purrr)
dsp <- purrr::partial(t_dsp, col_by = ADSL0$ARM)

rbind(
  dsp(ADSL0$COMPSTUD == "Y", row.name = "Completed study"),
  dsp(ADSL0$STUDONS == "Alive: In Follow-up", row.name = "Alive: In follow-up"),
  dsp(ADSL0$DISSTDFL == "Y",  row.name = "Discontinued study"),
  dsp(as.factor(ADSL0$STDDRS), indent = 1),
  rrow(),
  dsp(ADSL0$GOTTRT == "Y", row.name = "Received treatment"),
  dsp(ADSL0$DISTRTFL == "Y", row.name = "Discontinued treatment"),
  dsp(ADSL0$DRSCAT == "Safety", row.name = "Safety", indent = 1),
  dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT \%in\% "Safety", indent = 2),
  dsp(ADSL0$DRSCAT == "Other", row.name = "Other", indent = 1),
  dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT \%in\% "Other", indent = 2)
)

}
