% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_abn_shift.R
\name{t_el_abn_shift}
\alias{t_el_abn_shift}
\title{Count abnormality shift from baseline for each abnormality level}
\usage{
t_el_abn_shift(
  rri_ana,
  rri_base,
  value_ana,
  value_base,
  abnormal = c("HIGH", "LOW"),
  id,
  col_by,
  col_N = NULL,
  total = NULL
)
}
\arguments{
\item{rri_ana}{a vector specifying the Reference Range Indicator (e.g. High/Low/Normal). \cr
Fill blank with NA using \code{\link[=sas_na]{sas_na()}}.}

\item{rri_base}{a vector specifying the Reference Range Indicator (e.g. High/Low/Normal) at baseline \cr
for the corresponding records in \code{rri_ana}. Fill blank with NA using \code{\link[=sas_na]{sas_na()}}.}

\item{value_ana}{a vector of numeric measurement.}

\item{value_base}{a vector of numeric measurement at baseline \cr
for the corresponding records in \code{value_ana}.}

\item{abnormal}{a vector specifying the unique values that
represent abnormality in the order of display.}

\item{id}{a vector of subject identifier with length of \code{rri_ana}.}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{(\code{character} value)\cr
Denotes column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}
}
\value{
an \code{rtable}
}
\description{
This is the elementary function used in function \link{t_abn_shift}, where the calculation could be repeated \cr
for multiple post-baseline visits and multiple laboratory tests.
}
\examples{
library(dplyr)

ANL <- tibble(
  USUBJID = c("ID1", "ID1", "ID2", "ID2", "ID3", "ID3"),
  ANRIND = c("N", "L", "N", "N", "N", "H"),
  BNRIND = c("N", "N", "L", "L", "", ""),
  AVAL = c(1.5, 0.4, 1.8, 1.9, 2.4, 4.6),
  BASE = c(1.2, 1.2, 0.5, 0.5, NA, NA),
  ARM = factor(rep("ARM A", 6)),
)

t_el_abn_shift(
  rri_ana = sas_na(ANL$ANRIND),
  rri_base = sas_na(ANL$BNRIND),
  value_ana = ANL$AVAL,
  value_base = ANL$BASE,
  abnormal = c("H", "L"),
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = table(ANL$ARM),
  total = "All Patients"
)

t_el_abn_shift(
  rri_ana = sas_na(ANL$ANRIND),
  rri_base = sas_na(ANL$BNRIND),
  value_ana = ANL$AVAL,
  value_base = ANL$BASE,
  abnormal = c("H", "L"),
  id = ANL$USUBJID,
  col_by = by_all("All Patients"),
  col_N = nrow(ANL)
)

}
