% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ancova.R
\name{t_el_ancova}
\alias{t_el_ancova}
\title{Elementary table for analysis of covariance (ANCOVA).}
\usage{
t_el_ancova(formula, data, col_N, conf_level = 0.95)
}
\arguments{
\item{formula}{A \code{formula} corresponding to the investigated \code{\link[stats:lm]{linear model}}.
The left-hand side must include the dependent variable. The right-hand side must include the group variable
name wrapped in the special \code{arm()}.}

\item{data}{A \code{data.frame} which includes all the variables that are called in \code{formula}.}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{conf_level}{confidence level of the interval.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
This function returns an elementary summary table for an ANCOVA analysis, which compares the
covariate adjusted means of multiple groups with a reference group. Specifically, the
first level of the `arm()` factor variable on the right-hand side of the provided `formula`
is taken as the reference group.
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ADQS_filtered <- radqs(cached = TRUE) \%>\%
  dplyr::filter(PARAMCD == "FKSI-FWB", AVISIT == "WEEK 1 DAY 8")
t_el_ancova(
  formula = CHG ~ BASE + STRATA1 + arm(ARMCD),
  data = ADQS_filtered,
  col_N = table(ADSL$ARMCD)
)
}
\seealso{
\code{\link{s_ancova}}
}
