% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_count_unique.R
\name{t_el_count_unique}
\alias{t_el_count_unique}
\title{Count unique elements per cell in a vector}
\usage{
t_el_count_unique(
  x,
  col_by,
  col_N = NULL,
  total = NULL,
  na_rm = TRUE,
  denominator = c("N", "omit"),
  row_name = "number of unique elements"
)
}
\arguments{
\item{x}{a vector}

\item{col_by}{logical, factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vector with the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{string of column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{na_rm}{a logical value indicating whether \code{NA} values should be
removed from \code{x} prior to counting the unique elements per cell.}

\item{denominator}{either N or omit. With option N, the reference population
from \code{col_N} is used as the denominator. If \code{omit} is chosen,
the percentage is omitted and only the counts are returned.}

\item{row_name}{a string with the row name to display in the summary table
that is returned. Default is "number of unique elements."}
}
\value{
an \code{rtable}
}
\description{
Count unique elements per cell in a vector
}
\examples{

t_el_count_unique(
 x = paste("id", c(1, 4, 2, 3, 3), sep = "-"),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10)
)

t_el_count_unique(
 x = paste("id", c(1, 4, 2, 3, 3), sep = "-"),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 denominator = "omit"
)

# how missing values are handled
t_el_count_unique(
 x = c(1, 4, 2, 3, 3, NA),
 col_by = factor(c("A", "A", "B", "C", "C", "C")),
 col_N = c(2, 4, 10),
 na_rm = TRUE
)

library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADAE <- radae(cached = TRUE)

t_el_count_unique(ADAE$USUBJID, col_by = ADAE$ARM, col_N = table(ADSL$ARM),
  row_name = "Total number of patients with at least one AE")

}
\seealso{
\code{\link{t_count_unique}},
 \code{\link{t_events_per_term_id}}, \code{\link{t_el_events_per_term_id}},
 \code{\link{t_events_per_term_grade_id}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
