% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_disposition.R
\name{t_el_disposition}
\alias{t_el_disposition}
\title{Single element for disposition table}
\usage{
t_el_disposition(x = x, col_by, col_N = NULL, row.name = NULL,
  subset = NULL, show_n = FALSE, useNA = c("no", "ifany", "always"),
  drop_levels = NULL, denominator = "N")
}
\arguments{
\item{x}{A factor or logical vector}

\item{col_by}{logical factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default
\code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.}

\item{row.name}{Only applicable when x is a logical vector. A string to label the row name.
Default is "TRUE".}

\item{subset}{A logical vector with the same length as x that defines the subset. Applies to \code{x} and \code{subset}.}

\item{show_n}{Logic value to determine whether the "n" row is displayed. Default is FALSE. Only when x is of type factor}

\item{useNA}{choose whether missing data (NAs) should be displayed as a level.}

\item{drop_levels}{boolean whether to drop zero count levels}

\item{denominator}{either n or N for calculating the level associated
percentage. With option N, the reference population from \code{col_N} is used as the denominator.
With option n, the number of non-missing records from \code{x} is used as the denominator.}
}
\description{
Single element for disposition table
}
\examples{

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)

ADSL0[, c("COMPSTUD", "STUDONS")]
ADSL0 <- ADSL \%>\%
 mutate(
   COMPSTUD = sample(c('Y','N'),
                     size=nrow(ADSL),
                     replace = TRUE) \%>\% as.factor,
   STUDONS = sample(c('Alive: On Treatment', 'Alive: In Follow-up', NA),
                    size=nrow(ADSL),
                    replace = TRUE) \%>\% as.factor,
   STDDRS = sample(c('Death', 'Lost To Follow-Up',
                     'Protocol Violation', 'Withdrawal By Subject',
                     'Other'),
                   size=nrow(ADSL),
                   replace = TRUE) \%>\% as.factor,
   GOTTRT = ifelse(!is.na(ACTARMCD), 'Y', 'N') \%>\% as.factor,
   DISTRTFL = sample(c('Y','N'),
                     size=nrow(ADSL),
                     replace = TRUE) \%>\% as.factor,
   TRTDRS = sample(c('ADVERSE EVENT', 'PROGRESSIVE DISEASE',
                     'PHYSICIAN DECISION', 'LACK OF EFFICACY',
                     'OTHER'),
                   size=nrow(ADSL),
                   replace = TRUE) \%>\% as.factor,
   STUDONS = case_when(
     COMPSTUD == 'N' ~ STUDONS
   ),
   STDDRS = case_when(
     COMPSTUD == 'N' & is.na(STUDONS) ~ STDDRS
   ),
   DISSTDFL = case_when(
     !is.na(STDDRS) ~ "Y"
   ),
   DISTRTFL = case_when(
     GOTTRT == 'Y' ~ DISTRTFL
   ),
   TRTDRS = case_when(
     DISTRTFL == 'Y' ~ TRTDRS
   ),
   DRSCAT = case_when(
     TRTDRS \%in\% c('ADVERSE EVENT', 'PHYSICIAN DECISION') ~ "Safety",
     !is.na(TRTDRS) ~ "Other"
   )
 ) \%>\% var_relabel(COMPSTUD = "Complete Study",
                   STUDONS = "On-study Status",
                   DISSTDFL = "Discontinued Study",
                   STDDRS = "Reason for Study \\r\\nDiscontinuation",
                   GOTTRT = "Received Treatment",
                   DISTRTFL = "Discontinued Treatment",
                   TRTDRS = "Reason for Treatment \\r\\nDiscontinuation"
                   )

library(purrr)
dsp <- partial(t_el_disposition, col_by = ADSL0$ARM \%>\% by_add_total("All Patients"), denominator = "N")

rbind(
  dsp(ADSL0$COMPSTUD == "Y", row.name = "Completed study"),
  dsp(ADSL0$STUDONS == "Alive: In Follow-up", row.name = "Alive: In follow-up"),
  dsp(ADSL0$DISSTDFL == "Y",  row.name = "Discontinued study"),
  dsp(as.factor(ADSL0$STDDRS), indent = 1),
  rrow(),
  rrow("Show example of using n as denominator"),
  t_el_disposition(as.factor(ADSL0$STDDRS), col_by = ADSL0$ARM, denominator = "n",
     total = "All Patients", indent = 1),
  rrow(),
  dsp(ADSL0$GOTTRT == "Y", row.name = "Received treatment"),
  dsp(ADSL0$DISTRTFL == "Y", row.name = "Discontinued treatment"),
  dsp(ADSL0$DRSCAT == "Safety", row.name = "Safety", indent = 1),
  dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT \%in\% "Safety", indent = 2),
  dsp(ADSL0$DRSCAT == "Other", row.name = "Other", indent = 1),
  dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT \%in\% "Other", indent = 2)
)

# todo: remove subset
}
