% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_rsp.R
\name{t_el_forest_rsp}
\alias{t_el_forest_rsp}
\title{Elementary Table for Forest Response Plot}
\usage{
t_el_forest_rsp(
  rsp,
  col_by,
  strata_data = NULL,
  conf_int = 0.95,
  row_name = "",
  dense_header = FALSE
)
}
\arguments{
\item{rsp}{boolean vector, \code{TRUE} if subject is a responder \code{FALSE}
otherwise}

\item{col_by}{logical factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{strata_data}{data for stratification factors (categorical variables).
If \code{NULL}, no stratified analysis is performed.}

\item{conf_int}{confidence level of the interval}

\item{row_name}{name of row}

\item{dense_header}{Display the table headers in multiple rows.}
}
\value{
rtable with one row
}
\description{
Elementary Table for Forest Response Plot
}
\examples{
rsp = sample(c(TRUE, FALSE), 200, TRUE)
col_by = factor(sample(c("ARM A", "ARM B"), 200, TRUE), levels = c("ARM A", "ARM B"))
strata_data <- data.frame(
  STRATA1 = sample(c("STR1", "STR2"), 200, TRUE),
  STRATA2 = sample(c("low", "medium", "high"), 200, TRUE)
)
t_el_forest_rsp(
  rsp = rsp,
  col_by = col_by
)

t_el_forest_rsp(
  rsp = rsp,
  col_by = col_by,
  strata_data = strata_data
)

}
\seealso{
\code{\link{t_forest_rsp}}
}
