% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_tte.R
\name{t_el_forest_tte}
\alias{t_el_forest_tte}
\title{Elementary Table for Forest Time to Event Plot}
\usage{
t_el_forest_tte(tte, is_event, col_by, strata_data = NULL,
  ties = "exact", time_unit = "month", row_name = "",
  conf_int = 0.95, dense_header = TRUE)
}
\arguments{
\item{tte}{vector with numeric time to event data, currently no \code{NA} allowed}

\item{is_event}{is boolean, \code{TRUE} if event, \code{FALSE} if \code{tte}
is censored}

\item{col_by}{logical, factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{strata_data}{data for stratification factors (categorical variables).
If \code{NULL}, no stratified analysis is performed.}

\item{ties}{the method used for tie handling in \code{\link[survival]{coxph}}.}

\item{time_unit}{The unit of median survival time. Default is \code{months}.}

\item{row_name}{name of row}

\item{conf_int}{confidence level of the interval}

\item{dense_header}{Display the table headers in multiple rows.}
}
\value{
\code{rtable} with one row
}
\description{
Elementary Table for Forest Time to Event Plot
}
\examples{

n <- 200
tte <- rexp(n)
set.seed(1)
is_event <- sample(c(TRUE, FALSE), n, TRUE)
strata_data <- data.frame(
  STRATA1 = sample(c("STR1", "STR2"), n, TRUE),
  STRATA2 = sample(c("low", "medium", "high"), n, TRUE)
)
col_by = factor(sample(c("ARM A", "ARM B"), n, TRUE), levels = c("ARM A", "ARM B"))
t_el_forest_tte(
  tte = tte, is_event = is_event,
  col_by = col_by,
  conf_int = 0.9
)

t_el_forest_tte(
  tte = tte, is_event = is_event,
  col_by = col_by,
  strata_data = strata_data
)

t_el_forest_tte(
  tte = tte, is_event = is_event,
  col_by = factor(rep("ARM A", n), levels = c("ARM A", "ARM B"))
)

t_el_forest_tte(
  tte = tte, is_event = is_event,
  col_by = factor(rep("ARM B", n), levels = c("ARM A", "ARM B"))
)
}
\seealso{
\code{\link{t_forest_tte}}
}
