% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{t_el_multinomial_proportion}
\alias{t_el_multinomial_proportion}
\title{Proportions: outcomes by level}
\usage{
t_el_multinomial_proportion(
  rsp_multinomial,
  col_by,
  conf_level = 0.95,
  prop_ci_method = "waldcc",
  levels_without_ci = NULL
)
}
\arguments{
\item{rsp_multinomial}{(\code{factor} vector) The levels represent ordered response categories.
\code{NA} values or empty strings are not allowed.}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{conf_level}{confidence level of the interval.}

\item{prop_ci_method}{one of (\code{"waldcc"}, \code{"wald"}, \code{"clopper-pearson"},
\code{"wilson"}, \code{agresti-coull} or \code{jeffreys}).\cr
Specifies the method used to construct the confidence interval for proportion
of successful outcomes.}

\item{levels_without_ci}{(\code{character} vector) If any levels of \code{rsp_multinomial}
match, the corresponding rates are reported without a confidence interval.}
}
\value{
return_rtable
}
\description{
Table for proportion of outcomes per factor level
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADRS <- radrs(cached = TRUE) \%>\%
  dplyr::filter(PARAMCD == "BESRSPI")

t_el_multinomial_proportion(ADRS$AVALC, col_by = ADRS$ARMCD)

t_el_multinomial_proportion(ADRS$AVALC, col_by = ADRS$ARMCD, levels_without_ci = "NE")

# apply standard oncology labels, given factor input.
m_rsp <- ADRS$AVALC
levels(m_rsp) <- get_onco_rsp_label()[match(levels(m_rsp), names(get_onco_rsp_label()))]
t_el_multinomial_proportion(m_rsp, col_by = ADRS$ARMCD, levels_without_ci = "Not Evaluable (NE)")

t_el_multinomial_proportion(
  ADRS$AVALC,
  col_by = ADRS$ARMCD,
  prop_ci_method = "wilson",
  conf_level = 0.90
)

# single arm
t_el_multinomial_proportion(
  ADRS$AVALC,
  col_by = by_all("All"),
  prop_ci_method = "clopper-pearson"
)

# Add total
t_el_multinomial_proportion(
  ADRS$AVALC,
  col_by = by_add_total(ADRS$ARMCD, "All")
)
}
\seealso{
\code{\link{s_proportion}}, \code{\link{t_el_proportion}},
  \code{\link{t_binary_outcome}}
}
