% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{t_el_odds_ratio}
\alias{t_el_odds_ratio}
\title{OR: elementary summary table}
\usage{
t_el_odds_ratio(rsp, col_by, conf_level = 0.95, strata_data = NULL)
}
\arguments{
\item{rsp}{(\code{logical} vector)\cr
\code{TRUE} if subject is a responder \code{FALSE} otherwise}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{conf_level}{confidence level of the interval.}

\item{strata_data}{(\code{data.frame}) Optional stratification factors. If \code{NULL}, the unstratified analysis
will be performed.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
Elementary summary table for odds ratio estimation.
}
\details{
This function uses either logistic regression for unstratified analyses, or conditional logistic
  regression for stratified analyses.
  The Wald confidence interval with the specified confidence level is calculated. Note that
  for stratified analyses, there is currently no implementation for conditional likelihood confidence intervals,
  therefore we don't offer the option of a likelihood confidence interval.
  Note that when `rsp` contains only responders or non-responders, then the result values will be `NA`,
  because no odds ratio estimation is possible.
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADRS <- radrs(cached = TRUE) \%>\%
  dplyr::filter(PARAMCD == "BESRSPI")

t_el_odds_ratio(
  rsp = ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  conf_level = 0.9
)

t_el_odds_ratio(
  rsp = ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  strata_data = ADRS[, c("STRATA1", "STRATA2")]
)
}
\seealso{
\code{\link{s_odds_ratio}}, \code{\link{t_el_proportion}},
  \code{\link{t_binary_outcome}}
}
