% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{t_el_proportion}
\alias{t_el_proportion}
\title{Table for proportion of successful outcomes}
\usage{
t_el_proportion(rsp, col_by, conf_level = 0.95, prop_ci_method = "waldcc")
}
\arguments{
\item{rsp}{(\code{logical} vector)\cr
\code{TRUE} if subject is a responder \code{FALSE} otherwise}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{conf_level}{confidence level of the interval.}

\item{prop_ci_method}{one of (\code{"waldcc"}, \code{"wald"}, \code{"clopper-pearson"},
\code{"wilson"}, \code{agresti-coull} or \code{jeffreys}).\cr
Specifies the method used to construct the confidence interval for proportion
of successful outcomes.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
Table for proportion of successful outcomes
}
\details{
The following options are supported for confidence interval for a proportion:
\itemize{
\item The \code{wald} interval follows the usual textbook definition for a single proportion
confidence interval using the normal approximation.
\item The \code{waldcc} interval is similar to \code{wald} and in addition uses the continuity
correction \code{1/2n}.
\item The \code{clopper-pearson} interval calls \code{\link[stats]{binom.test}}. Also referred to
as the 'exact' method.
\item The \code{wilson} interval calls \code{\link[stats]{prop.test}} with option \code{correct=FALSE}.
Also referred to as Wilson score interval.
\item The \code{agresti-coull} interval was created by Alan Agresti and Brent Coull and can be understood (for 95\% CI)
as adding two successes and two failures to the data, and then using the Wald formula to construct a CI.
\item The \code{jeffreys} interval is an equal-tailed interval based on the non-informative Jeffreys prior for a
binomial proportion.
}
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADRS <- radrs(cached = TRUE) \%>\%
  dplyr::filter(PARAMCD == "BESRSPI")

t_el_proportion(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD
)

t_el_proportion(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  prop_ci_method = "wilson",
  conf_level = 0.90
)

# single arm
t_el_proportion(
  ADRS$AVALC == "CR",
  col_by = by_all("All"),
  prop_ci_method = "clopper-pearson"
)

# Add total
t_el_proportion(
  ADRS$AVALC == "CR",
  col_by = by_add_total(ADRS$ARMCD, "All")
)
}
\seealso{
\code{\link{s_proportion}},
  \code{\link{t_el_proportion_diff}}, \code{\link{t_el_test_proportion_diff}},
  \code{\link{t_binary_outcome}}
}
