% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{t_el_proportion_diff}
\alias{t_el_proportion_diff}
\title{Table for difference in proportions}
\usage{
t_el_proportion_diff(
  rsp,
  col_by,
  strata_data = NULL,
  conf_level = 0.95,
  diff_ci_method = c("wald", "waldcc", "cmh")
)
}
\arguments{
\item{rsp}{(\code{logical} vector)\cr
\code{TRUE} if subject is a responder \code{FALSE} otherwise}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{strata_data}{(\code{data.frame})\cr
Used for stratification factors. If \code{NULL}, no stratified
analysis is performed.}

\item{conf_level}{confidence level of the interval.}

\item{diff_ci_method}{one of (\code{"wald"}, \code{"waldcc"}, \code{"cmh"})\cr
Specifies the method used to construct the confidence interval for the difference
in proportion of outcomes.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
This function will calculate the difference in proportion of events between each
comparison group versus the reference group. The first level of \code{col_by} is
used as a reference group.
}
\details{
The following options are supported for difference of proportions confidence intervals:
\itemize{
  \item{Option \code{wald} calls \code{\link[stats]{prop.test}} with \code{correct=FALSE}}
  \item{Option \code{waldcc} includes continuity correction for Wald interval and
  calls \code{\link[stats]{prop.test}} with \code{correct=TRUE}}
  \item{Option \code{cmh} derives a CI for CMH-weighted difference of proportions and
   calls \code{\link{s_adj_proportion_diff}}.}
}
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADRS <- radrs(cached = TRUE) \%>\%
  dplyr::filter(PARAMCD == "BESRSPI")

t_el_proportion_diff(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  diff_ci_method = "wald"
)

t_el_proportion_diff(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  strata_data = ADRS[, c("STRATA1", "STRATA2")],
  diff_ci_method = "cmh",
  conf_level = 0.99
)

# all observations as non-responders
t_el_proportion_diff(
  rep(FALSE, nrow(ADRS)),
  col_by = ADRS$ARMCD,
  diff_ci_method = "wald"
)
}
\seealso{
\code{\link{s_proportion_diff}},
  \code{\link{t_el_proportion}}, \code{\link{t_el_test_proportion_diff}},
  \code{\link{t_binary_outcome}}
}
