% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{t_el_test_proportion_diff}
\alias{t_el_test_proportion_diff}
\title{Table for testing difference in proportions}
\usage{
t_el_test_proportion_diff(
  rsp,
  col_by,
  strata_data = NULL,
  test = c("chisq", "cmh")
)
}
\arguments{
\item{rsp}{(\code{logical} vector)\cr
\code{TRUE} if subject is a responder \code{FALSE} otherwise}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{strata_data}{(\code{data.frame})\cr
Used for stratification factors. If \code{NULL}, no stratified
analysis is performed.}

\item{test}{one of (\code{"chisq", "cmh"})\cr
Specifies the test used to calculate the p-value.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
This function will tabulate the p-value for testing the null hypothesis of no
difference between the proportion of events in the reference group as compared to
the proportion of events in each of the remaining group levels in \code{col_by}.
}
\details{
The following options are supported for testing difference of proportions:
\itemize{
  \item{Option \code{chisq} performs Chi-Squared test. Internally calls \code{\link[stats]{prop.test}}.}
  \item{Option \code{cmh} performs stratified Cochran-Mantel-Haenszel test.
  Internally calls \code{\link[stats]{mantelhaen.test}}.}
}
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADRS <- radrs(cached = TRUE) \%>\%
  dplyr::filter(PARAMCD == "BESRSPI")

t_el_test_proportion_diff(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  test = "chisq"
)

t_el_test_proportion_diff(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  strata_data = ADRS[, c("STRATA1", "STRATA2")],
  test = "cmh"
)

# all observations as non-responders
t_el_test_proportion_diff(
  rep(FALSE, nrow(ADRS)),
  col_by = ADRS$ARMCD,
  test = "chisq"
)

}
\seealso{
\code{\link{s_test_proportion_diff}}
  \code{\link{t_el_proportion}}, \code{\link{t_el_proportion_diff}},
  \code{\link{t_binary_outcome}}
}
