% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{t_event_count}
\alias{t_event_count}
\title{Count events}
\usage{
t_event_count(id, col_by, col_N)
}
\description{
Count events
}
\examples{
tbl <- t_event_count(
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  col_N = tapply(ASL$ARM, ASL$ARM, length)
)

}
