% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_events_patyear.R
\name{t_events_patyear}
\alias{t_events_patyear}
\title{event rate adjusted per person-years at Risk}
\usage{
t_events_patyear(
  events,
  time_in_years,
  col_by,
  col_N = NULL,
  total = NULL,
  conf_level = 0.95,
  conf_method = c("normal", "exact", "byar"),
  lambda = 100,
  event_type = "events"
)
}
\arguments{
\item{events}{(\code{logical} vector) \code{TRUE} if record is an event}

\item{time_in_years}{(\code{numeric} vector) time at risk in year}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{(\code{character} value)\cr
Denotes column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{conf_level}{confidence level of the interval.}

\item{conf_method}{confidence method (\code{"normal", "exact", "byar"}),
see details of \code{\link{ci_patyear}}}

\item{lambda}{a numeric rate factor, e.g. 100 for 100 person-years}

\item{event_type}{named (\code{character} value) to specify the type of event that is summarized}
}
\description{
Table for event rate adjusted for Patient-Years
}
\examples{
library(dplyr)
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADAETTE <- radaette(cached = TRUE)
ANL <- ADAETTE \%>\%
   dplyr::filter(PARAM == "Time to first occurrence of any adverse event")
tbl1 <- t_events_patyear(
   events = !(ANL$CNSR),
   time_in_years = ANL$AVAL,
   col_by = ANL$ARM,
   col_N = table(ADSL$ARM),
   conf_level = 0.95,
   conf_method = "normal",
   lambda = 100
)

ANL <- ANL \%>\% dplyr::filter(SEX \%in\% c("F", "M"))
tbl2 <- t_events_patyear(
   events = !(ANL$CNSR),
   time_in_years = ANL$AVAL,
   col_by = ANL$ARM,
   col_N = NULL,
   total = "Total Patients",
   conf_level = 0.9,
   conf_method = "byar",
   lambda = 1000,
   event_type = c("AE" = "adverse event")
)


}
