% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{t_events_per_term_grade_id}
\alias{t_events_per_term_grade_id}
\title{Adverse Events Table by Highest NCI CTCAE Grade}
\usage{
t_events_per_term_grade_id(terms, id, grade, col_by, col_N,
  total = "All Patients", grade_levels = 1:5)
}
\arguments{
\item{terms}{term information as character or factor vector or dataframe, however factor
levels are not repeated by class, only terms with count > 1 are listed per
class. Currently \code{terms} can only be a vector or dataframe with 1/2 column.
\code{var_relabel} is used as the character string used as a label in the column header
for each term.}

\item{id}{unique subject identifier. If a particular subject has no adverse
event then that information needs to be added to the \code{col_N} argument.}

\item{grade}{grade of adverse event as numeric. \code{var_relabel} is used as the character 
string used as a label in the column header for each grade.}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing. See 'Examples'.}

\item{col_N}{numeric vector with information of the number of patients in the
levels of \code{col_by}. This is useful if there are patients that have no
adverse events can be accounted for with this argument.}

\item{total}{character string that will be used as a label for a column with 
pooled total population, default is "All Patients". If the levels of col_by are 
the only columns of interest then total should be \code{NULL}}

\item{grade_levels}{numeric, ordered values of possible of grades in a form
of \code{x:y}, default is \code{1:5}.}
}
\value{
\code{rtable} object
}
\description{
\code{t_events_per_term_grade_id} returns adverse events sorted by highest NCI (National Cancer
 Institute) CTCAE (common terminology criteria for adverse avents) grade.
}
\details{
\code{t_events_per_term_grade_id} counts patients according to adverse events (AEs) of greatest
 intensity for system organ class (SOC) and overall rows and includes 
 percentages based on the total number of patients in the column heading 
 (i.e. "N=nnn"). If the intention is to use patients number from subject level
 dataset as N for percentage calculation then adeverse events dataset should
 be left joined to subject level dataset and the \code{col_by} variable should
 be dropped from adverse events dataset, see the example. Otherwise, N will be
 derived using adverse events dataset. At the preferred term (PT) level,
 multiple events within a patient of the same PT are counted once using the
 greatest intensity reported. 

\code{t_events_per_term_grade_id} removes any non-complete records, e.g. if class or term are 
 missing. If the intent is to preserve such records, then impute missing 
 values before using \code{t_events_per_term_grade_id}.
     
\code{t_events_per_term_grade_id} orders data by "All Patients" column from the most commonly
 reported SOC to the least frequent one. Within SOC, it sorts by decreasing
 frequency of PT. It brakes ties using SOC/PT names in alphabetical order.   

\code{t_events_per_term_grade_id} fills in \code{col_by} and \code{grade} with \code{0} value 
in case there was no AEs reported for particular \code{col_by} and/or 
\code{grade} category. Use \code{grade_levels} to modify the range of existing
grades. If data does not have any records with \code{grade} 5 and the intent 
is to show only grades 1-4 rows then use \code{grade_levels = 1:4}.

This is an equivalent of the STREAM output \code{\%stream_t_summary(templates = aet04)}
  (\url{http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_aet04.html})

This is an equivalent of the STREAM 1.17 output \code{\%stream_t_events_bygrade(templates = aet04)}
  (\url{<https://rochewiki.roche.com/confluence/pages/viewpage.action?pageId=294027501>})
}
\examples{

library(dplyr)
library(random.cdisc.data)

ASL <- rasl(10, seed = 1)
AAE <- raae(ASL, 4, seed = 2)

ANL <- left_join(AAE, ASL \%>\% select(USUBJID, STUDYID, ARM), by = c("USUBJID", "STUDYID")) \%>\%
  var_relabel(
    AEBODSYS = 'MedDRA System Organ Class',
    AEDECOD = 'MedDRA Preferred Term',
    AETOXGR = 'GRADE'
)

t_events_per_term_grade_id(
  terms = ANL$AEDECOD,
  id = ANL$USUBJID,
  grade = ANL$AETOXGR,
  col_by = ANL$ARM,
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients",
  grade_levels = 1:5
)

t_events_per_term_grade_id(
  terms = ANL \%>\% select(AEBODSYS, AEDECOD),
  id = ANL$USUBJID,
  grade = ANL$AETOXGR,
  col_by = ANL$ARM,
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients",
  grade_levels = 1:5
)


}
\author{
Edgar Manukyan (manukyae) \email{edgar.manukyan@roche.com}

Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}
}
