% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{t_events_per_term_id}
\alias{t_events_per_term_id}
\title{Adverse Events by System Organ Class and Preferred Term}
\usage{
t_events_per_term_id(terms, id, col_by, col_N, total = "All Patients")
}
\arguments{
\item{terms}{term information as character or factor dataframe, however factor
levels are not repeated by class, only terms with count > 1 are listed per
class. Currently \code{terms} can only be a dataframe with 2 columns.
\code{var_relabel} is used as the character string used as a label in the column header
for each term.}

\item{id}{unique subject identifier. If a particular subject has no adverse
event then that information needs to be added to the \code{col_N} argument.}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing. See 'Examples'.}

\item{col_N}{numeric vector with information of the number of patients in the
levels of \code{col_by}. This is useful if there are patients that have no
adverse events can be accounted for with this argument.}

\item{total}{character string that will be used as a label for a column with 
pooled total population, default is "All Patients". If the levels of col_by are 
the only columns of interest then total should be \code{NULL}}
}
\value{
\code{rtable} object
}
\description{
\code{t_events_per_term_id} returns adverse events summary table that corresponds to STREAM template AET02
}
\details{
This is an equivalent of the STREAM output \code{\%stream_t_summary(templates = aet02)}
  (\url{http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_aet02.html})

This is an equivalent of the STREAM 1.17 output \code{\%stream_t_events_basic(templates = aet02)}
  (\url{https://rochewiki.roche.com/confluence/pages/viewpage.action?pageId=294027342})
}
\examples{
# Simple example
library(tibble)
library(dplyr)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = factor(paste("ARM", LETTERS[rep(c(1,2), c(3,7))]))
)


ae_lookup <- tribble(
  ~CLASS,         ~TERM,   ~GRADE,
  "cl A",   "trm A_1/2",        1,
  "cl A",   "trm A_2/2",        2,  
  "cl B",   "trm B_1/3",        2,
  "cl B",   "trm B_2/3",        3,
  "cl B",   "trm B_3/3",        1,
  "cl C",   "trm C_1/1",        1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2,2,2,3,3,4,4,4,4,5,6,6,7,7)]
  ),
  ae_lookup[c(1,1,2,6,4,2,2,3,4,2,1,5,4,6),]
)

ANL <- left_join(AAE, ASL, by = "USUBJID")

terms <- data.frame(class = ANL$CLASS, term = ANL$TERM, stringsAsFactors = FALSE)
var_labels(terms) <- c('MedDRA System Organ Class', 'MedDRA Preferred Term')

tbl <- t_events_per_term_id(
  terms = terms,
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients"
)

tbl

# Example using dummy data
library(random.cdisc.data)
library(dplyr)

ASL <- radam("ASL", N = 10)
AAE <- radam("AAE", ADSL = ASL)

ANL <- left_join(AAE, ASL \%>\% select(USUBJID, STUDYID, ARM), by = c("STUDYID", "USUBJID"))

terms <- data.frame(class = ANL$AEBODSYS, term = ANL$AEDECOD, stringsAsFactors = FALSE)
var_labels(terms) <- c('MedDRA System Organ Class', 'MedDRA Preferred Term')

tbl <- t_events_per_term_id(terms = terms,
            id = ANL$USUBJID,
            col_by = factor(ANL$ARM),
            total = "All Patients"
            )

tbl

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
