% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{t_events_summary}
\alias{t_events_summary}
\title{Summary table for events}
\usage{
t_events_summary(x, id, col_by, col_N = tapply(col_by, col_by, length),
  total_label = NULL, at_least_one_label = NULL)
}
\arguments{
\item{x}{a dataframe}

\item{id}{unique subject identifier. If a particular subject has no adverse
event then that information needs to be added to the \code{col_N} argument.}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing. See 'Examples'.}

\item{col_N}{associated number of elements per level of \code{col_by}}

\item{total_label}{character string that will be used as a label in the row for
the total event count. If this is \code{NULL} then this row will be removed.
The default here is \code{NULL}.}

\item{at_least_one_label}{character string that will be used as a label in the row for
the total number with at least one event. If this is \code{NULL} then this row 
will be removed. The default here is \code{NULL}.}
}
\value{
an rtable
}
\description{
\code{t_events_summary} counts the number of unique elements per cell.
}
\examples{

x <- data.frame(x = c("t1", "t1", "t2", "t2", "t2"))
var_labels(x) <- "MedDRA Preferred Term"

t_events_summary(
 x = x,
 id = paste("id", c(1, 4, 2, 3, 3), sep = "-"),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 total_label = "Total number of events",
 at_least_one_label = "Total number of patients with at least one adverse event"
)

\dontrun{
#throws an error because x is not a dataframe
t_events_summary(
 x = c("t1", "t1", "t2", "t2", "t2"),
 id = paste("id", c(1, 4, 2, 3, 3), sep = "-"),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 total_label = "Total number of events",
 at_least_one_label = "Total number of patients with at least one adverse event"
)
}

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
