% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_rsp.R
\name{t_forest_rsp}
\alias{t_forest_rsp}
\title{Response Forest Plot Table}
\usage{
t_forest_rsp(response, event, arm, group_data, covariates = NULL)
}
\arguments{
\item{response}{Tumor Response data}

\item{event}{is boolean, \code{TRUE} if responder, \code{FALSE} if
non-responder}

\item{arm}{vector with arm information}

\item{group_data}{data frame with one column per sub-group variable}

\item{covariates}{set to NULL; currently not available for multivariate
survival analysis}
}
\description{
Response Forest Plot Table
}
\details{
Logistic model is used for odds ratio calculation
}
\examples{

library(random.cdisc.data)
ASL <- radam("ASL")
ARS <- radam("ARS", ADSL = ASL)

ARS_f <- ARS \%>\% filter(PARAMCD == "OVRSPI") 

ASL_f <- right_join(ASL \%>\% select(USUBJID, STUDYID, SEX, RACE, ARM),
                        ARS_f \%>\% select(USUBJID, STUDYID))

tbl <- forest_rsp(
  response = ARS_f$AVAL,
  event = ARS_f$AVALC \%in\% c("CR","PR"),
  arm = ASL_f$ARM, 
  group_data = ASL_f \%>\% select("SEX", "RACE")
)

tbl

\dontrun{
library(atezo.data)
library(dplyr) 
library(grid)
library(teal.oncology)
library(forcats)
devtools::install_github("Roche/rtables")
'\%needs\%' <- teal.oncology:::'\%needs\%'
ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)
ASL$temp <- c(rep("A", 200), rep("B", 250), rep(NA,1202-450))

ASL_f <- subset(ASL, 
                subset = (ITTWTFL == "N" & ARM \%in\% c("DUMMY A", "DUMMY C")), 
                select = c("USUBJID", "STUDYID", "ARM", "ARMCD", "SEX", "MLIVER", "TCICLVL2", "AGE4CAT", "RACE", "temp") )
ARS_f <- subset(ARS, 
                subset = (PARAMCD == "OVRSPI"),
                select = c("STUDYID", "USUBJID", "PARAMCD", "PARAM", "AVAL", "AVALC"))
ANL <- merge(ASL_f, ARS_f, by = c("STUDYID", "USUBJID"), all.x = TRUE, all.y = FALSE)

group_data <- subset(ASL_f, select = c("USUBJID", "STUDYID", "SEX", "MLIVER", "TCICLVL2", "AGE4CAT", "RACE"))

for(x in names(group_data)){
  attr(group_data[[x]], "label") <- attr(ASL[[x]], "label")
}   

names(group_data) <- labels_over_names(group_data)

head(group_data)

arm <- fct_relevel(ASL_f$ARM, "DUMMY C")

tbl <- forest_rsp(
          response = ANL$AVAL,
          event = ANL$AVALC \%in\% c("CR","PR"),
          arm = arm, 
          group_data = group_data[, -c(1,2), drop=FALSE]
)
Viewer(tbl)

}


  
}
\author{
Yuyao Song (songy24), \email{yuyao.song@roche.com}
}
