% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_tte.R
\name{t_forest_tte}
\alias{t_forest_tte}
\title{Time-to-event Forest Plot Table}
\usage{
t_forest_tte(tte, is_event, col_by, group_data = NULL, strata_data = NULL,
  total = "ALL", time_unit = "month", na.omit.group = TRUE)
}
\arguments{
\item{tte}{time to event data}

\item{is_event}{is boolean, \code{TRUE} if event, \code{FALSE} if \code{tte}
is censored}

\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group}

\item{group_data}{data frame with one column per grouping}
}
\description{
Time-to-event Forest Plot Table
}
\details{
Cox PH model is used for hazard ratio calculation
}
\examples{

devtools::install_github("Rpackages/random.cdisc.data", host='https://github.roche.com/api/v3')
library(random.cdisc.data)
library(dplyr)

ASL <- radam("ASL")
ATE <- radam("ATE", ADSL = ASL)

ATE_f <- ATE \%>\% filter(PARAMCD == "OS") 

ANL <- merge(ASL \%>\% select(USUBJID, STUDYID, SEX, RACE, ARM), ATE_f)

tbl <- t_forest_tte(
  tte = ANL$AVAL,
  is_event = ANL$CNSR == 0,
  col_by = factor(ANL$ARM), 
  group_data = as.data.frame(lapply(ANL[, c("SEX", "RACE")], as.factor))
)

Viewer(tbl)

}
\author{
Yuyao Song (songy24), \email{yuyao.song@roche.com}
}
