% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_km.R
\name{t_km}
\alias{t_km}
\title{Prepare K-M model annotation data with rtable format}
\usage{
t_km(fit_km)
}
\arguments{
\item{fit_km}{a class \code{\link{survfit}} object.}
}
\description{
An \code{\link[rtables]{rtable}} format of KM model data for further
annotation on top of Kaplan-Meier grob
}
\examples{

OS <- data.frame(AVAL = abs(rnorm(200)), 
                 CNSR = sample(c(0, 1), 200, TRUE), 
                 ARM = sample(LETTERS[1:3], 200, TRUE),
                 SEX = sample(c("M","F"), 200, TRUE),
                 RACE = sample(c("AA", "BB", "CC"), 200, TRUE),
                 ECOG = sample(c(0, 1), 200, TRUE))
fit_km <- survfit(Surv(AVAL, 1-CNSR) ~ ARM, data = OS, conf.type = "plain")
tbl <- t_km(fit_km)
tbl

}
\author{
Heng Wang (wangh107) \email{wang.heng@gene.com}
}
