% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_logistic.R
\name{t_logistic}
\alias{t_logistic}
\title{Multi-variable logistic regression table}
\usage{
t_logistic(formula, data, increments = NULL, conf_level = 0.95)
}
\arguments{
\item{formula}{a {\code{\link{glm}}} formula \cr
The formula object can be all main effect model formula, and a formula with one two-way interaction.}

\item{data}{a \code{data.frame} with all the variables specified in
\code{formula}.}

\item{increments}{(\code{named list})\cr
Used to specify numeric values of continuous variables in {\code{glm}} model {\code{formula}}
which interact with other variables. This is used to calculate the odds ratio when comparing
the other interaction variable effect. For example, for a model with ARM and AGE interaction,
{\code{increments = list(AGE = c(18, 65))}} will enable calculation of odds ratios of
comparison ARM vs. reference ARM at AGE = 18 and AGE = 65. If {\code{increments = NULL}}, then
default AGE value is ceiling of median.}

\item{conf_level}{(\code{numeric} value)\cr
Confidence level for Wald odds ratio confidence interval.}
}
\description{
Logistic regression for binary outcome with categorical/continuous covariates in model statement.
For each covariate category (if categorical) or specified values (if continuous), present degrees of freedom,
regression parameter estimate and standard error (SE) relative to reference group or category.
Report odds ratios for each covariate category or specified values and corresponding Wald
confidence intervals as default but allow user to specify other confidence levels.
Report p-value for Wald chi-square test of the null hypothesis that covariate has no effect on
response in model containing all specified covariates.
Allow option to include one two-way interaction and present similar output for
each interaction degree of freedom.
Note: For \code{glm} formula, the variable names need to be standard dataframe column name without
special characters. The big N is the total number of observations for complete cases.
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ADSL <- ADSL \%>\% dplyr::filter(SEX \%in\% c("F", "M"))
ADRS <- radrs(ADSL, seed = 2)

ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI") \%>\%
  dplyr::mutate(Response = ifelse(AVALC \%in\% c("PR", "CR"), 1, 0))
var_labels(ADRS_f) <- c(var_labels(ADRS), "Response" = "Tumor Response")

tbl <- t_logistic(
  formula = Response ~ ARM + AGE + SEX,
  data = ADRS_f )

tbl # or Viewer(tbl) for a html view

tbl2 <- t_logistic(
  formula = Response ~ ARM + AGE + BMRKR2 + ARM*BMRKR2,
  data = ADRS_f
)

tbl2

tbl3 <- t_logistic(
  formula = Response ~ ARM + AGE + BMRKR1 + ARM*BMRKR1,
  data = ADRS_f,
  increments = list("BMRKR1" = c(5, 10)),
  conf_level = 0.9
)
tbl3

tbl4 <- t_logistic(
  formula = Response ~ ARM + AGE + BMRKR1 + AGE*BMRKR1,
  data = ADRS_f,
  increments = list("AGE" = c(20, 65), "BMRKR1" = c(5, 10)),
  conf_level = 0.9
)
tbl4

}
