% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae_ctc.R
\name{t_max_grade_per_id}
\alias{t_max_grade_per_id}
\title{Count Max per Id Per Once}
\usage{
t_max_grade_per_id(grade, id, col_by, col_N = NULL, grade_levels = NULL,
  any_grade = "-Any Grade-")
}
\arguments{
\item{id}{character or factor}

\item{col_by}{(inherit)}

\item{col_N}{divider for percentage based on level of col_by}

\item{x}{a numeric vector}

\item{levels}{domain of x}

\item{any}{add row that counts any occurance}
}
\description{
E.g. as used in count ...
}
\examples{
id <- c(1,1,2,2,3,3,3,4)
t_max_grade_per_id(
  grade =  c(1,2,3,2,1,1,2,3),
  id = id,
  col_by = factor(LETTERS[id]),
  col_N = c(4,3,5,3)
)

t_max_grade_per_id(
  grade =  c(1,2,3,2,1,1,2,3),
  id = id,
  col_by = factor(LETTERS[id])
)

\dontrun{
t_max_grade_per_id(
  grade =  c(1,2,3),
  id = c(1,2,2),
  col_by = factor(LETTERS[1:3])
)
}
}
