% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_mmrm.R
\name{t_mmrm_fixed}
\alias{t_mmrm_fixed}
\title{Tabulate the fixed effect estimates of an MMRM fit.}
\usage{
t_mmrm_fixed(object, format = "xx.xxxx")
}
\arguments{
\item{object}{the MMRM model result produced by \code{\link{s_mmrm}}.}

\item{format}{\code{rtables} format for the numbers other than degrees of freedom and p-values
(default is \code{"xx.xxxx"}).}
}
\value{
\code{rtable} object with the fixed effect estimates, standard errors,
  degrees of freedom, t-statistics, and p-values.
}
\description{
Tabulate the fixed effect estimates of an MMRM fit.
}
\examples{
library(dplyr)
library(random.cdisc.data)

adsl <- radsl(cached = TRUE)
adqs <- radqs(cached = TRUE)
adqs_f <- adqs \%>\%
  dplyr::filter(PARAMCD=="FKSI-FWB" & !AVISIT \%in\% c("BASELINE")) \%>\%
  droplevels() \%>\%
  dplyr::mutate(ARM = factor(ARM, levels = c("B: Placebo", "A: Drug X", "C: Combination"))) \%>\%
  dplyr::mutate(AVISITN = rank(AVISITN) \%>\% as.factor() \%>\% as.numeric() \%>\% as.factor())

mmrm_results <- s_mmrm(
  vars = list(
    response = "AVAL",
    covariates = c("STRATA1", "BMRKR2"),
    id = "USUBJID",
    arm = "ARM",
    visit = "AVISIT"
  ),
  data = adqs_f,
  cor_struct = "random-quadratic",
  optimizer = "nloptwrap_bobyqa"
)
t_mmrm_fixed(mmrm_results)
}
