% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_mmrm.R
\name{t_mmrm_lsmeans}
\alias{t_mmrm_lsmeans}
\title{Tabulate the LS means of an MMRM model.}
\usage{
t_mmrm_lsmeans(
  object,
  col_N,
  show_relative = c("reduction", "increase", "none"),
  table_tree = TRUE
)
}
\arguments{
\item{object}{the MMRM model result produced by \code{\link{s_mmrm}}.}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{show_relative}{should the "reduction" (\code{control - treatment}, default) or the "increase"
(\code{treatment - control}) be shown for the relative change from baseline? It is also possible
to suppress this row in the table ("none").}

\item{table_tree}{(\code{logical} value)\cr
If \code{FALSE} then a single \code{rtable} gets returned, if \code{TRUE} a
\code{\link{node}} object gets returned with the elementary \code{rtable} objects.}
}
\value{
\code{rtable} object or table tree, depending on the the `table_tree` argument
}
\description{
This function summarizes adjusted \code{lsmeans} and standard error, as well as conducts
comparisons between groups' adjusted \code{lsmeans}, where the first level of the group
is the reference level.
}
\examples{
library(dplyr)
library(random.cdisc.data)

adsl <- radsl(cached = TRUE)
adqs <- radqs(cached = TRUE)
adqs_f <- adqs \%>\%
  dplyr::filter(PARAMCD=="FKSI-FWB" & !AVISIT \%in\% c("BASELINE")) \%>\%
  droplevels() \%>\%
  dplyr::mutate(ARM = factor(ARM, levels = c("B: Placebo", "A: Drug X", "C: Combination"))) \%>\%
  dplyr::mutate(AVISITN = rank(AVISITN) \%>\% as.factor() \%>\% as.numeric() \%>\% as.factor())

\dontrun{
# Chosen optimizer 'nloptwrap_bobyqa' led to problems during model fit
mmrm_results <- s_mmrm(
  vars = list(
    response = "AVAL",
    covariates = c("STRATA1", "BMRKR2"),
    id = "USUBJID",
    arm = "ARM",
    visit = "AVISIT"
  ),
  data = adqs_f,
  cor_struct = "unstructured",
  weights_emmeans = "proportional",
  optimizer = "nloptwrap_neldermead"  # Only to speed up this example.
)
t_mmrm_lsmeans(
  mmrm_results,
  col_N = table(adsl$ARM),
  show_relative = "increase",
  table_tree = FALSE
)
}
}
\seealso{
\code{\link{s_mmrm}}
}
