% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{t_one_count_per_id}
\alias{t_one_count_per_id}
\title{Count unique id}
\usage{
t_one_count_per_id(id, col_by, col_N)
}
\description{
Count unique id
}
\examples{
tbl <- t_one_count_per_id(
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  col_N = tapply(ASL$ARM, ASL$ARM, length)
)

}
