% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarize_by_visit.R
\name{t_summarize_by_visit}
\alias{t_summarize_by_visit}
\title{Summarize Information by Visit}
\usage{
t_summarize_by_visit(x, visit, col_by, fitst_as_blank = TRUE)
}
\description{
Summarize Information by Visit
}
\examples{

ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")   
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == 1] <- NA

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)



t_summarize_by_visit(x = ANL$AVAL, visit = ANL$VISIT, col_by = ANL$ARM)

t_summarize_by_visit(x = ANL$CHG, ANL$VISIT, ANL$ARM)

t_summarize_by_visit(x = c(ANL$AVAL, ANL$CHG), visit = c(ANL$VISIT, ANL$VISIT), factor(rep(ANL$ARM, ANL$ARM), levels=<...>)

}
