% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarize_byvisit.R
\name{t_summarize_byvisit}
\alias{t_summarize_byvisit}
\title{Summary Table by Visit}
\usage{
t_summarize_byvisit(data, visit, col_by)
}
\arguments{
\item{data}{data frame with numerical variables to be summarized. If the 
variable has a \code{label} attribute then it will be used as the column 
name.}

\item{visit}{factor with visit names ordered by desired display order in the
stacked table.}

\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
This function summarizes test results or change from baseline statistics by 
visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.
}
\details{

}
\examples{

ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")   
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

t_summarize_byvisit(data = ANL[c("AVAL")], visit = ANL$VISIT, col_by = ANL$ARM)
t_summarize_byvisit(data = ANL[c("PCHG")], visit = ANL$VISIT, col_by = ANL$ARM)

# Add label to variable instead showing variable name
ANL <- var_relabel(ANL, AVAL = "Value at\\nVisit", CHG = "Change from\\nBaseline", PCHG = "Percent Change\\nfrom Baseline")
t_summarize_byvisit(data = ANL[c("AVAL", "CHG")], visit = ANL$VISIT, col_by = ANL$ARM)

}
\author{
Chendi Liao (liaoc10) \email{chendi.liao@roche.com}
}
