% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarize_variables.R
\name{t_summarize_factor}
\alias{t_summarize_factor}
\title{Summarize Categorical Data}
\usage{
t_summarize_factor(x, col_by, useNA = c("no", "ifany", "always"),
  perc_denominator = c("n", "N"))
}
\arguments{
\item{x}{numeric variable}

\item{col_by}{a \code{\link{factor}} of length \code{nrow(x)} that defines
which levels in \code{col_by} define a column. If data should not be split
into columns use the \code{\link{no_by}} function.}

\item{useNA}{boolean, if \code{TRUE} then \code{NA} values in \code{x} get
turned into a factor level \code{"NA"}, if \code{FALSE} then the \code{NA}
values in \code{x} get dropped.}

\item{perc_denominator}{either n or N for calculating the level associated
percentage.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
Tabulate the number on non-missing observations, number of per level and
percentage
}
\examples{

t_summarize_factor(iris$Species, iris$Species)


x <-  factor(c("a", "b", "a", "a",  NA,  NA,  NA, "b", "b"))
cb <- factor(c("E", "E", "F", "F", "E", "E", "F", "F"))

t_summarize_factor(x, cb)

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
