% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarize_variables.R
\name{t_summarize_variables}
\alias{t_summarize_variables}
\title{Table of Variable Summaries}
\usage{
t_summarize_variables(data, col_by, total = NULL)
}
\arguments{
\item{data}{data frame}

\item{col_by}{a vector of length \code{nrow(x)} that defines which levels in
\code{col_by} define a column. If data should not be split into columns use
the \code{\link{no_by}} function.}

\item{total}{if not \code{NULL} then it must be a string and an addition
column will be added with the overall summaries}
}
\description{
Similiar as the demographic table in STREAM
}
\details{
Give a detailed description of what this function does.
}
\examples{

library(tern)
library(random.cdisc.data)
library(forcats)

ASL <- radam("ASL")

# control the label
attr(ASL$BAGE, "label") <- "Baseline Age of patient"
attr(ASL$SEX, "label")

# control categorical order
ASL$SEX <- fct_relevel(ASL$SEX, "M", "F", "UNDEFINED")

# control arm order
ASL$ARM <- fct_relevel(ASL$ARM, "ARM B", "ARM A")

t_summarize_variables(ASL \%>\% select(SEX, BAGE), col_by = ASL$ARM, total = "All Patients")

t_summarize_variables(iris \%>\% select(-Species), col_by  = iris$Species)

t_summarize_variables(iris \%>\% select(-Species), col_by  = no_by("All Species"))

\dontrun{
library(tern)
library(atezo.data)
library(dplyr)

ASL <- asl(com.roche.cdt30019.go29436.re)
tbl_expected <- get_demographic_table(com.roche.cdt30019.go29436.re)
Viewer(tbl_expected)

# order factor levels if they exist
of <- function(x, ...) {
  lvls <- if (is.factor(x)) levels(x) else unique(x)
  olvls <- intersect(unlist(list(...)), lvls)
  factor(as.character(x), levels = c(olvls, setdiff(lvls, olvls)))
}
add_lbl <- function(lbl, x) {
  structure(x, label = lbl)
}
lbl <- function(x) {
  attr(x, "label")
}

ADSL <- ASL \%>\%
 filter(ITTWTFL == 'Y') \%>\%
 mutate(ARM = of(ARM, "DUMMY C", "DUMMY B", "DUMMY A")) \%>\%
 mutate(SEX = add_lbl(lbl(SEX), recode_factor(SEX, M = "MALE", F = "FEMALE")))

tbl <- t_summarize_variables(
   data = subset(ADSL, select = toupper(c("sex", "mliver", "tciclvl2", "bage", "age4cat",
     "ethnic", "race", "bwt", "tobhx", "hist", "EGFRMUT",
     "alkmut", "krasmut", "becog"))),
   col_by = ADSL$ARM,
   total = "All Patients"
)
Viewer(tbl)

compare_rtables(tbl, tbl_expected)

# if all is good then
tbl <- tbl_expected
compare_rtables(tbl, tbl_expected)

}

}
\author{
Adrian Waddell \email{adrian.waddell@roche.com} and Xiao Yu Mo (mox5), \email{xiao_yu.mo@roche.com}
}
