% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.character}
\alias{t_summary.character}
\title{Summarize Character Data}
\usage{
\method{t_summary}{character}(x, col_by, col_N = table(col_by), ...)
}
\arguments{
\item{x}{a character vector}

\item{col_by}{a \code{\link{factor}} of length \code{nrow(x)} that defines
which levels in \code{col_by} define a column. If data should not be split
into columns use the \code{\link{no_by}} function.}

\item{col_N}{The column total for each column. If specified then
\code{\link{col_N}()} can be used on the cell data in order to retrieve the
column total. If \code{NULL} then no header row for the column is
displayed.}

\item{...}{arguments passed on to \code{\link{t_summary.factor}}}
}
\description{
Currently treated as for factors
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
