% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.Date}
\alias{t_summary.Date}
\title{Summarize Date Data}
\usage{
\method{t_summary}{Date}(x, col_by, col_N = NULL, total = NULL, ...)
}
\arguments{
\item{x}{(\code{Date}) object}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{(\code{character} value)\cr
Denotes column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{...}{(optional)\cr
additional arguments passed on to methods.
\itemize{
\item \code{q_type} an integer between 1 and 9 selecting one of the nine
quantile algorithms as detailed in \code{\link[stats:quantile]{quantile(), type argument}}
\item \code{conf_level} the confidence level between 0 and 1 (default: 0.95)
}}
}
\description{
Tabulate the range of dates.
}
\examples{
today <- Sys.Date()
tenweeks <- seq(today, length.out=10, by="1 week")
t_summary(
  x = tenweeks,
  col_by = by_all("all"),
  col_N = length(tenweeks)
)

t_summary(
  x = tenweeks,
  col_by = factor(LETTERS[c(1,1,1,2,2,3,3,3,4,4)])
)

t_summary(
  x = tenweeks,
  col_by = factor(LETTERS[c(1,1,1,2,2,3,3,3,4,4)]),
  total = "All"
)
}
\seealso{
\code{\link{t_summary}},
  \code{\link{t_summary.data.frame}}, \code{\link{t_summary.factor}},
  \code{\link{t_summary.logical}}, \code{\link{t_summary.numeric}},
  \code{\link{t_summary.character}}, \code{\link{t_summary_by}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
