% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary_variants.R
\name{t_summary.list}
\alias{t_summary.list}
\title{Helper function for t_summary_by}
\usage{
\method{t_summary}{list}(x_list, col_by_list, col_N = NULL,
  total = NULL, ..., table_tree = FALSE)
}
\arguments{
\item{x_list}{list of x}

\item{col_by_list}{list of \code{col_by}, one for each item of \code{x_list}}

\item{col_N}{a vector with the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{string of column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{...}{arguments passed on to methods}

\item{table_tree}{logical, if \code{FALSE} then a single \code{rtable} gets returned, if \code{TRUE} a
\code{\link{node}} object gets returned with the elementary \code{rtable} objects.}
}
\description{
You can use \code{...} to add arguments like \code{total}
}
\details{
If \code{col_N} is \code{NULL}, it gets it by adding up the numbers from all items in the \code{col_by_list}
}
