% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.logical}
\alias{t_summary.logical}
\title{Summarize Boolean Data}
\usage{
\method{t_summary}{logical}(
  x,
  col_by,
  col_N = NULL,
  total = NULL,
  row_name_true = "TRUE",
  row_name_false = "FALSE",
  ...
)
}
\arguments{
\item{x}{(\code{logical} vector)}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{(\code{character} value)\cr
Denotes column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{row_name_true}{(\code{character} value)\cr
Contains row.name for TRUE summary}

\item{row_name_false}{(\code{character} value)
Contains row.name for FALSE summary}

\item{...}{(optional)\cr
additional arguments passed on to \code{\link{t_summary.factor}}}
}
\description{
Boolean data will be converted to factor
}
\examples{

t_summary(
 x = c(TRUE,FALSE,NA,TRUE,FALSE,FALSE,FALSE,TRUE),
 col_by = factor(LETTERS[c(1,1,1,2,2,3,3,2)])
)

library(random.cdisc.data)
ADSL <- cadsl

ADSL$AGE[1:10] <- NA

with(
  ADSL,
  t_summary(
    AGE > 50,
    ARM,
    useNA = "ifany"
  )
)

with(
  ADSL,
  t_summary(
    AGE > 50,
    ARM,
    total = "All",
    denominator = "N",
    useNA = "ifany",
    row_name_true = "Baseline Age > 50",
    row_name_false = "Baseline Age <= 50"
  )
)
}
\seealso{
\code{\link{t_summary}},
  \code{\link{t_summary.numeric}}, \code{\link{t_summary.factor}},
  \code{\link{t_summary.data.frame}}, \code{\link{t_summary.Date}},
  \code{\link{t_summary.character}}, \code{\link{t_summary_by}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
