% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary_variants.R
\name{t_summary_true}
\alias{t_summary_true}
\title{Summarize the TRUE Counts}
\usage{
t_summary_true(x, col_by, col_N = NULL, total = NULL,
  row_name = deparse(substitute(x)), denominator = c("N", "n", "omit"))
}
\arguments{
\item{x}{factor variable}

\item{col_by}{logical factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default
\code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.}

\item{total}{string of column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{row_name}{name of row}

\item{denominator}{either n, N or omit. n and N are for calculating the level associated
percentage. With option N, the reference population from \code{col_N} is used as the denominator.
With option n, the number of non-missing records from \code{x} is used as the denominator.
If \code{omit} is chosen the percentage is omitted.}
}
\description{
Note that for this function the default \code{denominator = "N"} and not \code{n} as for the other \code{t_summary}
functions.
}
\examples{

with(iris, t_summary_true(Sepal.Length > 5.5, col_by = Species, total = "All Flowers"))
}
