% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_tte.R
\name{t_tte}
\alias{t_tte}
\title{Time to Event Table}
\usage{
t_tte(tte, is_event, event_descr, col_by, strata_data, time_points,
  time_unit = "month", ties = "exact")
}
\arguments{
\item{is_event}{boolean with \code{TRUE} if event and \code{FALSE} if censored}

\item{time_to_event}{time to event data}
}
\value{
a named list with one element per row where each element contains the
  data for the row
}
\description{
This is the description of the function
}
\examples{

library(random.cdisc.data)
library(forcats)

ASL <- radam("ASL")
ATE <- merge(ASL, radam("ATE", ADSL = ASL))

ATE_f <- subset(ATE, PARAMCD == "OS")


tbl <- t_tte(
  tte = ATE_f$AVAL,
  is_event = ATE_f$CNSR == 0,
  event_descr = factor(ATE_f$EVNTDESC),
  col_by = factor(ATE_f$ARM),
  strata_data = ATE_f[, c('SEX', 'RACE')],
  time_points = 6,
  time_unit = "month"
)

tbl

}
