% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot.R
\name{tab.list}
\alias{tab.list}
\title{create the time to event table}
\usage{
tab.list(adsl, data, groupvar, outcome, arm.ref, arm.comp)
}
\description{
create the time to event table
}
\examples{


\dontrun{
library(atezo.data)

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

tab.list(
  adsl = ASL,
  data = ATE,
  groupvar = c("BAGE","BECOG", "SEX"),
  outcome = "Overall Survival",
  arm.ref = "DUMMY A",
  arm.comp = c("DUMMY B", "DUMMY C")
)
}



}
