% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_trees.R
\name{table_tree}
\alias{table_tree}
\title{Create a tree object with rtable objects as leaf}
\usage{
table_tree(x)
}
\arguments{
\item{x}{a list or nested list of rtables, if \code{x} is an rtable then the rtable is returned}
}
\description{
Converts a list of rtables or a nested list of rtables into a \code{table_tree} object. Note that \code{NULL} objects
are used interchangably with rtable objects.
}
\details{
\code{table_tree} does not change the underlying data structure of the argument \code{x} other than adding
  the \code{"table_tree"} class attribute. The \code{table_tree} function also checks if the leafs are either
  \code{NULL} objects or objects of class \code{rtables}
}
\examples{

tbl <- rtable(header = 1:3, rrowl("row", letters[1:3]))

nl1 <- list(tbl1 = tbl, tbl2 = tbl, tbl3 = NULL)
tt1 <- table_tree(nl1)
tt1
summary(tt1)

nl2 <- list(A = nl1, B = nl1)
tt2 <- table_tree(nl2)
tt2
summary(tt2)

nl3 <- list(X = nl2, Y = nl2, Z = nl2)
tt3 <- table_tree(nl3)
summary(tt3)
\dontrun{
nl4 <- c(nl1, list(iris))
table_tree(nl4)
}

}
