% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_pairwise.R
\name{tabulate_pairwise}
\alias{tabulate_pairwise}
\title{Calculate statistics Pairwise with reference level}
\usage{
tabulate_pairwise(x, col_by, FUN, row.name, format = NULL, indent = 0)
}
\arguments{
\item{x}{a vector}

\item{col_by}{a factor first level is taken as reference level}

\item{FUN}{a function with two arguments, first argument is the subset of x
and the second argument is a factor with two levels from col_by}
}
\value{
an \code{rtable}
}
\description{
Calculate statistics Pairwise with reference level
}
\examples{
tabulate_pairwise(
  x = iris$Sepal.Length,
  col_by = iris$Species,
  FUN = function(xi, col_by_i) diff(tapply(xi, col_by_i, mean)),
  row.name = "diff mean"
)

tabulate_pairwise(
   x = data.frame(
        val = sample(c("X1", "X2"), 100, replace = TRUE),
        strata = sample(c("A", "B"), 100, replace = TRUE)
   ),
  col_by = factor(sample(c("BY1", "BY2", "BY3"), 100, replace = TRUE)),
  FUN = function(xi, col_by_i) {
            mhp <- mantelhaen.test(xi$val, col_by_i, xi$strata)
            mhp$p.value
            },
  row.name = "MH pvalue: stratified"
)

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
