% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_rbmi.R
\name{tabulate_rbmi}
\alias{tabulate_rbmi}
\alias{h_tidy_pool}
\alias{tidy.pool}
\alias{s_rbmi_lsmeans}
\alias{a_rbmi_lsmeans}
\alias{summarize_rbmi}
\title{Tabulation of RBMI Results}
\usage{
h_tidy_pool(x)

\method{tidy}{pool}(x)

s_rbmi_lsmeans(df, .in_ref_col, show_relative = c("reduction", "increase"))

a_rbmi_lsmeans(df, .in_ref_col, show_relative = c("reduction", "increase"))

summarize_rbmi(
  lyt,
  ...,
  table_names = "rbmi_summary",
  .stats = NULL,
  .formats = NULL,
  .indent_mods = NULL,
  .labels = NULL
)
}
\arguments{
\item{x}{(\code{list})}

\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}

\item{show_relative}{should the "reduction" (\code{control - treatment}, default) or the "increase"
(\code{treatment - control}) be shown for the relative change from baseline?}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{...}{additional arguments for the lower level functions.}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}
}
\value{
A dataframe
}
\description{
These functions can be used to produce tables from a fitted RBMI produced
}
\section{Functions}{
\itemize{
\item \code{h_tidy_pool}: Helper function to produce data frame with results of rbmi::pool
for a single visit

\item \code{tidy.pool}: Helper method (for \code{\link[broom:reexports]{broom::tidy()}}) to prepare a data frame from an
\code{pool} rbmi object containing the LS means and contrasts and multiple visits

\item \code{s_rbmi_lsmeans}: Statistics function which is extracting estimates from a tidied LS means
data frame.

\item \code{a_rbmi_lsmeans}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{summarize_rbmi}: Analyze function for tabulating LS means estimates from tidied
rbmi \code{pool} results.
}}

\examples{
\dontrun{
s_rbmi_lsmeans(df[8, ], .in_ref_col = FALSE)
}

\dontrun{
library(rtables)
library(dplyr)
library(broom)
library(rbmi)

path <- "~/NEST/rbmi_rds-export"
pool_obj <- readRDS(file.path(path, "pool_obj.RDS"))

h_tidy_pool(pool_obj$pars[1:3])
df <- tidy(pool_obj)
df

s_rbmi_lsmeans(df[2, ], .in_ref_col = FALSE)

afun <- make_afun(a_rbmi_lsmeans)
afun(df[2, ], .in_ref_col = FALSE)

basic_table() \%>\%
  split_cols_by("group", ref_group = levels(df$group)[1]) \%>\%
  split_rows_by("visit", split_label = "Visit", label_pos = "topleft") \%>\%
  summarize_rbmi() \%>\%
  build_table(df)
}

}
