% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{teal.oncology-package}
\alias{teal.oncology}
\alias{teal.oncology-package}
\title{Teal Modules useful in Oncology}
\description{
Teal Modules useful in Oncology
}
\author{
\strong{Maintainer}: Adrian Waddell \email{adrian.waddell@roche.com}

Authors:
\itemize{
  \item Heng Wang
  \item Mark Rothe
  \item Yuyao Song
  \item Chendi Liao
  \item Xiao Mu Yo
}

}
