% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{tidy.summary.coxph}
\alias{tidy.summary.coxph}
\title{Custom tidy method for \code{\link[survival:coxph]{survival::coxph()}} summary results.}
\usage{
\method{tidy}{summary.coxph}(x, ...)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{...}{additional arguments for the lower level functions.}
}
\description{
Tidy the \code{\link[survival:coxph]{survival::coxph()}} results into a \code{tibble} to extract model results.
}
\examples{
library(survival)
set.seed(1, kind = "Mersenne-Twister")
dta_bladder <- with(
  data = bladder[bladder$enum < 5, ],
  data.frame(
    time = stop,
    status = event,
    armcd = as.factor(rx),
    covar1 = as.factor(enum),
    covar2 = factor(
      sample(as.factor(enum)), levels = 1:4, labels = c("F", "F", "M", "M")
    )
  )
)
attr(dta_bladder$armcd, "label") <- "ARM"
attr(dta_bladder$covar1, "label") <- "A Covariate Label"
attr(dta_bladder$covar2, "label") <- "Sex (F/M)"
dta_bladder$age <- sample(20:60, size = nrow(dta_bladder), replace = TRUE)

formula <- "Surv(time, status) ~ armcd + covar1"
msum <- summary(survival::coxph(stats::as.formula(formula), data = dta_bladder))
broom::tidy(msum)

}
