% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_time_to_event_table.R
\name{time_to_event_table}
\alias{time_to_event_table}
\title{Time to Event Table}
\usage{
time_to_event_table(time_to_event, event, arm, big_n_arm, arm.ref, comp1.arm,
  comp2.arm, strata1, strata2, strata3, time_point)
}
\arguments{
\item{time_to_event}{time to event data}

\item{is_event}{boolean with \code{TRUE} if event and \code{FALSE} if censored}
}
\value{
a named list with one element per row where each element contains the
  data for the row
}
\description{
This is the description of the function
}
\examples{
\dontrun{
library(teal.oncology)
library(atezo.data)
library(dplyr)
library(survival)
library(broom)
library(stringr)
library(readr)

ASL <- atezo.data::asl(com.roche.cdt30019.go29436.re)
ATE <- atezo.data::ate(com.roche.cdt30019.go29436.re)

tte_tbl_stream <- atezo.data::get_time_to_event_table(com.roche.cdt30019.go29436.re)
teal.oncology::Viewer(tte_tbl_stream)

ASL_f <- dplyr::filter(ASL,ITTFL == "Y")
        
ATE_f <- dplyr::filter(ATE,ITTFL == "Y" & PARAMCD == "OS")

}
}
