% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_chgfbl_plot.R
\name{tm_chgfbl_plot}
\alias{tm_chgfbl_plot}
\title{Change from Baseline Plot Teal Module}
\usage{
tm_chgfbl_plot(label, paramcd, paramcd_choices = paramcd, arm_var = "ARM",
  arm_var_choices = arm_var, ytype = "CHG", ytype_choices = ytype,
  errbar = "SE", errbar_choices = errbar, ref_line_txt = NULL,
  plot_height = c(800, 400, 2000), font_size = c(20, 10, 30),
  n_rotate = TRUE, pre_output = NULL, post_output = NULL)
}
\arguments{
\item{label}{full name label of module}

\item{paramcd}{filter the rows in AQS given the paramcd value}

\item{paramcd_choices}{choices of possible paramcd}

\item{arm_var}{selected variable to use as arms}

\item{arm_var_choices}{choices of arm variables}

\item{ytype}{selected type of value to plot for y-axis}

\item{ytype_choices}{choices of possible ytype: AVAL=value at visit, CHG = change from baseline, PCHG = \% change from baseline}

\item{errbar}{selected type of error bar for the plot}

\item{errbar_choices}{choices of types of error bar: SE = standarnd error, SD = standard deviation, 95CI = 95\% confidence interval of the mean}

\item{ref_line_txt}{horizontal reference lines to display on plot, entered as text separated by comma}

\item{plot_height}{height of the change from baseline plot}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{arm_label}{label to display on plot for the selected arm variable}

\item{cex}{multiplier applied to overall fontsize}
}
\value{
an \code{\link[teal]{module}} object
}
\description{
This module produces a plot of the mean (of value at visit or
  change from baseline) by arm, that's similar to STREAM template MNG01
}
\details{
Package \code{forcats} used to re-format arm data into leveled
factors.
}
\examples{
 

\\donotrun{
library(atezo.data)
library(dplyr)
library(tidyr)
library(teal.oncology)
library(forcats)
library(ggplot2)
library(grid)
library(gridExtra)

ASL <- asl(com.roche.cdpt7722.wo29637.rl)
AQS <- aqs(com.roche.cdpt7722.wo29637.rl)

options(teal_logging = FALSE)


x <- teal::init(
  data = list(ASL = ASL, AQS = AQS),
  modules = root_modules(
    #tm_variable_browser(),
    #tm_data_table(),
    tm_chgfbl_plot(
       label = "Change from Baseline Plot",
       paramcd = "FATIGI",
       paramcd_choices = unique(AQS$PARAMCD),
       arm_var = "ARM",
       arm_var_choices = c("ARM", "ARMCD", "AGE65", "ICC"),
       ytype = "CHG",
       ytype_choices = c("CHG", "AVAL"),
       errbar = "SE",
       errbar_choices = c("SE", "95CI"),
       ref_line_txt = "2, -2"
   ))
)   
shinyApp(x$ui, x$server)

   
} 
}
\author{
Chendi Liao (liaoc10), \email{chendi.liao@roche.com}
}
