% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_demographic_table.R
\name{tm_demographic_table}
\alias{tm_demographic_table}
\title{Demographic Table Teal Module}
\usage{
tm_demographic_table(label, arm_var, arm_var_choices = arm_var,
  summarize_vars, summarize_vars_choices = summarize_vars,
  pre_output = NULL, post_output = NULL)
}
\description{
Demographic Table Teal Module
}
\examples{
\\donotrun{
library(atezo.data)
library(dplyr)
library(survival)

ASL <- asl(com.roche.cdt30019.go29436.re)

x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_demographic_table(
       label = "Demographic Table",
       arm_var = "ARM",
       arm_var_choices = c("ARM", "ARMCD"),
       summarize_vars =  toupper(c("sex", "mliver", "tciclvl2")),
       summarize_vars_choices = toupper(c("sex", "mliver", "tciclvl2", "bage", "age4cat",
            "ethnic", "race", "bwt", "tobhx", "hist", "EGFRMUT",
            "alkmut", "krasmut", "becog"))
   )
  )
)   
shinyApp(x$ui, x$server) 

  
} 


}
