% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot_rsp.R
\name{tm_forest_response}
\alias{tm_forest_response}
\title{Forest Response Plot teal module}
\usage{
tm_forest_response(label, paramcd = "OVRSPI", paramcd_choices = paramcd,
  arm_var = "ARM", arm_var_choices = arm_var, subgroup_var,
  subgroup_var_choices = subgroup_var, plot_height = c(700, 200, 2000),
  cex = 1.5,
  pre_output = helpText("graph needs to be of a certain width to be displayed"),
  post_output = NULL)
}
\description{
Forest Response Plot teal module
}
\examples{
 

\\donotrun{
library(atezo.data)
library(dplyr)
library(survival)
library(forcats)

ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)
 
ASL$BAGED <- ifelse(ASL$BAGE <= median(ASL$BAGE), "<=median", ">median")

x <- teal::init(
  data = list(ASL = ASL, ARS = ARS),
  modules = root_modules(
    tm_forest_response(
       label = "Forest Response",
       paramcd = "OVRSPI",
       paramcd_choices = c("BESRSPI","LSTASDI","MBESRSPI","MLSTASDI","OVRSPI"),
       plot_height = c(600, 200, 2000),
       subgroup_var = c("BAGED", "SEX", "BECOG"),
       arm_var = "ARM",
       arm_var_choices = c("ARM", "ARMCD", "ACTARM")
   )
  )
)   
shinyApp(x$ui, x$server) 

  
} 
}
