% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot.R
\name{tm_forest_survival}
\alias{tm_forest_survival}
\title{Forest Survival Plot teal module}
\usage{
tm_forest_survival(label, endpoint = "OS", endpoint_choices = "OS",
  plot_height = c(600, 200, 2000), pre_output = NULL, post_output = NULL)
}
\description{
Forest Survival Plot teal module
}
\examples{
 

\\donotrun{
library(atezo.data)
library(dplyr)
library(survival)

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)


x <- teal::init(
  data = list(ASL = ASL, ATE = ATE),
  modules = root_modules(
    tm_variable_browser(),
    tm_data_table(),
    tm_forest_survival(
       label = "Forest Survival",
       endpoint = "OS",
       endpoint_choices = unique(ATE$PARAMCD)
   )
  )
)   
shinyApp(x$ui, x$server) 

  
} 
}
