% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_kmplot_ADAM.R
\name{tm_kmplot_ADAM}
\alias{tm_kmplot_ADAM}
\title{teal module for KM plot}
\usage{
tm_kmplot_ADAM(label)
}
\description{
teal module for KM plot
}
\examples{

\dontrun{
### copy the code and save it as file app.R
library(atezo.data)
ASL <- asl(com.roche.cdt30019.go29436.re)
ATE <- ate(com.roche.cdt30019.go29436.re)
## Initialize Teal
x <- teal::init(
  data = list(ASL = ASL, ATE = ATE),
 modules = root_modules(
    tm_data_table(),
    tm_variable_browser(),
    tm_kmplot_ADAM( label = "KM PLOT" ) 
  )
)
## Initiate Shiny App
shinyApp(ui = x$ui, server = x$server)
}
}
