% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_percentage_cross_table.R
\name{tm_percentage_cross_table}
\alias{tm_percentage_cross_table}
\title{Cross table based on rtables}
\usage{
tm_percentage_cross_table(label = "Cross Table", x_var, v_var_choices, y_var,
  y_var_choices)
}
\description{
Cross table based on rtables
}
\examples{

\dontrun{
ASL <- data.frame(
  USUBJID = paste("p", 1:100, sep = "-"),
  STUDYID = "A",
  x = sample(c("Y", "N"), 100, TRUE),
  y = sample(c("A", "B", "C"), 100, TRUE),
  stringsAsFactors = FALSE
)

x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_percentage_cross_table("Cross Table", "x", "y")
  )
)

shinyApp(x$ui, x$server)



}
}
