% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_response_table.R
\name{tm_response_table}
\alias{tm_response_table}
\title{Response Table Teal Module}
\usage{
tm_response_table(label, paramcd = "OVRSPI", paramcd_choices = paramcd,
  arm.var = "ARM", arm.var_choices = arm.var, pre_output = NULL,
  post_output = NULL)
}
\arguments{
\item{label}{label of ..}

\item{paramcd}{filter the rows in ARS given the paramcd value}

\item{paramcd_choices}{choices of possible poaramcd}

\item{arm.var}{variable name for ARM variable in ASL}

\item{arm.var_choices}{choices of arm variables}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\value{
an \code{\link[teal]{module}} object
}
\description{
This modules ...
}
\details{
If necessary write more about that module here....
}
\examples{
 

\\donotrun{
library(atezo.data)
library(dplyr)
library(forcats)

ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

options(teal_logging = FALSE)


x <- teal::init(
  data = list(ASL = ASL, ARS = ARS),
  modules = root_modules(
    tm_variable_browser(),
    tm_data_table(),
    tm_response_table(
       label = "Response Table",
       paramcd = "OVRSPI",
       paramcd_choices = unique(ARS$PARAMCD),
       arm.var = "ARM",
       arm.var_choices = c("ARM", "ARMCD")
   )
  )
)   
shinyApp(x$ui, x$server) 

  
} 
}
