% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_time_to_event_table.R
\name{tm_time_to_event_table}
\alias{tm_time_to_event_table}
\title{Time To Event Table teal module}
\usage{
tm_time_to_event_table(label, time_points, time_points_choices = time_points,
  time_points_unit = "months", arm_var = "ARM", arm_var_choices = arm_var,
  ref_arm = NULL, strata_var = NULL, strata_var_choices = strata_var,
  paramcd = "OS", paramcd_choices = paramcd, pre_output = NULL,
  post_output = NULL)
}
\description{
Time To Event Table teal module
}
\examples{
 

\\donotrun{
library(atezo.data)
library(dplyr)
library(survival)


ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

x <- teal::init(
  data = list(ASL = ASL, ATE = ATE),
  modules = root_modules(
    tm_time_to_event_table(
       label = "Time To Event Table",
       time_points = 6,
       time_points_choices = c(6, 8),
       time_points_unit = "months",
       arm_var = "ARM",
       arm_var_choices = c("ARM", "ARMCD"),
       ref_arm = "DUMMY C",
       paramcd = "OS",
       paramcd_choices = unique(ATE$PARAMCD),
       strata_var = "SEX",
       strata_var_choices = c("SEX", "MLIVER", "TCICLVL2")
   )
  )
)   
shinyApp(x$ui, x$server) 

  
} 
}
