% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_venn.R
\name{tm_venn2}
\alias{tm_venn2}
\title{Venn2 teal module}
\usage{
tm_venn2(label, dataname, bm1_var, bm2_var, bm1_var_choices = bm1_var,
  bm2_var_choices = bm2_var, plot_height = c(600, 200, 2000), alpha = c(1,
  0, 1), pre_output = NULL, post_output = NULL)
}
\description{
Venn2 teal module
}
\examples{
 

N <- 100

var_biomarkers <- paste0("B", 1:10) 
sample_bm_data <- lapply(1:10, function(x)sample(c(TRUE, FALSE), N, replace = TRUE))
names(sample_bm_data) <- var_biomarkers

ASL <- do.call(data.frame, c(
  list(USUBJID = paste("ID", 1:N),STUDYID = "1"), sample_bm_data
))

x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_variable_browser(),
    tm_data_table(),
    tm_venn2("Venn Diagram", "ASL", "B1", "B2", var_biomarkers, var_biomarkers)
  )
)   
    
}
