% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\docType{methods}
\name{to_rtable}
\alias{to_rtable}
\alias{to_rtable,node-method}
\title{Converts the object to an rtable}
\usage{
to_rtable(x, ...)

\S4method{to_rtable}{node}(x, indent = 0)
}
\arguments{
\item{x}{object}

\item{...}{additional arguments to pass}

\item{indent}{indent to print with}
}
\description{
Converts the object to an rtable

Specific to nodes
}
\examples{
n11 <- node(name = "A", content = array(c(1:6), dim = c(2,3)), children = list())
n12 <- node(name = "B", content = array(c(1:6), dim = c(2,3)), children = list())
n13 <- node(name = "C", content = array(c(1:6), dim = c(2,3)), children = list())
n2 <- node(name = "D", content = c(1:3), children = list(n11, n12, n13))

to_rtable(n2) # not working because not a tree of rtables

}
