% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_string_with_names}
\alias{to_string_with_names}
\title{Create String Representation}
\usage{
to_string_with_names(x)
}
\arguments{
\item{x}{object}
}
\value{
string of the form \code{"key1: val1, keey2: val2, ..."}
}
\description{
Create a string representation of the object's contents with names.
}
\examples{
tern:::to_string_with_names(list(a = 1, 2))

}
