% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_fun}
\alias{try_fun}
\title{Make a safe evaluation of any function}
\usage{
try_fun(FUN, ..., timeout = 60 * 5)
}
\arguments{
\item{FUN}{(\code{character})\cr function name, without \code{::} pattern.}

\item{...}{(\code{character})\cr function arguments.}

\item{timeout}{(\code{numeric})\cr specifying the maximum number of seconds the expression is allowed to run
before being interrupted by the timeout. By default it is 5 minutes.}
}
\value{
Function assessment or an error.
}
\description{
Small utility function for better safe evaluation.
}
\examples{
\dontrun{
library(survival)
try_fun("clogit", I(Ozone > 140) ~ Solar.R, airquality)

try_fun("mean", 10)
}

}
