% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_tte_tbl.R
\name{tte_tbl}
\alias{tte_tbl}
\title{alternative tablist}
\usage{
tte_tbl(tte_data, ref_arm, comp_arm)
}
\arguments{
\item{time_to_event}{time to event data}

\item{event}{is boolean, \code{TRUE} if event, \code{FALSE} if time_to_event
is censored}

\item{group_by}{data frame with one column per grouping}

\item{arm}{vector with arm information}

\item{arm.ref}{a character vector defining which arms in arm should be taken 
as the reference}

\item{arm.comp}{a character vector defining which arms in arm should be taken 
as the comparison}
}
\description{
alternative tablist
}
\examples{

\dontrun{

library(atezo.data)
library(dplyr)
library(survival)
library(broom)
library(stringr)
source("./R/utils.R")

drop_label <- function(x){
     for (j in (1:ncol(x))){
          attr(x[[j]],"label") <- NULL
     }
     return(x)
}

ASL_ <- atezo.data::asl(com.roche.cdt30019.go29436.re)
ATE_ <- atezo.data::ate(com.roche.cdt30019.go29436.re)

ASL <- drop_label(ASL_)
ATE <- drop_label(ATE_)

ASL1 <- dplyr::select(ASL,USUBJID,ARM1,SEX,MLIVER,TCICLVL2,ITTFL) \%>\%
        dplyr::filter(ITTFL=="Y") \%>\%
        dplyr::select(-ITTFL)
        
ATE1 <- dplyr::select(ATE,USUBJID,PARAMCD,PARAM,EVNTDESC,CNSR,AVAL,AVALU) \%>\%
        dplyr::filter(PARAMCD == "OS") \%>\%
        dplyr::mutate(time   = AVAL,
                      status = ifelse(is.na(CNSR),NA,
                               ifelse(CNSR==1,0,1)))

ATE2 <- dplyr::inner_join(ASL1,ATE1,by=c("USUBJID"))

}
}
