% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_aggregation_functions.R
\name{ttest_two_arm}
\alias{ttest_two_arm}
\title{Performs 2-arm t-test between comparison group against the reference group.}
\usage{
ttest_two_arm(x, col_by, conf_level = 0.95, ...)
}
\arguments{
\item{x}{(\code{numeric} vector)}

\item{col_by}{(\code{factor} vector)\cr
Containing two or more levels. First level is taken as reference level}

\item{conf_level}{(\code{numeric} value)\cr
Number indicating confidence level. Default is 0.95
Must be greater than 0 and less than 1.}

\item{...}{(optional)\cr
Other arguments from \code{\link{stats}{t.test}}}
}
\value{
list of \code{lcl}, \code{ucl}, \code{diff}, \code{se}, \code{tvalue}, \code{pvalue}
}
\description{
Performs 2-arm t-test between comparison group against the reference group.
}
\examples{
ttest_two_arm(x = c(0,1,2,3,4,5,6,7,8,NA),
              col_by = factor(c("A", "A", "A", "A", "A", "B", "B", "B", "B", "B")),
              var.equal = TRUE)
}
