% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_response_table.R
\name{ui_response_table}
\alias{ui_response_table}
\title{UI part for response table teal module}
\usage{
ui_response_table(id, label, paramcd = "OVRSPI", paramcd_choices = paramcd,
  arm.var = "ARM", arm.var_choices = arm.var, pre_output, post_output)
}
\arguments{
\item{id}{namespace id}

\item{label}{label of ..}

\item{paramcd}{filter the rows in ARS given the paramcd value}

\item{paramcd_choices}{choices of possible poaramcd}

\item{arm.var}{variable name for ARM variable in ASL}

\item{arm.var_choices}{choices of arm variables}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
UI part for response table teal module
}
