% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{univariate}
\alias{univariate}
\title{Univariate formula special term}
\usage{
univariate(x)
}
\arguments{
\item{x}{A vector of variable name separated by comas.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The special term \code{univariate} indicate that the model should be fitted individually for
every variable included in univariate.
}
\details{
If provided alongside with pairwise specification, the model
\code{y ~ ARM + univariate(SEX, AGE, RACE)} lead to the study and comparison of the models
\itemize{
\item \code{y ~ ARM}
\item \code{y ~ ARM + SEX}
\item \code{y ~ ARM + AGE}
\item \code{y ~ ARM + RACE}
}
}
