% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rtables.R
\name{unlist_and_blank_na}
\alias{unlist_and_blank_na}
\title{Blank for Missing Input}
\usage{
unlist_and_blank_na(x)
}
\arguments{
\item{x}{(\code{vector})\cr input for a cell.}
}
\value{
Either empty character vector if all entries in \code{x} are missing (\code{NA}), or otherwise
the unlisted version of \code{x}
}
\description{
Helper function to use in tabulating model results.
}
\keyword{internal}
