% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_rtables.R
\name{unlist_rtables}
\alias{unlist_rtables}
\title{Unlist a Nested Lists with rtables as leafes}
\usage{
unlist_rtables(x)
}
\arguments{
\item{x}{a nested list of with rtables as leaf object}
}
\value{
a list of rtables
}
\description{
Often it is useful to flatten a nested lists with rtables as leafes to a list
of rtables. The algorithm used is a depth first tree traversal.
}
\examples{

l_tbls <- list(
  list(
     rtabulate(iris$Sepal.Length, iris$Species, mean),
     rtabulate(iris$Sepal.Length, iris$Species, sd)
  ),
  list(
     rtabulate(iris$Sepal.Width, iris$Species, mean)
  )  
)

unlist_rtables(l_tbls)

}
