% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{use_stuff}
\alias{use_stuff}
\title{Dummy function to use currently unused functions}
\usage{
use_stuff(...)
}
\arguments{
\item{...}{dummy arguments}
}
\description{
to delete eventually. The function was introduced in the refactor.
}
