% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_grob.R
\name{wrap.text}
\alias{wrap.text}
\title{Return split text that fits onto page width}
\usage{
wrap.text(txt, width = convertWidth(unit(1, "npc"), "inch", TRUE),
  gp = gpar())
}
\arguments{
\item{txt}{string (or vector of strins) to be split in multiple lines, not that
\code{'\n'} is also split into to lines}

\item{width}{max with of string, by default the width of the current viewport}

\item{gp}{graphical parameters for text}
}
\value{
a vector with the new strings
}
\description{
Return split text that fits onto page width
}
\examples{

library(grid)
grid.newpage()
pushViewport(plotViewport())
grid.rect()
text <- "g_uefg_ex02_il05_ip11_CABRTR72_NIS DRAFT Individual Patient Calculated ABR Bar Chart, Treated bleeds, intra−patient comparison, NIS"
grid.text(paste(wrap.text(text), collapse = "\\n"), just = "left", x = unit(0, "npc"))
}
