% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_text.R
\name{wrap_text}
\alias{wrap_text}
\title{return splitted text that fits onto page width}
\usage{
wrap_text(txt, width = convertWidth(unit(1, "npc"), "inch", TRUE),
  gp = gpar(), collapse = NULL)
}
\arguments{
\item{txt}{string (or vector of strins) to be split in multiple lines, not that 
\code{\\n} is also split into to lines}

\item{width}{max with of string, by default the width of the current viewport}

\item{gp}{graphical parameters for text}
}
\value{
e vector with the new strings
}
\description{
return splitted text that fits onto page width
}
\examples{
library(grid)
text <- "This is a test with many words and more"
wrap_text(txt = text, width = unit(4, "cm"))

}
