% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abnormal_by_marked.R
\name{abnormal_by_marked}
\alias{abnormal_by_marked}
\alias{s_count_abnormal_by_marked}
\alias{a_count_abnormal_by_marked}
\alias{count_abnormal_by_marked}
\title{Count patients with marked laboratory abnormalities}
\usage{
s_count_abnormal_by_marked(
  df,
  .var = "AVALCAT1",
  abnormal = c(Low = "LOW LOW"),
  variables = list(id = "USUBJID", direction = "ANRIND")
)

a_count_abnormal_by_marked(
  df,
  .var = "AVALCAT1",
  abnormal = c(Low = "LOW LOW"),
  variables = list(id = "USUBJID", direction = "ANRIND")
)

count_abnormal_by_marked(
  lyt,
  var,
  abnormal,
  variables,
  ...,
  table_names = abnormal,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{abnormal}{(named \code{string} or \code{character})\cr identifying the abnormal direction in \code{var}.}

\item{variables}{(named \code{list} of \code{string})\cr list of additional analysis variables.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{var}{(\code{string})\cr single variable name for the primary analysis variable.}

\item{...}{additional arguments for the lower level functions.}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
\code{\link[=s_count_abnormal_by_marked]{s_count_abnormal_by_marked()}} the single statistic \code{count_fraction}
with \verb{Single, not last}, \verb{Last or replicated} and \code{Any} results.

\code{\link[=a_count_abnormal_by_marked]{a_count_abnormal_by_marked()}} returns the corresponding list with formatted \code{\link[rtables:CellValue]{rtables::CellValue()}}.
}
\description{
Primary analysis variable \code{.var} indicates whether single, replicated
or last marked laboratory abnormality was observed (factor).
Additional analysis variables are \code{id} (character or factor) and \code{direction} indicating
the direction of the abnormality (factor).
Denominator is number of patients with at least one valid measurement during
treatment (post-baseline), and patients in the numerator are considered as follows:
\itemize{
\item For \verb{Single, not last} and \verb{Last or replicated}: Numerator is number of patients
with \verb{Single, not last} and \verb{Last or replicated} levels, respectively.
\item For \code{Any}: Numerator is the number of patients with either single or
replicated marked abnormalities.
}
}
\details{
Note that \verb{Single, not last} and \verb{Last or replicated} levels are
mutually exclusive. If a patient has abnormalities that meet both the \verb{Single, not last}
and \verb{Last or replicated} criteria, then the patient will be counted only under the \verb{Last or replicated} category.
}
\section{Functions}{
\itemize{
\item \code{s_count_abnormal_by_marked}: Statistics function which returns
the counts and fractions of patients with \verb{Single, not last}, \verb{Last or replicated} and \code{Any}
marked laboratory abnormalities for a single \code{abnormal} level.

\item \code{a_count_abnormal_by_marked}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{count_abnormal_by_marked}: Layout creating function which can be used for creating tables,
which can take statistics function arguments and additional format arguments (see below).
}}

\examples{

library(dplyr)

df <- data.frame(
  USUBJID = as.character(c(1, 1, 1, 1, 1, 2, 2, 2, 2, 2)),
  ARMCD = factor(c(rep("ARM A", 5),rep("ARM B", 5))),
  ANRIND = factor(c("NORMAL", "LOW", "HIGH", "HIGH HIGH","LOW LOW",
  "LOW", "LOW", "HIGH HIGH", "LOW", "HIGH HIGH")),
  ONTRTFL = c("", "Y", "Y", "Y", "Y", "Y", "Y", "Y", "Y", "Y"),
  PARAMCD = factor(rep("CRP", 10)),
  AVALCAT1 = factor(c("", "", "", "SINGLE", "REPLICATED", "", "", "LAST", "", "SINGLE")),
  stringsAsFactors = FALSE
)

# Select only post-baseline records.
df <- df \%>\% filter(ONTRTFL == "Y")
 s_count_abnormal_by_marked(df = df,
 .var = "AVALCAT1",
 abnormal = c(High = "HIGH HIGH"),
 variables = list(id = "USUBJID", direction = "ANRIND")
)

 s_count_abnormal_by_marked(df = df,
 .var = "AVALCAT1",
 abnormal = c(Low = "LOW LOW"),
 variables = list(id = "USUBJID", direction = "ANRIND")
)
# Use the Formatted Analysis function for `analyze()`. We need to ungroup `count_fraction` first
# so that the rtables formatting function `format_count_fraction()` can be applied correctly.
afun <- make_afun(a_count_abnormal_by_marked, .ungroup_stats = "count_fraction")
afun(
  df = df,
  .var = "AVALCAT1",
  abnormal = c(High = "HIGH HIGH"),
  variables = list(id = "USUBJID", direction = "ANRIND")
)
basic_table() \%>\%
split_cols_by("ARMCD") \%>\%
summarize_num_patients(var = "USUBJID",
                       .stats = "unique_count") \%>\%
count_abnormal_by_marked(var = "AVALCAT1" ,
                         abnormal = c(Low = "LOW LOW", High = "HIGH HIGH"),
                         variables = list(id = "USUBJID", direction = "ANRIND")) \%>\%
                         build_table(df = df)

}
