% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abnormal_by_worst_grade_worsen.R
\name{abnormal_by_worst_grade_worsen}
\alias{abnormal_by_worst_grade_worsen}
\alias{s_count_abnormal_lab_worsen_by_baseline}
\alias{a_count_abnormal_lab_worsen_by_baseline}
\alias{count_abnormal_lab_worsen_by_baseline}
\title{Patient Counts for Laboratory Events (Worsen From Baseline) by Highest Grade Post-Baseline}
\usage{
s_count_abnormal_lab_worsen_by_baseline(
  df,
  .var = "ATOXGR",
  variables = list(id = "USUBJID", baseline_var = "BTOXGR", direction_var = "GRADDR")
)

a_count_abnormal_lab_worsen_by_baseline(
  df,
  .var = "ATOXGR",
  variables = list(id = "USUBJID", baseline_var = "BTOXGR", direction_var = "GRADDR")
)

count_abnormal_lab_worsen_by_baseline(
  lyt,
  var,
  ...,
  table_names = NULL,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var, var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{variables}{(named \code{list} of \code{string}) \cr list of additional analysis variables including:
\itemize{
\item \code{id} (\code{string}): \cr subject variable name
\item \code{baseline_var} (\code{string}): \cr name of the data column containing baseline toxicity variable
\item \code{direction_var} (\code{string}): See \code{direction_var} for more detail
}}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{...}{additional arguments for the lower level functions.}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
\code{\link[=s_count_abnormal_lab_worsen_by_baseline]{s_count_abnormal_lab_worsen_by_baseline()}} returns the
counts and fraction of patients whose worst post-baseline lab grades are worse than
their baseline grades, for post-baseline worst grades "1", "2", "3", "4" and "Any".

\code{\link[=a_count_abnormal_lab_worsen_by_baseline]{a_count_abnormal_lab_worsen_by_baseline()}} returns
the corresponding list with formatted \code{\link[rtables:CellValue]{rtables::CellValue()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Patient count and fraction for laboratory events (worsen from baseline) shift table.
}
\section{Functions}{
\itemize{
\item \code{s_count_abnormal_lab_worsen_by_baseline}: Statistics function which calculates the
counts and fraction of patients whose worst post-baseline lab grades are worse than
their baseline grades, for post-baseline worst grades "1", "2", "3", "4" and "Any".

\item \code{a_count_abnormal_lab_worsen_by_baseline}: Formatted Analysis function which can be further customized by
calling \code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{count_abnormal_lab_worsen_by_baseline}: Layout
creating function which can be used for creating tables, which can take statistics
function arguments and additional format arguments (see below).
}}

\examples{
# Patients with worsening lab grade for CRP in the direction of low
tern:::s_count_abnormal_lab_worsen_by_baseline(
  df = df \%>\% filter(ARMCD == "ARM A" & PARAMCD == "CRP"),
  .var = "ATOXGR",
  variables = list(
    id = "USUBJID",
    baseline_var = "BTOXGR",
    direction_var = "GRADDR"
  )
)

tern:::a_count_abnormal_lab_worsen_by_baseline(
  df = df \%>\% filter(ARMCD == "ARM A" & PARAMCD == "CRP"),
  .var = "ATOXGR",
  variables = list(id = "USUBJID", baseline_var = "BTOXGR", direction_var = "GRADDR")
)

basic_table() \%>\%
  split_cols_by("ARMCD") \%>\%
  add_colcounts() \%>\%
  split_rows_by("PARAMCD") \%>\%
  split_rows_by("GRADDR") \%>\%
  count_abnormal_lab_worsen_by_baseline(
    var = "ATOXGR",
    variables = list(
      id = "USUBJID",
      baseline_var = "BTOXGR",
      direction_var = "GRADDR"
    )
  ) \%>\%
  append_topleft("Direction of Abnormality") \%>\%
  build_table(df = df, alt_counts_df = adsl)

}
\keyword{internal}
