% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R, R/g_forest.R
\name{add_footnotes<-}
\alias{add_footnotes<-}
\title{Add Footnotes}
\usage{
add_footnotes(x) <- value

add_footnotes(x) <- value
}
\arguments{
\item{x}{an object}

\item{value}{character vector}
}
\description{
This adds more footnotes.
}
\examples{
x <- table(iris$Species)
footnotes(x) <- "Species are equally distributed"
footnotes(x)
add_footnotes(x) <- "Add more footnotes"
footnotes(x)

x <- table(iris$Species)
footnotes(x) <- "Species are equally distributed"
footnotes(x)
add_footnotes(x) <- "Add more footnotes"
footnotes(x)
}
