% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTable.R
\name{addTable}
\alias{addTable}
\title{Add an rtable as Text Annotation to a Grid Grob}
\usage{
addTable(grob, tbl, x = unit(0.5, "npc"), y = unit(0.5, "npc"),
  just = c("left", "top"), vp = vpPath("plotArea", "topCurve"))
}
\arguments{
\item{grob}{a kmGrob for further editing}

\item{tbl}{an \code{\link[rtables]{rtable}}}

\item{x}{numeric vector or unit object specifying x-values}

\item{y}{numeric vector or unit object specifying y-values}

\item{just}{The justification of the text relative to its (x, y) location}

\item{vp}{grid viewport object for annotating a text grob}
}
\description{
utility function to edit a grid grob by adding a text grob
}
\examples{
library(dplyr)
library(random.cdisc.data)
ADSL <- radsl()
ADSL$RACE <- factor(sapply(as.character(ADSL$RACE), function(x) {
   if (nchar(x)>9) paste0(substr(x, 1,9), "...") else x
}))

ADTTE <- radtte(ADSL)
ADTTE_f <- subset(ADTTE, PARAMCD == "OS") 
                                    
fit_km <- survfit(Surv(AVAL, 1-CNSR) ~ ARM, data = ADTTE_f, conf.type = "plain")

a_kmgrob <- g_km(fit_km, xticks = 0.5, draw = FALSE)

fit_coxph <- coxph(Surv(AVAL, 1-CNSR) ~ ARM + strata(RACE), data = ADTTE_f, ties = "exact")
cox_tbl <- t_coxph(fit_coxph)

p_tbl <- addTable(
  a_kmgrob, 
  vp = vpPath("plotArea", "topCurve"), 
  x= unit(1, "lines"), y = unit(1, "lines"),
  just = c("left", "bottom"),
  tbl = cox_tbl 
)

grid.newpage()
grid.draw(p_tbl)

}
\author{
Heng Wang (wangh107) \email{wang.heng@gene.com}
}
