% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_km.R
\name{addTable}
\alias{addTable}
\title{Add text annotation on top of a kmgrob}
\usage{
addTable(kmgrob, vp, x = unit(0.5, "npc"), y = unit(0.5, "npc"), tbl,
  just = c("left", "top"))
}
\arguments{
\item{kmgrob}{a kmGrob for further editing.}

\item{vp}{a Grid veiwport object for annotating a text grob.}

\item{x}{A numeric vector or unit object specifying x-values.}

\item{y}{A numeric vector or unit object specifying y-values.}

\item{tbl}{An rtable.}

\item{just}{The justification of the text relative to its (x, y) location.}
}
\description{
utility function to edit a grid grob by adding a text grob
}
\examples{
library(scales)
library(dplyr)
OS <- data.frame(AVAL = abs(rnorm(200)), 
                 CNSR = sample(c(0, 1), 200, TRUE), 
                 ARM = sample(LETTERS[1:3], 200, TRUE),
                 SEX = sample(c("M","F"), 200, TRUE),
                 RACE = sample(c("AA", "BB", "CC"), 200, TRUE),
                 ECOG = sample(c(0, 1), 200, TRUE))
a_kmgrob <- kmGrob(formula_km = Surv(AVAL, 1-CNSR) ~ ARM, data = OS, xaxis_by = 0.5)
cox_tbl <- coxphAnnoData(Surv(AVAL, 1-CNSR) ~ ARM + strata(RACE), data = OS)
addTable(a_kmgrob, 
         vp = vpPath("plotarea", "topcurve"), 
         x= unit(1, "lines"), y = unit(1, "lines"),
         just = c("left", "bottom"),
         tbl = cox_tbl ) \%>\% 
         grid.draw()
         
         
}
