% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aesi_label}
\alias{aesi_label}
\title{Labels for Adverse Event Baskets}
\usage{
aesi_label(aesi, scope = NULL)
}
\arguments{
\item{aesi}{(\code{character})\cr with standardized MedDRA query name (e.g. SMQzzNAM) or customized query
name (e.g. CQzzNAM).}

\item{scope}{(\code{character})\cr with scope of query (e.g. SMQzzSC).}
}
\value{
A \code{string} with the standard label for the AE basket.
}
\description{
Labels for Adverse Event Baskets
}
\examples{
library(scda)
adae <- synthetic_cdisc_data("latest")$adae

# Standardized query label includes scope.
aesi_label(adae$SMQ01NAM, scope = adae$SMQ01SC)

# Customized query label.
aesi_label(adae$CQ01NAM)
}
