% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_diff.R
\name{afun_riskdiff}
\alias{afun_riskdiff}
\title{Analysis Function to Calculate Risk Difference Column Values}
\usage{
afun_riskdiff(
  df,
  labelstr = "",
  .var,
  .N_col,
  .N_row,
  .df_row,
  .spl_context,
  .all_col_counts,
  .stats,
  .indent_mods,
  afun,
  s_args = list(...)
)
}
\description{
Analysis Function to Calculate Risk Difference Column Values
}
