% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskdiff.R
\name{afun_riskdiff}
\alias{afun_riskdiff}
\title{Analysis Function to Calculate Risk Difference Column Values}
\usage{
afun_riskdiff(
  df,
  labelstr = "",
  .var,
  .N_col,
  .N_row,
  .df_row,
  .spl_context,
  .all_col_counts,
  .stats,
  .indent_mods,
  afun,
  s_args = list()
)
}
\arguments{
\item{df}{(\code{data.frame})\cr data set containing all analysis variables.}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions). See \code{\link[rtables:summarize_row_groups]{rtables::summarize_row_groups()}}
for more information.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.N_col}{(\code{integer})\cr column-wise N (column count) for the full column being analyzed that is typically
passed by \code{rtables}.}

\item{.N_row}{(\code{integer})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is typically passed by \code{rtables}.}

\item{.df_row}{(\code{data.frame})\cr data frame across all of the columns for the given row split.}

\item{.spl_context}{(\code{data.frame})\cr gives information about ancestor split states
that is passed by \code{rtables}.}

\item{.all_col_counts}{(\code{vector} of \code{integer})\cr each value represents a global count for a column. Values are
taken from \code{alt_counts_df} if specified (see \code{\link[rtables:build_table]{rtables::build_table()}}).}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels. Defaults to 0, which corresponds to the
unmodified default behavior. Can be negative.}

\item{afun}{(\verb{named list})\cr A named list containing one name-value pair where the name corresponds to the
name of the statistics function that should be used in calculations and the value is the corresponding analysis
function.}

\item{s_args}{(\verb{named list})\cr Additional arguments to be passed to the statistics function and analysis function
supplied in \code{afun}.}
}
\value{
A list of formatted \code{\link[rtables:CellValue]{rtables::CellValue()}}.
}
\description{
In the risk difference column, uses the statistics function associated with \code{afun} to calculates risk difference
values from arm X (reference group) and arm Y. These arms are specified when configuring the risk difference
column which is done using the \code{\link[=add_riskdiff]{add_riskdiff()}} split function in the previous call to \code{\link[rtables:split_cols_by]{rtables::split_cols_by()}}.
For all other columns, applies \code{afun} as usual. This function utilizes the \code{\link[=stat_propdiff_ci]{stat_propdiff_ci()}} function to
perform risk difference calculations.
}
\keyword{internal}
