% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskdiff.R
\name{afun_riskdiff}
\alias{afun_riskdiff}
\title{Analysis Function to Calculate Risk Difference Column Values}
\usage{
afun_riskdiff(
  df,
  labelstr = "",
  .var,
  .N_col,
  .N_row,
  .df_row,
  .spl_context,
  .all_col_counts,
  .stats,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL,
  na_str = default_na_str(),
  afun,
  s_args = list()
)
}
\arguments{
\item{df}{(\code{data.frame})\cr data set containing all analysis variables.}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions). See \code{\link[rtables:summarize_row_groups]{rtables::summarize_row_groups()}}
for more information.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.N_col}{(\code{integer})\cr column-wise N (column count) for the full column being analyzed that is typically
passed by \code{rtables}.}

\item{.N_row}{(\code{integer})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is typically passed by \code{rtables}.}

\item{.df_row}{(\code{data.frame})\cr data frame across all of the columns for the given row split.}

\item{.spl_context}{(\code{data.frame})\cr gives information about ancestor split states
that is passed by \code{rtables}.}

\item{.all_col_counts}{(\code{vector} of \code{integer})\cr each value represents a global count for a column. Values are
taken from \code{alt_counts_df} if specified (see \code{\link[rtables:build_table]{rtables::build_table()}}).}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics. See Details in \code{analyze_vars} for more
information on the \code{"auto"} setting.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels. Defaults to 0, which corresponds to the
unmodified default behavior. Can be negative.}

\item{na_str}{(\code{string})\cr string used to replace all \code{NA} or empty values in the output.}

\item{afun}{(named \code{list})\cr A named list containing one name-value pair where the name corresponds to
the name of the statistics function that should be used in calculations and the value is the corresponding
analysis function.}

\item{s_args}{(named \code{list})\cr Additional arguments to be passed to the statistics function and analysis
function supplied in \code{afun}.}
}
\value{
A list of formatted \code{\link[rtables:CellValue]{rtables::CellValue()}}.
}
\description{
In the risk difference column, this function uses the statistics function associated with \code{afun} to
calculates risk difference values from arm X (reference group) and arm Y. These arms are specified
when configuring the risk difference column which is done using the \code{\link[=add_riskdiff]{add_riskdiff()}} split function in
the previous call to \code{\link[rtables:split_cols_by]{rtables::split_cols_by()}}. For all other columns, applies \code{afun} as usual. This
function utilizes the \code{\link[=stat_propdiff_ci]{stat_propdiff_ci()}} function to perform risk difference calculations.
}
\seealso{
\itemize{
\item \code{\link[=stat_propdiff_ci]{stat_propdiff_ci()}} for details on risk difference calculation.
\item Split function \code{\link[=add_riskdiff]{add_riskdiff()}} which, when used as \code{split_fun} within \code{\link[rtables:split_cols_by]{rtables::split_cols_by()}} with
\code{riskdiff} argument set to \code{TRUE} in subsequent analyze functions calls, adds a risk difference column
to a table layout.
}
}
\keyword{internal}
