% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{argument_convention}
\alias{argument_convention}
\title{Dummy Function that Documents all Standard Arguments}
\arguments{
\item{na_rm}{(\code{logical} value)\cr
Indicating whether \code{NA} values should be
removed from \code{x} prior to counting the  'aa_trees.R' 'argument_convention.R' unique elements per cell.}

\item{tte}{(\code{numeric} vector)\cr
Contains time to event data, currently no \code{NA} allowed}

\item{is_event}{(\code{logical} value)\cr
\code{TRUE} if event, \code{FALSE} if \code{tte}
is censored}

\item{rsp}{(\code{logical} vector)\cr
\code{TRUE} if subject is a responder \code{FALSE} otherwise}

\item{conf_level}{confidence level of the interval.}

\item{prune_zero_rows}{\code{logical} whether to prune all zero rows}
}
\description{
The documentation to this function lists all the arguments in \code{tern} that are used repeatedly to express an
analysis
}
\details{
Although this function just returns \code{NULL} it has two uses, for the \code{tern} users it provides a
documentation of arguments that are commonly and consistently used in the framework. For the developer it adds a
single reference point to import the \code{roxygen} argument description with
\code{@inheritParams argument_convention}
}
