% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{argument_convention}
\alias{argument_convention}
\title{Standard Arguments}
\arguments{
\item{col_by}{(\code{factor})\cr defining column groups.}

\item{na_rm}{(\code{flag})\cr (deprecated) whether \code{NA} values should be removed from \code{x} prior to analysis.}

\item{na.rm}{(\code{flag})\cr whether \code{NA} values should be removed from \code{x} prior to analysis.}

\item{tte}{(\verb{\code{numeric}})\cr contains time-to-event duration values.}

\item{is_event}{(\code{logical})\cr \code{TRUE} if event, \code{FALSE} if time to event is censored.}

\item{rsp}{(\code{logical})\cr whether each subject is a responder or not.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{prune_zero_rows}{(\code{flag})\cr whether to prune all zero rows.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{variables}{(named \code{list} of \code{strings})\cr list of additional analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested by
a statistics function.}

\item{...}{additional arguments for the lower level functions.}
}
\description{
The documentation to this function lists all the arguments in \code{tern}
that are used repeatedly to express an analysis.
}
\details{
Although this function just returns \code{NULL} it has two uses, for
the \code{tern} users it provides a documentation of arguments that are
commonly and consistently used in the framework. For the developer it adds a
single reference point to import the \code{roxygen} argument description with:
\verb{@inheritParams argument_convention}
}
