% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{argument_convention}
\alias{argument_convention}
\title{Standard Arguments}
\arguments{
\item{...}{additional arguments for the lower level functions.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}

\item{.N_col}{(\code{count})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is passed by \code{rtables}.}

\item{.N_row}{(\code{count})\cr column-wise N (column count) for the full column that is passed by \code{rtables}.}

\item{.ref_group}{(\verb{data frame} or \code{vector})\cr the data corresponding to the reference group.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{col_by}{(\code{factor})\cr defining column groups.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{id}{(\code{string}) \cr subject variable name.}

\item{is_event}{(\code{logical})\cr \code{TRUE} if event, \code{FALSE} if time to event is censored.}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions).}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{na.rm}{(\code{flag})\cr whether \code{NA} values should be removed from \code{x} prior to analysis.}

\item{na_rm}{(\code{flag})\cr (deprecated) whether \code{NA} values should be removed from \code{x} prior to analysis.}

\item{prune_zero_rows}{(\code{flag})\cr whether to prune all zero rows.}

\item{rsp}{(\code{logical})\cr whether each subject is a responder or not.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{tte}{(\code{numeric})\cr contains time-to-event duration values.}

\item{var_labels}{character for label.}

\item{variables}{(named \code{list} of \code{string})\cr list of additional analysis variables.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{var}{(\code{string})\cr single variable name for the primary analysis variable.}

\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}
}
\description{
The documentation to this function lists all the arguments in \code{tern}
that are used repeatedly to express an analysis.
}
\details{
Although this function just returns \code{NULL} it has two uses, for
the \code{tern} users it provides a documentation of arguments that are
commonly and consistently used in the framework. For the developer it adds a
single reference point to import the \code{roxygen} argument description with:
\verb{@inheritParams argument_convention}
}
