% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{arm}
\alias{arm}
\title{Indicate Arm Variable in Formula}
\usage{
arm(x)
}
\arguments{
\item{x}{arm information}
}
\description{
We use \code{arm} to indicate the study arm variable in \code{tern} formulas.
}
