% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{arm}
\alias{arm}
\title{Indentity function to indicate specials in formulas}
\usage{
arm(x)
}
\arguments{
\item{x}{arm information}
}
\description{
We use \code{arm} to in tern formulas
}
\author{
boruvkaa
}
