% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_factor.R
\name{as_factor_keep_attributes}
\alias{as_factor_keep_attributes}
\title{Conversion of a Vector to a Factor}
\usage{
as_factor_keep_attributes(
  x,
  x_name = deparse(substitute(x)),
  na_level = "<Missing>",
  verbose = TRUE
)
}
\arguments{
\item{x}{(\code{atomic})\cr object to convert.}

\item{x_name}{(\code{string})\cr name of \code{x}.}

\item{na_level}{(\code{string})\cr the explicit missing level which should be used when
converting a character vector.}

\item{verbose}{defaults to \code{TRUE}. It prints out warnings and messages.}
}
\value{
The factor with same attributes (except class) as \code{x}. Does not do any modifications
if \code{x} is already a factor.
}
\description{
Converts \code{x} to a factor and keeps its attributes. Warns appropriately such that the user
can decide whether they prefer converting to factor manually (e.g. for full control of
factor levels).
}
\examples{
tern:::as_factor_keep_attributes(formatters::with_label(c(1, 1, 2, 3), "id"), verbose = FALSE)
tern:::as_factor_keep_attributes(c("a", "b", ""), "id", verbose = FALSE)

}
\keyword{internal}
