% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_factor_keep_attributes}
\alias{as_factor_keep_attributes}
\title{Convert to factor and keep attributes}
\usage{
as_factor_keep_attributes(x)
}
\arguments{
\item{x}{atomic}
}
\value{
factor with same attributes (except class) as x
}
\description{
Convert to factor and keep attributes
}
\examples{
as_factor_keep_attributes(with_label(c(1,1,2,3), "id"))
}
