% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertthat.R
\name{assertions}
\alias{assertions}
\alias{is_character_or_factor}
\alias{is_nonnegative_count}
\alias{is_variables}
\alias{is_df_with_variables}
\title{Additional assertions that work with `assert_that`.}
\usage{
is_character_or_factor(x)

is_nonnegative_count(x)

is_variables(x)

is_df_with_variables(df, variables)
}
\arguments{
\item{x}{object to test}

\item{df}{supposed data frame to test}

\item{variables}{supposed variables list to test}
}
\value{
logical whether the assertion holds (TRUE or FALSE). When used inside `assert_that` produces
  a meaningful error message.
}
\description{
Additional assertions that work with `assert_that`.
}
\section{Functions}{
\itemize{
\item \code{is_character_or_factor}: Check whether `x` is a character or factor vector.

\item \code{is_nonnegative_count}: Check whether `x` is a nonnegative count.

\item \code{is_variables}: Check whether `x` is a valid list of variable names.

\item \code{is_df_with_variables}: Check whether `df` is a data frame with the analysis `variables`.
}}

\examples{

# Check whether `x` is a character or factor vector.
is_character_or_factor(-1)
is_character_or_factor(c("a", "b"))
is_character_or_factor(factor(c("a", "b")))

# Check whether `x` is a nonnegative count.
is_nonnegative_count(-1)
is_nonnegative_count(0L)
is_nonnegative_count(10L)

# Check whether `x` is a valid list of variable names.
is_variables(list(val = "a"))
is_variables(list(1, 2))
is_variables(list("bla"))

# Check whether `df` contains the analysis `variables`.
is_df_with_variables(df = data.frame(a = 5, b = 3), variables = list(val = "a"))
}
