% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertthat.R
\name{assertions}
\alias{assertions}
\alias{is_character_or_factor}
\alias{is_nonnegative_count}
\alias{is_variables}
\alias{is_df_with_variables}
\alias{is_df_with_factors}
\alias{is_df_with_nlevels_factor}
\alias{is_equal_length}
\alias{is_proportion}
\alias{is_proportion_vector}
\alias{is_quantiles_vector}
\alias{is_valid_factor}
\alias{is_factor_no_na}
\alias{is_valid_character}
\alias{all_elements_in_ref}
\alias{has_tabletree_colnames}
\alias{is_df_with_no_na_level}
\title{Additional Assertions for \code{assert_that}}
\usage{
is_character_or_factor(x)

is_nonnegative_count(x)

is_variables(x)

is_df_with_variables(df, variables)

is_df_with_factors(df, variables)

is_df_with_nlevels_factor(df, variable, n_levels, relation = c("==", ">="))

is_equal_length(...)

is_proportion(x, include_boundaries = FALSE)

is_proportion_vector(x, include_boundaries = FALSE)

is_quantiles_vector(x, include_boundaries = FALSE)

is_valid_factor(x)

is_factor_no_na(x)

is_valid_character(x)

all_elements_in_ref(x, ref)

has_tabletree_colnames(x, col_names)

is_df_with_no_na_level(df, variables, na_level)
}
\arguments{
\item{x}{object to test}

\item{df}{(\verb{data frame})\cr input dataset.}

\item{variables}{(\verb{named list})\cr columns from dataset where you want to check if \code{na_level} exists.}

\item{variable}{(\code{string})\cr name of the single variable.}

\item{n_levels}{(\code{count})\cr number of levels to compare with.}

\item{relation}{(\code{string})\cr which relational operator to use for the comparison.}

\item{...}{a collection of objects to test.}

\item{include_boundaries}{(\code{logical})\cr whether to include boundaries when testing for proportions.}

\item{ref}{(\code{vector})\cr where matches from \code{x} are sought for \code{all_elements_in_ref}.}

\item{col_names}{(\code{character})\cr column names which should be present in the table.}

\item{na_level}{(\code{string})\cr the string user has been using to represent NA or missing data.}
}
\value{
\code{flag} whether the assertion holds (\code{TRUE} or \code{FALSE}). When used inside
\code{\link[assertthat:assert_that]{assertthat::assert_that()}} produces a meaningful error message.
}
\description{
We provide additional assertion functions which can be used together with \code{\link[assertthat:assert_that]{assertthat::assert_that()}}.
}
\section{Functions}{
\itemize{
\item \code{is_character_or_factor}: Check whether \code{x} is a character or factor vector.

\item \code{is_nonnegative_count}: Check whether \code{x} is a nonnegative count.

\item \code{is_variables}: Check whether \code{x} is a valid list of variable names.

\item \code{is_df_with_variables}: Check whether \code{df} is a data frame with the analysis \code{variables}.

\item \code{is_df_with_factors}: Check whether \code{df} is a data frame where the analysis \code{variables}
are all factors.

\item \code{is_df_with_nlevels_factor}: Check whether \code{df} is a data frame where the analysis \code{variable}
is a factor with \code{n_levels} number of levels.

\item \code{is_equal_length}: Check that objects provided are of same length.

\item \code{is_proportion}: Check whether \code{x} is a proportion: number between 0 and 1.

\item \code{is_proportion_vector}: Check whether \code{x} is a vector of proportions (numbers between 0 and 1).

\item \code{is_quantiles_vector}: Check whether \code{x} is a sorted vector of unique quantile proportions
(numbers between 0 and 1).

\item \code{is_valid_factor}: Check whether \code{x} is a valid factor (has levels and no empty string levels).

\item \code{is_factor_no_na}: Check whether \code{NA} has been handled when \code{x} is a factor.

\item \code{is_valid_character}: Check whether \code{x} is a valid character (has no NAs and no empty strings).

\item \code{all_elements_in_ref}: Check whether all elements of \code{x} are in a reference vector.

\item \code{has_tabletree_colnames}: Check whether rtables object \code{x} has the specified column names.

\item \code{is_df_with_no_na_level}: check if the input \code{df} has \code{na_level} in its \code{variables}.
}}

\examples{

# Check whether `x` is a character or factor vector.
is_character_or_factor(-1)
is_character_or_factor(c("a", "b"))
is_character_or_factor(factor(c("a", "b")))

# Check whether `x` is a nonnegative count.
is_nonnegative_count(-1)
is_nonnegative_count(0L)
is_nonnegative_count(10L)
is_nonnegative_count(0)
is_nonnegative_count(10)

# Check whether `x` is a valid list of variable names.
is_variables(list(val = "a"))
is_variables(list(val = c("a", "b")))
is_variables(list(1, 2))
is_variables(list("bla"))

# Check whether `df` contains the analysis `variables`.
is_df_with_variables(df = data.frame(a = 5, b = 3), variables = list(val = "a"))
is_df_with_variables(df = data.frame(a = 5, b = 3), variables = list(val = c("a", "b")))

# Check whether `df` contains all factor analysis `variables`.
is_df_with_factors(
  df = data.frame(a = factor("A", levels = c("A", "B")), b = 3),
  variables = list(val = "a")
)

# Check whether `df` contains a factor `variable` with `n_levels` levels.
is_df_with_nlevels_factor(
  df = data.frame(a = factor("A", levels = c("A", "B")), b = 3),
  variable = "a",
  n_levels = 2
)

#' # Check whether `x` is a valid list of variable names.
a <- 1
b <- NULL
c <- c(1, "car")
d <- 5
is_equal_length(a, b, c, d)

# Check whether `x` is between 0 and 1.
is_proportion(x = 0.3)
is_proportion(x = 1.3)
is_proportion(x = 0, include_boundaries = TRUE)

# Check whether `x` is a vector of numbers between 0 and 1.
is_proportion_vector(c(0.3, 0.1))
is_proportion_vector(c(1.3, 0.1))
is_proportion_vector(0, include_boundaries = TRUE)

# Check whether `x` is a vector of sorted quantile proportions between 0 and 1.
is_quantiles_vector(c(0.1, 0.3))
is_quantiles_vector(c(0.3, 0.1))
is_quantiles_vector(c(0.3, 0.3))
is_quantiles_vector(0, include_boundaries = TRUE)

# Check whether `x` is a valid factor.
is_valid_factor(-1)
is_valid_factor(factor(c("a", "b")))
is_valid_factor(factor(c("a", "")))
is_valid_factor(factor())

# Check whether `NA` has been handled when `x` is a factor.
is_factor_no_na(factor(c("a", NA)))
is_factor_no_na(factor(c("a", "<Missing>")))

# Check whether `x` is a valid character vector
is_valid_character(-1)
is_valid_character(c("a", "b"))
is_valid_character(c("a", ""))
is_valid_character("")

# Check whether all elements of `x` are in a reference vector.
all_elements_in_ref(c("a", "b"), c("a", "b", "c"))
all_elements_in_ref(c("a", "d"), c("a", "b", "c"))
all_elements_in_ref(c(1:3), c(1:5))

df <- data.frame(a = 1:3, b = 2:4)
df$a <- ifelse(df$a == 1, "<Missing>", df$a)
is_df_with_no_na_level(df, variables = list(a = "a"), na_level = "<Missing>")
is_df_with_no_na_level(df, variables = list(b = "b"), na_level = "<Missing>")
is_df_with_no_na_level(df, variables = list(a = "a", b = "b"), na_level = "<Missing>")
}
\keyword{internal}
