% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland_altman.R
\name{bland_altman}
\alias{bland_altman}
\alias{s_bland_altman}
\alias{g_bland_altman}
\title{Bland Altman analysis}
\usage{
s_bland_altman(x, y, conf_level = 0.95)

g_bland_altman(x, y, conf_level = 0.95)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{y}{('numeric')\cr vector of numbers we want to analyze, which we want to compare with x.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Functions of bland altman method to assess the agreement between two numerical vectors.
}
\section{Functions}{
\itemize{
\item \code{s_bland_altman()}: 

\item \code{g_bland_altman()}: 

}}
\examples{
x <- seq(1, 60, 5)
y <- seq(5, 50, 4)
conf_level <- 0.9
# Derive statistics that are needed for Bland Altman plot
s_bland_altman(x, y, conf_level = conf_level)
# Create a Bland Altman plot
g_bland_altman(x, y, conf_level = conf_level)
}
