% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mmrm.R
\name{build_mmrm_formula}
\alias{build_mmrm_formula}
\title{Helper function to build the MMRM formula}
\usage{
build_mmrm_formula(vars, cor_struct)
}
\arguments{
\item{vars}{variable list as per \code{\link{s_mmrm}}}

\item{cor_struct}{correlation structure as per \code{\link{s_mmrm}}}
}
\value{
the complete MMRM formula to use with \code{lmer}
}
\description{
Helper function to build the MMRM formula
}
