% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rtables.R
\name{c_label_n}
\alias{c_label_n}
\title{Content Row Function to Add Row Total to Labels}
\usage{
c_label_n(df, labelstr, .N_row)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions).}

\item{.N_row}{(\code{count})\cr column-wise N (column count) for the full column that is passed by \code{rtables}.}
}
\value{
\code{CellValue} that just has the right label.
}
\description{
This takes the label of the latest row split level and adds the row total in parentheses.
}
\note{
Important is here to not use \code{df} but \code{.N_row} in the implementation, because the former
is already split by columns and will refer to the first column of the data only.
}
\keyword{internal}
