% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data_frame}
\alias{check_data_frame}
\title{Check strata_data}
\usage{
check_data_frame(x)
}
\arguments{
\item{x}{data.frame with valid stratification data}

\item{n}{number of expected rows}
}
\description{
strata_data can be NULL or
}
