% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data_frame}
\alias{check_data_frame}
\title{Checks a Data Frame}
\usage{
check_data_frame(x, allow_missing = FALSE)
}
\arguments{
\item{x}{(\verb{data frame})\cr which should be checked}

\item{allow_missing}{(\code{flag})\cr whether missing data is allowed in the variables}
}
\value{
\code{TRUE} or fails
}
\description{
Checks a Data Frame
}
