% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mmrm.R
\name{check_mmrm_vars}
\alias{check_mmrm_vars}
\title{Helper function to check the MMRM variables}
\usage{
check_mmrm_vars(vars, data)
}
\arguments{
\item{vars}{variable list as per \code{\link{s_mmrm}}}

\item{data}{data frame that includes the variables}
}
\value{
a corresponding list of variable labels.
  For the covariates, the element \code{parts} contains a character vector with one element
  per covariates part (i.e. occurring variable in the original covariates specification) is
  included.
  If a label is not found, the variable name is instead used.
}
\description{
Helper function to check the MMRM variables
}
