% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_same_n}
\alias{check_same_n}
\title{Check Element Dimension}
\usage{
check_same_n(..., omit_null = TRUE)
}
\arguments{
\item{...}{data.frames or vectors}

\item{omit_null}{are \code{NULL} elements in \code{...} to be omitted from the check?}
}
\description{
Checks if the elements in \code{...} have the same dimension.
}
\keyword{internal}
