% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chgfbl_table.R
\name{chgfbl_data}
\alias{chgfbl_data}
\title{Change from Baseline Table Caclulation Function}
\usage{
chgfbl_data(data, arm_var = "ARM")
}
\arguments{
\item{data}{input data}

\item{arm_var}{variable to use for arm/grouping}
}
\description{
Change from Baseline Table Caclulation Function
}
\examples{

library(tern)
library(random.cdisc.data)

ASL <- radam("ASL")

AQS <- merge(ASL, radam("AQS", ADSL = ASL)) \%>\% 
   filter(PARAMCD == "FATIGI")

df <- chgfbl_data(data = ANL)

\dontrun{
library(dplyr)
library(tidyr)
library(atezo.data)
library(teal.oncology)
library(forcats)
library(ggplot2)
library(grid)
library(gridExtra)

ASL <- asl(com.roche.cdpt7722.wo29637.rl)
AQS <- aqs(com.roche.cdpt7722.wo29637.rl)

#parammap <- unique(AQS[c("PARAM", "PARAMCD")]) 

ASL_f <- ASL \%>\% select(STUDYID, USUBJID, ARM, ARMCD)
AQS_f <- AQS \%>\% select(STUDYID, USUBJID, PARAMCD, PARAM, AVAL, AVISIT, AVISITN, ADY, ABLFL, APBFL, CHG, BASE)
#AVALU variable not available in AQS
#AQS_f$AVALU <- "kg/m^2"

ANL <- left_join(ASL_f, AQS_f, by = c("STUDYID", "USUBJID")) \%>\% filter(PARAMCD == "FATIGI")

df <- chgfbl_data(data = ANL)
chgfbl_plot(data = df, ytype = "CHG", errbar = "SE", ref_line = c(2, -2), fontsize = 14)
chgfbl_plot(data = df, ytype = "AVAL", errbar = "95CI", ref_line = 3, yrange = c(0,3), n_rotate = T)

tbl <- chgfbl_table(data=df)

\dontrun{
Viewer(tbl)
}



\dontrun{
library(random.cdisc.data)
library(dplyr)
ASL <- radam("ASL")
ANL <- merge(ASL, radam("AQS", ADSL = ASL)) \%>\% filter(PARAMCD == "FATIGI")

df <- chgfbl_data(data = ANL)
chgfbl_plot(data = df, ytype = "CHG", errbar = "SE", ref_line = c(2, -2), fontsize = 14)

}
# now 
# chng_data <- ... # you already have this # do not export
# change_table(chng_data) # split by arm_name make columns and then combine
# change_plot(chng_data) # group_by(arm_name) make plotA

# maybe in future
# chng_table <- table...
# change_plot(chng_table)
# selection for chg_var can be "CHG" or "PCHG"
}
}
\author{
Chendi Liao (liaoc10), \email{chendi.liao@roche.com}
}
