% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clogit_with_tryCatch}
\alias{clogit_with_tryCatch}
\title{Wrapper function of survival::clogit so that when model fitting failed, a more useful message would show}
\usage{
clogit_with_tryCatch(formula, data, ...)
}
\description{
Wrapper function of survival::clogit so that when model fitting failed, a more useful message would show
}
