% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cols_by_groups.R
\name{combine_groups}
\alias{combine_groups}
\title{Reference and Treatment Group Combination}
\usage{
combine_groups(fct, ref = NULL, collapse = "/")
}
\arguments{
\item{fct}{(\code{factor})\cr the variable with levels which needs to be grouped.}

\item{ref}{(\code{string})\cr the reference level(s).}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.}
}
\value{
a \code{list} with first item \code{ref} (reference) and second item \code{trt}
(treatment).
}
\description{
Facilitate the re-combination of groups divided as reference and
treatment groups; it helps in arranging groups of
columns in the \code{rtables} framework and teal modules.
}
\examples{

library(rtables)

groups <- combine_groups(
  fct = DM$ARM,
  ref = c("B: Placebo")
)

basic_table() \%>\%
  split_cols_by_groups("ARM", groups) \%>\%
  add_colcounts() \%>\%
  summarize_vars("AGE") \%>\%
  build_table(DM)

}
