% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_rtables.R
\name{combine_rrows_with_indent}
\alias{combine_rrows_with_indent}
\title{Combines two lists of rrows to a list of rrows
Also keeps indent.}
\usage{
combine_rrows_with_indent(x, y)
}
\arguments{
\item{x}{first list of rows, row.names are taken from this list}

\item{y}{second list}
}
\value{
list of combined rows
}
\description{
todo: this should eventually replace rtables:::combine_rrows
}
