% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_tte.R
\name{control_coxph}
\alias{control_coxph}
\title{Control}
\usage{
control_coxph(
  pval_method = c("log-rank", "wald", "likelihood"),
  ties = c("efron", "breslow", "exact"),
  conf_level = 0.95
)
}
\arguments{
\item{pval_method}{The method to be passed as argument \code{pval_method} to function
\code{\link{s_coxph_pairwise}}. Default method is "log-rank", can also be set to "wald" or "likelihood".}

\item{ties}{The method to be passed as argument \code{ties} to \code{\link{coxph}}. Default is "efron",
can also be set to "breslow" or "exact".}

\item{conf_level}{(\code{numeric} value) passed as argument \code{conf_level} to function
\code{\link{s_coxph_pairwise}}.}
}
\value{
a list of components named as the arguments
}
\description{
Control function for \code{CoxPH} model
}
\details{
This is an auxiliary function for controlling arguments for \code{CoxPH} model, typically used internally to specify
details of \code{CoxPH} model for \code{\link{s_coxph_pairwise}}.
}
