% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_incidence_rate.R
\name{control_incidence_rate}
\alias{control_incidence_rate}
\title{Control function for incidence rate}
\usage{
control_incidence_rate(
  conf_level = 0.95,
  conf_type = c("normal", "normal_log", "exact", "byar"),
  time_unit = 1
)
}
\arguments{
\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{conf_type}{(\code{string}) \cr \code{normal} (default), \code{normal_log}, \code{exact}, or \code{byar}
for confidence interval type.}

\item{time_unit}{(\code{numeric}) \cr unit time in person-years.}
}
\value{
A list of components with the same name as the arguments.
}
\description{
This is an auxiliary function for controlling arguments for the incidence rate, used
internally to specify details in \code{\link[=s_incidence_rate]{s_incidence_rate()}}.
}
\examples{
control_incidence_rate(0.9, "exact", 100)

}
