% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_lineplot.R
\name{control_lineplot_vars}
\alias{control_lineplot_vars}
\title{Control Function for \code{g_lineplot} Function}
\usage{
control_lineplot_vars(
  x = "AVISIT",
  y = "AVAL",
  group_var = "ARM",
  paramcd = "PARAMCD",
  y_unit = "AVALU",
  subject_var = "USUBJID",
  strata = lifecycle::deprecated(),
  cohort_id = lifecycle::deprecated()
)
}
\arguments{
\item{x}{(\code{character})\cr x variable name.}

\item{y}{(\code{character})\cr y variable name.}

\item{group_var}{(\code{character} or \code{NA})\cr group variable name.}

\item{paramcd}{(\code{character} or \code{NA})\cr \code{paramcd} variable name.}

\item{y_unit}{(\code{character} or \code{NA})\cr \code{y_unit} variable name.}

\item{subject_var}{(\code{character} or \code{NA})\cr subject variable name.}

\item{strata}{(\code{character} or \code{NA})\cr deprecated - group variable name.}

\item{cohort_id}{(\code{character} or \code{NA})\cr deprecated - subject variable name.}
}
\value{
A named character vector of variable names.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Default values for \code{variables} parameter in \code{g_lineplot} function.
A variable's default value can be overwritten for any variable.
}
\examples{
control_lineplot_vars()
control_lineplot_vars(group_var = NA)

}
