% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_lineplot.R
\name{control_lineplot_vars}
\alias{control_lineplot_vars}
\title{Control Function for g_lineplot Function}
\usage{
control_lineplot_vars()
}
\value{
A character vector with names of variables.
}
\description{
Keeps default values for a vector of variables for g_lineplot function.
}
