% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_lineplot.R
\name{control_lineplot_vars}
\alias{control_lineplot_vars}
\title{Control Function for g_lineplot Function}
\usage{
control_lineplot_vars(
  x = "AVISIT",
  y = "AVAL",
  strata = "ARM",
  paramcd = "PARAMCD",
  y_unit = "AVALU"
)
}
\arguments{
\item{x}{(\code{character} scalar) x variable name. \cr}

\item{y}{(\code{character} scalar) y variable name. \cr}

\item{strata}{(\code{character} scalar or \code{NA}) strata variable name. \cr}

\item{paramcd}{(\code{character} scalar or \code{NA}) paramcd variable name. \cr}

\item{y_unit}{(\code{character} scalar or \code{NA}) y_unit variable name. \cr}
}
\value{
A named character vector with names of variables.
}
\description{
Default values for \code{variables} parameter in \code{g_lineplot} function.
A variable's default value can be overwritten for any variable.
}
\examples{
control_lineplot_vars()
control_lineplot_vars(strata = NA)
}
\author{
Wojciech Wojciak wojciech.wojciak@contractors.roche.com
}
