% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_cumulative.R
\name{count_cumulative}
\alias{count_cumulative}
\alias{h_count_cumulative}
\alias{s_count_cumulative}
\alias{s_count_nonmissing}
\title{Cumulative Counts with Thresholds}
\usage{
h_count_cumulative(
  x,
  threshold,
  lower_tail = TRUE,
  include_eq = TRUE,
  na.rm = TRUE,
  .N_col
)

s_count_cumulative(
  x,
  thresholds,
  lower_tail = TRUE,
  include_eq = TRUE,
  .N_col,
  ...
)

s_count_nonmissing(x)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{threshold}{(\code{number})\cr a cutoff value as threshold to count values of \code{x}.}

\item{lower_tail}{(\code{logical})\cr whether to count lower tail, default is \code{TRUE}.}

\item{include_eq}{(\code{logical})\cr whether to include value equal to the \code{threshold} in
count, default is \code{TRUE}.}

\item{na.rm}{(\code{flag})\cr whether \code{NA} values should be removed from \code{x} prior to analysis.}

\item{.N_col}{(\code{count})\cr denominator for percentage calculation.}

\item{thresholds}{(\code{numeric})\cr vector of cutoff value for the counts.}

\item{...}{additional arguments for the lower level functions.}
}
\value{
a named vector of 2:
\itemize{
\item \code{count}: the count of values less than, less or equal to, greater than, or
greater or equal to a threshold of user specification.
\item \code{percent}: the percentage of the count.
}

a named list:
\itemize{
\item \code{count_percent}: a list with each \code{thresholds} value as a component, each component
contains a vector for the count and percentage.
}

a list with named item:
\itemize{
\item \code{n}: the count of non-missing values.
}
}
\description{
Summarize cumulative counts of a (\code{numeric}) vector that is less than, less or equal to,
greater than, or greater or equal to user-specific thresholds.
}
\section{Functions}{
\itemize{
\item \code{h_count_cumulative}: Helper function to calculate count and percentage of
\code{x} values in the lower or upper tail given a threshold.

\item \code{s_count_cumulative}: Statistics function that produces a named lists given a
(\code{numeric}) vector of thresholds.

\item \code{s_count_nonmissing}: Statistics function to count non-missing values.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{
set.seed(1)
x <- c(sample(1:10, 10), NA)
.N_col <- length(x)
h_count_cumulative(x, 5, .N_col = .N_col)
h_count_cumulative(x, 5, lower_tail = FALSE, include_eq = FALSE, na.rm = FALSE, .N_col = .N_col)
h_count_cumulative(x, 0, lower_tail = FALSE, .N_col = .N_col)
h_count_cumulative(x, 100, lower_tail = FALSE, .N_col = .N_col)

s_count_cumulative(x, thresholds = c(0, 5, 11), .N_col = .N_col)
s_count_cumulative(x, thresholds = c(0, 5, 11), include_eq = FALSE, na.rm = FALSE, .N_col = .N_col)

set.seed(1)
x <- c(sample(1:10, 10), NA)
s_count_nonmissing(x)
}
