% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_cumulative.R
\name{count_cumulative}
\alias{count_cumulative}
\alias{h_count_cumulative}
\alias{s_count_cumulative}
\alias{a_count_cumulative}
\title{Cumulative Counts with Thresholds}
\usage{
h_count_cumulative(
  x,
  threshold,
  lower_tail = TRUE,
  include_eq = TRUE,
  na.rm = TRUE,
  .N_col
)

s_count_cumulative(
  x,
  thresholds,
  lower_tail = TRUE,
  include_eq = TRUE,
  .N_col,
  ...
)

a_count_cumulative(
  x,
  thresholds,
  lower_tail = TRUE,
  include_eq = TRUE,
  .N_col,
  ...
)

count_cumulative(
  lyt,
  vars,
  var_labels = vars,
  show_labels = "visible",
  ...,
  table_names = vars,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{threshold}{(\code{number})\cr a cutoff value as threshold to count values of \code{x}.}

\item{lower_tail}{(\code{logical})\cr whether to count lower tail, default is \code{TRUE}.}

\item{include_eq}{(\code{logical})\cr whether to include value equal to the \code{threshold} in
count, default is \code{TRUE}.}

\item{na.rm}{(\code{flag})\cr whether \code{NA} values should be removed from \code{x} prior to analysis.}

\item{.N_col}{(\code{count})\cr denominator for fraction calculation.}

\item{thresholds}{(\code{numeric})\cr vector of cutoff value for the counts.}

\item{...}{additional arguments for the lower level functions.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{var_labels}{character for label.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
A named vector of 2:
\itemize{
\item \code{count}: the count of values less than, less or equal to, greater than, or
greater or equal to a threshold of user specification.
\item \code{fraction}: the fraction of the count.
}

A named list:
\itemize{
\item \code{count_fraction}: a list with each \code{thresholds} value as a component, each component
contains a vector for the count and fraction.
}

\code{\link[=a_count_cumulative]{a_count_cumulative()}} returns the corresponding list with formatted \code{\link[rtables:CellValue]{rtables::CellValue()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Summarize cumulative counts of a (\code{numeric}) vector that is less than, less or equal to,
greater than, or greater or equal to user-specific thresholds.
}
\section{Functions}{
\itemize{
\item \code{h_count_cumulative}: Helper function to calculate count and fraction of
\code{x} values in the lower or upper tail given a threshold.

\item \code{s_count_cumulative}: Statistics function that produces a named lists given a
(\code{numeric}) vector of thresholds.

\item \code{a_count_cumulative}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{count_cumulative}: Layout creating function which can be be used for creating
summary tables for cumulative counts of a variable. The ellipsis (\code{...}) conveys
arguments to \code{s_count_cumulative()}, for instance \code{lower_tail = FALSE} if upper tail
should be accounted for.
}}

\examples{
set.seed(1, kind = "Mersenne-Twister")
x <- c(sample(1:10, 10), NA)
.N_col <- length(x)
h_count_cumulative(x, 5, .N_col = .N_col)
h_count_cumulative(x, 5, lower_tail = FALSE, include_eq = FALSE, na.rm = FALSE, .N_col = .N_col)
h_count_cumulative(x, 0, lower_tail = FALSE, .N_col = .N_col)
h_count_cumulative(x, 100, lower_tail = FALSE, .N_col = .N_col)
# Internal function - s_count_cumulative
\dontrun{
s_count_cumulative(x, thresholds = c(0, 5, 11), .N_col = .N_col)
s_count_cumulative(x, thresholds = c(0, 5, 11), include_eq = FALSE, na.rm = FALSE, .N_col = .N_col)
}

# Internal function - a_count_cumulative
\dontrun{
# Use the Formatted Analysis function for `analyze()`. We need to ungroup `count_fraction` first
# so that the `rtables` formatting function `format_count_fraction()` can be applied correctly.
afun <- make_afun(a_count_cumulative, .ungroup_stats = "count_fraction")
afun(x, thresholds = c(0, 5, 11), .N_col = .N_col)
}

library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  add_colcounts() \%>\%
  count_cumulative(
    vars = "AGE",
    thresholds = c(40, 60)
  ) \%>\%
  build_table(ADSL)
}
\keyword{internal}
