% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_occurrences.R
\name{count_occurrences}
\alias{count_occurrences}
\alias{s_count_occurrences}
\title{Occurrence Counts}
\usage{
s_count_occurrences(df, .N_col, .var = "MHDECOD", id = "USUBJID")

count_occurrences(lyt, vars, ...)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.N_col}{(\code{count})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is passed by \code{rtables}.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{id}{(\code{string}) \cr subject variable.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{...}{additional arguments for the lower level functions.}
}
\value{
A list with:
\itemize{
\item \code{count_percent}: list of counts and percentages with one element per occurrence.
}
}
\description{
Functions for analyzing frequencies and percentages of occurrences for patients with occurrence
data. Primary analysis variables are the dictionary terms. All occurrences are counted for total
counts. Multiple occurrences within patient at the lowest term level displayed in the table are
counted only once.
}
\section{Functions}{
\itemize{
\item \code{s_count_occurrences}: Statistics function which counts number of patients that report an
occurrence.

\item \code{count_occurrences}: Analyze Function that counts occurrences as part of rtables layouts.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{

df <- data.frame(
  USUBJID = as.character(c(1, 1, 2, 4, 4, 4)),
  MHDECOD = c("MH1", "MH2", "MH1", "MH1", "MH1", "MH3")
)

# Number of subjects in column being processed
N_per_col <- 4

# Count unique occurrences per subject
s_count_occurrences (df,
                    N_per_col,
                    .var = "MHDECOD",
                    id = "USUBJID")


df <- data.frame(
  USUBJID = as.character(c(1, 1, 2, 4, 4, 4,
                           6, 6, 6, 7, 7, 8)),
  MHDECOD = c("MH1", "MH2", "MH1", "MH1", "MH1", "MH3",
              "MH2", "MH2", "MH3", "MH1", "MH2", "MH4"),
  ARM = rep(c("A", "B"), each=6)
)
N_per_arm <- c(5L, 4L)

# Create table layout
lyt <- basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  add_colcounts() \%>\%
  count_occurrences(vars = "MHDECOD")

# Apply table layout to data and produce rtable object
lyt \%>\%
  build_table(df, col_counts = N_per_arm) \%>\%
  prune_table()

}
