% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_occurrences.R
\name{count_occurrences}
\alias{count_occurrences}
\alias{s_count_occurrences}
\alias{a_count_occurrences}
\title{Occurrence Counts}
\usage{
s_count_occurrences(df, .N_col, .var = "MHDECOD", id = "USUBJID")

a_count_occurrences(df, .N_col, .var = "MHDECOD", id = "USUBJID")

count_occurrences(
  lyt,
  vars,
  ...,
  .stats = "count_fraction",
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.N_col}{(\code{count})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is passed by \code{rtables}.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{id}{(\code{string}) \cr subject variable.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{...}{additional arguments for the lower level functions.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
A list with:
\itemize{
\item \code{count}: list of counts with one element per occurrence
\item \code{count_fraction}: list of counts and fractions with one element per occurrence.
}
}
\description{
Functions for analyzing frequencies and fractions of occurrences for patients with occurrence
data. Primary analysis variables are the dictionary terms. All occurrences are counted for total
counts. Multiple occurrences within patient at the lowest term level displayed in the table are
counted only once.
}
\section{Functions}{
\itemize{
\item \code{s_count_occurrences}: Statistics function which counts number of patients that report an
occurrence.

\item \code{a_count_occurrences}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{count_occurrences}: Analyze Function that counts occurrences as part of rtables layouts.
}}

\examples{
df <- data.frame(
  USUBJID = as.character(c(1, 1, 2, 4, 4, 4)),
  MHDECOD = c("MH1", "MH2", "MH1", "MH1", "MH1", "MH3")
)

N_per_col <- 4L

# Count unique occurrences per subject.
s_count_occurrences(
  df,
  N_per_col,
  .var = "MHDECOD",
  id = "USUBJID"
)

a_count_occurrences(
  df,
  N_per_col,
  .var = "MHDECOD",
  id = "USUBJID"
)


df <- data.frame(
  USUBJID = as.character(c(1, 1, 2, 4, 4, 4,
                           6, 6, 6, 7, 7, 8)),
  MHDECOD = c("MH1", "MH2", "MH1", "MH1", "MH1", "MH3",
              "MH2", "MH2", "MH3", "MH1", "MH2", "MH4"),
  ARM = rep(c("A", "B"), each=6)
)
N_per_arm <- c(5L, 4L)

# Create table layout
lyt <- basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  add_colcounts() \%>\%
  count_occurrences(vars = "MHDECOD", .stats = c("count"))

# Apply table layout to data and produce rtable object
lyt \%>\%
  build_table(df, col_counts = N_per_arm) \%>\%
  prune_table()

}
