% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_occurrences_by_grade.R
\name{count_occurrences_by_grade}
\alias{count_occurrences_by_grade}
\alias{summarize_occurrences_by_grade}
\alias{s_count_occurrences_by_grade}
\alias{a_count_occurrences_by_grade}
\title{Count occurrences by grade}
\usage{
count_occurrences_by_grade(
  lyt,
  var,
  id = "USUBJID",
  grade_groups = list(),
  remove_single = TRUE,
  only_grade_groups = FALSE,
  var_labels = var,
  show_labels = "default",
  riskdiff = FALSE,
  na_str = default_na_str(),
  nested = TRUE,
  ...,
  table_names = var,
  .stats = "count_fraction",
  .stat_names = NULL,
  .formats = list(count_fraction = format_count_fraction_fixed_dp),
  .indent_mods = NULL,
  .labels = NULL
)

summarize_occurrences_by_grade(
  lyt,
  var,
  id = "USUBJID",
  grade_groups = list(),
  remove_single = TRUE,
  only_grade_groups = FALSE,
  riskdiff = FALSE,
  na_str = default_na_str(),
  ...,
  .stats = "count_fraction",
  .stat_names = NULL,
  .formats = list(count_fraction = format_count_fraction_fixed_dp),
  .indent_mods = NULL,
  .labels = NULL
)

s_count_occurrences_by_grade(
  df,
  labelstr = "",
  .var,
  .N_row,
  .N_col,
  ...,
  id = "USUBJID",
  grade_groups = list(),
  remove_single = TRUE,
  only_grade_groups = FALSE,
  denom = c("N_col", "n", "N_row")
)

a_count_occurrences_by_grade(
  df,
  labelstr = "",
  ...,
  .stats = NULL,
  .stat_names = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{lyt}{(\code{PreDataTableLayouts})\cr layout that analyses will be added to.}

\item{id}{(\code{string})\cr subject variable name.}

\item{grade_groups}{(named \code{list} of \code{character})\cr list containing groupings of grades.}

\item{remove_single}{(\code{flag})\cr \code{TRUE} to not include the elements of one-element grade groups
in the the output list; in this case only the grade groups names will be included in the output. If
\code{only_grade_groups} is set to \code{TRUE} this argument is ignored.}

\item{only_grade_groups}{(\code{flag})\cr whether only the specified grade groups should be
included, with individual grade rows removed (\code{TRUE}), or all grades and grade groups
should be displayed (\code{FALSE}).}

\item{var_labels}{(\code{character})\cr variable labels.}

\item{show_labels}{(\code{string})\cr label visibility: one of "default", "visible" and "hidden".}

\item{riskdiff}{(\code{flag})\cr whether a risk difference column is present. When set to \code{TRUE}, \code{\link[=add_riskdiff]{add_riskdiff()}} must be
used as \code{split_fun} in the prior column split of the table layout, specifying which columns should be compared.
See \code{\link[=stat_propdiff_ci]{stat_propdiff_ci()}} for details on risk difference calculation.}

\item{na_str}{(\code{string})\cr string used to replace all \code{NA} or empty values in the output.}

\item{nested}{(\code{flag})\cr whether this layout instruction should be applied within the existing layout structure _if
possible (\code{TRUE}, the default) or as a new top-level element (\code{FALSE}). Ignored if it would nest a split.
underneath analyses, which is not allowed.}

\item{...}{additional arguments for the lower level functions.}

\item{table_names}{(\code{character})\cr this can be customized in the case that the same \code{vars} are analyzed multiple
times, to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.

Options are: \verb{'count_fraction', 'count_fraction_fixed_dp'}}

\item{.stat_names}{(\code{character})\cr names of the statistics that are passed directly to name single statistics
(\code{.stats}). This option is visible when producing \code{\link[rtables:data.frame_export]{rtables::as_result_df()}} with \code{make_ard = TRUE}.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics. See Details in \code{analyze_vars} for more
information on the \code{"auto"} setting.}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels. Defaults to 0, which corresponds to the
unmodified default behavior. Can be negative.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{df}{(\code{data.frame})\cr data set containing all analysis variables.}

\item{labelstr}{(\code{string})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions). See \code{\link[rtables:summarize_row_groups]{rtables::summarize_row_groups()}}
for more information.}

\item{.var, var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.N_row}{(\code{integer(1)})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is typically passed by \code{rtables}.}

\item{.N_col}{(\code{integer(1)})\cr column-wise N (column count) for the full column being analyzed that is typically
passed by \code{rtables}.}

\item{denom}{(\code{string})\cr choice of denominator for proportion. Options are:
\itemize{
\item \code{N_col}: total number of patients in this column across rows.
\item \code{n}: number of patients with any occurrences.
\item \code{N_row}: total number of patients in this row across columns.
}}
}
\value{
\itemize{
\item \code{count_occurrences_by_grade()} returns a layout object suitable for passing to further layouting functions,
or to \code{\link[rtables:build_table]{rtables::build_table()}}. Adding this function to an \code{rtable} layout will add formatted rows containing
the statistics from \code{s_count_occurrences_by_grade()} to the table layout.
}

\itemize{
\item \code{summarize_occurrences_by_grade()} returns a layout object suitable for passing to further layouting functions,
or to \code{\link[rtables:build_table]{rtables::build_table()}}. Adding this function to an \code{rtable} layout will add formatted content rows
containing the statistics from \code{s_count_occurrences_by_grade()} to the table layout.
}

\itemize{
\item \code{s_count_occurrences_by_grade()} returns a list of counts and fractions with one element per grade level or
grade level grouping.
}

\itemize{
\item \code{a_count_occurrences_by_grade()} returns the corresponding list with formatted \code{\link[rtables:CellValue]{rtables::CellValue()}}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The analyze function \code{\link[=count_occurrences_by_grade]{count_occurrences_by_grade()}} creates a layout element to calculate occurrence counts by grade.

This function analyzes primary analysis variable \code{var} which indicates toxicity grades. The \code{id} variable
is used to indicate unique subject identifiers (defaults to \code{USUBJID}). The user can also supply a list of
custom groups of grades to analyze via the \code{grade_groups} parameter. The \code{remove_single}  argument will
remove single grades from the analysis so that \emph{only} grade groups are analyzed.

If there are multiple grades recorded for one patient only the highest grade level is counted.

The summarize function \code{\link[=summarize_occurrences_by_grade]{summarize_occurrences_by_grade()}} performs the same function as
\code{\link[=count_occurrences_by_grade]{count_occurrences_by_grade()}} except it creates content rows, not data rows, to summarize the current table
row/column context and operates on the level of the latest row split or the root of the table if no row splits have
occurred.
}
\section{Functions}{
\itemize{
\item \code{count_occurrences_by_grade()}: Layout-creating function which can take statistics function
arguments and additional format arguments. This function is a wrapper for \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{summarize_occurrences_by_grade()}: Layout-creating function which can take content function arguments
and additional format arguments. This function is a wrapper for \code{\link[rtables:summarize_row_groups]{rtables::summarize_row_groups()}}.

\item \code{s_count_occurrences_by_grade()}: Statistics function which counts the
number of patients by highest grade.

\item \code{a_count_occurrences_by_grade()}: Formatted analysis function which is used as \code{afun}
in \code{count_occurrences_by_grade()}.

}}
\examples{
library(dplyr)

df <- data.frame(
  USUBJID = as.character(c(1:6, 1)),
  ARM = factor(c("A", "A", "A", "B", "B", "B", "A"), levels = c("A", "B")),
  AETOXGR = factor(c(1, 2, 3, 4, 1, 2, 3), levels = c(1:5)),
  AESEV = factor(
    x = c("MILD", "MODERATE", "SEVERE", "MILD", "MILD", "MODERATE", "SEVERE"),
    levels = c("MILD", "MODERATE", "SEVERE")
  ),
  stringsAsFactors = FALSE
)

df_adsl <- df \%>\%
  select(USUBJID, ARM) \%>\%
  unique()

# Layout creating function with custom format.
basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  add_colcounts() \%>\%
  count_occurrences_by_grade(
    var = "AESEV",
    .formats = c("count_fraction" = "xx.xx (xx.xx\%)")
  ) \%>\%
  build_table(df, alt_counts_df = df_adsl)

# Define additional grade groupings.
grade_groups <- list(
  "-Any-" = c("1", "2", "3", "4", "5"),
  "Grade 1-2" = c("1", "2"),
  "Grade 3-5" = c("3", "4", "5")
)

basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  add_colcounts() \%>\%
  count_occurrences_by_grade(
    var = "AETOXGR",
    grade_groups = grade_groups,
    only_grade_groups = TRUE
  ) \%>\%
  build_table(df, alt_counts_df = df_adsl)

# Layout creating function with custom format.
basic_table() \%>\%
  add_colcounts() \%>\%
  split_rows_by("ARM", child_labels = "visible", nested = TRUE) \%>\%
  summarize_occurrences_by_grade(
    var = "AESEV",
    .formats = c("count_fraction" = "xx.xx (xx.xx\%)")
  ) \%>\%
  build_table(df, alt_counts_df = df_adsl)

basic_table() \%>\%
  add_colcounts() \%>\%
  split_rows_by("ARM", child_labels = "visible", nested = TRUE) \%>\%
  summarize_occurrences_by_grade(
    var = "AETOXGR",
    grade_groups = grade_groups
  ) \%>\%
  build_table(df, alt_counts_df = df_adsl)

s_count_occurrences_by_grade(
  df,
  .N_col = 10L,
  .var = "AETOXGR",
  id = "USUBJID",
  grade_groups = list("ANY" = levels(df$AETOXGR))
)

a_count_occurrences_by_grade(
  df,
  .N_col = 10L,
  .N_row = 10L,
  .var = "AETOXGR",
  id = "USUBJID",
  grade_groups = list("ANY" = levels(df$AETOXGR))
)

}
\seealso{
Relevant helper function \code{\link[=h_append_grade_groups]{h_append_grade_groups()}}.
}
