% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_occurrences_by_grade.R
\name{count_occurrences_by_grade}
\alias{count_occurrences_by_grade}
\alias{h_append_grade_groups}
\alias{s_count_occurrences_by_grade}
\title{Occurrence Counts by Grade}
\usage{
h_append_grade_groups(grade_groups, refs)

s_count_occurrences_by_grade(
  df,
  .var,
  .N_col,
  id = "USUBJID",
  grade_groups = list(),
  labelstr = ""
)
}
\arguments{
\item{grade_groups}{(named \code{list} of \code{character})\cr containing groupings of grades.}

\item{refs}{(named \code{list} of \code{numeric})\cr where each name corresponds to a reference grade level
and each entry represents a count.}

\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.N_col}{(\code{count})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is passed by \code{rtables}.}

\item{id}{(\code{string}) \cr subject variable.}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions).}
}
\description{
Functions for analyzing frequencies and percentages of occurrences by grade for patients
with occurrence data. Multiple occurrences within one individual are counted once at the
greatest intensity/highest grade level.
}
\section{Functions}{
\itemize{
\item \code{h_append_grade_groups}: Helper function for \code{\link[=s_count_occurrences_by_grade]{s_count_occurrences_by_grade()}} to
insert grade groupings into list with individual grade frequencies.

\item \code{s_count_occurrences_by_grade}: Statistics function which given occurrence data counts the
number of patients by highest grade. Returns a list of counts and percentages with one element
per grade level or grade level grouping.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{
h_append_grade_groups(
  list(
    "Any Grade" = as.character(1:3),
     "Grade 1-2" = c("1", "2")
    ),
  list("1" = 10, "2" = 7, "3" = 2)
 )


df <- data.frame(
  USUBJID = as.character(c(1:5, 1)),
  ARM = factor(c("A", "A", "A", "B", "B", "A"), levels = c("A", "B")),
  AETOXGR = factor(c(1, 2, 3, 1, 2, 3), levels = c(1:5)),
  stringsAsFactors = FALSE
)

s_count_occurrences_by_grade(df,
                    .N_col = 10L, # nolint
                    .var = "AETOXGR",
                    id = "USUBJID",
                    grade_groups = list("ANY" = levels(df$AETOXGR)))

}
