% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_occurrences_by_grade.R
\name{count_occurrences_by_grade}
\alias{count_occurrences_by_grade}
\alias{h_append_grade_groups}
\alias{s_count_occurrences_by_grade}
\alias{a_count_occurrences_by_grade}
\alias{summarize_occurrences_by_grade}
\title{Occurrence Counts by Grade}
\usage{
h_append_grade_groups(grade_groups, refs)

s_count_occurrences_by_grade(
  df,
  .var,
  .N_col,
  id = "USUBJID",
  grade_groups = list(),
  labelstr = ""
)

a_count_occurrences_by_grade(
  df,
  .var,
  .N_col,
  id = "USUBJID",
  grade_groups = list(),
  labelstr = ""
)

count_occurrences_by_grade(
  lyt,
  var,
  ...,
  table_names = var,
  .stats = NULL,
  .formats = NULL,
  .indent_mods = NULL,
  .labels = NULL
)

summarize_occurrences_by_grade(
  lyt,
  var,
  ...,
  .stats = NULL,
  .formats = NULL,
  .indent_mods = NULL,
  .labels = NULL
)
}
\arguments{
\item{grade_groups}{(named \code{list} of \code{character})\cr containing groupings of grades.}

\item{refs}{(named \code{list} of \code{numeric})\cr where each name corresponds to a reference grade level
and each entry represents a count.}

\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.N_col}{(\code{count})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is passed by \code{rtables}.}

\item{id}{(\code{string}) \cr subject variable name.}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions).}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{var}{(\code{string})\cr single variable name for the primary analysis variable.}

\item{...}{additional arguments for the lower level functions.}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}
}
\description{
Functions for analyzing frequencies and fractions of occurrences by grade for patients
with occurrence data. Multiple occurrences within one individual are counted once at the
greatest intensity/highest grade level.
}
\section{Functions}{
\itemize{
\item \code{h_append_grade_groups}: Helper function for \code{\link[=s_count_occurrences_by_grade]{s_count_occurrences_by_grade()}} to
insert grade groupings into list with individual grade frequencies.

\item \code{s_count_occurrences_by_grade}: Statistics function which given occurrence data counts the
number of patients by highest grade. Returns a list of counts and fractions with one element
per grade level or grade level grouping.

\item \code{a_count_occurrences_by_grade}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{count_occurrences_by_grade}: Layout creating function which can be used for creating tables,
which can take statistics function arguments and additional format arguments (see below).

\item \code{summarize_occurrences_by_grade}: Layout creating function which adds content rows using the
statistics function and additional format arguments (see below).
}}

\examples{
h_append_grade_groups(
  list(
    "Any Grade" = as.character(1:3),
     "Grade 1-2" = c("1", "2")
    ),
  list("1" = 10, "2" = 7, "3" = 2)
 )


df <- data.frame(
  USUBJID = as.character(c(1:6, 1)),
  ARM = factor(c("A", "A", "A", "B", "B", "B", "A"), levels = c("A", "B")),
  AETOXGR = factor(c(1, 2, 3, 4, 1, 2, 3), levels = c(1:5)),
  AESEV = factor(
    x = c("MILD", "MODERATE", "SEVERE", "MILD",  "MILD", "MODERATE", "SEVERE"),
    levels = c("MILD", "MODERATE", "SEVERE")
  ),
  stringsAsFactors = FALSE
)

s_count_occurrences_by_grade(
  df,
  .N_col = 10L, # nolint
  .var = "AETOXGR",
  id = "USUBJID",
  grade_groups = list("ANY" = levels(df$AETOXGR))
)

a_count_occurrences_by_grade(
  df,
  .N_col = 10L, # nolint
  .var = "AETOXGR",
  id = "USUBJID",
  grade_groups = list("ANY" = levels(df$AETOXGR))
)


# Layout creating function with custom format.
basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  add_colcounts() \%>\%
  count_occurrences_by_grade(
    var = "AESEV",
    .formats = c("count_fraction" = "xx.xx (xx.xx\%)")
  ) \%>\%
  build_table(df, col_counts = c(3L, 3L))

# Define additional grade groupings.
grade_groups <-  list(
  "-Any-" = c("1", "2", "3", "4", "5"),
  "Grade 1-2" = c("1", "2"),
  "Grade 3-5" = c("3", "4", "5")
)

basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  add_colcounts() \%>\%
  count_occurrences_by_grade(
    var = "AETOXGR",
    grade_groups = grade_groups
  ) \%>\%
  build_table(df, col_counts = c(3L, 3L))


# Layout creating function with custom format.
basic_table() \%>\%
  add_colcounts() \%>\%
  split_rows_by("ARM", child_labels = "visible", nested = TRUE) \%>\%
  summarize_occurrences_by_grade(
  var = "AESEV",
  .formats = c("count_fraction" = "xx.xx (xx.xx\%)")) \%>\%
  build_table(df, col_counts = 10L)

basic_table() \%>\%
  add_colcounts() \%>\%
  split_rows_by("ARM", child_labels = "visible", nested = TRUE) \%>\%
  summarize_occurrences_by_grade(
    var = "AETOXGR",
    grade_groups = grade_groups
  ) \%>\%
  build_table(df, col_counts = 10L)

}
