% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_coxreg.R
\name{cox_regression}
\alias{cox_regression}
\alias{s_coxreg}
\alias{summarize_coxreg}
\title{Cox Proportional Hazards Regression}
\usage{
s_coxreg(df, .stats, which_vars = "all", var_nms = NULL)

summarize_coxreg(
  lyt,
  variables,
  control = control_coxreg(),
  at = list(),
  multivar = FALSE,
  common_var = "STUDYID",
  .stats = NULL,
  .formats = NULL,
  var_labels = NULL,
  split_fun = NULL,
  .indent_mods = NULL,
  .na_str = NA_character_,
  .section_div = NA_character_
)
}
\arguments{
\item{df}{(\code{data.frame})\cr data set containing all analysis variables.}

\item{.stats}{(\code{character})\cr the name of statistics to be reported among:
\itemize{
\item \code{n}: number of observations
\item \code{hr}: hazard ratio
\item \code{ci}: confidence interval
\item \code{pval}: p-value of the treatment effect
\item \code{pval_inter}: p-value of the interaction effect between the treatment and the covariate
}}

\item{which_vars}{(\code{character})\cr which rows should statistics be returned for from the given model.
Defaults to "all". Other options include "var_main" for main effects, "inter" for interaction effects,
and "multi_lvl" for multivariate model covariate level rows. When \code{which_vars} is "all" specific
variables can be selected by specifying \code{var_nms}.}

\item{var_nms}{(\code{character})\cr the \code{term} value of rows in \code{df} for which \code{.stats} should by returned. Typically
this is the name of a variable. If using variable labels, \code{var} should be a vector of both the desired
variable name and the variable label in that order to see all \code{.stats} related to that variable. When \code{which_vars}
is "var_main" \code{var_nms} should be only the variable name.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{variables}{(named \code{list} of \code{string})\cr list of additional analysis variables.}

\item{multivar}{(\code{flag})\cr if \code{TRUE}, the multi-variable Cox regression will run
and no interaction will be considered between the studied treatment and c
candidate covariate. Default is \code{FALSE} for univariate Cox regression including
an arm variable. When no arm variable is included in the univariate Cox regression,
then also \code{TRUE} should be used to tabulate the covariate effect estimates instead
of the treatment arm effect estimate across models.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{var_labels}{character for label.}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Fits a Cox regression model and estimate hazard ratio to describe the effect
size in a survival analysis.
}
\details{
Cox models are the most commonly used methods to estimate the magnitude of
the effect in survival analysis. It assumes proportional hazards: the ratio
of the hazards between groups (e.g., two arms) is constant over time.
This ratio is referred to as the "hazard ratio" (HR) and is one of the
most commonly reported metrics to describe the effect size in survival
analysis (NEST Team, 2020).
}
\section{Functions}{
\itemize{
\item \code{s_coxreg()}: transforms the tabulated results from \code{\link[=fit_coxreg_univar]{fit_coxreg_univar()}}
and \code{\link[=fit_coxreg_multivar]{fit_coxreg_multivar()}} into a list. Not much calculation is done here,
it rather prepares the data to be used by the layout creating function.

\item \code{summarize_coxreg()}: layout creating function.

}}
\note{
The usual formatting arguments for the \emph{layout creating} function
\code{summarize_coxreg} are not yet accepted (\code{.stats}, \code{.indent_mod}, \code{.formats},
\code{.labels}).
}
\examples{
library(survival)

# Testing dataset [survival::bladder].
set.seed(1, kind = "Mersenne-Twister")
dta_bladder <- with(
  data = bladder[bladder$enum < 5, ],
  data.frame(
    TIME = stop,
    STATUS = event,
    ARM = as.factor(rx),
    COVAR1 = as.factor(enum),
    COVAR2 = factor(
      sample(as.factor(enum)),
      levels = 1:4, labels = c("F", "F", "M", "M")
    )
  )
)
dta_bladder$AGE <- sample(20:60, size = nrow(dta_bladder), replace = TRUE)
dta_bladder$STUDYID <- factor("X")

plot(
  survfit(Surv(TIME, STATUS) ~ ARM + COVAR1, data = dta_bladder),
  lty = 2:4,
  xlab = "Months",
  col = c("blue1", "blue2", "blue3", "blue4", "red1", "red2", "red3", "red4")
)

# s_coxreg

# Univariate
u1_variables <- list(time = "TIME", event = "STATUS", arm = "ARM", covariates = c("COVAR1", "COVAR2"))
univar_model <- fit_coxreg_univar(variables = u1_variables, data = dta_bladder)
df1 <- broom::tidy(univar_model)
s_coxreg(df = df1, .stats = "hr")

# Univariate without treatment arm - only "COVAR2" covariate effects
u2_variables <- list(time = "TIME", event = "STATUS", covariates = c("COVAR1", "COVAR2"))
univar_covs_model <- fit_coxreg_univar(variables = u2_variables, data = dta_bladder)
df1_covs <- broom::tidy(univar_covs_model)
s_coxreg(df = df1_covs, .stats = "hr", var_nms = c("COVAR2", "Sex (F/M)"))

# Multivariate.
m1_variables <- list(time = "TIME", event = "STATUS", arm = "ARM", covariates = c("COVAR1", "COVAR2"))
multivar_model <- fit_coxreg_multivar(variables = m1_variables, data = dta_bladder)
df2 <- broom::tidy(multivar_model)
s_coxreg(df = df2, .stats = "hr")

# Multivariate without treatment arm - only "COVAR1" main effect
m2_variables <- list(time = "TIME", event = "STATUS", covariates = c("COVAR1", "COVAR2"))
multivar_covs_model <- fit_coxreg_multivar(variables = m2_variables, data = dta_bladder)
df2_covs <- broom::tidy(multivar_covs_model)
s_coxreg(df = df2_covs, .stats = "pval", which_vars = "var_main", var_nms = "COVAR1")

# summarize_coxreg

result_univar <- basic_table() \%>\%
  summarize_coxreg(
    variables = u1_variables,
    var_labels = c("A Covariate Label", "Sex (F/M)")
  ) \%>\%
  build_table(dta_bladder)
result_univar

result_multivar <- basic_table() \%>\%
  summarize_coxreg(
    variables = m1_variables,
    multivar = TRUE,
    var_labels = c("A Covariate Label", "Sex (F/M)")
  ) \%>\%
  build_table(dta_bladder)
result_multivar

result_univar_covs <- basic_table() \%>\%
  summarize_coxreg(
    variables = u2_variables,
    var_labels = c("A Covariate Label", "Sex (F/M)")
  ) \%>\%
  build_table(dta_bladder)
result_univar_covs

result_multivar_covs <- basic_table() \%>\%
  summarize_coxreg(
    variables = m2_variables,
    multivar = TRUE,
    var_labels = c("A Covariate Label", "Sex (F/M)")
  ) \%>\%
  build_table(dta_bladder)
result_multivar_covs

}
\seealso{
\link{fit_coxreg} for relevant fitting functions, \link{h_cox_regression} for relevant
helper functions, and \link{tidy_coxreg} for custom tidy methods.
}
