% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{cox_regression}
\alias{cox_regression}
\alias{h_coxreg}
\alias{s_coxreg}
\alias{fit_coxreg}
\title{Cox Proportional Hazards Regression}
\usage{
h_coxreg(variables, data, control = control_coxreg())

s_coxreg(df, .var)

fit_coxreg(lyt, conf_level)
}
\arguments{
\item{variables}{(named \code{list} of \code{strings})\cr list of additional analysis variables.}

\item{data}{(\verb{data frame})\cr the dataset containing the variables to fit the
models.}

\item{control}{(\code{list})\cr a list of parameters as returned by the
helper function \code{\link[=control_coxreg]{control_coxreg()}}.}

\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}
}
\description{
Fits a Cox regression model and estimate hazard ratio to describe the effect
size in a survival analysis.
}
\details{
Cox models are the most commonly used methods to estimate the magnitude of
the effect in survival analysis. It assumes proportional hazards: the ratio
of the hazards of between groups (e.g., two arms) is constant over time.
This ratio is referred to as the "hazard ratio" (HR) and is one of the
most commonly reported metrics to describe the effect size in survival
analysis \insertCite{NESTTeam2020}{tern}.
}
\section{Functions}{
\itemize{
\item \code{h_coxreg}: tabulates the result of a series of univariate Cox
regression models. \code{variables} corresponds to the names of variables found
in \code{data} includes, passed as a named list and corresponding to
\code{time}, \code{event}, \code{arm}, \code{covariates} and \code{strata} terms.

\item \code{s_coxreg}: transforms the tabulated results from \code{h_coxreg}
into a list. Not much calculation is done here, it rather prepares the data to
be used by the layout creating function.

\item \code{fit_coxreg}: layout creating function.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{

set.seed(1, kind = "Mersenne-Twister")
bladder <- survival::bladder
dta_bladder <- with(
  data = bladder[bladder$enum < 5, ],
  data.frame(
    time = stop,
    status = event,
    armcd = as.factor(rx),
    covar1 = as.factor(enum),
    covar2 = factor(
      sample(as.factor(enum)), levels = 1:4, labels = c("F", "F", "M", "M")
    )
  )
)
attr(dta_bladder$covar1, "label") <- "A Covar"
attr(dta_bladder$covar2, "label") <- "Sex (F/M)"

df <- h_coxreg(
  variables = list(
    time = "time", event = "status", arm = "armcd",
    covariates = c("covar1", "covar2")
  ),
  data = dta_bladder
)
s_coxreg(df = df, .var = "hr")


df <- h_coxreg(
  variables = list(
    time = "time", event = "status", arm = "armcd",
    covariates = c("covar1", "covar2")
  ),
  data = dta_bladder
)
s_coxreg(df = df, .var = "hr")


df <- h_coxreg(
  variables = list(
    time = "time", event = "status", arm = "armcd",
    covariates = c("covar1", "covar2")
  ),
  data = dta_bladder
)
result <- split_rows_by(lyt = NULL, "mod", child_labels = "visible") \%>\%
  fit_coxreg(conf_level = .95) \%>\%
  build_table(df = df)
result

}
\references{
\itemize{
\item \insertAllCited{}
\item \insertRef{Hughes2020a}{tern}
}
}
