% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot_data.R
\name{coxphAnnoData}
\alias{coxphAnnoData}
\title{Prepare Cox PH model annotation data with rtable format}
\usage{
coxphAnnoData(fit_coxph, info_coxph = "Cox Porportional Hazard Model")
}
\arguments{
\item{fit_coxph}{a class "coxph" object.}

\item{info_coxph}{label information for Cox PH model.}
}
\description{
An rtable format of Cox PH model data for further annotation on top of Kaplan-Meier grob
}
\examples{

OS <- data.frame(AVAL = abs(rnorm(200)), 
                 CNSR = sample(c(0, 1), 200, TRUE), 
                 ARM = sample(LETTERS[1:3], 200, TRUE),
                 SEX = sample(c("M","F"), 200, TRUE),
                 RACE = sample(c("AA", "BB", "CC"), 200, TRUE),
                 ECOG = sample(c(0, 1), 200, TRUE))
fit_coxph <- coxph(Surv(AVAL, 1-CNSR) ~ ARM + strata(RACE), data = OS, ties = "exact")
anno <- coxphAnnoData(fit_coxph)
anno
### add more annotation, e.g. Add log-rank test from a survdiff object
surv_diff <- survdiff(Surv(AVAL, 1-CNSR) ~ ARM + strata(RACE), data = OS)
logr_p <- pchisq(surv_diff$chisq, length(surv_diff$n) - 1, lower.tail = FALSE) 
anno_new <- anno \%>\% paste0(paste0("Log-rank test p-value: ", as.character(round(logr_p, 4))), "\\n", .)

}
