% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_variables.R
\name{create_afun_compare}
\alias{create_afun_compare}
\title{Constructor Function for \code{\link[=compare_vars]{compare_vars()}}}
\usage{
create_afun_compare(
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\description{
Constructor function which creates a combined Formatted Analysis function for use in
layout creating functions \code{\link[=compare_vars]{compare_vars()}}.
}
\note{
Since \code{\link[=a_compare]{a_compare()}} is generic and we want customization of the formatting arguments
via \code{\link[rtables:make_afun]{rtables::make_afun()}}, we need to create another temporary generic function, with
corresponding customized methods. Then in order for the methods to be found,
we need to wrap them in a combined \code{afun}. Since this is required by two layout creating
functions (and possibly others in the future), we provide a constructor that does this:
\code{\link[=create_afun_compare]{create_afun_compare()}}.
}
\examples{
# `create_afun_compare()` to create combined `afun`

afun <- create_afun_compare(
  .stats = c("n", "count_fraction", "mean_sd", "pval"),
  .indent_mods = c(pval = 1L)
)

lyt <- basic_table() \%>\%
  split_cols_by("ARMCD", ref_group = "ARM A") \%>\%
  analyze(
    "AGE",
    afun = afun,
    show_labels = "visible"
  )
build_table(lyt, df = ex_adsl)

lyt <- basic_table() \%>\%
  split_cols_by("ARMCD", ref_group = "ARM A") \%>\%
  analyze(
    "SEX",
    afun = afun,
    show_labels = "visible"
  )
build_table(lyt, df = ex_adsl)

}
\keyword{internal}
