% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abnormal_by_baseline.R
\name{d_count_abnormal_by_baseline}
\alias{d_count_abnormal_by_baseline}
\title{Description Function that produces the labels for \code{\link[=s_count_abnormal_by_baseline]{s_count_abnormal_by_baseline()}}.}
\usage{
d_count_abnormal_by_baseline(abnormal)
}
\description{
Description Function that produces the labels for \code{\link[=s_count_abnormal_by_baseline]{s_count_abnormal_by_baseline()}}.
}
\examples{
d_count_abnormal_by_baseline("LOW")

}
