% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_cumulative.R
\name{d_count_missed_doses}
\alias{d_count_missed_doses}
\title{Description of Cumulative Count for Missed Doses}
\usage{
d_count_missed_doses(thresholds)
}
\arguments{
\item{thresholds}{(\code{numeric})\cr vector of cutoff value for the counts.}
}
\value{
A named \code{character} vector.
}
\description{
This is a helper function that describes analysis in \code{count_missed_doses}
}
