% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_missed_doses.R
\name{d_count_missed_doses}
\alias{d_count_missed_doses}
\title{Description Function that Calculates Labels for \code{\link[=s_count_missed_doses]{s_count_missed_doses()}}.}
\usage{
d_count_missed_doses(thresholds)
}
\arguments{
\item{thresholds}{(vector of \code{count})\cr number of missed doses the patients at least had.}
}
\value{
\code{\link[=d_count_missed_doses]{d_count_missed_doses()}} returns a named \code{character} vector with the labels.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
