% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_multinomial_rsp.R
\name{d_onco_rsp_label}
\alias{d_onco_rsp_label}
\title{Standard Oncology Response}
\usage{
d_onco_rsp_label(x, na_level = "<Missing>")
}
\arguments{
\item{x}{(\code{character})\cr the standard oncology code to be described.}

\item{na_level}{(\code{string})\cr the string used to replace \code{NA} or empty values during data pre-processing using
\code{df_explicit_na}.}
}
\description{
Describe the oncology response in a standard way.
}
\examples{

d_onco_rsp_label(
  c("CR", "PR", "SD", "NON CR/PD", "PD", "NE", "<Missing>", "NE/Missing")
)
}
