% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportion.R
\name{d_proportion}
\alias{d_proportion}
\title{Description of the Proportion Summary}
\usage{
d_proportion(conf_level, method, long = FALSE)
}
\arguments{
\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{method}{(\code{string}) \cr
the method used to construct the confidence interval for proportion of
successful outcomes; one of \code{waldcc}, \code{wald}, \code{clopper-pearson}, \code{wilson},
\code{agresti-coull} or \code{jeffreys}.}

\item{long}{(\code{flag})\cr Whether a long or a short (default) description is required.}
}
\value{
String describing the analysis.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
This is a helper function that describes the analysis in \code{\link[=s_proportion]{s_proportion()}}.
}
