% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{d_proportion}
\alias{d_proportion}
\title{Describe the proportion summary}
\usage{
d_proportion(conf_level, prop_ci_method, long = FALSE)
}
\arguments{
\item{conf_level}{confidence level of the interval.}

\item{prop_ci_method}{one of (\code{"waldcc"}, \code{"wald"}, \code{"clopper-pearson"},
\code{"wilson"}, \code{agresti-coull} or \code{jeffreys}).\cr
Specifies the method used to construct the confidence interval for proportion
of successful outcomes.}

\item{long}{Whether a long or a short (default) description is required.}
}
\value{
String describing the analysis.
}
\description{
This is an auxiliary function that describes the analysis in `s_proportion`.
}
