% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportion.R
\name{d_proportion}
\alias{d_proportion}
\title{Description of the proportion summary}
\usage{
d_proportion(conf_level, method, long = FALSE, method_only = FALSE)
}
\arguments{
\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{method}{(\code{string})\cr the method used to construct the confidence interval
for proportion of successful outcomes; one of \code{waldcc}, \code{wald}, \code{clopper-pearson},
\code{wilson}, \code{wilsonc}, \code{strat_wilson}, \code{strat_wilsonc}, \code{agresti-coull} or \code{jeffreys}.}

\item{long}{(\code{flag})\cr whether a long or a short (default) description is
required. Ignored if \code{method_only = TRUE}.}

\item{method_only}{(\code{flag})\cr whether to return only the method description,
without the confidence interval part of the description.}
}
\value{
String describing the analysis.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is a helper function that describes the analysis in \code{\link[=s_proportion]{s_proportion()}}.
}
\examples{
d_proportion(0.95, "wald")
d_proportion(0.95, "wald", long = TRUE)
d_proportion(0.95, "wald", method_only = TRUE)

}
