% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{d_proportion_diff}
\alias{d_proportion_diff}
\title{Describe the difference between two proportions summary}
\usage{
d_proportion_diff(conf_level, diff_ci_method, long = FALSE)
}
\arguments{
\item{conf_level}{confidence level of the interval.}

\item{diff_ci_method}{one of (\code{"wald"}, \code{"waldcc"}, \code{"cmh"}, \code{"anderson-hauck"} or \code{"ha"},
\code{"newcombe"})\cr
Specifies the method used to construct the confidence interval for the difference
in proportion of outcomes.}

\item{long}{Whether a long or a short (default) description is required.}
}
\value{
String describing the analysis.
}
\description{
This is an auxiliary function that describes the analysis in \code{s_proportion_diff}.
}
