% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_diff.R
\name{d_proportion_diff}
\alias{d_proportion_diff}
\title{Description of Method Used for Proportion Comparison}
\usage{
d_proportion_diff(conf_level, method, long = FALSE)
}
\arguments{
\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{method}{(\code{string})\cr
the method used for the confidence interval estimation.}

\item{long}{(\code{logical})\cr
Whether a long or a short (default) description is required.}
}
\value{
String describing the analysis.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is an auxiliary function that describes the analysis in
\code{s_proportion_diff}.
}
