% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_subgroups.R
\name{d_rsp_subgroups_colvars}
\alias{d_rsp_subgroups_colvars}
\title{Labels for Column Variables in Binary Response by Subgroup Table}
\usage{
d_rsp_subgroups_colvars(vars, conf_level = NULL, method = NULL)
}
\arguments{
\item{vars}{(\code{character})\cr the name of statistics to be reported among
\code{n} (total number of observations per group), \code{n_rsp} (number of responders per group),
\code{prop} (proportion of responders), \code{n_tot} (total number of observations),
\code{or} (odds ratio), \code{ci} (confidence interval of odds ratio) and \code{pvalue} (p value of the effect).}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{method}{(\code{string})\cr
specifies the test used to calculate the p-value for the difference between
two proportions. For options, see \code{\link[=s_test_proportion_diff]{s_test_proportion_diff()}}. Default is \code{NULL}
so no test is performed.}
}
\value{
\code{list} of variables to tabulate and their labels.
}
\description{
Internal function to check variables included in
\link{tabulate_rsp_subgroups} and create column labels.
}
