% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_duration_subgroups.R
\name{d_survival_subgroups_colvars}
\alias{d_survival_subgroups_colvars}
\title{Labels for Column Variables in Survival Duration by Subgroup Table}
\usage{
d_survival_subgroups_colvars(vars, conf_level, method, time_unit = NULL)
}
\arguments{
\item{vars}{(\code{character})\cr the name of statistics to be reported among
\code{n_tot_events} (total number of events per group),
\code{n_events} (number of events per group),
\code{n_tot} (total number of observations per group),
\code{n} (number of observations per group),
\code{median} (median survival time),
\code{hr} (hazard ratio),
\code{ci} (confidence interval of hazard ratio) and
\code{pval} (p value of the effect).
Note, one of the statistics \code{n_tot} and \code{n_tot_events}, as well as both \code{hr} and \code{ci}
are required.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{method}{p-value method for testing hazard ratio = 1.}

\item{time_unit}{(\code{string})\cr label with unit of median survival time. Default \code{NULL} skips
displaying unit.}
}
\value{
\code{list} of variables to tabulate and their labels.
}
\description{
Internal function to check variables included in
\code{\link[=tabulate_survival_subgroups]{tabulate_survival_subgroups()}} and create column labels.
Note that at least one of \code{n_tot} and \code{n_tot_events} needs to be provided in \code{vars}.
}
