% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{d_test_proportion_diff}
\alias{d_test_proportion_diff}
\title{Describe the test between two proportions}
\usage{
d_test_proportion_diff(test)
}
\arguments{
\item{test}{one of (\code{"chisq", "cmh", "fisher", "schouten"})\cr
Specifies the test used to calculate the p-value.}
}
\value{
String describing the test from which the p-value is derived.
}
\description{
This is an auxiliary function that describes the analysis in \code{s_test_proportion_diff}.
}
