% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_proportion_diff.R
\name{d_test_proportion_diff}
\alias{d_test_proportion_diff}
\title{Description of the Difference Test Between Two Proportions}
\usage{
d_test_proportion_diff(method)
}
\arguments{
\item{method}{(\code{string})\cr
one of (\code{chisq}, \code{cmh}, \code{fisher}, \code{schouten}; specifies the test used
to calculate the p-value.}
}
\value{
\code{string} describing the test from which the p-value is derived.
}
\description{
This is an auxiliary function that describes the analysis in
\code{s_test_proportion_diff}.
}
