% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{day2month}
\alias{day2month}
\title{Conversion of Days to Months}
\usage{
day2month(x)
}
\arguments{
\item{x}{(\code{numeric})\cr time in days.}
}
\value{
A \code{numeric} vector with the time in months.
}
\description{
Conversion of Days to Months
}
\examples{
x <- c(403, 248, 30, 86)
day2month(x)
}
