% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_grob.R
\name{decorate_grob_factory}
\alias{decorate_grob_factory}
\title{Update Page Number}
\usage{
decorate_grob_factory(npages, ...)
}
\arguments{
\item{npages}{number of pages in total}

\item{...}{passed on to \code{\link[=decorate_grob]{decorate_grob()}}}
}
\value{
Closure that increments the page number.
}
\description{
Automatically updates page number.
}
\keyword{internal}
