% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_grob.R
\name{decorate_grob_factory}
\alias{decorate_grob_factory}
\title{Page number}
\usage{
decorate_grob_factory(npages, ...)
}
\arguments{
\item{npages}{number of pages in total}

\item{...}{passed on to \code{\link{decorate_grob}}}
}
\value{
closure that increments the page number
}
\description{
Automatically updates page number
}
\examples{
pf <- decorate_grob_factory(
  titles = "This is a test\nHello World",
  footnotes = "Here belong the footnotess",
  npages = 3
)

draw_grob(pf(NULL))
draw_grob(pf(NULL))
draw_grob(pf(NULL))
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
