% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_grob.R
\name{decorate_grob_factory}
\alias{decorate_grob_factory}
\title{Automatically updates page number}
\usage{
decorate_grob_factory(...)
}
\arguments{
\item{...}{passed on to decorate_grob}
}
\value{
closure that increments the page number
}
\description{
Automatically updates page number
}
\examples{
pf <- decorate_grob_factory(
  titles = "This is a test\\nHello World",
  footnotes = "Here belong the footnotess",
  npages = 3
)

grid.newpage(); grid.draw(pf(NULL))
grid.newpage(); grid.draw(pf(NULL))
grid.newpage(); grid.draw(pf(NULL))
\dontrun{
grid.newpage(); grid.draw(pf(NULL))
}
}
