% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_grob.R
\name{decorate_grob_set}
\alias{decorate_grob_set}
\title{Decorate Set of \code{grobs} and Add Page Numbering}
\usage{
decorate_grob_set(grobs, ...)
}
\arguments{
\item{grobs}{a list of grid grobs}

\item{...}{arguments passed on to \code{\link{decorate_grob}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Note that this uses the \code{decorate_grob_factory} function.
}
\examples{
library(ggplot2)
g <- with(iris, {
  list(
    ggplotGrob(qplot(Sepal.Length, Sepal.Width, col = Species)),
    ggplotGrob(qplot(Sepal.Length, Petal.Length, col = Species)),
    ggplotGrob(qplot(Sepal.Length, Petal.Width, col = Species)),
    ggplotGrob(qplot(Sepal.Width, Petal.Length, col = Species)),
    ggplotGrob(qplot(Sepal.Width, Petal.Width, col = Species)),
    ggplotGrob(qplot(Petal.Length, Petal.Width, col = Species))
  )
})
lg <- decorate_grob_set(grobs = g, titles = "Hello\nOne\nTwo\nThree", footnotes = "")
library(grid)
tern:::draw_grob(lg[[1]])
tern:::draw_grob(lg[[2]])
tern:::draw_grob(lg[[6]])
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
