% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_defaults_handling.R
\name{default_stats_and_formats}
\alias{default_stats_and_formats}
\alias{get_stats}
\alias{get_format_from_stats}
\alias{get_label_from_stats}
\alias{tern_default_formats}
\alias{tern_default_labels}
\title{Defaults for stats methods names and their relative formats/labels}
\usage{
get_stats(method_group, type = NULL, stats_in = NULL, add_pval = FALSE)

get_format_from_stats(stats, type = NULL, formats_in = NULL)

get_label_from_stats(stats, type = NULL, labels_in = NULL)

tern_default_formats(type = NULL)

tern_default_labels(type)
}
\arguments{
\item{method_group}{(\code{string}) \cr indicates the group of statistical methods that
we need the defaults from.}

\item{type}{(\code{string}) \cr type of data and result desired. Some stats defaults (and their formats and labels),
differ if you need to analyze different type of values. See Details for available options.}

\item{stats_in}{(\code{character}) \cr desired stats to be picked out from the selected method group.}

\item{stats}{(\code{character}) \cr statistical methods to get defaults formats for.}

\item{formats_in}{(named \code{vector}) \cr inserted formats to replace defaults.}
}
\value{
\itemize{
\item \code{get_stats()} returns a character vector with all default statistical methods.
}

\itemize{
\item \code{get_format_from_stats()} returns a named list of formats, they being a value from
\code{\link[formatters:list_formats]{formatters::list_valid_format_labels()}} or a custom function (e.g. \link{formatting_functions}).
}

\itemize{
\item \code{get_label_from_stats()} returns a named list of default labels (if present
otherwise \code{NULL}).
}

\itemize{
\item \code{tern_default_formats()} returns a complete named list of default formats for \code{tern}.
}

\itemize{
\item \code{tern_default_labels()} returns a complete named list of default labels for \code{tern}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Utility functions to get valid statistic methods for different method groups
(\code{.stats}) and their relative formats (\code{.formats}) and labels (\code{.labels}). This utility
is used across \code{tern}, but some of its working principles can be seen in \code{\link[=analyze_vars]{analyze_vars()}}.
}
\details{
Current choices for \code{type} are \code{counts} and \code{numeric} for \code{\link[=analyze_vars]{analyze_vars()}} and affect \code{get_stats()} and
\code{get_label_from_stats()}. Another option, \code{count_fraction_fixed_dp} affects only formats for count fractions.
}
\section{Functions}{
\itemize{
\item \code{get_stats()}: Get defaults statistical methods for different
groups of methods.

\item \code{get_format_from_stats()}: Get formats from vector of statistical methods. If not
present \code{NULL} is returned.

\item \code{get_label_from_stats()}: Get labels from vector of statistical methods.

\item \code{tern_default_formats()}: Function that produce the complete named list
of default formats for \code{tern}.

\item \code{tern_default_labels()}: Function that produce the complete named list
of default labels for \code{tern}.

}}
\note{
Formats in \code{tern} and \code{rtables} can be functions that take in the table cell value and
return a string. This is well documented in \code{vignette("custom_appearance", package = "rtables")}.
}
\seealso{
\link{formatting_functions}
}
