% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_table.R
\name{demographic_table}
\alias{demographic_table}
\title{Create Demographics Table from ADSL data}
\usage{
demographic_table(data, arm_var = "ARM", all.patients = TRUE,
  group_by_vars = c("SEX", "AGE", "RACE", "ETHNIC"))
}
\arguments{
\item{ADSL}{ADSL dataset with the following variables: USUBJID, STUDYID, AGE,
and SEX}
}
\description{
Similiar as in STREAM
}
\details{
Give a detailed description of what this function does.
}
\examples{
n <- 100
ASL <- data.frame(
   USUBJID = paste0("id-", 1:n),
   STUDYID = "study 1",
   AGE = 40 + rnorm(n, 0, 20),
   SEX = sample(c("M", "F", "UNDEFINED", NA), n, replace = TRUE),
   stringAsFactors = FALSE
)

demographic_table(ASL)

\dontrun{
library(atezo.data)
library(dplyr)

ASL <- asl(com.roche.cdt30019.go29436.re)
tbl_expected <- get_demographic_table(com.roche.cdt30019.go29436.re)
Viewer(tbl_expected)

# order factor levels if they exist
of <- function(x, ...) {
  lvls <- if (is.factor(x)) levels(x) else unique(x)
  olvls <- intersect(unlist(list(...)), lvls)
  factor(as.character(x), levels = c(olvls, setdiff(lvls, olvls)))
}
add_lbl <- function(lbl, x) {
  structure(x, label = lbl)
}
lbl <- function(x) {
  attr(x, "label")
}

ADSL <- ASL \%>\%
 filter(ITTWTFL == 'Y') \%>\%
 mutate(ARM = of(ARM, "DUMMY C", "DUMMY B", "DUMMY A")) \%>\%
 mutate(SEX = add_lbl(lbl(SEX), recode_factor(SEX, M = "MALE", F = "FEMALE")))



tbl <- demographic_table(
   data = ADSL,
   arm_var = "ARM",
   all.patients = TRUE,
   group_by = toupper(c("sex", "mliver", "tciclvl2", "bage", "age4cat",
     "ethnic", "race", "bwt", "tobhx", "hist", "EGFRMUT",
     "alkmut", "krasmut", "becog"))
)
Viewer(tbl)

compare_rtables(tbl, tbl_expected)

# if all is good then
tbl <- tbl_expected
compare_rtables(tbl, tbl_expected)

}

`\%needs\%` <- teal.oncology:::`\%needs\%`


}
\author{
Xiao Yu Mo (mox5), \email{xiao_yu.mo@roche.com}
}
