% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_demographic.R
\name{demographic_table}
\alias{demographic_table}
\title{Create Demographics Table from ASL data}
\usage{
demographic_table(ADSL)
}
\arguments{
\item{ADSL}{ASL dataset with the following variables: USUBJID, STUDYID, AGE,
and SEX}
}
\description{
Similiar as in STREAM
}
\details{
Give a detailed description of what this function does.
}
\examples{
n <- 100
ASL <- data.frame(
   USUBJID = paste0("id-", 1:n),
   STUDYID = "study 1",
   AGE = 40 + rnorm(n, 0, 20),
   SEX = sample(c("M", "F", "UNDEFINED", NA), n, replace = TRUE),
   stringAsFactor = FALSE
)

demographic_table(ASL)

\dontrun{
library(atezo.data)
ADSL <- asl(com.roche.atezo.cdpt7722.wo29637)

demographic_table(ADSL)
}


}
\author{
Xiao Yu Mo (mox5), \email{xiao_yu.mo@roche.com}
}
