% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desctool_binom_diff.R
\name{desctool_binom}
\alias{desctool_binom}
\title{Confidence Intervals for a Difference of Binomials}
\arguments{
\item{grp}{(\code{factor})\cr
vector assigning observations to one out of two groups
(e.g. reference and treatment group).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Several confidence intervals for the difference between proportions.
}
