% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desctools_binom_diff.R
\name{desctools_binom}
\alias{desctools_binom}
\alias{h_recycle}
\alias{desctools_binomci}
\title{Confidence Intervals for a Difference of Binomials}
\usage{
h_recycle(...)

desctools_binom(
  x1,
  n1,
  x2,
  n2,
  conf.level = 0.95,
  sides = c("two.sided", "left", "right"),
  method = c("ac", "wald", "waldcc", "score", "scorecc", "mn", "mee", "blj", "ha", "hal",
    "jp")
)

desctools_binomci(
  x,
  n,
  conf.level = 0.95,
  sides = c("two.sided", "left", "right"),
  method = c("wilson", "wald", "waldcc", "agresti-coull", "jeffreys", "modified wilson",
    "wilsoncc", "modified jeffreys", "clopper-pearson", "arcsine", "logit", "witting",
    "pratt", "midp", "lik", "blaker"),
  rand = 123,
  tol = 1e-05
)
}
\arguments{
\item{conf.level}{\cr confidence level, defaults to 0.95}

\item{sides}{\cr a character string specifying the side of the confidence interval. Must be one of "two-sided" (default), "left" or "right".}

\item{method}{\cr character string specifying which method to use. Can be one out of: "wald", "wilson", "wilsoncc", "agresti-coull", "jeffreys", "modified wilson", "modified jeffreys", "clopper-pearson", "arcsine.
", "logit", "witting", "pratt", "midp", "lik" and "blaker"

@return A matric with 3 columns containing:
\itemize{
\item \code{est}: estimate of proportion difference.
\item \code{lwrci}: lower end of the confidence interval
\item \code{upci}:  upper end of the confidence interval.
}}

\item{x}{\cr number of successes}

\item{n}{\cr number of trials}

\item{grp}{(\code{factor})\cr
vector assigning observations to one out of two groups
(e.g. reference and treatment group).}
}
\value{
A named list of 3 values:
\itemize{
\item \code{est}: estimate of proportion difference.
\item \code{lwrci}: estimate of lower end of the confidence interval
\item \code{upci}: estimate of upper end of the confidence interval.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Several confidence intervals for the difference between proportions.
}
\section{Functions}{
\itemize{
\item \code{h_recycle()}: This function recycles all supplied elements to the maximal dimension.

\item \code{desctools_binom()}: Several Confidence Intervals for the difference between proportions.

\item \code{desctools_binomci()}: Compute confidence intervals for binomial proportions.

}}
\examples{
# Internal function -
\dontrun{

set.seed(2)
rsp <- sample(c(TRUE, FALSE), replace = TRUE, size = 20)
grp <- factor(c(rep("A", 10), rep("B", 10)))
tbl <- table(grp, factor(rsp, levels = c(TRUE, FALSE)))
desctools_binom(
  tbl[1], sum(tbl[1], tbl[3]), tbl[2], sum(tbl[2], tbl[4]),
  conf.level = 0.90, method = "waldcc"
)
}

}
\keyword{internal}
