% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_explicit_na.R
\name{df_explicit_na}
\alias{df_explicit_na}
\title{Prepare data to be used in a teal app}
\usage{
df_explicit_na(
  data,
  omit_columns = NULL,
  char_as_factor = TRUE,
  na_level = "<Missing>"
)
}
\arguments{
\item{data}{(\code{data.frame}) Any \code{cdisc} data set}

\item{omit_columns}{(\code{character}) vector of columns that should not be touched
by this function}

\item{char_as_factor}{(\code{logical}) Whether to make all \code{character} variables
inside the \code{data.frame} that are not omit_columnsd \code{factor} variables.}

\item{na_level}{(\code{character}) that is used to replace all \code{NA} levels
inside non-omit_columnsd columns}
}
\value{
The \code{data.frame} inserted with the desired changes made.
}
\description{
teal does not take care of missing data provided by the app author. This is a helper
function to encode \code{NA} values with a specific character char.
}
\details{
Additionally teal requires character variables to be converted to factors. This can
be executed by setting the \code{char_as_factor} variable to \code{TRUE}.
}
\examples{

my_data <- data.frame(
  x = c("A", "B", NA, "C"),
  y = c("D", "E", "F", "E"),
  z = c(1, 2, 3, 4)
)

df_explicit_na(my_data)

library(random.cdisc.data)
ADSL <- radsl(N = 100, na_percentage = 0.1)
df_explicit_na(ADSL, omit_columns = c("USUBJID", "STUDYID"))


}
