% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_explicit_na.R
\name{df_explicit_na}
\alias{df_explicit_na}
\title{Prepare Data to be Used in a Teal App}
\usage{
df_explicit_na(
  data,
  omit_columns = NULL,
  char_as_factor = TRUE,
  na_level = "<Missing>"
)
}
\arguments{
\item{data}{(\code{data.frame}) Any \code{cdisc} data set}

\item{omit_columns}{(\code{character}) vector of columns that should not be touched
by this function}

\item{char_as_factor}{(\code{logical}) Whether to make all \code{character} variables
inside the \code{data.frame} that are not \code{omit_columns} \code{factor} variables.}

\item{na_level}{(\code{character}) that is used to replace all \code{NA} levels
inside non - \code{omit_column} columns}
}
\value{
The \code{data.frame} inserted with the desired changes made.
}
\description{
teal does not take care of missing data provided by the app author. This is a helper
function to encode \code{NA} values with a specific character char.
}
\details{
Additionally teal requires character variables to be converted to factors. This can
be executed by setting the \code{char_as_factor} variable to \code{TRUE}.
}
\examples{

my_data <- data.frame(
  v = c("A", NA, NA, NA),
  w = c("A", "B", NA, "C"),
  x = c("D", "E", "F", NA),
  y = c("G", "H", "I", NA),
  z = c(1, 2, 3, 4)
)

df_explicit_na(my_data, omit_columns = c("x", "y"))

}
