% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_equal_float}
\alias{.is_equal_float}
\title{Utility function to check if a float value is equal to another float value}
\usage{
.is_equal_float(x, y)
}
\arguments{
\item{x}{(\code{numeric(1)})\cr a float number.}

\item{y}{(\code{numeric(1)})\cr a float number.}
}
\value{
\code{TRUE} if identical, otherwise \code{FALSE}.
}
\description{
Uses \code{.Machine$double.eps} as the tolerance for the comparison.
}
\keyword{internal}
