% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_grid.R
\name{draw_grob}
\alias{draw_grob}
\title{Draw grob
Draw grob on device page}
\usage{
draw_grob(grob, newpage = TRUE, vp = NULL)
}
\arguments{
\item{grob}{grid object}

\item{newpage}{draw on a new page}

\item{vp}{a \code{\link{viewport}} object (or \code{NULL}).}
}
\description{
Draw grob
Draw grob on device page
}
\examples{
library(dplyr)
rect <- rectGrob(width = unit(0.5, "npc"), height = unit(0.5, "npc"))
rect \%>\% draw_grob(vp = viewport(angle = 45))

num <- lapply(1:10, textGrob)
num \%>\%
  arrange_grobs(grobs = .) \%>\%
  draw_grob()
showViewport()
}
