% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_grid.R
\name{draw_grob}
\alias{draw_grob}
\title{Draw \code{grob}}
\usage{
draw_grob(grob, newpage = TRUE, vp = NULL)
}
\arguments{
\item{grob}{grid object}

\item{newpage}{draw on a new page}

\item{vp}{a \code{\link{viewport}} object (or \code{NULL}).}
}
\description{
Draw grob on device page.
}
\examples{
library(dplyr)
library(grid)

# Internal function - draw_grob
\dontrun{
rect <- rectGrob(width = grid::unit(0.5, "npc"), height = grid::unit(0.5, "npc"))
rect \%>\% tern:::draw_grob(vp = grid::viewport(angle = 45))

num <- lapply(1:10, textGrob)
num \%>\%
  tern:::arrange_grobs(grobs = .) \%>\%
  tern:::draw_grob()
showViewport()
}

}
\keyword{internal}
