% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_example.R
\name{dummy_summation}
\alias{dummy_summation}
\alias{h_dummy_sum}
\alias{s_dummy_sum}
\alias{dummy_sum}
\title{Documentation Example for Helper, Statistics and Analyze Functions}
\usage{
h_dummy_sum(x)

s_dummy_sum(x, .var = "x")

dummy_sum(lyt, vars)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}
}
\value{
The statistics are:
\itemize{
\item \code{sum}: the sum of all numbers, after removing \code{NA} values.
}
}
\description{
This illustrates the conventions on how to document helper, statistics and analyze functions
together.
}
\section{Functions}{
\itemize{
\item \code{h_dummy_sum}: Helper Function which does the actual summation.

\item \code{s_dummy_sum}: Statistics Function which sums the numbers, after removing \code{NA} values.

\item \code{dummy_sum}: Analyze Function which adds the sum analysis to
the input layout. Note that additional formatting arguments can be used
here.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{
h_dummy_sum(1:10)

s_dummy_sum(c(1, 2))

basic_table() \%>\%
  split_cols_by("Species") \%>\%
  dummy_sum("Sepal.Length") \%>\%
  build_table(iris)

basic_table() \%>\%
  split_cols_by("Species") \%>\%
  dummy_sum(c("Sepal.Length", "Petal.Length")) \%>\%
  build_table(iris)

}
