% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_multinomial_rsp.R
\name{estimate_multinomial_rsp}
\alias{estimate_multinomial_rsp}
\alias{s_multinomial_response}
\alias{estimate_multinomial_response}
\title{Estimation of Proportions per Level of Factor}
\usage{
s_multinomial_response(
  x,
  conf_level = 0.95,
  method = c("waldcc", "wald", "clopper-pearson", "wilson", "agresti-coull",
    "jeffreys"),
  long = FALSE
)

estimate_multinomial_response(
  lyt,
  var,
  vars = var,
  show_labels = "hidden",
  ...
)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{method}{(\code{string}) \cr
the method used to construct the confidence interval for proportion of
successful outcomes; one of \code{waldcc}, \code{wald}, \code{clopper-pearson}, \code{wilson},
\code{agresti-coull} or \code{jeffreys}.}

\item{long}{(\code{flag})\cr a long description is required.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{var}{(\code{string})\cr single variable name.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{...}{other arguments are ultimately conveyed to \code{\link[=s_multinomial_response]{s_multinomial_response()}}.}
}
\description{
Estimate the proportion along with confidence interval of a proportion
regarding the level of a factor.
}
\section{Functions}{
\itemize{
\item \code{s_multinomial_response}: statistics function estimating a
proportion per factor level along with its confidence interval for the level.

\item \code{estimate_multinomial_response}: used in a \code{rtables} pipeline.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{

# Basic use of the statistics function
s_multinomial_response(c(0, 1, 0, 1))


# Use of the layout creating function.
dta_test <- data.frame(
  USUBJID = paste0("S", 1:12),
  ARM     = rep(LETTERS[1:3], each = 4),
  AVAL    = c(A = c(1, 1, 1, 1), B = c(0, 0, 1, 1), C = c(0, 0, 0, 0))
)
dta_test$AVALC <- as.factor(c(
  "Complete Response (CR)", "Partial Response (PR)"
)[dta_test$AVAL + 1])

lyt <- split_cols_by(lyt = NULL, var = "ARM") \%>\%
  estimate_multinomial_response(var = "AVALC")

as_html(build_table(lyt = lyt, df = dta_test)) \%>\% Viewer

}
