% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_multinomial_rsp.R
\name{estimate_multinomial_rsp}
\alias{estimate_multinomial_rsp}
\alias{d_onco_rsp_label}
\alias{s_length_proportion}
\alias{a_length_proportion}
\alias{estimate_multinomial_response}
\title{Estimation of Proportions per Level of Factor}
\usage{
d_onco_rsp_label(x)

s_length_proportion(x, .N_col, ...)

a_length_proportion(x, .N_col, ...)

estimate_multinomial_response(
  lyt,
  var,
  ...,
  show_labels = "hidden",
  table_names = var,
  .stats = "prop_ci",
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{x}{(\code{character})\cr the standard oncology code to be described.}

\item{.N_col}{(\code{count})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is passed by \code{rtables}.}

\item{...}{additional arguments for the lower level functions.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{var}{(\code{string})\cr single variable name for the primary analysis variable.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
See \code{\link[=s_proportion]{s_proportion()}} for statistics and additional possible arguments.
}
\description{
Estimate the proportion along with confidence interval of a proportion
regarding the level of a factor.
}
\section{Functions}{
\itemize{
\item \code{d_onco_rsp_label}: Standard Oncology Response

Describe the oncology response in a standard way.

\item \code{s_length_proportion}: Statistics function which takes the length of the input \code{x} and takes that
as the number of successes, and the column number \code{.N_col} as the total number, and feeds that into
\code{\link[=s_proportion]{s_proportion()}}.

\item \code{a_length_proportion}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{estimate_multinomial_response}: Analyze Function which adds the multinomial proportion analysis to
the input layout. Note that additional formatting arguments can be used
here.
}}

\examples{

d_onco_rsp_label(
  c("CR", "PR", "SD", "NON CR/PD", "PD", "NE", "Missing", "<Missing>", "NE/Missing")
)

# Adding some values not considered in d_onco_rsp_label

d_onco_rsp_label(
  c("CR", "PR", "hello", "hi")
)
s_length_proportion(rep("CR", 10), .N_col = 100)
s_length_proportion(factor(character(0)), .N_col = 100)
a_length_proportion(rep("CR", 10), .N_col = 100)
a_length_proportion(factor(character(0)), .N_col = 100)

library(dplyr)
# Use of the layout creating function.
dta_test <- data.frame(
  USUBJID = paste0("S", 1:12),
  ARM     = factor(rep(LETTERS[1:3], each = 4)),
  AVAL    = c(A = c(1, 1, 1, 1), B = c(0, 0, 1, 1), C = c(0, 0, 0, 0))
) \%>\% mutate(
  AVALC = factor(AVAL,
    levels = c(0, 1),
    labels = c("Complete Response (CR)", "Partial Response (PR)")
  )
)


lyt <- basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  estimate_multinomial_response(var = "AVALC")

tbl <- build_table(lyt, dta_test)

html <- as_html(tbl)
html
\dontrun{
Viewer(html)
}
}
