% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportion.R
\name{estimate_proportions}
\alias{estimate_proportions}
\alias{s_proportion}
\alias{a_proportion}
\alias{estimate_proportion}
\title{Estimation of Proportions}
\usage{
s_proportion(
  x,
  conf_level = 0.95,
  method = c("waldcc", "wald", "clopper-pearson", "wilson", "wilsonc", "agresti-coull",
    "jeffreys"),
  long = FALSE
)

a_proportion(
  x,
  conf_level = 0.95,
  method = c("waldcc", "wald", "clopper-pearson", "wilson", "wilsonc", "agresti-coull",
    "jeffreys"),
  long = FALSE
)

estimate_proportion(
  lyt,
  vars,
  ...,
  show_labels = "hidden",
  table_names = vars,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{x}{(\code{logical})\cr whether each subject is a responder or not.
\code{TRUE} represents a successful outcome.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{method}{(\code{string}) \cr
the method used to construct the confidence interval for proportion of
successful outcomes; one of \code{waldcc}, \code{wald}, \code{clopper-pearson}, \code{wilson},
\code{agresti-coull} or \code{jeffreys}.}

\item{long}{(\code{flag})\cr a long description is required.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{...}{other arguments are ultimately conveyed to \code{\link[=s_proportion]{s_proportion()}}.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Estimate the proportion of responders within a studied population.
}
\section{Functions}{
\itemize{
\item \code{s_proportion}: statistics function estimating a
proportion along with its confidence interval.

\item \code{a_proportion}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{estimate_proportion}: used in a \code{rtables} pipeline.
}}

\examples{
s_proportion(c(1, 0, 1, 0))

a_proportion(c(1, 0, 1, 0))

dta_test <- data.frame(
  USUBJID = paste0("S", 1:12),
  ARM     = rep(LETTERS[1:3], each = 4),
  AVAL    = c(A = c(1, 1, 1, 1), B = c(0, 0, 1, 1), C = c(0, 0, 0, 0))
)

basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  estimate_proportion(vars = "AVAL") \%>\%
  build_table(df = dta_test)

}
