% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\alias{ex_adsl}
\title{Simulated CDISC Data for Examples}
\format{
rds (data.frame)

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300 rows and 14 columns.
}
\usage{
ex_adsl
}
\description{
Simulated CDISC Data for Examples
}
\section{Functions}{
\itemize{
\item \code{ex_adsl}: ADSL data

}}
\keyword{datasets}
