% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\alias{ex_adsl}
\alias{ex_adae}
\alias{ex_adlb}
\alias{ex_adrs}
\alias{ex_adtte}
\title{Simulated CDISC Data for Examples}
\format{
rds (data.frame)

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 200 rows and 21 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 541 rows and 44 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4200 rows and 50 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1600 rows and 28 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1000 rows and 28 columns.
}
\usage{
ex_adsl

ex_adae

ex_adlb

ex_adrs

ex_adtte
}
\description{
Simulated CDISC Data for Examples
}
\section{Functions}{
\itemize{
\item \code{ex_adsl}: ADSL data

\item \code{ex_adae}: ADAE data

\item \code{ex_adlb}: ADLB data

\item \code{ex_adrs}: ADRS data

\item \code{ex_adtte}: ADTTE data

}}
\keyword{datasets}
