% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\alias{tern_ex_adsl}
\alias{tern_ex_adae}
\alias{tern_ex_adlb}
\alias{tern_ex_adpp}
\alias{tern_ex_adrs}
\alias{tern_ex_adtte}
\title{Simulated CDISC Data for Examples}
\format{
\code{rds} (\code{data.frame})

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 200 rows and 21 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 541 rows and 42 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4200 rows and 50 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 522 rows and 25 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1600 rows and 29 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1000 rows and 28 columns.
}
\usage{
tern_ex_adsl

tern_ex_adae

tern_ex_adlb

tern_ex_adpp

tern_ex_adrs

tern_ex_adtte
}
\description{
Simulated CDISC Data for Examples
}
\section{Functions}{
\itemize{
\item \code{tern_ex_adsl}: ADSL data

\item \code{tern_ex_adae}: ADAE data

\item \code{tern_ex_adlb}: ADLB data

\item \code{tern_ex_adpp}: ADPP data

\item \code{tern_ex_adrs}: ADRS data

\item \code{tern_ex_adtte}: ADTTE data

}}
\keyword{datasets}
