% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{explicit_na}
\alias{explicit_na}
\title{Substitute missing data with a string or factor level}
\usage{
explicit_na(x, label = "<Missing>")
}
\arguments{
\item{x}{factor or character vector}

\item{label}{character string that missing data should be replaced with}
}
\description{
Substitute missing data with a string or factor level
}
\examples{

explicit_na(c(NA, "a", "b"))
is.na(explicit_na(c(NA, "a", "b")))

explicit_na(factor(c(NA, "a", "b")))
is.na(explicit_na(factor(c(NA, "a", "b"))))

explicit_na(sas_na(c("a", "")))

}
