% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{extreme_format}
\alias{extreme_format}
\alias{h_get_format_threshold}
\alias{h_format_threshold}
\title{Formatting Extreme Values}
\usage{
h_get_format_threshold(digits = 2L)

h_format_threshold(x, digits = 2L)
}
\arguments{
\item{digits}{(\code{integer})\cr number of decimal places to display.}

\item{x}{(\code{number})\cr value to format.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{Rtables} Formatting Functions that handle extreme values.
}
\details{
For each input, apply a format to the specified number of \code{digits}. If the value is
below a threshold, it returns "<0.01" e.g. if the number of \code{digits} is 2. If the value is
above a threshold, it returns ">999.99" e.g. if the number of \code{digits} is 2.
If it is zero, then returns "0.00".
}
\section{Functions}{
\itemize{
\item \code{h_get_format_threshold}: Internal helper function to calculate the threshold and create formatted strings
used in Formatting Functions. Returns a list with elements \code{threshold} and \code{format_string}.

\item \code{h_format_threshold}: Internal helper function to apply a threshold format to a value.
Creates a formatted string to be used in Formatting Functions.
}}

\examples{
h_get_format_threshold(2L)

h_format_threshold(0.001)
h_format_threshold(1000)

}
\seealso{
Other formatting functions: 
\code{\link{format_count_fraction}()},
\code{\link{format_fraction_threshold}()},
\code{\link{format_fraction}()},
\code{\link{formatting_functions}}
}
\concept{formatting functions}
