% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_conf_level}
\alias{f_conf_level}
\title{Utility function to create label for confidence interval}
\usage{
f_conf_level(conf_level)
}
\arguments{
\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}
}
\value{
a \code{string}
}
\description{
Utility function to create label for confidence interval
}
\keyword{internal}
