% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_factor.R
\name{fct_collapse_only}
\alias{fct_collapse_only}
\title{Collapsing of Factor Levels and Keeping Only Those New Group Levels}
\usage{
fct_collapse_only(.f, ...)
}
\arguments{
\item{.f}{(\code{factor} or \code{character})\cr original vector.}

\item{...}{(named \code{character} vectors)\cr levels in each vector provided will be collapsed into
the new level given by the respective name.}
}
\value{
The modified factor with collapsed levels. Values and levels which are not included
in the given character vectors input are discarded. The returned factor has levels in the order
given.
}
\description{
This collapses levels and only keeps those new group levels, in the order provided.
}
\examples{
fct_collapse_only(factor(c("a", "b", "c", "d")), TRT = "b", CTRL = c("c", "d"))

}
\seealso{
\code{\link[forcats:fct_collapse]{forcats::fct_collapse()}}, \code{\link[=fct_discard]{fct_discard()}}, \code{\link[forcats:fct_relevel]{forcats::fct_relevel()}} which are used
internally.
}
