% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_factor.R
\name{fct_discard}
\alias{fct_discard}
\title{Discard Certain Levels from a Factor}
\usage{
fct_discard(x, discard)
}
\arguments{
\item{x}{(\code{factor})\cr the original factor.}

\item{discard}{(\code{character})\cr which levels to discard.}
}
\value{
The modified factor with observations as well as levels from \code{discard} dropped.
}
\description{
This discards the observations as well as the levels specified from a factor.
}
\examples{
fct_discard(factor("a", "b", "c"), "c")

}
