% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmrm.R
\name{fit_lme4}
\alias{fit_lme4}
\title{Fitting \code{lme4} Model}
\usage{
fit_lme4(formula, data, optimizer = "automatic", n_cores = 1L)
}
\arguments{
\item{formula}{(\code{formula})\cr the MMRM formula (it could also be another \code{lme4} formula).}

\item{data}{a \code{data.frame} with all the variables specified in
\code{vars}. Records with missing values in any independent variables
will be excluded.}

\item{optimizer}{a string specifying the optimization algorithm which should be used. By default, "automatic"
will (if necessary) try all possible optimization algorithms and choose the best result. If another algorithm
is chosen and does not give a valid result, an error will occur.}

\item{n_cores}{(\code{count})\cr number of cores to parallelize over the "automatic" optimizer search.}
}
\value{
the \code{\link[lmerTest:lmerModLmerTest-class]{lmerTest::lmerModLmerTest}} object.
}
\description{
Helper function to fit the MMRM with \code{lme4} and \code{lmerTest.}
}
