% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mmrm.R
\name{fit_lme4_single_optimizer}
\alias{fit_lme4_single_optimizer}
\title{Fitting \code{lme4} Model}
\usage{
fit_lme4_single_optimizer(
  formula,
  data,
  optimizer = c("automatic", "nloptwrap_neldermead", "nloptwrap_bobyqa", "nlminbwrap",
    "bobyqa", "neldermead", "nmkbw", "optimx_lbfgsb")
)
}
\arguments{
\item{formula}{(\code{formula})\cr the \code{lme4} formula.}

\item{data}{a \code{data.frame} with all the variables specified in
\code{vars}. Records with missing values in any independent variables
will be excluded.}

\item{optimizer}{a string specifying the optimization algorithm which should be used. By default, "automatic"
will (if necessary) try all possible optimization algorithms and choose the best result. If another algorithm
is chosen and does not give a valid result, an error will occur.}
}
\value{
the fitted \code{\link[lmerTest:lmerModLmerTest-class]{lmerTest::lmerModLmerTest}} object, with additional attributes \code{messages} and \code{optimizer}.
}
\description{
Internal helper function to fit an lme4 model with a single optimizer, while capturing messages and warnings.
}
\note{
While we are not directly importing functions from \code{dfoptim} or \code{optimx} here, we rely on them
being available to \code{lme4} (which has these packages only in "Suggests"). Therefore we note the imports below.
}
