% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R
\name{footnotes<-}
\alias{footnotes<-}
\title{Attribute}
\usage{
footnotes(x) <- value
}
\arguments{
\item{x}{an object}

\item{value}{character vector}
}
\description{
Assign value to attribute footnote of object x
}
\examples{
x <- table(iris$Species)
footnotes(x) <- "Species are equally distributed"
attributes(x)
}
