% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R, R/g_forest.R
\name{footnotes<-}
\alias{footnotes<-}
\title{Assign Footnotes}
\usage{
footnotes(x) <- value

footnotes(x) <- value
}
\arguments{
\item{x}{an object}

\item{value}{character vector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Assign value to attribute footnote of object \code{x}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated here. Please look at R/footnotes for the correct function.
}
\examples{
x <- table(iris$Species)
footnotes(x) <- "Species are equally distributed"
attributes(x)
x <- table(iris$Species)
footnotes(x) <- "Species are equally distributed"
attributes(x)
}
