% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_forest.R
\name{footnotes}
\alias{footnotes}
\title{Retrieve value from attribute footnote of object x}
\usage{
footnotes(x)
}
\arguments{
\item{x}{an object}
}
\description{
Retrieve value from attribute footnote of object x
}
\examples{
x <- table(iris$Species)
footnotes(x) <- "Species are equally distributed"
footnotes(x)

}
