% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R, R/g_forest.R
\name{footnotes}
\alias{footnotes}
\title{Retrieve Footnotes}
\usage{
footnotes(x)

footnotes(x)
}
\arguments{
\item{x}{an object}
}
\description{
Retrieve value from attribute \code{footnote} of object \code{x}.
}
\examples{
x <- table(iris$Species)
footnotes(x) <- "Species are equally distributed"
footnotes(x)

x <- table(iris$Species)
footnotes(x) <- "Species are equally distributed"
footnotes(x)

}
