% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_rsp.R
\name{forest_rsp}
\alias{forest_rsp}
\title{Response Forest Plot Table}
\usage{
forest_rsp(response, event, arm, group_data, covariates = NULL)
}
\arguments{
\item{response}{Tumor Response data}

\item{event}{is boolean, \code{TRUE} if responder, \code{FALSE} if non-responder}

\item{arm}{vector with arm information}

\item{group_data}{data frame with one column per sub-group variable}

\item{covariates}{set to NULL; currently not available for multivariate survival analysis}
}
\description{
Response Forest Plot Table
}
\details{
Logistic model is used for odds ratio calculation
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr) 
library(grid)
library(teal.oncology)
library(forcats)
'\%needs\%' <- teal.oncology:::'\%needs\%'
ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)
ASL$temp <- c(rep("A", 200), rep("B", 250), rep(NA,1202-450))

ARS_f <- ARS \%>\% filter(PARAMCD == "OVRSPI") \%>\% 
                 filter(ITTWTFL == "N") \%>\% 
                 filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\%
                 select(c("USUBJID", "STUDYID", "SEX", "ICLEVEL", "TC3IC3", "ARM", "AVAL", "AVALC"))
ASL_f <- ASL \%>\% filter(ITTWTFL == "N") \%>\% 
             filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\% 
             select(c("USUBJID", "STUDYID", "SEX", "MLIVER", "TCICLVL2", "AGE4CAT", "RACE", "temp"))

group_data <- left_join(ASL_f, ATE_f \%>\% select(c("USUBJID", "STUDYID")))
names(group_data) <- labels_over_names(group_data)

head(group_data)

arm <- fct_relevel(ARS_f$ARM, "DUMMY C")

tbl <- forest_rsp(
          response = ARS_f$AVAL,
          event = ARS_f$AVALC \%in\% c("CR","PR"),
          arm = arm, 
          group_data = group_data[, -c(1,2), drop=FALSE]
)
Viewer(tbl)

}


  
}
\author{
Yuyao Song (songy24), \email{yuyao.song@roche.com}
}
