% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_rsp.R
\name{forest_rsp}
\alias{forest_rsp}
\title{alternative tablist}
\usage{
forest_rsp(response, event, arm, group_by, covariARSs = NULL)
}
\arguments{
\item{event}{is boolean, \code{TRUE} if event, \code{FALSE} if time_to_event
is censored}

\item{arm}{vector with arm information}

\item{group_by}{data frame with one column per grouping}

\item{time_to_event}{Response data}

\item{arm.ref}{a character vector defining which arms in arm should be taken 
as the reference}

\item{arm.comp}{a character vector defining which arms in arm should be taken 
as the comparison}
}
\description{
alternative tablist
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr) 
library(grid)
'\%needs\%' <- teal.oncology:::'\%needs\%'
ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

resp_tbl_stream <- get_forest_response_table(com.roche.cdt30019.go29436.re)
Viewer(resp_tbl_stream )

ARS_f <- ARS \%>\% filter(PARAMCD == "OVRSPI") \%>\% 
                 filter(ITTWTFL == "Y") \%>\% 
                 filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\%
                 select(c("USUBJID", "STUDYID", "SEX", "ICLEVEL", "TC3IC3", "ARM", "AVAL", "AVALC"))
ASL_f <- ASL \%>\% filter(ITTWTFL == "Y") \%>\% filter(ARM \%in\% c("DUMMY A", "DUMMY C"))

ARS_f_resp <- ARS_f \%>\% filter(AVALC \%in\% c("CR", "PR"))
ARS_f_IC1 <- ARS_f \%>\% filter(TCLEVEL == "1")


group_by <- ARS_f[c("USUBJID", "STUDYID", "SEX", "ICLEVEL", "TC3IC3")]
names(group_by) <- c(Reduce(cbind,lapply(group_by, function(x){attr(x,"label")})))

head(group_by)

# Dummy C First (comparison in survfit and glm)
arm <- fct_relevel(ATE_f$ARM, "DUMMY C")

tbl <- forest_rsp(
          response = ARS_f$AVAL,
          event = ARS_f$AVALC \%in\% c("CR","PR"),
          arm = arm, 
          group_by = group_by[, -c(1,2), drop=FALSE]
)
Viewer(tbl)

compare_rtables(tbl, surv_tbl_stream, comp.attr = FALSE)

plot(tbl)

}


  
}
