% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_tte.R
\name{forest_tte}
\alias{forest_tte}
\title{Time-to-event Forest Plot Table}
\usage{
forest_tte(time_to_event, event, arm, group_data, covariates = NULL)
}
\arguments{
\item{time_to_event}{time to event data}

\item{event}{is boolean, \code{TRUE} if event, \code{FALSE} if time_to_event
is censored}

\item{arm}{vector with arm information}

\item{group_data}{data frame with one column per grouping}

\item{covariates}{set to NULL; currently not available for multivariate survival analysis}
}
\description{
Time-to-event Forest Plot Table
}
\details{
Cox PH model is used for hazard ratio calculation
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr)
library(forcats)
library(survival)
library(teal.oncology)

'\%needs\%' <- teal.oncology:::'\%needs\%'

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

tbl_stream <- get_forest_survival_table(com.roche.cdt30019.go29436.re)
Viewer(tbl_stream)


ATE_f <- ATE \%>\% filter(PARAMCD == "OS") \%>\% 
             filter(ITTWTFL == "Y") \%>\% 
             filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\%
             select(c("USUBJID", "STUDYID", "AVAL", "CNSR", "ARMCD"))
ASL_f <- ASL \%>\% filter(ITTWTFL == "Y") \%>\% 
             filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\% 
             select(c("USUBJID", "STUDYID", "SEX", "MLIVER", "TCICLVL2", "AGE4CAT", "RACE"))


group_data <- left_join(ASL_f, ATE_f \%>\% select(c("USUBJID", "STUDYID")))
group_data$MLIVER <- factor(group_data$MLIVER, levels = c("Y", "N"), labels = c("Yes", "No"))
group_data$TCICLVL2 <- factor(group_data$TCICLVL2, levels = c("TC3 or IC2/3", "TC0/1/2 and IC0/1"), labels = c("TC3 or IC2/3", "TC0/1/2 and IC0/1"))#' 
group_data$SEX <- factor(group_data$SEX, levels = c("F", "M"), labels = c("FEMALE", "MALE"))
group_data$AGE4CAT <- factor(group_data$AGE4CAT, levels = c("<65", "65 to 74", "75 to 84", ">=85"), labels = c("<65", "65 to 74", "75 to 84", ">=85"))
group_data$RACE <- factor(group_data$RACE, levels = c("ASIAN", "BLACK OR AFRICAN AMERICAN", "WHITE", "UNKNOWN"), labels = c("ASIAN", "BLACK OR AFRICAN AMERICAN", "WHITE", NA))
names(group_data) <- labels_over_names(group_data)
head(group_data)

arm <- fct_relevel(ATE_f$ARMCD, "C")

tbl <- forest_tte(
   time_to_event = ATE_f$AVAL,
   event = ATE_f$CNSR == 0,
   group_data = group_data[, -c(1,2), drop=FALSE],
   arm = arm
)
Viewer(tbl)

Viewer(tbl, tbl_stream)

compare_rtables(tbl, tbl_stream, comp.attr = FALSE)

forest_tte_plot(tbl)

}

# forest_tte(Surv(AVAL ~ I(CNSR != 'N') ~ ARM + SEX, data = ATE))
}
\author{
Yuyao Song (songy24), \email{yuyao.song@roche.com}
}
