% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_tte.R
\name{forest_tte}
\alias{forest_tte}
\title{alternative tablist}
\usage{
forest_tte(time_to_event, event, arm, group_by, covariates = NULL)
}
\arguments{
\item{time_to_event}{time to event data}

\item{event}{is boolean, \code{TRUE} if event, \code{FALSE} if time_to_event
is censored}

\item{arm}{vector with arm information}

\item{group_by}{data frame with one column per grouping}

\item{arm.ref}{a character vector defining which arms in arm should be taken 
as the reference}

\item{arm.comp}{a character vector defining which arms in arm should be taken 
as the comparison}
}
\description{
alternative tablist
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr)
library(survival)
library(forcats)
library(forestplot) 

'\%needs\%' <- teal.oncology:::'\%needs\%'

surv_tbl_stream <- get_forest_survival_table(com.roche.cdt30019.go29436.re)
Viewer(surv_tbl_stream )

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

ATE_f <- ATE \%>\% filter(PARAMCD == "OS") \%>\% 
             filter(ITTWTFL == "Y") \%>\% 
             filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\%
             select(c("USUBJID", "STUDYID", "AVAL", "CNSR", "ARM"))
ASL_f <- ASL \%>\% filter(ITTWTFL == "Y") \%>\% 
             filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\% 
             select(c("USUBJID", "STUDYID", "SEX", "MLIVER", "TCICLVL2", "AGE4CAT", "RACE", "BECOG"))


group_by <- left_join(ASL_f, ATE_f \%>\% select(c("USUBJID", "STUDYID")))
old_grp_names <- names(group_by) 
new_grp_names <- c(Reduce(cbind,lapply(group_by, function(x){
     if (is.null(attr(x,"label"))) NA else attr(x,"label")
      })
     ))
names(group_by) <- ifelse (is.na (new_grp_names), old_grp_names, new_grp_names)

head(group_by)

# Dummy C First (comparison in survfit and glm)
arm <- fct_relevel(ATE_f$ARM, "DUMMY C")

# Collapse Factors
# fct_collapse(arm, "ARM A/B" = c("ARM A", "ARM B"))
# rename a level
# arm <- fct_recode(ATE_f$ARM, "Treat ARM" = "DUMMY A")

tbl <- forest_tte(
   time_to_event = ATE_f$AVAL,
   event = ATE_f$CNSR == 0,
   group_by = group_by[, -c(1,2), drop=FALSE],
   arm = arm
)
tbl <- forest_tte_table(tbl)
tbl
Viewer(tbl)
fplot <- forest_tte_plot(tbl)
compare_rtables(tbl, surv_tbl_stream, comp.attr = FALSE)

plot(tbl)


}

# forest_tte(Surv(AVAL ~ I(CNSR != 'N') ~ ARM + SEX, data = ATE))
}
