% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_tte.R
\name{forest_tte}
\alias{forest_tte}
\title{Time-to-event Forest Plot Table}
\usage{
forest_tte(time_to_event, event, arm, group_data, covariates = NULL)
}
\arguments{
\item{time_to_event}{time to event data}

\item{event}{is boolean, \code{TRUE} if event, \code{FALSE} if time_to_event
is censored}

\item{arm}{vector with arm information}

\item{group_data}{data frame with one column per grouping}

\item{covariates}{set to NULL; currently not available for multivariate survival analysis}
}
\description{
Time-to-event Forest Plot Table
}
\details{
Cox PH model is used for hazard ratio calculation
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr)
library(forcats)
library(survival)
library(teal.oncology)

'\%needs\%' <- teal.oncology:::'\%needs\%'

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

ATE_f <- ATE \%>\% filter(PARAMCD == "OS") \%>\% 
             filter(ITTWTFL == "Y") \%>\% 
             filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\%
             select(c("USUBJID", "STUDYID", "AVAL", "CNSR", "ARMCD"))
ASL_f <- ASL \%>\% filter(ITTWTFL == "Y") \%>\% 
             filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\% 
             select(c("USUBJID", "STUDYID", "SEX", "MLIVER", "AGE4CAT", "BECOG"))


group_data <- left_join(ASL_f, ATE_f \%>\% select(c("USUBJID", "STUDYID")))
names(group_data) <- labels_over_names(group_data)
head(group_data)

arm <- fct_relevel(ATE_f$ARMCD, "C")

tbl <- forest_tte(
   time_to_event = ATE_f$AVAL,
   event = ATE_f$CNSR == 0,
   group_data = group_data[, -c(1,2), drop=FALSE],
   arm = arm
)
Viewer(tbl)

}

# forest_tte(Surv(AVAL ~ I(CNSR != 'N') ~ ARM + SEX, data = ATE))
}
\author{
Yuyao Song (songy24), \email{yuyao.song@roche.com}
}
