% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_functions.R
\name{format_auto}
\alias{format_auto}
\title{Automatic formats from data significant digits}
\usage{
format_auto(res, x, x_stat, ...)
}
\arguments{
\item{res}{(\code{numeric}) \cr results from statistical evaluation. It can be more than one
element (e.g. for \code{.stats = "mean_sd"}).}

\item{x}{(\code{numeric}) \cr all the data the statistics was created upon. Used only to find
significant digits.}

\item{x_stat}{(\code{string}) \cr string indicating the current statistical method used.}

\item{...}{\code{rtables} standard for all format functions}
}
\value{
A string that \code{rtables} prints in a table cell.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Formatting function for the majority of default methods used in \code{\link[=analyze_vars]{analyze_vars()}}.
For non-derived values, the significant digits of data is used (e.g. range), while derived
values have one more digits (measure of location and dispersion like mean, standard deviation).
This function can be called internally with "auto" like, for example,
\code{.formats = c("mean" = "auto")}.
}
\examples{
x_todo <- c(0.001, 0.2, 0.0011000, 3, 4)
res <- c(mean(x_todo[1:3]), sd(x_todo[1:3]))
format_auto(res = res, x = x_todo, x_stat = "mean_sd")
format_auto(range(x_todo), x_todo, "range")
no_sc_x <- c(0.0000001, 1)
format_auto(no_sc_x, no_sc_x, "range")

}
\seealso{
Other formatting functions: 
\code{\link{extreme_format}},
\code{\link{format_count_fraction_fixed_dp}()},
\code{\link{format_count_fraction}()},
\code{\link{format_extreme_values_ci}()},
\code{\link{format_extreme_values}()},
\code{\link{format_fraction_fixed_dp}()},
\code{\link{format_fraction_threshold}()},
\code{\link{format_fraction}()},
\code{\link{format_xx}()},
\code{\link{formatting_functions}}
}
\concept{formatting functions}
\keyword{internal}
