% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_extreme_values}
\alias{format_extreme_values}
\title{Formatting a Single Extreme Value}
\usage{
format_extreme_values(digits = 2L)
}
\arguments{
\item{digits}{(\code{integer})\cr number of decimal places to display.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Create Formatting Function for a single extreme value.
}
\examples{
format_fun <- format_extreme_values(2L)
format_fun(x = 0.127)
format_fun(x = Inf)
format_fun(x = 0)
format_fun(x = 0.009)

}
