% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_extreme_values_ci}
\alias{format_extreme_values_ci}
\title{Formatting Extreme Values Part of a Confidence Interval}
\usage{
format_extreme_values_ci(digits = 2L)
}
\arguments{
\item{digits}{(\code{integer})\cr number of decimal places to display.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Formatting Function for extreme values part of a confidence interval. Values
are formatted as e.g. "(xx.xx, xx.xx)" if if the number of \code{digits} is 2.
}
\examples{
format_fun <- format_extreme_values_ci(2L)
format_fun(x = c(0.127, Inf))
format_fun(x = c(0, 0.009))

}
