% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_fraction}
\alias{format_fraction}
\title{Formatting Fraction and Percentage}
\usage{
format_fraction(x, ...)
}
\arguments{
\item{x}{(\code{integer})\cr with elements \code{num} and \code{denom}.}

\item{...}{required for \code{rtables} interface.}
}
\value{
a string in the format \verb{num / denom (ratio \%)}. If \code{num} is 0 the format is \code{num / denom}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Formats a fraction together with ratio in percent.
}
\examples{
format_fraction(x = c(num = 2L, denom = 3L))
format_fraction(x = c(num = 0L, denom = 3L))
}
\seealso{
Other formatting functions: 
\code{\link{extreme_format}},
\code{\link{format_count_fraction}()},
\code{\link{format_fraction_threshold}()},
\code{\link{formatting_functions}}
}
\concept{formatting functions}
