% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_fraction}
\alias{format_fraction}
\title{Formatting Fraction and Percentage}
\usage{
format_fraction(x, ...)
}
\arguments{
\item{x}{(\code{integer})\cr with elements \code{num} and \code{denom}.}

\item{...}{required for rtables interface.}
}
\value{
a string in the format \verb{num / denom (ratio \%)}. If \code{num} is 0 the format is \code{num / denom}.
}
\description{
Formats a fraction together with ratio in percent.
}
\examples{
format_fraction(x = c(num = 2L, denom = 3L))
format_fraction(x = c(num = 0L, denom = 3L))

}
\seealso{
Other formatting functions: 
\code{\link{formatting_functions}}
}
\concept{formatting functions}
