% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_functions.R
\name{format_sigfig}
\alias{format_sigfig}
\title{Formatting Numeric Values By Significant Figures}
\usage{
format_sigfig(sigfig)
}
\arguments{
\item{sigfig}{(\code{integer})\cr number of significant figures to display.}

\item{x}{(\code{numeric})\cr value.}

\item{...}{required for \code{rtables} interface.}
}
\value{
An \code{rtables} formatting function.
}
\description{
Format numeric values to print with a specified number of significant figures.
}
\examples{
fmt_3sf <- format_sigfig(3)
fmt_3sf(1.658)
fmt_3sf(1e1)

fmt_5sf <- format_sigfig(5)
fmt_5sf(0.57)
fmt_5sf(0.000025645)

}
\seealso{
Other formatting functions: 
\code{\link{extreme_format}},
\code{\link{format_count_fraction_fixed_dp}()},
\code{\link{format_count_fraction}()},
\code{\link{format_extreme_values_ci}()},
\code{\link{format_extreme_values}()},
\code{\link{format_fraction_fixed_dp}()},
\code{\link{format_fraction_threshold}()},
\code{\link{format_fraction}()},
\code{\link{format_xx}()},
\code{\link{formatting_functions}}
}
\concept{formatting functions}
