% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{format_wrap}
\alias{format_wrap}
\alias{format_wrap_df}
\alias{format_wrap_x}
\title{Construct analysis functions with additional formatting arguments for use with \code{analyze}.}
\usage{
format_wrap_df(sfun, indent_mods, formats)

format_wrap_x(sfun, indent_mods, formats)
}
\arguments{
\item{sfun}{the original statistics function returning a named list of results}

\item{indent_mods}{named vector with default (nonnegative) indent modifications}

\item{formats}{named vector with default formats}
}
\value{
the constructed analysis function
}
\description{
The returned function has first argument \code{df} or \code{x} and uses any additional arguments for the
original \code{sfun}. It has additional four formatting arguments:
\describe{
\item{.stats}{character vector to select statistics from the \code{sfun} result}
\item{.indent_mods}{named vector with custom (nonnegative) indent modifications}
\item{.formats}{named vector with custom formats}
\item{.labels}{named vector with custom labels}
}
Note that \code{.indent_mods}, \code{.formats} and \code{.labels} don't need to contain an element for each
statistic, as they have defaults initialized either from the arguments below (for indents and formats)
or from the \code{sfun} results (for labels).
}
\section{Functions}{
\itemize{
\item \code{format_wrap_df}: the wrapper for \code{sfun} having \code{df} as first argument.

\item \code{format_wrap_x}: the wrapper for \code{sfun} having \code{x} as first argument.
}}

\examples{
sfun <- function(df) {
  list(
    nrows = nrow(df),
    ncols = length(df)
  )
}
afun <- format_wrap_df(
  sfun,
  indent_mods = c(nrows = 0L, ncols = 2L),
  formats = c(nrows = "xx.", ncols = "xx.xx")
)
df <- data.frame(
  a = c(1, 2),
  b = c(3, 4)
)
afun(df)
afun(
  df,
  .indent_mods = c(nrows = 3L),
  .stats = "nrows",
  .formats = c(nrows = "xx.xx"),
  .labels = c(nrows = "Number of rows")
)
sfun <- function(x) {
  list(
    n = length(x),
    mean = mean(x),
    median = median(x)
  )
}
afun <- format_wrap_x(
  sfun,
  indent_mods = c(n = 0L, mean = 2L, median = 1L),
  formats = c(n = "xx.", mean = "xx.xx", median = "xx")
)
x <- c(1, 0, -1, 2, 5, 3, 2.5, 7.1)
afun(x)
afun(
  x,
  .indent_mods = c(median = 3L),
  .stats = c("n", "median"),
  .formats = c(median = "xx.xx"),
  .labels = c(n = "Number of numbers")
)
}
