% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_functions.R
\name{formatting_functions}
\alias{formatting_functions}
\title{Formatting Functions}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

See below for the list of formatting functions created in \code{tern} to work with \code{rtables}.

Other available formats can be listed via \code{\link[formatters:list_formats]{formatters::list_valid_format_labels()}}. Additional
custom formats can be created via the \code{\link[formatters:sprintf_format]{formatters::sprintf_format()}} function.
}
\seealso{
Other formatting functions: 
\code{\link{extreme_format}},
\code{\link{format_auto}()},
\code{\link{format_count_fraction_fixed_dp}()},
\code{\link{format_count_fraction_lt10}()},
\code{\link{format_count_fraction}()},
\code{\link{format_extreme_values_ci}()},
\code{\link{format_extreme_values}()},
\code{\link{format_fraction_fixed_dp}()},
\code{\link{format_fraction_threshold}()},
\code{\link{format_fraction}()},
\code{\link{format_sigfig}()},
\code{\link{format_xx}()}
}
\concept{formatting functions}
