% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{formatting_functions}
\alias{formatting_functions}
\title{Additional Formatting Functions}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This summarizes the additional Formatting Functions to work with \code{rtables}.
}
\seealso{
Other formatting functions: 
\code{\link{extreme_format}},
\code{\link{format_count_fraction}()},
\code{\link{format_fraction_threshold}()},
\code{\link{format_fraction}()}
}
\concept{formatting functions}
