% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_forest.R
\name{g_forest}
\alias{g_forest}
\title{Forest plot (table + graph)}
\usage{
g_forest(x, arm.ref = "ReferenceAAAvery longtitle",
  arm.comp = "Treatment AAAverylongtitle", anl.model = "logistic",
  time.unit = "month", padx = unit(0, "lines"), cex = 1)
}
\arguments{
\item{x}{rtable from forest_rsp function or from forest_tte function}

\item{arm.ref}{string used to label reference arm}

\item{arm.comp}{string used to label comparable arm}

\item{anl.model}{the regression method used for analysis. Valid options are \t{survival} and \t{logistic}}

\item{time.unit}{string used to display the unit of median survival time. default is "month".}

\item{padx}{gap between two columns}

\item{cex}{multiplier applied to overall fontsize}
}
\description{
Forest plot (table + graph)
}
\examples{

library(random.cdisc.data)
ASL <- radam("ASL")
ARS <- radam("ARS", ADSL = ASL)

ARS_f <- ARS \%>\% filter(PARAMCD == "OVRSPI") 

ASL_f <- right_join(ASL \%>\% select(USUBJID, STUDYID, SEX, RACE, ARM),
                        ARS_f \%>\% select(USUBJID, STUDYID))

tbl <- t_forest_rsp(
  response = ARS_f$AVAL,
  event = ARS_f$AVALC \%in\% c("CR","PR"),
  arm = ASL_f$ARM, 
  group_data = ASL_f \%>\% select("SEX", "RACE")
)

g_forest(tbl)

\dontrun{
library(atezo.data)
library(dplyr) 
library(grid)
library(teal.oncology)
library(forcats)

'\%needs\%' <- teal.oncology:::'\%needs\%'
ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

tbl_stream <- get_forest_response_table(com.roche.cdt30019.go29436.re)
Viewer(tbl_stream)

ARS_f <- ARS \%>\% filter(PARAMCD == "OVRSPI") \%>\% 
                 filter(ITTWTFL == "Y") \%>\% 
                 filter(ARM \%in\% c("DUMMY A", "DUMMY C")) \%>\%
                 select(c("USUBJID", "STUDYID", "SEX", "ICLEVEL", "TC3IC3", "ARM", "AVAL", "AVALC"))
ASL_f <- ASL \%>\% filter(ITTWTFL == "Y") \%>\% filter(ARM \%in\% c("DUMMY A", "DUMMY C"))

group_data <- ARS_f[c("USUBJID", "STUDYID", "ICLEVEL", "TC3IC3")]
names(group_data) <- labels_over_names(group_data)

head(group_data)

arm <- fct_relevel(ARS_f$ARM, "DUMMY C")

tbl <- t_forest_rsp(
          response = ARS_f$AVAL,
          event = ARS_f$AVALC \%in\% c("CR","PR"),
          arm = arm, 
          group_data = group_data[, -c(1,2), drop=FALSE]
)

Viewer(tbl)

library(grid)
g_forest(tbl, arm.ref = "ReferenceAAAvery longtitle", arm.comp = "Treatment AAAverylongtitle", padx=unit(0, "lines"))

}
}
\author{
Yuyao Song (songy24), \email{yuyao.song@roche.com}
}
