% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_forest.R
\name{g_forest}
\alias{g_forest}
\title{Forest plot}
\usage{
g_forest(tbl, i_col_est, i_col_ci, header_forest, padx = unit(1, "lines"))
}
\arguments{
\item{tbl}{an rtable object}

\item{i_col_est}{column index with estimator}

\item{i_col_ci}{column index with confidence intervals}

\item{header_forest}{how to label the forest}

\item{padx}{gap between two columns}

\item{cex}{multiplier applied to overall fontsize}
}
\description{
Create a forest plot from any \code{\link[rtables]{rtable}} object that has a
column with a single value and a column with 2 values
}
\examples{

library(random.cdisc.data)

ASL <- radam("ASL")
ASL$RACE <- factor(sapply(as.character(ASL$RACE), function(x) if (nchar(x)>9) paste0(substr(x, 1,9), "...") else x))
ATE <- radam("ATE", ADSL = ASL)

ATE_f <- subset(ATE, PARAMCD == "OS") 

ANL <- merge(ASL, ATE_f)

tbl <- t_forest_tte(
  tte = ANL$AVAL,
  is_event = ANL$CNSR == 0,
  col_by = factor(ANL$ARM), 
  group_data = as.data.frame(lapply(ANL[, c("SEX", "RACE")], as.factor)),
  dense_header = TRUE
)

## note plot requires a certain width
g_forest(
  tbl = tbl,
  i_col_est = 8,
  i_col_ci = 9,
  header_forest = c("Treatement\\nBetter", "Comparison\\nBetter")
)


}
\author{
Yuyao Song (songy24) \email{yuyao.song@roche.com}
}
