% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_forest.R
\name{g_forest}
\alias{g_forest}
\title{Create a Forest Plot based on a Table}
\usage{
g_forest(
  tbl,
  col_x,
  col_ci,
  vline = NULL,
  forest_header = NULL,
  xlim = NULL,
  logx = FALSE,
  x_at = NULL,
  width_row_names = NULL,
  width_columns = NULL,
  width_forest = unit(1, "null"),
  col_symbol_size = NULL,
  draw = TRUE,
  newpage = TRUE
)
}
\arguments{
\item{tbl}{(\code{rtable})}

\item{col_x}{(\code{integer})\cr column index with estimator.}

\item{col_ci}{(\code{integer})\cr column index with confidence intervals.}

\item{vline}{(\code{number})\cr
x coordinate for vertical line, if \code{NULL} then the line is omitted.}

\item{forest_header}{(\code{character}, length 2)\cr
text displayed to the left and right of \code{vline}, respectively.
If \code{vline = NULL} then \code{forest_header} needs to be \code{NULL} too.}

\item{xlim}{(\code{numeric})\cr limits for x axis.}

\item{logx}{(\code{flag})\cr show the x-values on logarithm scale.}

\item{x_at}{(\verb{numeric)\\cr x-tick locations, if }NULL` they get automatically chosen.}

\item{width_row_names}{(\code{unit})\cr
width for row names. If \code{NULL} the widths get automatically calculated. See \code{\link[grid:unit]{grid::unit()}}.}

\item{width_columns}{(\code{unit})\cr
widths for the table columns. If \code{NULL} the widths get automatically calculated. See \code{\link[grid:unit]{grid::unit()}}.}

\item{width_forest}{(\code{unit})\cr
width for the forest column. If \code{NULL} the widths get automatically calculated. See \code{\link[grid:unit]{grid::unit()}}.}

\item{col_symbol_size}{(\code{integer})\cr
column index from \code{tbl} containing data to be used to determine relative
size for estimator plot symbol. Typically, the symbol size is proportional to the
sample size used to calculate the estimator. If \code{NULL}, the same symbol
size is used for all subgroups.}

\item{draw}{(\code{flag})\cr whether the plot should be drawn.}

\item{newpage}{(\code{flag})\cr
if \code{draw = TRUE} should plot be drawn on a new page.}
}
\description{
Create a forest plot from any \code{\link[rtables:rtable]{rtables::rtable()}} object that has a
column with a single value and a column with 2 values.
}
\examples{

library(random.cdisc.data)
library(dplyr)
library(forcats)
library(rtables)

adrs <- radrs(cached = TRUE)

n_records <- 20
adrs_labels <- var_labels(adrs)
adrs <- adrs \%>\%
  filter(PARAMCD == "BESRSPI") \%>\%
  filter(ARM \%in\% c("A: Drug X", "B: Placebo")) \%>\%
  slice(seq_len(n_records)) \%>\%
  droplevels() \%>\%
  mutate(
    # Reorder levels of factor to make the placebo group the reference arm.
    ARM = fct_relevel(ARM, "B: Placebo"),
    rsp = AVALC == "CR"
  )
var_labels(adrs) <- c(adrs_labels, "Response")

df <- extract_rsp_subgroups(
  variables = list(rsp = "rsp", arm = "ARM", subgroups = c("SEX", "STRATA2")),
  data = adrs
)

# Full commonly used response table.
tbl <- basic_table() \%>\%
  tabulate_rsp_subgroups(df)

p <- g_forest(
  tbl = tbl,
  col_x = 6,
  col_ci = 7,
  vline = 1,
  forest_header = c("Treatement\nBetter", "Comparison\nBetter"),
  xlim = c(.1, 10),
  logx = TRUE,
  x_at = c(.1, 1, 10),
  draw = FALSE
)

draw_grob(p)

# Odds rato only table.
tbl_or <- basic_table() \%>\%
  tabulate_rsp_subgroups(df, vars = c("n_tot", "or", "ci"))
tbl_or

p <- g_forest(
  tbl = tbl_or,
  col_x = 2,
  col_ci = 3,
  vline = 1,
  forest_header = c("Treatement\nBetter", "Comparison\nBetter"),
  xlim = c(.1, 10),
  logx = TRUE,
  x_at = c(.1, 1, 10),
  draw = FALSE
)

draw_grob(p)

# Works with any rtable

tbl <- rtable(
  header = c("E", "CI", "N"),
  rrow("", 1, c(.8, 1.2), 200),
  rrow("", 1.2, c(1.1, 1.4), 50)
)

g_forest(
  tbl = tbl,
  col_x = 1,
  col_ci = 2,
  col_symbol_size = 3
)

tbl <- rtable(
  header = rheader(
    rrow("", rcell("A", colspan = 2)),
    rrow("", "c1", "c2")
  ),
  rrow("row 1", 1, c(.8, 1.2)),
  rrow("row 2", 1.2, c(1.1, 1.4))
)

g_forest(
  tbl = tbl,
  col_x = 1,
  col_ci = 2,
  vline = 1,
  forest_header = c("Hello", "World")
)

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
