% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_km.R
\name{g_km}
\alias{g_km}
\title{Kaplan-Meier Plot}
\usage{
g_km(fit_km, xticks = NULL, col = NULL, lty = NULL,
  title = "Kaplan - Meier Plot", xlab = "Days", gp = NULL,
  vp = NULL, name = NULL, draw = TRUE, newpage = TRUE)
}
\arguments{
\item{fit_km}{a class \code{\link{survfit}} object.}

\item{xticks}{break interval of x-axis. It takes a numeric vector or \code{NULL}.}

\item{col}{a vector of color for each line.}

\item{lty}{a vector of line type for each curve.}

\item{title}{title for plot.}

\item{xlab}{a string for label of x-axis}

\item{draw}{draw the plot on device.}

\item{newpage}{open a new draw page.}
}
\description{
Create a KM plot for any \code{\link[survival]{survfit}} object.
}
\examples{
library(random.cdisc.data)
library(tern)
ADSL <- radsl()
ADSL$RACE <- factor(sapply(as.character(ADSL$RACE), function(x) {
   if (nchar(x)>9) paste0(substr(x, 1,9), "...") else x
}))
ADTTE <- radtte(ADSL)

ADTTE_f <- subset(ADTTE, PARAMCD == "OS") 

fit_km <- survfit(Surv(AVAL, 1-CNSR) ~ ARM, data = ADTTE_f, conf.type = "plain")

g_km(fit_km = fit_km)

g_km(fit_km = fit_km, col = c("black", "red", "blue"))
 
fit_km <- survfit(Surv(AVAL, 1-CNSR) ~ 1, data = ADTTE, conf.type = "plain")
g_km(fit_km, xlab = "Duration (Days)", col = "blue", lty = "dashed")

}
\author{
Heng Wang (wangh107) \email{wang.heng@gene.com}
}
