% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_km.R
\name{g_km}
\alias{g_km}
\title{Kaplan-Meier Plot}
\usage{
g_km(fit_km, col = NULL, xticks = NULL, title = "Kaplan - Meier Plot",
  xlab = "Days", draw = TRUE, newpage = TRUE)
}
\arguments{
\item{fit_km}{a class \code{\link{survfit}} object}

\item{col}{a vector of color for each line}

\item{xticks}{break interval of x-axis. It takes a numeric vector or
\code{NULL}.}

\item{title}{plot title}

\item{xlab}{a string for label of x-axis}

\item{draw}{boolean, should plot be drawn}

\item{newpage}{boolean if \code{draw=TRUE} should plot be drawn on a new page}
}
\description{
Create a KM plot for any \code{\link[survival]{survfit}} object.
}
\examples{
library(random.cdisc.data)

ASL <- radam("ASL")
ASL$RACE <- factor(sapply(as.character(ASL$RACE), function(x) {
   if (nchar(x)>9) paste0(substr(x, 1,9), "...") else x
}))
ATE <- radam("ATE", ADSL = ASL)

ATE_f <- subset(ATE, PARAMCD == "OS") 

ANL <- merge(ASL, ATE_f, by = c("USUBJID", "STUDYID"))
fit_km <- survfit(Surv(AVAL, 1-CNSR) ~ ARM, data = ANL, conf.type = "plain")

g_km(fit_km = fit_km)

g_km(fit_km = fit_km, col = c("black", "red"))
 
fit_km <- survfit(Surv(AVAL, 1-CNSR) ~ 1, data = ANL, conf.type = "plain")
g_km(fit_km, xlab = "Duration (Days)")

}
\author{
Heng Wang (wangh107) \email{wang.heng@gene.com}
}
