% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_km.R
\name{g_km}
\alias{g_km}
\title{Kaplan-Meier Plot}
\usage{
g_km(
  fit_km,
  xticks = NULL,
  col = NA,
  lty = 1,
  lwd = 1,
  censor_show = TRUE,
  pch = 3,
  size = unit(0.5, "char"),
  max_time = NULL,
  title = "Kaplan - Meier Plot",
  xlab = "Days",
  ylab = "Survival Probability",
  draw = TRUE,
  newpage = TRUE,
  gp = NULL,
  vp = NULL,
  name = NULL
)
}
\arguments{
\item{fit_km}{(\code{\link{survfit}} object)}

\item{xticks}{numeric vector of xticks, single number with spacing
between ticks or \code{NULL}.}

\item{col}{(\code{character} vector)\cr
lines colors. Length of a vector should be equal to number of strata
from \code{\link{survfit}} object}

\item{lty}{(\code{integer} vector)\cr
line type. Length of a vector should be equal to number of strata
from \code{\link{survfit}} object}

\item{lwd}{(\code{numeric} vector)\cr
line width. Length of a vector should be equal to number of strata
from \code{\link{survfit}} object}

\item{censor_show}{(\code{logical} value)\cr
Whether to show censored.}

\item{pch}{(\code{numeric} vector)\cr
point type for censored.}

\item{size}{(\code{numeric} vector)\cr
size of censored point, a class of \code{unit}.}

\item{max_time}{(\code{numeric})\cr
maximum value of X axis range (\code{NULL} for default).}

\item{title}{(\code{character} value)\cr
title for plot.}

\item{xlab}{(\code{character} value)\cr
label of x-axis}

\item{ylab}{(\code{character} value)\cr
label of y-axis.}

\item{draw}{(\code{logical} value)\cr
Whether to draw the plot on device.}

\item{newpage}{(\code{logical} value)\cr
Whether to open a new draw page.}

\item{gp}{(abject of class \code{gpar}).}

\item{vp}{a grid \code{viewport}.}

\item{name}{(\code{character} value)\cr
Used to uniquely identify the object.}
}
\description{
Create a KM plot for any \code{\link[survival]{survfit}} object.
}
\examples{
library(random.cdisc.data)
library(dplyr)
library(survival)

ADSL <- cadsl
levels(ADSL$RACE) <- strtrim(levels(ADSL$RACE), 10)

ADTTE <- cadtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

formula <- Surv(AVAL, 1 - CNSR) ~ ARM
fit_km <- survfit(formula, data = ADTTE_f, conf.type = "plain")

g_km(fit_km = fit_km)
plot(fit_km)

p <- g_km(fit_km = fit_km, col = c("black", "red", "blue"), lty = c(1, 2, 3), draw = FALSE)

# Refactor
# g_tkm <- textGrob(
#   label = toString(t_km(formula, data = ADTTE_f), gap = 1),
#   x = unit(1, "npc") - unit(2, "lines"),
#   y = unit(1, "npc") - unit(2, "lines"),
#   just = c("right", "top"),
#   gp = gpar(fontfamily = "mono", fontsize = 8, fontface = "bold")
# )
#
# p_t <- addGrob(p, gTree(children = gList(g_tkm), vp = vpPath("mainPlot", "kmCurve", "curvePlot")))
#
# grid.newpage()
# grid.draw(p_t)

g_km(fit_km, xlab = "Duration (Days)")
}
\author{
Heng Wang (wangh107) \email{wang.heng@gene.com}
}
