% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_mmrm.R
\name{g_mmrm_diagnostic}
\alias{g_mmrm_diagnostic}
\title{Diagnostic plots for an MMRM model.}
\usage{
g_mmrm_diagnostic(
  object,
  type = c("fit-residual", "q-q-residual"),
  z_threshold = NULL
)
}
\arguments{
\item{object}{the MMRM model result produced by \code{\link{s_mmrm}}.}

\item{type}{string specifying the type of diagnostic plot to be produced:
\describe{
\item{fit-residual}{A fitted vs residuals plot, grouped by visits. This allows to see if there is remaining
structure in the residuals that might be captured by adding additional covariates to the model.}
\item{q-q-residual}{A Q-Q plot for the residuals (i.e. sorted standardized residuals vs. normal quantiles),
grouped by visits. Observations with an absolute standardized residual above \code{z_threshold} will be
labeled.}
}}

\item{z_threshold}{optional number indicating the normal quantile threshold for the Q-Q plot.}
}
\value{
a \code{ggplot2} plot
}
\description{
This function produces diagnostic plots.
}
\details{
Here we use marginal fitted values and residuals. That is, only the fixed effects are used
to estimate fitted values, and the difference of those fitted values vs. the observed data are
the residuals. This is in line with the MMRM philosophy where random effects are only used to
model the marginal residual distribution.
}
\examples{
library(dplyr)
library(random.cdisc.data)

adsl <- radsl(cached = TRUE)
adqs <- radqs(cached = TRUE)
adqs_f <- adqs \%>\%
  dplyr::filter(PARAMCD=="FKSI-FWB" & !AVISIT \%in\% c("BASELINE")) \%>\%
  droplevels() \%>\%
  dplyr::mutate(ARM = factor(ARM, levels = c("B: Placebo", "A: Drug X", "C: Combination"))) \%>\%
  dplyr::mutate(AVISITN = rank(AVISITN) \%>\% as.factor() \%>\% as.numeric() \%>\% as.factor())

mmrm_results <- s_mmrm(
  vars = list(
    response = "AVAL",
    covariates = c("STRATA1", "BMRKR2"),
    id = "USUBJID",
    arm = "ARM",
    visit = "AVISIT"
  ),
  data = adqs_f,
  cor_struct = "random-quadratic",
  weights_emmeans = "equal"
)
g_mmrm_diagnostic(mmrm_results)
g_mmrm_diagnostic(mmrm_results, type = "q-q-residual")
}
\seealso{
\code{\link{g_mmrm_lsmeans}} for plotting the LS means and contrasts.
}
