% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_waterfall.R
\name{g_waterfall}
\alias{g_waterfall}
\title{Horizontal Waterfall Plot}
\usage{
g_waterfall(height, id, col = NULL, xlab = NULL, ylab = NULL,
  col.legend.title = NULL, title = NULL)
}
\arguments{
\item{height}{numeric vector to be plotted as the waterfall bars}

\item{id}{vector of IDs used as the x-axis label for the waterall bars}

\item{col}{vector of a categorical variable for bar coloring}

\item{xlab}{x label. Default is \code{ID}.}

\item{ylab}{y label. Default is \code{Value}.}

\item{col.legend.title}{A string to be displayed as legend title.}

\item{title}{A string to be displayed as plot title.}
}
\description{
This basic waterfall plot visualizes a quantity \code{height} ordered by value with some
markup
}
\examples{
g_waterfall(height = c(3,5,-1), id = letters[1:3])

g_waterfall(height = c(3,5,-1), id = letters[1:3], col = c("red", "green", "red"))

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(seed = 1)
ADSL_f <- ADSL \%>\%
  select(USUBJID, STUDYID, ARM, ARMCD, SEX)
  
ADRS <- radrs(ADSL, seed = 2)
ADRS_f <- subset(ADRS, PARAMCD == "OVRINV") \%>\% 
   mutate(pchg = rnorm(1200, 10, 50))

ADRS_f <- head(ADRS_f, 30)
ADRS_f <- ADRS_f[!duplicated(ADRS_f$USUBJID),]
head(ADRS_f)

g_waterfall(
  height = ADRS_f$pchg,
  id = ADRS_f$USUBJID,
  col = ADRS_f$AVALC
)

g_waterfall(
  height = ADRS_f$pchg,
  id = ADRS_f$USUBJID,
  col = ADRS_f$AVALC
)

g_waterfall(
  height = ADRS_f$pchg,
  id = paste("asdfdsfdsfsd",ADRS_f$USUBJID),
  col = ADRS_f$SEX
)

g_waterfall(
  height = ADRS_f$pchg,
  id = paste("asdfdsfdsfsd",ADRS_f$USUBJID),
  xlab = "ID",
  ylab = "Percentage Change",
  title = "Waterfall plot"
)

}
\author{
Yuyao Song (songy24) \email{yuyao.song@roche.com}
}
