% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_waterfall.R
\name{g_waterfall}
\alias{g_waterfall}
\title{Horizontal Waterfall Plot}
\usage{
g_waterfall(
  height,
  id,
  col_var = NULL,
  col = getOption("tern.color"),
  xlab = NULL,
  ylab = NULL,
  col_legend_title = NULL,
  title = NULL
)
}
\arguments{
\item{height}{(\code{numeric} vector)\cr
Contains values to be plotted as the waterfall bars}

\item{id}{(vector)\cr
Contains of IDs used as the x-axis label for the waterfall bars}

\item{col_var}{(vector)\cr
Categorical variable for bar coloring}

\item{col}{(\code{character})\cr colors.}

\item{xlab}{(\code{character} value)\cr
x label. Default is \code{ID}.}

\item{ylab}{(\code{character} value)\cr
y label. Default is \code{Value}.}

\item{col_legend_title}{(\code{character} value)\cr
Text to be displayed as legend title.}

\item{title}{(\code{character} value)\cr
Text to be displayed as plot title.}
}
\value{
(\code{ggplot} object)\cr
Waterfall plot
}
\description{
This basic waterfall plot visualizes a quantity \code{height} ordered by value with some
markup.
}
\examples{
g_waterfall(height = c(3, 5, -1), id = letters[1:3])

g_waterfall(
  height = c(3, 5, -1),
  id = letters[1:3],
  col = letters[1:3]
)

library(scda)
library(dplyr)

ADSL <- synthetic_cdisc_data("latest")$adsl
ADSL_f <- ADSL \%>\%
  select(USUBJID, STUDYID, ARM, ARMCD, SEX)

ADRS <- synthetic_cdisc_data("latest")$adrs
ADRS_f <- ADRS \%>\%
  filter(PARAMCD == "OVRINV") \%>\%
  mutate(pchg = rnorm(n(), 10, 50))

ADRS_f <- head(ADRS_f, 30)
ADRS_f <- ADRS_f[!duplicated(ADRS_f$USUBJID), ]
head(ADRS_f)

g_waterfall(
  height = ADRS_f$pchg,
  id = ADRS_f$USUBJID,
  col_var = ADRS_f$AVALC
)

g_waterfall(
  height = ADRS_f$pchg,
  id = ADRS_f$USUBJID,
  col_var = ADRS_f$AVALC
)

g_waterfall(
  height = ADRS_f$pchg,
  id = paste("asdfdsfdsfsd", ADRS_f$USUBJID),
  col_var = ADRS_f$SEX
)

g_waterfall(
  height = ADRS_f$pchg,
  id = paste("asdfdsfdsfsd", ADRS_f$USUBJID),
  xlab = "ID",
  ylab = "Percentage Change",
  title = "Waterfall plot"
)
}
\author{
Yuyao Song (songy) \email{yuyao.song@roche.com}
}
