% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_wf.R
\name{g_wf}
\alias{g_wf}
\title{Horizontal Waterfall Plot}
\usage{
g_wf(height, id, col = NULL, xlab = NULL, ylab = NULL,
  col.legend.title = NULL, plot.title = NULL)
}
\arguments{
\item{height}{numeric vector to be plotted as the waterfall bars}

\item{id}{vector of IDs used as the x-axis label for the waterall bars}

\item{col}{vector of a categorical variable for bar coloring}

\item{xlab}{x label. Default is \code{ID}.}

\item{ylab}{y label. Default is \code{Value}.}

\item{col.legend.title}{A string to be displayed as legend title.}

\item{plot.title}{A string to be displayed as plot title.}
}
\description{
This basic waterfall plot visualizes a quatity \code{height} ordered by value with some
markup
}
\examples{


g_wf(height = c(3,5,-1), id = letters[1:3])

g_wf(height = c(3,5,-1), id = letters[1:3], col = c("red", "green", "red"))


library(random.cdisc.data)
library(dplyr)

ASL <- radam("ASL")
ARS <- radam("ARS", ADSL = ASL)

ARS_f <- subset(ARS, PARAMCD == "OVRINV")

ASL_f <- ASL \%>\%
  select(USUBJID, STUDYID, ARM, ARMCD, SEX)

apchg <- ARS_f \%>\%
  mutate(pchg = rnorm(100, 10, 50))
  # Merge pchange and besr response
asld <- right_join(apchg, ASL_f \%>\% select(STUDYID, USUBJID, SEX, ARMCD))


asld <- head(asld, 30)

head(asld)

g_wf(
  height = asld$pchg,
  id = asld$USUBJID,
  col = asld$AVALC
)

g_wf(
  height = asld$pchg,
  id = paste("asdfdsfdsfsd",asld$USUBJID),
  col = asld$SEX
)

g_wf(
  height = asld$pchg,
  id = paste("asdfdsfdsfsd",asld$USUBJID),
  xlab = "ID",
  ylab = "Percentage Change",
  plot.title = "Waterfall plot"
)

}
\author{
Yuyao Song (songy24) \email{yuyao.song@roche.com}
}
