% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_wf.R
\name{g_wf}
\alias{g_wf}
\title{Horizontal Waterfall Plot}
\usage{
g_wf(y, id = 1:length(y), col = "steelblue", xlab = "ID",
  ylab = "Value", legend.title = NULL)
}
\arguments{
\item{y}{variable name of the numeric vector to be plotted from \code{data}}

\item{xlab}{x label. Default is \code{ID}.}

\item{ylab}{y label. Default is \code{Value}.}

\item{legend.title}{A string to be displayed as legend title.}

\item{data}{data frame}

\item{id_var}{ID variable name used as the x-axis label for each bar from \code{data}}

\item{col_var}{name of the categorical variable for bar coloring and legend from \code{data}}

\item{title}{A string to be displayed as plot title.}

\item{add_id}{variables to be added to the summary table.}
}
\description{
The waterfall plot visualizes a quatity \code{y} ordered by value with some
markup
}
\examples{

library(random.cdisc.data)
library(dplyr)

ASL <- radam("ASL")
ARS <- radam("ARS", ADSL = ASL)

ARS_f <- subset(ARS, PARAMCD == "OVRINV")

ASL_f <- ASL \%>\%
  select(USUBJID, STUDYID, ARM, ARMCD, SEX)

apchg <- ARS_f \%>\%
  mutate(pchg = rnorm(100, 10, 50))
  # Merge pchange and besr response
asld <- right_join(apchg, ASL_f \%>\% select(STUDYID, USUBJID, SEX, ARMCD))


asld <- head(asld, 30)

head(asld)

with(asld,
  g_wf(
    y = asld$pchg,
    id = asld$USUBJID,
    col = asld$AVALC,
    
  )
)



}
\author{
Yuyao Song (songy24) \email{yuyao.song@roche.com}
}
