% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mmrm.R
\name{get_mmrm_lsmeans}
\alias{get_mmrm_lsmeans}
\title{Extract LS Means from MMRM}
\usage{
get_mmrm_lsmeans(fit, vars, conf_level, weights)
}
\arguments{
\item{fit}{result of \code{\link[=fit_lme4]{fit_lme4()}}.}

\item{vars}{(named \code{list} of \code{string} or \code{character})\cr specifying the variables in the MMRM.
The following elements need to be included as character vectors and match corresponding columns
in \code{data}:
\describe{
\item{response}{the response variable}
\item{covariates}{the additional covariate terms (might also include interactions)}
\item{id}{the subject ID variable}
\item{arm}{the treatment group variable (factor)}
\item{visit}{the visit variable (factor)}
}
Note that the main effects and interaction of \code{arm} and \code{visit} are by default included in the model.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{weights}{(\code{string})\cr type of weights to be used for the LS means,
see \code{\link[emmeans:emmeans]{emmeans::emmeans()}} for details.}
}
\value{
A list with the LS means \code{estimates} and \code{contrasts} results between the treatment
and control arm groups at the different visits.
}
\description{
Helper function to extract the LS means from an MMRM fit.
}
