% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pool_object.R
\name{get_pool_object}
\alias{get_pool_object}
\title{Get cached \code{rbmi} S3 \code{pool} object}
\usage{
get_pool_object()
}
\description{
Get cached \code{rbmi} S3 \code{pool} object
}
\examples{
get_pool_object()
}
