% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_rsp.R
\name{glm_subgroup}
\alias{glm_subgroup}
\title{alternative tablist}
\usage{
glm_subgroup(response, event, arm, arm.ref, arm.comp = setdiff(arm, arm.ref),
  group_by, covariARSs = NULL)
}
\arguments{
\item{event}{is boolean, \code{TRUE} if event, \code{FALSE} if time_to_event
is censored}

\item{arm}{vector with arm information}

\item{arm.ref}{a character vector defining which arms in arm should be taken 
as the reference}

\item{arm.comp}{a character vector defining which arms in arm should be taken 
as the comparison}

\item{group_by}{data frame with one column per grouping}

\item{time_to_event}{Response data}
}
\description{
alternative tablist
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr) 
library(grid)
'\%needs\%' <- teal.oncology:::'\%needs\%'
ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

surv_tbl_stream <- get_forest_response_table(com.roche.cdt30019.go29436.re)
Viewer(surv_tbl_stream )

ARS_f <- ARS \%>\% filter(PARAMCD == "OVRSPI") \%>\% filter(ITTWTFL == "Y") \%>\% filter(ARM \%in\% c("DUMMY A", "DUMMY C"))
ASL_f <- ASL \%>\% filter(ITTWTFL == "Y") \%>\% filter(ARM \%in\% c("DUMMY A", "DUMMY C"))

ARS_f_resp <- ARS_f \%>\% filter(AVALC \%in\% c("CR", "PR"))
ARS_f_IC1 <- ARS_f \%>\% filter(TCLEVEL == "1")

group_by <- merge(
 ARS_f[c("USUBJID", "STUDYID")],
 ASL_f[c("USUBJID", "STUDYID", "TCLEVEL", "ICLEVEL", "TC3IC3")],
 all.y = TRUE, all.x = FALSE
)

head(group_by)

tbl <- glm_subgroup(
          response = ARS_f$AVAL,
          event = ARS_f$AVALC \%in\% c("CR","PR"),
          arm = ARS_f$ARM, 
          group_by = group_by[, -c(1,2), drop=FALSE],
          arm.ref = "DUMMY C",
          arm.comp = "DUMMY A"
)
Viewer(tbl)

compare_rtables(tbl, surv_tbl_stream, comp.attr = FALSE)
}


  
}
