% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{h_coxreg_univ_extract}
\alias{h_coxreg_univ_extract}
\title{Tabulation of Univariate Cox Regressions}
\usage{
h_coxreg_univ_extract(
  mod,
  var,
  data,
  effect = "armcd",
  control = control_coxreg()
)
}
\arguments{
\item{mod}{model}

\item{var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{data}{(\verb{data frame})\cr the dataset containing the variables to fit the
models.}

\item{effect}{(\code{string})\cr the treatment variable.}

\item{control}{(\code{list})\cr a list of parameters as returned by the
helper function \code{\link[=control_coxreg]{control_coxreg()}}.}
}
\description{
Tabulate the result of a Cox regression model.
}
\examples{

# Example, see `?survival::bladder`.
library(survival)
dta_test <- with(
  data = bladder[bladder$enum < 5, ],
  data.frame(
    time = stop,
    status = event,
    armcd = as.factor(rx),
    covar1 = as.factor(enum)
  )
)
mod <- coxph(Surv(time, status) ~ armcd + covar1, data = dta_test)
h_coxreg_univ_extract(mod, var = "covar1", data = dta_test)

# Example, see `?survival::aml`.
plot(
  survfit(Surv(time, status) ~ x, data = aml),
  lty=2:3, xlab="Months"
)
mod <- coxph(Surv(time, status) ~ x, data = aml)
h_coxreg_univ_extract(mod, effect = "x", var = "x", data = aml)

}
