% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{h_coxreg_univ_formulas}
\alias{h_coxreg_univ_formulas}
\title{Helper for Cox Regression Formula}
\usage{
h_coxreg_univ_formulas(variables, interaction = FALSE)
}
\arguments{
\item{variables}{(named \code{list} of \code{strings})\cr list of additional analysis variables.}

\item{interaction}{(\code{flag})\cr if \code{TRUE}, the model includes the
interaction between the studied treatment and candidate covariate.}
}
\value{
It \code{character} vector coercible into formulas (e.g \code{\link[=as.formula]{as.formula()}}).
}
\description{
Creates a list of formulas. It is used internally by \code{\link[=h_coxreg]{h_coxreg()}}
for the comparison of univariate Cox regression models.
}
\examples{

# `h_coxreg_univ_formulas`

## Simple formulas.
h_coxreg_univ_formulas(
  variables = list(
    time = "time", event = "status", arm = "armcd", covariates = c("X", "y")
  )
)

## Addition of an optional strata.
h_coxreg_univ_formulas(
  variables = list(
    time = "time", event = "status", arm = "armcd", covariates = c("X", "y"),
    strata = "SITE"
  )
)

## Inclusion of the interaction term.
h_coxreg_univ_formulas(
  variables = list(
    time = "time", event = "status", arm = "armcd", covariates = c("X", "y"),
    strata = "SITE"
  ),
  interaction = TRUE
)

}
