% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_plot.R
\name{h_data_plot}
\alias{h_data_plot}
\title{Helper function: tidy survival fit}
\usage{
h_data_plot(fit_km, max_time = NULL)
}
\arguments{
\item{fit_km}{(\code{survfit})\cr result of \code{\link[survival:survfit]{survival::survfit()}}.}

\item{max_time}{(\code{numeric})\cr maximum value of X axis range (\code{NULL} for
default)}
}
\description{
Convert the survival fit data into a data frame designed for plotting
within \code{g_km}.
}
\examples{

\dontrun{
library(random.cdisc.data)
library(dplyr)
library(survival)

radtte(cached = TRUE) \%>\%
  filter(PARAMCD == "OS") \%>\%
  survfit(form = Surv(AVAL, 1 - CNSR) ~ ARMCD, data = .) \%>\%
  tern:::h_data_plot()
}

}
