% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_plot.R
\name{h_data_plot}
\alias{h_data_plot}
\title{Helper function: tidy survival fit}
\usage{
h_data_plot(fit_km, armval = "All", max_time = NULL)
}
\arguments{
\item{fit_km}{(\code{survfit})\cr result of \code{\link[survival:survfit]{survival::survfit()}}.}

\item{armval}{(\code{string}) \cr used as strata name when treatment arm
variable only has one level. Default is "All".}

\item{max_time}{(\code{numeric})\cr maximum value to show on X axis.
Only data values less than or up to to this threshold value will be plotted.(\code{NULL} for
default)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This starts from the \code{\link[broom:reexports]{broom::tidy()}} result, and then:
\itemize{
\item post-processes the \code{strata} column into a factor,
\item extends each stratum by an additional first row with time 0
and probability 1 so that downstream plot lines start at those coordinates,
\item adds a \code{censor} column,
\item filters the rows before \code{max_time}.
}
}
\details{
Convert the survival fit data into a data frame designed for plotting
within \code{g_km}.
}
\examples{
\dontrun{
library(scda)
library(dplyr)
library(survival)

# Test with multiple arms
synthetic_cdisc_data("latest")$adtte \%>\%
  filter(PARAMCD == "OS") \%>\%
  survfit(form = Surv(AVAL, 1 - CNSR) ~ ARMCD, data = .) \%>\%
  h_data_plot()

# Test with single arm
synthetic_cdisc_data("latest")$adtte \%>\%
  filter(PARAMCD == "OS", ARMCD == "ARM B") \%>\%
  survfit(form = Surv(AVAL, 1 - CNSR) ~ ARMCD, data = .) \%>\%
  h_data_plot(armval = "ARM B")
}

}
