% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_plot.R
\name{h_decompose_gg}
\alias{h_decompose_gg}
\title{\code{ggplot} Decomposition}
\usage{
h_decompose_gg(gg)
}
\arguments{
\item{gg}{(\code{ggplot})\cr a graphic to decompose.}
}
\description{
The elements composing the \code{ggplot} are extracted and organized in a
list containing:
the panel (\code{panel}),
the y-axis and its label (\code{yaxis}, \code{ylab}),
idem for the x-axis (\code{xaxis}, \code{xlab}),
the legend (\code{guide}).
}
\examples{
\dontrun{
library(scda)
library(dplyr)
library(survival)
library(grid)

fit_km <- synthetic_cdisc_data("latest")$adtte \%>\%
  filter(PARAMCD == "OS") \%>\%
  survfit(form = Surv(AVAL, 1 - CNSR) ~ ARMCD, data = .)
data_plot <- h_data_plot(fit_km = fit_km)
xticks <- h_xticks(data = data_plot)
gg <- h_ggkm(
  data = data_plot,
  yval = "Survival",
  censor_show = TRUE,
  xticks = xticks, xlab = "Days", ylab = "Survival Probability",
  title = "tt"
)

g_el <- h_decompose_gg(gg)
grid::grid.newpage()
grid.rect(gp = grid::gpar(lty = 1, col = "red", fill = "gray85", lwd = 5))
grid::grid.draw(g_el$panel)

grid::grid.newpage()
grid.rect(gp = grid::gpar(lty = 1, col = "royalblue", fill = "gray85", lwd = 5))
grid::grid.draw(with(g_el, cbind(ylab, yaxis)))
}

}
\keyword{internal}
