% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_diff.R
\name{h_find_ci_bound_uniroot}
\alias{h_find_ci_bound_uniroot}
\title{Root-finding CI bounds from one-sided p-value functions}
\usage{
h_find_ci_bound_uniroot(
  p_value_function,
  cutoff,
  direction = c("increasing", "decreasing"),
  interval = c(-1, 1),
  tol = 1e-06,
  maxiter = 1000
)
}
\arguments{
\item{p_value_function}{(\code{function}) one-sided p-value function in terms of \code{d}.}

\item{cutoff}{(\code{number}) one-sided significance level threshold.}

\item{direction}{(\code{string}) one of \code{"increasing"} or \code{"decreasing"}.}

\item{interval}{(\code{numeric}) 2-element search interval.}

\item{tol}{(\code{number}) tolerance for \code{\link[stats:uniroot]{stats::uniroot()}}.}

\item{maxiter}{(\code{integer}) maximum number of \code{\link[stats:uniroot]{stats::uniroot()}} iterations.}
}
\value{
A \code{number} with the requested CI bound, or \code{NA_real_} if no bound exists.
}
\description{
This function is an internal helper for \code{\link[=prop_diff_uncond_exact]{prop_diff_uncond_exact()}}.
}
\keyword{internal}
