% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_patient_plot.R
\name{h_g_ipp}
\alias{h_g_ipp}
\title{Helper Function To Create Simple Line Plot over Time}
\usage{
h_g_ipp(
  df,
  xvar,
  yvar,
  xlab,
  ylab,
  id_var,
  title = "Individual Patient Plots",
  subtitle = "",
  caption = NULL,
  add_baseline_hline = FALSE,
  yvar_baseline = "BASE",
  ggtheme = h_set_nest_theme(10),
  col = getOption("tern.color")
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Function that generates a simple line plot displaying parameter trends over time.
}
\examples{
library(scda)
library(dplyr)

# Select a small sample of data to plot.
adlb <- synthetic_cdisc_data("latest")$adlb \%>\%
  filter(PARAMCD == "ALT", !(AVISIT \%in\% c("SCREENING", "BASELINE"))) \%>\%
  slice(1:36)

p <- h_g_ipp(
  df = adlb,
  xvar = "AVISIT",
  yvar = "AVAL",
  xlab = "Visit",
  id_var = "USUBJID",
  ylab = "SGOT/ALT (U/L)",
  add_baseline_hline = TRUE
)
p

}
