% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_plot.R
\name{h_ggkm}
\alias{h_ggkm}
\title{Helper function: KM plot}
\usage{
h_ggkm(
  data,
  xticks = NULL,
  yval,
  censor_show,
  xlab,
  ylab,
  title,
  lwd = 1,
  lty = NULL,
  pch = "|",
  size = 3,
  col = NULL,
  ggtheme = NULL
)
}
\arguments{
\item{data}{(\code{data.frame})\cr survival data as pre-processed by \code{h_data_plot}.}

\item{xticks}{(\code{numeric}, \code{number}, or \code{NULL})\cr
numeric vector of ticks or single number with spacing
between ticks on the x axis.
If \code{NULL} (default), \code{\link[labeling:extended]{labeling::extended()}} is used to determine
an optimal tick position on the x axis.}

\item{yval}{(\code{string})\cr value of y-axis. Should be either
\code{Survival} (default) or \code{Failure} probability.}

\item{censor_show}{(\code{flag})\cr whether to show censored.}

\item{xlab}{(\code{string})\cr label of x-axis.}

\item{ylab}{(\code{string})\cr label of y-axis.}

\item{title}{(\code{string})\cr title for plot.}

\item{lwd}{(\code{numeric})\cr line width. Length of a vector should be equal
to number of strata from \code{\link[survival:survfit]{survival::survfit()}}.}

\item{lty}{(\code{numeric})\cr line type. Length of a vector should be equal
to number of strata from \code{\link[survival:survfit]{survival::survfit()}}.}

\item{pch}{(\code{numeric})\cr point type for censored.}

\item{size}{(\code{numeric})\cr size of censored point, a class of \code{unit}.}

\item{col}{(\code{character})\cr lines colors. Length of a vector should be equal
to number of strata from \code{\link[survival:survfit]{survival::survfit()}}.}

\item{ggtheme}{(\code{theme})\cr a graphical theme as provided by \code{ggplot2} to
control outlook of the Kaplan-Meier curve.}
}
\description{
Draw the Kaplan-Meier plot using \code{ggplot2}.
}
\examples{
\dontrun{

library(random.cdisc.data)
library(dplyr)
library(survival)

fit_km <- radtte(cached = TRUE) \%>\%
  filter(PARAMCD == "OS") \%>\%
  survfit(form = Surv(AVAL, 1 - CNSR) ~ ARMCD, data = .)
data_plot <- h_data_plot(fit_km = fit_km)
xticks <- h_xticks(data = data_plot)
gg <- h_ggkm(
  data = data_plot,
  censor_show = TRUE,
  xticks = xticks,
  xlab = "Days",
  yval = "Survival",
  ylab = "Survival Probability",
  title = "Survival"
)

gg
}

}
