% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_plot.R
\name{h_grob_coxph}
\alias{h_grob_coxph}
\title{Helper Function: CoxPH Grob
Grob of rtable output from \link{h_tbl_coxph_pairwise}}
\usage{
h_grob_coxph(
  ...,
  x = 0,
  y = 0,
  ttheme = gridExtra::ttheme_default(base_size = 12, padding = grid::unit(c(1, 0.5),
    "lines"), core = list(bg_params = list(fill = c("grey95", "grey90"), alpha = 0.5)))
)
}
\arguments{
\item{...}{arguments will be passed to \code{\link[=h_tbl_coxph_pairwise]{h_tbl_coxph_pairwise()}}.}

\item{x}{a \code{numeric} value between 0 and 1 specifying x-location.}

\item{y}{a \code{numeric} value between 0 and 1 specifying y-location.}

\item{ttheme}{(\code{list})\cr see \code{\link[gridExtra:tableGrob]{gridExtra::ttheme_default()}}.}
}
\description{
Helper Function: CoxPH Grob
Grob of rtable output from \link{h_tbl_coxph_pairwise}
}
\examples{
\dontrun{
library(scda)
library(dplyr)
library(survival)
library(grid)

grid::grid.newpage()
grid.rect(gp = grid::gpar(lty = 1, col = "pink", fill = "gray85", lwd = 1))
data <- synthetic_cdisc_data("latest")$adtte \%>\%
  filter(PARAMCD == "OS") \%>\%
  mutate(is_event = CNSR == 0)
tbl_grob <- tern:::h_grob_coxph(
   df = data,
   variables = list(tte = "AVAL", is_event = "is_event", arm = "ARMCD"),
   control_coxph_pw = control_coxph(conf_level = 0.9), x = 0.5, y = 0.5)
grid::grid.draw(tbl_grob)
}

}
