% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_plot.R
\name{h_grob_median_surv}
\alias{h_grob_median_surv}
\title{Helper Function: Survival Estimation Grob}
\usage{
h_grob_median_surv(fit_km, ttheme = gridExtra::ttheme_default())
}
\arguments{
\item{ttheme}{(\code{list})\cr see \code{\link[gridExtra:tableGrob]{gridExtra::ttheme_default()}}.}
}
\description{
The survival fit is transformed in a grob containing a table with groups in
rows characterized by N, median and 95\% confidence interval.
}
\examples{
\dontrun{
library(random.cdisc.data)
library(dplyr)
library(survival)

grid.newpage()
grid.rect(gp = gpar(lty = 1, col = "pink", fill = "gray85", lwd = 1))
radtte(cached = TRUE) \%>\%
  filter(PARAMCD == "OS") \%>\%
  survfit(form = Surv(AVAL, 1 - CNSR) ~ ARMCD, data = .) \%>\%
  h_grob_median_surv \%>\%
  grid.draw
}

}
