% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_plot.R
\name{h_grob_median_surv}
\alias{h_grob_median_surv}
\title{Helper Function: Survival Estimation Grob}
\usage{
h_grob_median_surv(
  fit_km,
  armval = "All",
  x = 0.9,
  y = 0.9,
  ttheme = gridExtra::ttheme_default()
)
}
\arguments{
\item{fit_km}{(\code{survfit})\cr result of \code{\link[survival:survfit]{survival::survfit()}}.}

\item{armval}{(\code{string}) \cr used as strata name when treatment arm
variable only has one level. Default is "All".}

\item{x}{a \code{numeric} value between 0 and 1 specifying x-location.}

\item{y}{a \code{numeric} value between 0 and 1 specifying y-location.}

\item{ttheme}{(\code{list})\cr see \code{\link[gridExtra:tableGrob]{gridExtra::ttheme_default()}}.}
}
\description{
The survival fit is transformed in a grob containing a table with groups in
rows characterized by N, median and 95\% confidence interval.
}
\examples{
\dontrun{
library(scda)
library(dplyr)
library(survival)
library(grid)

grid::grid.newpage()
grid.rect(gp = grid::gpar(lty = 1, col = "pink", fill = "gray85", lwd = 1))
synthetic_cdisc_data("latest")$adtte \%>\%
  filter(PARAMCD == "OS") \%>\%
  survfit(form = Surv(AVAL, 1 - CNSR) ~ ARMCD, data = .) \%>\%
  h_grob_median_surv() \%>\%
  grid::grid.draw()
}

}
