% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_diff.R
\name{h_miettinen_nurminen_var}
\alias{h_miettinen_nurminen_var}
\title{Variance Estimate Following Miettinen and Nurminen}
\usage{
h_miettinen_nurminen_var(est1, est2, x1, x2, n1, n2)
}
\arguments{
\item{est1}{(\code{numeric(1)}) \cr
Estimated proportion for group 1. Used together with \code{est2} to define the
risk difference.
May be \code{NA_real_} when an estimate cannot be calculated.}

\item{est2}{(\code{numeric(1)}) \cr
Estimated proportion for group 2. Used together with \code{est1} to define the
risk difference.
May be \code{NA_real_} when an estimate cannot be calculated.}

\item{x1}{(\code{numeric}) \cr
Number of responders in group 1 for each stratum.
Must have length at least 1.}

\item{x2}{(\code{numeric}) \cr
Number of responders in group 2 for each stratum.
Must have the same length as \code{x1}.}

\item{n1}{(\code{numeric}) \cr
Number of observations in group 1 for each stratum.
Must have the same length as \code{x1}.}

\item{n2}{(\code{numeric}) \cr
Number of observations in group 2 for each stratum.
Must have the same length as \code{x1}.}
}
\value{
A named \code{list} with elements:
\itemize{
\item \code{p1_est}: constrained maximum likelihood estimate of the proportion in
group 1 for each stratum.
\item \code{p2_est}: constrained maximum likelihood estimate of the proportion in
group 2 for each stratum.
\item \code{var_est}: Miettinen-Nurminen variance estimate for each stratum.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Calculates the \insertCite{MiettinenNurminen1985;textual}{tern} variance
estimate for the difference between two proportions. The estimate is based on
the constrained maximum likelihood estimates of the two proportions under the
specified risk difference and is used to obtain the standard error for the
Miettinen-Nurminen confidence interval.
}
\details{
The risk difference is defined as \code{est2} - \code{est1}. For each stratum, the
function calculates the constrained maximum likelihood estimate for the
proportion in group 1 and obtains the corresponding estimate for group 2
by adding the risk difference. The variance is then calculated from these
estimates using the Miettinen-Nurminen variance formula.

The variance is returned as \code{NA_real_} for strata where the variance cannot
be calculated.

The variable names in this function follow the notation in the original
paper by \insertCite{MiettinenNurminen1985;textual}{tern}, cf. Appendix 1.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=prop_diff_cmh]{prop_diff_cmh()}}, \code{\link[=h_prop_cmh]{h_prop_cmh()}}, \code{\link[=h_miettinen_nurminen_stratified_ci]{h_miettinen_nurminen_stratified_ci()}}
}
\keyword{internal}
