% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_prop_diff.R
\name{h_prop_diff}
\alias{h_prop_diff}
\alias{prop_diff_wald}
\alias{prop_diff_ha}
\alias{prop_diff_nc}
\alias{prop_diff_cmh}
\title{Helper Functions to Calculate Proportion Difference}
\usage{
prop_diff_wald(rsp, grp, conf_level, correct)

prop_diff_ha(rsp, grp, conf_level)

prop_diff_nc(rsp, grp, conf_level, correct = FALSE)

prop_diff_cmh(rsp, grp, strata, conf_level = 0.95)
}
\arguments{
\item{rsp}{(\code{logical})\cr whether each subject is a responder or not.}

\item{grp}{(\code{factor})\cr
vector assigning observations to one out of two groups
(e.g. reference and treatment group).}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{correct}{\code{logical}\cr
include the continuity correction.}

\item{strata}{(\code{factor})\cr
with one level per stratum and same length as \code{rsp}.}
}
\description{
Helper Functions to Calculate Proportion Difference
}
\section{Functions}{
\itemize{
\item \code{prop_diff_wald}: The Wald interval follows the usual textbook
definition for a single proportion confidence interval using the normal
approximation. It is possible to include a continuity correction for Wald's
interval.

\item \code{prop_diff_ha}: Anderson-Hauck confidence interval.

\item \code{prop_diff_nc}: Newcombe confidence interval. It is based on
the Wilson score confidence interval for a single binomial proportion.

\item \code{prop_diff_cmh}: Calculates the weighted difference.
This is defined as the difference in response rates between the
experimental treatment group and the control treatment group, adjusted
for stratification factors by applying Cochran-Mantel-Haenszel (CMH)
weights. For the CMH chi-squared test, use \code{\link[stats:mantelhaen.test]{stats::mantelhaen.test()}}.
}}

\examples{
# Wald confidence interval
set.seed(2)
rsp <- sample(c(TRUE, FALSE), replace = TRUE, size = 20)
grp <- c(rep("A", 10), rep("B", 10))
tern:::prop_diff_wald(rsp = rsp, grp = grp, conf_level = 0.90, correct = FALSE)

# Anderson-Hauck confidence interval
## "Mid" case: 3/4 respond in group A, 1/2 respond in group B.
rsp <- c(TRUE, FALSE, FALSE, TRUE, TRUE, TRUE)
grp <- factor(c("A", "B", "A", "B", "A", "A"), levels = c("B", "A"))
tern:::prop_diff_ha(rsp = rsp, grp = grp, conf_level = 0.90)

## Edge case: Same proportion of response in A and B.
rsp <- c(TRUE, FALSE, TRUE, FALSE)
grp <- factor(c("A", "A", "B", "B"), levels = c("A", "B"))
tern:::prop_diff_ha(rsp = rsp, grp = grp, conf_level = 0.6)

# Newcombe confidence interval

set.seed(1)
rsp <- c(
  sample(c(TRUE, FALSE), size = 40, prob = c(3 / 4, 1 / 4), replace = TRUE),
  sample(c(TRUE, FALSE), size = 40, prob = c(1 / 2, 1 / 2), replace = TRUE)
)
grp <- factor(rep(c("A", "B"), each = 40), levels = c("B", "A"))
table(rsp, grp)
tern:::prop_diff_nc(rsp = rsp, grp = grp, conf_level = 0.9)

# Cochran-Mantel-Haenszel confidence interval

set.seed(2)
rsp <- sample(c(TRUE, FALSE), 100, TRUE)
grp <- sample(c("Placebo", "Treatment"), 100, TRUE)
grp <- factor(grp, levels = c("Placebo", "Treatment"))
strata_data <- data.frame(
  "f1" = sample(c("a", "b"), 100, TRUE),
  "f2" = sample(c("x", "y", "z"), 100, TRUE),
  stringsAsFactors = TRUE
)

tern:::prop_diff_cmh(
  rsp = rsp, grp = grp, strata = interaction(strata_data),
  conf_level = 0.90
)

}
\seealso{
\code{\link[=prop_diff]{prop_diff()}} for implementation of these helper functions.
}
\keyword{internal}
