% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_diff_test.R
\name{h_prop_diff_test}
\alias{h_prop_diff_test}
\alias{prop_chisq}
\alias{prop_cmh}
\alias{prop_schouten}
\alias{prop_fisher}
\title{Helper Functions to Test Proportion Differences}
\usage{
prop_chisq(tbl)

prop_cmh(ary)

prop_schouten(tbl)

prop_fisher(tbl)
}
\arguments{
\item{tbl}{(\code{matrix})\cr
with two groups in rows and the binary response (\code{TRUE}/\code{FALSE}) in
columns.}

\item{ary}{(\code{array}, 3 dimensions)\cr
with two groups in rows, the binary response (\code{TRUE}/\code{FALSE}) in
columns, the strata in the third dimension.}
}
\description{
Helper functions to implement various tests on the difference between two
proportions.
}
\section{Functions}{
\itemize{
\item \code{prop_chisq()}: performs Chi-Squared test.
Internally calls \code{\link[stats:prop.test]{stats::prop.test()}}.

\item \code{prop_cmh()}: performs stratified Cochran-Mantel-Haenszel test.
Internally calls \code{\link[stats:mantelhaen.test]{stats::mantelhaen.test()}}. Note that strata with less than two observations
are automatically discarded.

\item \code{prop_schouten()}: performs the Chi-Squared test with Schouten
correction.

\item \code{prop_fisher()}: performs the Fisher's exact test.
Internally calls \code{\link[stats:fisher.test]{stats::fisher.test()}}.

}}
\examples{
# Non-stratified proportion difference test

## Data
A <- 20
B <- 20
set.seed(1)
rsp <- c(
  sample(c(TRUE, FALSE), size = A, prob = c(3 / 4, 1 / 4), replace = TRUE),
  sample(c(TRUE, FALSE), size = A, prob = c(1 / 2, 1 / 2), replace = TRUE)
)
grp <- c(rep("A", A), rep("B", B))
tbl <- table(grp, rsp)

## Chi-Squared test
# Internal function - prop_chisq
\dontrun{
prop_chisq(tbl)
}

# Stratified proportion difference test

## Data
rsp <- sample(c(TRUE, FALSE), 100, TRUE)
grp <- factor(rep(c("A", "B"), each = 50))
strata <- factor(rep(c("V", "W", "X", "Y", "Z"), each = 20))
tbl <- table(grp, rsp, strata)

## Cochran-Mantel-Haenszel test
# Internal function - prop_cmh
\dontrun{
prop_cmh(tbl)
}

## Chi-Squared test + Schouten correction.
# Internal function - prop_schouten
\dontrun{
prop_schouten(tbl)
}

## Fisher's exact test
# Internal function - prop_fisher
\dontrun{
prop_fisher(tbl)
}

}
\seealso{
[prop_diff_test())] for implementation of these helper functions.

[prop_diff_test())]: R:prop_diff_test())

For information on the Schouten correction (Schouten, 1980),
visit https://onlinelibrary.wiley.com/doi/abs/10.1002/bimj.4710220305.
}
\keyword{internal}
