% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desctool_binom_diff.R
\name{h_recycle}
\alias{h_recycle}
\title{Recycle list of parameters}
\usage{
h_recycle(...)
}
\description{
This function recycles all supplied elements to the maximal dimension.
}
\keyword{internal}
