% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_subgroups.R
\name{h_response_subgroups}
\alias{h_response_subgroups}
\alias{h_proportion_df}
\alias{h_proportion_subgroups_df}
\alias{h_odds_ratio_df}
\alias{h_odds_ratio_subgroups_df}
\title{Helper Functions for Tabulating Binary Response by Subgroup}
\usage{
h_proportion_df(rsp, arm)

h_proportion_subgroups_df(variables, data, label_all = "All Patients")

h_odds_ratio_df(rsp, arm, conf_level = 0.95, method = NULL)

h_odds_ratio_subgroups_df(
  variables,
  data,
  conf_level = 0.95,
  method = NULL,
  label_all = "All Patients"
)
}
\arguments{
\item{rsp}{(\code{logical})\cr whether each subject is a responder or not.}

\item{arm}{(\code{factor})\cr the treatment group variable.}

\item{variables}{(named \code{list} of \code{string})\cr list of additional analysis variables.}

\item{data}{(\verb{data frame})\cr the dataset containing the variables to summarize.}

\item{label_all}{(\code{string})\cr label for the total population analysis.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{method}{(\code{string})\cr
specifies the test used to calculate the p-value for the difference between
two proportions. For options, see \code{\link[=s_test_proportion_diff]{s_test_proportion_diff()}}. Default is \code{NULL}
so no test is performed.}
}
\description{
Helper functions that tabulate in a data frame statistics such as response rate
and odds ratio for population subgroups.
}
\details{
Main functionality is to prepare data for use in a layout creating function.
}
\section{Functions}{
\itemize{
\item \code{h_proportion_df}: Helper to prepare a data frame of binary responses by arm.

\item \code{h_proportion_subgroups_df}: Summarizes proportion of binary responses by arm and across subgroups
in a data frame. \code{variables} corresponds to the names of variables found in \code{data}, passed as a named list and
requires elements \code{rsp}, \code{arm} and \code{subgroups}.

\item \code{h_odds_ratio_df}: Helper to prepare a data frame with estimates of
the odds ratio between a treatment and a control arm.

\item \code{h_odds_ratio_subgroups_df}: Summarizes estimates of the odds ratio between a treatment and a control
arm across subgroups in a data frame. \code{variables} corresponds to the names of variables found in
\code{data}, passed as a named list and requires elements \code{rsp}, \code{arm} and \code{subgroups}.
}}

\examples{

# Testing dataset.
library(random.cdisc.data)
library(dplyr)

adrs <- radrs(cached = TRUE)
adrs_labels <- var_labels(adrs)

adrs <- adrs \%>\%
  filter(PARAMCD == "BESRSPI") \%>\%
  filter(ARM \%in\% c("A: Drug X", "B: Placebo")) \%>\%
  droplevels() \%>\%
  mutate(
    # Reorder levels of factor to make the placebo group the reference arm.
    ARM = forcats::fct_relevel(ARM, "B: Placebo"),
    rsp = AVALC == "CR"
  )
var_labels(adrs) <- c(adrs_labels, "Response")


h_proportion_df(
  c(TRUE, FALSE, FALSE),
  arm = factor(c("A", "A", "B"), levels = c("A", "B"))
)


h_proportion_subgroups_df(
  variables = list(rsp = "rsp", arm = "ARM", subgroups = c("SEX", "BMRKR2")),
  data = adrs
)


h_odds_ratio_df(
  c(TRUE, FALSE, FALSE, TRUE),
  arm = factor(c("A", "A", "B", "B"), levels = c("A", "B"))
)
# Include p-value.
h_odds_ratio_df(adrs$rsp, adrs$ARM, method = "chisq")


h_odds_ratio_subgroups_df(
  variables = list(rsp = "rsp", arm = "ARM", subgroups = c("SEX", "BMRKR2")),
  data = adrs
)

}
