% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_survival_duration_subgroups.R
\name{h_split_by_subgroups}
\alias{h_split_by_subgroups}
\title{Split Dataframe by Subgroups}
\usage{
h_split_by_subgroups(data, subgroups, groups_lists = list())
}
\arguments{
\item{data}{(\verb{data frame})\cr dataset to split.}

\item{subgroups}{(\code{character})\cr names of factor variables from \code{data} used to create subsets.
Unused levels not present in \code{data} are dropped. Note that the order in this vector
determines the order in the downstream table.}

\item{groups_lists}{(named \code{list} of \code{list})\cr optionally contains for each \code{subgroups} variable a
list, which specifies the new group levels via the names and the
levels that belong to it in the character vectors that are elements of the list.}
}
\value{
A list with subset data (\code{df}) and metadata about the subset (\code{df_labels}).
}
\description{
Split a dataframe into a non-nested list of subsets.
}
\details{
Main functionality is to prepare data for use in forest plot layouts.
}
\examples{

}
\keyword{"A"),}
\keyword{"A",}
\keyword{"B"),}
\keyword{"B",}
\keyword{"C"}
\keyword{"C")}
\keyword{"C"))}
\keyword{"C")),}
\keyword{"C",}
\keyword{"D"),}
\keyword{"D",}
\keyword{"z")}
\keyword{"z"),}
\keyword{)}
\keyword{<-}
\keyword{=}
\keyword{c("A",}
\keyword{c("D",}
\keyword{c("y",}
\keyword{c(1:5),}
\keyword{data}
\keyword{data.frame(}
\keyword{df}
\keyword{df,}
\keyword{factor(c("A",}
\keyword{factor(c("C",}
\keyword{for",}
\keyword{formatters::var_labels(df)}
\keyword{groups_lists}
\keyword{internal}
\keyword{levels}
\keyword{library(rtables)}
\keyword{list(}
\keyword{list("AB"}
\keyword{names(df))}
\keyword{paste("label}
\keyword{subgroups}
\keyword{tern::h_split_by_subgroups(}
\keyword{x}
\keyword{y}
\keyword{z}
