% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_stack_by_baskets.R
\name{h_stack_by_baskets}
\alias{h_stack_by_baskets}
\title{Helper Function to create a new \code{SMQ} variable in ADAE by stacking
\code{SMQ} and/or \code{CMQ} records.}
\usage{
h_stack_by_baskets(
  df,
  baskets = grep("^(SMQ|CQ).+NAM$", names(df), value = TRUE),
  smq_varlabel = "Standardized MedDRA Query",
  keys = c("STUDYID", "USUBJID", "ASTDTM", "AESEQ", "AETERM")
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{baskets}{(\code{character})\cr variable names of the selected Standardized/Customized queries.}

\item{smq_varlabel}{(\code{string})\cr a label for the new variable created.}

\item{keys}{(\code{character})\cr names of the key variables to be returned
along with the new variable created.}
}
\description{
Helper Function to create a new \code{SMQ} variable in ADAE that
consists of all adverse events belonging to selected
Standardized/Customized queries.
The new dataset will only contain records of the adverse events
belonging to any of the selected baskets.
}
\examples{

library(scda)

adae <- synthetic_cdisc_data("rcd_2021_05_05")$adae[1:20 , ]
h_stack_by_baskets(df = adae)

}
