% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_biomarkers_subgroups.R
\name{h_tab_one_biomarker}
\alias{h_tab_one_biomarker}
\title{Helper Function for Tabulation of a Single Biomarker Result}
\usage{
h_tab_one_biomarker(df, afuns, colvars)
}
\arguments{
\item{df}{(\code{data.frame})\cr results for a single biomarker.}

\item{afuns}{(named \code{list} of \code{function})\cr analysis functions.}

\item{colvars}{(\code{list} with \code{vars} and \code{labels})\cr variables to tabulate and their labels.}
}
\value{
The \code{rtables} table object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is used by \code{\link[=h_tab_surv_one_biomarker]{h_tab_surv_one_biomarker()}} and \code{\link[=h_tab_rsp_one_biomarker]{h_tab_rsp_one_biomarker()}},
please see there for examples.
}
