% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_plot.R
\name{h_tbl_coxph_pairwise}
\alias{h_tbl_coxph_pairwise}
\title{Helper Function: Pairwise CoxPH table}
\usage{
h_tbl_coxph_pairwise(df, variables, control_coxph_pw = control_coxph())
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{variables}{(named \code{list}) of variable names. Details are: \cr
\itemize{
\item \code{tte}: variable indicating time-to-event duration values (\code{numeric}).
\item \code{is_event}: event variable (\code{logical}) \cr \code{TRUE} if event, \code{FALSE} if time to event is censored.
\item \code{arm}: the treatment group variable (\code{factor}).
\item \code{strat}: (\code{character} or \code{NULL}) variable names indicating stratification factors.
}}

\item{control_coxph_pw}{(\code{list}) \cr parameters for comparison details, specified by using \cr
the helper function \code{\link[=control_coxph]{control_coxph()}}. Some possible parameter options are: \cr
\itemize{
\item \code{pval_method}: (\code{string}) \cr p-value method for testing hazard ratio = 1.
Default method is "log-rank", can also be set to "wald" or "likelihood".
\item \code{ties}: (\code{string}) \cr specifying the method for tie handling. Default is "efron",
can also be set to "breslow" or "exact". See more in \code{\link[survival:coxph]{survival::coxph()}}
\item \code{conf_level}: (\code{proportion})\cr confidence level of the interval for HR.
}}
}
\description{
Create an rtable of pairwise stratified or unstratified CoxPH analysis results.
}
\examples{
\dontrun{
library(random.cdisc.data)
library(dplyr)

adtte <- radtte(cached = TRUE) \%>\%
  filter(PARAMCD == "OS") \%>\%
  mutate(is_event = CNSR == 0)

tern:::h_tbl_coxph_pairwise(
  df = adtte,
  variables = list(tte = "AVAL", is_event = "is_event", arm = "ARM"),
  control_coxph_pw = control_coxph(conf_level = 0.9))
}

}
