% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_plot.R
\name{h_tbl_median_surv}
\alias{h_tbl_median_surv}
\title{Helper Function: Survival Estimations}
\usage{
h_tbl_median_surv(fit_km)
}
\description{
Transform a survival fit to a table with groups in rows characterized
by N, median and confidence interval.
}
\examples{
\dontrun{
library(random.cdisc.data)
library(dplyr)

adtte <- radtte(cached = TRUE) \%>\%
  filter(PARAMCD == "OS")

fit <- survival::survfit(
  form = survival::Surv(AVAL, 1 - CNSR) ~ ARMCD,
  data = adtte
)
h_tbl_median_surv(fit_km = fit)
}

}
