% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abnormal_by_worst_grade_worsen.R
\name{h_worsen_counter}
\alias{h_worsen_counter}
\title{Helper Function to Analyze Patients for \code{\link[=s_count_abnormal_lab_worsen_by_baseline]{s_count_abnormal_lab_worsen_by_baseline()}}}
\usage{
h_worsen_counter(df, id, .var, baseline_var, direction_var)
}
\arguments{
\item{baseline_var}{(\code{string}) \cr baseline lab grade variable}
}
\value{
\code{\link[=h_worsen_counter]{h_worsen_counter()}} returns the counts and fraction of patients
whose worst post-baseline lab grades are worse than their baseline grades, for
post-baseline worst grades "1", "2", "3", "4" and "Any".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Helper function to count the number of patients and the fraction of patients according to
highest post-baseline lab grade variable \code{.var}, baseline lab grade variable \code{baseline_var},
and the direction of interest specified in \code{direction_var}.
}
\examples{
# `h_worsen_counter`
h_worsen_counter(
  df \%>\% filter(PARAMCD == "CRP" & GRADDR == "Low"),
  id = "USUBJID",
  .var = "ATOXGR",
  baseline_var = "BTOXGR",
  direction_var = "GRADDR"
)

}
