% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_plot.R
\name{h_xticks}
\alias{h_xticks}
\title{Helper function: x tick positions}
\usage{
h_xticks(data, xticks = NULL)
}
\arguments{
\item{data}{(\code{data.frame})\cr survival data as pre-processed by \code{h_data_plot}.}

\item{xticks}{(\code{numeric} or \code{number})\cr
numeric vector of ticks or single number with spacing
between ticks on the x axis.
If \code{NULL}, \code{\link[labeling:extended]{labeling::extended()}} is used to determine
an optimal tick position on the x axis.}
}
\description{
Calculate the positions of ticks on the x-axis. However, if \code{xticks} already
exists it is kept as is. It is based on the same function \code{ggplot2} relies on,
and is required in the graphic and the patient-at-risk annotation table.
}
\examples{

\dontrun{
library(random.cdisc.data)
library(dplyr)
library(survival)

data <- radtte(cached = TRUE) \%>\%
  filter(PARAMCD == "OS") \%>\%
  survfit(form = Surv(AVAL, 1 - CNSR) ~ ARMCD, data = .) \%>\%
  h_data_plot()

h_xticks(data)
h_xticks(data, xticks = seq(0, 3000, 500))
h_xticks(data, xticks = 500)
}

}
