% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_incidence_rate.R
\name{incidence_rate}
\alias{incidence_rate}
\alias{h_incidence_rate_normal}
\alias{h_incidence_rate_normal_log}
\alias{h_incidence_rate_exact}
\alias{h_incidence_rate_byar}
\alias{h_incidence_rate}
\alias{s_incidence_rate}
\alias{estimate_incidence_rate}
\alias{a_incidence_rate}
\title{Incidence rate}
\usage{
h_incidence_rate_normal(person_years, n_events, alpha = 0.05)

h_incidence_rate_normal_log(person_years, n_events, alpha = 0.05)

h_incidence_rate_exact(person_years, n_events, alpha = 0.05)

h_incidence_rate_byar(person_years, n_events, alpha = 0.05)

h_incidence_rate(person_years, n_events, control = control_incidence_rate())

s_incidence_rate(
  df,
  .var,
  n_events,
  is_event,
  control = control_incidence_rate()
)

estimate_incidence_rate(
  lyt,
  vars,
  ...,
  show_labels = "hidden",
  table_names = vars,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)

a_incidence_rate(
  df,
  .var,
  n_events,
  is_event,
  control = control_incidence_rate()
)
}
\arguments{
\item{person_years}{(\code{numeric}) \cr total person-years at risk.}

\item{n_events}{(\code{integer}) \cr number of events observed.}

\item{alpha}{(\code{numeric}) \cr two-sided alpha-level for confidence interval.}

\item{control}{(\code{list}) \cr parameters for estimation details, specified by using
the helper function \code{\link[=control_incidence_rate]{control_incidence_rate()}}. Possible parameter options are: \cr
\itemize{
\item \code{conf_level}: (\code{proportion}) \cr confidence level for the estimated incidence rate.
\item \code{conf_type}: (\code{string}) \cr \code{normal} (default), \code{normal_log}, \code{exact}, or \code{byar}
for confidence interval type.
\item \code{time_unit_input}: (\code{string}) \cr \code{day}, \code{month}, or \code{year} (default)
indicating time unit for data input.
\item \code{time_unit_output}: (\code{numeric}) \cr time unit for desired output (in person-years).
}}

\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{is_event}{(\code{logical})\cr \code{TRUE} if event, \code{FALSE} if time to event is censored.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{...}{additional arguments for the lower level functions.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
The statistics are:
\itemize{
\item \code{person_years}: total person-years at risk
\item \code{n_events}: total number of events observed
\item \code{rate}: estimated incidence rate
\item \code{rate_ci}: confidence interval for the incidence rate
}
}
\description{
Estimate the event rate adjusted for person-years at risk, otherwise known
as incidence rate. Primary analysis variable is the person-years at risk.
}
\section{Functions}{
\itemize{
\item \code{h_incidence_rate_normal}: helper function to estimate the incidence rate and
associated confidence interval based on the normal approximation for the
incidence rate. Unit is one person-year.

\item \code{h_incidence_rate_normal_log}: helper function to estimate the incidence rate and
associated confidence interval based on the normal approximation for the
logarithm of the incidence rate. Unit is one person-year.

\item \code{h_incidence_rate_exact}: helper function to estimate the incidence rate and
associated exact confidence interval. Unit is one person-year.

\item \code{h_incidence_rate_byar}: helper function to estimate the incidence rate and
associated Byar's confidence interval. Unit is one person-year.

\item \code{h_incidence_rate}: incidence_rate helper function to estimate the incidence rate and
associated confidence interval.

\item \code{s_incidence_rate}: statistics function which estimates the incidence rate and the
associated confidence interval.

\item \code{estimate_incidence_rate}: layout creating function which adds analyze rows using the statistics
function \code{s_incidence_rate} and desired format.

\item \code{a_incidence_rate}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.
}}

\examples{
h_incidence_rate_normal(200, 2)

h_incidence_rate_normal_log(200, 2)

h_incidence_rate_exact(200, 2)

h_incidence_rate_byar(200, 2)

h_incidence_rate(200, 2)

h_incidence_rate(
  200,
  2,
  control_incidence_rate(
    conf_level = 0.9,
    conf_type = "normal_log",
    time_unit_output = 100)
)


library(dplyr)

df <- data.frame(
  USUBJID = as.character(seq(6)),
  CNSR = c(0, 1, 1, 0, 0, 0),
  AVAL = c(10.1, 20.4, 15.3, 20.8, 18.7, 23.4),
  ARM = factor(c("A", "A", "A", "B", "B", "B"))
) \%>\%
  mutate(is_event = CNSR == 0) \%>\%
  mutate(n_events = as.integer(is_event))

s_incidence_rate(
  df,
  .var = "AVAL",
  n_events = "n_events",
  control = control_incidence_rate(
    time_unit_input = "month",
    time_unit_output = 100
  )
)

basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  add_colcounts() \%>\%
  estimate_incidence_rate(
    vars = "AVAL",
    n_events = "n_events",
    control = control_incidence_rate(
      time_unit_input = "month",
      time_unit_output = 100
    )
  ) \%>\%
  build_table(df)

a_incidence_rate(
  df,
  .var = "AVAL",
  n_events = "n_events",
  control = control_incidence_rate(time_unit_input = "month", time_unit_output = 100)
)

}
