% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_plot.R
\name{kaplan_meier}
\alias{kaplan_meier}
\title{Kaplan-Meier Plot}
\arguments{
\item{data}{(\code{data.frame})\cr survival data as pre-processed by \code{h_data_plot}.}

\item{fit_km}{(\code{survfit})\cr see \code{\link[survival:survfit]{survival::survfit()}}.}

\item{xticks}{(\code{numeric} or \code{number})\cr
numeric vector of ticks or single number with spacing
between ticks on the x axis.
If \code{NULL}, \code{\link[labeling:extended]{labeling::extended()}} is used to determine
an optimal tick position on the x axis.}

\item{censor_show}{(\code{flag})\cr whether to show censored.}

\item{xlab}{(\code{string})\cr label of x-axis.}

\item{ylab}{(\code{string})\cr label of y-axis.}

\item{title}{(\code{string})\cr title for plot.}

\item{draw}{(\code{flag})\cr draw the plot on device.}

\item{newpage}{(\code{flag})\cr open a new page on the graphic device.}

\item{col}{(\code{character})\cr lines colors. Length of a vector should be equal
to number of strata from \code{\link[survival:survfit]{survival::survfit()}}.}

\item{lty}{(\code{numeric})\cr line type. Length of a vector should be equal
to number of strata from \code{\link[survival:survfit]{survival::survfit()}}.}

\item{lwd}{(\code{numeric})\cr line width. Length of a vector should be equal
to number of strata from \code{\link[survival:survfit]{survival::survfit()}}.}

\item{pch}{(\code{numeric})\cr point type for censored.}

\item{size}{(\code{numeric})\cr size of censored point, a class of \code{unit}.}

\item{max_time}{(\code{numeric})\cr maximum value of X axis range (\code{NULL} for
default)}

\item{ggtheme}{(\code{theme})\cr a graphical theme as provided by \code{ggplot2} to
control outlook of the Kaplan-Meier curve.}

\item{annot_at_risk}{(\code{flag})\cr compute and add the annotation table
reporting the number of patient at risk matching the main grid of the
Kaplan-Meier curve.}

\item{annot_surv_med}{(\code{flag})\cr compute and add the annotation table
on the top right corner of the Kaplan-Meier curve estimating the
median survival time per group.}
}
\description{
Kaplan-Meier Plot
}
