% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot_data.R
\name{kmCurveData}
\alias{kmCurveData}
\title{Prepare necessary data for Kaplan-Meier Plot}
\usage{
kmCurveData(fit_km, xticks = NULL)
}
\arguments{
\item{fit_km}{a class \code{\link{survfit}} object.}

\item{xticks}{break interval of x-axis. It takes a numeric vector or \code{NULL}.}
}
\description{
Return a list of data for primitive element of grid drawing
}
\examples{


OS <- data.frame(AVAL = abs(rnorm(200)), 
                 CNSR = sample(c(0, 1), 200, TRUE), 
                 ARM = sample(LETTERS[1:3], 200, TRUE),
                 SEX = sample(c("M","F"), 200, TRUE),
                 RACE = sample(c("AA", "BB", "CC"), 200, TRUE),
                 ECOG = sample(c(0, 1), 200, TRUE))
fit_km <- survfit(Surv(AVAL, 1-CNSR) ~ ARM, data = OS, conf.type = "plain")
kmCurveData(fit_km, xticks = c(0.5, 0.8, 1.5))


}
