% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_km.R
\name{kmGrob}
\alias{kmGrob}
\title{Define a kmplot grob}
\usage{
kmGrob(title = "Kaplan - Meier Plot", ...)
}
\arguments{
\item{title}{title for plot.}

\item{...}{All other arguments will be passed to a call to the \code{kmCurveData()} function.}
}
\description{
create a grid graphical object for basic KM plot using data from a call of function \code{kmCurveData}
}
\examples{
OS <- data.frame(AVAL = abs(rnorm(200)), 
                 CNSR = sample(c(0, 1), 200, TRUE), 
                 ARM = sample(LETTERS[1:3], 200, TRUE),
                 SEX = sample(c("M","F"), 200, TRUE),
                 RACE = sample(c("AA", "BB", "CC"), 200, TRUE),
                 ECOG = sample(c(0, 1), 200, TRUE))
kmGrob(formula_km = Surv(AVAL, 1-CNSR) ~ ARM, data = OS)

}
