% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot.R
\name{kmplot}
\alias{kmplot}
\title{our fast km plot}
\usage{
kmplot(formula_km, data, add_km = TRUE, add_coxph = TRUE,
  formula_coxph = formula_km,
  info_coxph = "Cox Proportional Model: Unstratified Analysis", add = FALSE,
  title = "Kaplan - Meier Plot")
}
\description{
our fast km plot
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr)
library(forcats)
library(survival)
ATE <- ate(com.roche.cdt30019.go29436.re)

ANL <- ATE \%>\%
  filter(PARAMCD == "OS") \%>\%
  select(AVAL, CNSR, ARM, SEX, RACE, MLIVER) \%>\%
  mutate(ARM = fct_relevel(ARM, "DUMMY C"))

kmplot(Surv(AVAL, 1-CNSR) ~ ARM, data = ANL)

kmplot(Surv(AVAL, 1-CNSR) ~ ARM, data = ANL, add_coxph = TRUE, add_km = TRUE)


## facet by row
nplots <- 2
# new plot
grid.newpage()

# margins
pushViewport(plotViewport(margin = c(3, 10, 2, 2)))

# layout
pushViewport(viewport(layout = grid.layout(ncol = 1, nrow = 2*nplots-1,
   heights = unit(head(rep(c(1, 7), nplots), -1), head(rep(c("null", "lines"), nplots), -1))
)))

# now add first plot
pushViewport(viewport(layout.pos.row = 1))
kmplot(Surv(AVAL, 1-CNSR) ~ ARM, data = ANL, add_coxph = TRUE, add = TRUE)
popViewport()

#pushViewport(viewport(layout.pos.row = 2))
#grid.rect(gp = gpar(fill = "thistle"))
#popViewport()

pushViewport(viewport(layout.pos.row = 3))
grid.text("MY TITLE", y = unit(1, "npc") + unit(1, "lines"), gp = gpar(fontface = "bold", fontsize = 16))
kmplot(Surv(AVAL, 1-CNSR) ~ ARM, data = ANL, add_coxph = TRUE, add = TRUE)
popViewport()


#survminer::ggsurvplot(fit, risk.table = TRUE, break.time.by = 2)

}

}
