% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot_regular.R
\name{kmPlot}
\alias{kmPlot}
\title{kmPlot}
\usage{
kmPlot(time_to_event, event, arm, arm.ref, arm.rest = setdiff(arm, arm.ref),
  stratum = NULL, facet_by = NULL, n_col = 2, ...)
}
\arguments{
\item{time_to_event}{time to event values}

\item{event}{boolean, \code{TRUE} if event and \code{FALSE} if censored}

\item{arm.ref}{character: reference arm}
}
\description{
descr
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr)
library(survival)
library(ggplot2)
library(survminer)

ATE <-  ate(com.roche.cdt30019.go29436.re)

ATE_filtered <- ATE \%>\% filter(PARAMCD == "OS")
kmPlot(
   time_to_event = ATE_filtered$AVAL,
   event = ATE_filtered$CNSR == 0,
   arm = ATE_filtered$ARM,
   arm.ref =  "DUMMY B"  
   )
   
kmPlot(
   time_to_event = ATE_filtered$AVAL,
   event = ATE_filtered$CNSR == 0,
   arm = ATE_filtered$ARM,
   facet_by = ATE_filtered$SEX,
   arm.ref =  "DUMMY B" ,
   arm.rest =  c("DUMMY A", "DUMMY C" )
   )
 }
}
