% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_transpose.R
\name{list_transpose}
\alias{list_transpose}
\title{Transpose a list by changing depht 2 elements with depth 1 elements}
\usage{
list_transpose(x)
}
\description{
Transpose a list by changing depht 2 elements with depth 1 elements
}
\examples{

x <- list(c1 = 1:3, c2 = 11:13, c3 = 111:113)

}
