% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{logistic_regression_cols}
\alias{logistic_regression_cols}
\title{Logistic Regression Multi-Variable Column Layout Function}
\usage{
logistic_regression_cols(lyt, conf_level = 0.95)
}
\arguments{
\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Layout creating function for a multi-variable column layout summarizing
logistic regression results.
}
