% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{logistic_summary_by_flag}
\alias{logistic_summary_by_flag}
\title{Logistic Regression Summary Table Constructor Function}
\usage{
logistic_summary_by_flag(flag_var)
}
\arguments{
\item{flag_var}{(\code{string})\cr variable name identifying which row should be used in this
content function.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Constructor for content functions to be used to summarize
logistic regression results.
}
