% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{lt_ae_term}
\alias{lt_ae_term}
\title{Only By Term}
\usage{
lt_ae_term(term, id, col_by, col_N = tapply(col_by, col_by, length),
  total = "All Patients", term_label)
}
\description{
Only By Term
}
\examples{
l_tbls <- lt_ae_term( 
  term =  ANL$TERM,
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients",
  term_label = "MedDRA Preferred Term"
)
do.call(fast_stack_rtables, l_tbls)
}
