% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{month2day}
\alias{month2day}
\title{Conversion of Months to Days}
\usage{
month2day(x)
}
\arguments{
\item{x}{(\code{numeric})\cr time in months.}
}
\value{
A \code{numeric} vector with the time in days.
}
\description{
Conversion of Months to Days. This is an approximative calculation because it
considers each month as having an average of 30.4375 days.
}
\examples{
x <- c(13.25, 8.15, 1, 2.834)
month2day(x)

}
