% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_available}
\alias{n_available}
\title{Number of Available (Non-Missing Entries) in a Vector}
\usage{
n_available(x)
}
\arguments{
\item{x}{(\code{vector})\cr where to count the non-missing values.}
}
\value{
Number of non-missing values.
}
\description{
Small utility function for better readability.
}
\examples{
# Internal function - n_available
\dontrun{
n_available(c(1, NA, 2))
}

}
\keyword{internal}
