% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{na_as_level}
\alias{na_as_level}
\title{Replace NA values by NA string level}
\usage{
na_as_level(x, na_level = "NA")
}
\arguments{
\item{x}{factor, possibly with NAs}

\item{na_level}{name of new level for NAs}
}
\value{
factor with additional NA level
}
\description{
Replace NA values by NA string level
}
\examples{
na_as_level(factor(c(1, 1, 2)))
na_as_level(factor(c(1, 1, NA)), "na")
}
