% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\docType{class}
\name{node-class}
\alias{node-class}
\alias{node-validity}
\alias{node}
\title{An S4 Class to represent nodes in a tree}
\usage{
node(name, content, children)
}
\arguments{
\item{name}{name of node}

\item{content}{content of node}

\item{children}{children of node}
}
\value{
object of class node
}
\description{
From Wikipedia: A tree data structure can be defined recursively as a collection of nodes (starting at a root node),
where each node is a data structure consisting of a value, together with a list of references to nodes (the
"children"), with the constraints that no reference is duplicated, and none points to the root.

A node is also known as a tree. This function makes sure that the children are properly named.
They can then be properly accessed through node_children[['childName']]
See the class node for more info about parameters.
}
\details{
Use the function \code{\link{node}} to create objects of this class.

Note: changing the names of the children after calling this function will cause the children list names
to be no longer in-sync. (or should accessor update child names??)
}
\section{Slots}{

\describe{
\item{\code{name}}{name of node}

\item{\code{content}}{of node (e.g. \code{data.frame} or \code{rtable} objects)}

\item{\code{children}}{a list of \code{node} objects}
}}

\examples{
n11 <- node(name = "A", content = array(c(1:6), dim = c(2,3)), children = list())
n12 <- node(name = "B", content = array(c(1:6), dim = c(2,3)), children = list())
n13 <- node(name = "C", content = array(c(1:6), dim = c(2,3)), children = list())
n2 <- node(name = "D", content = c(1:3), children = list(n11, n12, n13))
# incorrect example (children with same name):
#node(name = "A", content = c(1:3), children = list("A"))
}
