% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{node_format_data}
\alias{node_format_data}
\title{Format}
\usage{
node_format_data(
  gap_to_children = NULL,
  children_gap = NULL,
  children_indent = NULL,
  content_indent = NULL,
  left_header = NULL
)
}
\arguments{
\item{gap_to_children}{row gap between content and children}

\item{children_gap}{row gaps between children}

\item{children_indent}{indentation of children wrt to node with respect to current indent level}

\item{content_indent}{indentation of content of node with respect to current indent level}

\item{left_header}{header to add to the left of the \code{rtable}, is not inherited}
}
\description{
Node Format Data
}
\examples{
tbl <- function(x) {
  rtable(c("A", "B"), rrow(x, "-", "-"))
}

to_rtable(node("AAA", tbl("r1")))
to_rtable(node("AAA", tbl("r1"), children = list(
   node("c1", tbl("r2")),
   node("c1", tbl("r2")),
   node("c1", tbl("r2"))
)))


to_rtable(node("AAA", tbl("r1"), children = list(
   node("c1", tbl("r2")),
   node("c1", tbl("r2")),
   node("c1", tbl("r2"))
), format_data = node_format_data(children_indent = 0)))
}
