% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{odds_ratio}
\alias{odds_ratio}
\alias{or_glm}
\title{Odds Ratio Estimation}
\usage{
or_glm(data, conf_level)
}
\arguments{
\item{data}{(\verb{data frame})\cr
expected with three variables \code{rsp}, \code{grp}.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}
}
\description{
Compares bivariate responses between two groups in terms of odds ratios
along with a confidence interval.
}
\details{
This function uses either logistic regression for unstratified
analyses, or conditional logistic regression for stratified analyses.
The Wald confidence interval with the specified confidence level is
calculated. Note that, for stratified analyses, there is currently no
implementation for conditional likelihood confidence intervals,
therefore the likelihood confidence interval as an option is not yet
available. Besides, when \code{rsp} contains only responders or non-responders,
then the result values will be \code{NA}, because no odds ratio estimation is
possible.
}
\section{Functions}{
\itemize{
\item \code{or_glm}: estimates the odds ratio based on \code{\link[stats:glm]{stats::glm()}}.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{

data <- data.frame(
  rsp = as.logical(c(1, 1, 0, 1, 0, 0)),
  grp = letters[c(1, 1, 1, 2, 2, 2)]
)

or_glm(data, conf_level = 0.95)

}
