% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{pool_obj}
\alias{pool_obj}
\title{Catched \code{rbmi} pool S3 object}
\format{
An object of class \code{pool} of length 5.
}
\usage{
data(pool_obj)
}
\description{
Catched \code{rbmi} pool S3 object
}
\keyword{datasets}
