% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_proportion_diff.R
\name{prop_diff_test}
\alias{prop_diff_test}
\alias{prop_chisq}
\alias{prop_schouten}
\alias{prop_fisher}
\alias{prop_cmh}
\title{Difference Test for Two Proportions}
\usage{
prop_chisq(tbl)

prop_schouten(tbl)

prop_fisher(tbl)

prop_cmh(ary)
}
\arguments{
\item{tbl}{(\code{matrix})\cr
with two groups in rows and the binary response (\code{TRUE}/\code{FALSE}) in
columns.}

\item{ary}{(\code{array}, 3 dimensions)\cr
with two groups in rows, the binary response (\code{TRUE}/\code{FALSE}) in
columns, the strata in the third dimension.}
}
\description{
Various tests were implemented to test the difference between two
proportions.
}
\section{Functions}{
\itemize{
\item \code{prop_chisq}: performs Chi-Squared test.
Internally calls \code{\link[stats:prop.test]{stats::prop.test()}}.

\item \code{prop_schouten}: performs the Chi-Squared test with Schouten
correction.

\item \code{prop_fisher}: performs the Fisher's exact test.
Internally calls \code{\link[stats:fisher.test]{stats::fisher.test()}}.

\item \code{prop_cmh}: performs stratified Cochran-Mantel-Haenszel test.
Internally calls \code{\link[stats:mantelhaen.test]{stats::mantelhaen.test()}}.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{

# Non-stratified proportion difference test

## Data
A <- 20
B <- 20
set.seed(1)
rsp <- c(
  sample(c(TRUE, FALSE), size = A, prob = c(3/4, 1/4), replace = TRUE),
  sample(c(TRUE, FALSE), size = A, prob = c(1/2, 1/2), replace = TRUE)
)
grp <- c(rep("A", A), rep("B", B))
tbl <- table(grp, rsp)

## Chi-Squared test
prop_chisq(tbl)


## Chi-Squared test + Schouten correction.
prop_schouten(tbl)


## Fisher's exact test
prop_fisher(tbl)


# Stratified proportion difference test

## Data
rsp <- sample(c(TRUE, FALSE), 100, TRUE)
grp <- factor(rep(c("A", "B"), each = 50))
strata <- factor(rep(c("V", "W", "X", "Y", "Z"), each = 20))
tbl <- table(grp, rsp, strata)

## Cochran-Mantel-Haenszel test
prop_cmh(tbl)

}
