% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_diff_test.R
\name{prop_diff_test}
\alias{prop_diff_test}
\alias{prop_chisq}
\alias{prop_schouten}
\alias{prop_fisher}
\alias{prop_cmh}
\alias{s_test_proportion_diff}
\alias{a_test_proportion_diff}
\alias{test_proportion_diff}
\title{Difference Test for Two Proportions}
\usage{
prop_chisq(tbl)

prop_schouten(tbl)

prop_fisher(tbl)

prop_cmh(ary)

s_test_proportion_diff(
  df,
  .var,
  .ref_group,
  .in_ref_col,
  variables = list(strata = NULL),
  method = c("chisq", "schouten", "fisher", "cmh")
)

a_test_proportion_diff(
  df,
  .var,
  .ref_group,
  .in_ref_col,
  variables = list(strata = NULL),
  method = c("chisq", "schouten", "fisher", "cmh")
)

test_proportion_diff(
  lyt,
  vars,
  ...,
  show_labels = "hidden",
  table_names = vars,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{tbl}{(\code{matrix})\cr
with two groups in rows and the binary response (\code{TRUE}/\code{FALSE}) in
columns.}

\item{ary}{(\code{array}, 3 dimensions)\cr
with two groups in rows, the binary response (\code{TRUE}/\code{FALSE}) in
columns, the strata in the third dimension.}

\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.ref_group}{(\verb{data frame} or \code{vector})\cr the data corresponding to the reference group.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}

\item{variables}{(named \code{list} of \code{string})\cr list of additional analysis variables.}

\item{method}{(\code{string})\cr
one of (\code{chisq}, \code{cmh}, \code{fisher}, \code{schouten}; specifies the test used
to calculate the p-value.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{...}{other arguments are passed to \code{\link[=s_test_proportion_diff]{s_test_proportion_diff()}}.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
Named \code{list} with a single item \code{pval} with an attribute \code{label}
describing the method used. The p-value tests the null hypothesis that
proportions in two groups are the same.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Various tests were implemented to test the difference between two
proportions.
}
\section{Functions}{
\itemize{
\item \code{prop_chisq()}: performs Chi-Squared test.
Internally calls \code{\link[stats:prop.test]{stats::prop.test()}}.

\item \code{prop_schouten()}: performs the Chi-Squared test with Schouten correction.

\item \code{prop_fisher()}: performs the Fisher's exact test.
Internally calls \code{\link[stats:fisher.test]{stats::fisher.test()}}.

\item \code{prop_cmh()}: performs stratified Cochran-Mantel-Haenszel test.
Internally calls \code{\link[stats:mantelhaen.test]{stats::mantelhaen.test()}}. Note that strata with less than two observations
are automatically discarded.

\item \code{s_test_proportion_diff()}: Statistics function which test the difference
between two proportions.

\item \code{a_test_proportion_diff()}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{test_proportion_diff()}: Layout creating function which can be used for
creating tables, which can take statistics function arguments and
additional format arguments.

}}
\examples{
# Non-stratified proportion difference test

## Data
A <- 20
B <- 20
set.seed(1)
rsp <- c(
  sample(c(TRUE, FALSE), size = A, prob = c(3 / 4, 1 / 4), replace = TRUE),
  sample(c(TRUE, FALSE), size = A, prob = c(1 / 2, 1 / 2), replace = TRUE)
)
grp <- c(rep("A", A), rep("B", B))
tbl <- table(grp, rsp)

## Chi-Squared test
# Internal function - prop_chisq
\dontrun{
prop_chisq(tbl)
}

## Chi-Squared test + Schouten correction.
# Internal function - prop_schouten
\dontrun{
prop_schouten(tbl)
}

## Fisher's exact test
# Internal function - prop_fisher
\dontrun{
prop_fisher(tbl)
}

# Stratified proportion difference test

## Data
rsp <- sample(c(TRUE, FALSE), 100, TRUE)
grp <- factor(rep(c("A", "B"), each = 50))
strata <- factor(rep(c("V", "W", "X", "Y", "Z"), each = 20))
tbl <- table(grp, rsp, strata)

## Cochran-Mantel-Haenszel test
# Internal function - prop_cmh
\dontrun{
prop_cmh(tbl)
}


# Statistics function
dta <- data.frame(
  rsp = sample(c(TRUE, FALSE), 100, TRUE),
  grp = factor(rep(c("A", "B"), each = 50)),
  strat = factor(rep(c("V", "W", "X", "Y", "Z"), each = 20))
)

# Internal function - s_test_proportion_diff
\dontrun{
s_test_proportion_diff(
  df = subset(dta, grp == "A"),
  .var = "rsp",
  .ref_group = subset(dta, grp == "B"),
  .in_ref_col = FALSE,
  variables = list(strata = "strat"),
  method = "cmh"
)
}

# Internal function - a_test_proportion_diff
\dontrun{
a_test_proportion_diff(
  df = subset(dta, grp == "A"),
  .var = "rsp",
  .ref_group = subset(dta, grp == "B"),
  .in_ref_col = FALSE,
  variables = list(strata = "strat"),
  method = "cmh"
)
}

# With `rtables` pipelines.
l <- basic_table() \%>\%
  split_cols_by(var = "grp", ref_group = "B") \%>\%
  test_proportion_diff(
    vars = "rsp",
    method = "cmh", variables = list(strata = "strat")
  )

build_table(l, df = dta)
}
\seealso{
For information on the Schouten correction (Schouten, 1980),
visit https://onlinelibrary.wiley.com/doi/abs/10.1002/bimj.4710220305.
}
\keyword{internal}
