% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{range_noinf}
\alias{range_noinf}
\title{Re-implemented \code{\link[=range]{range()}} Default S3 method for numerical objects}
\usage{
range_noinf(x, na.rm = FALSE, finite = FALSE)
}
\arguments{
\item{x}{(\code{numeric})\cr a sequence of numbers for which the range is computed.}

\item{na.rm}{(\code{logical})\cr indicating if \code{NA} should be omitted.}

\item{finite}{(\code{logical})\cr indicating if non-finite elements should be removed.}
}
\value{
A 2-element vector of class \code{numeric}.
}
\description{
This function returns \code{c(NA, NA)} instead of \code{c(-Inf, Inf)} for zero-length data
without any warnings.
}
\keyword{internal}
