% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{range_noinf}
\alias{range_noinf}
\title{Re-implemented \code{\link[base]{range.default}} default S3 method for numerical objects only.
It returns \code{c(NA, NA)} instead of \code{c(-Inf, Inf)} for zero-length data
without any warnings.}
\usage{
range_noinf(x, na.rm = FALSE, finite = FALSE)
}
\arguments{
\item{x}{numeric S3 class, a sequence of numbers for which the range is computed.}

\item{na.rm}{logical, indicating if \code{\link[base]{NA}}'s should be
    omitted.}

\item{finite}{logical, indicating if all non-finite elements should
    be omitted.}
}
\value{
2-elements vector of class numeric.
}
\description{
Re-implemented \code{\link[base]{range.default}} default S3 method for numerical objects only.
It returns \code{c(NA, NA)} instead of \code{c(-Inf, Inf)} for zero-length data
without any warnings.
}
\examples{
range_noinf(1:5)
range_noinf(c(1:5, NA, NA), na.rm = TRUE)
range_noinf(numeric(), na.rm = TRUE)
range_noinf(c(1:5, NA, NA, Inf), na.rm = TRUE, finite = TRUE)
range_noinf(Inf)
range_noinf(Inf, na.rm = TRUE, finite = TRUE)
range_noinf(c(Inf, NA), na.rm = FALSE, finite = TRUE)
range_noinf(c(1, Inf, NA), na.rm = FALSE, finite = TRUE)

}
