% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{range_noinf}
\alias{range_noinf}
\title{Re-implemented \code{\link[base:range]{range.default}} default S3 method for numerical objects only.
It returns \code{c(NA, NA)} instead of \code{c(-Inf, Inf)} for zero-length data
without any warnings.}
\usage{
range_noinf(x, na.rm = FALSE, finite = FALSE)
}
\arguments{
\item{x}{numeric S3 class, a sequence of numbers for which the range is computed.}

\item{na.rm}{(\code{logical})\cr indicating if NA should be omitted.}

\item{finite}{(\code{logical})\cr indicating if non-finite elements should be removed.}
}
\value{
2-elements vector of class numeric.
}
\description{
Re-implemented \code{\link[base:range]{range.default}} default S3 method for numerical objects only.
It returns \code{c(NA, NA)} instead of \code{c(-Inf, Inf)} for zero-length data
without any warnings.
}
\examples{
# Internal function - range_noinf
\dontrun{
tern:::range_noinf(1:5)
tern:::range_noinf(c(1:5, NA, NA), na.rm = TRUE)
tern:::range_noinf(numeric(), na.rm = TRUE)
tern:::range_noinf(c(1:5, NA, NA, Inf), na.rm = TRUE, finite = TRUE)
tern:::range_noinf(Inf)
tern:::range_noinf(Inf, na.rm = TRUE, finite = TRUE)
tern:::range_noinf(c(Inf, NA), na.rm = FALSE, finite = TRUE)
tern:::range_noinf(c(1, Inf, NA), na.rm = FALSE, finite = TRUE)
}

}
\keyword{internal}
