% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reapply_varlabels}
\alias{reapply_varlabels}
\title{Reapply Variable Labels}
\usage{
reapply_varlabels(x, varlabels, ...)
}
\arguments{
\item{x}{(\code{vector})\cr vector of elements that needs new labels.}

\item{varlabels}{(\code{character})\cr vector of labels for \code{x}.}

\item{...}{further parameters to be added to the list.}
}
\value{
\code{x} with variable labels reapplied.
}
\description{
This is a helper function that is used in tests.
}
