% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mmrm.R
\name{refit_lme4_all_optimizers}
\alias{refit_lme4_all_optimizers}
\title{Refit an lme4 model with all possible optimizers and return the best result.}
\usage{
refit_lme4_all_optimizers(original_fit)
}
\arguments{
\item{original_fit}{The original model fit coming from \code{\link{fit_lme4_single_optimizer}}.}
}
\value{
The "best" model fit, defined as a converging fit without any warnings or messages, leading
  to the highest log-likelihood. If no optimizer succeeds, then an error is thrown.
}
\description{
Refit an lme4 model with all possible optimizers and return the best result.
}
