% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_response_table.R
\name{response_table}
\alias{response_table}
\title{Create the Response Table}
\usage{
response_table(response, val_rsp, val_nrsp, arm, arm_ref, arm_comp,
  incl_missing)
}
\arguments{
\item{response}{Tumor response data}

\item{val_rsp}{Character vector, defining list of response values to be used as responders}

\item{val_nrsp}{Character vector, defining list of response values to be used as non-responders}

\item{arm}{Arm information data}

\item{arm_ref}{Character vector, defining which arm(s) from the list of arms should be used as reference group}

\item{arm_comp}{Character vector, defining which arm(s) from the list of arms should be used as comparison group}

\item{incl_missing}{Boolean value, \code{TRUE} if missing values should be considered non-responders, 
\code{FALSE} if missing response should be removed from analysis}
}
\description{
Create the Response Table
}
\examples{


\dontrun{
library(atezo.data)
library(dplyr)
'\%needs\%' <- teal.oncology:::'\%needs\%'

ASL <- asl(com.roche.cdt30019.go29436.re)
ARS <- ars(com.roche.cdt30019.go29436.re)

ARS_f <- ARS \%>\% filter(PARAMCD == "BESRSPI")

response_table(
   response     = ARS_f$AVALC,
   val_rsp      = c("CR","PR"),
   val_nrsp     = c("SD", "NON CR/PD", "PD", "NE"),
   arm          = ARS_f$ARM,
   arm_ref      = "DUMMY A",
   arm_comp     = "DUMMY B",
   incl_missing = T
)

} 

}
