% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_table.R
\name{response_table}
\alias{response_table}
\title{Create the Response Table}
\usage{
response_table(response, value.resp = c("CR", "PR"), value.nresp = c("SD",
  "NON CR/PD", "PD", "NE"), arm, style = 1)
}
\arguments{
\item{response}{Tumor response data}

\item{arm}{Arm information data}

\item{style}{Must be 1 or 2, \code{1} if only want to display rates summary for each response value category,
\code{2} if want to display rates, difference in rate and odds ratio for each response value category.}

\item{value.rsp}{Character vector, defining list of response values to be used as responders}

\item{value.nrsp}{Character vector, defining list of response values to be used as non-responders}

\item{arm.ref}{Character vector, defining which arm(s) from the list of arms should be used as reference group}

\item{arm.comp}{Character vector, defining which arm(s) from the list of arms should be used as comparison group}

\item{arm.comp.combine}{Logical, \code{TRUE} if want all non-ref arms to be combined into one comparison group, 
\code{FALSE} if want each of the non-ref arms to be a separate comparison group}
}
\description{
Create the Response Table
}
\details{
We use the \code{PropCIs} to calculate the CIs because...
}
\examples{


\dontrun{
library(dplyr)
library(shiny)
library(atezo.data)
library(teal.oncology)
library(PropCIs)
library(forcats)


ARS <- ars(com.roche.cdt30019.go29436.re) \%>\% select(c("USUBJID", "STUDYID", "ARM", "PARAMCD", "AVALC")) \%>\% 
       filter(PARAMCD == "OVRSPI")

#If want to include missing as non-responders
ARS$AVALC[ARS$AVALC==""] <- "NE"


tbl_stream <- get_response_table(com.roche.cdt30019.go29436.re)
Viewer(tbl_stream)

tbl <- response_table(response = ARS$AVALC, arm = ARS$ARM, arm.ref = "DUMMY B", arm.comp.combine=F)
Viewer(tbl)

response_table(response = ARS$AVALC, arm = ARS$ARM, arm.ref = "DUMMY B", arm.comp.combine=T)
response_table(response = ARS$AVALC, arm = ARS$ARM, arm.ref = "DUMMY B", arm.comp.combine=F, style = 2)
response_table(response = ARS$AVALC, arm = ARS$ARM, arm.ref = "DUMMY B", arm.comp.combine=T, style = 2)


ars_f2 <-  ars(com.roche.cdt30019.go29436.re) \%>\%
 filter(ITTGEFL=='Y', ITTWTFL=='Y', PARAMCD=='OVRSPI')

tbl <- response_table(
  response = vapply(ars_f2$AVALC, function(x) {if (x == "") "NE" else x}, character(1)),
  arm = ars_f2$ARM,
  arm.ref = "DUMMY C",
  arm.comp.combine = FALSE
)

Viewer(tbl)

compare_rtables(tbl, tbl_stream, comp.attr = FALSE)
compare_rtables(tbl, tbl_stream)

tbl[2,3]
tbl_stream[2,3]

Arm variable is armcd1
}


'\%needs\%' <- teal.oncology:::'\%needs\%' # for debugging purpuses
}
\author{
Chendi Liao (liaoc10), \email{chendi.liao@roche.com}
}
