% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_response_table.R
\name{response_table}
\alias{response_table}
\title{Create the Response Table}
\usage{
response_table(time_to_event, event, arm)
}
\arguments{
\item{time_to_event}{time to event variable}

\item{event}{is a boolean with \code{TRUE} if event and \code{FALSE} if
censored}

\item{arm}{arm data}
}
\description{
Create the Response Table
}
\examples{

\dontrun{
library(atezo.data)

ATE <- ate(com.roche.cdt30019.go29436.re)

library(dplyr)

ATE_F <- ATE \%>\% filter(PARAMCD == "OS")

response_table(
   time_to_event = ATE_F$AVAL,
   event = ATE_F$CNSR == 0,
   arm = ATE_F$ARM
)

get_response_table(com.roche.cdt30019.go29436.re)

} 

}
