% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_table.R
\name{response_table_ADAM}
\alias{response_table_ADAM}
\title{Response Table with ADaM data structure}
\usage{
response_table_ADAM(ASL, ARS, paramcd = "OVRSPI", arm.var = "ARM", arm.ref,
  arm.comp, arm.comp.combine = TRUE, value.resp = c("CR", "PR"),
  value.nresp = c("SD", "NON CR/PD", "PD", "NE"), incl.missing = TRUE,
  strata.var = NULL, style = 1)
}
\arguments{
\item{ASL}{dataset with following variables: USUBJID, STUDYID, the specified
grouping variable}

\item{ARS}{ARS dataset containing the following variables: USUBJID, STUDYID,
PARAMCD, AVALC}

\item{paramcd}{Name of overall response parameter}

\item{arm.var}{Name of variable with arm information}

\item{arm.ref}{Character vector, defining which arm(s) from the list of arms
should be used as reference group}

\item{arm.comp}{Character vector, defining which arm(s) from the list of arms
should be used as comparison group}

\item{arm.comp.combine}{Logical, \code{TRUE} if want all non-ref arms to be
combined into one comparison group, \code{FALSE} if want each of the
non-ref arms to be a separate comparison group}

\item{incl.missing}{Logical, \code{TRUE} if missing values should be
considered non-responders, \code{FALSE} if missing response should be
removed from analysis}

\item{style}{Must be 1 or 2, \code{1} if only want to display rates summary
for each response value category, \code{2} if want to display rates,
difference in rate and odds ratio for each response value category.}

\item{value.rsp}{Character vector, defining list of response values to be
used as responders}

\item{value.nrsp}{Character vector, defining list of response values to be
used as non-responders}

\item{strata}{Character vector, defining stratification factor for stratified 
analysis, default is NULL}
}
\description{
Response Table with ADaM data structure
}
\details{
Package \code{forcats} was used to re-format arm data into leveled factors.
Package \code{epiR} was used to compute 2-by-2 contingency table statistics
and analysis with stratification factors.
}
\examples{
\dontrun{
library(dplyr)
library(atezo.data)
library(teal.oncology)
library(forcats)
library(epiR)

'\%needs\%' <- teal.oncology:::'\%needs\%'

ASL <- asl(com.roche.cdt30019.go29436.re)
ARS <- ars(com.roche.cdt30019.go29436.re)

teal.oncology:::response_table_ADAM(ASL, ARS,
                    paramcd = "OVRSPI",
                    arm.ref = "DUMMY C",
                    arm.comp = c("DUMMY A", "DUMMY B"),
                    arm.comp.combine = FALSE,
                    value.resp = c("CR", "PR"),
                    value.nresp = c("SD", "NON CR/PD", "PD", "NE"),
                    incl.missing = TRUE,
                    strata.var = "SEX",
                    style = 1
                    )
}

}
