% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_response_table.R
\name{response_table_adam}
\alias{response_table_adam}
\title{with ASL and ATE adam data}
\usage{
response_table_adam(ASL, ATE)
}
\arguments{
\item{ASL}{is the asl dataset}

\item{ATE}{is the ate dataset}
}
\description{
with ASL and ATE adam data
}
\examples{

\dontrun{
response_table_adam(ASL, ATE)

}
}
