% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_response_table.R
\name{response_table_ADAM}
\alias{response_table_ADAM}
\title{Response Table with ADaM data structure}
\usage{
response_table_ADAM(ASL, ARS, response_par = "BESRSPI", arm_var = "ARM",
  arm.ref, arm.comp, val_rsp = c("CR", "PR"), val_nrsp = c("SD",
  "NON CR/PD", "PD", "NE"), incl_missing = T)
}
\arguments{
\item{ASL}{ASL dataset with following variables: USUBJID, STUDYID, the specified grouping variable}

\item{ARS}{ARS dataset containing the following variables: USUBJID, STUDYID, PARAMCD, AVALC}

\item{arm_var}{Name of the variable to use as testing arms}

\item{arm.ref}{Character vector, defining which arm(s) from the list of arms should be used as reference group}

\item{arm.comp}{Character vector, defining which arm(s) from the list of arms should be used as comparison group}

\item{val_rsp}{Character vector, defining list of response values to be used as responders}

\item{val_nrsp}{Character vector, defining list of response values to be used as non-responders}

\item{incl_missing}{Boolean value, \code{TRUE} if missing values should be considered non-responders, 
\code{FALSE} if missing response should be removed from analysis}
}
\description{
Response Table with ADaM data structure
}
\examples{

\dontrun{
library(atezo.data)
'\%needs\%' <- teal.oncology:::'\%needs\%'
source("C:/Users/liaoc10/Desktop/teal.onco_miscfiles/rtable.R")

ASL <- asl(com.roche.cdt30019.go29436.re)
ARS <- ars(com.roche.cdt30019.go29436.re)

response_table_ADAM(ASL, ARS, 
                    arm.ref = "DUMMY A", 
                    arm.comp = "DUMMY B")

}

}
