% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{rsplit}
\alias{rsplit}
\title{Recursively split over variables in a data.frame}
\usage{
rsplit(x, by, total = NULL)
}
\arguments{
\item{x}{vector or data frame containing values to be divided into groups.}

\item{by}{factor or a data.frame with factors}

\item{total}{label with name of \code{x} added to the split, if \code{NULL} only splitted data is given}
}
\value{
nested list
}
\description{
Recursively split over variables in a data.frame
}
\examples{
library(magrittr)
df_by <- expand.grid(u = c("ii", "iii"), x = LETTERS[1:4], y = letters[1:3], z = letters[21:22])
x <- data.frame(num = 1:nrow(df_by))

rsplit(x, df_by[[1]]) \%>\%
   data_tree() \%>\%
   summary()

rsplit(x, df_by[1]) \%>\% data_tree() \%>\% summary()
rsplit(x, df_by[1:2]) \%>\% data_tree() \%>\% summary()
rsplit(x, df_by[1:3]) \%>\% data_tree() \%>\% summary()
rsplit(x, df_by[1:4]) \%>\% data_tree() \%>\% summary()


y <- rsplit(x, df_by[1], total = "-Overall-")

y \%>\% data_tree() \%>\% summary()

identical(y[[1]], x)

y <- rsplit(x, df_by[1:4], total = "-total-")
y \%>\% data_tree() \%>\% summary()
identical( y$`-total-`$`-total-`$`-total-`$`-total-`, x)

}
\seealso{
\code{\link{split}}
}
