% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{rsplit}
\alias{rsplit}
\title{Recursively split over variables in a data.frame}
\usage{
rsplit(x, by)
}
\arguments{
\item{x}{vector or data frame containing values to be divided into groups.}

\item{by}{factor or a data.frame with factors}
}
\value{
nested list
}
\description{
Recursively split over variables in a data.frame
}
\examples{
df_by <- expand.grid(u = c("ii", "iii"), x = LETTERS[1:4], y = letters[1:3], z = letters[21:22])
x <- data.frame(num = 1:nrow(df_by))
rsplit(x, df_by[[1]])
rsplit(x, df_by[1])
rsplit(x, df_by[1:2])
rsplit(x, df_by[1:3])
rsplit(x, df_by[1:4])
}
\seealso{
\code{\link{split}}
}
