% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{rsplit}
\alias{rsplit}
\title{Split recursively}
\usage{
rsplit(x, by_lst)
}
\arguments{
\item{x}{object to split}

\item{by_lst}{list of by objects to recursively split by}
}
\value{
nested list, first level corresponding to split by by_lst[[1]], second to recursive split by by_lst[[2]] etc.
}
\description{
This is an extension of \code{\link{esplit}} to the case when by is a list to split by recursively
}
\examples{
by_lst <- list(factor(c("M", "M", "F", "F", "F")), factor(c("O", "Y", "Y", "Y", "Y")))
tern:::rsplit(
  list(data.frame(x = 1:5, y = 6:10), data.frame(z1 = 11:15, z2 = 16:20)),
  by_lst
)
}
