% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtable.R
\name{rtable}
\alias{rtable}
\title{Creating Reporting Tables}
\usage{
rtable(col.names, format = NULL, ...)
}
\arguments{
\item{col.names}{vector with column names}

\item{format}{a valid format string or a format function for
\code{\link{rcell}}s. To get a list of all valid format strings use 
\code{\link{get_rcell_formats}}.}

\item{...}{each element is an \code{\link{rrow}} object}
}
\value{
a \code{rtable} object
}
\description{
Reporting tables allow multiple values per cell, cell formatting and mergin 
cells. Currently \code{rtable}s can be converted to html.
}
\details{
The rtable objects can be currently expported to text with to 
\code{\link{toString.rtable}} and to html with \code{\link{toHTML.rtable}}.
In future we would like to add a \code{toGridGrob} function.

Note that the formats propagate to the \code{\link{rrow}}s and 
\code{\link{rcells}} if these do not specify their own format.
}
\examples{

tbl <- rtable(
  col.names = c("Treatement\\nN=100", "Comparison\\nN=300"),
  format = "xx (xx.xx\%)",
  rrow("Response", c(104, .2), c(100, .4)),
  rrow("Non-Response", c(23, .4), c(43, .5432432423423)),
  rrow(),
  rrow("this is a very long section header"),
  rrow("HR", rcell(3.23, "xx.xx", colspan = 2)),
  rrow("95\% CI", indent = 1, rcell(c(2.432, 4.3214), format = "(xx.x, xx.x)", colspan = 2))
)

tbl

row.names(tbl)
names(tbl)


tbl[1,2]

tbl[3,2]
tbl[5,1]
tbl[5,2]


dim(tbl)
nrow(tbl)
ncol(tbl)

tbl[[1]][[1]]

as_html(tbl)

Viewer(tbl)

# colspans

tbl2 <- rtable(
  c("A", "B", "C", "D", "E"),
  format = "xx",
  rrow("r1", 1, 2, 3, 4, 5),
  rrow("r2", rcell("sp2", colspan = 2), "sp1", rcell("sp2-2", colspan = 2))
)

tbl2

Viewer(tbl2)

tbl2[1,3]
tbl2[2,1]
tbl2[2,2]
tbl2[2,3]
tbl2[2,4]
tbl2[2,5]


# custom format
my_format <- function(x, output) {
   paste(x, collapse = "/")
}
tbl3 <- rtable(
  c("A", "B"),
  format = my_format,
  rrow("row1", c(1,2,3,4), letters[1:10])
)
tbl3
 
}
