% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rtable}
\alias{rtable}
\title{Creating Reporting Tables}
\usage{
rtable(col.names, format = NULL, ...)
}
\arguments{
\item{col.names}{vector with column names}

\item{format}{a valid format string}

\item{...}{each element is an \code{\link{rrow}} object}
}
\value{
a \code{rtable} object
}
\description{
Reporting tables allow multiple values per cell, cell formatting and mergin 
cells. Currently \code{rtable}s can be converted to html.
}
\examples{

tbl <- rtable(
  col.names = c("Treatement\\nN=100", "Comparison\\nN=300"),
  format = "xx (xx.xx\%)",
  rrow("Response", c(104, .2), c(100, .4)),
  rrow("Non-Response", c(23, .4), c(43, .5)),
  rrow(),
  rrow("this is a very long section header"),
  rrow("HR", rcell(3.23, "xx.xx", colspan = 2)),
  rrow("95\% CI", indent = 2, rcell(3.23, format = "xx.x", colspan = 2))
)

tbl

row.names(tbl)
names(tbl)


tbl[1,2]

tbl[3,2]
tbl[5,1]
tbl[5,2]


dim(tbl)
nrow(tbl)
ncol(tbl)

tbl[[1]][[1]]

as_html(tbl)

Viewer(tbl)

# colspans

tbl2 <- rtable(
  c("A", "B", "C", "D", "E"),
  format = "xx",
  rrow("r1", 1, 2, 3, 4, 5),
  rrow("r2", merge_cells(2, "sp2"), "sp1", merge_cells(2, "sp2-2"))
)

Viewer(tbl2)

tbl2[1,3]
tbl2[2,1]
tbl2[2,2]
tbl2[2,3]
tbl2[2,4]
tbl2[2,5]

}
