% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rtable}
\alias{rtable}
\title{create a table}
\usage{
rtable(col.names, format = NULL, ...)
}
\arguments{
\item{col.names}{vector with column names}

\item{...}{each element is a row}
}
\description{
create a table
}
\examples{

tbl <- rtable(c("Treatement", "Comparison"),
  cell_format("xx (xx.xx\%)"),
  rrow("Response", c(104, .2), c(100, .4)),
  rrow("Non-Response", c(23, .4), c(43, .5)),
  empty_row(),
  rrow("more"),
  rrow("HR", merge_cells(2, cf(3.23, "xx.xx"))),
  rrow("95\% CI", indent = 2, merge_cells(2, cf(3.23, "xx.xx")))
)

tbl

dim(tbl)
nrow(tbl)
ncol(tbl)



as_html(tbl)

Viewer(tbl)
}
