% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rtable}
\alias{rtable}
\title{create a table}
\usage{
rtable(col.names, format = NULL, ...)
}
\arguments{
\item{col.names}{vector with column names}

\item{...}{each element is a row}
}
\description{
create a table
}
\examples{

tbl <- rtable(
  col.names = c("Treatement\\nN=100", "Comparison\\nN=300"),
  format = "xx (xx.xx\%)",
  rrow("Response", c(104, .2), c(100, .4)),
  rrow("Non-Response", c(23, .4), c(43, .5)),
  empty_row(),
  rrow("this is a very long section header"),
  rrow("HR", merge_cells(2, cf(3.23, "xx.xx"))),
  rrow("95\% CI", indent = 2, merge_cells(2, 3.23), format = "xx.x")
)

tbl

tbl[1,2]

tbl[3,2]
tbl[5,1]
tbl[5,2]


dim(tbl)
nrow(tbl)
ncol(tbl)

tbl[[1]][[1]]

as_html(tbl)

Viewer(tbl)

# colspans

tbl2 <- rtable(
  c("A", "B", "C", "D", "E"),
  format = "xx",
  rrow("r1", 1, 2, 3, 4, 5),
  rrow("r2", merge_cells(2, "sp2"), "sp1", merge_cells(2, "sp2-2"))
)

Viewer(tbl2)

tbl2[1,3]
tbl2[2,1]
tbl2[2,2]
tbl2[2,3]
tbl2[2,4]
tbl2[2,5]

}
