% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables_access.R
\name{rtables_access}
\alias{rtables_access}
\alias{h_row_counts}
\alias{h_row_fractions}
\alias{h_col_counts}
\title{rtables Access Helper Functions}
\usage{
h_row_counts(table_row, col_names)

h_row_fractions(table_row, col_names)

h_col_counts(table, col_names)
}
\arguments{
\item{table_row}{(\code{TableRow})\cr an analysis row in a occurrence table.}

\item{col_names}{(\code{character})\cr the names of the columns to extract from.}

\item{table}{(\code{VTableNodeInfo})\cr an occurrence table or row.}
}
\description{
These are a couple of functions that help with accessing the data in rtables objects.
Currently these work for occurrence tables, which are defined
as having a count as the first element and a fraction as the second element in each cell.
}
\section{Functions}{
\itemize{
\item \code{h_row_counts}: helper function to extract counts from specified columns
in a \code{TableRow}.

\item \code{h_row_fractions}: helper function to extract fractions from specified columns
in a \code{TableRow}.

\item \code{h_col_counts}: helper function to extract column counts from specified columns
in a table.
}}

\seealso{
\verb{[prune_occurrences]} for usage of these functions.
}
