% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables_access.R
\name{rtables_access}
\alias{rtables_access}
\alias{h_row_counts}
\alias{h_row_fractions}
\alias{h_col_counts}
\alias{is_leaf_table}
\alias{h_content_first_row}
\title{\code{rtables} Access Helper Functions}
\usage{
h_row_counts(
  table_row,
  col_names = NULL,
  col_indices = h_col_indices(table_row, col_names)
)

h_row_fractions(
  table_row,
  col_names = NULL,
  col_indices = h_col_indices(table_row, col_names)
)

h_col_counts(
  table,
  col_names = NULL,
  col_indices = h_col_indices(table, col_names)
)

is_leaf_table(table)

h_content_first_row(table)
}
\arguments{
\item{table_row}{(\code{TableRow})\cr an analysis row in a occurrence table.}

\item{col_names}{(\code{character})\cr the names of the columns to extract from.}

\item{col_indices}{(\code{integer})\cr the indices of the columns to extract from. If \code{col_names} are provided,
then these are inferred from the names of \code{table_row}. (Note that this currently only works well with a single
column split.)}

\item{table}{(\code{VTableNodeInfo})\cr an occurrence table or row.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These are a couple of functions that help with accessing the data in \code{rtables} objects.
Currently these work for occurrence tables, which are defined
as having a count as the first element and a fraction as the second element in each cell.
}
\section{Functions}{
\itemize{
\item \code{h_row_counts()}: helper function to extract counts from specified columns
in a \code{TableRow}.

\item \code{h_row_fractions()}: helper function to extract fractions from specified columns
in a \code{TableRow}.

\item \code{h_col_counts()}: Helper function to extract column counts from specified columns
in a table.

\item \code{is_leaf_table()}: Helper function which says whether current table is a leaf in the tree.

\item \code{h_content_first_row()}: Helper function to get first row of content table of current table.

}}
\seealso{
\verb{[prune_occurrences]} for usage of these functions.
}
\keyword{internal}
