% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{s_cox_multivariate}
\alias{s_cox_multivariate}
\title{Multivariate Cox Model - summarized results}
\usage{
s_cox_multivariate(
  formula,
  data,
  conf_level = 0.95,
  pval_method = c("wald", "likelihood"),
  ...
)
}
\arguments{
\item{formula}{A \code{formula} corresponding to the investigated \code{\link[survival:Surv]{survival model}}
including covariates.}

\item{data}{A \code{data.frame} which includes the variable in formula and covariates.}

\item{conf_level}{The level of confidence for the hazard ration interval estimations. Default is 0.95.}

\item{pval_method}{The method used for the estimation of p.values, should be one of \code{"wald"} (default) or
\code{"likelihood"}.}

\item{...}{Optional parameters passed to \code{\link[survival:coxph]{coxph()}}
\itemize{
\item \code{ties} a character string specifying the method for tie handling, one of \code{exact} (default), \code{efron}, \code{breslow}.
}}
}
\description{
Analyses based on multivariate Cox model are usually not performed for the Controlled Substance Reporting or
regulatory documents but serve exploratory purposes only (e.g., for publication). In practice, the model usually
includes only the main effects (without interaction terms). It produces the hazard ratio estimates for each of the
covariates included in the model.
The analysis follows the same principles (e.g., stratified vs. unstratified analysis and tie handling) as the
usual Cox model analysis. Since there is usually no pre-specified hypothesis testing for such analysis,
the p.values need to be interpreted with caution. (\strong{Statistical Analysis of Clinical Trials Data with R},
\verb{NEST's bookdown})
}
\details{
The output is limited to single effect terms. Work in ongoing for estimation of interaction terms
but is out of scope as defined by the  Global Data Standards Repository
(\strong{\code{GDS_Standard_TLG_Specs_Tables_2.doc}}).
}
\examples{
library(scda)
library(dplyr)

ADTTE <- synthetic_cdisc_data("latest")$adtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS") # _f: filtered
ADTTE_f <- dplyr::filter(
  ADTTE_f,
  PARAMCD == "OS"
  & SEX \%in\% c("F", "M")
  & RACE \%in\% c("ASIAN", "BLACK OR AFRICAN AMERICAN", "WHITE")
)
ADTTE_f$SEX <- droplevels(ADTTE_f$SEX)
ADTTE_f$RACE <- droplevels(ADTTE_f$RACE)
\dontrun{
s_cox_multivariate(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ (ARMCD + RACE + AGE)^2, data = ADTTE_f
)
}
}
