% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_coxreg.R
\name{s_coxph_pairwise}
\alias{s_coxph_pairwise}
\title{Pairwise model}
\usage{
s_coxph_pairwise(
  formula,
  data,
  conf_level = 0.95,
  pval_method = c("log-rank", "wald", "likelihood"),
  ...
)
}
\arguments{
\item{formula}{(\code{formula})\cr
Specifies \code{\link[survival:Surv]{survival model}}.
The arm variable needs to be wrapped in \code{\link{arm}}. The
\code{\link[survival]{strata}} special will only be used for the stratified analysis. If there is not
\code{\link[survival]{strata}} specification then the stratified analysis is omitted.}

\item{data}{(\code{data.frame})\cr
Contains all the variable that are used in \code{formula}}

\item{conf_level}{(\code{numeric} value)\cr
level for computation of the confidence intervals.
If set to \code{FALSE} no confidence intervals are printed}

\item{pval_method}{(one of (\code{"log-rank", "likelihood", "wald"}))\cr
Specifies the method used to calculate the p-value. Default value is \code{"log-rank"}.}

\item{...}{(optional)\cr
Other arguments will be passed to \code{\link[survival]{coxph}}}
}
\value{
a list of dataframes with unstratified and/or stratified analysis results
}
\description{
Run Pairwise (ARM) \code{CoxPH} model for unstratified and stratified analysis
}
\examples{
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADTTE <- radtte(cached = TRUE)
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

cox_res <-  s_coxph_pairwise(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARM) + strata(RACE),
  data = ADTTE_f,
  pval_method = "wald",  ties = "exact"
)
names(cox_res)
cox_res[[2]]$stratified

cox_res2 <- s_coxph_pairwise(
  formula =  Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARM),
  data = ADTTE_f,
  conf_level = 0.8, pval_method = "likelihood",  ties = "breslow"
)
cox_res2[[2]]$stratified
cox_res2[[2]]$unstratified
}
