% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{s_odds_ratio}
\alias{s_odds_ratio}
\title{OR: summary}
\usage{
s_odds_ratio(rsp, col_by, conf_level = 0.95, strat = NULL)
}
\arguments{
\item{rsp}{(\code{logical} vector)\cr
\code{TRUE} if subject is a responder \code{FALSE} otherwise}

\item{col_by}{(\code{factor} value)\cr for defining column group}

\item{conf_level}{confidence level of the interval.}

\item{strat}{Optional factor with one level per stratum and same length as \code{rsp}. If `NULL`,
then the unstratified analysis is computed.}
}
\value{
A \code{data.frame} with columns:
\describe{
  \item{level}{The level of the `arm` variable which is compared with the reference level.}
  \item{odds_ratio}{The odds ratio estimate.}
  \item{ci_lower}{The lower bound of the confidence interval.}
  \item{ci_upper}{The upper bound of the confidence interval.}
}
The attribute `conf_level` saves the used confidence level.
}
\description{
Summary function for odds ratio estimation.
}
\details{
This function uses either logistic regression for unstratified analyses, or conditional logistic
  regression for stratified analyses.
  The Wald confidence interval with the specified confidence level is calculated. Note that
  for stratified analyses, there is currently no implementation for conditional likelihood confidence intervals,
  therefore we don't offer the option of a likelihood confidence interval.
  Note that when `rsp` contains only responders or non-responders, then the result values will be `NA`,
  because no odds ratio estimation is possible.
}
\examples{
set.seed(1)
rsp <- sample(c(TRUE, FALSE), 100, TRUE)
trt <- sample(c("Placebo", "Treatment", "Combination"), 100, TRUE)
trt <- factor(trt, levels = c("Placebo", "Treatment", "Combination"))
strata_data <- data.frame(
  "f1" = sample(c("a", "b"), 100, TRUE),
  "f2" = sample(c("x", "y", "z"), 100, TRUE),
  stringsAsFactors = TRUE
)

s_odds_ratio(rsp, trt)
s_odds_ratio(rsp, trt, strat = interaction(strata_data))
}
