% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{s_proportion}
\alias{s_proportion}
\title{Outcomes summary}
\usage{
s_proportion(
  x,
  conf_level = 0.95,
  prop_ci_method = c("waldcc", "wald", "clopper-pearson", "wilson", "agresti-coull",
    "jeffreys")
)
}
\arguments{
\item{x}{\code{logical} vector. \code{TRUE} represents a successful outcome.}

\item{conf_level}{confidence level of the interval.}

\item{prop_ci_method}{one of (\code{"waldcc"}, \code{"wald"}, \code{"clopper-pearson"},
\code{"wilson"}, \code{agresti-coull} or \code{jeffreys}).\cr
Specifies the method used to construct the confidence interval for proportion
of successful outcomes.}
}
\value{
Named list with analysis summary:
\describe{
\item{prop}{Proportion of successful outcomes in sample.}
\item{prop_ci}{Confidence interval for \code{prop}.}
\item{label_ci}{Label for confidence interval in \code{prop_ci}.}
}
}
\description{
Summary for proportion of successful outcomes
}
\details{
The following options are supported for confidence interval for a proportion:
\itemize{
\item The \code{wald} interval follows the usual textbook definition for a single proportion
confidence interval using the normal approximation.
\item The \code{waldcc} interval is similar to \code{wald} and in addition uses the continuity
correction \code{1/2n}.
\item The \code{clopper-pearson} interval calls \code{\link[stats]{binom.test}}. Also referred to
as the 'exact' method.
\item The \code{wilson} interval calls \code{\link[stats]{prop.test}} with option \code{correct=FALSE}.
Also referred to as Wilson score interval.
\item The \code{agresti-coull} interval was created by Alan Agresti and Brent Coull and can be understood (for 95\% CI)
as adding two successes and two failures to the data, and then using the Wald formula to construct a CI.
\item The \code{jeffreys} interval is an equal-tailed interval based on the non-informative Jeffreys prior for a
binomial proportion.
}
}
\note{
Note that the \code{prop} element of the analysis summary retains the maximum likelihood estimator, also
when the \code{agresti-coull} method is used, while e.g. other packages might return instead the modified
proportion estimator (obtained by adding artificial success and failures to the data). This behavior is consistent
with the software SAS.
}
\examples{
set.seed(1)
rsp <- sample(c(TRUE, FALSE), 100, TRUE)
s_proportion(rsp)

s_proportion(rsp, conf_level = 0.9, prop_ci_method = "clopper-pearson")
}
\seealso{
\code{\link{t_el_proportion}},
\code{\link{s_proportion_diff}}, \code{\link{s_test_proportion_diff}}
}
