% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{sas_na}
\alias{sas_na}
\title{Convert empty strings and strings with only whitespaces to NA}
\usage{
sas_na(x, empty = TRUE, whitespaces = TRUE)
}
\arguments{
\item{x}{factor or character vector}

\item{empty}{boolean if true empty strings get replaced by \code{NA}}

\item{whitespaces}{boolean, if true then strings made from whitespaces only get replaced with \code{NA}}
}
\description{
SAS imports missing data as empty strings or strings with whitespaces only. This helper function can be used to
convert these values to \code{NA}s.
}
\examples{

sas_na(c("1", "", " ", "   ", "b"))
sas_na(factor(c("", " ", "b")))

is.na(sas_na(c("1", "", " ", "   ", "b")))
}
