% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_occurrences.R
\name{score_occurrences}
\alias{score_occurrences}
\title{Occurrence Table Sorting}
\usage{
score_occurrences(rtable_object)
}
\arguments{
\item{rtable_object}{(\code{rtable}) \cr rtable object name.}
}
\value{
A sorted rtable object.
}
\description{
Functions to sort occurrence tables.
}
\section{Functions}{
\itemize{
\item \code{score_occurrences}: Sorting function that works on occurrence data tabulated in rtables.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{

library(assertthat)
library(random.cdisc.data)
library(rtables)
ADSL <- radsl(cached = TRUE)
ADAE <- radae(cached = TRUE)

N_per_arm <- table(ADSL$ARM)

lyt <- basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  add_colcounts() \%>\%
  summarize_num_patients(var = "USUBJID",
                         .stats = c("unique"),
                         .labels = c("Total number of patients with at least one event")) \%>\%
  split_rows_by("AEBODSYS", child_labels = "visible", nested = FALSE)  \%>\%
  summarize_num_patients(var = "USUBJID",
                         .labels = c("Total number of patients with at least one event",
                                     "Total number of events")) \%>\%
 count_occurrences(vars="AEDECOD")

rtable_object <- build_table(lyt, ADAE, col_counts = N_per_arm) \%>\%
 prune_table()

rtable_object_sorted <- rtable_object \%>\%
sort_at_path(path =  c("AEBODSYS", "*", "AEDECOD"), scorefun = score_occurrences)

rtable_object_sorted

}
