% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helper.R
\name{stat_mean_ci}
\alias{stat_mean_ci}
\title{Mean CI for \code{ggplot2}}
\usage{
stat_mean_ci(x, conf_level = 0.95, na.rm = TRUE, n_lim = 2)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{na.rm}{(\code{flag})\cr whether \code{NA} values should be removed from \code{x} prior to analysis.}

\item{n_lim}{(\code{number})\cr threshold number of non-missing \code{x} to estimate
the confidence interval for mean.}
}
\description{
Convenient function for the addition of mean confidence intervals as
error bars.
}
\examples{
require(ggplot2)
d <- ggplot(mtcars, aes(cyl, mpg)) + geom_point()
d + stat_summary(
  fun.data = stat_mean_ci, geom = "errorbar",
  aes(group = 1, linetype = "mean")
)

d + stat_summary(
  fun.data = function(x) stat_mean_ci(x, conf_level = 0.5),
  geom = "errorbar", aes(group = 1, linetype = "mean")
)

}
