% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_mean_ci}
\alias{stat_mean_ci}
\title{Mean CI for \code{ggplot2}}
\usage{
stat_mean_ci(x, conf_level = 0.95, na.rm = TRUE, n_min = 2, gg_helper = TRUE)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{na.rm}{(\code{flag})\cr whether \code{NA} values should be removed from \code{x} prior to analysis.}

\item{n_min}{(\code{number})\cr a minimum number of non-missing \code{x} to estimate
the confidence interval for mean.}

\item{gg_helper}{(\code{logical})\cr \code{TRUE} when output should be aligned
for the use with ggplot.}
}
\description{
Convenient function for calculating the mean confidence interval.
It is dedicated to use as a ggplot helper function for plotting
confidence intervals as error bars.
}
\examples{

stat_mean_ci(sample(10), gg_helper = FALSE)

require(ggplot2)

p <- ggplot(mtcars, aes(cyl, mpg)) + geom_point()

p + stat_summary(
  fun.data = stat_mean_ci,
  geom = "errorbar"
)

p + stat_summary(
  fun.data = stat_mean_ci,
  fun.args = list(conf_level = 0.5)
  geom = "errorbar"
)

}
