% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helper.R
\name{stat_median_ci}
\alias{stat_median_ci}
\title{Median CI for ggplot2}
\usage{
stat_median_ci(x, conf_level = 0.95, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of values.}

\item{conf_level}{Level of confidence for the interval (aka 1 - alpha)}

\item{na.rm}{Remove the missing data.}
}
\description{
Convenient function for the addition of the median confidence intervals as
error bars.
}
\details{
The function was adapted from \verb{DescTools/versions/0.99.35/source}
}
\examples{
require(ggplot2)
d <- ggplot(mtcars, aes(cyl, mpg)) + geom_point()
d + stat_summary(
  fun.data = stat_median_ci, geom = "errorbar",
  aes(group = 1, linetype = "median")
)
}
