% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ancova.R
\name{summarize_ancova}
\alias{summarize_ancova}
\alias{s_ancova}
\alias{a_ancova}
\title{Summary for analysis of covariance (ANCOVA).}
\usage{
s_ancova(
  df,
  .var,
  .df_row,
  variables,
  .ref_group,
  .in_ref_col,
  conf_level,
  interaction_y = FALSE,
  interaction_item = NULL
)

a_ancova(
  df,
  .var,
  .df_row,
  variables,
  .ref_group,
  .in_ref_col,
  conf_level,
  interaction_y = FALSE,
  interaction_item = NULL
)

summarize_ancova(
  lyt,
  vars,
  var_labels,
  ...,
  show_labels = "visible",
  table_names = vars,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL,
  interaction_y = FALSE,
  interaction_item = NULL
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.df_row}{(\verb{data frame})\cr data frame across all of the columns for the given row split.}

\item{variables}{(named \code{list} of \code{string})\cr list of additional analysis variables.}

\item{.ref_group}{(\verb{data frame} or \code{vector})\cr the data corresponding to the reference group.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{interaction_y}{(\code{character})\cr a selected item inside of the interaction_item column
which will be used to select the specific ANCOVA results. if the interaction is not
needed, the default option is FALSE}

\item{interaction_item}{(\code{character})\cr name of the variable that should have interactions
with arm. if the interaction is not needed, the default option is NULL}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{var_labels}{character for label.}

\item{...}{additional arguments for the lower level functions.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
A named list of 5 statistics:
\itemize{
\item \code{n}: count of complete sample size for the group.
\item \code{lsmean}: estimated marginal means in the group.
\item \code{lsmean_diff}: difference in estimated marginal means in comparison to the reference
group. If working with the reference group, this will be empty.
\item \code{lsmean_diff_ci}: confidence level for difference in estimated marginal means in
comparison to the reference group.
\item \code{pval}: p-value (not adjusted for multiple comparisons).
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Summarize results of ANCOVA. This can be used to analyze multiple endpoints and/or
multiple timepoints within the same response variable \code{.var}.
}
\section{Functions}{
\itemize{
\item \code{s_ancova()}: Statistics function that produces a named list of results
of the investigated linear model.

\item \code{a_ancova()}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{summarize_ancova()}: Layout creating function which can be be used for creating
summary tables for analysis of covariance (ANCOVA).

}}
\examples{
library(dplyr)

df <- iris \%>\% filter(Species == "virginica")
.df_row <- iris
.var <- "Petal.Length"
variables <- list(arm = "Species", covariates = "Sepal.Length * Sepal.Width")
.ref_group <- iris \%>\% filter(Species == "setosa")
conf_level <- 0.95

# Internal function - s_ancova
\dontrun{
s_ancova(
  df, .var, .df_row, variables, .ref_group,
  .in_ref_col = FALSE,
  conf_level, interaction_y = FALSE, interaction_item = NULL
)
}

# Internal function - a_ancova
\dontrun{
a_ancova(
  df, .var, .df_row, variables, .ref_group,
  .in_ref_col = FALSE,
  interaction_y = FALSE, interaction_item = NULL, conf_level
)
}

basic_table() \%>\%
  split_cols_by("Species", ref_group = "setosa") \%>\%
  add_colcounts() \%>\%
  summarize_ancova(
    vars = "Petal.Length",
    variables = list(arm = "Species", covariates = NULL),
    table_names = "unadj",
    conf_level = 0.95, var_labels = "Unadjusted comparison",
    .labels = c(lsmean = "Mean", lsmean_diff = "Difference in Means")
  ) \%>\%
  summarize_ancova(
    vars = "Petal.Length",
    variables = list(arm = "Species", covariates = c("Sepal.Length", "Sepal.Width")),
    table_names = "adj",
    conf_level = 0.95, var_labels = "Adjusted comparison (covariates: Sepal.Length and Sepal.Width)"
  ) \%>\%
  build_table(iris)
}
\keyword{internal}
