% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ancova.R
\name{summarize_ancova}
\alias{summarize_ancova}
\alias{h_ancova}
\alias{s_ancova}
\alias{a_ancova}
\title{Summary for analysis of covariance (ANCOVA).}
\usage{
h_ancova(.var, .df_row, variables)

s_ancova(df, .var, .df_row, variables, .ref_group, .in_ref_col, conf_level)

a_ancova(df, .var, .df_row, variables, .ref_group, .in_ref_col, conf_level)

summarize_ancova(
  lyt,
  vars,
  var_labels,
  ...,
  show_labels = "visible",
  table_names = vars,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.df_row}{(\verb{data frame})\cr data set that includes all the variables that are called
in \code{.var} and \code{variables}.}

\item{variables}{(named \code{list} of \code{strings})\cr list of additional analysis variables, with
expected elements:
\itemize{
\item \code{arm}: (\code{string})\cr group variable, for which the covariate adjusted means of multiple
groups will be summarized. Specifically, the first level of \code{arm} variable is taken as the
reference group.
\item \code{covariates}: (\code{character})\cr a vector that can contain single variable names (such as
\code{"X1"}), and/or interaction terms indicated by \code{"X1 * X2"}.
}}

\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.ref_group}{(\verb{data frame} or \code{vector})\cr the data corresponding to the reference group.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{var_labels}{character for label.}

\item{...}{additional arguments for the lower level functions.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
A named list of 5 statistics:
\itemize{
\item \code{n}: count of complete sample size for the group.
\item \code{lsmean}: estimated marginal means in the group.
\item \code{lsmean_diff}: difference in estimated marginal means in comparison to the reference
group. If working with the reference group, this will be empty.
\item \code{lsmean_diff_ci}: confidence level for difference in estimated marginal means in
comparison to the reference group.
\item \code{pval}: p-value (not adjusted for multiple comparisons).
}
}
\description{
Summarize results of ANCOVA. This can be used to analyze multiple endpoints and/or
multiple timepoints within the same response variable \code{.var}.
}
\section{Functions}{
\itemize{
\item \code{h_ancova}: Helper function to return results of a linear model.

\item \code{s_ancova}: Statistics function that produces a named list of results
of the investigated linear model.

\item \code{a_ancova}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{summarize_ancova}: Layout creating function which can be be used for creating
summary tables for analysis of covariance (ANCOVA).
}}

\examples{
h_ancova(
  .var = "Sepal.Length",
  .df_row = iris,
  variables = list(arm = "Species", covariates = c("Petal.Length * Petal.Width", "Sepal.Width"))
)

library(random.cdisc.data)
library(dplyr)
adsl <- radsl(cached = TRUE)
adqs <- radqs(cached = TRUE)
adqs_single <- adqs \%>\%
  filter(
    AVISIT == "WEEK 1 DAY 8",  # single time point
    PARAMCD == "FKSI-FWB"  # single end point
  ) \%>\%
  mutate(CHG = ifelse(BMEASIFL == "Y", CHG, NA))  # only analyze evaluable population

df <- adqs_single \%>\% dplyr::filter(ARMCD == "ARM B")
.var <- "CHG"
.df_row <- adqs_single
variables <- list(arm = "ARMCD", covariates = "SEX * AGE")
.ref_group <- adqs_single \%>\% dplyr::filter(ARMCD == "ARM A")
conf_level <- 0.95
s_ancova(df, .var, .df_row, variables, .ref_group, .in_ref_col = FALSE, conf_level)

a_ancova(df, .var, .df_row, variables, .ref_group, .in_ref_col = FALSE, conf_level)

library(random.cdisc.data)
library(dplyr)
adsl <- radsl(cached = TRUE)
adqs <- radqs(cached = TRUE)

adqs_single <- adqs \%>\%
  filter(
    AVISIT == "WEEK 1 DAY 8",  # single time point
    PARAMCD == "FKSI-FWB"  # single end point
  ) \%>\%
  mutate(CHG = ifelse(BMEASIFL == "Y", CHG, NA))  # only analyze evaluable population
adqs_multi <- filter(adqs, AVISIT == "WEEK 1 DAY 8")
n_per_arm <- table(adsl$ARM)

basic_table() \%>\%
  split_cols_by("ARMCD", ref_group = "ARM A") \%>\%
  add_colcounts() \%>\%
  summarize_ancova(
    vars = "CHG",
    variables = list(arm = "ARMCD", covariates = NULL),
    conf_level = 0.95, var_labels = "Unadjusted comparison",
    .labels = c(lsmean = "Mean", lsmean_diff = "Difference in Means")
  ) \%>\%
  summarize_ancova(
    vars = "CHG",
    variables = list(arm = "ARMCD", covariates = c("BASE", "STRATA1")),
    conf_level = 0.95, var_labels = "Adjusted comparison (covariates BASE and STRATA1)"
  ) \%>\%
  build_table(adqs_single, col_counts = n_per_arm)

basic_table() \%>\%
  split_cols_by("ARMCD", ref_group = "ARM A") \%>\%
  split_rows_by("PARAMCD") \%>\%
  summarize_ancova(
    vars = "CHG",
    variables = list(arm = "ARMCD", covariates = c("BASE", "STRATA1")),
    conf_level = 0.95, var_labels = "Adjusted mean"
  ) \%>\%
  build_table(adqs_multi, col_counts = n_per_arm)

}
