% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_colvars.R
\name{summarize_colvars}
\alias{summarize_colvars}
\title{Summarize Variables in Columns}
\usage{
summarize_colvars(
  lyt,
  .stats = c("n", "mean_sd", "median", "range", "count_fraction"),
  ...
)
}
\arguments{
\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{...}{arguments passed to \code{s_summary()}.}
}
\description{
This Analyze Function uses the new S3 generic function \code{\link[=s_summary]{s_summary()}} to summarize
different variables that are arranged in columns.
Additional standard formatting arguments are available.
}
\note{
Currently a warning is thrown when building the table. This will be fixed in an upcoming
new rtables version. Similarly, it is not yet possible to select statistics with \code{.stats}, which
leads to an error.
}
\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{

dta_test <- data.frame(
  USUBJID = rep(1:6, each = 3),
  PARAMCD = rep("lab", 6*3),
  AVISIT = rep(paste0("V", 1:3), 6),
  ARM = rep(LETTERS[1:3], rep(6, 3)),
  AVAL = c(9:1, rep(NA, 9)),
  CHG = c(1:9, rep(NA, 9))
)

## Default output within a `rtables` pipeline.
l <- basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  split_rows_by("AVISIT") \%>\%
  split_cols_by_multivar(vars = c("AVAL", "CHG")) \%>\%
  summarize_colvars()
tab <- testthat::expect_warning(build_table(l, dta_test))
tab

## Selection of statistics does not work yet, but formats and labels do.
l <- basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  split_rows_by("AVISIT") \%>\%
  split_cols_by_multivar(vars = c("AVAL", "CHG")) \%>\%
  summarize_colvars(
    # .stats = c("n", "mean_sd")  # not possible yet
    .formats = c("mean_sd" = "xx.x, xx.x"),
    .labels = c(n = "n", mean_sd = "Mean, SD")
  )
tab <- testthat::expect_warning(build_table(l, dta_test))
tab

## Use arguments interpreted by `s_summary`.
l <- basic_table() \%>\%
  split_cols_by("ARM") \%>\%
  split_rows_by("AVISIT") \%>\%
  split_cols_by_multivar(vars = c("AVAL", "CHG")) \%>\%
  summarize_colvars(na.rm = FALSE)
tab <- testthat::expect_warning(build_table(l, dta_test))
tab

}
\seealso{
\code{\link[rtables:split_cols_by_multivar]{rtables::split_cols_by_multivar()}}
}
