% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_glm_count.R
\name{summarize_glm_count}
\alias{summarize_glm_count}
\alias{h_glm_poisson}
\alias{h_glm_quasipoisson}
\alias{h_glm_count}
\alias{h_ppmeans}
\title{Summary for Poisson Negative Binomial.}
\usage{
h_glm_poisson(.var, .df_row, variables, weights)

h_glm_quasipoisson(.var, .df_row, variables, weights)

h_glm_count(.var, .df_row, variables, distribution, weights)

h_ppmeans(obj, .df_row, arm, conf_level)
}
\arguments{
\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.df_row}{(\verb{data frame})\cr data set that includes all the variables that are called
in \code{.var} and \code{variables}.}

\item{variables}{(named \code{list} of \code{strings})\cr list of additional analysis variables, with
expected elements:
\itemize{
\item \code{arm}: (\code{string})\cr group variable, for which the covariate adjusted means of multiple
groups will be summarized. Specifically, the first level of \code{arm} variable is taken as the
reference group.
\item \code{covariates}: (\code{character})\cr a vector that can contain single variable names (such as
\code{"X1"}), and/or interaction terms indicated by \code{"X1 * X2"}.
\item \code{offset}: (\code{numeric})\cr a numeric vector or scalar adding an offset to the predictions.
}}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{`weights`(`numeric`)\cr}{a character value or numeric vector specifying weights used in averaging predictions.}

\item{`distribution`(`character`)\cr}{a character value specifying the distribution used in the regression.}
}
\description{
Summarize results of a Poisson Negative Binomial Regression. This can be used to analyze multiple endpoints and/or
multiple timepoints within the same response variable \code{.var}.
}
\section{Functions}{
\itemize{
\item \code{h_glm_poisson}: Helper function to return results of a poisson model.

\item \code{h_glm_quasipoisson}: Helper function to return results of a quasipoisson model.

\item \code{h_glm_count}: Helper function to return the results of the selected model (poisson, quasipossion, negative binomial).

\item \code{h_ppmeans}: Helper function to return the results of the selected model (poisson, quasipossion, negative binomial).
}}

\examples{
h_glm_poisson(
  .var = "AVAL",
  .df_row = an1,
  variables = list(arm = "arm", offset = "lnTMATRSK", covariates = NULL)
)
h_glm_quasipoisson(
  .var = "AVAL",
  .df_row = ##INSERT DF HERE,
  variables = list(arm = "TRT02P", offset = "lnTMATRSK", covariates = NULL)
)
}
