% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_glm_count.R
\name{summarize_glm_count}
\alias{summarize_glm_count}
\alias{s_glm_count}
\alias{a_glm_count}
\title{Summary for Poisson Negative Binomial.}
\usage{
s_glm_count(
  df,
  .var,
  .df_row,
  variables,
  .ref_group,
  .in_ref_col,
  distribution,
  conf_level,
  rate_mean_method,
  weights,
  scale = 1
)

a_glm_count(
  df,
  .var,
  .df_row,
  variables,
  .ref_group,
  .in_ref_col,
  distribution,
  conf_level,
  rate_mean_method,
  weights,
  scale = 1
)

summarize_glm_count(
  lyt,
  vars,
  var_labels,
  ...,
  show_labels = "visible",
  table_names = vars,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.df_row}{(\verb{data frame})\cr data frame across all of the columns for the given row split.}

\item{variables}{(named \code{list} of \code{string})\cr list of additional analysis variables.}

\item{.ref_group}{(\verb{data frame} or \code{vector})\cr the data corresponding to the reference group.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{var_labels}{character for label.}

\item{...}{additional arguments for the lower level functions.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
A named list of 5 statistics:
\itemize{
\item \code{n}: count of complete sample size for the group.
\item \code{rate}: estimated event rate per follow-up time.
\item \code{rate_ci}: confidence level for estimated rate per follow-up time.
\item \code{rate_ratio}: Ratio of event rates in each treatment arm to the reference arm.
\item \code{rate_ratio_ci}: confidence level for the rate ratio.
\item \code{pval}: p-value.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Summarize results of a Poisson Negative Binomial Regression.
This can be used to analyze count and/or frequency data using a linear model.
}
\section{Functions}{
\itemize{
\item \code{s_glm_count()}: Statistics function that produces a named list of results
of the investigated poisson model.

\item \code{a_glm_count()}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{summarize_glm_count()}: Layout creating function which can be be used for creating
summary tables for analysis of count data using generalized linear models (poisson, quasipoisson).

}}
\examples{

# Internal function - s_change_from_baseline
\dontrun{
s_glm_count(
  df = anl \%>\%
    filter(ARMCD == "ARM B"),
  .df_row = anl,
  .var = "AVAL",
  .in_ref_col = TRUE,
  variables = list(arm = "ARMCD", offset = "lgTMATRSK", covariates = c("REGION1")),
  conf_level = 0.95,
  distribution = "quasipoisson",
  rate_mean_method = "ppmeans"
)
}


# Internal function - s_change_from_baseline
\dontrun{
a_glm_count(
  df = anl \%>\%
    filter(ARMCD == "ARM A"),
  .var = "AVAL",
  .df_row = anl,
  variables = list(arm = "ARMCD", offset = "lgTMATRSK", covariates = c("REGION1")),
  .ref_group = "ARM B", .in_ref_col = TRUE,
  conf_level = 0.95,
  distribution = "poisson",
  rate_mean_method = "ppmeans"
)
}


library(scda)
library(dplyr)
anl <- synthetic_cdisc_dataset("latest", "adtte") \%>\%
  filter(PARAMCD == "TNE")
anl$AVAL_f <- as.factor(anl$AVAL)

lyt <- basic_table() \%>\%
  split_cols_by("ARM", ref_group = "B: Placebo") \%>\%
  add_colcounts() \%>\%
  summarize_vars(
    "AVAL_f",
    var_labels = "Number of exacerbations per patient",
    .stats = c("count_fraction"),
    .formats = c("count_fraction" = "xx (xx.xx\%)"),
    .label = c("Number of exacerbations per patient")
  ) \%>\%
  summarize_glm_count(
    vars = "AVAL",
    variables = list(arm = "ARM", offset = "lgTMATRSK", covariates = NULL),
    conf_level = 0.95,
    distribution = "poisson",
    rate_mean_method = "emmeans",
    var_labels = "Unadjusted exacerbation rate (per year)",
    table_names = "unadj",
    .stats = c("rate"),
    .labels = c(rate = "Rate")
  ) \%>\%
  summarize_glm_count(
    vars = "AVAL",
    variables = list(arm = "ARM", offset = "lgTMATRSK", covariates = c("REGION1")),
    conf_level = 0.95,
    distribution = "quasipoisson",
    rate_mean_method = "ppmeans",
    var_labels = "Adjusted (QP) exacerbation rate (per year)",
    table_names = "adj",
    .stats = c("rate", "rate_ci", "rate_ratio", "rate_ratio_ci", "pval"),
    .labels = c(
      rate = "Rate", rate_ci = "Rate CI", rate_ratio = "Rate Ratio",
      rate_ratio_ci = "Rate Ratio CI", pval = "p value"
    )
  )
build_table(lyt = lyt, df = anl)
}
\keyword{internal}
