% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_num_patients.R
\name{summarize_num_patients}
\alias{summarize_num_patients}
\alias{s_num_patients}
\alias{s_num_patients_content}
\title{Number of patients}
\usage{
s_num_patients(x, .N_col)

s_num_patients_content(df, labelstr = "", .N_col, .var)

summarize_num_patients(
  lyt,
  var,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL
)
}
\arguments{
\item{x}{(\code{character} or \code{factor}) \cr vector of patient IDs.}

\item{.N_col}{(\code{count})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is passed by \code{rtables}.}

\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions).}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}
}
\value{
A list with:
\itemize{
\item \code{unique}: vector of count and percentage.
\item \code{nonunique} : vector of count.
}
}
\description{
Count the number of unique and non-unique patients in a column (variable).
}
\section{Functions}{
\itemize{
\item \code{s_num_patients}: Statistics function which counts the number of
unique patients, the corresponding percentage taken with respect to the
total number of patients, and the number of non-unique patients.

\item \code{s_num_patients_content}: Counts the number of unique patients in a column
(variable), the corresponding percentage taken with respect to the total
number of patients, and the number of non-unique patients in the column.
Function serves as a wrapper that carries over both expected arguments \code{df}
and \code{labelstr} in \code{cfun} of \code{\link[=summarize_row_groups]{summarize_row_groups()}}.

\item \code{summarize_num_patients}: Layout creating function which adds content rows using the statistics
function \code{\link[=s_num_patients_content]{s_num_patients_content()}} and desired format.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{
# Use the statistics function to count number of unique and nonunique patients.
s_num_patients(x = as.character(c(1, 2, 1, 4, NA)), .N_col = 5L)
# Count number of unique and non-unique patients.
df <- data.frame(
  USUBJID = as.character(c(1, 2, 1, 4, NA)),
  AGE = c(10, 15, 10, 17, 8)
)
s_num_patients_content(df, .N_col = 5, .var = "USUBJID")

}
