% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variables_numeric_in_cols.R
\name{summarize_numeric_in_columns}
\alias{summarize_numeric_in_columns}
\alias{summary_numeric_in_cols}
\alias{summarize_vars_numeric_in_cols}
\title{Summary numeric variables in columns}
\usage{
summary_numeric_in_cols(x, labelstr = "", custom_label = NULL, ...)

summarize_vars_numeric_in_cols(
  lyt,
  var,
  ...,
  .stats = c("n", "mean", "sd", "se", "cv", "geom_cv"),
  .labels = c(n = "n", mean = "Mean", sd = "SD", se = "SE", cv = "CV (\%)", geom_cv =
    "CV \% Geometric Mean"),
  .indent_mods = NULL,
  col_split = TRUE
)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions).}

\item{custom_label}{(\code{string} or \code{NULL})\cr if provided and \code{labelstr} is
empty then this will be used as the row label.}

\item{...}{additional arguments for the lower level functions.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{var}{(\code{string})\cr single variable name for the primary analysis variable.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}

\item{col_split}{(\code{flag})\cr whether the columns should be split.}
}
\value{
A named list of all statistics returned by \code{\link[=s_summary.numeric]{s_summary.numeric()}}.
See \code{\link[=s_summary.numeric]{s_summary.numeric()}} to be aware of all available statistics.
}
\description{
These functions can be used to produce summary tables for PK datasets.
}
\section{Functions}{
\itemize{
\item \code{summary_numeric_in_cols}: a wrapper of \code{\link[=s_summary.numeric]{s_summary.numeric()}}
function that produces a named list of statistics to include as columns.

\item \code{summarize_vars_numeric_in_cols}: Layout creating function which can be used for creating
summary tables in columns, primarily used for PK data sets.
}}

\examples{

library(scda)
ADPC <- scda::synthetic_cdisc_data("latest")$adpc
summary_numeric_in_cols(ADPC$AGE, custom_label = "stats")


lyt <- basic_table() \%>\%
  split_rows_by(var = "ARM") \%>\%
  split_rows_by(var = "SEX") \%>\%
  summarize_vars_numeric_in_cols(var = "AGE", col_split = TRUE)
result <- build_table(lyt = lyt, df = ADPC)
result

# By selecting just some statistics and ad-hoc labels
lyt <- basic_table() \%>\%
  split_rows_by(var = "ARM") \%>\%
  split_rows_by(var = "SEX") \%>\%
  summarize_vars_numeric_in_cols(
  var = "AGE",
  .stats = c("n", "cv", "geom_mean", "mean_ci"),
  .labels = c(n = "myN", cv = "myCV", geom_mean = "myGeomMean", mean_ci = "Mean (95\%CI)"),
  col_split = TRUE)
result <- build_table(lyt = lyt, df = ADPC)
result

lyt <- basic_table() \%>\%
  summarize_vars_numeric_in_cols(
  var = "AGE",
  col_split = TRUE,
  custom_label = "some custom label")
  result <- build_table(lyt, df = ADPC)
result

# PKPT03
ADPC$PARAMUNIT <- paste0(ADPC$PARAM, "(", ADPC$AVALU, ")")
lyt <- basic_table() \%>\%
  split_rows_by(var = "PARAMUNIT") \%>\%
  summarize_vars_numeric_in_cols(var = "AVAL", col_split = TRUE)
  result <- build_table(lyt, df = ADPC)
result

# PKCT01FDS
ADPC$NRELTM1 <- as.factor(ADPC$NRELTM1)
lyt <- basic_table() \%>\%
  split_rows_by(var = "VISIT") \%>\%
  split_rows_by(var = "NRELTM1") \%>\%
  summarize_vars_numeric_in_cols(var = "AVAL", col_split = TRUE)
  result <- build_table(lyt, df = ADPC)
result

}
\seealso{
\link{summarize_vars}.
}
