% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_patients_exposure_in_cols.R
\name{summarize_patients_exposure_in_cols}
\alias{summarize_patients_exposure_in_cols}
\alias{s_count_patients_sum_exposure}
\title{Counting Patients Summing Exposure Across All Patients in Columns}
\usage{
s_count_patients_sum_exposure(
  df,
  .var = "AVAL",
  id = "USUBJID",
  labelstr = "",
  .N_col,
  custom_label = NULL
)

summarize_patients_exposure_in_cols(
  lyt,
  var,
  ...,
  .stats = c("n_patients", "sum_exposure"),
  .labels = c(n_patients = "Patients", sum_exposure = "Person time"),
  .indent_mods = NULL,
  col_split = TRUE
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{id}{(\code{string}) \cr subject variable name.}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions).}

\item{.N_col}{(\code{count})\cr row-wise N (row group count) for the group of observations being analyzed
(i.e. with no column-based subsetting) that is passed by \code{rtables}.}

\item{custom_label}{(\code{string} or \code{NULL})\cr if provided and \code{labelstr} is empty then this will
be used as label.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{var}{(\code{string})\cr single variable name for the primary analysis variable.}

\item{...}{additional arguments for the lower level functions.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}

\item{col_split}{(\code{flag})\cr whether the columns should be split.
Set to \code{FALSE} when the required column split has been done already earlier in the layout pipe.}
}
\value{
\code{\link[=s_count_patients_sum_exposure]{s_count_patients_sum_exposure()}} returns a list with the statistics:\cr
\itemize{
\item \code{n_patients}: number of unique patients in \code{df}.
\item \code{sum_exposure}: sum of \code{.var} across all patients in \code{df}.
}
}
\description{
Counting the number of patients and summing analysis value (i.e exposure values) across all patients
when a column table layout is required.
}
\section{Functions}{
\itemize{
\item \code{s_count_patients_sum_exposure}: Statistics function which counts numbers
of patients and the sum of exposure across all patients.

\item \code{summarize_patients_exposure_in_cols}: Layout creating function which adds the count
statistics of patients and the sum of analysis value in the column layout as content rows.
}}

\examples{
set.seed(1)
df <- data.frame(
  USUBJID = c(paste("id", seq(1, 12), sep = "")),
  ARMCD = c(rep("ARM A", 6), rep("ARM B", 6)),
  SEX = c(rep("Female", 6), rep("Male", 6)),
  AVAL = as.numeric(sample(seq(1, 20), 12)),
  stringsAsFactors = TRUE
)
adsl <- data.frame(
  USUBJID = c(paste("id", seq(1, 12), sep = "")),
  ARMCD = c(rep("ARM A", 2), rep("ARM B", 2)),
  SEX = c(rep("Female", 2), rep("Male", 2)),
  stringsAsFactors = TRUE
)

s_count_patients_sum_exposure(df = df, .N_col = nrow(adsl))
s_count_patients_sum_exposure(
  df = df,
  .N_col = nrow(adsl),
  custom_label = "some user's custom label"
)

lyt <- basic_table() \%>\%
  split_cols_by("ARMCD", split_fun = add_overall_level("Total", first = FALSE)) \%>\%
  summarize_patients_exposure_in_cols(var = "AVAL", col_split = TRUE) \%>\%
  split_rows_by("SEX") \%>\%
  summarize_patients_exposure_in_cols(var = "AVAL", col_split = FALSE)
result <- build_table(lyt, df = df, alt_counts_df = adsl)
result

lyt2 <- basic_table() \%>\%
  split_cols_by("ARMCD", split_fun = add_overall_level("Total", first = FALSE)) \%>\%
  summarize_patients_exposure_in_cols(
    var = "AVAL", col_split = TRUE,
    .stats = "n_patients", custom_label = "some custom label"
  ) \%>\%
  split_rows_by("SEX") \%>\%
  summarize_patients_exposure_in_cols(var = "AVAL", col_split = FALSE)
result2 <- build_table(lyt2, df = df, alt_counts_df = adsl)
result2

lyt3 <- basic_table() \%>\%
  summarize_patients_exposure_in_cols(var = "AVAL", col_split = TRUE)
result3 <- build_table(lyt3, df = df, alt_counts_df = adsl)
result3

lyt4 <- basic_table() \%>\%
  summarize_patients_exposure_in_cols(var = "AVAL", col_split = TRUE, .stats = "sum_exposure")
result4 <- build_table(lyt4, df = df, alt_counts_df = adsl)
result4
}
