% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pk.R
\name{summarize_pk_in_cols}
\alias{summarize_pk_in_cols}
\title{Title}
\usage{
summarize_pk_in_cols(
  lyt,
  var,
  ...,
  .stats = c("n", "mean_sd"),
  .labels = c(n = "n", mean_sd = "mean (sd)"),
  .indent_mods = NULL,
  col_split = TRUE
)
}
\arguments{
\item{col_split}{}
}
\value{

}
\description{
Title
}
