% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variables_in_cols.R
\name{summary_in_cols}
\alias{summary_in_cols}
\alias{summarize_variables_in_columns}
\alias{summary_in_cols.numeric}
\alias{summary_in_cols.factor}
\alias{summary_in_cols.character}
\alias{summarize_vars_in_cols}
\title{Summary numeric variables in columns}
\usage{
summary_in_cols(x, labelstr = "", custom_label = NULL)

\method{summary_in_cols}{numeric}(x, labelstr = "", custom_label = NULL, ...)

\method{summary_in_cols}{factor}(x, labelstr = "", custom_label = NULL, ...)

\method{summary_in_cols}{character}(x, labelstr = "", custom_label = NULL, ...)

summarize_vars_in_cols(
  lyt,
  var,
  var_type = "numeric",
  ...,
  .stats = c("n", "mean", "sd", "se", "cv", "geom_cv"),
  .labels = c(n = "n", mean = "Mean", sd = "SD", se = "SE", cv = "CV (\%)", geom_cv =
    "CV \% Geometric Mean"),
  .formats = NULL,
  .indent_mods = NULL,
  col_split = TRUE
)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers we want to analyze.}

\item{labelstr}{(\code{character})\cr label of the level of the parent split currently being summarized
(must be present as second argument in Content Row Functions).}

\item{custom_label}{(\code{string} or \code{NULL})\cr if provided and \code{labelstr} is
empty then this will be used as the row label.}

\item{...}{additional arguments for the lower level functions.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{var}{(\code{string})\cr single variable name for the primary analysis variable.}

\item{var_type}{Variable type, "numeric" by default.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}

\item{col_split}{(\code{flag})\cr whether the columns should be split.}
}
\value{
A named list of all statistics returned by \code{\link[=s_summary.numeric]{s_summary.numeric()}}.
See \code{\link[=s_summary.numeric]{s_summary.numeric()}} to be aware of all available statistics.

A named list of all statistics returned by \code{\link[=s_summary.factor]{s_summary.factor()}}.
See \code{\link[=s_summary.factor]{s_summary.factor()}} to be aware of all available statistics.

A named list of all statistics returned by \code{\link[=s_summary.character]{s_summary.character()}}.
See \code{\link[=s_summary.character]{s_summary.character()}} to be aware of all available statistics.
}
\description{
These functions can be used to produce summary tables for PK datasets.
}
\section{Functions}{
\itemize{
\item \code{summary_in_cols}: \code{summary_in_cols} is a S3 generic
function to produce an object description.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\item \code{summary_in_cols.numeric}: a wrapper of \code{\link[=s_summary.numeric]{s_summary.numeric()}}
function that produces a named list of statistics to include as columns.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\item \code{summary_in_cols.factor}: a wrapper of
\code{\link[=s_summary.factor]{s_summary.factor()}} function that produces a named list of statistics to include as columns.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\item \code{summary_in_cols.character}: a wrapper of \code{\link[=s_summary.character]{s_summary.character()}}
function that produces a named list of statistics to include as columns.

\item \code{summarize_vars_in_cols}: Layout creating
function which can be used for creating summary tables in columns, primarily used for PK data sets.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}}

\examples{

library(scda)
library(dplyr)
ADPP <- scda::synthetic_cdisc_data("latest")$adpp \%>\% h_pkparam_sort()
summary_in_cols(ADPP$AGE, custom_label = "stats")

lyt <- basic_table() \%>\%
  split_rows_by(var = "ARM", label_pos = "topleft") \%>\%
  split_rows_by(var = "SEX", label_pos = "topleft") \%>\%
  summarize_vars_in_cols(var = "AGE", col_split = TRUE)
result <- build_table(lyt = lyt, df = ADPP)
result

# By selecting just some statistics and ad-hoc labels
lyt <- basic_table() \%>\%
  split_rows_by(var = "ARM", label_pos = "topleft") \%>\%
  split_rows_by(var = "SEX", label_pos = "topleft") \%>\%
  summarize_vars_in_cols(
    var = "AGE",
    .stats = c("n", "cv", "geom_mean", "mean_ci", "median", "min", "max"),
    .labels = c(
      n = "myN",
      cv = "myCV",
      geom_mean = "myGeomMean",
      mean_ci = "Mean (95\%CI)",
      median = "Median",
      min = "Minimum",
      max = "Maximum"
    ),
    col_split = TRUE
  )
result <- build_table(lyt = lyt, df = ADPP)
result

lyt <- basic_table() \%>\%
  summarize_vars_in_cols(
    var = "AGE",
    col_split = TRUE,
    custom_label = "some custom label"
  )
result <- build_table(lyt, df = ADPP)
result

# PKPT03
lyt <- basic_table() \%>\%
  split_rows_by(var = "TLG_DISPLAY", split_label = "PK Parameter", label_pos = "topleft") \%>\%
  summarize_vars_in_cols(
    var = "AVAL",
    col_split = TRUE,
    .stats = c("n", "mean", "sd", "cv", "geom_mean", "geom_cv", "median", "min", "max"),
    .labels = c(
      n = "n",
      mean = "Mean",
      sd = "SD",
      cv = "CV (\%)",
      geom_mean = "Geometric Mean",
      geom_cv = "CV \% Geometric Mean",
      median = "Median",
      min = "Minimum",
      max = "Maximum"
    )
  )
result <- build_table(lyt, df = ADPP)
result
}
\seealso{
\link{summarize_vars}.
}
