% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{summary.node}
\alias{summary.node}
\title{Method overwrite}
\usage{
\method{summary}{node}(object, ...)
}
\arguments{
\item{object}{node object}

\item{...}{ignored currently}
}
\value{
a summary of the tree, see \code{\link{basic_node_info}}
}
\description{
Overwrite s3 method because s3 and s4 with same names cannot coexist
}
