% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,node-method}
\title{Create a summary of the object}
\usage{
summary(x, ...)

\S4method{summary}{node}(x, index = numeric(0))
}
\arguments{
\item{x}{node object}

\item{...}{other args}

\item{index}{path to node (excluding node name itself)}
}
\description{
Create a summary of the object

Creates a summary of a tree to first-depth.
It displays the dimension and class of each entry in an rtable.
}
\examples{
n11 <- node(name = "A", content = array(c(1:6), dim = c(2,3)), children = list())
n2 <- node(name = "D", content = c(1:3), children = list(
  node(name = "A", content = array(c(1:6), dim = c(2,3)), children = list()),
  node(name = "B", content = array(c(1:6), dim = c(2,3)), children = list()),
  node(name = "C", content = array(c(1:6), dim = c(2,3)), children = list())
))
summary(n11)
summary(n2)
}
