% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variables.R
\name{summary_formats}
\alias{summary_formats}
\title{Format Function for Descriptive Statistics}
\usage{
summary_formats(type = "numeric")
}
\arguments{
\item{type}{(\code{string})\cr choice of a summary data type.
Only \code{counts} and \code{numeric} types are currently supported.}
}
\description{
Returns format patterns for descriptive statistics.
The format is understood by the \code{rtables}.
}
