% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variables.R
\name{summary_labels}
\alias{summary_labels}
\title{Label Function for Descriptive Statistics}
\usage{
summary_labels()
}
\description{
Returns labels of descriptive statistics for numeric variables.
}
\keyword{internal}
