% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot.R
\name{surv_subgroup}
\alias{surv_subgroup}
\title{alternative tablist}
\usage{
surv_subgroup(time_to_event, event, arm, arm.ref, arm.comp = setdiff(arm,
  arm.ref), group_by, covariates = NULL)
}
\arguments{
\item{time_to_event}{time to event data}

\item{event}{is boolean, \code{TRUE} if event, \code{FALSE} if time_to_event
is censored}

\item{arm}{vector with arm information}

\item{arm.ref}{a character vector defining which arms in arm should be taken 
as the reference}

\item{arm.comp}{a character vector defining which arms in arm should be taken 
as the comparison}

\item{group_by}{data frame with one column per grouping}
}
\description{
alternative tablist
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr)
library(survival)

'\%needs\%' <- teal.oncology:::'\%needs\%'

surv_tbl_steam <- get_forest_survival(com.roche.cdt30019.go29436.re)

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

ATE_f <- ATE \%>\% filter(PARAMCD == "OS")

ASL$BAGED <- ifelse(ASL$BAGE <= median(ASL$BAGE), "<=median", ">median")

group_by <- merge(
 ATE_f[c("USUBJID", "STUDYID")],
 ASL[c("USUBJID", "STUDYID", "BAGED", "SEX", "BECOG")],
 all.x = TRUE, all.y = FALSE
)

head(group_by)

tbl <- surv_subgroup(
   time_to_event = ATE_f$AVAL,
   event = ATE_f$CNSR == 0,
   arm = ATE_f$ARM, 
   group_by = group_by[, -c(1,2), drop=FALSE],
   arm.ref = "DUMMY A"
)
tbl
Viewer(tbl)

compare_rtables(tbl, surv_tbl_steam)

plot(tbl)


}

# surv_subgroup(Surv(AVAL ~ I(CNSR != 'N') ~ ARM + SEX, data = ATE))
}
