% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot_tte.R
\name{surv_subgroup}
\alias{surv_subgroup}
\title{alternative tablist}
\usage{
surv_subgroup(time_to_event, event, arm, arm.ref, arm.comp = setdiff(arm,
  arm.ref), group_by, covariates = NULL)
}
\arguments{
\item{time_to_event}{time to event data}

\item{event}{is boolean, \code{TRUE} if event, \code{FALSE} if time_to_event
is censored}

\item{arm}{vector with arm information}

\item{arm.ref}{a character vector defining which arms in arm should be taken 
as the reference}

\item{arm.comp}{a character vector defining which arms in arm should be taken 
as the comparison}

\item{group_by}{data frame with one column per grouping}
}
\description{
alternative tablist
}
\examples{

\dontrun{
library(atezo.data)
library(dplyr)
library(survival)

'\%needs\%' <- teal.oncology:::'\%needs\%'

surv_tbl_stream <- get_forest_survival_table(com.roche.cdt30019.go29436.re)
Viewer(surv_tbl_stream )

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

ATE_f <- ATE \%>\% filter(PARAMCD == "OS") \%>\% filter(ITTWTFL == "Y") \%>\% filter(ARM \%in\% c("DUMMY A", "DUMMY C"))
ASL_f <- ASL \%>\% filter(ITTWTFL == "Y") \%>\% filter(ARM \%in\% c("DUMMY A", "DUMMY C"))


group_by <- merge(
 ATE_f[c("USUBJID", "STUDYID")],
 ASL_f[c("USUBJID", "STUDYID", "SEX", "MLIVER", "TCICLVL2", "AGE4CAT", "RACE")],
 all.y = TRUE, all.x = FALSE
)

group_by$MLIVER <- factor(group_by$MLIVER, levels(factor(group_by$MLIVER))[c(2, 1)])
group_by$TCICLVL2 <- factor(group_by$TCICLVL2, levels(factor(group_by$TCICLVL2))[c(2, 1)])
group_by$AGE4CAT <- factor(group_by$AGE4CAT, levels(factor(group_by$AGE4CAT))[c(1, 3, 4, 2)])
group_by$RACE <- factor(group_by$RACE, levels(factor(group_by$RACE))[c(2, 3, 6, 1, 4, 5)])

head(group_by)

tbl <- surv_subgroup(
   time_to_event = ATE_f$AVAL,
   event = ATE_f$CNSR == 0,
   arm = ATE_f$ARM, 
   group_by = group_by[, -c(1,2), drop=FALSE],
   arm.ref = "DUMMY C",
   arm.comp = "DUMMY A"
)
tbl
Viewer(tbl)

compare_rtables(tbl, surv_tbl_stream, comp.attr = FALSE)

plot(tbl)


}

# surv_subgroup(Surv(AVAL ~ I(CNSR != 'N') ~ ARM + SEX, data = ATE))
}
