% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_coxph_pairwise.R
\name{survival_coxph_pairwise}
\alias{survival_coxph_pairwise}
\alias{s_coxph_pairwise}
\alias{coxph_pairwise}
\title{Pairwise CoxPH model}
\usage{
s_coxph_pairwise(
  df,
  .ref_group,
  .in_ref_col,
  .var,
  is_event,
  strat = NULL,
  control = control_coxph()
)

coxph_pairwise(lyt, vars, var_labels = "CoxPH", show_labels = "visible", ...)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.ref_group}{(\verb{data frame} or \code{vector})\cr the data corresponding to the reference group.}

\item{.in_ref_col}{(\code{logical})\cr
\code{TRUE} when working with the reference level, \code{FALSE} otherwise.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested by
a statistics function.}

\item{is_event}{(\code{logical})\cr \code{TRUE} if event, \code{FALSE} if time to event is censored.}

\item{strat}{(\code{character} or \code{NULL}) variable names indicating stratification factors.}

\item{control}{(\code{list}) \cr parameters for comparison details, specified by using \cr
the helper function \link{control_coxph}. Some possible parameter options are: \cr
\itemize{
\item \code{pval_method}: (\code{string}) \cr p-value method for testing hazard ratio = 1.
Default method is "log-rank", can also be set to "wald" or "likelihood".
\item \code{ties}: (\code{string}) \cr specifying the method for tie handling. Default is "efron",
can also be set to "breslow" or "exact". see more in \code{\link[survival:coxph]{survival::coxph()}}
\item \code{conf_level}: (\code{proportion})\cr confidence level of the interval for HR.
}}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{var_labels}{character for label.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{...}{additional arguments for the lower level functions.}
}
\value{
The statistics are:
\itemize{
\item \code{pvalue} : p-value to test HR = 1.
\item \code{hr} : hazard ratio.
\item \code{hr_ci} : confidence interval for hazard ratio.
}
}
\description{
Summarize p-value, HR and CIs from stratified or unstratified CoxPH model.
}
\section{Functions}{
\itemize{
\item \code{s_coxph_pairwise}: Statistics Function which analyzes HR, CIs of HR and p-value with coxph model.

\item \code{coxph_pairwise}: Analyze Function which adds the pairwise coxph analysis
to the input layout. Note that additional formatting arguments can be used here.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{
library(random.cdisc.data)
library(dplyr)
ADTTE <- radtte(cached = TRUE)
ADTTE_f <- ADTTE \%>\%
 dplyr::filter(PARAMCD == "OS") \%>\%
 dplyr::mutate(is_event = CNSR == 0)
df <- ADTTE_f \%>\% dplyr::filter(ARMCD == "ARM A")
df_ref_group <- ADTTE_f \%>\% dplyr::filter(ARMCD == "ARM B")
s_coxph_pairwise(df, df_ref_group,.in_ref_col = FALSE, .var = "AVAL", is_event = "is_event")


split_cols_by(lyt = NULL, var = "ARMCD", ref_group = "ARM A") \%>\%
 add_colcounts() \%>\%
 coxph_pairwise(
    vars = "AVAL",
    is_event = "is_event",
    var_labels = "Unstratified Analysis"
 ) \%>\%
 build_table(df = ADTTE_f)

split_cols_by(lyt = NULL, var = "ARMCD", ref_group = "ARM A") \%>\%
 add_colcounts() \%>\%
 coxph_pairwise(
    vars = "AVAL",
    is_event = "is_event",
    var_labels = "Stratified Analysis",
    strat = "SEX",
    control = control_coxph(pval_method = "wald")
 ) \%>\%
 build_table(df = ADTTE_f)

}
