% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_coxph_pairwise.R
\name{survival_coxph_pairwise}
\alias{survival_coxph_pairwise}
\alias{s_coxph_pairwise}
\alias{a_coxph_pairwise}
\alias{coxph_pairwise}
\title{Pairwise CoxPH model}
\usage{
s_coxph_pairwise(
  df,
  .ref_group,
  .in_ref_col,
  .var,
  is_event,
  strat = NULL,
  control = control_coxph()
)

a_coxph_pairwise(
  df,
  .ref_group,
  .in_ref_col,
  .var,
  is_event,
  strat = NULL,
  control = control_coxph()
)

coxph_pairwise(
  lyt,
  vars,
  ...,
  var_labels = "CoxPH",
  show_labels = "visible",
  table_names = vars,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.ref_group}{(\verb{data frame} or \code{vector})\cr the data corresponding to the reference group.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{is_event}{(\code{logical})\cr \code{TRUE} if event, \code{FALSE} if time to event is censored.}

\item{strat}{(\code{character} or \code{NULL}) variable names indicating stratification factors.}

\item{control}{(\code{list}) \cr parameters for comparison details, specified by using \cr
the helper function \code{\link[=control_coxph]{control_coxph()}}. Some possible parameter options are: \cr
\itemize{
\item \code{pval_method}: (\code{string}) \cr p-value method for testing hazard ratio = 1.
Default method is "log-rank", can also be set to "wald" or "likelihood".
\item \code{ties}: (\code{string}) \cr specifying the method for tie handling. Default is "efron",
can also be set to "breslow" or "exact". See more in \code{\link[survival:coxph]{survival::coxph()}}
\item \code{conf_level}: (\code{proportion})\cr confidence level of the interval for HR.
}}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{...}{additional arguments for the lower level functions.}

\item{var_labels}{character for label.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
The statistics are:
\itemize{
\item \code{pvalue} : p-value to test HR = 1.
\item \code{hr} : hazard ratio.
\item \code{hr_ci} : confidence interval for hazard ratio.
}
}
\description{
Summarize p-value, HR and CIs from stratified or unstratified CoxPH model.
}
\section{Functions}{
\itemize{
\item \code{s_coxph_pairwise}: Statistics Function which analyzes HR, CIs of HR and p-value with coxph model.

\item \code{a_coxph_pairwise}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{coxph_pairwise}: Analyze Function which adds the pairwise coxph analysis
to the input layout. Note that additional formatting arguments can be used here.
}}

\examples{
library(random.cdisc.data)
library(dplyr)

ADTTE <- radtte(cached = TRUE)
ADTTE_f <- ADTTE \%>\%
  filter(PARAMCD == "OS") \%>\%
  mutate(is_event = CNSR == 0)
df <- ADTTE_f \%>\%
  filter(ARMCD == "ARM A")
df_ref_group <- ADTTE_f \%>\%
  filter(ARMCD == "ARM B")

s_coxph_pairwise(df, df_ref_group, .in_ref_col = FALSE, .var = "AVAL", is_event = "is_event")

a_coxph_pairwise(df, df_ref_group, .in_ref_col = FALSE, .var = "AVAL", is_event = "is_event")


split_cols_by(lyt = NULL, var = "ARMCD", ref_group = "ARM A") \%>\%
  add_colcounts() \%>\%
  coxph_pairwise(
    vars = "AVAL",
    is_event = "is_event",
    var_labels = "Unstratified Analysis"
  ) \%>\%
  build_table(df = ADTTE_f)

split_cols_by(lyt = NULL, var = "ARMCD", ref_group = "ARM A") \%>\%
  add_colcounts() \%>\%
  coxph_pairwise(
    vars = "AVAL",
    is_event = "is_event",
    var_labels = "Stratified Analysis",
    strat = "SEX",
    control = control_coxph(pval_method = "wald")
  ) \%>\%
  build_table(df = ADTTE_f)
}
