% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_duration_subgroups.R
\name{survival_duration_subgroups}
\alias{survival_duration_subgroups}
\alias{extract_survival_subgroups}
\alias{a_survival_subgroups}
\alias{tabulate_survival_subgroups}
\title{Tabulate Survival Duration by Subgroup}
\usage{
extract_survival_subgroups(
  variables,
  data,
  control = control_coxph(),
  label_all = "All Patients"
)

a_survival_subgroups(
  .formats = list(n = "xx", n_events = "xx", median = "xx.x", n_tot = "xx", hr =
    list(format_extreme_values(2L)), ci = list(format_extreme_values_ci(2L)), pval =
    "x.xxxx | (<0.0001)")
)

tabulate_survival_subgroups(lyt, vars, control = control_coxph())
}
\arguments{
\item{variables}{(named \code{list} of \code{string})\cr list of additional analysis variables.}

\item{data}{(\verb{data frame})\cr the dataset containing the variables to summarize.}

\item{control}{(\code{list}) \cr parameters for comparison details, specified by using \cr
the helper function \code{\link[=control_coxph]{control_coxph()}}. Some possible parameter options are: \cr
\itemize{
\item \code{pval_method}: (\code{string}) \cr p-value method for testing hazard ratio = 1.
Default method is "log-rank", can also be set to "wald" or "likelihood".
\item \code{ties}: (\code{string}) \cr specifying the method for tie handling. Default is "efron",
can also be set to "breslow" or "exact". See more in \code{\link[survival:coxph]{survival::coxph()}}
\item \code{conf_level}: (\code{proportion})\cr confidence level of the interval for HR.
}}

\item{label_all}{(\code{string})\cr label for the total population analysis.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr the name of statistics to be reported among
\code{n} (total number of observations per group), \code{n_events} (number of events per group),
\code{median} (median survival time), \code{n_tot} (total number of observations),
\code{hr} (hazard ratio), \code{ci} (confidence interval of hazard ratio) and \code{pvalue} (p value of the effect).}
}
\description{
Tabulate statistics such as median survival time and hazard ratio for population subgroups.
}
\details{
These functions create a layout starting from a data frame which contains
the required statistics. Tables typically used as part of forest plot.
}
\section{Functions}{
\itemize{
\item \code{extract_survival_subgroups}: Prepares estimates of median survival times and treatment hazard ratios for
population subgroups in data frames. Simple wrapper for \code{\link[=h_survtime_subgroups_df]{h_survtime_subgroups_df()}} and \code{\link[=h_coxph_subgroups_df]{h_coxph_subgroups_df()}}.
Result is a list of two data frames: \code{survtime} and \code{hr}.
\code{variables} corresponds to the names of variables found in \code{data}, passed as a named list and requires elements
\code{tte}, \code{is_event}, \code{arm}, \code{subgroups} and optionally \code{strat}.

\item \code{a_survival_subgroups}: Formatted Analysis function used to format the results of
\code{\link[=extract_survival_subgroups]{extract_survival_subgroups()}}. Returns is a list of Formatted Analysis functions with one element per statistic.

\item \code{tabulate_survival_subgroups}: layout creating function.
}}

\examples{

# Testing dataset.
library(random.cdisc.data)
library(dplyr)

adtte <- radtte(cached = TRUE)

# Save variable labels before data processing steps.
adtte_labels <- var_labels(adtte)

adtte <- adtte \%>\%
  filter(
    PARAMCD == "OS",
    ARM \%in\% c("B: Placebo", "A: Drug X"),
    SEX \%in\% c("M", "F")
  ) \%>\%
  mutate(
    # Reorder levels of ARM to display reference arm before treatment arm.
    ARM = droplevels(forcats::fct_relevel(ARM, "B: Placebo")),
    SEX = droplevels(SEX),
    is_event = CNSR == 0
  ) \%>\%
  var_relabel(
    ARM = adtte_labels["ARM"],
    SEX = adtte_labels["SEX"],
    is_event = "Event Flag"
  )


df <- extract_survival_subgroups(
  variables = list(
    tte = "AVAL",
    is_event = "is_event",
    arm = "ARM", subgroups = c("SEX", "BMRKR2")
  ),
  data = adtte
)
df

a_survival_subgroups(.formats = list("n" = "xx", "median" = "xx.xx"))


## Table of survival times by subgroup.
basic_table() \%>\%
  tabulate_survival_subgroups(vars = c("n", "median")) \%>\%
  build_table(df$survtime)

## Table of hazard ratios by subgroup.
basic_table() \%>\%
  tabulate_survival_subgroups(vars = c("n_tot", "hr", "ci")) \%>\%
  build_table(df$hr)

}
