% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_time.R
\name{survival_time}
\alias{survival_time}
\alias{s_surv_time}
\alias{a_surv_time}
\alias{surv_time}
\title{Survival Time Analysis}
\usage{
s_surv_time(df, .var, is_event, control = control_surv_time())

a_surv_time(df, .var, is_event, control = control_surv_time())

surv_time(
  lyt,
  vars,
  ...,
  var_labels = "Time to Event",
  table_names = vars,
  .stats = c("median", "median_ci", "quantiles", "range_censor", "range_event"),
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{is_event}{(\code{logical})\cr \code{TRUE} if event, \code{FALSE} if time to event is censored.}

\item{control}{a (\code{list}) of parameters for comparison details, specified by using \cr
the helper function \link{control_surv_time}. Some possible parameter options are: \cr
\itemize{
\item \code{conf_level}: (\code{proportion})\cr confidence level of the interval for survival time.
\item \code{conf_type}: (\code{string}) \cr "plain" (default), "log", "log-log" for confidence interval type, \cr
see more in \code{\link[survival:survfit]{survival::survfit()}}. Note option, "none" is not supported.
\item \code{quantiles}: numeric vector of length two to specify the quantiles of survival time.
}}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{...}{additional arguments for the lower level functions.}

\item{var_labels}{character for label.}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars} are analyzed multiple times,
to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
The statistics are:
\itemize{
\item \code{median} : median survival time.
\item \code{median_ci} : confidence interval for median time.
\item \code{quantiles} : survival time for two specified quantiles.
\item \code{range_censor} : survival time range for censored observations.
\item \code{range_event} : survival time range for observations with events.
\item \code{range} : survival time range for all observations.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Summarize median survival time and CIs, percentiles of survival times, \cr
survival time range of censored/event patients.
}
\section{Functions}{
\itemize{
\item \code{s_surv_time}: Statistics Function which analyzes survival times.
\code{range_censor} and \code{range_event}.

\item \code{a_surv_time}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{surv_time}: Analyze Function which adds the survival times analysis
to the input layout. Note that additional formatting arguments can be used here.
}}

\examples{
library(scda)
library(dplyr)

ADTTE <- synthetic_cdisc_data("latest")$adtte
ADTTE_f <- ADTTE \%>\%
  filter(PARAMCD == "OS") \%>\%
  mutate(
    AVAL = day2month(AVAL),
    is_event = CNSR == 0
  )
df <- ADTTE_f \%>\% filter(ARMCD == "ARM A")

# Internal function - s_surv_time
\dontrun{
tern:::s_surv_time(df, .var = "AVAL", is_event = "is_event")
}

# Internal function - a_surv_time
\dontrun{
tern:::a_surv_time(df, .var = "AVAL", is_event = "is_event")
}

basic_table() \%>\%
  split_cols_by(var = "ARMCD") \%>\%
  add_colcounts() \%>\%
  surv_time(
    vars = "AVAL",
    var_labels = "Survival Time (Months)",
    is_event = "is_event",
    control = control_surv_time(conf_level = 0.9, conf_type = "log-log")
  ) \%>\%
  build_table(df = ADTTE_f)
}
\keyword{internal}
