% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_timepoint.R
\name{survival_timepoint}
\alias{survival_timepoint}
\alias{s_surv_timepoint}
\alias{a_surv_timepoint}
\alias{s_surv_timepoint_diff}
\alias{a_surv_timepoint_diff}
\alias{surv_timepoint}
\title{Survival Time Point Analysis}
\usage{
s_surv_timepoint(
  df,
  .var,
  time_point,
  is_event,
  control = control_surv_timepoint()
)

a_surv_timepoint(
  df,
  .var,
  time_point,
  is_event,
  control = control_surv_timepoint()
)

s_surv_timepoint_diff(
  df,
  .var,
  .ref_group,
  .in_ref_col,
  time_point,
  control = control_surv_timepoint(),
  ...
)

a_surv_timepoint_diff(
  df,
  .var,
  .ref_group,
  .in_ref_col,
  time_point,
  control = control_surv_timepoint(),
  ...
)

surv_timepoint(
  lyt,
  vars,
  ...,
  table_names_suffix = "",
  var_labels = "Time",
  show_labels = "visible",
  method = c("surv", "surv_diff", "both"),
  .stats = c("pt_at_risk", "event_free_rate", "rate_ci", "rate_diff", "rate_diff_ci",
    "ztest_pval"),
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{time_point}{(\code{number}) \cr survival time point of interest.}

\item{is_event}{(\code{logical})\cr \code{TRUE} if event, \code{FALSE} if time to event is censored.}

\item{control}{a (\code{list}) of parameters for comparison details, specified by using \cr
the helper function \link{control_surv_timepoint}. Some possible parameter options are: \cr
\itemize{
\item \code{conf_level}: (\code{proportion})\cr confidence level of the interval for survival rate.
\item \code{conf_type}: (\code{string}) \cr "plain" (default), "log", "log-log" for confidence interval type, \cr
see more in \code{\link[survival:survfit]{survival::survfit()}}. Note that the option "none" is no longer supported.
\item \code{time_point}: (\code{number}) \cr survival time point of interest.
}}

\item{.ref_group}{(\verb{data frame} or \code{vector})\cr the data corresponding to the reference group.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}

\item{...}{additional arguments for the lower level functions.}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{table_names_suffix}{(\code{string})\cr optional suffix for the \code{table_names} used for the rtables to
avoid warnings from duplicate table names.}

\item{var_labels}{character for label.}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}

\item{method}{(\code{string})\cr either \code{surv} (survival estimations),
\code{surv_diff} (difference in survival with the control) or \code{both}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}
}
\value{
The statistics are:
\itemize{
\item \code{pt_at_risk} : patients remaining at risk.
\item \code{event_free_rate} : event free rate (\%).
\item \code{rate_se} : standard error of event free rate.
\item \code{rate_ci} : confidence interval for event free rate.
}

The statistics are:
\itemize{
\item \code{rate_diff} : event free rate difference between two groups.
\item \code{rate_diff_ci} : confidence interval for the difference.
\item \code{ztest_pval} : p-value to test the difference is 0.
}
}
\description{
Summarize patient's survival rate and difference of survival rates between groups at a time point.
}
\section{Functions}{
\itemize{
\item \code{s_surv_timepoint}: Statistics Function which analyzes survival rate.

\item \code{a_surv_timepoint}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{s_surv_timepoint_diff}: Statistics Function which analyzes difference between two survival rates.

\item \code{a_surv_timepoint_diff}: Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.

\item \code{surv_timepoint}: Analyze Function which adds the survival rate analysis to the input layout.
Note that additional formatting arguments can be used here.
}}

\examples{
library(scda)
library(dplyr)

ADTTE <- synthetic_cdisc_data("latest")$adtte
ADTTE_f <- ADTTE \%>\%
  filter(PARAMCD == "OS") \%>\%
  mutate(
    AVAL = day2month(AVAL),
    is_event = CNSR == 0
  )
df <- ADTTE_f \%>\%
  filter(ARMCD == "ARM A")

s_surv_timepoint(df, .var = "AVAL", time_point = 7, is_event = "is_event")
a_surv_timepoint(df, .var = "AVAL", time_point = 7, is_event = "is_event")
df_ref_group <- ADTTE_f \%>\%
  filter(ARMCD == "ARM B")

s_surv_timepoint_diff(df, df_ref_group, .in_ref_col = TRUE, .var = "AVAL", is_event = "is_event")
s_surv_timepoint_diff(
  df,
  df_ref_group,
  .in_ref_col = FALSE,
  .var = "AVAL",
  time_point = 7,
  is_event = "is_event"
)
a_surv_timepoint_diff(
  df,
  df_ref_group,
  .in_ref_col = FALSE,
  .var = "AVAL",
  time_point = 7,
  is_event = "is_event"
)

# Survival at given time points.
basic_table() \%>\%
  split_cols_by(var = "ARMCD", ref_group = "ARM A") \%>\%
  add_colcounts() \%>\%
  surv_timepoint(
    vars = "AVAL",
    var_labels = "Months",
    is_event = "is_event",
    time_point = 7
  ) \%>\%
  build_table(df = ADTTE_f)

# Difference in survival at given time points.
basic_table() \%>\%
  split_cols_by(var = "ARMCD", ref_group = "ARM A") \%>\%
  add_colcounts() \%>\%
  surv_timepoint(
    vars = "AVAL",
    var_labels = "Months",
    is_event = "is_event",
    time_point = 9,
    method = "surv_diff",
    .indent_mods = c("rate_diff" = 0L, "rate_diff_ci" = 2L, "ztest_pval" = 2L)
  ) \%>\%
  build_table(df = ADTTE_f)

# Survival and difference in survival at given time points.
basic_table() \%>\%
  split_cols_by(var = "ARMCD", ref_group = "ARM A") \%>\%
  add_colcounts() \%>\%
  surv_timepoint(
    vars = "AVAL",
    var_labels = "Months",
    is_event = "is_event",
    time_point = 9,
    method = "both"
  ) \%>\%
  build_table(df = ADTTE_f)
}
