% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_timepoint.R
\name{survival_timepoint}
\alias{survival_timepoint}
\alias{s_surv_timepoint}
\alias{surv_timepoint}
\alias{s_surv_timepoint_diff}
\alias{surv_timepoint_diff}
\title{Survival Time Point Analysis}
\usage{
s_surv_timepoint(df, .var, is_event, control = control_surv_timepoint())

surv_timepoint(
  lyt,
  vars,
  var_labels = "Months",
  .stats = c("pt_at_risk", "event_free_rate", "rate_ci"),
  ...
)

s_surv_timepoint_diff(
  df,
  .var,
  .ref_group,
  .in_ref_col,
  conf_level = 0.95,
  ...
)

surv_timepoint_diff(
  lyt,
  vars,
  var_labels = "Time",
  show_labels = "visible",
  ...
)
}
\arguments{
\item{df}{(\verb{data frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{is_event}{(\code{logical})\cr \code{TRUE} if event, \code{FALSE} if time to event is censored.}

\item{control}{a (\code{list}) of parameters for comparison details, specified by using \cr
the helper function \link{control_surv_timepoint}. Some possible parameter options are: \cr
\itemize{
\item \code{conf_level}: (\code{proportion})\cr confidence level of the interval for survival rate.
\item \code{conf_type}: (\code{string}) \cr "plain" (default), "none", "log", "log-log" for confidence interval type, \cr
see more in \code{\link[survival:survfit]{survival::survfit()}}
\item \code{time_point}: (\code{number}) \cr survival time point of interest.
}}

\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{vars}{(\code{character})\cr variable names for the primary analysis variable to be iterated over.}

\item{var_labels}{character for label.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{...}{additional arguments for the lower level functions.}

\item{.ref_group}{(\verb{data frame} or \code{vector})\cr the data corresponding to the reference group.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}

\item{conf_level}{confidence level for difference of two risk rates}

\item{show_labels}{label visibility: one of "default", "visible" and "hidden".}
}
\value{
The statistics are:
\itemize{
\item \code{pt_at_risk} : patients remaining at risk.
\item \code{event_free_rate} : event free rate (\%).
\item \code{rate_se} : standard error of event free rate.
\item \code{rate_ci} : confidence interval for event free rate.
}

The statistics are:
\itemize{
\item \code{rate_diff} : event free rate difference between two groups.
\item \code{rate_diff_ci} : confidence interval for the difference.
\item \code{ztest_pval} : p-value to test the difference is 0.
}
}
\description{
Summarize patient's survival rate and difference of survival rates between groups at a time point.
}
\section{Functions}{
\itemize{
\item \code{s_surv_timepoint}: Statistics Function which analyzes survival rate.

\item \code{surv_timepoint}: Analyze Function which adds the survival rate analysis to the input layout.
Note that additional formatting arguments can be used here.

\item \code{s_surv_timepoint_diff}: Statistics Function which analyzes difference between two survival rates.

\item \code{surv_timepoint_diff}: Analyze Function which adds the difference between two survival rates analysis
to the input layout. Note that additional formatting arguments can be used here.
}}

\section{Formatting arguments}{

These additional formatting arguments can be passed to the layout creating function:
\describe{
\item{.stats}{(\code{character})\cr names of the statistics to use}
\item{.indent_mods}{(\code{integer})\cr named vector of indent modifiers for the labels}
\item{.formats}{(\code{character} or \code{list})\cr named vector of formats for the statistics}
\item{.labels}{(\code{character})\cr named vector of labels for the statistics (without indent)}
}
}

\examples{
library(random.cdisc.data)
library(dplyr)
ADTTE <- radtte(cached = TRUE)
ADTTE_f <- ADTTE \%>\%
 dplyr::filter(PARAMCD == "OS") \%>\%
 dplyr::mutate(is_event = CNSR == 0)
df <- ADTTE_f \%>\% dplyr::filter(ARMCD == "ARM A")
s_surv_timepoint(df, .var = "AVAL", is_event = "is_event")


split_cols_by(lyt = NULL, var = "ARMCD", ref_group = "ARM A") \%>\%
 add_colcounts() \%>\%
 surv_timepoint(
   vars = "AVAL",
   var_labels = "7 Months",
   is_event = "is_event",
   control = control_surv_timepoint(time_point = 7)
 ) \%>\%
 build_table(df = ADTTE_f)


df_ref_group <- ADTTE_f \%>\% dplyr::filter(ARMCD == "ARM B")
s_surv_timepoint_diff(df, df_ref_group, .in_ref_col = TRUE, .var = "AVAL", is_event = "is_event")
s_surv_timepoint_diff(df, df_ref_group, .in_ref_col = FALSE, .var = "AVAL", is_event = "is_event")


split_cols_by(lyt = NULL, var = "ARMCD", ref_group = "ARM A") \%>\%
 add_colcounts() \%>\%
 surv_timepoint_diff(
   vars = "AVAL",
   var_labels = "9 Months",
   is_event = "is_event",
   control = control_surv_timepoint(time_point = 9),
   .indent_mods = c("rate_diff" = 0L, "rate_diff_ci" = 2L, "ztest_pval" = 2L)
 ) \%>\%
 build_table(df = ADTTE_f)

split_cols_by(lyt = NULL, var = "ARMCD", ref_group = "ARM A") \%>\%
 add_colcounts() \%>\%
 surv_timepoint(
   vars = "AVAL",
   var_labels = "7 Months",
   is_event = "is_event",
   control = control_surv_timepoint(time_point = 7)
 ) \%>\%
 surv_timepoint_diff(
   vars = "AVAL",
   show_label = "hidden",
   is_event = "is_event",
   control = control_surv_timepoint(time_point = 7)
 ) \%>\%
 build_table(df = ADTTE_f)

}
