% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae_ctc.R
\name{t_ae_ctc}
\alias{t_ae_ctc}
\title{Adverse Events Table by Highest NCI CTCAE Grade}
\usage{
t_ae_ctc(class, term, id, grade, col_by, total = "All Patients",
  grade_levels = 1:5)
}
\arguments{
\item{class}{system organ class variable.}

\item{term}{preferred term variable.}

\item{id}{unique subject identifier variable. If a particular subject has no
adverse event then the subject \code{id} should be listed where
\code{class} and \code{term} should be set to missing (i.e. \code{NA}).}

\item{grade}{grade of adverse event variable.}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing. See 'Examples'.}

\item{total}{character string that will be used as a label for a column with 
pooled total population, default is "All Patients".}

\item{grade_levels}{ordered values of possible of grades in a form of
\code{x:y}, default is \code{1:5}. This assures a proper fill in for
grades, see 'Details'.}
}
\description{
\code{t_ae_ctc} returns adverse events sorted by highest NCI (National Cancer
 Institute) CTCAE (common terminology criteria for adverse avents) grade.
}
\details{
\code{t_ae_ctc} counts patients according to adverse events (AEs) of greatest
 intensity for system organ class (SOC) and overall rows and includes 
 percentages based on the total number of patients in the column heading 
 (i.e. "N=nnn"). If the intention is to use patients number from subject level
 dataset as N for percentage calculation then adeverse events dataset should
 be left joined to subject level dataset and the \code{col_by} variable should
 be dropped from adverse events dataset, see the example. Otherwise, N will be
 derived using adverse events dataset. At the preferred term (PT) level,
 multiple events within a patient of the same PT are counted once using the
 greatest intensity reported. 

\code{t_ae_ctc} removes any non-complete records, e.g. if class or term are 
 missing. If the intent is to preserve such records, then impute missing 
 values before using \code{t_ae_ctc}.
     
\code{t_ae_ctc} orders data by "All Patients" column from the most commonly
 reported SOC to the least frequent one. Within SOC, it sorts by decreasing
 frequency of PT. It brakes ties using SOC/PT names in alphabetical order.   

\code{t_ae_ctc} fills in \code{col_by} and \code{grade} with \code{0} value 
in case there was no AEs reported for particular \code{col_by} and/or 
\code{grade} category. Use \code{grade_levels} to modify the range of existing
grades. If data does not have any records with \code{grade} 5 and the intent 
is to show only grades 1-4 rows then use \code{grade_levels = 1:4}.
}
\examples{
# Simple example
library(tibble)
library(dplyr)
ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = paste("ARM", LETTERS[rep(c(1,2), c(3,7))])
)


ae_lookup <- tribble(
~CLASS,         ~TERM,   ~GRADE,
"cl A",   "trm A_1/2",        1,
"cl A",   "trm A_2/2",        2,  
"cl B",   "trm B_1/3",        2,
"cl B",   "trm B_2/3",        3,
"cl B",   "trm B_3/3",        1,
"cl C",   "trm C_1/1",        1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2,2,2,3,3,4,4,4,4,5,6,6,7,7)]
  ),
  ae_lookup[c(1,1,2,6,4,2,2,3,4,2,1,5,4,6),]
)

ANL <- left_join(ASL, AAE, by = "USUBJID")

tbl <- t_ae_ctc(
  class = ANL$CLASS,
  term =  ANL$TERM,
  id = ANL$USUBJID,
  grade = ANL$GRADE,
  col_by = factor(ANL$ARM),
  total = "All Patients",
  grade_levels = 1:3
)

tbl


library(random.cdisc.data)
library(dplyr)

ASL <- radam("ASL", N = 10)
AAE <- radam("AAE", ADSL = ASL)

ANL <- left_join(AAE, ASL \%>\% select(USUBJID, STUDYID, ARM), by = c("STUDYID", "USUBJID"))

tbl <- with(ANL,
            t_ae_ctc(
              class = AEBODSYS,
              term =  AEDECOD,
              id = USUBJID,
              grade = AETOXGR,
              col_by = factor(ARM),
              total = "All Patients",
              grade_levels = 1:5
            )
)

tbl

}
\author{
Edgar Manukyan (manukyae) \email{edgar.manukyan@roche.com}

Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
