% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoints.R
\name{t_binary_outcome}
\alias{t_binary_outcome}
\title{Summary table for binary outcome}
\usage{
t_binary_outcome(
  rsp,
  col_by,
  strata_data = NULL,
  conf_level = 0.95,
  prop_ci_method = "waldcc",
  unstrat_analysis = control_binary_comparison(),
  strat_analysis = control_binary_comparison(diff_test = "cmh")
)
}
\arguments{
\item{rsp}{(\code{logical} vector)\cr
\code{TRUE} if subject is a responder \code{FALSE} otherwise}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{strata_data}{(\code{data.frame})\cr
Used for stratification factors. If \code{NULL}, no stratified
analysis is performed.}

\item{conf_level}{confidence level of the interval.}

\item{prop_ci_method}{one of (\code{"waldcc"}, \code{"wald"}, \code{"clopper-pearson"},
\code{"wilson"}, \code{agresti-coull} or \code{jeffreys}).\cr
Specifies the method used to construct the confidence interval for proportion
of successful outcomes.}

\item{unstrat_analysis}{list specified by calling \code{\link{control_binary_comparison}} for controlling the
"Unstratified Analysis" section. If \code{NULL}, this table section is excluded.}

\item{strat_analysis}{list specified by calling \code{\link{control_binary_comparison}} for controlling the
"Stratified Analysis" section. Only applicable if stratification factors from \code{strata_data} are provided.
If \code{NULL}, this table section is excluded.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
This is a convenience wrapper function for creating a summary of common analyses
for binary outcome variables. Options include reporting unstratified
analysis, stratified analysis or both.
}
\details{
Table subsections are matched by name, so to remove a subsection set only the corresponding arguments of
\code{unstrat_analysis} or \code{strat_analysis} to \code{NULL}.
}
\examples{
library(random.cdisc.data)
library(dplyr)
ADRS <- radrs(cached = TRUE) \%>\%
  dplyr::filter(PARAMCD == "BESRSPI")

# Unstratified analysis only with default options for `unstrat_analysis`.
t_binary_outcome(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD
)

# Unstratified + stratified analysis with default options for
# `unstrat_analysis` and `strat_analysis`.
t_binary_outcome(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  strata_data = ADRS[, c("STRATA1", "STRATA2")]
)

# Stratified analysis only, using CMH weights and corresponding test statistic.
# Exclude Odds Ratio estimates.
t_binary_outcome(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  strata_data = ADRS[, c("STRATA1", "STRATA2")],
  unstrat_analysis = NULL,
  strat_analysis = control_binary_comparison(diff_ci = "cmh", diff_test = "cmh", odds_ratio = FALSE)
 )

# Exclude any comparisons between arms,
# specify non-default CI method for proportions and CI level.
t_binary_outcome(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  prop_ci_method = "clopper-pearson",
  unstrat_analysis = NULL,
  conf_level = 0.90
 )
}
\seealso{
\code{\link{control_binary_comparison}}, \code{\link{t_el_proportion}},
\code{\link{t_el_proportion_diff}}, \code{\link{t_el_test_proportion_diff}}
}
