% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_count_true.R
\name{t_count_true.data.frame}
\alias{t_count_true.data.frame}
\title{Tabulate \code{TRUE} counts in a data.frame}
\usage{
\method{t_count_true}{data.frame}(
  x,
  col_by,
  col_N = NULL,
  total = NULL,
  denominator = c("N", "n", "omit"),
  table_tree = FALSE,
  ...
)
}
\arguments{
\item{x}{data frame with only logical variables to be summarized.
If a variable has a \code{label} attribute then it will be used for the row name.}

\item{col_by}{logical, factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vector with the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{string of column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{denominator}{either \code{"n"}, \code{"N"} or \code{"omit"}. \code{"n"} and \code{"N"} are for calculating
the level associated percentage. With option \code{"N"}, the reference population from \code{col_N} is used as
the denominator. With option \code{"n"}, the number of non-missing records from \code{x} is used as
the denominator. If \code{"omit"} is chosen the percentage is omitted.}

\item{table_tree}{logical, if \code{FALSE} then a single \code{rtable} gets returned, if \code{TRUE} a
\code{\link{node}} object gets returned with the elementary \code{rtable} objects.}

\item{...}{not used arguments}
}
\description{
Note that for this function the default \code{denominator = "N"}
and not \code{n} as for the other \code{t_summary} functions.
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ANL <- ADSL \%>\%
  dplyr::transmute(
    SAFPOP = SAFFL == 'Y' ,
    DISCSTUD =  !is.na(DCSREAS)
  )

t_count_true(ANL, col_by = ADSL$ARM)

# Add variable labels
ANL <- ANL \%>\%
  var_relabel(
    SAFPOP = "Safety Population",
    DISCSTUD = "Discontinued study"
  )

t_count_true(ANL, col_by = ADSL$ARM)
t_count_true(ANL, col_by = ADSL$ARM, denominator = "omit")
}
\seealso{
\code{\link{t_count_true}},
\code{\link{t_count_true.logical}},
\code{\link{t_summary.logical}}, \code{\link{t_summary.factor}},
}
