% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_count_true.R
\name{t_count_true.default}
\alias{t_count_true.default}
\title{Return an \code{rtable} with a no Method Message}
\usage{
\method{t_count_true}{default}(
  x,
  col_by,
  col_N = NULL,
  total = NULL,
  denominator = c("N", "n", "omit"),
  ...
)
}
\arguments{
\item{x}{a logical vector}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{(\code{character} value)\cr
Denotes column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{denominator}{(one of \code{("n", "N", "omit"})\cr
Value set to \code{"n"} or \code{"N"} are for calculating the level associated percentage.
With option \code{"N"}, the reference population from \code{col_N} is used as
the denominator. With option \code{"n"}, the number of non-missing records from \code{x} is used as
the denominator. If \code{"omit"} is chosen the percentage is omitted.}

\item{...}{not used arguments}
}
\value{
\code{rtable}
}
\description{
If there is no explicit method for an object an \code{rtable} with one row with an
appropriate message is returned.
}
\examples{
t_count_true(1:5, factor(LETTERS[c(1,2,1,1,2)]))
}
