% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_events_term_id.R
\name{t_count_unique}
\alias{t_count_unique}
\title{Count Unique Elements Per Cell}
\usage{
t_count_unique(x, col_by, col_N = NULL, na_rm = TRUE,
  row_name = "number of unique elements")
}
\arguments{
\item{x}{a vector}

\item{col_by}{logical factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default
\code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.}

\item{na_rm}{a logical value indicating whether \code{NA} values should be
removed from \code{x} prior to counting the unique elements per cell.}

\item{row_name}{a string with the row name to display in the summary table
that is returned. Default is "number of unique elements."}
}
\value{
an \code{rtable}
}
\description{
\code{t_count_unique} counts the number of unique elements per cell.
}
\examples{

t_count_unique(
 x = paste("id", c(1, 4, 2, 3, 3), sep = "-"),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10)
)

t_count_unique(
 x = c(1, 4, 2, 3, 3, NA),
 col_by = factor(c("A", "A", "B", "C", "C", "C")),
 col_N = c(2, 4, 10),
 row_name = "Unique Records",
 na_rm = FALSE
)
}
\seealso{
\code{\link{t_events_per_term_id}}, \code{\link{t_el_events_per_term_id}},
  \code{\link{t_events_per_term_grade_id}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
