% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{t_count_unique}
\alias{t_count_unique}
\title{Count Unique Elements Per Cell}
\usage{
t_count_unique(x, col_by, col_N = NULL,
  row.name = "number of unique elements", indent = 0)
}
\arguments{
\item{x}{a vector}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing. See 'Examples'.}

\item{col_N}{associated number of elements per level of \code{col_by}}
}
\value{
an rtable
}
\description{
Count the number of unique elements per cell.
}
\examples{

t_count_unique(
 x = paste("id", c(1, 4, 2, 3, 3), sep = "-"),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10)
)

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
