% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_events_term_id.R
\name{t_count_unique}
\alias{t_count_unique}
\title{Count Unique Elements Per Cell}
\usage{
t_count_unique(x, col_by, col_n, na_rm = TRUE,
  row_name = "number of unique elements", indent = 0)
}
\arguments{
\item{x}{a vector}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing. See 'Examples'.}

\item{col_n}{associated number of elements per level of \code{col_by}}

\item{na_rm}{a logical value indicating whether \code{NA} values should be
removed from \code{x} pror to counting the unique elements per cell}

\item{row_name}{a string with the row name to disply in the summary table
that is returned. Default is "number of unique elements."}

\item{indent}{non-negative integer where 0 means that the row should not be indented}
}
\value{
an rtable
}
\description{
\code{t_count_unique} counts the number of unique elements per cell.
}
\examples{

t_count_unique(
 x = paste("id", c(1, 4, 2, 3, 3), sep = "-"),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_n = c(2, 4, 10)
)

t_count_unique(
 x = c(1, 4, 2, 3, 3, NA),
 col_by = factor(c("A", "A", "B", "C", "C", "C")),
 col_n = c(2, 4, 10),
 row_name = "Unique Records",
 indent = 1,
 na_rm = FALSE
)
}
\seealso{
\code{\link{t_events_per_term_id}}, \code{\link{t_events_summary}},
  \code{\link{lt_events_per_term_id_2}}, \code{\link{t_events_per_term_grade_id}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
