% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{t_count_unique}
\alias{t_count_unique}
\title{Count Unique Elements Per Cell}
\usage{
t_count_unique(x, col_by, col_N = NULL,
  row.name = "number of unique elements", indent = 0)
}
\arguments{
\item{x}{a vector}

\item{col_N}{associated number of elements per level of \code{col_by}}
}
\value{
an rtable
}
\description{
Count the number of unique elements per cell.
}
\examples{

t_count_unique(
 x = paste("id", c(1, 4, 2, 3, 3), sep = "-"),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10)
)

}
