% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot_data.R
\name{t_coxph}
\alias{t_coxph}
\title{Prepare Cox PH model annotation data with rtable format}
\usage{
t_coxph(fit_coxph, info_coxph = "Cox Porportional Hazard Model")
}
\arguments{
\item{fit_coxph}{a class \code{\link{coxph}} object.}

\item{info_coxph}{label information for Cox PH model.}
}
\description{
An \code{\link[rtables]{rtable}} format of \code{\link{coxph}} object for further annotation on top of Kaplan-Meier grob
}
\examples{
library(dplyr)
library(rtables)
OS <- data.frame(AVAL = abs(rnorm(200)), 
                 CNSR = sample(c(0, 1), 200, TRUE), 
                 ARM = sample(LETTERS[1:3], 200, TRUE),
                 SEX = sample(c("M","F"), 200, TRUE),
                 RACE = sample(c("AA", "BB", "CC"), 200, TRUE),
                 ECOG = sample(c(0, 1), 200, TRUE))
fit_coxph <- coxph(Surv(AVAL, 1-CNSR) ~ ARM + strata(RACE), data = OS, ties = "exact")
tbl <- t_coxph(fit_coxph)
Viewer(tbl)

}
