% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_coxph.R
\name{t_coxph}
\alias{t_coxph}
\title{Proportional Hazards Regression Model Fit Summary Table}
\usage{
t_coxph(fit_coxph, info_coxph = "Cox Porportional Hazard Model")
}
\arguments{
\item{fit_coxph}{a class \code{\link{survival}{coxph}} object.}

\item{info_coxph}{label information for Cox PH model.}
}
\description{
An \code{\link[rtables]{rtable}} format of \code{\link[survival]{coxph}}
object for further annotation on top of Kaplan-Meier grob
}
\examples{
library(random.cdisc.data)

ADSL <- cadsl
ADSL$RACE <- factor(
  vapply(as.character(ADSL$RACE),
         function(x) {
           if (nchar(x)>9) paste0(substr(x, 1,9), "...") else x
         },
         character(1)
  )
)

ADTTE <- cadtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")
\donttest{
fit_coxph <- coxph(formula = Surv(time = AVAL, time2 = 1-CNSR) ~ ARM + strata(RACE),
  data = ADTTE_f, ties = "exact"
)
tbl <- t_coxph(fit_coxph)
tbl
}

#todoL why is there no space in the table by summary: e.g. "B: Placebo" -> "ARM B: Placebo"
}
\author{
Heng Wang (wangh107) \email{wang.heng@gene.com}
}
