% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_coxph.R
\name{t_coxph}
\alias{t_coxph}
\title{Proportional Hazards Regression Model Fit Summary Table}
\usage{
t_coxph(fit_coxph, info_coxph = "Cox Porportional Hazard Model")
}
\arguments{
\item{fit_coxph}{a class \code{\link{coxph}} object.}

\item{info_coxph}{label information for Cox PH model.}
}
\description{
An \code{\link[rtables]{rtable}} format of \code{\link[survival]{coxph}}
object for further annotation on top of Kaplan-Meier grob
}
\examples{

OS <- data.frame(
  AVAL = abs(rnorm(200)), 
  CNSR = sample(c(0, 1), 200, TRUE), 
  ARM = sample(LETTERS[1:3], 200, TRUE),
  SEX = sample(c("M","F"), 200, TRUE),
  RACE = sample(c("AA", "BB", "CC"), 200, TRUE),
  ECOG = sample(c(0, 1), 200, TRUE)
)
                 
fit_coxph <- coxph(Surv(AVAL, 1-CNSR) ~ ARM + strata(RACE), data = OS, ties = "exact")
tbl <- t_coxph(fit_coxph)
tbl

}
\author{
Heng Wang (wangh107) \email{wang.heng@gene.com}
}
