% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_coxph.R
\name{t_coxph}
\alias{t_coxph}
\title{Proportional Hazards Regression Model Fit Summary Table}
\usage{
t_coxph(
  formula,
  data,
  conf_level = 0.95,
  pval_method = c("log-rank", "wald", "likelihood"),
  ...
)
}
\arguments{
\item{formula}{(\code{formula})\cr
Specifies \code{\link[survival:Surv]{survival model}}.
The arm variable needs to be wrapped in \code{\link{arm}}. The
\code{\link[survival]{strata}} special will only be used for the stratified analysis. If there is not
\code{\link[survival]{strata}} specification then the stratified analysis is omitted.}

\item{data}{(\code{data.frame})\cr
Contains all the variable that are used in \code{formula}}

\item{conf_level}{(\code{numeric} value)\cr
level for computation of the confidence intervals.
If set to \code{FALSE} no confidence intervals are printed}

\item{pval_method}{(one of (\code{"log-rank", "likelihood", "wald"}))\cr
Specifies the method used to calculate the p-value. Default value is \code{"log-rank"}.}

\item{...}{(optional)\cr
Other arguments will be passed to \code{\link[survival]{coxph}}}
}
\description{
An \code{\link[rtables]{rtable}} format of \code{\link[tern]{s_coxph_pairwise}}
results for further annotation on top of Kaplan-Meier grob
}
\examples{
library(random.cdisc.data)
ADSL <- cadsl
ADTTE <- cadtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

tbl <- t_coxph(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f
)
tbl

tbl <- t_coxph(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + strata(SEX),
  data = ADTTE_f,
  conf_level = 0.8, pval_method = "wald", ties = "exact"
)
tbl
}
\author{
Heng Wang (wangh107) \email{wang.heng@gene.com}
}
