% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_dsp.R
\name{t_dsp}
\alias{t_dsp}
\title{Single element for disposition table}
\usage{
t_dsp(x = x, col_by, col_N = table(col_by), row.name = NULL,
  indent = 0, subset = NULL, show_n = FALSE, ...)
}
\arguments{
\item{x}{a factor or logical vector.}

\item{col_by}{The group variable to define columns.}

\item{col_N}{The column total for each group that is displayed in the table header with (N=xx).}

\item{row.name}{Only applicable when x is a logical vector. A string to label the row name. Default is "TRUE".}

\item{indent}{An integer specifing the spaces before row.name.}

\item{subset}{A logical vector with the same length as x that defines the subset.}

\item{show_n}{Logic value to determine whether the "n" row is displayed. Default is FALSE.}

\item{...}{arguments passed on to methods}
}
\description{
Single element for disposition table
}
\examples{

library(random.cdisc.data)
library(dplyr)
library(utils.nest)

ADSL <- radsl(seed = 1)

ADSL0 <- ADSL \%>\%
 mutate(
   COMPSTUD = sample(c('Y','N'),
                     size=nrow(ADSL),
                     replace = TRUE) \%>\% as.factor,
   STUDONS = sample(c('Alive: On Treatment', 'Alive: In Follow-up', NA),
                    size=nrow(ADSL),
                    replace = TRUE) \%>\% as.factor,
   STDDRS = sample(c('Death', 'Lost To Follow-Up',
                     'Protocol Violation', 'Withdrawal By Subject',
                     'Other'),
                   size=nrow(ADSL),
                   replace = TRUE) \%>\% as.factor,
   GOTTRT = ifelse(!is.na(ACTARMCD), 'Y', 'N') \%>\% as.factor,
   DISTRTFL = sample(c('Y','N'),
                     size=nrow(ADSL),
                     replace = TRUE) \%>\% as.factor,
   TRTDRS = sample(c('ADVERSE EVENT', 'PROGRESSIVE DISEASE',
                     'PHYSICIAN DECISION', 'LACK OF EFFICACY',
                     'OTHER'),
                   size=nrow(ADSL),
                   replace = TRUE) \%>\% as.factor,
   STUDONS = case_when(
     COMPSTUD == 'N' ~ STUDONS
   ),
   STDDRS = case_when(
     COMPSTUD == 'N' & is.na(STUDONS) ~ STDDRS
   ),
   DISSTDFL = case_when(
     !is.na(STDDRS) ~ "Y"
   ),
   DISTRTFL = case_when(
     GOTTRT == 'Y' ~ DISTRTFL
   ),
   TRTDRS = case_when(
     DISTRTFL == 'Y' ~ TRTDRS
   ),
   DRSCAT = case_when(
     TRTDRS \%in\% c('ADVERSE EVENT', 'PHYSICIAN DECISION') ~ "Safety",
     !is.na(TRTDRS) ~ "Other"
   )
 ) \%>\% var_relabel(COMPSTUD = "Complete Study",
                   STUDONS = "On-study Status",
                   DISSTDFL = "Discontinued Study",
                   STDDRS = "Reason for Study \\r\\nDiscontinuation",
                   GOTTRT = "Received Treatment",
                   DISTRTFL = "Discontinued Treatment",
                   TRTDRS = "Reason for Treatment \\r\\nDiscontinuation"
                   )

library(purrr)
dsp <- purrr::partial(t_dsp, col_by = ADSL0$ARM, denominator = "N", total = "All Patients")

rbind(
  dsp(ADSL0$COMPSTUD == "Y", row.name = "Completed study"),
  dsp(ADSL0$STUDONS == "Alive: In Follow-up", row.name = "Alive: In follow-up"),
  dsp(ADSL0$DISSTDFL == "Y",  row.name = "Discontinued study"),
  dsp(as.factor(ADSL0$STDDRS), indent = 1),
  rrow(),
  dsp(ADSL0$GOTTRT == "Y", row.name = "Received treatment"),
  dsp(ADSL0$DISTRTFL == "Y", row.name = "Discontinued treatment"),
  dsp(ADSL0$DRSCAT == "Safety", row.name = "Safety", indent = 1),
  dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT \%in\% "Safety", indent = 2),
  dsp(ADSL0$DRSCAT == "Other", row.name = "Other", indent = 1),
  dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT \%in\% "Other", indent = 2)
)

}
