% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_disposition.R
\name{t_el_disposition}
\alias{t_el_disposition}
\title{Single element for disposition table}
\usage{
t_el_disposition(
  x,
  col_by,
  col_N = NULL,
  total = NULL,
  row.name = NULL,
  subset = NULL,
  show_n = FALSE,
  useNA = c("no", "ifany", "always"),
  drop_levels = NULL,
  denominator = "N"
)
}
\arguments{
\item{x}{(\code{factor} or \code{logical} vector)}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{(\code{character} value)\cr
Denotes column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{row.name}{(\code{character} value)\cr
Only applicable when x is a logical vector. A string to label the row name. Default is "TRUE".}

\item{subset}{(\code{logical} vector)\cr
Contains the same length as x that defines the subset
to take from \code{x}. NAs are cast to false.}

\item{show_n}{(\code{logical} value)
Determines whether the \code{"n"} row is displayed. Default is \code{FALSE}.
Only when x is of type factor}

\item{useNA}{(\code{logical})\cr
choose whether missing data (NAs) should be displayed as a level.}

\item{drop_levels}{(\code{logical} value)\cr
Whether to drop zero count levels}

\item{denominator}{(one of \code{("n", "N", "omit"})\cr
Value set to \code{"n"} or \code{"N"} are for calculating the level associated percentage.
With option \code{"N"}, the reference population from \code{col_N} is used as
the denominator. With option \code{"n"}, the number of non-missing records from \code{x} is used as
the denominator. If \code{"omit"} is chosen the percentage is omitted.}
}
\description{
Creates a disposition table
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)

ADSL0 <- ADSL \%>\%
  mutate(
    COMPSTUD = sample(c('Y','N'),
                      size=nrow(ADSL),
                      replace = TRUE) \%>\% as.factor,
    STUDONS = sample(c('Alive: On Treatment', 'Alive: In Follow-up', NA),
                     size=nrow(ADSL),
                     replace = TRUE) \%>\% as.factor,
    STDDRS = sample(c('Death', 'Lost To Follow-Up',
                      'Protocol Violation', 'Withdrawal By Subject',
                      'Other'),
                    size=nrow(ADSL),
                    replace = TRUE) \%>\% as.factor,
    GOTTRT = ifelse(!is.na(ACTARMCD), 'Y', 'N') \%>\% as.factor,
    DISTRTFL = sample(c('Y','N'),
                      size=nrow(ADSL),
                      replace = TRUE) \%>\% as.factor,
    TRTDRS = sample(c('ADVERSE EVENT', 'PROGRESSIVE DISEASE',
                      'PHYSICIAN DECISION', 'LACK OF EFFICACY',
                      'OTHER'),
                    size=nrow(ADSL),
                    replace = TRUE) \%>\% as.factor,
    STUDONS = case_when(
      COMPSTUD == 'N' ~ STUDONS
    ),
    STDDRS = case_when(
      COMPSTUD == 'N' & is.na(STUDONS) ~ STDDRS
    ) \%>\% as.factor,
    DISSTDFL = case_when(
      !is.na(STDDRS) ~ "Y"
    ),
    DISTRTFL = case_when(
      GOTTRT == 'Y' ~ DISTRTFL
    ),
    TRTDRS = case_when(
      DISTRTFL == 'Y' ~ TRTDRS
    ),
    DRSCAT = case_when(
      TRTDRS \%in\% c('ADVERSE EVENT', 'PHYSICIAN DECISION') ~ "Safety",
      !is.na(TRTDRS) ~ "Other"
    )
  ) \%>\% var_relabel(COMPSTUD = "Complete Study",
                    STUDONS = "On-study Status",
                    DISSTDFL = "Discontinued Study",
                    STDDRS = "Reason for Study \r\nDiscontinuation",
                    GOTTRT = "Received Treatment",
                    DISTRTFL = "Discontinued Treatment",
                    TRTDRS = "Reason for Treatment \r\nDiscontinuation"
  )

library(purrr)
t_el_disposition_indent <- function(..., indent = 0) indent(t_el_disposition(...), indent)
dsp <- partial(
  t_el_disposition_indent,
  col_by = ADSL0$ARM,
  total = "All Patients",
  denominator = "N"
)
# Only using rtables
# old way of doing things with indent variable
rbind(
  dsp(ADSL0$COMPSTUD == "Y", row.name = "Completed study"),
  dsp(ADSL0$STUDONS == "Alive: In Follow-up", row.name = "Alive: In follow-up"),
  dsp(ADSL0$DISSTDFL == "Y",  row.name = "Discontinued study"),
  dsp(ADSL0$STDDRS, indent = 1),
  rrow(),
  rrow("Show example of using n as denominator"),
  indent(t_el_disposition(
    ADSL0$STDDRS,
    col_by = ADSL0$ARM,
    total = "All Patients",
    denominator = "n"
  ), 1),
  rrow(),
  dsp(ADSL0$GOTTRT == "Y", row.name = "Received treatment"),
  dsp(ADSL0$DISTRTFL == "Y", row.name = "Discontinued treatment"),
  dsp(ADSL0$DRSCAT == "Safety", row.name = "Safety", indent = 1),
  dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT \%in\% "Safety", indent = 2),
  dsp(ADSL0$DRSCAT == "Other", row.name = "Other", indent = 1),
  dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT \%in\% "Other", indent = 2)
)

# new way:
treatment_section_rtable <- rbind(
  dsp(ADSL0$GOTTRT == "Y", row.name = "Received treatment"),
  dsp(ADSL0$DISTRTFL == "Y", row.name = "Discontinued treatment"),
  indent(
    rbind(
      dsp(ADSL0$DRSCAT == "Safety", row.name = "Safety"),
      indent(dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT \%in\% "Safety"), 1),
      dsp(ADSL0$DRSCAT == "Other", row.name = "Other"),
      indent(dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT \%in\% "Other"), 1)
    ),
    1
  )
)
treatment_section_rtable

# easier to manipulate for their purposes by other users
treatment_section_tree <- node(
  invisible_node_name("treatment"),
  content = NULL,
  format_data = list(children_gap = 0),
  children = list(
    node(
      invisible_node_name("received_treatment"),
      content = dsp(ADSL0$GOTTRT == "Y", row.name = "Received treatment")
    ),
    node(
      invisible_node_name("discontinuedtreatment"),
      dsp(ADSL0$DISTRTFL == "Y", row.name = "Discontinued treatment"),
      format_data = list(gap_to_children = 0, children_gap = 0, children_indent = 1),
      children = list(
        node(
          invisible_node_name("safety"),
          rbind(
            dsp(ADSL0$DRSCAT == "Safety", row.name = "Safety"),
            indent(dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT \%in\% "Safety"), 1)
          )
        ),
        node(
          invisible_node_name("other"),
          rbind(
            dsp(ADSL0$DRSCAT == "Other", row.name = "Other"),
            indent(dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT \%in\% "Other"), 1)
          )
        )
      )
    )
  )
)
to_rtable(treatment_section_tree)

to_rtable(invisible_node(
  name = "t_el_disp",
  children = list(
    node(
      name = invisible_node_name("Study status"),
      content = rbind(
        dsp(ADSL0$COMPSTUD == "Y", row.name = "Completed study"),
        dsp(ADSL0$STUDONS == "Alive: In Follow-up", row.name = "Alive: In follow-up")
      ),
      children = list(
        node(
          name = invisible_node_name("Reasons for study discontinuation (using capital N)"),
          rbind(
            dsp(ADSL0$DISSTDFL == "Y",  row.name = "Discontinued study"),
            indent(dsp(ADSL0$STDDRS), 1)
          )
        ),
        node(
          name = "Reasons for study discontinuation (using small n)",
          content = t_el_disposition(
            ADSL0$STDDRS,
            col_by = ADSL0$ARM,
            total = "All Patients",
            denominator = "n"
          )
        )
      )
    ),
    treatment_section_tree
  )
))
}
