% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_events_term_id.R
\name{t_el_events_per_term_id}
\alias{t_el_events_per_term_id}
\title{Summary table for events}
\usage{
t_el_events_per_term_id(id, col_by, col_N = NULL,
  total_events = "Total number of events",
  subjects_with_events = "Total number of patients with at least one adverse event")
}
\arguments{
\item{id}{unique subject identifier. If a particular subject has no adverse
event then that information needs to be added to the \code{col_N} argument.}

\item{col_by}{logical factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default
\code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.}

\item{total_events}{character string that will be used as a label in the row
for the total event count. If this is \code{NULL} then this row will be
removed.}

\item{subjects_with_events}{character string that will be used as a label in
the row for the total number with at least one event. If this is
\code{NULL} then this row will be removed.}
}
\value{
an \code{rtable}
}
\description{
\code{t_el_events_per_term_id} counts the number of unique elements per cell.
}
\examples{
t_el_events_per_term_id(
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10)
)
#todo: rename total_events -> total_events_label
#todo: rename subjects_with_events -> subjects_with_events_label
}
\seealso{
\code{\link{t_count_unique}}, \code{\link{t_events_per_term_id}},
  \code{\link{t_events_per_term_grade_id}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
