% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_rsp.R
\name{t_el_forest_rsp}
\alias{t_el_forest_rsp}
\title{Elementary Table for Forest Response Plot}
\usage{
t_el_forest_rsp(rsp, col_by, row_name = "", dense_header = FALSE)
}
\arguments{
\item{rsp}{boolean, \code{TRUE} if subject is a responder \code{FALSE}
otherwise}

\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group.}

\item{row_name}{name of row}

\item{dense_header}{Display the table headers in multiple rows.}
}
\value{
rtable with one row
}
\description{
Elementary Table for Forest Response Plot
}
\examples{

t_el_forest_rsp(
  rsp = sample(c(TRUE, FALSE), 200, TRUE),
  col_by = factor(sample(c("ARM A", "ARM B"), 200, TRUE), levels = c("ARM A", "ARM B"))
)

}
