% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_rsp.R
\name{t_el_forest_rsp}
\alias{t_el_forest_rsp}
\title{Elementary Table for Forest Response Plot}
\usage{
t_el_forest_rsp(
  rsp,
  col_by,
  strata_data = NULL,
  conf_level = 0.95,
  row_name = "",
  dense_header = FALSE
)
}
\arguments{
\item{rsp}{(\code{logical} vector)\cr
\code{TRUE} if subject is a responder \code{FALSE} otherwise}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{strata_data}{(\code{character}, \code{factor} or \code{data.frame})\cr
Used for \code{\link[survival:strata]{stratification factors}}
If \code{NULL}, no stratified analysis is performed.}

\item{conf_level}{(\code{numeric} value) confidence level of the interval}

\item{row_name}{(\code{character} value) name of row}

\item{dense_header}{(\code{logical} value) Display the table headers in multiple rows.}
}
\value{
\code{rtable} with one row
}
\description{
Elementary Table for Forest Response Plot
}
\examples{
rsp = sample(c(TRUE, FALSE), 200, TRUE)
col_by = factor(sample(c("ARM A", "ARM B"), 200, TRUE), levels = c("ARM A", "ARM B"))
strata_data <- data.frame(
  STRATA1 = sample(c("STR1", "STR2"), 200, TRUE),
  STRATA2 = sample(c("low", "medium", "high"), 200, TRUE)
)
t_el_forest_rsp(
  rsp = rsp,
  col_by = col_by
)

t_el_forest_rsp(
  rsp = rsp,
  col_by = col_by,
  strata_data = strata_data
)
}
\seealso{
\code{\link{t_forest_rsp}}
}
