% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_tte.R
\name{t_el_forest_tte}
\alias{t_el_forest_tte}
\title{Elementary Table for Forest Time to Event Plot}
\usage{
t_el_forest_tte(
  tte,
  is_event,
  col_by,
  strata_data = NULL,
  ties = "exact",
  time_unit = "month",
  row_name = "",
  conf_level = 0.95,
  dense_header = TRUE
)
}
\arguments{
\item{tte}{(\code{numeric} vector)\cr
Contains time to event data, currently no \code{NA} allowed}

\item{is_event}{(\code{logical} value)\cr
\code{TRUE} if event, \code{FALSE} if \code{tte}
is censored}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{strata_data}{(\code{character}, \code{factor} or \code{data.frame})\cr
Used for \code{\link[survival:strata]{stratification factors}}
(categorical variables). If \code{NULL}, no stratified analysis is performed.}

\item{ties}{(\code{character} value) the method used for tie handling in \code{\link[survival]{coxph}}.}

\item{time_unit}{(\code{character} value) The unit of median survival time. Default is \code{"months"}.
If \code{NULL} unit will be skipped.}

\item{row_name}{(\code{character} value) name of row}

\item{conf_level}{(\code{numeri} value) confidence level of the interval}

\item{dense_header}{(\code{logical} value) Display the table headers in multiple rows.}
}
\value{
\code{rtable} with one row
}
\description{
Elementary Table for Forest Time to Event Plot
}
\examples{

n <- 200
tte <- rexp(n)
set.seed(1)
is_event <- sample(c(TRUE, FALSE), n, TRUE)
strata_data <- data.frame(
  STRATA1 = sample(c("STR1", "STR2"), n, TRUE),
  STRATA2 = sample(c("low", "medium", "high"), n, TRUE)
)
col_by = factor(sample(c("ARM A", "ARM B"), n, TRUE), levels = c("ARM A", "ARM B"))
t_el_forest_tte(
  tte = tte,
  is_event = is_event,
  col_by = col_by,
  conf_level = 0.9
)

t_el_forest_tte(
  tte = tte,
  is_event = is_event,
  col_by = col_by,
  strata_data = strata_data
)

t_el_forest_tte(
  tte = tte,
  is_event = is_event,
  col_by = factor(rep("ARM A", n),
  levels = c("ARM A", "ARM B"))
)

t_el_forest_tte(
  tte = tte, is_event = is_event,
  col_by = factor(rep("ARM B", n), levels = c("ARM A", "ARM B"))
)
}
\seealso{
\code{\link{t_forest_tte}}
}
