% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_events_term_id.R
\name{t_events_per_term_id}
\alias{t_events_per_term_id}
\title{Basic Events Table}
\usage{
t_events_per_term_id(
  terms,
  id,
  col_by,
  col_N = NULL,
  total = NULL,
  event_type = "event",
  table_tree = FALSE
)
}
\arguments{
\item{terms}{(\code{character} or \code{factor} vector, or \code{data.frame})\cr
Represents events information. Multi-level
nesting is possible when \code{terms} is a \code{data.frame} and columns should be ordered with the first
column specifying the highest level term and the last column the lowest level term.}

\item{id}{(vector)\cr
contains subject identifier. Length of \code{id} must be the same as the length or number of rows of
\code{terms}.}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{col_N}{(\code{numeric} vector)\cr
Contains information of the number of patients in the levels of \code{col_by}. This is useful
if there are patients that have no adverse events can be accounted for with this argument.}

\item{total}{(\code{character} value)\cr
Denotes column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{event_type}{(\code{character} value)\cr
Specifies type of event that is summarized (e.g. adverse event, treatment). Default is \code{"event"}.}

\item{table_tree}{(\code{logical} value)\cr
If \code{FALSE} then a single \code{rtable} gets returned, if \code{TRUE} a
\code{\link{node}} object gets returned with the elementary \code{rtable} objects.}
}
\value{
an \code{\link{rtable}} object.
}
\description{
This function summarizes number of unique subjects with events and total number of events.
It creates basic summary of events and can be used for any events data like Adverse Events,
concomitant medication, medical history, etc.
}
\details{
Implementation examples are to apply \code{t_events_per_term_id} on Adverse Event data
to create Adverse Events summary table
(\code{AET02},
\href{http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_aet02.html#example-report-outputs-aet02-aet02}{
STREAM2.x},
\href{https://rochewiki.roche.com/confluence/pages/viewpage.action?pageId=294027342}{STREAM1.17} ),
or apply on Concomitant Medication data to create Concomitant Treatment summary table
(\code{CMT01},
\href{http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_cmt01.html#example-report-outputs-cmt01-cmt01}{
STREAM2.x},
\href{https://rochewiki.roche.com/confluence/pages/viewpage.action?pageId=294027342}{STREAM1.17}).


\code{t_events_per_term_id} includes percentages based on the total number of subjects
in the column heading (i.e. \code{"N=nnn"}). \code{col_N} can be explicitly specified to
get N for percentage calculation from either events dataset or additional dataset such as
subject level dataset. See the example.

Multiple occurrences of the same event in an individual \code{id} are counted only once when counting
number of subjects.

\code{t_events_per_term_id} doesn't deal with data with any non-complete records (has \code{NA}),
e.g. if any terms are missing. Impute missing values before using \code{t_events_per_term_id}.

\code{t_events_per_term_id} orders data by overall frequency across all columns from the most commonly
 reported higher level term to the least frequent one. Within each group of higher level term,
 it sorts by decreasing frequency of lower level term. It brakes ties using \code{terms} names in
 alphabetical order.
}
\examples{

library(dplyr)

t_events_per_term_id(
 terms = with_label(factor(c("t1", "t1", "t2", "t2", "t2")), "Term"),
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 total = "All Patients"
)

\dontrun{
# terms can't be empty strings or NA
t_events_per_term_id(
 terms = with_label(factor(c("", "t1", "t2", "t2", "t2")), "Term"),
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 total = "All Patients"
)

t_events_per_term_id(
 terms = with_label(factor(c(NA, "t1", "t2", "t2", "t2")), "Term"),
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 total = "All Patients"
)

}

# fill in "" with explicit NA
t_events_per_term_id(
 terms = explicit_na(sas_na(factor(c("", "", "t2", "t1", "t2")))),
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 total = "All Patients"
)

library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADAE <- radae(cached = TRUE)

t_events_per_term_id(
  terms = ADAE$AEDECOD,
  id = ADAE$USUBJID,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  event_type = "adverse event"
)

t_events_per_term_id(
  terms = ADAE$AEDECOD,
  id = ADAE$USUBJID,
  col_by = ADAE$ARM,
  total = "All Patients",
  col_N = table(ADSL$ARM),
  event_type = "adverse event"
)

t_events_per_term_id(
  terms = ADAE[, c("AEBODSYS", "AEDECOD")],
  id = ADAE$USUBJID,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients"
)

ADSL <- radsl(cached = TRUE)
ADCM <- radcm(cached = TRUE) \%>\%
  dplyr::filter(ATIREL == "CONCOMITANT")

t_events_per_term_id(
  terms = ADCM[, c("CMCLAS", "CMDECOD")],
  id = ADCM$USUBJID,
  col_by = ADCM$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients",
  event_type = "treatment"
)


# Table Tree

summary(t_events_per_term_id(
  terms = ADAE$AEDECOD,
  id = ADAE$USUBJID,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients",
  table_tree = TRUE
))


tbls <- t_events_per_term_id(
 terms = with_label(factor(c("t1", "t1", "t2", "t2", "t2")), "Term"),
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 table_tree = TRUE
)
summary(tbls)


tbls <- t_events_per_term_id(
  terms = ADAE[, c("AEBODSYS", "AEDECOD")],
  id = ADAE$USUBJID,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients",
  table_tree = TRUE
)
summary(tbls)
tbls[[1]]
tbls[['cl A.1']]


}
\seealso{
\code{\link{t_el_count_unique}}, \code{\link{t_el_events_per_term_id}},
  \code{\link{t_events_per_term_grade_id}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
