% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_events.R
\name{t_events_summary}
\alias{t_events_summary}
\title{Summary table for events}
\usage{
t_events_summary(term, id, col_by, col_N,
  total_events = "Total number of events",
  subjects_with_events = "Total number of patients with at least one adverse event")
}
\arguments{
\item{term}{a character vector with optional label attribute}

\item{id}{unique subject identifier. If a particular subject has no adverse 
event then that information needs to be added to the \code{col_N} argument.}

\item{col_by}{group variable that will be used for a column header.
\code{col_by} has to be a factor and can not be missing. See 'Examples'.}

\item{col_N}{associated number of elements per level of \code{col_by}}

\item{total_events}{character string that will be used as a label in the row
for the total event count. If this is \code{NULL} then this row will be
removed.}

\item{subjects_with_events}{character string that will be used as a label in
the row for the total number with at least one event. If this is
\code{NULL} then this row will be removed.}
}
\value{
an rtable
}
\description{
\code{t_events_summary} counts the number of unique elements per cell.
}
\examples{

t_events_summary(
 term = with_label(c("t1", "t1", "t2", "t2", "t2"), "Term"),
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10)
)


t_events_summary(
 term = with_label(c("t1", "t1", "t2", "t2", "t2"), "Term"),
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 total_events = NULL
)


t_events_summary(
 term = NULL,
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10)
)
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
