% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_rsp.R
\name{t_forest_rsp}
\alias{t_forest_rsp}
\title{Response Forest Plot Table}
\usage{
t_forest_rsp(rsp, col_by, group_data = NULL, total = "ALL",
  na.omit.group = TRUE)
}
\arguments{
\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group}

\item{group_data}{data frame with one column per grouping variable}

\item{rep}{Tumor Response data}
}
\description{
Response Forest Plot Table
}
\details{
Logistic model is used for odds ratio calculation
}
\examples{

library(random.cdisc.data)
ASL <- radam("ASL")
ARS <- radam("ARS", ADSL = ASL)
ASL$PSUEDO <- ASL$SEX

ARS_f <- ARS \%>\% filter(PARAMCD == "OVRSPI") 
ANL <- merge(ASL \%>\% select(USUBJID, STUDYID, SEX, RACE, PSUEDO, ARM), ARS_f)
head(ANL)

tbl <- t_forest_rsp(
  rsp = ANL$AVALC \%in\% c("CR", "PR"),
  col_by = factor(ANL$ARM), 
  group_data = as.data.frame(lapply(ANL[, c("SEX", "RACE")], as.factor))
)

Viewer(tbl)
   
}
\author{
Yuyao Song (songy24), \email{yuyao.song@roche.com}
}
