% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_rsp.R
\name{t_forest_rsp}
\alias{t_forest_rsp}
\title{Response Forest Plot Table}
\usage{
t_forest_rsp(rsp, col_by, group_data = NULL, total = "ALL",
  na.omit.group = TRUE, dense_header = FALSE)
}
\arguments{
\item{rsp}{boolean, \code{TRUE} if subject is a responder \code{FALSE}
otherwise}

\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group}

\item{group_data}{data frame with one column per grouping variable}

\item{total}{character with the row name of the analysis run on all data. If
\code{NULL} analysis is omitted}

\item{na.omit.group}{boolean, do not list NAs}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
Response Forest Plot Table
}
\details{
Logistic model is used for odds ratio calculation.

The returned table contains one row per analysis applied on a subset of data
(indicated by the row name). The analysis is summarized with the following 9
columns:

\describe{
  \item{1}{\emph{Total n} the total number of subjects included in analysis population}
  \item{3-4}{Summary of responders in reference arm, \emph{n} and \emph{Responders} are the 
  total number of patients and the number of responders in reference arm, respectively.
  \code{Response.Rate} is the percentage of responders in reference arm.}
  \item{5-7}{same statistics as for reference arm now for comparison arm}
  \item{8}{\emph{Odds Ratio} ranges from 0 to infinity, estimated by applying 
  univariate logistic regression with binary response status (responder or non-responder) 
  as the the outcome and arm variable as the explanatory variable. Odds ratio greater than 1 
  indicates better performance in comparison arm; odds ratio less than 1 indicates better performance in reference arm.}
  \item{9}{\emph{95 \% CI} The 95% confidence interval indicates the level of uncertainty 
  around the measure of effect (Odds Ratio). Because only a small sample of the overall 
  population is included in the analysis, by having an upper and lower confidence limit 
  we can infer that the true treatment effect lies in between. If the 95% confidence interval 
  includes 1, then we say that the difference between two arms is not significant at a significance level of 0.05.}
}
}
\examples{
library(random.cdisc.data)

ASL <- radam("ASL")
ARS <- radam("ARS", ADSL = ASL)

ARS_f <- subset(ARS, PARAMCD == "OVRSPI")
ANL <- merge(ASL, ARS_f)

tbl <- t_forest_rsp(
  rsp = ANL$AVALC \%in\% c("CR", "PR"),
  col_by = factor(ANL$ARM), 
  group_data = ANL[, c("SEX", "RACE")]
)

tbl

\dontrun{
Viewer(tbl)
}
   
}
\author{
Yuyao Song (songy24) \email{yuyao.song@roche.com}
}
