% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_rsp.R
\name{t_forest_rsp}
\alias{t_forest_rsp}
\title{Response Table as used for Forest Plot}
\usage{
t_forest_rsp(rsp, col_by, group_data = NULL, total = "ALL",
  na_omit_group = TRUE, dense_header = FALSE)
}
\arguments{
\item{rsp}{is a boolean vector. If \code{TRUE}, observation is a response,
otherwise, \code{FALSE}.}

\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group.}

\item{group_data}{data frame with one variable per grouping}

\item{total}{character string that will be used as a label for a column with total population. If the levels of
\code{col_by} are the only columns of interest then total should be \code{NULL}.}

\item{na_omit_group}{is boolean. Default is \code{TRUE}, do not display \code{NA} as a category.}

\item{dense_header}{Display the table headers in multiple rows. Default is \code{FALSE}.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
The Response forest plot table summarizes response data by groups. The
function returns sample sizes and responder counts and response rates for
each analysis arm, as well as a odds ratio and the corresponding 95\%
confidence interval from a univariate logistic model.
}
\details{
Logistic regression is used for odds ratio calculation.

Each row in the returned table contains the analysis statistics for a
subgroup of data (indicated by the row name). The summary table is consist of
the following 9 columns:

\describe{

  \item{1}{\emph{Total n} the total number of subjects included in analysis
  population}

  \item{3-4}{Summary of responders in reference arm, \emph{n} and
  \emph{Responders} are the total number of patients and the number of
  responders in reference arm, respectively. \code{Response.Rate} is the
  percentage of responders in reference arm.}

  \item{5-7}{same statistics as for reference arm now for comparison arm}

  \item{8}{\emph{Odds Ratio} ranges from 0 to infinity, estimated by applying
  univariate logistic regression. Binary response status (responder or
  non-responder) is the outcome and arm is the explanatory variable. Odds
  ratio greater than 1 indicates better performance in comparison arm; odds
  ratio less than 1 indicates better performance in reference arm.}

  \item{9}{\emph{95 \% CI} The 95% confidence interval indicates the level of
  uncertainty around the measure of effect (Odds Ratio). Because only a small
  sample of the overall population is included in the analysis, by having an
  upper and lower confidence limit we can infer that the true treatment
  effect lies in between. If the 95% confidence interval includes 1, then we
  say that the difference between two arms is not significant at a
  significance level of 0.05.}

}
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADSL <- cadsl
ADSL$'FAKE Name > -1.3 Flag' <- rep(c('Y', 'N'), 50)

ADRS <- radrs(ADSL, seed = 2)
ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI") \%>\%
  dplyr::filter(ARM != 'C: Combination') \%>\%
  mutate(ARM = droplevels(ARM))

tbl <- t_forest_rsp(
  rsp = ADRS_f$AVALC \%in\% c("CR", "PR"),
  col_by = factor(ADRS_f$ARM),
  group_data = ADRS_f[, c("SEX", "RACE", "FAKE Name > -1.3 Flag")]
)

tbl

\dontrun{
Viewer(tbl)
}
}
\seealso{
\code{\link{t_rsp}}
}
\author{
Yuyao Song (songy24) \email{yuyao.song@roche.com}
}
