% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_tte.R
\name{t_forest_tte}
\alias{t_forest_tte}
\title{Time-to-event Table as used for Forest Plot}
\usage{
t_forest_tte(tte, is_event, col_by, by, total = "All",
  strata_data = NULL, ties = "exact", time_unit = "month",
  dense_header = FALSE, table_tree = FALSE)
}
\arguments{
\item{tte}{vector with numeric time to event data, currently no \code{NA} allowed}

\item{is_event}{is boolean, \code{TRUE} if event, \code{FALSE} if \code{tte}
is censored}

\item{col_by}{logical factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{by}{\code{list} or \code{data.frame} with one factor variable to calculate the \code{t_el_forest_tte}}

\item{strata_data}{currently not supported}

\item{ties}{the method used for tie handling in \code{\link[survival]{coxph}}.}

\item{time_unit}{The unit of median survival time. Default is \code{months}.}

\item{dense_header}{Display the table headers in multiple rows.}

\item{table_tree}{logical, if \code{FALSE} then a single \code{rtable} gets returned, if \code{FALSE} a
\code{table_tree} gets returned with is either a list or a nested list of elementary \code{rtable} objects.}
}
\description{
The time-to-event forest plot table summarizes time-to-event data by groups. The function returns event counts and
median survival time for each analysis arm, as well as a hazard ratio and the corresponding 95\% confidence interval
from a Cox proportional hazard model.
}
\details{
Cox proportional hazard model is used for hazard ratio calculation

The returned table contains one row per analysis within a subgroup of data
(indicated by the row name). The analysis is summarized with the following 9
columns:

\describe{
  \item{1}{\emph{Total n} the total number of subjects included in analysis population}
  \item{2-4}{Survival statistics for reference arm, \emph{n} and \emph{Events} are the total number
of patients and the number of events in reference arm, respectively.
  \code{Median (month)} is the survival time estimated by Kaplan-Meier method. Time
 unit can be modified per study needs.}
  \item{5-7}{same analysis as for reference arm now for comparison arm}
  \item{8}{\emph{Hazard Ratio} ranges from 0 to infinity. The hazard ratio is an estimate of
 the ratio of the hazard rate in the comparison group versus that in the reference group.
   Univariate Cox proportional hazard model is applied to obtain the estimated hazard ratio.
  Hazard ratio > 1 implies better treatment effect in reference arm, and hazard ratio < 1 when
  comparison arm is better. }
  \item{9}{\emph{95\% Wald CI} The 95% confidence interval indicates the level of uncertainty
  around the measure of effect (Hazard Ratio). Because only a small sample of the overall
  population is included in the analysis, by having an upper and lower confidence limit
  we can infer that the true treatment effect lies in between. If the 95% confidence interval
  includes 1, then we say that the difference between two arms is not significant at a significance level of 0.05.}
}
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(seed = 1)

ADTTE <- radtte(ADSL, seed = 2)
ADTTE_f <- ADTTE \%>\%
  dplyr::filter(PARAMCD == "OS", ARMCD \%in\% c("ARM B", "ARM A")) \%>\%
  mutate(ARMCD = droplevels(ARMCD))

tbl <- t_forest_tte(
  tte = ADTTE_f$AVAL,
  is_event = ADTTE_f$CNSR == 0,
  col_by = ADTTE_f$ARMCD,
  by = droplevels(ADTTE_f[, c("SEX", "RACE")]),
  ties = "exact",
  dense_header = TRUE
)

tbl

\dontrun{
Viewer(tbl)
}

# table tree
tbls <- t_forest_tte(
  tte = ADTTE_f$AVAL,
  is_event = ADTTE_f$CNSR == 0,
  col_by = ADTTE_f$ARMCD,
  by = droplevels(ADTTE_f[, c("SEX", "RACE")]),
  ties = "exact",
  dense_header = TRUE,
  table_tree = TRUE
)
summary(tbls)
tbl <- rbind_table_tree(lapply(tbls, rbindl_rtables))

}
\seealso{
\code{\link{t_tte}}
}
\author{
Yuyao Song (songy24) \email{yuyao.song@roche.com}
}
