% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_tte.R
\name{t_forest_tte}
\alias{t_forest_tte}
\title{Time-to-event Forest Plot Table}
\usage{
t_forest_tte(tte, is_event, col_by, group_data = NULL, strata_data = NULL,
  total = "ALL", time_unit = "month", na.omit.group = TRUE,
  dense_header = FALSE)
}
\arguments{
\item{tte}{time to event data}

\item{is_event}{is boolean, \code{TRUE} if event, \code{FALSE} if \code{tte}
is censored}

\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group}

\item{group_data}{data frame with one column per grouping}
}
\description{
Time-to-event Forest Plot Table
}
\details{
Cox PH model is used for hazard ratio calculation


The returned table contains one row per analysis applied on a subset of data
(indicated by the row name). The analysis is summarized with the following 9
columns:

\describe{
  \item{1}{\emph{Total n} the total number of subjects used for analysis}
  \item{2-4}{analysis for reference arm, \emph{n}, \emph{Events},
  \code{Median (moneth)}}
  \item{5-7}{same analysis as for reference arm now for comparison arm}
  \item{8}{\emph{Hazard Ratio}}
  \item{9}{\emph{95\% Wald CI}}
}
}
\examples{

library(random.cdisc.data)

ASL <- radam("ASL")
ATE <- radam("ATE", ADSL = ASL)

ATE_f <- subset(ATE, PARAMCD == "OS") 

ANL <- merge(ASL, ATE_f)

tbl <- t_forest_tte(
  tte = ANL$AVAL,
  is_event = ANL$CNSR == 0,
  col_by = factor(ANL$ARM), 
  group_data = as.data.frame(lapply(ANL[, c("SEX", "RACE")], as.factor))
)

tbl

}
\author{
Yuyao Song (songy24) \email{yuyao.song@roche.com}
}
