% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forest_tte.R
\name{t_forest_tte}
\alias{t_forest_tte}
\title{Time-to-event Forest Plot Table}
\usage{
t_forest_tte(tte, is_event, col_by, group_data = NULL, strata_data = NULL,
  total = "ALL", time_unit = "month", na.omit.group = TRUE,
  dense_header = FALSE)
}
\arguments{
\item{tte}{time to event data}

\item{is_event}{is boolean, \code{TRUE} if event, \code{FALSE} if \code{tte}
is censored}

\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group}

\item{group_data}{data frame with one column per grouping}
}
\description{
Time-to-event Forest Plot Table
}
\details{
Cox propotionl hazard model is used for hazard ratio calculation


The returned table contains one row per analysis within a subgroup of data
(indicated by the row name). The analysis is summarized with the following 9
columns:

\describe{
  \item{1}{\emph{Total n} the total number of subjects included in analysis population}
  \item{2-4}{Survival statistics for reference arm, \emph{n} and \emph{Events} are the total number of patients and the number of events in reference arm, respectively.
  \code{Median (month)} is the survival time estimated by Kaplan-Meier method. Time unit can be modified per study needs.}
  \item{5-7}{same analysis as for reference arm now for comparison arm}
  \item{8}{\emph{Hazard Ratio} ranges from 0 to infinity. The hazard ratio is an estimate of the ratio of the hazard rate in the comparison group versus that in the reference group.
   Univariate Cox proportional hazard model is applied to obtain the estimated hazard ratio. 
  Hazard ratio > 1 implies better treatment effect in reference arm, and hazard ratio < 1 when comparion arm is better. }
  \item{9}{\emph{95\% Wald CI} The 95% confidence interval indicates the level of uncertainty 
  around the measure of effect (Hazard Ratio). Because only a small sample of the overall 
  population is included in the analysis, by having an upper and lower confidence limit 
  we can infer that the true treatment effect lies in between. If the 95% confidence interval 
  includes 1, then we say that the difference between two arms is not significant at a significance level of 0.05.}
}
}
\examples{

library(random.cdisc.data)

ASL <- radam("ASL")
ATE <- radam("ATE", ADSL = ASL)

ATE_f <- subset(ATE, PARAMCD == "OS") 

ANL <- merge(ASL, ATE_f)

tbl <- t_forest_tte(
  tte = ANL$AVAL,
  is_event = ANL$CNSR == 0,
  col_by = factor(ANL$ARM), 
  group_data = as.data.frame(lapply(ANL[, c("SEX", "RACE")], as.factor)),
  dense_header = TRUE
)

tbl
Viewer(tbl)

# RealWorld example
library(haven)
ASL <- read_sas("Y:/cdt30099/s30081i/libraries/aslpd.sas7bdat")
ARS <- read_sas("Y:/cdt30099/s30081i/libraries/ars.sas7bdat")
ATE <- read_sas("Y:/cdt30099/s30081i/libraries/ate.sas7bdat")
APDTE <- read_sas("Y:/cdt30099/s30081i/libraries/apdte.sas7bdat")

ATE <- rbind(ATE, APDTE)
ARS <- merge(ARS, subset(ASL, select = c("STUDYID", "USUBJID", "STRAT1", "STRAT2", "STRAT3")), by = c("STUDYID", "USUBJID"))
ATE <- merge(ATE, subset(ASL, select = c("STUDYID", "USUBJID", "STRAT1", "STRAT2", "STRAT3")), by = c("STUDYID", "USUBJID"))

ATE_f <- subset(ATE, PARAMCD == "OS") 
ANL <- merge(ASL, ATE_f)

tbl <- t_forest_tte(
  tte = ANL$AVAL,
  is_event = ANL$CNSR == 0,
  col_by = factor(ANL$ARM), 
  group_data = as.data.frame(lapply(ANL[, c("SEX", "RACE")], as.factor)),
  dense_header = TRUE
)

tbl
Viewer(tbl)

}
\author{
Yuyao Song (songy24) \email{yuyao.song@roche.com}
}
