% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_km.R
\name{t_km}
\alias{t_km}
\title{Kaplan-Meier Fit Summary Table}
\usage{
t_km(fit_km)
}
\arguments{
\item{fit_km}{a class \code{\link{survfit}} object.}
}
\description{
An \code{\link[rtables]{rtable}} format of KM model data for further
annotation on top of Kaplan-Meier grob
}
\examples{
library(random.cdisc.data)

ADSL <- cadsl
ADSL$RACE <- factor(
  vapply(as.character(ADSL$RACE),
         function(x) {
           if (nchar(x)>9) paste0(substr(x, 1,9), "...") else x
         },
         character(1)
  )
)

ADTTE <- cadtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

fit_km <- survfit(Surv(AVAL, 1-CNSR) ~ ARM, data = ADTTE_f, conf.type = "plain")
tbl <- t_km(fit_km)
tbl

fit_km <- survfit(Surv(AVAL, 1-CNSR) ~ 1, data = ADTTE_f, conf.type = "plain")
tbl <- t_km(fit_km)
tbl
}
\seealso{
\code{\link{g_km}}
}
\author{
Heng Wang (wangh107) \email{wang.heng@gene.com}
}
