% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_km.R
\name{t_km}
\alias{t_km}
\title{Kaplan-Meier Fit Summary Table}
\usage{
t_km(formula, data, conf.int = 0.95, ...)
}
\arguments{
\item{formula}{(\code{formula})\cr
Specifies \code{\link[survival:Surv]{survival model}}}

\item{data}{(\code{data.frame})\cr with all the variable that are used in
\code{formula}}

\item{conf.int}{(\code{numeric} value)\cr
level for computation of the confidence intervals.}

\item{...}{(optional)\cr
additional parameters passed to \code{\link{survfit}}}
}
\description{
An \code{\link[rtables]{rtable}} format of KM model data for further
annotation on top of Kaplan-Meier grob
}
\examples{
library(random.cdisc.data)

ADSL <- cadsl
ADSL$RACE <- factor(
  vapply(as.character(ADSL$RACE),
         function(x) {
           if (nchar(x)>9) paste0(substr(x, 1,9), "...") else x
         },
         character(1)
  )
)

ADTTE <- cadtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

formula <- Surv(AVAL, 1-CNSR) ~ ARM
tbl <- t_km(formula, data = ADTTE_f, conf.type = "plain")
tbl

tbl <- t_km(formula, data = ADTTE_f, conf.int = 0.8)
tbl
}
\seealso{
\code{\link{g_km}}
}
\author{
Heng Wang (wangh107) \email{wang.heng@gene.com}
}
