% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_events_term_grade_id.R
\name{t_max_grade_per_id}
\alias{t_max_grade_per_id}
\title{Tabulate maximum grade per id by \code{col_by}}
\usage{
t_max_grade_per_id(
  grade,
  id,
  col_by,
  col_N = NULL,
  total = NULL,
  any_grade = "Any Grade",
  prune_zero_rows = TRUE
)
}
\arguments{
\item{grade}{`factor` grade of adverse event.
It is assumed that intensity corresponds to the order of the factor levels.}

\item{id}{a vector with id values}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{(\code{character} value)\cr
Denotes column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{any_grade}{add a row that counts any occurrence, it is named \code{Any
Grade} by default}

\item{prune_zero_rows}{\code{logical} whether to prune all zero rows}
}
\description{
\code{t_max_grade_per_id} is used for deriving adverse events tables, these are returned
as nested lists.
}
\examples{
t_max_grade_per_id(
  grade = factor(c(1, 2, 3)),
  id = c(1,1,1),
  col_by = factor(rep("A", 3)),
  col_N = 3
)

# don't prune zero rows
t_max_grade_per_id(
  grade = factor(c(1, 2, 3)),
  id = c(1,1,1),
  col_by = factor(rep("A", 3)),
  col_N = 3,
  prune_zero_rows = FALSE
)

id <- c(1,1,2,2,3,3,3,4)
grade <- factor(c("a", "b", "c", "b", "a", "a", "b", "c"))
levels(grade) <- c("d", "c", "b", "a")
t_max_grade_per_id(
  grade = grade,
  id = id,
  col_by = factor(LETTERS[id]),
  col_N = c(4, 3, 5, 3),
  any_grade = "Any Class"
)

t_max_grade_per_id(
  grade = factor(c(1, 2, 3, 2, 1, 1, 2, 3)),
  id = id,
  col_by = factor(LETTERS[id]),
  col_N = c(4, 3, 5, 4)
)

\dontrun{
# throws an error because each id can only have one col_by
# t_max_grade_per_id(
  #   grade = factor(c(1, 2, 3)),
  #   id = c(1, 2, 2),
  #   col_by = factor(LETTERS[1:3]),
  #   col_N = c(15, 10, 12)
  # )
}

\dontrun{
# throws an error because grade should not be NA
# t_max_grade_per_id(
#   grade =  factor(c(1, 2, NA)),
#   id = c(1, 2, 3),
#   col_by = factor(LETTERS[1:3]),
#   col_N = c(15, 10, 12)
# )
}
}
\seealso{
\code{\link{t_max_grade_per_id}}, \code{\link{t_events_per_term_grade_id}}
}
