% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_events_term_grade_id.R
\name{t_max_grade_per_id}
\alias{t_max_grade_per_id}
\title{Tabulate maximum grade per id by \code{col_by}}
\usage{
t_max_grade_per_id(grade, id, col_by, col_N = table(col_by),
  grade_levels = NULL, any_grade = "-Any Grade-")
}
\arguments{
\item{grade}{a numeric vector with grade values}

\item{id}{a vector with id values}

\item{col_by}{a factor with values used for column names}

\item{col_N}{a vector with total n for each level of \code{col_by}}

\item{grade_levels}{a numeric vector used for naming rows for each level of
grade}

\item{any_grade}{add a row that counts any occurrence, it is named \code{-Any
Grade-} by default}
}
\description{
\code{t_max_grade_per_id} is used for deriving adverse events tables, these are returned
as nested lists.
}
\examples{

t_max_grade_per_id(
  grade =  c(1,2,3),
  id = c(1,1,1),
  col_by = factor(rep("A", 3)),
  col_N = 3
)

id <- c(1,1,2,2,3,3,3,4)
t_max_grade_per_id(
  grade =  c(1,2,3,2,1,1,2,3),
  id = id,
  col_by = factor(LETTERS[id]),
  col_N = c(4,3,5,3)
)

t_max_grade_per_id(
  grade =  c(1,2,3,2,1,1,2,3),
  id = id,
  col_by = factor(LETTERS[id]),
  col_N = c(4,3,5,4)
)

\dontrun{
# throws an error because each id can only have one col_by
t_max_grade_per_id(
  grade =  c(1,2,3),
  id = c(1,2,2),
  col_by = factor(LETTERS[1:3])
)
}

\dontrun{
# throws an error because grade NA is not in grade_levels 
t_max_grade_per_id(
  grade =  c(1,2,NA),
  id = c(1,2,3),
  col_by = factor(LETTERS[1:3])
)
}

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}

Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}

Heng Wang (wangh107) \email{wang.heng@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
