% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_mmrm.R
\name{t_mmrm}
\alias{t_mmrm}
\title{Mix model with repeated measurements (MMRM) model}
\usage{
t_mmrm(
  formula = AVAL ~ arm(ARM) + visit(AVISIT) + ARM * VISIT,
  data,
  id_var = "USUBJID",
  arm_var = "ARM",
  visit_var = "AVISIT",
  col_N,
  mode = c("df.error", "auto", "boot-satterthwaite"),
  conf.level = 0.95,
  weights_emmeans = "proportional",
  corStruct = NULL,
  table_tree = TRUE
)
}
\arguments{
\item{formula}{a gls formula.}

\item{data}{a \code{data.frame} with all the variables specified in
\code{formula}. Records with missing values in any independent variables
will be excluded.}

\item{id_var}{a character describing the variable name used as subject IDs,
"USUBJID" by default.}

\item{arm_var}{a character describing the variable name for \code{\link{arm}},
"ARM" by default. The arm variable must be factor in \code{data}}

\item{visit_var}{a character describing the variable name for visit,
"AVISIT" by default. This variable must be factor in data.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default
\code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.}

\item{mode}{algorithm for degree of freedom: "auto", "df.error" or
"boot-satterthwaite".}

\item{conf.level}{confidence level. Must be number greater than 0 and less
than 1.}

\item{weights_emmeans}{argument from emmeans, "proportional" by default.}

\item{corStruct}{\code{NULL} by default or a string with the name of \link[nlme]{corClasses}.}

\item{table_tree}{logical, if \code{FALSE} then a single \code{rtable} gets returned, if \code{TRUE} a
\code{\link{node}} object gets returned with the elementary \code{rtable} objects.}
}
\value{
rtable object or table tree, depending on the the `table_tree` argument
}
\description{
The MMRM table function summarizes MMRM test results by visit and groups. The
function produces adjusted lsmeans and standard error, as well as conducts
comparisons between groups' adjusted means, where the first level of the group
is the reference level.
}
\details{
\url{https://pages.github.roche.com/NEST/docs/bookdown/NEST/statistics_clinical_trials/master/mixed-effect-model-repeat-measurement.html}

Create rtable of MMRM test results
}
\examples{
library(dplyr)
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADQS <- radqs(cached = TRUE)
ADQS_f <- ADQS \%>\%
  filter(PARAMCD=="FKSI-FWB" & !AVISIT \%in\% c("BASELINE")) \%>\%
  droplevels() \%>\%
  mutate(ARM = factor(ARM, levels = c("B: Placebo", "A: Drug X", "C: Combination"))) \%>\%
  mutate(AVISITN = rank(AVISITN) \%>\% as.factor() \%>\% as.numeric() \%>\% as.factor())

t_mmrm(formula = AVAL ~ ARM + AVISIT + STRATA1 + BMRKR2 + BASE + ARM*AVISIT,
       data = ADQS_f,
       id_var = "USUBJID",
       arm_var = "ARM",
       visit_var = "AVISIT",
       col_N = table(ADSL$ARM),
       mode = "boot-satterthwaite",
       conf.level = 0.95,
       weights_emmeans = "proportional",
       corStruct = "corSymm",
       table_tree = FALSE
)

}
\seealso{
\code{\link{a_mmrm}}
}
