% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_rsp.R
\name{t_rsp}
\alias{t_rsp}
\title{Response Table Caclulation Function}
\usage{
t_rsp(rsp, col_by, parition_rsp_by = NULL, strata_data = NULL)
}
\arguments{
\item{rsp}{boolean, \code{TRUE} if subject is a responder \code{FALSE}
otherwise}

\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group}

\item{parition_rsp_by}{factor to partition \code{rsp} and run additional
analysis}

\item{strata_data}{data for stratification factors (categorical variables),
if \code{NULL} no stratified analysis is performed}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
TODO: Description
}
\details{
If no stratification, chi-squared test is performed by default to test for
equality of proportions in response rate between reference and comparison
arms.

If stratification factors are added, Cochran-Mantel-Haensel test statistics
are reported for the test of equal proportions, odds ratio and its
corresponding 95% confidence interval.
}
\examples{

library(random.cdisc.data)

ASL <- radam("ASL", N = 1000, start_with = list(RACE = c("white", "asian")))
ARS <- radam("ARS", ADSL = ASL)

ASL$ARM <- factor(sample(LETTERS[1:3], nrow(ASL), TRUE))
ASL$RACE <- factor(ASL$RACE)

ANL <- merge(ASL, subset(ARS, PARAMCD == "OVRSPI"))

tbl <- t_rsp(
 rsp = ANL$AVALC \%in\% c("CR", "PR"),
 col_by = relevel(factor(ANL$ARMCD), "ARM B", "ARM A"),
 parition_rsp_by = droplevels(factor(ANL$AVALC, levels = c("CR", "PR", "SD", "NON CR/PD", "PD", "NE"))),
 strata_data = ANL[, c("SEX", "RACE")]
)

tbl

}
\author{
Chendi Liao (liaoc10) \email{chendi.liao@roche.com}
}
