% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarize_by_visit.R
\name{t_summarize_by_visit}
\alias{t_summarize_by_visit}
\title{Summary Table by Visit}
\usage{
t_summarize_by_visit(data, visit, id, col_by, col_N = NULL)
}
\arguments{
\item{data}{(\code{data.frame})\cr
with numerical variables to be summarized. If the
variable has a \code{label} attribute then it will be used as the
sub-header column name.}

\item{visit}{(\code{factor} vector)\cr
with visit names ordered by desired display order in the
  stacked table.}

\item{id}{(\code{vector})\cr
unique subject identifier variable.}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
This function summarizes test results or change from baseline statistics by
visit. Corresponds to STREAM table templates \code{EGT01}, \code{VST01} and \code{LBT01}.
}
\details{
Every variable in \code{data} will be mapped to a sub-column under
  \code{col_by} column. Multiple variables will be shown side-by-side with
  the variable label (if present) or variable name as the sub-column header.

  Currently \emph{n}, \emph{Mean (SD)}, \emph{Median}, \emph{IQR} and
  \emph{Min-Max} will be derived for each variable from \code{data} per each
  factor in \code{col_by}.
}
\examples{

library(random.cdisc.data)
library(dplyr)
ADSL <- radsl(cached = TRUE)
ADVS <- radvs(cached = TRUE)

tbl1 <- t_summarize_by_visit(
  data = ADVS[c("AVAL")],
  visit = ADVS$AVISIT,
  col_by = ADVS$ARM,
  id = ADVS$USUBJID,
  col_N = table(ADSL$ARM)
)
\dontrun{
Viewer(tbl1)
}

tbl2 <- t_summarize_by_visit(
  data = ADVS[c("PCHG")],
  visit = ADVS$AVISIT,
  col_by = ADVS$ARM,
  id = ADVS$USUBJID,
  col_N = table(ADSL$ARM)
)
\dontrun{
Viewer(tbl2)
}

tbl3 <- t_summarize_by_visit(
  data = ADVS[c("AVAL", "CHG")],
  visit = ADVS$AVISIT,
  col_by = ADVS$ARM,
  id = ADVS$USUBJID,
  col_N = table(ADSL$ARM)
)
\dontrun{
Viewer(tbl3)
}


ADSL <- cadsl
ADQS <- cadqs \%>\%
  dplyr::filter(PARAMCD == "BFIALL")

tbl <- t_summarize_by_visit(
  data = ADQS[c("AVAL", "CHG")],
  visit = ADQS$AVISIT,
  col_by = ADQS$ARM,
  id = ADQS$USUBJID,
  col_N = table(ADSL$ARM)
)

\dontrun{
Viewer(tbl)
}
}
\author{
Chendi Liao (liaoc10) \email{chendi.liao@roche.com}
}
