% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarize_by_visit.R
\name{t_summarize_by_visit}
\alias{t_summarize_by_visit}
\title{Summary Table by Visit}
\usage{
t_summarize_by_visit(data, visit, id, col_by, col_N = NULL)
}
\arguments{
\item{data}{data frame with numerical variables to be summarized. If the
variable has a \code{label} attribute then it will be used as the
sub-header column name.}

\item{visit}{factor with visit names ordered by desired display order in the
stacked table.}

\item{id}{unique subject identifier variable.}

\item{col_by}{logical factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default
\code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
This function summarizes test results or change from baseline statistics by
visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.
}
\details{
Every variable in \code{data} will be mapped to a sub-column under
  \code{col_by} column. Multiple variables will be shown side-by-side with
  the variable label (if present) or variable name as the sub-column header.

  Currently \emph{n}, \emph{Mean (SD)}, \emph{Median}, \emph{IQR} and
  \emph{Min-Max} will be derived for each variable from \code{data} per each
  factor in \code{col_by}.
}
\examples{
# EXAMPLE 1
library(random.cdisc.data)

ADSL <- cadsl
ADVS <- cadvs

t_summarize_by_visit(data = ADVS[c("AVAL")], visit = ADVS$AVISIT, col_by = ADVS$ARM,
  id = ADVS$USUBJID, col_N = table(ADSL$ARM))

t_summarize_by_visit(data = ADVS[c("PCHG")], visit = ADVS$AVISIT, col_by = ADVS$ARM,
  id = ADVS$USUBJID, col_N = table(ADSL$ARM))

t_summarize_by_visit(
  data = ADVS[c("AVAL", "CHG")],
  visit = ADVS$AVISIT,
  col_by = ADVS$ARM,
  id = ADVS$USUBJID,
  col_N = table(ADSL$ARM)
)

# EXAMPLE 2
library(random.cdisc.data)
library(dplyr)

ADSL <- cadsl
ADQS <- cadqs \%>\%
  dplyr::filter(PARAMCD == "BFIALL")

tbl <- t_summarize_by_visit(
  data = ADQS[c("AVAL", "CHG")],
  visit = ADQS$AVISIT,
  col_by = ADQS$ARM,
  id = ADQS$USUBJID,
  col_N = table(ADSL$ARM)
)

tbl

\dontrun{
Viewer(tbl)
}

}
\author{
Chendi Liao (liaoc10) \email{chendi.liao@roche.com}
}
