% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarize_by_visit.R
\name{t_summarize_by_visit}
\alias{t_summarize_by_visit}
\title{Summary Table by Visit}
\usage{
t_summarize_by_visit(data, visit, col_by)
}
\arguments{
\item{data}{data frame with numerical variables to be summarized. If the 
variable has a \code{label} attribute then it will be used as the column 
name.}

\item{visit}{factor with visit names ordered by desired display order in the
stacked table.}

\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
This function summarizes test results or change from baseline statistics by 
visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.
}
\details{

}
\examples{

ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")   
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

ANL <- var_relabel(ANL, AVAL = "Value at\\nVisit", CHG = "Change from\\nBaseline")



t_summarize_by_visit(data = ANL$AVAL, visit = ANL$VISIT, col_by = ANL$ARM)

t_summarize_by_visit(data = ANL$CHG, visit = ANL$VISIT, col_by = ANL$ARM)

arm <- as.character(ANL$ARM)
ARM <- factor(c(arm, paste0(arm,"-chng")), levels = c("ARM A", "ARM A-chng", "ARM B", "ARM B-chng"))

t_summarize_by_visit(data = c(ANL$AVAL, ANL$CHG), visit = rep(ANL$VISIT, 2), ARM)

}
\author{
Chendi Liao (liaoc10) \email{chendi.liao@roche.com}
}
