% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarize_by_visit.R
\name{t_summarize_by_visit}
\alias{t_summarize_by_visit}
\title{Summary Table by Visit}
\usage{
t_summarize_by_visit(data, visit, id, col_by)
}
\arguments{
\item{data}{data frame with numerical variables to be summarized. If the 
variable has a \code{label} attribute then it will be used as the 
sub-header column name.}

\item{visit}{factor with visit names ordered by desired display order in the 
stacked table.}

\item{id}{unique subject identifier variable.}

\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
This function summarizes test results or change from baseline statistics by 
visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.
}
\details{
Every variable in \code{data} will be mapped to a sub-column under 
  \code{col_by} column. Multiple variables will be shown side-by-side with
  the variable label (if present) or variable name as the sub-column header.
  
  Currently \emph{n}, \emph{Mean (SD)}, \emph{Median}, \emph{IQR} and 
  \emph{Min-Max} will be derived for each variable from \code{data} per each 
  factor in \code{col_by}.
}
\examples{

# EXAMPLE 1
ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")   
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

t_summarize_by_visit(data = ANL[c("AVAL")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)
t_summarize_by_visit(data = ANL[c("PCHG")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

# Add label to variable instead showing variable name
ANL <- var_relabel(ANL, AVAL = "Value at\\nVisit",
                        CHG = "Change from\\nBaseline",
                        PCHG = "Percent Change\\nfrom Baseline")
                        
t_summarize_by_visit(
  data = ANL[c("AVAL", "CHG")],
  visit = ANL$VISIT,
  col_by = ANL$ARM,
  id = ANL$USUBJID
)


# EXAMPLE 2
library(random.cdisc.data)
library(dplyr)

AQS <- radam("AQS") \%>\%
   filter(PARAMCD == "BFIALL") \%>\% 
   mutate(
     ARM = factor(sample(c("ARM A", "ARM B", "ARM C"), nrow(.), replace = TRUE)),
     AVISIT = factor(AVISIT, levels = unique(AVISIT))
   )

tbl <- t_summarize_by_visit(
  data = AQS[c("AVAL", "CHG")],
  visit = AQS$AVISIT,
  col_by = AQS$ARM,
  id = AQS$USUBJID
)
tbl
\dontrun{
Viewer(tbl)
}

}
\author{
Chendi Liao (liaoc10) \email{chendi.liao@roche.com}
}
