% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarize_variables.R
\name{t_summarize_variables}
\alias{t_summarize_variables}
\title{Table of Variable Summaries}
\usage{
t_summarize_variables(data, col_by, total = NULL)
}
\arguments{
\item{data}{data frame}

\item{col_by}{a vector of length \code{nrow(x)} that defines which levels in
\code{col_by} define a column. If data should not be split into columns use
the \code{\link{no_by}} function.}

\item{total}{if not \code{NULL} then it must be a string and an addition
column will be added with the overall summaries}
}
\description{
Similiar as the demographic table in STREAM
}
\details{
Give a detailed description of what this function does.
}
\examples{

library(tern)
library(random.cdisc.data)
library(forcats)

ASL <- radam("ASL")

# control the label
attr(ASL$BAGE, "label") <- "Baseline Age of patient"
attr(ASL$SEX, "label")

# control categorical order
ASL$SEX <- fct_relevel(ASL$SEX, "M", "F")

# control arm order
ASL$ARM <- fct_relevel(ASL$ARM, "ARM B", "ARM A")

t_summarize_variables(ASL \%>\% select(SEX, BAGE), col_by = ASL$ARM, total = "All Patients")

t_summarize_variables(iris \%>\% select(-Species), col_by  = iris$Species)

t_summarize_variables(iris \%>\% select(-Species), col_by  = no_by("All Species"))

}
\author{
Adrian Waddell \email{adrian.waddell@roche.com} and Xiao Yu Mo (mox5), \email{xiao_yu.mo@roche.com}
}
