% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.character}
\alias{t_summary.character}
\title{Summarize Character Data}
\usage{
\method{t_summary}{character}(x, col_by, col_N = table(col_by),
  total = NULL, ...)
}
\arguments{
\item{x}{a character vector}

\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default}

\item{total}{character string that will be used as a label for a column with total population. If the levels of
\code{col_by} are the only columns of interest then total should be \code{NULL}.}

\item{...}{arguments passed on to \code{\link{t_summary.factor}}}
}
\description{
Currently treated as factors
}
\seealso{
\code{\link{t_summary}},
  \code{\link{t_summary.data.frame}}, \code{\link{t_summary.factor}},
  \code{\link{t_summary.logical}}, \code{\link{t_summary.Date}},
  \code{\link{t_summary.numeric}}, \code{\link{t_summary_by}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
