% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.character}
\alias{t_summary.character}
\title{Summarize Character Data}
\usage{
\method{t_summary}{character}(x, ...)
}
\arguments{
\item{x}{a character vector}

\item{...}{arguments passed on to methods}
}
\description{
Currently treated as factors
}
\seealso{
\code{\link{t_summary}},
  \code{\link{t_summary.data.frame}}, \code{\link{t_summary.factor}},
  \code{\link{t_summary.logical}}, \code{\link{t_summary.Date}},
  \code{\link{t_summary.numeric}}, \code{\link{t_summary_by}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
