% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.character}
\alias{t_summary.character}
\title{Summarize Character Data}
\usage{
\method{t_summary}{character}(x, col_by, col_N = table(col_by),
  total = NULL, ...)
}
\arguments{
\item{x}{a character vector}

\item{col_by}{a \code{\link{factor}} of length \code{nrow(x)} that defines
which levels in \code{col_by} define a column. If data should not be split
into columns use the \code{\link{no_by}} function.}

\item{col_N}{The column total for each group that is displayed in the table header with (N=xx).}

\item{total}{character string that will be used as a label for a column with pooled total population. If the levels of \code{col_by} are the only columns of interest then total should be \code{NULL}.}

\item{...}{arguments passed on to \code{\link{t_summary.factor}}}
}
\description{
Currently treated as factors
}
\seealso{
\code{\link{t_summary}},
  \code{\link{t_summary.data.frame}}, \code{\link{t_summary.factor}},
  \code{\link{t_summary.logical}}, \code{\link{t_summary.Date}},
  \code{\link{t_summary.numeric}}, \code{\link{t_summary_by}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
