% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.data.frame}
\alias{t_summary.data.frame}
\title{Variables Summary Table}
\usage{
\method{t_summary}{data.frame}(x, col_by, total = NULL, ...)
}
\arguments{
\item{x}{data frame with variables to be summarized as described in the
details section. If the variable has a \code{label} attribute then it will
be used for the row name.}

\item{col_by}{a \code{\link{factor}} of length \code{nrow(x)} that defines
which levels in \code{col_by} define a column. If data should not be split
into columns use the \code{\link{no_by}} function.}

\item{total}{if not \code{NULL} then it must be a string and an addition
column will be added with the overall summaries}

\item{...}{not used arguments}
}
\description{
Similiar as the demographic table in STREAM
}
\details{
Every variable in \code{x} will be mapped to a summary table using
\code{\link{t_summary}} and then be stacked.
}
\examples{

library(random.cdisc.data)

ASL <- radam("ASL")

# control the label
ASL <- var_relabel(ASL, BAGE = "Baseline Age of patient")

# control categorical order
ASL$SEX <- relevel(ASL$SEX, "M", "F")

t_summary(ASL[, c("SEX", "BAGE")], col_by = ASL$ARM, total = "All Patients")

t_summary(ASL[, c("SEX", "BAGE")], col_by = ASL$ARM, total = "All Patients",
                      useNA = 'always')

ASL$SEX[1:10] <- NA

t_summary(ASL[, c("SEX", "BAGE")], col_by = ASL$ARM, total = "All Patients",
                      useNA_factors = 'ifany')

# with iris data

t_summary(iris[, -5], col_by  = iris$Species)

t_summary(iris, col_by  = no_by("All Species"))


x <- factor(c("A", NA, "B", "B", "A"))
cb <- factor(c("I", "I", "I", "II", "II"))

t_summary(x, col_by = cb)

t_summary(data.frame(x), col_by = cb, useNA = "ifany")

}
\seealso{
\code{\link{t_summary}}, \code{\link{t_summary.factor}}, \code{\link{t_summary.numeric}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
