% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.data.frame}
\alias{t_summary.data.frame}
\title{Variables Summary Table}
\usage{
\method{t_summary}{data.frame}(x, col_by, col_N = default_col_N(col_by),
  total = NULL, ..., table_tree = FALSE)
}
\arguments{
\item{x}{data frame with variables to be summarized as described in the
details section. If the variable has a \code{label} attribute then it will
be used for the row name.}

\item{col_by}{logical factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default
\code{table(col_by)}.}

\item{total}{(deprecated in favor of \code{col_by}) character string that will be used as a label for a column with total population. If the levels of
\code{col_by} are the only columns of interest then total should be \code{NULL}.}

\item{...}{arguments passed on to methods}

\item{table_tree}{boolean, whether the compound table function should return a table_tree object or the elementary
table should be stacked.}
}
\description{
Similar as the demographic table in STREAM
}
\details{
You can use the label attribute of the columns in the data frame to modify the node value / displayed value,
e.g. AGE -> "Baseline Age of patient"


Every variable in \code{x} will be mapped to a summary table using
\code{\link{t_summary}} and then be stacked.
}
\examples{
# with iris data
t_summary(iris$Sepal.Length, iris$Species)

library(random.cdisc.data)
library(dplyr)
library(rtables)
ADSL <- radsl(N = 100, seed = 1)

t_summary(ADSL[, c("SEX", "AGE")], col_by  = by_all("All"), col_N = nrow(ADSL))

# control the label
ADSL <- var_relabel(ADSL, AGE = "Baseline Age of patient")

# control categorical order
ADSL$SEX <- relevel(ADSL$SEX, "M", "F")

t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients")
t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients", useNA = 'always')

ADSL$SEX[1:10] <- NA
t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients", useNA = 'ifany')
t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients", denominator = "N",
          useNA = 'ifany')

ADSL_AGE65 <- ADSL \%>\%
  dplyr::filter(AGE > 65)
t_summary(ADSL_AGE65[, c("AGE", "SEX")], ADSL_AGE65$ARM, total = "All Patients",
          col_N = table(ADSL_AGE65$ARM), denominator = "N", drop_levels = TRUE)


tbls <- t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, table_tree = TRUE)
summary(tbls)
rbind_table_tree(tbls)

}
\seealso{
\code{\link{t_summary}},
  \code{\link{t_summary.numeric}}, \code{\link{t_summary.factor}},
  \code{\link{t_summary.logical}}, \code{\link{t_summary.Date}},
  \code{\link{t_summary.character}}, \code{\link{t_summary_by}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
