% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.data.frame}
\alias{t_summary.data.frame}
\title{Variables Summary Table}
\usage{
\method{t_summary}{data.frame}(x, col_by, col_N = table(col_by),
  total = NULL, ...)
}
\arguments{
\item{x}{data frame with variables to be summarized as described in the
details section. If the variable has a \code{label} attribute then it will
be used for the row name.}

\item{col_by}{a \code{\link{factor}} of length \code{nrow(x)} that defines
which levels in \code{col_by} define a column. If data should not be split
into columns use the \code{\link{no_by}} function.}

\item{col_N}{The column total for each group that is displayed in the table header with (N=xx).}

\item{total}{if not \code{NULL} then it must be a string and an addition
column will be added with the overall summaries}

\item{...}{arguments passed on to methods}
}
\description{
Similiar as the demographic table in STREAM
}
\details{
Every variable in \code{x} will be mapped to a summary table using
\code{\link{t_summary}} and then be stacked.
}
\examples{
# with iris data
t_summary(iris$Sepal.Length, iris$Species)

library(random.cdisc.data)
library(dplyr)
ADSL <- radsl(N = 100, seed = 1) 

t_summary(ADSL[, c("SEX", "AGE")], col_by  = no_by("All"), col_N = nrow(ADSL))

# control the label
ADSL <- var_relabel(ADSL, AGE = "Baseline Age of patient")

# control categorical order
ADSL$SEX <- relevel(ADSL$SEX, "M", "F")

t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients")
t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients",
         useNA = 'always')
         
ADSL$SEX[1:10] <- NA
t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients", 
         useNA = 'ifany')
t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients",
         denominator = "N", useNA = 'ifany')
         
ADSL_AGE65 <- ADSL \%>\% filter(AGE > 65)
t_summary(ADSL_AGE65[, c("AGE", "SEX")], ADSL_AGE65$ARM, total = "All Patients", 
         col_N = table(ADSL$ARM), denominator = "N", drop_levels = TRUE)

}
\seealso{
\code{\link{t_summary}}, \code{\link{t_summary.factor}}, \code{\link{t_summary.numeric}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
