% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.Date}
\alias{t_summary.Date}
\title{Summarize Date Data}
\usage{
\method{t_summary}{Date}(x, col_by, col_N = table(col_by), total = NULL,
  ...)
}
\arguments{
\item{x}{a Date object}

\item{col_by}{a \code{\link{factor}} of length \code{nrow(x)} that defines
which levels in \code{col_by} define a column. If data should not be split
into columns use the \code{\link{no_by}} function.}

\item{col_N}{The column total for each column. If specified then
\code{\link{col_N}()} can be used on the cell data in order to retrieve the
column total. If \code{NULL} then no header row for the column is
displayed.}

\item{...}{arguments passed on to \code{\link{t_summary.factor}}}
}
\description{
Currently treated as for factors
}
\examples{
(today <- Sys.Date())
(tenweeks <- seq(today, length.out=10, by="1 week")) 
t_summary(tenweeks, no_by("all"), length(tenweeks))

t_summary(tenweeks, factor(LETTERS[c(1,1,1,2,2,3,3,3,4,4)]))

t_summary(tenweeks, factor(LETTERS[c(1,1,1,2,2,3,3,3,4,4)]), total = "All")

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
