% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.Date}
\alias{t_summary.Date}
\title{Summarize Date Data}
\usage{
\method{t_summary}{Date}(x, col_by, col_N = NULL, total = NULL, ...)
}
\arguments{
\item{x}{a Date object}

\item{col_by}{logical factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default
\code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.}

\item{total}{string of column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{...}{arguments passed on to methods}
}
\description{
Tabulate the range of dates.
}
\examples{
today <- as.Date("2000-01-01")
library(dplyr)

today <- Sys.Date()
tenweeks <- seq(today, length.out=10, by="1 week")
t_summary(tenweeks, by_all("all"), length(tenweeks))

t_summary(tenweeks, factor(LETTERS[c(1,1,1,2,2,3,3,3,4,4)]))

t_summary(tenweeks, col_by = factor(LETTERS[c(1,1,1,2,2,3,3,3,4,4)]), total = "All")
}
\seealso{
\code{\link{t_summary}},
  \code{\link{t_summary.data.frame}}, \code{\link{t_summary.factor}},
  \code{\link{t_summary.logical}}, \code{\link{t_summary.numeric}},
  \code{\link{t_summary.character}}, \code{\link{t_summary_by}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
