% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.default}
\alias{t_summary.default}
\title{Return an \code{rtable} with a no Method Message}
\usage{
\method{t_summary}{default}(x, col_by, col_N = get_N(col_by), ...)
}
\arguments{
\item{x}{data frame with variables to be summarized as described in the
details section. If the variable has a \code{label} attribute then it will
be used for the row name.}

\item{col_by}{logical factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default
\code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.}

\item{...}{not used arguments}
}
\value{
rtable
}
\description{
If there is no explicit method for an object an \code{rtable} with one row with an
appropriate message is returned.
}
\examples{
t_summary(structure(1:5, class = "aaa"), factor(LETTERS[c(1,2,1,1,2)]))

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
