% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.factor}
\alias{t_summary.factor}
\title{Summarize Categorical Data}
\usage{
\method{t_summary}{factor}(
  x,
  col_by,
  col_N = NULL,
  total = NULL,
  useNA = c("ifany", "no", "always"),
  denominator = c("N", "n", "omit"),
  drop_levels = FALSE,
  ...
)
}
\arguments{
\item{x}{(\code{factor}) variable}

\item{col_by}{(\code{logical}, \code{factor} or \code{data.frame})\cr
Contains reference and comparison group information, the
first \code{level} indicates the reference group. See the functions
\code{\link[rtables]{col_by_to_matrix}} and \code{\link[rtables]{by_add_total}}.}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{(\code{character} value)\cr
Denotes column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{useNA}{(\code{logical})\cr
choose whether missing data (NAs) should be displayed as a level.}

\item{denominator}{(one of \code{("n", "N", "omit"})\cr
Value set to \code{"n"} or \code{"N"} are for calculating the level associated percentage.
With option \code{"N"}, the reference population from \code{col_N} is used as
the denominator. With option \code{"n"}, the number of non-missing records from \code{x} is used as
the denominator. If \code{"omit"} is chosen the percentage is omitted.}

\item{drop_levels}{(\code{logical} value)\cr
Whether to drop zero count levels}

\item{...}{(optional)\cr
additional arguments passed on to methods}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
Tabulate the number of non-missing observations, number of observations per level and
percentage.
}
\examples{
# with iris data
t_summary(iris$Species, iris$Species)

library(random.cdisc.data)
ADSL <- cadsl

t_summary(
  x = ADSL$SEX,
  col_by = ADSL$ARM,
  total = "All"
)

t_summary(
  x = ADSL$SEX,
  col_by = ADSL$ARM,
  useNA = "always"
)

ADSL$SEX[1:10] <- NA
t_summary(
  x = ADSL$SEX,
  col_by = ADSL$ARM,
  total = "All",
  denominator = "N",
  useNA = "ifany"
 )

t_summary(
  x = ADSL$SEX,
  col_by = ADSL$ARM,
  total = "All",
  denominator = "n",
  useNA = "no"
)
}
\seealso{
\code{\link{t_summary}},
  \code{\link{t_summary.data.frame}}, \code{\link{t_summary.numeric}},
  \code{\link{t_summary.logical}}, \code{\link{t_summary.Date}},
  \code{\link{t_summary.character}}, \code{\link{t_summary_by}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
