% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.factor}
\alias{t_summary.factor}
\title{Summarize Categorical Data}
\usage{
\method{t_summary}{factor}(x, col_by, col_N = table(col_by), useNA = c("no",
  "ifany", "always"), denominator = c("n", "N"), drop_levels = FALSE, ...)
}
\arguments{
\item{x}{numeric variable}

\item{col_by}{a \code{\link{factor}} of length \code{nrow(x)} that defines
which levels in \code{col_by} define a column. If data should not be split
into columns use the \code{\link{no_by}} function.}

\item{col_N}{The column total for each column. If specified then
\code{\link{col_N}()} can be used on the cell data in order to retrieve the
column total. If \code{NULL} then no header row for the column is
displayed.}

\item{useNA}{choose whether missing data (NAs) should be displayed as a level}

\item{denominator}{either n or N for calculating the level associated
percentage. With option N, the reference population from \code{col_N} is used as the denominator. 
With option n, the number of non-missing records from \code{x} is used as the denominator.}

\item{drop_levels}{boolean whether to drop zero count levels}

\item{...}{not used arguments}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
Tabulate the number of non-missing observations, number of observations per level and
percentage.
}
\examples{

t_summary(iris$Species, iris$Species)


x <-  factor(c("a", "b", "a", "a",  NA,  NA,  NA, "b", "b"))
cb <- factor(c("E", "E", "F", "F", "E", "E", "F", "F", "F"))

t_summary(x, cb)

t_summary(x, cb, useNA = "always")

t_summary(x, cb, denominator = "N")

t_summary(x, cb, useNA = "always",  denominator = "N")

t_summary(x, cb, useNA = "always",  denominator = "n")

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
