% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.logical}
\alias{t_summary.logical}
\title{Summarize Boolean Data}
\usage{
\method{t_summary}{logical}(x, col_by, row.name.TRUE = "TRUE",
  row.name.FALSE = "FALSE", denominator = c("n", "N"), ...)
}
\arguments{
\item{x}{a factor}

\item{col_by}{a \code{\link{factor}} of length \code{nrow(x)} that defines
which levels in \code{col_by} define a column. If data should not be split
into columns use the \code{\link{no_by}} function.}

\item{row.name.TRUE}{character string with row.name for TRUE summary}

\item{row.name.FALSE}{character string with row.name for FALSE summary}

\item{denominator}{either n or N for calculating the level associated
percentage.}

\item{...}{arguments passed on to \code{\link{t_summary.factor}}}
}
\description{
Summarize Boolean Data
}
\examples{
t_summary(
  x = c(TRUE,FALSE,NA,TRUE,FALSE,FALSE,FALSE,TRUE),
  col_by = factor(LETTERS[c(1,1,1,2,2,3,3,2)])
)

t_summary(
  x = c(TRUE,FALSE,NA,TRUE,FALSE,FALSE,FALSE,TRUE),
  col_by = factor(LETTERS[c(1,1,1,2,2,3,3,2)]),
  denominator = "N"
)

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
