% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.logical}
\alias{t_summary.logical}
\title{Summarize Boolean Data}
\usage{
\method{t_summary}{logical}(x, col_by, col_N = table(col_by),
  total = NULL, row_name_true = "TRUE", row_name_false = "FALSE",
  ...)
}
\arguments{
\item{x}{a logical vector}

\item{col_by}{factor with reference and comparison group information, the
first \code{level} indicates the reference group.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default}

\item{total}{character string that will be used as a label for a column with total population. If the levels of
\code{col_by} are the only columns of interest then total should be \code{NULL}.}

\item{row_name_true}{character string with row.name for TRUE summary}

\item{row_name_false}{character string with row.name for FALSE summary}

\item{...}{arguments passed on to \code{\link{t_summary.factor}}}
}
\description{
Boolean data will be converted to factor
}
\examples{
t_summary(
 x = c(TRUE,FALSE,NA,TRUE,FALSE,FALSE,FALSE,TRUE),
 col_by = factor(LETTERS[c(1,1,1,2,2,3,3,2)])
)

library(random.cdisc.data)
ADSL <- radsl(N = 100, seed = 1)

ADSL$AGE[1:10] <- NA

with(ADSL, t_summary(AGE > 65, ARM, useNA = "ifany"))
with(ADSL, t_summary(AGE > 65, ARM, denominator = "N", total = "All", useNA = "ifany",
                     row.name.TRUE = "Baseline Age > 65", row.name.FALSE = "Baseline Age <= 65"))
}
\seealso{
\code{\link{t_summary}},
  \code{\link{t_summary.numeric}}, \code{\link{t_summary.factor}},
  \code{\link{t_summary.data.frame}}, \code{\link{t_summary.Date}},
  \code{\link{t_summary.character}}, \code{\link{t_summary_by}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
