% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.numeric}
\alias{t_summary.numeric}
\title{Summarize Numeric Variables}
\usage{
\method{t_summary}{numeric}(x, col_by, col_N = NULL, total = NULL,
  f_numeric = c("count_n", "mean_sd", "median", "range"), ...)
}
\arguments{
\item{x}{numeric variable}

\item{col_by}{logical, factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vector with the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{total}{string of column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{f_numeric}{a combination of the analysis functions to be evaluated \code{"count_n", "mean_sd", "median",
"q1_q3", "range", "se"}}

\item{...}{arguments passed on to methods}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
Tabulate the number of non-missing observations, mean, standard error, median, and range
for different groups.
}
\examples{
library(dplyr)

# with iris data
t_summary(iris$Sepal.Length, iris$Species)

library(random.cdisc.data)
library(rtables)
ADSL <- cadsl

t_summary(ADSL$AGE, ADSL$ARM)
t_summary(ADSL$AGE, col_by = col_by_to_matrix(ADSL$ARM), total = "All")
t_summary(ADSL$AGE, col_by = ADSL$ARM, total = "All")

ADSL$AGE[1:10] <- NA
t_summary(ADSL$AGE, by_all("All"), col_N = nrow(ADSL))
}
\seealso{
\code{\link{t_summary}},
  \code{\link{t_summary.data.frame}}, \code{\link{t_summary.factor}},
  \code{\link{t_summary.logical}}, \code{\link{t_summary.Date}},
  \code{\link{t_summary.character}}, \code{\link{t_summary_by}}
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
