% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary.numeric}
\alias{t_summary.numeric}
\title{Summarize Numeric Variables}
\usage{
\method{t_summary}{numeric}(x, col_by, ...)
}
\arguments{
\item{x}{numeric variable}

\item{col_by}{a \code{\link{factor}} of length \code{nrow(x)} that defines
which levels in \code{col_by} define a column. If data should not be split
into columns use the \code{\link{no_by}} function.}

\item{...}{arguments passed to \code{FUN}}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
Tabulate the number on non-missing observations, mean, sd, median, and range
for different groups.
}
\examples{

t_summary(iris$Sepal.Length, iris$Species)

}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
