% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary}
\alias{t_summary}
\title{Summarize an Object for Different Groups}
\usage{
t_summary(x, col_by, col_N = get_N(col_by), total = NULL, ...)
}
\arguments{
\item{x}{an object to dispatch on}

\item{col_by}{logical factor or data.frame with reference and comparison group information, the
first \code{level} indicates the reference group. See the functions \code{factor_to_matrix_by}
and \code{add_total_by}.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default
\code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.}

\item{total}{string of column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{...}{arguments passed on to methods}
}
\description{
Summarize an Object for Different Groups
}
\examples{
t_summary(iris$Sepal.Length, iris$Species)

library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

t_summary(ADSL$AGE, ADSL$ARMCD)
t_summary(ADSL[, c("AGE", "SEX", "RACE")], ADSL$ARMCD)
t_summary(ADSL[, c("AGE", "SEX", "RACE")], ADSL$ARMCD, total = "All Patients")

with(ADSL, t_summary(AGE > 65, ARMCD))
}
\seealso{
\code{\link{t_summary.data.frame}},
  \code{\link{t_summary.numeric}}, \code{\link{t_summary.factor}},
  \code{\link{t_summary.logical}}, \code{\link{t_summary.Date}},
  \code{\link{t_summary.character}}, \code{\link{t_summary_by}}
}
