% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary}
\alias{t_summary}
\title{Summarize an Object for Different Groups}
\usage{
t_summary(x, col_by, col_N, ...)
}
\arguments{
\item{x}{an object to dispatch on}

\item{col_by}{a \code{\link{factor}} of length \code{nrow(x)} that defines
which levels in \code{col_by} define a column. If data should not be split
into columns use the \code{\link{no_by}} function.}

\item{col_N}{The column total for each column. If specified then
\code{\link{col_N}()} can be used on the cell data in order to retrieve the
column total. If \code{NULL} then no header row for the column is
displayed.}

\item{...}{arguments passed on to methods}
}
\description{
Summarize an Object for Different Groups
}
\details{
Note that N refers to the number of observations and n refers to the numer of
non-missing observations
}
\examples{
t_summary(iris$Sepal.Length, iris$Species)

library(random.cdisc.data)
ADSL <- radsl(N = 100, seed = 1) 

t_summary(ADSL$AGE, ADSL$ARMCD)
t_summary(ADSL[, c("AGE", "SEX", "RACE")], ADSL$ARMCD)
with(ADSL, t_summary(AGE > 65, ARMCD))

}
