% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary.R
\name{t_summary}
\alias{t_summary}
\title{Summarize an Object for Different Groups}
\usage{
t_summary(x, col_by, col_N = table(col_by), ...)
}
\arguments{
\item{x}{an object to dispatch on}

\item{col_by}{a \code{\link{factor}} of length \code{nrow(x)} that defines
which levels in \code{col_by} define a column. If data should not be split
into columns use the \code{\link{no_by}} function.}

\item{col_N}{The column total for each group that is displayed in the table header with (N=xx).}

\item{...}{arguments passed on to methods}
}
\description{
Summarize an Object for Different Groups
}
\examples{
t_summary(iris$Sepal.Length, iris$Species)

library(random.cdisc.data)
ADSL <- radsl(N = 100, seed = 1) 

t_summary(ADSL$AGE, ADSL$ARMCD)
t_summary(ADSL[, c("AGE", "SEX", "RACE")], ADSL$ARMCD)
with(ADSL, t_summary(AGE > 65, ARMCD))

}
\seealso{
\code{\link{t_summary.numeric}}, \code{\link{t_summary.factor}}, \code{\link{t_summary_by}}
}
