% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary_by.R
\name{t_summary_by}
\alias{t_summary_by}
\title{Summarize an Object for Different Groups with by Variable}
\usage{
t_summary_by(x, by, col_by, col_N, total = NULL, ...)
}
\arguments{
\item{x}{vector}

\item{by}{a \code{factor} of length \code{nrow(x)} with no missing values. The levels of \code{by} define the summary sub-groups in the table.}

\item{col_by}{a \code{factor} of length \code{nrow(x)} with no missing values. The levels of \code{col_by} define the columns in the table.}

\item{col_N}{a \code{table} object with the reference population used for the header of the table. See examples below.}

\item{total}{character string that will be used as a label for a column with pooled total population. If the levels of \code{col_by} are the only columns of interest then total should be \code{NULL}.}

\item{...}{arguments passed on to methods}
}
\description{
Summarize an Object for Different Groups with by Variable
}
\examples{

t_summary_by(
  x = sample(c(TRUE, FALSE), 100, TRUE),
  by = factor(sample(paste("day", 1:5), 100, TRUE)),
  col_by = factor(sample(paste("ARM", LETTERS[1:3]), 100, TRUE)),
  col_N = table(rep(paste("ARM", LETTERS[1:3]), 10)),
  total = "All Patients"
)

t_summary_by(
  x = sample(c(TRUE, FALSE), 100, TRUE),
  by = factor(sample(paste("day", 1:5), 100, TRUE)),
  col_by = factor(sample(paste("ARM", LETTERS[1:3]), 100, TRUE)),
  col_N = table(rep(paste("ARM", LETTERS[1:3]), 10)),
  total = "All Patients",
  denominator = "N"
)


ANL <- data.frame(
 aval = sample(letters[1:5], 100, TRUE),
 avisit = factor(sample(c(
   "Screening", "Day 5", "Day 10"
 ), 100, TRUE), levels = c("Screening", "Day 5", "Day 10")),
 col_by = factor(sample(paste("ARM", LETTERS[1:3]), 100, TRUE))
)
ANL <- var_relabel(ANL, aval = "Analysis Value", avisit = "Analysis Visit" )
levels(ANL$aval) <- c(levels(ANL$aval), "z")

t_summary_by(
  x = ANL$aval,
  by = ANL$avisit,
  col_by = ANL$col_by,
  col_N = c(10, 15, 20)
)

t_summary_by(
  x = ANL$aval,
  by = ANL$avisit,
  col_by = ANL$col_by,
  col_N = c(10, 15, 20),
  total = "All",
  denominator = "N",
  drop_levels = TRUE
)
}
