% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary_variants.R
\name{t_summary_by}
\alias{t_summary_by}
\title{Summarize an Object for Different Groups with by Variable}
\usage{
t_summary_by(x, row_by, col_by, col_N = NULL, total = NULL, ...,
  table_tree = FALSE)
}
\arguments{
\item{x}{vector}

\item{row_by}{a \code{factor} of length \code{nrow(x)} with no missing values. The levels of \code{by} define
the summary sub-groups in the table.}

\item{col_by}{a \code{factor} of length \code{nrow(x)} with no missing values. The levels of \code{col_by}
define the columns in the table.}

\item{col_N}{a vecor with the reference population per level of \code{col_by}. This often by default
\code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.}

\item{total}{string of column name of an added total column using \code{\link[rtables]{by_add_total}} to
\code{col_by} and \code{\link{col_N_add_total}} to \code{col_N}. If \code{NULL} no total column is added.}

\item{...}{arguments passed on to methods}

\item{table_tree}{logical, if \code{FALSE} then a single \code{rtable} gets returned, if \code{TRUE} a
\code{\link{node}} object gets returned with the elementary \code{rtable} objects.}
}
\description{
This is a wrapper around the basic \code{t_summary} functions with the difference that it has a row_by
argument. The row_by argument can either be a "by" object or it can be a "r_by" object in which case
each list item is interpreted recursively as is typical for the \code{terms} item
in \code{t_events_term_grade_id}.
}
\details{
For every level of the variable \code{by} a summary table using \code{\link{t_summary}} will be created.
The individual tables are then stacked together.
}
\examples{
library(dplyr)
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)

t_summary_by(
 x = ADSL$SEX,
 row_by = ADSL$COUNTRY,
 col_by = ADSL$ARMCD,
 total = "All Patients",
 drop_levels = TRUE
)


ADSL$SEX[1:5] <- NA

t_summary_by(
 x = ADSL$SEX,
 row_by = ADSL$COUNTRY,
 col_by = ADSL$ARMCD,
 total = "All Patients",
 drop_levels = TRUE,
 useNA = "ifany"
)

ADSL <- ADSL \%>\% select(STUDYID, USUBJID, ARMCD)

ADQS <- radqs(cached = TRUE)
ADQS_f <- ADQS \%>\%
  dplyr::filter(PARAMCD=="BFIALL")

t_summary_by(
 x = ADQS_f$AVAL,
 row_by = ADQS_f$AVISIT,
 col_by = ADQS_f$ARMCD,
 total = "All Patients",
 col_N = table(ADSL$ARMCD),
)

ADQS_f$AVALCAT1 <- factor(ifelse(ADQS_f$AVAL >= 0, "Positive", "Negative"),
  levels = c("Positive", "Negative"))
ADQS_f <- var_relabel(ADQS_f, AVALCAT1 = "Result" )

t_summary_by(
 x = ADQS_f$AVALCAT1,
 row_by = ADQS_f$AVISIT,
 col_by = ADQS_f$ARMCD,
 total = "All Patients",
 col_N = table(ADSL$ARMCD),
)

ADSL <- radsl(cached = TRUE) #todo: replace by seed version
ADLB <- radlb(cached = TRUE) #todo: replace by seed version

# Recursive case
t_summary_by(
  x = ADLB$AVAL,
  row_by = r_by(ADLB[, c("PARAM", "AVISIT")]),
  col_by = ADLB$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients"
)


# Table Tree
ADSL <- radsl(cached = TRUE)

tbls <- t_summary_by(
 x = ADSL$SEX,
 row_by = ADSL$COUNTRY,
 col_by = ADSL$ARMCD,
 total = "All Patients",
 table_tree = TRUE
)
summary(tbls)
to_rtable(tbls)

# Other examples
ADQS <- radqs(ADSL, seed = 2) \%>\%
  mutate(PARAM = as.factor(PARAM)) \%>\%
  var_relabel(
    PARAM = "Questionnaire Parameter",
    AVISIT = "Visit"
  )

# we suppress warnings because some are NaN (change CHG at screening visit)
suppressWarnings(t_summary_by(
  x = ADQS[,c("AVAL", "CHG")],
  row_by = r_by(ADQS[, c("PARAM", "AVISIT")]),
  col_by = ADQS$ARMCD,
  col_N = table(ADSL$ARMCD),
  total = "All Patients"
))

ADQS_f <- ADQS \%>\%
  dplyr::filter(PARAMCD == "BFIALL") \%>\%
  dplyr::mutate(
    AVALCAT1 = factor(
      ifelse(AVAL >= 50, "Positive", "Negative"),
      levels = c("Positive", "Negative")
    )
  ) \%>\%
  var_relabel(
    AVALCAT1 = "Result",
    AVISIT = "Visit"
  )

t_summary_by(
  x = ADQS_f \%>\% select(c("AVALCAT1", "AVAL")),
  row_by = ADQS_f$AVISIT,
  col_by = ADQS_f$ARMCD,
  col_N = table(ADSL$ARMCD),
  total = "All Patients"
)
}
\seealso{
\code{\link{t_summary}}, \code{\link{t_summary.data.frame}},
  \code{\link{t_summary.numeric}}, \code{\link{t_summary.factor}},
  \code{\link{t_summary.logical}}, \code{\link{t_summary.Date}},
  \code{\link{t_summary.character}}
}
\author{
Jana Stoilova (stoilovs) \email{jana.stoilova@roche.com}
}
