% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summary_by.R
\name{t_summary_by}
\alias{t_summary_by}
\title{Summarize an Object for Different Groups with by Variable}
\usage{
t_summary_by(x, by, col_by, col_N, total = NULL, ...)
}
\arguments{
\item{x}{vector}

\item{by}{a \code{factor} of length \code{nrow(x)} with no missing values. The levels of \code{by} define the summary sub-groups in the table.}

\item{col_by}{a \code{factor} of length \code{nrow(x)} with no missing values. The levels of \code{col_by} define the columns in the table.}

\item{col_N}{a \code{table} object with the reference population used for the header of the table. See examples below.}

\item{total}{character string that will be used as a label for a column with pooled total population. If the levels of \code{col_by} are the only columns of interest then total should be \code{NULL}.}

\item{...}{arguments passed on to methods}
}
\description{
Summarize an Object for Different Groups with by Variable
}
\details{
For every level of the variable \code{by} a summary table using \code{\link{t_summary}} will be created.
The individual tables are then stacked together.
}
\examples{
library(dplyr)
library(random.cdisc.data)

ADSL <- radsl(N = 30, seed = 1)

t_summary_by(
 x = ADSL$SEX,
 by = ADSL$COUNTRY,
 col_by = ADSL$ARMCD,
 col_N = table(ADSL$ARMCD),
 total = "All Patients",
 drop_levels = TRUE
)

ADSL$SEX[1:5] <- NA

t_summary_by(
 x = ADSL$SEX,
 by = ADSL$COUNTRY,
 col_by = ADSL$ARMCD,
 col_N = table(ADSL$ARMCD),
 total = "All Patients",
 drop_levels = TRUE,
 useNA = "ifany"
)

ADSL <- ADSL \%>\% select(STUDYID, USUBJID, ARMCD)

ADQS <- radqs(ADSL, seed = 2)
ADQS_f <- ADQS \%>\%
  dplyr::filter(PARAMCD=="BFIALL")

t_summary_by(
 x = ADQS_f$AVAL,
 by = ADQS_f$AVISIT,
 col_by = ADQS_f$ARMCD,
 col_N = table(ADSL$ARMCD),
 total = "All Patients"
)

ADQS_f$AVALCAT1 <- factor(ifelse(ADQS_f$AVAL >= 0, "Positive", "Negative"),
  levels = c("Positive", "Negative"))
ADQS_f <- var_relabel(ADQS_f, AVALCAT1 = "Result" )

t_summary_by(
 x = ADQS_f$AVALCAT1,
 by = ADQS_f$AVISIT,
 col_by = ADQS_f$ARMCD,
 col_N = table(ADSL$ARMCD),
 total = "All Patients"
)
}
\seealso{
\code{\link{t_summary}}, \code{\link{t_summary.data.frame}},
  \code{\link{t_summary.numeric}}, \code{\link{t_summary.factor}},
  \code{\link{t_summary.logical}}, \code{\link{t_summary.Date}},
  \code{\link{t_summary.character}}
}
\author{
Jana Stoilova (stoilovs) \email{jana.stoilova@roche.com}
}
