% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_tte.R
\name{t_tte}
\alias{t_tte}
\title{Time to Event Table}
\usage{
t_tte(formula, data, event_descr, time_points, time_unit = "month",
  ties = "exact")
}
\arguments{
\item{event_descr}{a factor that partitions the the events into earliest
contributing event}

\item{time_points}{numeric vector with times displayed in the time point
analysis, if \code{NULL} this section of the table will be omitted}

\item{time_unit}{a string with the unit of the \code{tte} argument}

\item{ties}{passed forward to \code{\link[survival]{coxph}}}

\item{tte}{vector with numeric time to event data, currently no \code{NA} allowed}

\item{is_event}{boolean with \code{TRUE} if event and \code{FALSE} if censored}

\item{strata_data}{a data frame with factors that are used for
stratification. If \code{NULL} then no stratified analysis is displayed.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
This is the description of the function
}
\details{
The time to event section is derived from a a kaplan meier estimator for the
model \code{Surv(tte, is_event) ~ col_by}.

The stratified and unstratified analysis is evaluated pair-wise (reference to
comparison) and \code{\link[survival]{survdiff}} is used to get the p-value
whereas \code{\link[survival]{coxph}} is used to calculate the hazard ratio
and confidence interval.

The time point analysis is based on the kaplan meier estimator.
}
\examples{

library(random.cdisc.data)
library(forcats)

ASL <- radam("ASL")
ATE <- radam("ATE", ADSL = ASL)

ATE_f <- subset(ATE, PARAMCD == "OS")
 
ANL <- merge(ASL, ATE_f, all.x =TRUE, all.y = FALSE, by = c("USUBJID", "STUDYID"))


tbl <- t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ANL,
  event_descr = factor(EVNTDESC),
  time_points = c(6, 2000),
  time_unit = "month"
)

tbl


}
