% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_tte.R
\name{t_tte}
\alias{t_tte}
\title{Time to Event Table}
\usage{
t_tte(
  formula,
  data,
  col_N,
  event_descr = NULL,
  time_points = NULL,
  time_unit = "month",
  conf_level = 0.95,
  probs_survfit = c(0.25, 0.75),
  conf_type_survfit = c("plain", "none", "log", "log-log"),
  comparison = control_coxph(),
  table_tree = FALSE
)
}
\arguments{
\item{formula}{(\code{formula})\cr
Specifies \code{\link[survival:Surv]{survival model}}.
The arm variable needs to be wrapped in \code{\link{arm}}. The
\code{\link[survival]{strata}} special will only be used for the stratified analysis. If there is not
\code{\link[survival]{strata}} specification then the stratified analysis is omitted.}

\item{data}{(\code{data.frame})\cr
Contains all the variable that are used in \code{formula}}

\item{col_N}{(\code{integer} vector)\cr
Contains the reference population per level of \code{col_by}. This by default is often
set to \code{get_N(col_by)}. Names of the list are ignored, they are added in the order they appear.
The default may not appropriate and should be modified by the user when \code{col_by} contains
more than one record per subject.}

\item{event_descr}{(\code{factor} vector) that partitions the the events into earliest contributing event. The
variable name can be provided unquoted in which case it is looked up in \code{data} and then the calling
environment.}

\item{time_points}{(\code{numeric} vector) with times displayed in the time point analysis, if \code{NULL} this
section of the table will be omitted}

\item{time_unit}{(\code{character} value) with the unit of the \code{tte} argument}

\item{conf_level}{(\code{numeric} value or named \code{numeric} vector) either a single number or a named vector of
confidence levels where the names are \code{survfit} and \code{ztest}. If confidence level for comparison needs
to be specified, please use \code{conf_level} for \code{comparison}.}

\item{probs_survfit}{(\code{numeric} vector of length two) to specify the quantiles in \code{\link[stats]{quantile}}}

\item{conf_type_survfit}{(one of \code{"plain"} (default), \code{"none"}, \code{"log"}, \code{"log-log"}). Used in
\code{conf.type} in \code{\link[survival]{survfit}}}

\item{comparison}{a (\code{list}) of parameters for comparison details, specified by using the helper function
  \code{\link{control_coxph}}. If \code{NULL}, the comparison table section ("Unstratified Analysis" and
"Stratified Analysis") will be excluded. Some possible parameter options are:
\itemize{
  \item{\code{pval_method}} The method to be passed as argument \code{pval_method} to function
   \code{\link{s_coxph_pairwise}}. Default method is "log-rank", can also be set to "wald" or "likelihood".
   \item{\code{ties}} The method to be passed as argument \code{ties} to \code{\link{coxph}}. Default is "efron",
  can also be set to "breslow" or "exact".
  \item{\code{conf_level}} (\code{numeric} value) passed as argument \code{conf_level} to function
  \code{\link{s_coxph_pairwise}}.
}}

\item{table_tree}{(\code{logical} value)\cr
If \code{FALSE} then a single \code{rtable} gets returned, if \code{TRUE} a
\code{\link{node}} object gets returned with the elementary \code{rtable} objects.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
The time to event table summarizes time to event data with different models
as described in the details section.
}
\details{
The time to event section is derived from a Kaplan-Meier estimator for the
\code{formula} argument with the strata special dropped.

The stratified and unstratified analysis is evaluated pair-wise (reference to
comparison) and \code{\link[survival]{coxph}} is used to get the p-value, calculate the hazard ratio
and confidence interval.

The time point analysis is based on the Kaplan-Meier estimator.
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

ADTTE <- radtte(cached = TRUE)
ADTTE_f <- subset(ADTTE, PARAMCD == "EFS" & BMEASIFL == "Y")

t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM),
  data = ADTTE_f,
  col_N = table(ADSL$ARM)
)

t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM),
  data = ADTTE_f,
  col_N = table(ADSL$ARM),
  event_descr = factor(ADTTE_f$EVNTDESC)
)

t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM),
  data = ADTTE_f,
  col_N = table(ADSL$ARM),
  event_descr = factor(ADTTE_f$EVNTDESC),
  time_points = c(6, 12, 360),
  time_unit = "month",
)

t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM),
  data = ADTTE_f,
  col_N = table(ADSL$ARM),
  event_descr = factor(ADTTE_f$EVNTDESC),
  time_points = c(6, 12, 360),
  conf_level = c(survfit = 0.91, ztest = 0.99)
)

t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f,
  col_N = table(ADSL$ARM),
  event_descr = factor(ADTTE_f$EVNTDESC),
  time_points = c(6, 12, 24),
  time_unit = "month",
  conf_type_survfit = "log-log",
  conf_level = 0.9,
  comparison = control_coxph(pval_method = "wald", ties = "exact", conf_level = 0.99)
)

# Set comparison to NULL to not display comparison
t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM),
  data = ADTTE_f,
  col_N = table(ADSL$ARM),
  event_descr = factor(ADTTE_f$EVNTDESC),
  time_points = c(6, 12),
  comparison = NULL
)

# table_tree = TRUE
tbl3 <- t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f,
  col_N = table(ADSL$ARM),
  event_descr = factor(ADTTE_f$EVNTDESC),
  time_points = c(6, 2000),
  time_unit = "month",
  table_tree = TRUE
)
summary(tbl3)

to_rtable(tbl3)

# when there's single arm
library(dplyr)
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE) \%>\% filter(ARM == "C: Combination") \%>\% droplevels()
ADTTE <- radtte(cached = TRUE)
ADTTE_f <- ADTTE \%>\%
  filter(PARAMCD == "EFS" & BMEASIFL == "Y" & ARM == "C: Combination") \%>\%
  droplevels()

t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM),
  data = ADTTE_f,
  col_N = table(ADSL$ARM),
  time_points = c(6, 12),
  comparison = NULL
)

}
