% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_tte.R
\name{t_tte}
\alias{t_tte}
\title{Time to Event Table}
\usage{
t_tte(tte, is_event, event_descr, col_by, strata_data, time_points,
  time_unit = "month", ties = "exact")
}
\arguments{
\item{tte}{vector with numeric time to event data, currently no \code{NA} allowed}

\item{is_event}{boolean with \code{TRUE} if event and \code{FALSE} if censored}

\item{event_descr}{a factor that partitions the the events into earliest
contributing event}

\item{strata_data}{a data frame with factors that are used for
stratification. If \code{NULL} then no stratified analysis is displayed.}

\item{time_points}{numeric vector with times displayed in the time point
analysis, if \code{NULL} this section of the table will be omitted}

\item{time_unit}{a string with the unit of the \code{tte} argument}

\item{ties}{passed forward to \code{\link[survival]{coxph}}}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
This is the description of the function
}
\details{
The time to event section is derived from a a kaplan meier estimator for the
model \code{Surv(tte, is_event) ~ col_by}.

The stratified and unstratified analysis is evaluated pair-wise (reference to
comparison) and \code{\link[survival]{survdiff}} is used to get the p-value
whereas \code{\link[survival]{coxph}} is used to calculate the hazard ratio
and confidence interval.

The time point analysis is based on the kaplan meier estimator.
}
\examples{

library(random.cdisc.data)
library(forcats)

ASL <- radam("ASL")
ATE <- merge(ASL, radam("ATE", ADSL = ASL))

ATE_f <- subset(ATE, PARAMCD == "OS")

tbl <- t_tte(
  tte = ATE_f$AVAL,
  is_event = ATE_f$CNSR == 0,
  event_descr = factor(ATE_f$EVNTDESC),
  col_by = factor(ATE_f$ARM),
  strata_data = ATE_f[, c('SEX', 'RACE')],
  time_points = c(6, 2000),
  time_unit = "month"
)

tbl

}
