% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_tte.R
\name{t_tte}
\alias{t_tte}
\title{Time to Event Table}
\usage{
t_tte(formula, data, event_descr, time_points, time_unit = "month",
  ties = "exact", table_tree = FALSE)
}
\arguments{
\item{formula}{a survival formula, the arm variable needs to be wrapped in
\code{\link{arm}}. The \code{\link[survival]{strata}} special will only be
used for the stratified analysis. If there is not
\code{\link[survival]{strata}} specification then the stratified analysis
is omitted.}

\item{data}{a \code{data.frame} with all the variable that are used in
\code{formula}}

\item{event_descr}{a factor that partitions the the events into earliest
contributing event. The variable name can be provided unquoted in which
case it is looked up in \code{data} and then the calling environment.}

\item{time_points}{numeric vector with times displayed in the time point
analysis, if \code{NULL} this section of the table will be omitted}

\item{time_unit}{a string with the unit of the \code{tte} argument}

\item{ties}{passed forward to \code{\link[survival]{coxph}}}

\item{table_tree}{boolean, whether the compound table function should return a table_tree object or the elementary
table should be stacked.}
}
\value{
an \code{\link[rtables]{rtable}} object
}
\description{
The time to event table summarizes time to event data with different models
as described in the details section.
}
\details{
The time to event section is derived from a Kaplan-Meier estimator for the
\code{formula} argument with the strata special dropped.

The stratified and unstratified analysis is evaluated pair-wise (reference to
comparison) and \code{\link[survival]{survdiff}} is used to get the p-value
whereas \code{\link[survival]{coxph}} is used to calculate the hazard ratio
and confidence interval.

The time point analysis is based on the Kaplan-Meier estimator.
}
\examples{
\donttest{
library(random.cdisc.data)
ADSL <- cadsl

ADTTE <- cadtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

tbl <- t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f,
  event_descr = factor(EVNTDESC),
  time_points = c(6, 2000),
  time_unit = "month"
)

tbl

Viewer(tbl)

# table_tree = TRUE
tbl <- t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f,
  event_descr = factor(EVNTDESC),
  time_points = c(6, 2000),
  time_unit = "month",
  table_tree = TRUE
)
summary(tbl)
}
}
\seealso{
\code{\link{t_forest_tte}}
}
\author{
Mark Rothe (rothem1)

Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
