% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_pairwise.R
\name{tabulate_pairwise}
\alias{tabulate_pairwise}
\title{Calculate statistics Pairwise with reference level}
\usage{
tabulate_pairwise(x, col_by, FUN, row.name, ...)
}
\arguments{
\item{x}{a vector}

\item{col_by}{a factor first level is taken as reference level}

\item{FUN}{a function with two arguments, first argument is the subset of x
and the second argument is a factor with two levels from col_by}

\item{...}{arguments passed to rrow}
}
\value{
an rtable
}
\description{
Calculate statistics Pairwise with reference level
}
\examples{
tabulate_pairwise(
  x = iris$Sepal.Length,
  col_by = iris$Species,
  FUN = function(xi, col_by_i) diff(tapply(xi, col_by_i, mean)),
  row.name = "diff mean"
)

}
