% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_pairwise.R
\name{tabulate_pairwise}
\alias{tabulate_pairwise}
\title{Calculate statistics Pairwise with reference level}
\usage{
tabulate_pairwise(x, col_by, FUN, row.name, format = NULL, indent = 0)
}
\arguments{
\item{x}{a vector}

\item{col_by}{a factor first level is taken as reference level}

\item{FUN}{a function with two arguments, first argument is the subset of x
and the second argument is a factor with two levels from col_by}

\item{row.name}{string with row name}

\item{format}{a valid format string or a format function for
\code{\link{rcell}}s. To get a list of all valid format strings use
\code{\link{list_rcell_format_labels}}. If format is \code{NULL} then the elements
of a cell get pasted separated by a comma.}

\item{indent}{non-negative integer where 0 means that the row should not be
indented}
}
\value{
an rtable
}
\description{
Calculate statistics Pairwise with reference level
}
\examples{
tern:::tabulate_pairwise(
  x = iris$Sepal.Length,
  col_by = iris$Species,
  FUN = function(xi, col_by_i) diff(tapply(xi, col_by_i, mean)),
  row.name = "diff mean"
)
}
\author{
Adrian Waddell (waddella) \email{adrian.waddell@roche.com}
}
