% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tern-package}
\alias{tern}
\alias{tern-package}
\title{tern Package}
\description{
tern is an analysis package to create tables, listings and graphs to analyze clinical trials data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/tern}
  \item Report bugs at \url{https://github..com/insightsengineering/tern/issues}
}

}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

Authors:
\itemize{
  \item Saibah Chohan \email{saibah.chohan@roche.com}
  \item Francois Collin \email{francois.collin@contractors.roche.com}
  \item Nick Paszty \email{paszty.nicholas@gene.com}
  \item Ting Qi \email{qit3@gene.com}
  \item Daniel Sabanes Bove \email{daniel.sabanes_bove@roche.com}
  \item Jana Stoilova \email{jana.stoilova@roche.com}
  \item Adrian Waddell \email{adrian.waddell@roche.com}
  \item Heng Wang \email{wang.heng@gene.com}
  \item Godwin Yung \email{yungg@gene.com}
  \item Carreras Maximo \email{maximo.carreras@roche.com}
  \item Zhao Rui \email{zhaor8@gene.com}
  \item Wojciech Wojciak \email{wojciech.wojciak@contractors.roche.com}
}

Other contributors:
\itemize{
  \item Roche [copyright holder, funder]
  \item Audrey Boruvka \email{audrey.boruvka@roche.com} [contributor]
}

}
