% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tern-package}
\alias{tern}
\alias{tern-package}
\title{TLG creation functions}
\description{
TLG creation functions
}
\author{
\strong{Maintainer}: Adrian Waddell \email{adrian.waddell@roche.com}

Authors:
\itemize{
  \item Heng Wang
  \item Mark Rothe
  \item Yuyao Song
  \item Chendi Liao
  \item Xiao Mu Yo
}

}
