% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tern-package}
\alias{tern}
\alias{tern-package}
\title{TLG creation functions}
\description{
Functions to create TLG
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pages.github.roche.com/NEST/tern/}
  \item \url{https://github.roche.com/NEST/tern/}
  \item Report bugs at \url{https://github.roche.com/NEST/tern/issues}
}

}
\author{
\strong{Maintainer}: NEST \email{nestcicd@roche.com}

Authors:
\itemize{
  \item Adrian Waddell \email{adrian.waddell@roche.com}
  \item Maximilian Mordig \email{maximilian.mordig@roche.com}
  \item Heng Wang \email{wang.heng@gene.com}
  \item Jennifer Li \email{li.jing@gene.com}
  \item Yuyao Song \email{yuyao.song@roche.com}
  \item Chendi Liao \email{chendi.liao@roche.com}
  \item Jana Stoilova \email{jana.stoilova@roche.com}
  \item Edgar Manukyan \email{edgar.manukyan@roche.com}
  \item Ting Qi \email{qit3@gene.com}
}

Other contributors:
\itemize{
  \item Roche [copyright holder, funder]
  \item Audrey Boruvka \email{audrey.boruvka@roche.com} [contributor]
}

}
