% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tern-package}
\alias{tern}
\alias{tern-package}
\title{tern Package}
\description{
Package to create tables, listings and graphs to analyze clinical trials data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/tern}
  \item Report bugs at \url{https://github.com/insightsengineering/tern/issues}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com}

Authors:
\itemize{
  \item Daniel Sabanés Bové \email{daniel.sabanes_bove@roche.com}
  \item Jana Stoilova \email{jana.stoilova@roche.com}
  \item Heng Wang \email{wang.heng@gene.com}
  \item Francois Collin
  \item Adrian Waddell \email{adrian.waddell@gene.com}
  \item Pawel Rucki \email{pawel.rucki@roche.com}
  \item Chendi Liao \email{chendi.liao@roche.com}
  \item Jennifer Li \email{li.jing@gene.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
