% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tern-package}
\alias{tern}
\alias{tern-package}
\title{tern is an analysis package to create tables, listings and graphs to analyze clinical trials data.}
\description{
Functions to create TLG
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pages.github.roche.com/NEST/tern/}
  \item \url{https://github.roche.com/NEST/tern/}
  \item Report bugs at \url{https://github.roche.com/NEST/tern/issues}
}

}
\author{
\strong{Maintainer}: NEST \email{nestcicd@roche.com}

Authors:
\itemize{
  \item Saibah Chohan \email{saibah.chohan@roche.com}
  \item Francois Collin \email{francois.collin@contractors.roche.com}
  \item Nick Paszty \email{paszty.nicholas@gene.com}
  \item Ting Qi \email{qit3@gene.com}
  \item Daniel Sabanes Bove \email{daniel.sabanes_bove@roche.com}
  \item Jana Stoilova \email{jana.stoilova@roche.com}
  \item Adrian Waddell \email{adrian.waddell@roche.com}
  \item Heng Wang \email{wang.heng@gene.com}
  \item Godwin Yung \email{yungg@gene.com}
}

Other contributors:
\itemize{
  \item Roche [copyright holder, funder]
  \item Audrey Boruvka \email{audrey.boruvka@roche.com} [contributor]
}

}
