% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tern-package}
\alias{tern}
\alias{tern-package}
\title{tern Package}
\description{
Package to create tables, listings and graphs to analyze clinical trials data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/tern/}
  \item \url{https://github.com/pharmaverse/tern/}
  \item Report bugs at \url{https://github.com/pharmaverse/tern/issues}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com} (\href{https://orcid.org/0000-0001-7566-2787}{ORCID})

Authors:
\itemize{
  \item Joe Zhu \email{joe.zhu@roche.com} (\href{https://orcid.org/0000-0001-7566-2787}{ORCID})
  \item Daniel Sabanés Bové \email{daniel.sabanes_bove@rconis.com}
  \item Jana Stoilova \email{stoilova@dnli.com}
  \item Davide Garolini \email{davide.garolini@roche.com} (\href{https://orcid.org/0000-0002-1445-1369}{ORCID})
  \item Emily de la Rua \email{emilydelarua@gmail.com} (\href{https://orcid.org/0009-0000-8738-5561}{ORCID})
  \item Abinaya Yogasekaram \email{ayogasek@gmail.com} (\href{https://orcid.org/0009-0005-2083-1105}{ORCID})
  \item Heng Wang \email{Heng.Wang21@gilead.com}
  \item Francois Collin
  \item Adrian Waddell \email{adrian.waddell@gmail.com}
  \item Pawel Rucki \email{pawel.rucki@roche.com}
  \item Chendi Liao
  \item Jennifer Li
}

Other contributors:
\itemize{
  \item David Munoz Tord \email{david.munoztord@mailbox.org} [contributor]
  \item Wojciech Wojciak \email{wojciech.wojciak@gmail.com} [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
