% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_time_to_event_table.R
\name{time_to_event_table}
\alias{time_to_event_table}
\title{Time to Event Table}
\usage{
time_to_event_table(ATE)
}
\arguments{
\item{ATE}{time to event data.frame}
}
\description{
This is the description of the function
}
\examples{

time_to_event_table(NULL)
}
