% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_event_table.R
\name{time_to_event_table}
\alias{time_to_event_table}
\title{Time to Event Table}
\usage{
time_to_event_table(time_to_event, event, arm, earliest_contributing_event,
  strata_data, time_points, time_unit = "month", ties = "exact")
}
\arguments{
\item{time_to_event}{time to event data}

\item{is_event}{boolean with \code{TRUE} if event and \code{FALSE} if censored}
}
\value{
a named list with one element per row where each element contains the
  data for the row
}
\description{
This is the description of the function
}
\examples{

library(random.cdisc.data)
library(forcats)

ASL <- radam("ASL")
ATE <- merge(ASL, radam("ATE", ADSL = ASL))

ATE_f <- subset(ATE, PARAMCD == "OS")

tbl <- time_to_event_table(
  time_to_event = ATE_f$AVAL,
  event = ATE_f$CNSR == 0,
  earliest_contributing_event = ATE_f$EVNTDESC,
  arm = factor(ATE_f$ARM),
  strata_data = ATE_f \%>\% select(SEX, RACE),
  time_points = 6,
  time_unit = "month"
)

tbl

\dontrun{

library(atezo.data)
library(dplyr)
library(forcats)
 
ASL <- asl(com.roche.cdt30019.go29436.re)
ATE <- ate(com.roche.cdt30019.go29436.re)

tte_tbl_stream <- get_time_to_event_table(com.roche.cdt30019.go29436.re)
Viewer(tte_tbl_stream)

ATE_f <- ATE \%>\%
 filter(ITTFL == "Y", PARAMCD == "OS") \%>\%
 mutate(ARM_ANL = fct_relevel(ARM1, "DUMMY C", "DUMMY B", "DUMMY A"))


tbl <- time_to_event_table(
  time_to_event = ATE_f$AVAL,
  event = ATE_f$CNSR == 0,
  earliest_contributing_event = ATE_f$EVNTDESC,
  arm = ATE_f$ARM_ANL,
  strata_data = ATE_f \%>\% select(SEX, MLIVER, TCICLVL2),
  time_points = 6,
  time_unit = "month"
)

Viewer(tbl)

Viewer(tbl, tte_tbl_stream)

compare_rtables(tbl, tte_tbl_stream, comp.attr = FALSE)

tbl[13,1]
tte_tbl_stream[13,1]
}

}
