% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_time_to_event_table.R
\name{time_to_event_table}
\alias{time_to_event_table}
\title{Time to Event Table}
\usage{
time_to_event_table(time_to_event, event, arm, arm.ref, vars_strat)
}
\arguments{
\item{time_to_event}{time to event data}

\item{is_event}{boolean with \code{TRUE} if event and \code{FALSE} if censored}
}
\value{
a named list with one element per row where each element contains the
  data for the row
}
\description{
This is the description of the function
}
\examples{
\dontrun{
library(atezo.data)
library(dplyr)

ATE <- ate(com.roche.cdt30019.go29436.re)

tbl_stream <- get_time_to_event_table(com.roche.cdt30019.go29436.re)
Viewer(tbl_stream)

ATE_f <- ATE \%>\% filter(PARAMCD == "OS")

tbl <- time_to_event_table(
    time_to_event = ATE_f$AVAL,
    event = (ATE_f$CNSR == 0)*1,
    arm = ATE_f$ARM1,
    arm.ref = "DUMMY C"
)

tbl

compare_rtables(tbl, tbl_stream)


}
}
