% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_time_to_event_table.R
\name{time_to_event_table}
\alias{time_to_event_table}
\title{Time to Event Table}
\usage{
time_to_event_table(time_to_event, event, arm, arm.ref,
  arm.rest = setdiff(arm, arm.ref))
}
\arguments{
\item{time_to_event}{time to event data}

\item{is_event}{boolean with \code{TRUE} if event and \code{FALSE} if censored}
}
\value{
a named list with one element per row where each element contains the
  data for the row
}
\description{
This is the description of the function
}
\examples{
\dontrun{
library(atezo.data)
library(dplyr)

ATE <- ate(com.roche.cdt30019.go29436.re)


ATE_f <- ATE \%>\% filter(PARAM == "OS")

tbl <- time_to_event_table(
    time_to_event = ATE_f$AVAL,
    is_event = (ATE_f$CNSR == "N"),
    arm = ATE_f$ARM,
    arm.ref = "DUMMY A",
    arm.rest = c("DUMMY C", "DUMMY B")
)

tbl_stream <- get_time_to_event_table(com.roche.cdt30019.go29436.re)

compare_table(tbl, tbl_stream)

time_to_event_table(NULL)

}
}
