% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_event_table.R
\name{time_to_event_table}
\alias{time_to_event_table}
\title{Time to Event Table}
\usage{
time_to_event_table(time_to_event, event, arm, big_n_arm, arm.ref, comp1.arm,
  comp2.arm, strata1, strata2, strata3, time_point, desc_event)
}
\arguments{
\item{time_to_event}{time to event data}

\item{is_event}{boolean with \code{TRUE} if event and \code{FALSE} if censored}
}
\value{
a named list with one element per row where each element contains the
  data for the row
}
\description{
This is the description of the function
}
\examples{
\dontrun{

library(teal.oncology)
library(atezo.data)
library(dplyr)
library(survival)
library(broom)
library(stringr)
library(readr)

ASL <- atezo.data::asl(com.roche.cdt30019.go29436.re)
ATE <- atezo.data::ate(com.roche.cdt30019.go29436.re)

tte_tbl_stream <- atezo.data::get_time_to_event_table(com.roche.cdt30019.go29436.re)

ASL_f <- dplyr::filter(ASL,ITTFL == "Y")
        
ATE_f <- dplyr::filter(ATE,ITTFL == "Y" & PARAMCD == "OS")

tbl <- time_to_event_table(
  time_to_event = ATE_f$AVAL,
  event = ifelse(is.na(ATE_f$CNSR),NA,
                 ifelse(ATE_f$CNSR==1,0,1)),
  arm = ATE_f$ARM1,
  big_n_arm = ASL_f$ARM1,
  arm.ref = "DUMMY C",
  comp1.arm = "DUMMY B",
  comp2.arm = "DUMMY A",
  strata1 = as.factor(ATE_f$SEX),
  strata2 = as.factor(ATE_f$MLIVER),
  strata3 = as.factor(ATE_f$TCICLVL2),
  time_point = as.numeric(6),
  desc_event = ATE_f$EVNTDESC
)

teal.oncology::Viewer(tbl)

teal.oncology::Viewer(tte_tbl_stream)

teal.oncology::compare_rtables(tbl, tte_tbl_stream)

}

To do ... 
- create the teal module ... e.g. ...
  https://github.roche.com/Rpackages/teal.oncology/blob/songy24-forestpl/R/tm_forestplot_tte.R#L314

}
