% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_event_table.R
\name{time_to_event_table}
\alias{time_to_event_table}
\title{Time to Event Table}
\usage{
time_to_event_table(time_to_event, event, is_earliest_contr_event_death, arm,
  arm.ref, strata_data)
}
\arguments{
\item{time_to_event}{time to event data}

\item{is_event}{boolean with \code{TRUE} if event and \code{FALSE} if censored}
}
\value{
a named list with one element per row where each element contains the
  data for the row
}
\description{
This is the description of the function
}
\examples{
\dontrun{

library(atezo.data)
library(dplyr)
library(survival)
 
ASL <- atezo.data::asl(com.roche.cdt30019.go29436.re)
ATE <- atezo.data::ate(com.roche.cdt30019.go29436.re)

tte_tbl_stream <- get_time_to_event_table(com.roche.cdt30019.go29436.re)
Viewer(tte_tbl_stream)

ATE_f <- ATE \%>\% filter(ITTFL == "Y", PARAMCD == "OS")

tbl <- time_to_event_table(
  time_to_event = ATE_f$AVAL,
  event = ATE_f$CNSR == 0,
  arm = ATE_f$ARM1,
  is_earliest_contr_event_death = ATE_f$EVNTDESC == "Death",
  arm.ref = "DUMMY C",
  strata_data = ATE_f \%>\% select(SEX, MLIVER, TCICLVL2)
)

Viewer(tbl)

Viewer(tte_tbl_stream)

compare_rtables(tbl, tte_tbl_stream, comp.attr = FALSE)

tbl[13,1]
tte_tbl_stream[13,1]
}

}
