% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_event_table.R
\name{time_to_event_table}
\alias{time_to_event_table}
\title{Time to Event Table}
\usage{
time_to_event_table(time_to_event, event, arm, is_earliest_contr_event_death,
  strata_data, time_points)
}
\arguments{
\item{time_to_event}{time to event data}

\item{is_event}{boolean with \code{TRUE} if event and \code{FALSE} if censored}
}
\value{
a named list with one element per row where each element contains the
  data for the row
}
\description{
This is the description of the function
}
\examples{
\dontrun{

library(atezo.data)
library(dplyr)
library(forcats)
 
ASL <- asl(com.roche.cdt30019.go29436.re)
ATE <- ate(com.roche.cdt30019.go29436.re)

tte_tbl_stream <- get_time_to_event_table(com.roche.cdt30019.go29436.re)
Viewer(tte_tbl_stream)

ATE_f <- ATE \%>\%
 filter(ITTFL == "Y", PARAMCD == "OS") \%>\%
 mutate(ARM_ANL = fct_relevel(ARM1, "DUMMY C", "DUMMY B", "DUMMY A"))


tbl <- time_to_event_table(
  time_to_event = ATE_f$AVAL,
  event = ATE_f$CNSR == 0,
  arm = ATE_f$ARM_ANL,
  is_earliest_contr_event_death = ATE_f$EVNTDESC == "Death",
  strata_data = ATE_f \%>\% select(SEX, MLIVER, TCICLVL2),
  time_points = setNames(6, paste(6, "Months"))
)

Viewer(tbl)

Viewer(tbl, tte_tbl_stream)

compare_rtables(tbl, tte_tbl_stream, comp.attr = FALSE)

tbl[13,1]
tte_tbl_stream[13,1]
}

}
