% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_chgfbl_table.R
\name{tm_chgfbl_table}
\alias{tm_chgfbl_table}
\title{Change from Baseline Table Teal Module}
\usage{
tm_chgfbl_table(label, paramcd, paramcd_choices = paramcd, arm_var = "ARM",
  arm_var_choices = arm_var, pre_output = NULL, post_output = NULL)
}
\arguments{
\item{label}{full name label of module}

\item{paramcd}{filter the rows in AQS given the paramcd value}

\item{paramcd_choices}{choices of possible poaramcd}

\item{arm_var}{selected variable to use as arms}

\item{arm_var_choices}{choices of arm variables}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\value{
an \code{\link[teal]{module}} object
}
\description{
This module produces a summary table for test results and change
from baseline by visit, that matches STREAM template LBT01
}
\details{
Package \code{forcats} used to re-format arm data into leveled factors.
}
\examples{
 

\\donotrun{
library(atezo.data)
library(dplyr)
library(tidyr)
library(teal.oncology)
library(forcats)

ASL <- asl(com.roche.cdpt7722.wo29637.rl)
AQS <- aqs(com.roche.cdpt7722.wo29637.rl)

options(teal_logging = FALSE)


x <- teal::init(
  data = list(ASL = ASL, AQS = AQS),
  modules = root_modules(
    #tm_variable_browser(),
    #tm_data_table(),
    tm_chgfbl_table(
       label = "Change from Baseline Table",
       paramcd = "FATIGI",
       paramcd_choices = unique(AQS$PARAMCD),
       arm_var = "ARM",
       arm_var_choices = c("ARM", "ARMCD", "AGE65", "ICC"),
   ))
)   
shinyApp(x$ui, x$server) 

  
} 
}
\author{
Chendi Liao (liaoc10), \email{chendi.liao@roche.com}
}
