% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_demographic_table.R
\name{tm_demographic_table}
\alias{tm_demographic_table}
\title{Demographic Table Teal Module}
\usage{
tm_demographic_table(label, group_by_vars,
  group_by_vars_choices = group_by_vars, pre_output = NULL,
  post_output = NULL)
}
\description{
Demographic Table Teal Module
}
\examples{
\\donotrun{
library(atezo.data)
library(dplyr)
library(survival)

ASL <- asl(com.roche.cdt30019.go29436.re)

x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_variable_browser(),
    tm_data_table(),
    tm_demographic_table(
       label = "Demographic Table",
       group_by_vars = toupper(c("sex", "mliver", "tciclvl2", "bage", "age4cat",
            "ethnic", "race", "bwt", "tobhx", "hist", "EGFRMUT",
            "alkmut", "krasmut", "becog"))
   )
  )
)   
shinyApp(x$ui, x$server) 

  
} 


}
