% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot_rsp.R
\name{tm_forest_response}
\alias{tm_forest_response}
\title{Forest Response Plot teal module}
\usage{
tm_forest_response(label, paramcd = "OVRSPI", paramcd_choices = paramcd,
  arm_var = "ARM", arm_var_choices = arm_var, subgroup_var,
  subgroup_var_choices = subgroup_var, plot_height = c(700, 200, 2000),
  cex = 1.5,
  pre_output = helpText("graph needs to be of a certain width to be displayed"),
  post_output = NULL)
}
\arguments{
\item{label}{a character string displayed as module label}

\item{paramcd}{default response type from PARAMCD}

\item{paramcd_choices}{a vector of possible \code{paramcd}}

\item{arm_var}{default variable name used as the arm variable}

\item{arm_var_choices}{a character vector for the choices of \code{arm_var}}

\item{subgroup_var}{a vector of variable names used as the default subgroups}

\item{subgroup_var_choices}{a vector of variable names to choose the \code{subgroup_var} from}

\item{plot_height}{height of the forest plot}

\item{cex}{multiplier applied to overall fontsize}

\item{pre_output}{text displayed at the top of the plot}

\item{post_output}{text displayed at the bottom of the plot}
}
\description{
Forest Response Plot teal module
}
\examples{
  

\\donotrun{
library(teal.oncology)
library(atezo.data)
library(dplyr)
library(forcats)

ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)
 
ASL$BAGED <- ifelse(ASL$BAGE <= median(ASL$BAGE), "<=median", ">median")

x <- teal::init(
  data = list(ASL = ASL, ARS = ARS),
  modules = root_modules(
    tm_forest_response(
       label = "Forest Response",
       paramcd = "OVRSPI",
       paramcd_choices = c("BESRSPI","LSTASDI","MBESRSPI","MLSTASDI","OVRSPI"),
       plot_height = c(600, 200, 2000),
       subgroup_var = c("BAGED", "SEX", "BECOG"),
       subgroup_var_choices = names(ASL),
       arm_var = "ARM",
       arm_var_choices = c("ARM", "ARMCD", "ACTARM")
   )
  )
)   
shinyApp(x$ui, x$server) 

  
} 
}
\author{
Yuyao Song (songy24), \email{yuyao.song@roche.com}
}
