% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot_rsp.R
\name{tm_forest_response}
\alias{tm_forest_response}
\title{Forest Response Plot teal module}
\usage{
tm_forest_response(label, paramcd, paramcd_choices = paramcd, arm.ref,
  arm.ref_choices = arm.ref, arm.comp, arm.comp_choices = arm.comp,
  subgroup_var, subgroup_var_choices = subgroup_var, response,
  response_choices = response, plot_height = c(600, 200, 2000),
  pre_output = NULL, post_output = NULL)
}
\description{
Forest Response Plot teal module
}
\examples{
 

\\donotrun{
library(atezo.data)
library(dplyr)
library(survival)

ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

arms <- unique(ASL$ARM)
x <- teal::init(
  data = list(ASL = ASL, ARS = ARS),
  modules = root_modules(
    tm_variable_browser(),
    tm_data_table(),
    tm_forest_response(
       label = "Forest Response",
       paramcd = "OVRSPI",
       paramcd_choices = unique(ARS$PARAMCD),
       response = "CR",
       response_choices = unique(ARS$AVALC),
       subgroup_var = c("BAGED", "SEX", "BECOG"),
       arm.ref = arms[1],
       arm.ref_choices = arms,
       arm.comp = arms[-1],
       arm.comp_choices = arms
   )
  )
)   
shinyApp(x$ui, x$server) 

  
} 
}
