% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot_tte.R
\name{tm_forest_survival}
\alias{tm_forest_survival}
\title{Forest Survival Plot Teal Module}
\usage{
tm_forest_survival(label, arm_var = "ARM", arm_var_choices = arm_var,
  subgroup_var, subgroup_var_choices = subgroup_var, paramcd = "OS",
  paramcd_choices = paramcd, plot_height = c(600, 200, 2000), cex = 1.5,
  pre_output = helpText("graph needs to be of a certain width to be displayed"),
  post_output = NULL)
}
\arguments{
\item{label}{a character string displayed as module label}

\item{arm_var}{default variable name used as the arm variable}

\item{arm_var_choices}{a character vector for the choices of \code{arm_var}}

\item{subgroup_var}{a vector of variable names used as the default subgroups}

\item{paramcd}{default enpoint from PARAMCD}

\item{paramcd_choices}{a vector of possible \code{paramcd}}

\item{plot_height}{height of the forest plot}

\item{cex}{multiplier applied to overall fontsize}

\item{pre_output}{text displayed at the top of the plot}

\item{post_output}{text displayed at the bottom of the plot}

\item{subgroup_var_choicesa}{a vector of variable names to choose the \code{subgroup_var} from}
}
\description{
Forest Survival Plot Teal Module
}
\examples{
 

\\donotrun{
library(atezo.data)
library(teal.oncology)
library(dplyr)
library(forcats)

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)
ASL$BAGED <- ifelse(ASL$BAGE <= median(ASL$BAGE), "<=median", ">median")

options(teal_logging = FALSE)

arms <- unique(ASL$ARM)
x <- teal::init(
  data = list(ASL = ASL, ATE = ATE),
  modules = root_modules(
    tm_forest_survival(
       label = "Forest Survival",
       paramcd = "OS",
       paramcd_choices = unique(ATE$PARAMCD),
       subgroup_var = c("BAGED", "SEX", "BECOG"),
       arm_var = "ARM",
       arm_var_choices = c("ARM", "ARMCD", "ACTARM")
   )
  )
)   
shinyApp(x$ui, x$server) 

  
} 
}
\author{
Yuyao Song (songy24), \email{yuyao.song@roche.com}
}
