% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot.R
\name{tm_forest_survival}
\alias{tm_forest_survival}
\title{Forest Survival Plot teal module}
\usage{
tm_forest_survival(label, dataname, plot_height = c(600, 200, 2000),
  pre_output = NULL, post_output = NULL)
}
\description{
Forest Survival Plot teal module
}
\examples{
 

\\donotrun{
library(atezo.data)
library(dplyr)
library(survival)

'\%needs\%' <- teal.oncology:::'\%needs\%'

surv_tbl_stream <- get_forest_survival_table(com.roche.cdt30019.go29436.re)
Viewer(surv_tbl_stream )

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)


x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_variable_browser(),
    tm_data_table(),
    tm_venn2("Venn Diagram", "ASL", "B1", "B2", var_biomarkers, var_biomarkers)
  )
)   
shinyApp(x$ui, x$server) 

  
} 
}
