% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_tte.R
\name{tm_forest_survival}
\alias{tm_forest_survival}
\title{Forest Survival Plot teal module}
\usage{
tm_forest_survival(label, arm.ref, arm.ref_choices = arm.ref, arm.comp,
  arm.comp_choices = arm.comp, baseline_risk,
  baseline_risk_choices = baseline_risk, endpoint = "OS",
  endpoint_choices = "OS", plot_height = c(600, 200, 2000),
  pre_output = NULL, post_output = NULL)
}
\description{
Forest Survival Plot teal module
}
\examples{
 

\\donotrun{
library(atezo.data)
library(dplyr)
library(survival)

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

arms <- unique(ASL$ARM)
x <- teal::init(
  data = list(ASL = ASL, ATE = ATE),
  modules = root_modules(
    tm_variable_browser(),
    tm_data_table(),
    tm_forest_survival(
       label = "Forest Survival",
       endpoint = "OS",
       endpoint_choices = unique(ATE$PARAMCD),
       baseline_risk = c("BAGED", "SEX", "BECOG"),
       arm.ref = arms[1],
       arm.ref_choices = arms,
       arm.comp = arms[-1],
       arm.comp_choices = arms
   )
  )
)   
shinyApp(x$ui, x$server) 

  
} 
}
