% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_forestplot_tte.R
\name{tm_forest_survival}
\alias{tm_forest_survival}
\title{Forest Survival Plot teal module}
\usage{
tm_forest_survival(label, arm_var = "ARM", arm_var_choices = arm_var,
  subgroup_var, subgroup_var_choices = subgroup_var, paramcd = "OS",
  paramcd_choices = paramcd, plot_height = c(600, 200, 2000),
  pre_output = helpText("graph needs to be of a certain width to be displayed"),
  post_output = NULL)
}
\description{
Forest Survival Plot teal module
}
\examples{
 

\\donotrun{
library(atezo.data)
library(teal.oncology)
library(dplyr)
library(forcats)

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)
ASL$BAGED <- ifelse(ASL$BAGE <= median(ASL$BAGE), "<=median", ">median")

options(teal_logging = FALSE)

arms <- unique(ASL$ARM)
x <- teal::init(
  data = list(ASL = ASL, ATE = ATE),
  modules = root_modules(
    tm_forest_survival(
       label = "Forest Survival",
       paramcd = "OS",
       paramcd_choices = unique(ATE$PARAMCD),
       subgroup_var = c("BAGED", "SEX", "BECOG"),
       arm_var = "ARM",
       arm_var_choices = c("ARM", "ARMCD", "ACTARM")
   )
  )
)   
shinyApp(x$ui, x$server) 

  
} 
}
