% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_time_to_event_table.R
\name{tm_time_to_event_table}
\alias{tm_time_to_event_table}
\title{Time To Event Table teal module}
\usage{
tm_time_to_event_table(label, arm.ref, arm.ref_choices = arm.ref, arm.comp,
  arm.comp_choices = arm.comp, strata_var, strata_var_choices = strata_var,
  endpoint = "OS", endpoint_choices = "OS", pre_output = NULL,
  post_output = NULL)
}
\description{
Time To Event Table teal module
}
\examples{
 

\\donotrun{
library(atezo.data)
library(dplyr)
library(survival)

ATE <- ate(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

arms <- unique(ASL$ARM)
x <- teal::init(
  data = list(ASL = ASL, ATE = ATE),
  modules = root_modules(
    tm_variable_browser(),
    tm_data_table(),
    tm_time_to_event_table(
       label = "Time To Event Table",
       endpoint = "OS",
       endpoint_choices = unique(ATE$PARAMCD),
       strata_var = c("SEX", "MLIVER", "TCICLVL2"),
       arm.ref = arms[1],
       arm.ref_choices = arms,
       arm.comp = arms[-1],
       arm.comp_choices = arms
   )
  )
)   
shinyApp(x$ui, x$server) 

  
} 
}
