% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_trees.R
\name{to_rtable}
\alias{to_rtable}
\alias{to_rtable,node-method}
\alias{to_rtable,rtable-method}
\title{Object: convert}
\usage{
to_rtable(x)

\S4method{to_rtable}{node}(x)

\S4method{to_rtable}{rtable}(x)
}
\arguments{
\item{x}{object}
}
\description{
Converts the object to an \code{rtable}

Specific to nodes

Trivial (identity)
}
\examples{
n11 <- node(name = "A", content = array(c(1:6), dim = c(2,3)), children = list())
n12 <- node(name = "B", content = array(c(1:6), dim = c(2,3)), children = list())
n13 <- node(name = "C", content = array(c(1:6), dim = c(2,3)), children = list())
n2 <- node(name = "D", content = c(1:3), children = list(n11, n12, n13))

# to_rtable(n2) # not working because not a tree of rtables

to_rtable(node(
  invisible_node_name("received_treatment"),
  content = t_summary(structure(1:5, class = "aaa"), factor(LETTERS[c(1,2,1,1,2)]))
))
}
