% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rtables.R
\name{to_string_matrix}
\alias{to_string_matrix}
\title{Convert Table into Matrix of Strings}
\usage{
to_string_matrix(x, with_spaces = FALSE, print_txt_to_copy = FALSE)
}
\arguments{
\item{x}{\code{rtables} table.}

\item{with_spaces}{Should the tested table keep the indentation and other relevant spaces?}

\item{print_txt_to_copy}{Utility to have a way to copy the input table directly
into the expected variable instead of copying it too manually.}
}
\value{
A \code{matrix} of \code{string}s.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Helper function to use mostly within tests. \code{with_spaces}parameter allows
to test not only for content but also indentation and table structure.
\code{print_txt_to_copy} instead facilitate the testing development by returning a well
formatted text that needs only to be copied and pasted in the expected output.
}
