% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{to_string_matrix}
\alias{to_string_matrix}
\title{Convert Table into Matrix of Strings}
\usage{
to_string_matrix(x)
}
\arguments{
\item{x}{the table}
}
\value{
A matrix of strings
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Helper function to use mostly within tests.
}
