% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{to_string_matrix}
\alias{to_string_matrix}
\title{Convert a table into a matrix of strings.}
\usage{
to_string_matrix(x)
}
\arguments{
\item{x}{the table}
}
\value{
matrix of strings
}
\description{
Helper function to use mostly within tests.
}
