% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_aggregation_functions.R
\name{ttest_ci_one_arm}
\alias{ttest_ci_one_arm}
\title{Confidence level for mean estimate (one-arm t-test)}
\usage{
ttest_ci_one_arm(x, conf_level = 0.95, ...)
}
\arguments{
\item{x}{(\code{numeric} vector)}

\item{conf_level}{(\code{numeric} value)
Specifies confidence level. Must be greater than 0 and less than 1.}

\item{...}{(optional)
Other arguments from \code{\link{stats}{t.test}}}
}
\value{
a vector of 2 numeric numbers representing the lower and upper bounds
}
\description{
The calculation of confidence interval for an estimation of population mean
 will use \code{qt} function with degree of freedom of \code{n - 1}
}
\examples{
ttest_ci_one_arm(c(1,2,3,4,5,6,7,8, NA), conf_level = 0.95, mu = 1, alternative = "less")
}
