% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_aggregation_functions.R
\name{ttest_ci_one_arm}
\alias{ttest_ci_one_arm}
\title{Function: CI}
\usage{
ttest_ci_one_arm(x, conf_level = 0.95)
}
\arguments{
\item{x}{(\code{numeric} vector)}

\item{conf_level}{confidence level of the interval.}
}
\value{
a vector of 2 numeric numbers representing the lower and upper bounds.
Missing values are discarded before the CI calculation.
If not at least 2 \code{x} values are available, then an empty \code{rcell} object is returned.
}
\description{
Confidence level for mean estimate (based on one-arm t-test statistic)
}
\details{
The calculation of a two-sided confidence interval for an estimation of population mean
will use the \code{qt} function with \code{n - 1} degrees of freedom.
}
\examples{
ttest_ci_one_arm(c(1, 2, 3, 4, NA), conf_level = 0.95)
ttest_ci_one_arm(as.numeric(c(NA, NA)))
}
