% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_aggregation_functions.R
\name{ttest_two_arm}
\alias{ttest_two_arm}
\title{ttest_two_arm Performs 2-arm t-test between comparison group against the reference group.}
\usage{
ttest_two_arm(x, col_by, conf.level = 0.95, ...)
}
\arguments{
\item{x}{numeric vector}

\item{col_by}{a factor with 2 or more levels. First level is taken as reference level}

\item{conf.level}{an number indicating confidence level. Default is 0.95
Must be greater than 0 and less than 1.}

\item{...}{Other arguments from \code{\link{stats}{t.test}}}
}
\value{
list of
  lcl, ucl, diff, se, tvalue, pvalue
}
\description{
ttest_two_arm Performs 2-arm t-test between comparison group against the reference group.
}
\examples{
ttest_two_arm(x = c(0,1,2,3,4,5,6,7,8,NA),
              col_by = factor(c("A", "A", "A", "A", "A", "B", "B", "B", "B", "B")),
              var.equal = TRUE)
}
