% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_factor.R
\name{ungroup_stats}
\alias{ungroup_stats}
\title{Ungroup Non-Numeric Statistics}
\usage{
ungroup_stats(x, .formats, .labels, .indent_mods, .in_ref_col = FALSE)
}
\arguments{
\item{x}{(\verb{named list} of \code{numeric})\cr list of numeric statistics containing the statistics to ungroup.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels. Defaults to 0, which corresponds to the
unmodified default behavior. Can be negative.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}
}
\value{
A \code{list} with modified elements \code{x}, \code{.formats}, \code{.labels}, and \code{.indent_mods}.
}
\description{
Ungroups grouped non-numeric statistics within input vectors \code{.formats}, \code{.labels}, and \code{.indent_mods}.
}
\seealso{
\code{\link[=a_summary]{a_summary()}} which uses this function internally.
}
\keyword{internal}
