% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniroot_catch_na}
\alias{uniroot_catch_na}
\title{Find a root while returning NA for missing function values}
\usage{
uniroot_catch_na(f, interval, ...)
}
\arguments{
\item{f}{(\code{function})\cr function for which the root is sought.}

\item{interval}{(\code{numeric(2)})\cr end points of the interval to be searched.}

\item{...}{further arguments passed to \code{\link[stats:uniroot]{stats::uniroot()}}.}
}
\value{
A numeric scalar containing the found root, or \code{NA_real_} if \code{f} returns
\code{NA} at either endpoint of \code{interval}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A wrapper around \code{\link[stats:uniroot]{stats::uniroot()}} that returns \code{NA_real_} when \code{f} is \code{NA}
at either end of \code{interval}.
}
\seealso{
\code{\link[stats:uniroot]{stats::uniroot()}}
}
\keyword{internal}
