% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{use_stuff}
\alias{use_stuff}
\title{Dummy Function to use Currently Unused Functions}
\usage{
use_stuff(...)
}
\arguments{
\item{...}{dummy arguments}
}
\description{
The function was introduced at the start of the rewriting of tern,
and will be deleted before the first release.
}
