% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_split_funs.R
\name{utils_split_funs}
\alias{utils_split_funs}
\alias{ref_group_last}
\alias{keep_level_order}
\title{Custom Split Functions}
\usage{
ref_group_last(df, spl, vals = NULL, labels = NULL, trim = FALSE, .spl_context)

keep_level_order(
  df,
  spl,
  vals = NULL,
  labels = NULL,
  trim = FALSE,
  .spl_context
)
}
\value{
\itemize{
\item \code{ref_group_last} returns an utility function that puts the reference group
as last and needs to be assigned to \code{split_fun}.
}

\itemize{
\item \code{keep_level_order} returns an utility function that keeps the original levels'.
It needs to be assigned to \code{split_fun}.
}
}
\description{
Collection of useful functions that are expanding on the core list of functions
provided by \code{rtables}. See \code{?rtables::custom_split_funs} and \code{\link[rtables:make_split_fun]{rtables::make_split_fun()}}
for more information on how to make a custom split function. All these functions
work with \code{\link[=split_rows_by]{split_rows_by()}} argument \code{split_fun} to modify the way the split
happens.
}
\section{Functions}{
\itemize{
\item \code{ref_group_last()}: split function to place reference group facet last
during post-processing stage.

\item \code{keep_level_order()}: split function to keep original order of factor
levels in the split.

}}
\examples{
library(dplyr)

dat <- data.frame(
  x = factor(letters[1:5], levels = letters[5:1]),
  y = 1:5
)
# ref_group_last

# With rtables layout functions
basic_table() \%>\%
  split_cols_by("x", ref_group = "c", split_fun = ref_group_last) \%>\%
  analyze("y") \%>\%
  build_table(dat)

# With tern layout funcitons
adtte_f <- tern_ex_adtte \%>\%
  filter(PARAMCD == "OS") \%>\%
  mutate(
    AVAL = day2month(AVAL),
    is_event = CNSR == 0
  )

basic_table() \%>\%
  split_cols_by(var = "ARMCD", ref_group = "ARM B", split_fun = ref_group_last) \%>\%
  add_colcounts() \%>\%
  surv_time(
    vars = "AVAL",
    var_labels = "Survival Time (Months)",
    is_event = "is_event",
  ) \%>\%
  build_table(df = adtte_f)

# keep_level_order

}
\seealso{
\code{\link[rtables:make_split_fun]{rtables::make_split_fun()}}
}
