% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{var_labels<-}
\alias{var_labels<-}
\title{Sets column labels of a \code{data.frame}.}
\usage{
var_labels(x) <- value
}
\arguments{
\item{x}{(\code{data.frame}) object with columns}

\item{value}{(\code{character}) labels}
}
\value{
\code{x}
}
\description{
Sets column labels of a \code{data.frame}.
}
\examples{
var_labels(iris) <- colnames(iris)
}
\keyword{internal}
