% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables.R
\name{var_labels}
\alias{var_labels}
\title{Extracts dataset and variable labels from a dataset.}
\usage{
var_labels(data, fill = FALSE)
}
\arguments{
\item{data}{(\code{data.frame}) table to extract the labels from}

\item{fill}{(\code{logical(1)}) if \code{TRUE}, the function will return variable names for columns with non-existent labels;
otherwise will return \code{NA} for them}
}
\value{
column labels
}
\description{
Extracts dataset and variable labels from a dataset.
}
\keyword{internal}
