% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_venn.R
\name{venn2}
\alias{venn2}
\title{Create a Venn Diagram Plot with 2 groups}
\usage{
venn2(x, y, xlab, ylab)
}
\arguments{
\item{x}{boolean has biomarker or not}

\item{y}{boolean has biomarker of not}
}
\value{
list with absolute and percentage cross table
}
\description{
Create a Venn Diagram Plot with 2 groups
}
\examples{

n <- 100
tGE <- sample(c(TRUE, FALSE), n, replace=TRUE, prob = c(.2, .8))
IHC <- sample(c(TRUE, FALSE), n, replace=TRUE, prob = c(.6, .4))

\dontrun{
library(atezo.data)
ASL <- asl(com.roche.cdpt7722.wo29637.rl)

IC <- ASL$IC \%in\% c("2", "3")
TC <- ASL$TC \%in\% c(2, 3)

x <- venn2(x = IC, y = TC)
plot(x)

venn2(x = IC, y = TC, "biomarker IT", "biamrker TC")

}


x = c(rep(T,5), rep(F,3), rep(T, 4), rep(F, 9))
y = c(rep(T,5), rep(F,3), rep(F, 4), rep(T, 9))
table(x,y)
y <- venn2(x,y, "X", "Y")
plot(y)


# if too few then table is plotted
x = c(F, F,F,F, T,T)
y = c(F, T,T,T, F,F)
table(x,y)
y <- venn2(x, y, "X", "Y")
plot(y)

}
