% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_text.R
\name{wrap_text}
\alias{wrap_text}
\title{Split String}
\usage{
wrap_text(
  txt,
  width = grid::convertWidth(grid::unit(1, "npc"), "inch", TRUE),
  gp = grid::gpar(),
  collapse = NULL
)
}
\arguments{
\item{txt}{string (or vector of strins) to be split in multiple lines, not that
\code{\\n} is also split into to lines}

\item{width}{max with of string, by default the width of the current viewport}

\item{gp}{graphical parameters for text}
}
\value{
e vector with the new strings
}
\description{
Return splitted text that fits onto page width.
}
\examples{
text <- "This is a test with many words and more"

# Internal function - wrap_text
\dontrun{
tern:::wrap_text(txt = text, width = grid::unit(4, "cm"), collapse = "\n")
tern:::wrap_text(txt = text, width = grid::unit(5, "cm"), collapse = "\n")
}

}
\author{
Adrian Waddell
}
\keyword{internal}
